(Deprecated) Fit the MMRM Model

(Deprecated) Fit the MMRM Model

a_mmrm(
  data,
  formula = AVAL ~ ARM + AVISIT + ARM * VISIT,
  id_var = "USUBJID",
  arm_var = "ARM",
  visit_var = "AVISIT",
  mode = c("df.error", "auto", "boot-satterthwaite"),
  conf_level = 0.95,
  weights_emmeans = "proportional",
  corStruct = NULL
)

Arguments

data
a data.frame with all the variables specified in formula. Records with missing values in any independent variables will be excluded.

formula
a gls formula.

id_var
a character describing the variable name used as subject IDs, "USUBJID" by default.

arm_var
a character describing the variable name for arm, "ARM" by default. The arm variable must be factor in data

visit_var
a character describing the variable name for visit, "AVISIT" by default. This variable must be factor in data.

mode
algorithm for degree of freedom: auto, df.error or boot-satterthwaite.

conf_level
confidence level. Must be number greater than 0 and less than 1.

weights_emmeans
argument from emmeans, "proportional" by default.

corStruct
NULL by default or a string with the name of corClasses.

Value

a dataframe with MMRM results

Warning

This function has been deprecated. Please use s_mmrm instead.

Examples

library(random.cdisc.data)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

ADSL <- radsl(cached = TRUE)
ADQS <- radqs(cached = TRUE)
ADQS_f <- ADQS %>%
  dplyr::filter(PARAMCD=="FKSI-FWB" & !AVISIT %in% c("BASELINE")) %>%
  droplevels() %>%
  dplyr::mutate(ARM = factor(ARM, levels = c("B: Placebo", "A: Drug X", "C: Combination"))) %>%
  dplyr::mutate(AVISITN = rank(AVISITN) %>% as.factor() %>% as.numeric() %>% as.factor())

if (FALSE) {
mmrm_results <- a_mmrm(
  data = ADQS_f,
  formula = AVAL ~ ARM + AVISIT + STRATA1 + BMRKR2 + ARM * AVISIT,
  id_var = "USUBJID",
  arm_var = "ARM",
  visit_var = "AVISIT",
  mode = "boot-satterthwaite",
  conf_level = 0.95,
  weights_emmeans = "proportional",
  corStruct = "corSymm"
)

names(mmrm_results)

mmrm_results["contrast"]
mmrm_results["estimate"]
}