Patient Counts with the Most Extreme Post-Baseline Toxicity Grade by Baseline NCI-CTCAE Grade

Patient count and fraction for laboratory test shift table, highest NCI-CTCAE grade post-baseline by baseline NCI-CTCAE grade, according to user-defined grouping_list, is generated. Per GDSR standard:

for abnormality in the high direction, grouping_list should be defined as:
list("Not High" = c(0, -1, -2, -3, -4), "1" = 1, "2" = 2, "3" = 3, "4" = 4, "Missing" = "<Missing>")

for abnormality in the low direction, grouping_list should be defined as:
list("Not Low" = c(0, 1, 2, 3, 4), "1" = -1, "2" = -2, "3" = -3, "4" = -4, "Missing" = "<Missing>")

For a particular baseline NCI-CTCAE grade group that is in the grouping_list, for example, "Not High", assuming abnormality is in the high direction, the total number of patients who are "Not High" at baseline is counted, that is the number of patients who have a baseline NCI-CTCAE grades of 0, -1, -2, -3 or -4. In addition, for these patients who are "Not High" at baseline, the counts and fractions of patients with post-baseline NCI-CTCAE grades in each of the groups in group_listing are produced, i.e., "Not High", "1", "2", "3", "4" and "Missing". Note the order of grouping_list controls the output order.

Care must be taken when specifying the grouping_list. Each element in the grouping_list must be specified using a = sign; on the left of the = sign is the name of the group; on the right of = sign is NCI-CTCAE grade(s) to be included in that group, i.e., "NAME" = c(1, 2, 3, 4).

User may define custom functions for extracting the worst lab grades. At the minimum, the output df should contain relevant subject-level variables derived from ADSL (e.g. USUBJID, ACTARM) and analysis variables sourced from ADLB:PARAM or PARAMCD, ATOXGR and BTOXGR. One worst lab grade per patient per lab is expected for each of ATOXGR and BTOXGR unless summarizing the worst grade by visit in which case a visit variable such as AVISIT would be necessary.

h_adsl_adlb_merge_using_worst_flag(
  adsl,
  adlb,
  worst_flag = c(WGRHIFL = "Y"),
  by_visit = FALSE,
  no_fillin_visits = c("SCREENING", "BASELINE")
)

h_group_counter(df, id, .var, grouping_list)

s_count_abnormal_by_worst_grade_by_baseline(
  df,
  .var = "ATOXGR",
  grouping_list = list(Missing = "<Missing>", `Not High` = c(0, -1, -2, -3, -4), `1` =
    1, `2` = 2, `3` = 3, `4` = 4),
  variables = list(id = "USUBJID", by_var = "BTOXGR", by_grade = "Missing")
)

a_count_abnormal_by_worst_grade_by_baseline(
  df,
  .var = "ATOXGR",
  grouping_list = list(Missing = "<Missing>", `Not High` = c(0, -1, -2, -3, -4), `1` =
    1, `2` = 2, `3` = 3, `4` = 4),
  variables = list(id = "USUBJID", by_var = "BTOXGR", by_grade = "Missing")
)

count_abnormal_by_worst_grade_by_baseline(
  lyt,
  var,
  grouping_list,
  variables = list(id = "USUBJID", by_var = "BTOXGR", by_grade_list = NULL),
  ...,
  table_names = NULL,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)

Arguments

adsl
(data frame) ADSL dataframe

adlb
(data frame) ADLB dataframe

worst_flag
(named vector) Worst post-baseline lab flag variable

by_visit
(logical) defaults to FALSE to generate worst grade per patient. If worst grade per patient per visit is specified for worst_flag, then by_visit should be TRUE to generate worst grade patient per visit.

no_fillin_visits
(named character) Visits that are not considered for post-baseline worst toxicity grade. Defaults to c("SCREENING", "BASELINE").

df
(data frame)
data set containing all analysis variables.

id
(string)
subject variable name.

.var
(string)
single variable name that is passed by rtables when requested by a statistics function.

grouping_list
(named list) defining how inner NCI-CTCAE grades should be grouped.

variables
(named list of string) list of additional analysis variables including:

id (string):
subject variable name

by_var (string):
name of the data column containing the outer level toxicity variable within which toxicity variable .var is counted. Inputs for by_var and .var can be switched, e.g. .var = ATOXGR and by_var = BTOXGR, this generates post-baseline grade by baseline grade table, or .var = BTOXGR and by_var = ATOXGR, this generates baseline grade by post-baseline grade table.

by_grade (character):
outer level toxicity group of interest for s_count_abnormal_by_worst_grade_by_baseline(). by_grade must be identical to the name of an element in the grouping_list

by_grade_list (list):
indicating all outer groups of interest for count_abnormal_by_worst_grade_by_baseline(). If by_grade_list`` is not specified, then all groups in grouping_listwill be used.by_grade_listfollows the same format asgrouping_list`.

lyt
(layout)
input layout where analyses will be added to.

var
(string)
single variable name for the primary analysis variable.

...
additional arguments for the lower level functions.

table_names
(character)
this can be customized in case that the same vars are analyzed multiple times, to avoid warnings from rtables.

.stats
(character)
statistics to select for the table.

.formats
(named character or list)
formats for the statistics.

.labels
(named character)
labels for the statistics (without indent).

.indent_mods
(named integer)
indent modifiers for the labels.

Value

h_adsl_adlb_merge_using_worst_flag() returns a df containing variables shared between adlb and adsl along with variables relevant for analysis: PARAM, PARAMCD, ATOXGR, and BTOXGR. Optionally AVISIT, AVISITN are included when by_visit = TRUE and no_fillin_visits = c("SCREENING", "BASELINE").

h_group_counter() returns the total number of patients, and counts and fractions for each group identified in the grouping_list in a named list, with names from grouping_list.

s_count_abnormal_by_worst_grade_by_baseline() returns the total patient count, and the counts and fraction of patients' worst post-baseline NCI-CTCAE grades in each of the groups in group_listing, for a particular baseline NCI-CTCAE grade group, by_grade. Elements of by_grade_list must be in the set of names of elements in grouping_list, or

a_count_abnormal_by_worst_grade() returns the corresponding list with formatted rtables::CellValue().

Functions

h_adsl_adlb_merge_using_worst_flag: Helper function to filter ADSL and ADLB datasets to create a dataset with the worst post-baseline lab using (worst_flag) flag. Missing data will be created for the following situations:

patients who are present in ADSL but no lab data in ADLB (both baseline and post-baseline)

patients who do not have any post-baseline lab values

patients without any post-baseline values flagged as the worst

h_group_counter: Helper function to group patients (identified using id) according to post-baseline toxicity in .var, then the counts and fractions are produced according to grouping_list, along with total number of patients.

s_count_abnormal_by_worst_grade_by_baseline: Statistics function, for each NCI-CTCAE grade group specified in by_grade, this function calculates the total, counts and fraction of patients' worst NCI-CTCAE grades in .var falling into each of the groups in group_listing.

a_count_abnormal_by_worst_grade_by_baseline: Formatted Analysis function which can be further customized by calling rtables::make_afun() on it. It is used as afun in rtables::analyze().

count_abnormal_by_worst_grade_by_baseline: Layout creating function which can be used for creating tables, which can take statistics function arguments and additional format arguments (see below).

Examples

library(random.cdisc.data)
adlb <- radlb(cached = TRUE)
adsl <- radsl(cached = TRUE)

# `h_adsl_adlb_merge_using_worst_flag`
adlb_out <- h_adsl_adlb_merge_using_worst_flag(adsl, adlb, worst_flag = c("WGRHIFL" = "Y"))

# `h_adsl_adlb_merge_using_worst_flag` by visit example
adlb_out_by_visit <- h_adsl_adlb_merge_using_worst_flag(
  adsl,
  adlb,
  worst_flag = c("WGRLOVFL" = "Y"),
  by_visit = TRUE
)

# `h_group_counter`
h_group_counter(
  df = adlb_out,
  id = "USUBJID",
  .var = "ATOXGR",
  grouping_list = list(
    `Missing`= "<Missing>",
    `Not High` = c(0, -1, -2, -3, -4),
    `1` = 1,
    `2` = 2,
    `3` = 3,
    `4` = 4
    )
  )

#> $count
#> $count$Total
#> [1] 400
#> 
#> 
#> $count_fraction
#> $count_fraction$Missing
#> [1] 0 0
#> 
#> $count_fraction$`Not High`
#> [1] 374.000   0.935
#> 
#> $count_fraction$`1`
#> [1] 128.00   0.32
#> 
#> $count_fraction$`2`
#> [1] 123.0000   0.3075
#> 
#> $count_fraction$`3`
#> [1] 111.0000   0.2775
#> 
#> $count_fraction$`4`
#> [1] 93.0000  0.2325
#> 
#> 

library(dplyr)
# Count worst post-baseline lab for patients who were `Not Low` at baseline.
s_count_abnormal_by_worst_grade_by_baseline(
  df = adlb_out %>% filter(ARMCD == "ARM A" & PARAMCD == "CRP"),
  .var = "ATOXGR",
  grouping_list = list(
    "Missing" = "<Missing>",
    "Not Low" = c(0, 1, 2, 3, 4),
    "1" = -1,
    "2" = -2,
    "3" = -3,
    "4" = -4
    ),
  variables = list(
    id = "USUBJID",
    by_var = "BTOXGR",
    by_grade = "Not Low"
  )
)

#> $count
#> $count$Total
#> [1] 0
#> 
#> 
#> $count_fraction
#> $count_fraction$Missing
#> [1] 0 0
#> 
#> $count_fraction$`Not Low`
#> [1] 0 0
#> 
#> $count_fraction$`1`
#> [1] 0 0
#> 
#> $count_fraction$`2`
#> [1] 0 0
#> 
#> $count_fraction$`3`
#> [1] 0 0
#> 
#> $count_fraction$`4`
#> [1] 0 0
#> 
#> 

afun <- make_afun(a_count_abnormal_by_worst_grade_by_baseline, .ungroup_stats = "count_fraction")
afun(
df = adlb_out %>% filter(ARMCD == "ARM A" & PARAMCD == "CRP"),
.var = "ATOXGR",
grouping_list = list(
  "Missing" = "<Missing>",
  "Not High" = c(0, -1, -2, -3, -4),
  "1" = 1,
   "2" = 2,
   "3" = 3,
   "4" = 4
   ),
variables = list(id = "USUBJID",
                 by_var = "BTOXGR",
                 by_grade = "4" )
)

#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>   row_name formatted_cell indent_mod row_label
#> 1    count              6          0     count
#> 2  Missing              0          0   Missing
#> 3 Not High      5 (83.3%)          0  Not High
#> 4        1              0          0         1
#> 5        2      1 (16.7%)          0         2
#> 6        3              0          0         3
#> 7        4              0          0         4

basic_table() %>%
  split_cols_by("ARMCD") %>%
  add_colcounts() %>%
  split_rows_by("PARAMCD") %>%
  count_abnormal_by_worst_grade_by_baseline(
    var = "ATOXGR",
    grouping_list = list(
      "Missing" = "<Missing>",
      "Not High" = c(0, -1, -2, -3, -4),
      "1" = 1,
      "2" = 2,
      "3" = 3,
      "4" = 4
      ),
    variables = list(
      id = "USUBJID",
      by_var = "BTOXGR"
      )
    ) %>%
  build_table(df = adlb_out, alt_counts_df = adsl)

#>                  ARM A        ARM B        ARM C   
#>                 (N=134)      (N=134)      (N=132)  
#> ---------------------------------------------------
#> ALT                                                
#>   Missing                                          
#>     Total          0            0            0     
#>     Missing        0            0            0     
#>     Not High       0            0            0     
#>     1              0            0            0     
#>     2              0            0            0     
#>     3              0            0            0     
#>     4              0            0            0     
#>   Not High                                         
#>     Total         121          118          118    
#>     Missing        0            0            0     
#>     Not High   72 (59.5%)   70 (59.3%)    59 (50%) 
#>     1          16 (13.2%)    13 (11%)    17 (14.4%)
#>     2          13 (10.7%)   12 (10.2%)   18 (15.3%)
#>     3           9 (7.4%)    14 (11.9%)   12 (10.2%)
#>     4          11 (9.1%)     9 (7.6%)    12 (10.2%)
#>   1                                                
#>     Total          4            4            3     
#>     Missing        0            0            0     
#>     Not High    1 (25%)      1 (25%)     2 (66.7%) 
#>     1           1 (25%)         0            0     
#>     2           1 (25%)         0            0     
#>     3              0         1 (25%)     1 (33.3%) 
#>     4           1 (25%)      2 (50%)         0     
#>   2                                                
#>     Total          4            4            4     
#>     Missing        0            0            0     
#>     Not High    3 (75%)      3 (75%)      3 (75%)  
#>     1              0         1 (25%)      1 (25%)  
#>     2           1 (25%)         0            0     
#>     3              0            0            0     
#>     4              0            0            0     
#>   3                                                
#>     Total          2            5            5     
#>     Missing        0            0            0     
#>     Not High    2 (100%)     2 (40%)      4 (80%)  
#>     1              0         2 (40%)         0     
#>     2              0         1 (20%)         0     
#>     3              0            0            0     
#>     4              0            0         1 (20%)  
#>   4                                                
#>     Total          3            3            2     
#>     Missing        0            0            0     
#>     Not High   1 (33.3%)    2 (66.7%)     2 (100%) 
#>     1          1 (33.3%)        0            0     
#>     2              0            0            0     
#>     3          1 (33.3%)    1 (33.3%)        0     
#>     4              0            0            0     
#> CRP                                                
#>   Missing                                          
#>     Total          0            0            0     
#>     Missing        0            0            0     
#>     Not High       0            0            0     
#>     1              0            0            0     
#>     2              0            0            0     
#>     3              0            0            0     
#>     4              0            0            0     
#>   Not High                                         
#>     Total         115          115          114    
#>     Missing        0            0            0     
#>     Not High   60 (52.2%)   68 (59.1%)   70 (61.4%)
#>     1          17 (14.8%)   13 (11.3%)    9 (7.9%) 
#>     2          12 (10.4%)   14 (12.2%)   13 (11.4%)
#>     3          16 (13.9%)   10 (8.7%)    13 (11.4%)
#>     4          10 (8.7%)    10 (8.7%)     9 (7.9%) 
#>   1                                                
#>     Total          5            5            5     
#>     Missing        0            0            0     
#>     Not High    4 (80%)      3 (60%)      5 (100%) 
#>     1              0            0            0     
#>     2              0         2 (40%)         0     
#>     3              0            0            0     
#>     4           1 (20%)         0            0     
#>   2                                                
#>     Total          6            5            5     
#>     Missing        0            0            0     
#>     Not High    3 (50%)      3 (60%)      4 (80%)  
#>     1              0         1 (20%)         0     
#>     2          2 (33.3%)        0            0     
#>     3              0            0            0     
#>     4          1 (16.7%)     1 (20%)      1 (20%)  
#>   3                                                
#>     Total          2            3            5     
#>     Missing        0            0            0     
#>     Not High    2 (100%)    1 (33.3%)     4 (80%)  
#>     1              0            0            0     
#>     2              0            0            0     
#>     3              0        1 (33.3%)     1 (20%)  
#>     4              0        1 (33.3%)        0     
#>   4                                                
#>     Total          6            6            3     
#>     Missing        0            0            0     
#>     Not High   5 (83.3%)    4 (66.7%)     3 (100%) 
#>     1              0        1 (16.7%)        0     
#>     2          1 (16.7%)        0            0     
#>     3              0        1 (16.7%)        0     
#>     4              0            0            0     
#> IGA                                                
#>   Missing                                          
#>     Total          0            0            0     
#>     Missing        0            0            0     
#>     Not High       0            0            0     
#>     1              0            0            0     
#>     2              0            0            0     
#>     3              0            0            0     
#>     4              0            0            0     
#>   Not High                                         
#>     Total         119          123          119    
#>     Missing        0            0            0     
#>     Not High   61 (51.3%)   69 (56.1%)   74 (62.2%)
#>     1          24 (20.2%)   12 (9.8%)    13 (10.9%)
#>     2          13 (10.9%)   20 (16.3%)   12 (10.1%)
#>     3          11 (9.2%)    11 (8.9%)    16 (13.4%)
#>     4          10 (8.4%)    11 (8.9%)     4 (3.4%) 
#>   1                                                
#>     Total          2            4            6     
#>     Missing        0            0            0     
#>     Not High    2 (100%)     2 (50%)     5 (83.3%) 
#>     1              0            0            0     
#>     2              0            0        1 (16.7%) 
#>     3              0            0            0     
#>     4              0         2 (50%)         0     
#>   2                                                
#>     Total          4            4            3     
#>     Missing        0            0            0     
#>     Not High    3 (75%)      2 (50%)     1 (33.3%) 
#>     1              0         2 (50%)         0     
#>     2              0            0        1 (33.3%) 
#>     3              0            0        1 (33.3%) 
#>     4           1 (25%)         0            0     
#>   3                                                
#>     Total          5            2            2     
#>     Missing        0            0            0     
#>     Not High    3 (60%)         0         1 (50%)  
#>     1           1 (20%)         0            0     
#>     2              0            0         1 (50%)  
#>     3           1 (20%)      2 (100%)        0     
#>     4              0            0            0     
#>   4                                                
#>     Total          4            1            2     
#>     Missing        0            0            0     
#>     Not High    3 (75%)      1 (100%)     1 (50%)  
#>     1              0            0            0     
#>     2           1 (25%)         0         1 (50%)  
#>     3              0            0            0     
#>     4              0            0            0