Add text annotation on top of a kmgrob

utility function to edit a grid grob by adding a text grob

addTable(kmgrob, tbl, x = unit(0.5, "npc"), y = unit(0.5, "npc"),
  just = c("left", "top"), vp = vpPath("plotArea", "topCurve"))

Arguments

kmgrob
a kmGrob for further editing.

tbl
An rtable.

x
A numeric vector or unit object specifying x-values.

y
A numeric vector or unit object specifying y-values.

just
The justification of the text relative to its (x, y) location.

vp
a Grid veiwport object for annotating a text grob.

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

OS <- data.frame(AVAL = abs(rnorm(200)),
                 CNSR = sample(c(0, 1), 200, TRUE),
                 ARM = sample(LETTERS[1:3], 200, TRUE),
                 SEX = sample(c("M","F"), 200, TRUE),
                 RACE = sample(c("AA", "BB", "CC"), 200, TRUE),
                 ECOG = sample(c(0, 1), 200, TRUE))
fit_km <- survfit(Surv(AVAL, 1-CNSR) ~ ARM, data = OS, conf.type = "plain")
a_kmgrob <- kmGrob(fit_km, xticks = 0.5)
fit_coxph <- coxph(Surv(AVAL, 1-CNSR) ~ ARM + strata(RACE), data = OS, ties = "exact")
cox_tbl <- coxphAnnoData(fit_coxph)

#> Error in coxphAnnoData(fit_coxph): could not find function "coxphAnnoData"

addTable(a_kmgrob,
         vp = vpPath("plotArea", "topCurve"),
         x= unit(1, "lines"), y = unit(1, "lines"),
         just = c("left", "bottom"),
         tbl = cox_tbl ) %>%
         grid.draw()

#> Error in toString(tbl, gap = 1): object 'cox_tbl' not found