Conversion of a Vector to a Factor

Converts x to a factor and keeps its attributes. Warns appropriately such that the user can decide whether they prefer converting to factor manually (e.g. for full control of factor levels).

Usage

as_factor_keep_attributes(
  x,
  x_name = deparse(substitute(x)),
  na_level = "<Missing>"
)

Arguments

x

(atomic)
object to convert.

x_name

(string)
name of x.

na_level

(string)
the explicit missing level which should be used when converting a character vector.

Value

The factor with same attributes (except class) as x. Does not do any modifications if x is already a factor.

Examples

tern:::as_factor_keep_attributes(formatters::with_label(c(1, 1, 2, 3), "id"))
#> Warning: automatically converting numeric variable formatters::with_label(c(1, 1, 2, 3), "id") to factor, better manually convert to factor to avoid failures
#> [1] 1 1 2 3
#> attr(,"label")
#> [1] id
#> Levels: 1 2 3
tern:::as_factor_keep_attributes(c("a", "b", ""), "id")
#> Warning: automatically converting character variable id to factor, better manually convert to factor to avoid failures
#> [1] a         b         <Missing>
#> Levels: a b <Missing>