Check proportion difference arguments

Verifies that and/or convert arguments into valid values to be used in the estimation of difference in responder proportions.

Usage

check_diff_prop_ci(rsp, grp, strata = NULL, conf_level, correct = NULL)

Arguments

rsp

(logical)
vector indicating whether each subject is a responder or not.

grp

(factor)
vector assigning observations to one out of two groups (e.g. reference and treatment group).

strata

(factor)
variable with one level per stratum and same length as rsp.

conf_level

(proportion)
confidence level of the interval.

correct

(flag)
whether to include the continuity correction. For further information, see stats::prop.test().