MMRM Variables Helper

Helper function to check the MMRM variables.

check_mmrm_vars(vars, data)

Arguments

vars
(named list of string or character)
specifying the variables in the MMRM. The following elements need to be included as character vectors and match corresponding columns in data:

response

the response variable

covariates

the additional covariate terms (might also include interactions)

id

the subject ID variable

arm

the treatment group variable (factor)

visit

the visit variable (factor)

Note that the main effects and interaction of arm and visit are by default included in the model.

data
a data.frame with all the variables specified in vars. Records with missing values in any independent variables will be excluded.

Value

a corresponding list of variable labels. For the covariates, the element parts contains a character vector with one element per covariates part (i.e. occurring variable in the original covariates specification) is included. If a label is not found, the variable name is instead used.