Plotting data generated from chgfbl_data into change from baseline mean plot using ggplot2 and gridExtra

Plotting data generated from chgfbl_data into change from baseline mean plot using ggplot2 and gridExtra

chgfbl_plot(data, ytype = "CHG", errbar = "SE", ref_line = NULL,
  yrange = NULL, fontsize = 16, n_rotate = TRUE)

Arguments

data
data generated by chgfbl_data function

ytype
type of value to plot for y-axis, possible values can be "AVAL", "CHG" or "PCHG"

errbar
type of error bar to draw, possible values can be "SE", "SD", "95CI"

ref_line
horizontal reference lines, must be either NULL (default), a single number or a numeric vector

yrange
Range of y-axis to display in plot. If empty (default), range automatically controlled by ggplot2. Or must be a numeric vector of length of 2, with the first number smaller than second number.

fontsize
Font size of text to be displayed on the plot

n_rotate
logical value, FALSE is default, if TRUE text in the frequency table is rotated 90 degrees counterclockwise.

Examples

library(tern)
library(random.cdisc.data)

ASL <- radam("ASL")

ANL <- merge(ASL, radam("AQS", ADSL = ASL)) %>%
   filter(PARAMCD == "FATIGI")

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Error in filter(., PARAMCD == "FATIGI"): object 'PARAMCD' not found

df <- chgfbl_data(data = ANL)

#> Error in chgfbl_data(data = ANL): object 'ANL' not found

chgfbl_plot(data = df, ytype = "CHG", errbar = "SE", ref_line = c(2, -2), fontsize = 14)

#> Error in data$mean: object of type 'closure' is not subsettable