Wrap data generated from chgfbl_data into rtable format, to create a test result and change from baseline by visit summary table

Wrap data generated from chgfbl_data into rtable format, to create a test result and change from baseline by visit summary table

chgfbl_table(data)

Arguments

data
data generated by chgfbl_data function

Examples

library(tern)
library(random.cdisc.data)

ASL <- radam("ASL")

ANL <- merge(ASL, radam("AQS", ADSL = ASL)) %>%
   filter(PARAMCD == "FATIGI")

df <- chgfbl_data(data = ANL)
chgfbl_table(df)

#>                        ARM A                 ARM A                 ARM B                 ARM B     
#>                    Value at Visit         Change from          Value at Visit         Change from  
#>                        (N=43)               Baseline               (N=57)               Baseline   
#>                                              (N=43)                                      (N=57)    
#> ---------------------------------------------------------------------------------------------------
#> WEEK 1
#>   n                      43                                          57                            
#>   Mean (SD)           -0.1 (1)                                     0 (1)                           
#>   Median                 0                                          -0.1                           
#>   IQR               -0.99 - 0.64                                -0.54 - 0.69                       
#>   Min - Max         -1.92 - 1.84                                -2.46 - 2.79                       
#> 
#> WEEK 2
#>   n                      43                    43                    57                    57      
#>   Mean (SD)           -0.2 (1)             -0.1 (1.5)            -0.1 (1.2)             0 (1.5)    
#>   Median                -0.1                   0                    -0.2                  -0.2     
#>   IQR               -0.75 - 0.23          -0.97 - 0.89          -0.97 - 0.69          -0.98 - 1.47 
#>   Min - Max         -2.45 - 1.94          -3.91 - 3.68          -1.88 - 3.59          -3.21 - 3.24 
#> 
#> WEEK 3
#>   n                      38                    38                    50                    50      
#>   Mean (SD)           -0.2 (1)             -0.1 (1.7)             -0.1 (1)              0 (1.7)    
#>   Median                -0.1                  -0.2                  -0.2                  -0.2     
#>   IQR                -1 - 0.42            -1.63 - 1.21          -0.61 - 0.39          -1.27 - 1.28 
#>   Min - Max         -2.32 - 1.48          -3.6 - 3.26           -2.21 - 2.45          -2.68 - 4.14 
#> 
#> WEEK 4
#>   n                      30                    30                    42                    42      
#>   Mean (SD)          -0.1 (1.3)            0.1 (1.8)             0.1 (0.9)             0.2 (1.2)   
#>   Median                -0.3                  0.2                   -0.1                  0.1      
#>   IQR               -0.94 - 0.88          -1.47 - 1.6           -0.43 - 0.39          -0.61 - 0.83 
#>   Min - Max         -2.39 - 2.13          -2.7 - 3.13           -2.16 - 2.01          -2.01 - 3.34 
#> 
#> WEEK 5
#>   n                      27                    27                    38                    38      
#>   Mean (SD)          0.1 (0.9)             0.4 (1.6)             -0.1 (1.1)             0 (1.7)    
#>   Median                0.2                   0.4                   -0.2                  -0.3     
#>   IQR               -0.41 - 0.74          -0.68 - 1.59          -0.67 - 0.61           -0.8 - 1.3  
#>   Min - Max         -1.97 - 1.71          -2.72 - 3.49          -2.72 - 1.76          -3.92 - 3.35 
#> 
#> WEEK 6
#>   n                      25                    25                    32                    32      
#>   Mean (SD)          0.1 (1.1)             0.3 (1.7)             0.1 (0.8)             0.1 (1.5)   
#>   Median                0.1                   0.9                   0.1                    0       
#>   IQR               -0.74 - 0.77          -0.73 - 1.28          -0.33 - 0.44          -1.05 - 1.2  
#>   Min - Max         -1.6 - 3.45           -3.26 - 4.25          -1.61 - 1.96          -2.71 - 3.14 
#> 
#> WEEK 7
#>   n                      22                    22                    27                    27      
#>   Mean (SD)            0 (1)                0.5 (1)              -0.1 (1.1)            -0.3 (1.4)  
#>   Median                 0                    0.7                    0                    -0.3     
#>   IQR               -0.91 - 0.85          -0.54 - 1.24          -0.92 - 0.73          -1.25 - 0.69 
#>   Min - Max         -1.63 - 1.63          -1.33 - 2.4           -1.99 - 2.1           -2.88 - 3.72 
#> 
#> WEEK 8
#>   n                      19                    19                    23                    23      
#>   Mean (SD)           0 (0.9)              0.5 (1.5)             -0.3 (0.9)            -0.4 (1.5)  
#>   Median                -0.1                  0.7                   -0.2                  -0.5     
#>   IQR               -0.5 - 0.43           -0.19 - 1.28          -1.02 - 0.32          -1.3 - 0.26  
#>   Min - Max         -1.88 - 1.85          -2.87 - 3.75          -1.78 - 1.41          -2.91 - 2.87 
#> 
#> WEEK 9
#>   n                      16                    16                    15                    15      
#>   Mean (SD)          -0.2 (0.9)             0 (1.4)               0 (0.9)              -0.1 (1.3)  
#>   Median                -0.2                   0                    -0.2                  0.2      
#>   IQR               -0.62 - 0.24          -1.14 - 1.04          -0.58 - 0.6           -1.35 - 0.79 
#>   Min - Max         -1.94 - 1.12          -2.48 - 2.32          -1.59 - 1.32          -1.88 - 1.97 
#> 
#> WEEK 10
#>   n                      9                     9                     9                     9       
#>   Mean (SD)           -0.1 (1)             0.2 (1.9)              0 (0.9)               0 (2.1)    
#>   Median                 0                    0.6                   -0.4                  -0.5     
#>   IQR               -0.22 - 0.31          -0.79 - 1.35          -0.56 - 0.72          -1.23 - 1.06 
#>   Min - Max         -1.53 - 1.68          -2.91 - 3.1           -1.3 - 1.44           -2.42 - 4.44 
#> 
#> WEEK 11
#>   n                      5                     5                     5                     5       
#>   Mean (SD)          -0.2 (0.6)            -0.1 (1.3)             0.2 (1)              -0.5 (1.2)  
#>   Median                -0.1                  -0.1                  -0.1                  -0.4     
#>   IQR              -0.64 - -0.05          -1.26 - 1.02          -0.17 - 0.84         -0.91 - -0.28 
#>   Min - Max         -0.92 - 0.71          -1.44 - 1.32          -1.26 - 1.42          -2.03 - 1.17 
#>