Control Function for Descriptive Statistics

Sets a list of parameters for summaries of descriptive statistics. Typically used internally to specify details for s_summary.

Usage

control_summarize_vars(
  conf_level = 0.95,
  quantiles = c(0.25, 0.75),
  quantile_type = 2,
  test_mean = 0
)

Arguments

conf_level

(proportion)
confidence level of the interval.

quantiles

(numeric)
of length two to specify the quantiles to calculate.

quantile_type

(numeric)
between 1 and 9 selecting quantile algorithms to be used.
Default is set to 2 as this matches the default quantile algorithm in SAS proc univariate set by QNTLDEF=5. This differs from R's default. See more about type in stats::quantile().

test_mean

(numeric)
to test against the mean under the null hypothesis when calculating p-value.

Value

A list of components with the same names as the arguments.