Remove Shared Variables

Variables are considered shared if they have the same variable name

drop_shared_variables(x, y, keep)

Arguments

x
a data.frame

y
a data.frame

keep
optional, a vector with variable names that should not be removed

Value

a data.frame

Examples

drop_shared_variables(iris, iris[, 1:3])

#>     Petal.Width    Species
#> 1           0.2     setosa
#> 2           0.2     setosa
#> 3           0.2     setosa
#> 4           0.2     setosa
#> 5           0.2     setosa
#> 6           0.4     setosa
#> 7           0.3     setosa
#> 8           0.2     setosa
#> 9           0.2     setosa
#> 10          0.1     setosa
#> 11          0.2     setosa
#> 12          0.2     setosa
#> 13          0.1     setosa
#> 14          0.1     setosa
#> 15          0.2     setosa
#> 16          0.4     setosa
#> 17          0.4     setosa
#> 18          0.3     setosa
#> 19          0.3     setosa
#> 20          0.3     setosa
#> 21          0.2     setosa
#> 22          0.4     setosa
#> 23          0.2     setosa
#> 24          0.5     setosa
#> 25          0.2     setosa
#> 26          0.2     setosa
#> 27          0.4     setosa
#> 28          0.2     setosa
#> 29          0.2     setosa
#> 30          0.2     setosa
#> 31          0.2     setosa
#> 32          0.4     setosa
#> 33          0.1     setosa
#> 34          0.2     setosa
#> 35          0.2     setosa
#> 36          0.2     setosa
#> 37          0.2     setosa
#> 38          0.1     setosa
#> 39          0.2     setosa
#> 40          0.2     setosa
#> 41          0.3     setosa
#> 42          0.3     setosa
#> 43          0.2     setosa
#> 44          0.6     setosa
#> 45          0.4     setosa
#> 46          0.3     setosa
#> 47          0.2     setosa
#> 48          0.2     setosa
#> 49          0.2     setosa
#> 50          0.2     setosa
#> 51          1.4 versicolor
#> 52          1.5 versicolor
#> 53          1.5 versicolor
#> 54          1.3 versicolor
#> 55          1.5 versicolor
#> 56          1.3 versicolor
#> 57          1.6 versicolor
#> 58          1.0 versicolor
#> 59          1.3 versicolor
#> 60          1.4 versicolor
#> 61          1.0 versicolor
#> 62          1.5 versicolor
#> 63          1.0 versicolor
#> 64          1.4 versicolor
#> 65          1.3 versicolor
#> 66          1.4 versicolor
#> 67          1.5 versicolor
#> 68          1.0 versicolor
#> 69          1.5 versicolor
#> 70          1.1 versicolor
#> 71          1.8 versicolor
#> 72          1.3 versicolor
#> 73          1.5 versicolor
#> 74          1.2 versicolor
#> 75          1.3 versicolor
#> 76          1.4 versicolor
#> 77          1.4 versicolor
#> 78          1.7 versicolor
#> 79          1.5 versicolor
#> 80          1.0 versicolor
#> 81          1.1 versicolor
#> 82          1.0 versicolor
#> 83          1.2 versicolor
#> 84          1.6 versicolor
#> 85          1.5 versicolor
#> 86          1.6 versicolor
#> 87          1.5 versicolor
#> 88          1.3 versicolor
#> 89          1.3 versicolor
#> 90          1.3 versicolor
#> 91          1.2 versicolor
#> 92          1.4 versicolor
#> 93          1.2 versicolor
#> 94          1.0 versicolor
#> 95          1.3 versicolor
#> 96          1.2 versicolor
#> 97          1.3 versicolor
#> 98          1.3 versicolor
#> 99          1.1 versicolor
#> 100         1.3 versicolor
#> 101         2.5  virginica
#> 102         1.9  virginica
#> 103         2.1  virginica
#> 104         1.8  virginica
#> 105         2.2  virginica
#> 106         2.1  virginica
#> 107         1.7  virginica
#> 108         1.8  virginica
#> 109         1.8  virginica
#> 110         2.5  virginica
#> 111         2.0  virginica
#> 112         1.9  virginica
#> 113         2.1  virginica
#> 114         2.0  virginica
#> 115         2.4  virginica
#> 116         2.3  virginica
#> 117         1.8  virginica
#> 118         2.2  virginica
#> 119         2.3  virginica
#> 120         1.5  virginica
#> 121         2.3  virginica
#> 122         2.0  virginica
#> 123         2.0  virginica
#> 124         1.8  virginica
#> 125         2.1  virginica
#> 126         1.8  virginica
#> 127         1.8  virginica
#> 128         1.8  virginica
#> 129         2.1  virginica
#> 130         1.6  virginica
#> 131         1.9  virginica
#> 132         2.0  virginica
#> 133         2.2  virginica
#> 134         1.5  virginica
#> 135         1.4  virginica
#> 136         2.3  virginica
#> 137         2.4  virginica
#> 138         1.8  virginica
#> 139         1.8  virginica
#> 140         2.1  virginica
#> 141         2.4  virginica
#> 142         2.3  virginica
#> 143         1.9  virginica
#> 144         2.3  virginica
#> 145         2.5  virginica
#> 146         2.3  virginica
#> 147         1.9  virginica
#> 148         2.0  virginica
#> 149         2.3  virginica
#> 150         1.8  virginica