Prepares Response Data for Population Subgroups in Data Frames

Prepares response rates and odds ratios for population subgroups in data frames. Simple wrapper for h_odds_ratio_subgroups_df() and h_proportion_subgroups_df(). Result is a list of two data frames: prop and or. variables corresponds to the names of variables found in data, passed as a named list and requires elements rsp, arm and optionally subgroups and strat. groups_lists optionally specifies groupings for subgroups variables.

Usage

extract_rsp_subgroups(
  variables,
  data,
  groups_lists = list(),
  conf_level = 0.95,
  method = NULL,
  label_all = "All Patients"
)

Arguments

variables

(named list of string)
list of additional analysis variables.

data

(data.frame)
the dataset containing the variables to summarize.

groups_lists

(named list of list)
optionally contains for each subgroups variable a list, which specifies the new group levels via the names and the levels that belong to it in the character vectors that are elements of the list.

conf_level

(proportion)
confidence level of the interval.

method

(string)
specifies the test used to calculate the p-value for the difference between two proportions. For options, see s_test_proportion_diff(). Default is NULL so no test is performed.

label_all

(string)
label for the total population analysis.

See also

Examples

library(dplyr)
library(forcats)

adrs <- tern_ex_adrs
adrs_labels <- formatters::var_labels(adrs)

adrs_f <- adrs %>%
  filter(PARAMCD == "BESRSPI") %>%
  filter(ARM %in% c("A: Drug X", "B: Placebo")) %>%
  droplevels() %>%
  mutate(
    # Reorder levels of factor to make the placebo group the reference arm.
    ARM = fct_relevel(ARM, "B: Placebo"),
    rsp = AVALC == "CR"
  )
formatters::var_labels(adrs_f) <- c(adrs_labels, "Response")

# Unstratified analysis.
df <- extract_rsp_subgroups(
  variables = list(rsp = "rsp", arm = "ARM", subgroups = c("SEX", "BMRKR2")),
  data = adrs_f
)
df
#> $prop
#>           arm  n n_rsp      prop     subgroup    var    var_label row_type
#> 1  B: Placebo 73    52 0.7123288 All Patients    ALL All Patients  content
#> 2   A: Drug X 69    60 0.8695652 All Patients    ALL All Patients  content
#> 3  B: Placebo 40    30 0.7500000            F    SEX          SEX analysis
#> 4   A: Drug X 38    30 0.7894737            F    SEX          SEX analysis
#> 5  B: Placebo 33    22 0.6666667            M    SEX          SEX analysis
#> 6   A: Drug X 31    30 0.9677419            M    SEX          SEX analysis
#> 7  B: Placebo 24    16 0.6666667          LOW BMRKR2       BMRKR2 analysis
#> 8   A: Drug X 26    24 0.9230769          LOW BMRKR2       BMRKR2 analysis
#> 9  B: Placebo 23    18 0.7826087       MEDIUM BMRKR2       BMRKR2 analysis
#> 10  A: Drug X 26    21 0.8076923       MEDIUM BMRKR2       BMRKR2 analysis
#> 11 B: Placebo 26    18 0.6923077         HIGH BMRKR2       BMRKR2 analysis
#> 12  A: Drug X 17    15 0.8823529         HIGH BMRKR2       BMRKR2 analysis
#> 
#> $or
#>   arm n_tot        or       lcl        ucl conf_level     subgroup    var
#> 1       142  2.692308 1.1339760   6.392129       0.95 All Patients    ALL
#> 2        78  1.250000 0.4337195   3.602559       0.95            F    SEX
#> 3        64 15.000000 1.8009265 124.935693       0.95            M    SEX
#> 4        50  6.000000 1.1253817  31.989146       0.95          LOW BMRKR2
#> 5        49  1.166667 0.2905018   4.685379       0.95       MEDIUM BMRKR2
#> 6        43  3.333333 0.6124760  18.141301       0.95         HIGH BMRKR2
#>      var_label row_type
#> 1 All Patients  content
#> 2          SEX analysis
#> 3          SEX analysis
#> 4       BMRKR2 analysis
#> 5       BMRKR2 analysis
#> 6       BMRKR2 analysis
#> 

# Stratified analysis.
df_strat <- extract_rsp_subgroups(
  variables = list(rsp = "rsp", arm = "ARM", subgroups = c("SEX", "BMRKR2"), strat = "STRATA1"),
  data = adrs_f
)
df_strat
#> $prop
#>           arm  n n_rsp      prop     subgroup    var    var_label row_type
#> 1  B: Placebo 73    52 0.7123288 All Patients    ALL All Patients  content
#> 2   A: Drug X 69    60 0.8695652 All Patients    ALL All Patients  content
#> 3  B: Placebo 40    30 0.7500000            F    SEX          SEX analysis
#> 4   A: Drug X 38    30 0.7894737            F    SEX          SEX analysis
#> 5  B: Placebo 33    22 0.6666667            M    SEX          SEX analysis
#> 6   A: Drug X 31    30 0.9677419            M    SEX          SEX analysis
#> 7  B: Placebo 24    16 0.6666667          LOW BMRKR2       BMRKR2 analysis
#> 8   A: Drug X 26    24 0.9230769          LOW BMRKR2       BMRKR2 analysis
#> 9  B: Placebo 23    18 0.7826087       MEDIUM BMRKR2       BMRKR2 analysis
#> 10  A: Drug X 26    21 0.8076923       MEDIUM BMRKR2       BMRKR2 analysis
#> 11 B: Placebo 26    18 0.6923077         HIGH BMRKR2       BMRKR2 analysis
#> 12  A: Drug X 17    15 0.8823529         HIGH BMRKR2       BMRKR2 analysis
#> 
#> $or
#>   arm n_tot        or       lcl        ucl conf_level     subgroup    var
#> 1       142  2.645025 1.1213486   6.239058       0.95 All Patients    ALL
#> 2        78  1.170999 0.4085455   3.356394       0.95            F    SEX
#> 3        64 14.585738 1.7619446 120.743724       0.95            M    SEX
#> 4        50  5.455383 1.0330065  28.810276       0.95          LOW BMRKR2
#> 5        49  1.233205 0.3025472   5.026636       0.95       MEDIUM BMRKR2
#> 6        43  3.699720 0.6662397  20.545054       0.95         HIGH BMRKR2
#>      var_label row_type
#> 1 All Patients  content
#> 2          SEX analysis
#> 3          SEX analysis
#> 4       BMRKR2 analysis
#> 5       BMRKR2 analysis
#> 6       BMRKR2 analysis
#> 

# Grouping of the BMRKR2 levels.
df_grouped <- extract_rsp_subgroups(
  variables = list(rsp = "rsp", arm = "ARM", subgroups = c("SEX", "BMRKR2")),
  data = adrs_f,
  groups_lists = list(
    BMRKR2 = list(
      "low" = "LOW",
      "low/medium" = c("LOW", "MEDIUM"),
      "low/medium/high" = c("LOW", "MEDIUM", "HIGH")
    )
  )
)
df_grouped
#> $prop
#>           arm  n n_rsp      prop        subgroup    var    var_label row_type
#> 1  B: Placebo 73    52 0.7123288    All Patients    ALL All Patients  content
#> 2   A: Drug X 69    60 0.8695652    All Patients    ALL All Patients  content
#> 3  B: Placebo 40    30 0.7500000               F    SEX          SEX analysis
#> 4   A: Drug X 38    30 0.7894737               F    SEX          SEX analysis
#> 5  B: Placebo 33    22 0.6666667               M    SEX          SEX analysis
#> 6   A: Drug X 31    30 0.9677419               M    SEX          SEX analysis
#> 7  B: Placebo 24    16 0.6666667             low BMRKR2       BMRKR2 analysis
#> 8   A: Drug X 26    24 0.9230769             low BMRKR2       BMRKR2 analysis
#> 9  B: Placebo 47    34 0.7234043      low/medium BMRKR2       BMRKR2 analysis
#> 10  A: Drug X 52    45 0.8653846      low/medium BMRKR2       BMRKR2 analysis
#> 11 B: Placebo 73    52 0.7123288 low/medium/high BMRKR2       BMRKR2 analysis
#> 12  A: Drug X 69    60 0.8695652 low/medium/high BMRKR2       BMRKR2 analysis
#> 
#> $or
#>   arm n_tot        or       lcl        ucl conf_level        subgroup    var
#> 1       142  2.692308 1.1339760   6.392129       0.95    All Patients    ALL
#> 2        78  1.250000 0.4337195   3.602559       0.95               F    SEX
#> 3        64 15.000000 1.8009265 124.935693       0.95               M    SEX
#> 4        50  6.000000 1.1253817  31.989146       0.95             low BMRKR2
#> 5        99  2.457983 0.8853707   6.823900       0.95      low/medium BMRKR2
#> 6       142  2.692308 1.1339760   6.392129       0.95 low/medium/high BMRKR2
#>      var_label row_type
#> 1 All Patients  content
#> 2          SEX analysis
#> 3          SEX analysis
#> 4       BMRKR2 analysis
#> 5       BMRKR2 analysis
#> 6       BMRKR2 analysis
#>