Prepares Survival Data Estimates for Multiple Biomarkers in a Single Data Frame

Prepares estimates for number of events, patients and median survival times, as well as hazard ratio estimates, confidence intervals and p-values, for multiple biomarkers across population subgroups in a single data frame. variables corresponds to the names of variables found in data, passed as a named list and requires elements tte, is_event, biomarkers (vector of continuous biomarker variables) and optionally subgroups and strat. groups_lists optionally specifies groupings for subgroups variables.

Usage

extract_survival_biomarkers(
  variables,
  data,
  groups_lists = list(),
  control = control_coxreg(),
  label_all = "All Patients"
)

Arguments

variables

(named list of string)
list of additional analysis variables.

data

(data.frame)
the dataset containing the variables to summarize.

groups_lists

(named list of list)
optionally contains for each subgroups variable a list, which specifies the new group levels via the names and the levels that belong to it in the character vectors that are elements of the list.

control

(list)
a list of parameters as returned by the helper function control_coxreg().

label_all

(string)
label for the total population analysis.

See also

Examples

# Typical analysis of two continuous biomarkers `BMRKR1` and `AGE`,
# in multiple regression models containing one covariate `RACE`,
# as well as one stratification variable `STRATA1`. The subgroups
# are defined by the levels of `BMRKR2`.

library(dplyr)

adtte <- tern_ex_adtte
adtte_labels <- formatters::var_labels(adtte)

adtte_f <- adtte %>%
  filter(PARAMCD == "OS") %>%
  mutate(
    AVALU = as.character(AVALU),
    is_event = CNSR == 0
  )
labels <- c("AVALU" = adtte_labels[["AVALU"]], "is_event" = "Event Flag")
formatters::var_labels(adtte_f)[names(labels)] <- labels

df <- extract_survival_biomarkers(
  variables = list(
    tte = "AVAL",
    is_event = "is_event",
    biomarkers = c("BMRKR1", "AGE"),
    strata = "STRATA1",
    covariates = "SEX",
    subgroups = "BMRKR2"
  ),
  data = adtte_f
)
df
#>   biomarker biomarker_label n_tot n_tot_events   median        hr       lcl
#> 1    BMRKR1          BMRKR1   200          142 753.5176 0.9778269 0.9279033
#> 2       AGE             AGE   200          142 753.5176 0.9902722 0.9655108
#> 3    BMRKR1          BMRKR1    70           52 691.8542 0.9135352 0.8296837
#> 4       AGE             AGE    70           52 691.8542 0.9992661 0.9544867
#> 5    BMRKR1          BMRKR1    68           48 654.1798 1.0085287 0.9195976
#> 6       AGE             AGE    68           48 654.1798 0.9691179 0.9118318
#> 7    BMRKR1          BMRKR1    62           42 835.0535 1.0102921 0.9038195
#> 8       AGE             AGE    62           42 835.0535 0.9830767 0.9444710
#>        ucl conf_level      pval     pval_label     subgroup    var    var_label
#> 1 1.030437       0.95 0.4016894 p-value (Wald) All Patients    ALL All Patients
#> 2 1.015669       0.95 0.4492774 p-value (Wald) All Patients    ALL All Patients
#> 3 1.005861       0.95 0.0656222 p-value (Wald)          LOW BMRKR2       BMRKR2
#> 4 1.046146       0.95 0.9749606 p-value (Wald)          LOW BMRKR2       BMRKR2
#> 5 1.106060       0.95 0.8569068 p-value (Wald)       MEDIUM BMRKR2       BMRKR2
#> 6 1.030003       0.95 0.3129519 p-value (Wald)       MEDIUM BMRKR2       BMRKR2
#> 7 1.129308       0.95 0.8569876 p-value (Wald)         HIGH BMRKR2       BMRKR2
#> 8 1.023260       0.95 0.4037047 p-value (Wald)         HIGH BMRKR2       BMRKR2
#>   row_type
#> 1  content
#> 2  content
#> 3 analysis
#> 4 analysis
#> 5 analysis
#> 6 analysis
#> 7 analysis
#> 8 analysis

# Here we group the levels of `BMRKR2` manually.
df_grouped <- extract_survival_biomarkers(
  variables = list(
    tte = "AVAL",
    is_event = "is_event",
    biomarkers = c("BMRKR1", "AGE"),
    strata = "STRATA1",
    covariates = "SEX",
    subgroups = "BMRKR2"
  ),
  data = adtte_f,
  groups_lists = list(
    BMRKR2 = list(
      "low" = "LOW",
      "low/medium" = c("LOW", "MEDIUM"),
      "low/medium/high" = c("LOW", "MEDIUM", "HIGH")
    )
  )
)
df_grouped
#>   biomarker biomarker_label n_tot n_tot_events   median        hr       lcl
#> 1    BMRKR1          BMRKR1   200          142 753.5176 0.9778269 0.9279033
#> 2       AGE             AGE   200          142 753.5176 0.9902722 0.9655108
#> 3    BMRKR1          BMRKR1    70           52 691.8542 0.9135352 0.8296837
#> 4       AGE             AGE    70           52 691.8542 0.9992661 0.9544867
#> 5    BMRKR1          BMRKR1   138          100 658.9503 0.9615235 0.9022940
#> 6       AGE             AGE   138          100 658.9503 0.9932597 0.9590645
#> 7    BMRKR1          BMRKR1   200          142 753.5176 0.9778269 0.9279033
#> 8       AGE             AGE   200          142 753.5176 0.9902722 0.9655108
#>        ucl conf_level      pval     pval_label        subgroup    var
#> 1 1.030437       0.95 0.4016894 p-value (Wald)    All Patients    ALL
#> 2 1.015669       0.95 0.4492774 p-value (Wald)    All Patients    ALL
#> 3 1.005861       0.95 0.0656222 p-value (Wald)             low BMRKR2
#> 4 1.046146       0.95 0.9749606 p-value (Wald)             low BMRKR2
#> 5 1.024641       0.95 0.2264495 p-value (Wald)      low/medium BMRKR2
#> 6 1.028674       0.95 0.7051617 p-value (Wald)      low/medium BMRKR2
#> 7 1.030437       0.95 0.4016894 p-value (Wald) low/medium/high BMRKR2
#> 8 1.015669       0.95 0.4492774 p-value (Wald) low/medium/high BMRKR2
#>      var_label row_type
#> 1 All Patients  content
#> 2 All Patients  content
#> 3       BMRKR2 analysis
#> 4       BMRKR2 analysis
#> 5       BMRKR2 analysis
#> 6       BMRKR2 analysis
#> 7       BMRKR2 analysis
#> 8       BMRKR2 analysis