Insertion of explicit missing values in a factor

This inserts explicit missing values in a factor based on a condition. Additionally, existing NA values will be explicitly converted to given na_level.

Usage

fct_explicit_na_if(x, condition, na_level = "<Missing>")

Arguments

x

(factor)
the original factor.

condition

(logical)
positions at which to insert missing values.

na_level

(string)
which level to use for missing values.

Value

A modified factor with inserted and existing NA converted to na_level.

See also

Examples

fct_explicit_na_if(factor(c("a", "b", NA)), c(TRUE, FALSE, FALSE))
#> [1] <Missing> b         <Missing>
#> Levels: a b <Missing>