Deprecated by tern.mmrm::fit_mmrm: MMRM Analysis

Does the MMRM analysis. Multiple other functions can be called on the result to produce tables and graphs.

Usage

fit_mmrm(
  vars = list(response = "AVAL", covariates = c(), id = "USUBJID", arm = "ARM", visit =
    "AVISIT"),
  data,
  conf_level = 0.95,
  cor_struct = "unstructured",
  weights_emmeans = "proportional",
  optimizer = "automatic",
  parallel = FALSE
)

Arguments

vars

(named list of string or character)
specifying the variables in the MMRM. The following elements need to be included as character vectors and match corresponding columns in data:

response

the response variable

covariates

the additional covariate terms (might also include interactions)

id

the subject ID variable

arm

the treatment group variable (factor)

visit

the visit variable (factor)

Note that the main effects and interaction of arm and visit are by default included in the model.

data

a data.frame with all the variables specified in vars. Records with missing values in any independent variables will be excluded.

conf_level

(proportion)
confidence level of the interval.

cor_struct

a string specifying the correlation structure, defaults to "unstructured". See the details.

weights_emmeans

argument from emmeans::emmeans(), "proportional" by default.

optimizer

a string specifying the optimization algorithm which should be used. By default, "automatic" will (if necessary) try all possible optimization algorithms and choose the best result. If another algorithm is chosen and does not give a valid result, an error will occur.

parallel

flag that controls whether "automatic" optimizer search can use available free cores on the machine (not default).

Value

An mmrm object which is a list with MMRM results:

fit

The lmerModLmerTest object which was fitted to the data. Note that the attribute optimizer contains the finally used optimization algorithm, which can be useful for refitting the model later on.

cov_estimate

The matrix with the covariance matrix estimate.

diagnostics

A list with model diagnostic statistics (REML criterion, AIC, corrected AIC, BIC).

lsmeans

This is a list with data frames estimates and contrasts.

vars

The variable list.

labels

Corresponding list with variable labels extracted from data.

ref_level

The reference level for the arm variable, which is always the first level.

conf_level

The confidence level which was used to construct the confidence intervals.

Details

Only Satterthwaite adjusted degrees of freedom (d.f.) are supported, because they match the results obtained in SAS (confirmed for unstructured and compound symmetry correlation structures).

For the correlation structure (cor_struct), the user can choose among the following options, sorted in descending number of variance parameters:

unstructured

Unstructured covariance matrix. This is the most flexible choice and default. If there are T visits, then T * (T+1) / 2 variance parameters are used. Note: the current actual implementation uses one more variance parameter, which does not have any effect of the results. Therefore we report here the actually relevant number of parameters.

random-quadratic

Random quadratic spline for the random effects of the time variable. 7 variance parameters are used.

random-slope

Random slope for the random effects of the time variable. 4 variance parameters are used.

compound-symmetry

Constant correlation between visits. 2 variance parameters are used.

For the optimizer, the user can choose among the following alternatives to the recommended "automatic":

nloptwrap_neldermead

NLopt version of the Nelder-Mead algorithm (via package nloptr)

nloptwrap_bobyqa

NLopt version of the BOBYQA algorithm (via package nloptr)

bobyqa

BOBYQA algorithm (via package minqa)

nlminbwrap

nlminb algorithm (wrapper for nlminb)

neldermead

lme4 version of the Nelder-Mead algorithm with box constraints (via package lme4)

nmkbw

Nelder-Mead algorithm (via package dfoptim)

optimx_lbfgsb

L-BFGS-B algorithm (via package optimx)

Note

The ordering of the input data sets can lead to slightly different numerical results or different convergence behavior. This is a known observation with the used package lme4. However, once convergence is achieved, the results are reliable up to numerical precision.

Examples

library(scda)
library(dplyr)
library(rtables)

adsl <- synthetic_cdisc_data("latest")$adsl
adqs <- synthetic_cdisc_data("latest")$adqs
adqs_f <- adqs %>%
  filter(PARAMCD == "FKSI-FWB" & !AVISIT %in% c("BASELINE")) %>%
  droplevels() %>%
  mutate(ARMCD = factor(ARMCD, levels = c("ARM B", "ARM A", "ARM C"))) %>%
  mutate(
    AVISITN = rank(AVISITN) %>%
      as.factor() %>%
      as.numeric() %>%
      as.factor()
  )
formatters::var_labels(adqs_f) <- formatters::var_labels(adqs)
if (FALSE) {
# sometimes results in failure to converge with 1 negative eigenvalue
# in the event that it fails to converge, change the optimizer to "automatic"
mmrm_results <- fit_mmrm(
  vars = list(
    response = "AVAL",
    covariates = c("STRATA1", "BMRKR2"),
    id = "USUBJID",
    arm = "ARMCD",
    visit = "AVISIT"
  ),
  data = adqs_f,
  cor_struct = "unstructured",
  weights_emmeans = "equal",
  optimizer = "nloptwrap_neldermead" # Only to speed up this example.
)
}