Formatting Fraction with Lower Threshold

Formats a fraction when the second element of the input x is the fraction. It applies a lower threshold, below which it is just stated that the fraction is smaller than that.

format_fraction_threshold(threshold)

Arguments

threshold

(proportion)
lower threshold.

Value

An rtables Formatting Function that takes numeric input x where the second element is the fraction that is formatted. If the fraction is above or equal to the threshold, then it is displayed in percentage. If it is positive but below the threshold, it returns "<1" e.g. if the threshold is 0.01. If it is zero, then just "0" is returned.

See also

Examples

format_fun <- format_fraction_threshold(0.05)
format_fun(x = c(20, 0.1))
#> [1] 10
format_fun(x = c(2, 0.01))
#> [1] "<5"
format_fun(x = c(0, 0))
#> [1] "0"