Create a Forest Plot based on a Table

Usage

g_forest(
  tbl,
  col_x = attr(tbl, "col_x"),
  col_ci = attr(tbl, "col_ci"),
  vline = 1,
  forest_header = attr(tbl, "forest_header"),
  xlim = c(0.1, 10),
  logx = TRUE,
  x_at = c(0.1, 1, 10),
  width_row_names = NULL,
  width_columns = NULL,
  width_forest = grid::unit(1, "null"),
  col_symbol_size = attr(tbl, "col_symbol_size"),
  draw = TRUE,
  newpage = TRUE
)

Arguments

tbl

(rtable)

col_x

(integer)
column index with estimator. By default tries to get this from tbl attribute col_x, otherwise needs to be manually specified.

col_ci

(integer)
column index with confidence intervals. By default tries to get this from tbl attribute col_ci, otherwise needs to be manually specified.

vline

(number)
x coordinate for vertical line, if NULL then the line is omitted.

forest_header

(character, length 2)
text displayed to the left and right of vline, respectively. If vline = NULL then forest_header needs to be NULL too. By default tries to get this from tbl attribute forest_header.

xlim

(numeric)
limits for x axis.

logx

(flag)
show the x-values on logarithm scale.

x_at

(numeric)\cr x-tick locations, if NULL` they get automatically chosen.

width_row_names

(unit)
width for row names. If NULL the widths get automatically calculated. See grid::unit().

width_columns

(unit)
widths for the table columns. If NULL the widths get automatically calculated. See grid::unit().

width_forest

(unit)
width for the forest column. If NULL the widths get automatically calculated. See grid::unit().

col_symbol_size

(integer)
column index from tbl containing data to be used to determine relative size for estimator plot symbol. Typically, the symbol size is proportional to the sample size used to calculate the estimator. If NULL, the same symbol size is used for all subgroups. By default tries to get this from tbl attribute col_symbol_size, otherwise needs to be manually specified.

draw

(flag)
whether the plot should be drawn.

newpage

(flag)
whether the plot should be drawn on a new page. Only considered if draw = TRUE is used.

Value

(gtree) object containing the forest plot and table

Details

Create a forest plot from any rtables::rtable() object that has a column with a single value and a column with 2 values.

Examples