Helper Function: Survival Estimations

Transform a survival fit to a table with groups in rows characterized by N, median and confidence interval.

h_tbl_median_surv(fit_km, armval = "All")

Arguments

fit_km

(survfit)
result of survival::survfit().

armval

(string)
used as strata name when treatment arm variable only has one level. Default is "All".

Examples

if (FALSE) {
library(scda)
library(dplyr)
library(survival)

adtte <- synthetic_cdisc_data("latest")$adtte %>%
  filter(PARAMCD == "OS")

fit <- survfit(
  form = Surv(AVAL, 1 - CNSR) ~ ARMCD,
  data = adtte
)
h_tbl_median_surv(fit_km = fit)
}