Install a particular version of tern with the recommended versions of terns dependencies

For production use one should install tern using this function so that matching versions of tern and random.cdisc.data are installed. For a first time installation use the instructions on GitHub repository README.mdhttp://github.roche.com/Rpackages/tern.

install_tern(ref = NULL, dependencies = c("all", "depends", "none"),
  eval = TRUE)

Arguments

ref
either NULL, 'master', 'devel', or a version tag (i.e. 'v0.0.1'). If NULL then the current installed version is used.

dependencies
one of 'all', 'depends' or 'none'

eval
execute the installation. This is only useful to install ref='master', 'devel', or a version the preceeds the current installed version

Details

Currently the R packages tern, rtables, random.cdisc.data, teal.oncology, and teal are evolving rapidly and we currently do not guarantee backwards compatibility. Therefore tern and teal.oncology will provide the install_<pkg> functions which provide install instructions for recommended package versions.

Examples

install_tern(eval = FALSE)

#> Error in install_tern(eval = FALSE): no recommended version combination for tern v0.5.0 is  stored in the current installed package tern

install_tern('master', 'depends', eval = FALSE)

#> {
#>     devtools::install_github("Roche/rtables", ref = "master")
#>     devtools::install_github("Rpackages/tern", ref = "master", 
#>         host = "https://github.roche.com/api/v3")
#> }

install_tern('devel', 'none', eval = FALSE)

#> {
#>     devtools::install_github("Rpackages/tern", ref = "devel", 
#>         host = "https://github.roche.com/api/v3")
#> }