List of Adverse Events Terms Tables

lt_ae_max_grade_class_term returns a nested list of adverse events tables by unique id (t_count_unique).

lt_ae_class_term(terms, id, col_by, col_N = tapply(col_by, col_by, length),
  total = "All Patients")

Arguments

terms
term information as character or factor dataframe, however factor levels are not repeated by class, only terms with count > 1 are listed per class. Currently terms can only be a dataframe with 2 columns. var_relabel is used as the character string used as a label in the column header for each term.

id
unique subject identifier. If a particular subject has no adverse event then that information needs to be added to the col_N argument.

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing. See 'Examples'.

col_N
numeric vector with information of the number of patients in the levels of col_by. This is useful if there are patients that have no adverse events can be accounted for with this argument.

total
character string that will be used as a label for a column with pooled total population, default is "All Patients". If the levels of col_by are the only columns of interest then total should be NULL

Examples

# Simple example
library(tibble)
library(dplyr)

ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = factor(paste("ARM", LETTERS[rep(c(1,2), c(3,7))]))
)

ae_lookup <- tribble(
  ~CLASS,         ~TERM,   ~GRADE,
  "cl A",   "trm A_1/2",        1,
  "cl A",   "trm A_2/2",        2,
  "cl B",   "trm B_1/3",        2,
  "cl B",   "trm B_2/3",        3,
  "cl B",   "trm B_3/3",        1,
  "cl C",   "trm C_1/1",        1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2,2,2,3,3,4,4,4,4,5,6,6,7,7)]
  ),
  ae_lookup[c(1,1,2,6,4,2,2,3,4,2,1,5,4,6),]
)

ANL <- left_join(AAE, ASL[, c("USUBJID", "ARM")], by ="USUBJID")

terms <- data.frame(class = ANL$CLASS, term = ANL$TERM, stringsAsFactors = FALSE)
var_labels(terms) <- c("Class label", "Term label")

l_tbls <- lt_ae_class_term(
  terms = terms,
  id = ANL$USUBJID,
  col_by = ANL$ARM,
  col_N = tapply(ASL$ARM, ASL$ARM, length),
  total = "All Patients"
)
recursive_stack_rtables(l_tbls)

#> Class label                                                                  ARM A               ARM B            All Patients
#>   Term label                                                                 (N=3)               (N=7)               (N=10)   
#> ------------------------------------------------------------------------------------------------------------------------------
#> - Any Adverse Event -
#>   Overall total number of events                                               5                   9                   14     
#>   Overall total number of patients with at least one adverse event         2 (66.67%)          4 (57.14%)           6 (60%)   
#> 
#> cl A
#>   Total number of events                                                       3                   4                   7      
#>   Total number of patients with at least one adverse event                 1 (33.33%)          3 (42.86%)           4 (40%)   
#>   trm A_1/2                                                                1 (33.33%)          1 (14.29%)           2 (20%)   
#>   trm A_2/2                                                                1 (33.33%)          2 (28.57%)           3 (30%)   
#> 
#> cl B
#>   Total number of events                                                       1                   4                   5      
#>   Total number of patients with at least one adverse event                 1 (33.33%)          3 (42.86%)           4 (40%)   
#>   trm B_2/3                                                                1 (33.33%)          2 (28.57%)           3 (30%)   
#>   trm B_1/3                                                                  0 (0%)            1 (14.29%)           1 (10%)   
#>   trm B_3/3                                                                  0 (0%)            1 (14.29%)           1 (10%)   
#> 
#> cl C
#>   Total number of events                                                       1                   1                   2      
#>   Total number of patients with at least one adverse event                 1 (33.33%)          1 (14.29%)           2 (20%)   
#>   trm C_1/1                                                                1 (33.33%)          1 (14.29%)           2 (20%)   

# NOT RUN {
#throws an error because terms is not a dataframe
l_tbls <- lt_ae_class_term(
  terms =  ANL$TERM,
  id = ANL$USUBJID,
  col_by = ANL$ARM,
  col_N = tapply(ASL$ARM, ASL$ARM, length),
  total = "All Patients"
)
# }