rtabulate-friendly aggregation functions (a_*)

These functions enhance the control of missing data, including rcell dedicated formats and are intended to be used with rtables::rtabulate. Included:

a_mean_sd mean along with standard deviation.

a_median median.

a_n_true_and_freq number of positive cases and frequency.

a_count number of items.

a_q1q3 first and third quartiles.

a_iqr interquartile range.

a_range range.

Calculates median

Calculates proportions

Calculates counts

Calculates First and Third quartiles

Calculates Interquartile range

Calculates Range

a_mean_sd(x, na.rm = TRUE, na = NA, rcell = FALSE)

a_median(x, na.rm = TRUE, na = NA, rcell = FALSE)

a_n_true_and_freq(x, na.rm = TRUE, na = NA, rcell = FALSE)

a_count(x, na.rm = TRUE, na = NA, rcell = FALSE)

a_q1q3(x, na.rm = TRUE, na = NA, rcell = FALSE, type = 7)

a_iqr(x, na.rm = TRUE, na = NA, rcell = FALSE, type = 7)

a_range(x, na.rm = TRUE, na = NA, rcell = FALSE)

Arguments

x
A vector of numeric values to be aggregated.

na.rm
A logical value indicating whether NA values should be stripped before the computation proceeds. Note that in a_count, it indicates whether or not to return NA if all data are missing.

na
The string to use for missing values.

rcell
A logical value indicated when the return value should be a formatted rcell.

type
An integer between 1 and 9 to select the quantile algorithm, see stats::quantile().

Value

Either a numeric or a rcell.

Details

These functions offer similar control over the management of missing values, and the possibility to return a rcell object.

Note

These functions come as a replacement of non-exported functions:

a_mean_sd replaces mean_sd, mean_sd2, mean_sd3.

a_median replaces median_t2, median_t3.

a_n_true_and_freq replaces positives_and_proportion, positives_and_proportion2, positives_and_proportion3.

a_count replaces n_count, n_not_na2, n_not_na3.

a_q1q3 replaces q1_q3.

a_iqr replaces iqr_num, iqr_num2, iqr_num3.

a_range range range_t2, range_t3.

Examples

# Data
# ----

dta <- list(
    a = c(1, 2, 3),
    b = c(1, 2, NA),
    c = c(1, 1, NA),
    d = c(1),
    e = c(),
    f = c(NA, NA),
    g = NULL
  )

# Mean and standard deviation
# ---------------------------
lapply(dta, a_mean_sd)

#> $a
#> [1] 2 1
#> 
#> $b
#> [1] 1.5000000 0.7071068
#> 
#> $c
#> [1] 1 0
#> 
#> $d
#> [1]  1 NA
#> 
#> $e
#> [1] NA NA
#> 
#> $f
#> [1] NA NA
#> 
#> $g
#> [1] NA NA
#> 

# Median
# ------
lapply(dta, a_median, rcell = TRUE, na = "NE")

#> $a
#> 2
#> 
#> $b
#> 1.5
#> 
#> $c
#> 1
#> 
#> $d
#> 1
#> 
#> $e
#> NE
#> 
#> $f
#> NE
#> 
#> $g
#> NE
#> 

# Count positive cases and proportion
# -----------------------------------
booleans <- list(
  a = c(TRUE, TRUE, TRUE),
  b = c(FALSE, FALSE, FALSE),
  c = c(TRUE, FALSE, NA),
  d = c(),
  e = NULL
)

lapply(
  booleans,
  a_n_true_and_freq,
  na.rm = FALSE,
  na = "NE",
  rcell = TRUE
  )

#> Absolute and relative frequencies can't be estimated, x is not logical.

#> Absolute and relative frequencies can't be estimated, x is not logical.

#> $a
#> 3 (1)
#> 
#> $b
#> 0 (0)
#> 
#> $c
#> NE (NE)
#> 
#> $d
#> NE (NE)
#> 
#> $e
#> NE (NE)
#> 

# Count items with explicit rules for NA
# --------------------------------------
lapply(dta, a_count, na.rm = FALSE, rcell = TRUE, na = "NE")

#> $a
#> 3
#> 
#> $b
#> 2
#> 
#> $c
#> 2
#> 
#> $d
#> 1
#> 
#> $e
#> NE
#> 
#> $f
#> NE
#> 
#> $g
#> NE
#> 

# First and third quartiles
# -------------------------
dta_quant <- list(
  a = c(1, 1, 2, 3, 4, 4, 5, 7, 8, 9),    # n = 10
  b = c(1, 1, 2, 3, 4, 4, 5, 7, 8, 9, 9), # n = 11
  c = c(1, 1, 2, 3, 4, 4, 5, 7, 8, NA),   # n = 9 + 1 NA
  d = c(NA, NA),
  e = c(),
  f = NULL
)
lapply(dta_quant, a_q1q3, na = "NE", rcell = TRUE, na.rm = FALSE)

#> $a
#> 2.25 - 6.5
#> 
#> $b
#> 2.5 - 7.5
#> 
#> $c
#> NE - NE
#> 
#> $d
#> NE - NE
#> 
#> $e
#> NE - NE
#> 
#> $f
#> NE - NE
#> 

# Interquartile range
# --------------------
lapply(dta_quant, a_iqr, rcell = TRUE)

#> $a
#> 4.25
#> 
#> $b
#> 5
#> 
#> $c
#> 3
#> 
#> $d
#> NA
#> 
#> $e
#> NA
#> 
#> $f
#> NA
#> 

# Range
# -----
lapply(dta_quant, a_range, rcell = TRUE, na = "not avail.")

#> $a
#> 1 - 9
#> 
#> $b
#> 1 - 9
#> 
#> $c
#> 1 - 8
#> 
#> $d
#> not avail. - not avail.
#> 
#> $e
#> not avail. - not avail.
#> 
#> $f
#> not avail. - not avail.
#>