Stack Tables Stored in a nested list of depth 2

recursive_stack_rtables expects a list with lists of rtables to be stacked. Sometimes these tables have repeated information at the top and hence the first n rows can be optionally removed from the tables that are not first in the lists.

recursive_stack_rtables(x)

Arguments

x
list with lists of rtables

Value

rtable

Examples

l_tbls <- list(
  list(
     rtabulate(iris$Sepal.Length, iris$Species, mean),
     rtabulate(iris$Sepal.Length, iris$Species, sd)
  ),
  list(
     rtabulate(iris$Sepal.Width, iris$Species, mean),
     rtabulate(iris$Sepal.Width, iris$Species, sd)
  ),
  list(
     rtabulate(iris$Petal.Length, iris$Species, mean),
     rtabulate(iris$Petal.Length, iris$Species, sd)
  )
)

tern:::recursive_stack_rtables(l_tbls)

#>                  setosa                 versicolor                virginica    
#> -------------------------------------------------------------------------------
#> mean              5.006                    5.936                    6.588      
#> 
#> sd          0.352489687213451        0.516171147063863        0.635879593274432
#> 
#> mean              3.428                    2.77                     2.974      
#> 
#> sd          0.379064369096289        0.313798323378411        0.322496638172637
#> 
#> mean              1.462                    4.26                     5.552      
#> 
#> sd          0.173663996480184        0.469910977239958        0.551894695663983