Response Table Caclulation Function

Response Table Caclulation Function

response_table(response, value.resp = c("CR", "PR"), value.nresp = c("SD",
  "NON CR/PD", "PD", "NE"), arm, strata_data = NULL, style = 1)

Arguments

response
Tumor response data

arm
Arm information data as factor

strata_data
data for stratification factors (categorical variables)

style
Must be 1 or 2, 1 if only want to display rates summary for each response value category, 2 if want to display rates, difference in rate and odds ratio for each response value category.

value.rsp
Character vector, defining list of response values to be used as responders

value.nrsp
Character vector, defining list of response values to be used as non-responders

Details

We use the epiR package to compute 2-by-2 contingency table statistics and analysis with stratification factors.

If no stratification, chi-squared test is performed by default to test for equality of proportions in response rate between reference and comparison arms.

If stratification factors are added, Cochran-Mantel-Haensel test statistics are reported for the test of equal proportions, odds ratio and its corresponding 95

Examples

library(random.cdisc.data)
library(forcats)

ASL <- radam("ASL")
ARS <- radam("ARS", ADSL = ASL)

ANL <- merge(ASL, ARS) %>%
  filter(PARAMCD == "OVRSPI")

tbl <- response_table(
 response = ANL$AVALC,
 value.resp = c("CR", "PR"),
 value.nresp = setdiff(ANL$AVALC, c("CR", "PR")),
 arm = fct_relevel(factor(ANL$ARMCD), "ARM B", "ARM A")
)

tbl

#>                                                                                 ARM B                 ARM A     
#>                                                                                 (N=39)                (N=61)    
#> ----------------------------------------------------------------------------------------------------------------
#> Responders                                                                    21 (53.8%)             25 (41%)   
#> Non-Responders                                                                18 (46.2%)             36 (59%)   
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                                 (37.18, 69.91)        (28.55, 54.32)
#> 
#> Difference in Response Rates                                                               -12.86               
#>   95% CI for difference (Wald without correction)#                                     (-32.79, 7.07)           
#>   95% CI for difference (Wald with continuity correction)#                             (-34.89, 9.17)           
#>   p-value## (Chi-squared)                                                                  0.2081               
#> 
#> Odds Ratio#                                                                                 0.6                 
#>   95% CI#                                                                               (0.26, 1.34)            
#> 
#> Complete Response (CR)                                                        5 (12.8%)             10 (16.4%)  
#>   95% CI (Wald)                                                              (4.3, 27.43)         (8.15, 28.09) 
#> 
#> Partial Response (PR)                                                          16 (41%)             15 (24.6%)  
#>   95% CI (Wald)                                                             (25.57, 57.9)         (14.46, 37.29)
#> 
#> Stable Disease (SD)                                                            3 (7.7%)             15 (24.6%)  
#>   95% CI (Wald)                                                             (1.62, 20.87)         (14.46, 37.29)
#> 
#> Progressive Disease (PD)                                                      6 (15.4%)             10 (16.4%)  
#>   95% CI (Wald)                                                             (5.86, 30.53)         (8.15, 28.09) 
#> 
#> Missing or unevaluable                                                        9 (23.1%)              11 (18%)   
#> 
#> # USE WITH CAUTION! One or more cells has a frequency count of <= 5.

# NOT RUN {
library(dplyr)
library(atezo.data)
library(teal.oncology)
library(forcats)
library(epiR)

ARS <- ars(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

# Stream example

tbl_stream <- get_response_table(com.roche.cdt30019.go29436.re)
Viewer(tbl_stream)

ANL <- ARS %>%
  filter(PARAMCD == "OVRSPI", ITTGEFL == 'Y', ITTWTFL == 'Y') %>%
  mutate(AVALC = ifelse(AVALC == "", "NE", AVALC))

tbl <- response_table(
 response = ANL$AVALC,
 value.resp = c("CR", "PR"),
 value.nresp = setdiff(ANL$AVALC, c("CR", "PR")),
 arm = fct_relevel(ANL$ARMCD, "C", "B", "A")
)

Viewer(tbl, tbl_stream)

compare_rtables(tbl, tbl_stream, comp.attr = FALSE)

# Other examples

paramcd = "OVRSPI"
var_arm = "ARM"
ref_arm = "DUMMY C"
comp_arm = c("DUMMY B", "DUMMY A")
combine_arm = FALSE
responders = c("CR", "PR")
var_strata = NULL

ANL <- ARS %>%
  filter(PARAMCD == paramcd, ARM %in% c(ref_arm, comp_arm))

# If want to include missing as non-responders
ANL$AVALC[ANL$AVALC==""] <- "NE"

arm1 <- factor(ANL$ARM)
ARM <- fct_relevel(arm1, ref_arm, comp_arm)

tbl <- response_table(
 response = ANL$AVALC,
 value.resp = c("CR", "PR"),
 value.nresp = setdiff(ANL$AVALC, c("CR", "PR")),
 arm = ARM,
 strata_data = if (!is.null(var_strata)) ANL[var_strata] else NULL,
 style = 1
)

Viewer(tbl)
#clean up
rm(paramcd, var_arm, ref_arm, comp_arm, combine_arm, responders, var_strata)

# Example 2
paramcd = "OVRSPI"
var_arm = "ARM"
ref_arm = c("DUMMY C", "DUMMY B")
comp_arm = "DUMMY A"
combine_arm = TRUE
responders = c("CR", "PR")
var_strata = "SEX"

refname <- paste0(ref_arm, collapse = "/")
ARM <- do.call(fct_collapse, setNames(list(arm1, ref_arm), c("f", refname))) %>%
  fct_relevel(refname, comp_arm)

#style 2 
tbl2 <- response_table(response = ANL$AVALC,
                       value.resp = c("CR", "PR"),
                       value.nresp = setdiff(ANL$AVALC, c("CR", "PR")),
                       arm = ARM,
                       strata_data = if (!is.null(var_strata)) ANL[var_strata] else NULL,
                       style = 2)
Viewer(tbl2)

# }

'%needs%' <- teal.oncology:::'%needs%' # for debugging purpuses