Response Table with ADaM data structure

Response Table with ADaM data structure

response_table_ADAM(ASL, ARS, paramcd = "OVRSPI", arm.var = "ARM", arm.ref,
  arm.comp, arm.comp.combine = TRUE, value.resp = c("CR", "PR"),
  value.nresp = c("SD", "NON CR/PD", "PD", "NE"), incl.missing = TRUE,
  strata.var = NULL, style = 1)

Arguments

ASL
dataset with following variables: USUBJID, STUDYID, the specified grouping variable

ARS
ARS dataset containing the following variables: USUBJID, STUDYID, PARAMCD, AVALC

paramcd
Name of overall response parameter

arm.var
Name of variable with arm information

arm.ref
Character vector, defining which arm(s) from the list of arms should be used as reference group

arm.comp
Character vector, defining which arm(s) from the list of arms should be used as comparison group

arm.comp.combine
Logical, TRUE if want all non-ref arms to be combined into one comparison group, FALSE if want each of the non-ref arms to be a separate comparison group

incl.missing
Logical, TRUE if missing values should be considered non-responders, FALSE if missing response should be removed from analysis

style
Must be 1 or 2, 1 if only want to display rates summary for each response value category, 2 if want to display rates, difference in rate and odds ratio for each response value category.

value.rsp
Character vector, defining list of response values to be used as responders

value.nrsp
Character vector, defining list of response values to be used as non-responders

strata
Character vector, defining stratification factor for stratified analysis, default is NULL

Details

Package forcats was used to re-format arm data into leveled factors. Package epiR was used to compute 2-by-2 contingency table statistics and analysis with stratification factors.

Examples

# NOT RUN {
library(dplyr)
library(atezo.data)
library(teal.oncology)
library(forcats)
library(epiR)

'%needs%' <- teal.oncology:::'%needs%'

ASL <- asl(com.roche.cdt30019.go29436.re)
ARS <- ars(com.roche.cdt30019.go29436.re)

teal.oncology:::response_table_ADAM(ASL, ARS,
                    paramcd = "OVRSPI",
                    arm.ref = "DUMMY C",
                    arm.comp = c("DUMMY A", "DUMMY B"),
                    arm.comp.combine = FALSE,
                    value.resp = c("CR", "PR"),
                    value.nresp = c("SD", "NON CR/PD", "PD", "NE"),
                    incl.missing = TRUE,
                    strata.var = "SEX",
                    style = 1
                    )
# }