Replace format of rtables cells

A Generic function to replace format of given rtables cells. This is convenient when, for instance, some data should rather be blanked during a post-processing step.

Replace format of `rtables`

rreplace_format(
  x,
  row = NULL,
  col = NULL,
  old = FALSE,
  new = function(x, output) ""
)

# S3 method for rtable
rreplace_format(
  x,
  row = NULL,
  col = NULL,
  old = FALSE,
  new = function(x, output) ""
)

Arguments

x
A rtable.

row
A vector of row indexes. Default `NULL` applies to all rows.

col
A vector of columns indexes. Default `NULL` applies to all columns.

old
A targeted `rtable` format to be replaced. Default `FALSE`, indicates a replacement independent from current format, otherwise all `rtables` compatible format are accepted.

new
The replacement format.

Examples

library(random.cdisc.data)
library(tern)
library(dplyr)

ADSL <- radsl(cached = TRUE)
ADLB <- radlb(cached = TRUE)
saved_labels <-  var_labels(ADLB)

# For illustration purpose, the example focuses on "Alanine Aminotransferase
# Measurement" starting from baseline, while excluding visit at week 1 for
# subjects who were prescribed a placebo.
ADLB <- subset(
  ADLB,
  PARAM == "Alanine Aminotransferase Measurement" &
    !(ARM == "B: Placebo" & AVISIT == "WEEK 1 DAY 8") & AVISIT != "SCREENING"
)

# Preprocessing.
ADLB_l <- split(
  ADLB, f = ifelse(ADLB$AVISIT == "BASELINE", "Baseline", "Follow-up")
)
ADLB_l <- lapply(ADLB_l, `var_labels<-`, saved_labels)
tbls1 <- lapply(X = ADLB_l, function(x) {

  tbl <- t_summary_by(
    x = compare_in_header(x[, c("AVAL", "CHG")]),
    row_by = x$AVISIT,
    col_by = x$ARM,
    col_N = table(ADSL$ARM),
    f_numeric = c("count_n", "mean_sd3", "median_t3", "range_t3")
  )
  return(tbl)

})

rreplace_format(tbls1$Baseline, col = 3)

#> Analysis Visit                                    A: Drug X                                                                 B: Placebo                                                              C: Combination                          
#>                                                    (N=134)                                                                   (N=134)                                                                   (N=132)                              
#>                              Analysis Value                Absolute Change from Baseline               Analysis Value                Absolute Change from Baseline               Analysis Value                Absolute Change from Baseline
#> --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> BASELINE
#>   n                                134                                  134                                                                       134                                  132                                  132             
#>   Mean (SD)                   49.57 (8.32)                             0 (0)                                                                     0 (0)                             50.9 (7.82)                             0 (0)            
#>   Median                          49.57                                  0                                                                         0                                  50.77                                  0              
#>   Min - Max                   23.97 - 70.85                            0 - 0                                                                     0 - 0                            27.63 - 67.45                            0 - 0            

rreplace_format(tbls1$Baseline, row = 2, old = NULL, new = "xx%")

#> Analysis Visit                                    A: Drug X                                                                 B: Placebo                                                              C: Combination                          
#>                                                    (N=134)                                                                   (N=134)                                                                   (N=132)                              
#>                              Analysis Value                Absolute Change from Baseline               Analysis Value                Absolute Change from Baseline               Analysis Value                Absolute Change from Baseline
#> --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> BASELINE
#>   n                              13400%                               13400%                               13400%                               13400%                               13200%                               13200%            
#>   Mean (SD)                   49.57 (8.32)                             0 (0)                            50.31 (8.34)                             0 (0)                             50.9 (7.82)                             0 (0)            
#>   Median                          49.57                                  0                                  50.15                                  0                                  50.77                                  0              
#>   Min - Max                   23.97 - 70.85                            0 - 0                            26.22 - 79.12                            0 - 0                            27.63 - 67.45                            0 - 0