Split list recursively according to by and return the associated tree

Split list recursively according to by and return the associated tree

rsplit_to_tree(lst, by_lst, drop_empty_levels = TRUE, non_leaves_null = FALSE)

Arguments

lst
list to split, esplit will be applied to all list elements

by_lst
list of columns, each of which is a factor to recursively split by

drop_empty_levels
whether to drop empty levels, this happens often when you have, e.g. two factors, one with levels (clA, clB), the other with levels (clA_1, clA_2, clB_1, clB_2) and only the combinations clA-clA_1, clA-clA_2, clB-clB_1, clB-clB_2 appear out of the eight combinations.

non_leaves_null
whether to assign NULL content to any non-leaves

Value

node object

Examples

tree <- rsplit_to_tree(
  1:5,
  by_lst = list(factor(c("M", "M", "F", "F", "F")), factor(c("O", "Y", "Y", "Y", "Y")))
)
summary(tree)

#>                index            content             dim            hide label          format  
#> -----------------------------------------------------------------------------------------------
#> root                            integer                              FALSE                     
#>   F              1              integer                              FALSE                     
#>     Y           1, 1            integer                              FALSE                     
#>   M              2              integer                              FALSE                     
#>     O           2, 1            integer                              FALSE                     
#>     Y           2, 2            integer                              FALSE                     

cat(displayable(tree))

#> root:
#>   1, 2, 3, 4, 5
#>   F:
#>     3, 4, 5
#>     Y:
#>       3, 4, 5
#>   M:
#>     1, 2
#>     O:
#>       1
#>     Y:
#>       2

by_lst <- list(
  factor(c(rep("clA", 4), rep("clB", 4))),
  factor(c(rep(c("A1", "A2"), 2), rep(c("B1", "B2"), 2)))
)
summary(rsplit_to_tree(1:8, by_lst))

#>                 index            content             dim            hide label          format  
#> ------------------------------------------------------------------------------------------------
#> root                             integer                              FALSE                     
#>   clA             1              integer                              FALSE                     
#>     A1           1, 1            integer                              FALSE                     
#>     A2           1, 2            integer                              FALSE                     
#>   clB             2              integer                              FALSE                     
#>     B1           2, 1            integer                              FALSE                     
#>     B2           2, 2            integer                              FALSE                     

summary(rsplit_to_tree(1:8, by_lst, drop_empty_levels = FALSE))

#>                 index            content             dim            hide label          format  
#> ------------------------------------------------------------------------------------------------
#> root                             integer                              FALSE                     
#>   clA             1              integer                              FALSE                     
#>     A1           1, 1            integer                              FALSE                     
#>     A2           1, 2            integer                              FALSE                     
#>     B1           1, 3            integer                              FALSE                     
#>     B2           1, 4            integer                              FALSE                     
#>   clB             2              integer                              FALSE                     
#>     A1           2, 1            integer                              FALSE                     
#>     A2           2, 2            integer                              FALSE                     
#>     B1           2, 3            integer                              FALSE                     
#>     B2           2, 4            integer                              FALSE