Summary for analysis of covariance (ANCOVA).

Summary for analysis of covariance (ANCOVA).

s_ancova(formula, data, conf_level = 0.95)

Arguments

formula
A formula corresponding to the investigated linear model. The left-hand side must include the dependent variable. The right-hand side must include the group variable name wrapped in the special arm().

data
A data.frame which includes all the variables that are called in formula.

conf_level
The confidence level of the resulting confidence intervals (default 0.95).

Value

Named list with analysis results as summary_emm objects, which are data.frame objects with a special print method:

sum_fit

Summary of the estimated marginal means in the groups.

sum_contrasts

Summary of the estimated contrasts of the marginal means. Here the first level of the arm variable is taken as the reference or control group, and all other groups are compared with this control group. Specifically, the differences of the marginal means are estimated.

Examples

# Estimate the adjusted means of sepal length in the different iris species,
# adjusting for covariates:
result <- s_ancova(
  formula = Sepal.Length ~ arm(Species) + Sepal.Width + Petal.Length + Petal.Width,
  data = iris
)
result

#> $sum_fit
#>  Species    emmean     SE  df lower.CL upper.CL n_total n_complete
#>  setosa       6.43 0.1914 144     6.05     6.80      50         50
#>  versicolor   5.70 0.0649 144     5.57     5.83      50         50
#>  virginica    5.40 0.1487 144     5.11     5.70      50         50
#> 
#> Confidence level used: 0.95 
#> 
#> $sum_contrasts
#>  contrast            estimate    SE  df lower.CL upper.CL t.ratio p.value
#>  versicolor - setosa   -0.724 0.240 144    -1.20   -0.249 -3.013  0.0031 
#>  virginica - setosa    -1.023 0.334 144    -1.68   -0.364 -3.067  0.0026 
#> 
#> Confidence level used: 0.95 
#> 

# We can see that the adjusted mean sepal length in the "setosa" species is significantly
# higher than in the "versicolor" and "virginica" species (p-values 0.003 in both differences).