Multivariate Cox Model - summarized results

Analyses based on multivariate Cox model are usually not performed for the Controlled Substance Reporting or regulatory documents but serve exploratory purposes only (e.g., for publication). In practice, the model usually includes only the main effects (without interaction terms). It produces the hazard ratio estimates for each of the covariates included in the model. The analysis follows the same principles (e.g., stratified vs. unstratified analysis and tie handling) as the usual Cox model analysis. Since there is usually no pre-specified hypothesis testing for such analysis, the p.values need to be interpreted with caution. (Statistical Analysis of Clinical Trials Data with R, NEST's bookdown)

s_cox_multivariate(
  formula,
  data,
  conf_level = 0.95,
  pval_method = c("wald", "likelihood"),
  ...
)

Arguments

formula

A formula corresponding to the investigated survival model including covariates.

data

A data.frame which includes the variable in formula and covariates.

conf_level

The level of confidence for the hazard ration interval estimations. Default is 0.95.

pval_method

The method used for the estimation of p.values, should be one of "wald" (default) or "likelihood".

...

Optional parameters passed to coxph()

ties a character string specifying the method for tie handling, one of exact (default), efron, breslow.

Details

The output is limited to single effect terms. Work in ongoing for estimation of interaction terms but is out of scope as defined by the Global Data Standards Repository (GDS_Standard_TLG_Specs_Tables_2.doc).

Examples

library(scda)
library(dplyr)

ADTTE <- synthetic_cdisc_data("latest")$adtte
ADTTE_f <- subset(ADTTE, PARAMCD == "OS") # _f: filtered
ADTTE_f <- filter(
  ADTTE_f,
  PARAMCD == "OS" &
    SEX %in% c("F", "M") &
    RACE %in% c("ASIAN", "BLACK OR AFRICAN AMERICAN", "WHITE")
)
ADTTE_f$SEX <- droplevels(ADTTE_f$SEX)
ADTTE_f$RACE <- droplevels(ADTTE_f$RACE)
if (FALSE) {
s_cox_multivariate(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ (ARMCD + RACE + AGE)^2, data = ADTTE_f
)
}