Pairwise model

Run Pairwise (ARM) CoxPH model for unstratified and stratified analysis

s_coxph_pairwise(
  formula,
  data,
  conf_level = 0.95,
  pval_method = c("log-rank", "wald", "likelihood"),
  ...
)

Arguments

formula
(formula)
Specifies survival model. The arm variable needs to be wrapped in arm. The strata special will only be used for the stratified analysis. If there is not strata specification then the stratified analysis is omitted.

data
(data.frame)
Contains all the variable that are used in formula

conf_level
(numeric value)
level for computation of the confidence intervals. If set to FALSE no confidence intervals are printed

pval_method
(one of ("log-rank", "likelihood", "wald"))
Specifies the method used to calculate the p-value. Default value is "log-rank".

...
(optional)
Other arguments will be passed to coxph

Value

a list of dataframes with unstratified and/or stratified analysis results

Examples

library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADTTE <- radtte(cached = TRUE)
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")

cox_res <-  s_coxph_pairwise(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARM) + strata(RACE),
  data = ADTTE_f,
  pval_method = "wald",  ties = "exact"
)
names(cox_res)

#> [1] "A: Drug X"      "B: Placebo"     "C: Combination"

cox_res[[2]]$stratified

#> $pvalue
#> [1] 0.8437855
#> 
#> $hr
#> [1] 0.9699648
#> 
#> $hr_ci
#> lower .95 upper .95 
#> 0.7161865 1.3136684 
#> 
#> $conf_level
#> [1] 0.95
#> 
#> $pval_method
#> [1] "wald"
#> 

cox_res2 <- s_coxph_pairwise(
  formula =  Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARM),
  data = ADTTE_f,
  conf_level = 0.8, pval_method = "likelihood",  ties = "breslow"
)
cox_res2[[2]]$stratified

#> NULL

cox_res2[[2]]$unstratified

#> $pvalue
#> [1] 0.8510999
#> 
#> $hr
#> [1] 1.028742
#> 
#> $hr_ci
#> lower .80 upper .80 
#> 0.8477249 1.2484110 
#> 
#> $conf_level
#> [1] 0.8
#> 
#> $pval_method
#> [1] "likelihood"
#>