event rate adjusted per person-years at Risk

Table for event rate adjusted for Patient-Years

s_events_patyear(
  events,
  time_in_years,
  col_by,
  col_N = NULL,
  total = NULL,
  conf_level = 0.95,
  conf_method = c("normal", "exact", "byar"),
  lambda = 100
)

Arguments

events
(logical vector) TRUE if record is an event

time_in_years
(numeric vector) time at risk in year

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

conf_level
confidence level of the interval.

conf_method
confidence method ("normal", "exact", "byar"), see details of ci_patyear

lambda
a numeric rate factor, e.g. 100 for 100 person-years

Examples

library(dplyr)
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADAETTE <- radaette(cached = TRUE)
ANL <- ADAETTE %>%
   dplyr::filter(PARAM == "Time to first occurrence of any adverse event")
s_patyear <- s_events_patyear(
   events = !(ANL$CNSR),
   time_in_years = ANL$AVAL,
   col_by = ANL$ARM,
   col_N = table(ADSL$ARM),
   total = "All Patients",
   conf_level = 0.95,
   conf_method = "normal",
   lambda = 100
)

s_patyear$events_by_col

#> $`A: Drug X`
#> [1] 80
#> 
#> $`B: Placebo`
#> [1] 106
#> 
#> $`C: Combination`
#> [1] 81
#> 
#> $`All Patients`
#> [1] 267
#>