Summary

Summary of logistic regression with one two-way interaction term

s_logistic_interaction(glm_model, conf_level = 0.95, increments = NULL)

Arguments

glm_model
(glm model object)
The model object can be all main effect model, and a model with one two-way interaction.

conf_level
(numeric value)
Confidence level for Wald odds ratio confidence interval.

increments
(named list)
Used to specify numeric values of continuous variables in glm model formula which interact with other variables. This is used to calculate the odds ratio when comparing the other interaction variable effect. For example, for a model with ARM and AGE interaction, increments = list(AGE = c(18, 65)) will enable calculation of odds ratios of comparison ARM vs. reference ARM at AGE = 18 and AGE = 65. If increments = NULL, then default AGE value is ceiling of median.

Examples

library(random.cdisc.data)
library(dplyr)
library(purrr)

ADSL <- radsl(cached = TRUE)
ADSL <- ADSL %>% mutate(
 SEX = as.character(SEX),
 SEX = case_when(!SEX %in% c("F", "M") ~ "U",
               TRUE ~ SEX))
ADRS <- radrs(ADSL, seed = 2)
ADRS_f <- subset(ADRS, PARAMCD == "BESRSPI") %>%
  mutate(Response = case_when(AVALC %in% c("PR", "CR") ~ 1,
                              TRUE ~ 0))
glm_model <- glm(
 formula = Response ~ ARM + AGE + SEX + ARM*SEX,
 data = ADRS_f,
 family = "binomial")
s_logistic_interaction(glm_model, conf_level = 0.90)

#> $N
#> [1] 400
#> 
#> $results
#> $results$ARM
#> $results$ARM$main
#>     df p-value
#> ARM  2 0.73452
#> 
#> $results$ARM$predictor
#> $results$ARM$predictor$term_type
#> [1] "categorical"
#> 
#> $results$ARM$predictor$interact_with
#> [1] "SEX"
#> 
#> $results$ARM$predictor$counts_by_level
#> 
#>      A: Drug X     B: Placebo C: Combination 
#>            134            134            132 
#> 
#> $results$ARM$predictor$term_ref_level
#> [1] "A: Drug X"
#> 
#> $results$ARM$predictor$term_comp_level
#> [1] "B: Placebo"     "C: Combination"
#> 
#> 
#> $results$ARM$summary
#> $results$ARM$summary$`B: Placebo`
#> $results$ARM$summary$`B: Placebo`$summary_comp_level
#>                     coef        se   p-value df
#> ARMB: Placebo -0.6930258 0.8822656 0.4321564  1
#> 
#> $results$ARM$summary$`B: Placebo`$summary_with_interaction
#> $results$ARM$summary$`B: Placebo`$summary_with_interaction$F
#>          or      lcl      ucl
#> 1 0.5000607 0.117159 2.134371
#> 
#> $results$ARM$summary$`B: Placebo`$summary_with_interaction$M
#>          or        lcl       ucl
#> 1 0.1000208 0.01686135 0.5933189
#> 
#> 
#> 
#> $results$ARM$summary$`C: Combination`
#> $results$ARM$summary$`C: Combination`$summary_comp_level
#>                       coef       se   p-value df
#> ARMC: Combination 16.89404 2118.627 0.9936377  1
#> 
#> $results$ARM$summary$`C: Combination`$summary_with_interaction
#> $results$ARM$summary$`C: Combination`$summary_with_interaction$F
#>         or lcl ucl
#> 1 21726408   0 Inf
#> 
#> $results$ARM$summary$`C: Combination`$summary_with_interaction$M
#>         or lcl ucl
#> 1 15759485   0 Inf
#> 
#> 
#> 
#> 
#> 
#> $results$AGE
#> $results$AGE$main
#>     df p-value
#> AGE  1 0.80309
#> 
#> $results$AGE$predictor
#> $results$AGE$predictor$term_type
#> [1] "continuous"
#> 
#> $results$AGE$predictor$term_ref_level
#> NULL
#> 
#> $results$AGE$predictor$term_comp_level
#> NULL
#> 
#> $results$AGE$predictor$counts_by_level
#> NULL
#> 
#> 
#> $results$AGE$summary
#>            coef         se   p-value       or df       lcl      ucl
#> AGE 0.008827023 0.03540005 0.8030896 1.008866  1 0.9517995 1.069354
#> 
#> 
#> $results$SEX
#> $results$SEX$main
#>     df p-value
#> SEX  1 0.79825
#> 
#> $results$SEX$predictor
#> $results$SEX$predictor$term_type
#> [1] "categorical"
#> 
#> $results$SEX$predictor$interact_with
#> [1] "ARM"
#> 
#> $results$SEX$predictor$counts_by_level
#> 
#>   F   M 
#> 231 169 
#> 
#> $results$SEX$predictor$term_ref_level
#> [1] "F"
#> 
#> $results$SEX$predictor$term_comp_level
#> [1] "M"
#> 
#> 
#> $results$SEX$summary
#> $results$SEX$summary$M
#> $results$SEX$summary$M$summary_comp_level
#>           coef       se   p-value df
#> SEXM 0.3170556 1.240401 0.7982541  1
#> 
#> $results$SEX$summary$M$summary_with_interaction
#> $results$SEX$summary$M$summary_with_interaction$`A: Drug X`
#>         or     lcl      ucl
#> 1 1.373079 0.17849 10.56275
#> 
#> $results$SEX$summary$M$summary_with_interaction$`B: Placebo`
#>          or        lcl       ucl
#> 1 0.2746395 0.09437353 0.7992374
#> 
#> $results$SEX$summary$M$summary_with_interaction$`C: Combination`
#>          or lcl ucl
#> 1 0.9959776   0 Inf
#> 
#> 
#> 
#> 
#> 
#> $results$`ARM:SEX`
#> $results$`ARM:SEX`$main
#>         df p-value
#> ARM:SEX  2 0.51346
#> 
#> $results$`ARM:SEX`$predictor
#> $results$`ARM:SEX`$predictor$term_type
#> [1] "interaction"
#> 
#> 
#> $results$`ARM:SEX`$summary
#>                                   coef          se   p-value df
#> B: Placebo * M , n = 52     -1.6093515    1.393834 0.2482452  1
#> C: Combination * M , n = 62 -0.3210861 3091.239514 0.9999171  1
#> 
#> 
#> 

glm_model2 <- glm(
 formula = Response ~ ARM + AGE + ARM*AGE,
 data = ADRS_f,
 family = "binomial")
s_logistic_interaction(glm_model2, conf_level = 0.90)

#> $N
#> [1] 400
#> 
#> $results
#> $results$ARM
#> $results$ARM$main
#>     df p-value
#> ARM  2 0.50272
#> 
#> $results$ARM$predictor
#> $results$ARM$predictor$term_type
#> [1] "categorical"
#> 
#> $results$ARM$predictor$interact_with
#> [1] "AGE"
#> 
#> $results$ARM$predictor$counts_by_level
#> 
#>      A: Drug X     B: Placebo C: Combination 
#>            134            134            132 
#> 
#> $results$ARM$predictor$term_ref_level
#> [1] "A: Drug X"
#> 
#> $results$ARM$predictor$term_comp_level
#> [1] "B: Placebo"     "C: Combination"
#> 
#> 
#> $results$ARM$summary
#> $results$ARM$summary$`B: Placebo`
#> $results$ARM$summary$`B: Placebo`$summary_comp_level
#>                    coef       se   p-value df
#> ARMB: Placebo -4.113091 3.507083 0.2408779  1
#> 
#> $results$ARM$summary$`B: Placebo`$summary_with_interaction
#> $results$ARM$summary$`B: Placebo`$summary_with_interaction$`34`
#>          or        lcl       ucl
#> 1 0.2131818 0.06777549 0.6705444
#> 
#> 
#> 
#> $results$ARM$summary$`C: Combination`
#> $results$ARM$summary$`C: Combination`$summary_comp_level
#>                       coef      se   p-value df
#> ARMC: Combination 14.46513 7273.15 0.9984131  1
#> 
#> $results$ARM$summary$`C: Combination`$summary_with_interaction
#> $results$ARM$summary$`C: Combination`$summary_with_interaction$`34`
#>         or lcl ucl
#> 1 18121525   0 Inf
#> 
#> 
#> 
#> 
#> 
#> $results$AGE
#> $results$AGE$main
#>     df p-value
#> AGE  1  0.4434
#> 
#> $results$AGE$predictor
#> $results$AGE$predictor$term_type
#> [1] "continuous"
#> 
#> $results$AGE$predictor$interact_with
#> [1] "ARM"
#> 
#> 
#> $results$AGE$summary
#> $results$AGE$summary$summary_term
#>            coef         se   p-value df
#> AGE -0.06610236 0.08624352 0.4434016  1
#> 
#> $results$AGE$summary$summary_with_interaction
#> $results$AGE$summary$summary_with_interaction$`A: Drug X`
#>         or       lcl      ucl
#> 1 0.936035 0.8122393 1.078699
#> 
#> $results$AGE$summary$summary_with_interaction$`B: Placebo`
#>         or       lcl      ucl
#> 1 1.009456 0.9459243 1.077255
#> 
#> $results$AGE$summary$summary_with_interaction$`C: Combination`
#>   or           lcl           ucl
#> 1  1 5.045044e-144 1.982143e+143
#> 
#> 
#> 
#> 
#> $results$`ARM:AGE`
#> $results$`ARM:AGE`$main
#>         df p-value
#> ARM:AGE  2 0.72847
#> 
#> $results$`ARM:AGE`$predictor
#> $results$`ARM:AGE`$predictor$term_type
#> [1] "interaction"
#> 
#> 
#> $results$`ARM:AGE`$summary
#>                               coef           se   p-value df
#> B: Placebo, n = 134     0.07551414   0.09486712 0.4260325  1
#> C: Combination, n = 132 0.06610236 200.59771408 0.9997371  1
#> 
#> 
#> 

glm_model3 <- glm(
 formula = Response ~ ARM + AGE + BMRKR1 + BMRKR1*AGE,
 data = ADRS_f,
 family = "binomial")
s_logistic_interaction(glm_model3 , conf_level = 0.90)

#> $N
#> [1] 400
#> 
#> $results
#> $results$ARM
#> $results$ARM$main
#>     df  p-value
#> ARM  2 0.086012
#> 
#> $results$ARM$predictor
#> $results$ARM$predictor$term_type
#> [1] "categorical"
#> 
#> $results$ARM$predictor$term_ref_level
#> [1] "A: Drug X"
#> 
#> $results$ARM$predictor$term_comp_level
#> [1] "B: Placebo"     "C: Combination"
#> 
#> $results$ARM$predictor$counts_by_level
#> 
#>      A: Drug X     B: Placebo C: Combination 
#>            134            134            132 
#> 
#> 
#> $results$ARM$summary
#>                     coef           se    p-value           or df        lcl
#> B: Placebo     -1.466571    0.6620967 0.02675729 2.307153e-01  1 0.07764384
#> C: Combination 16.777124 1541.7908469 0.99131792 1.932914e+07  1 0.00000000
#>                      ucl
#> B: Placebo     0.6855603
#> C: Combination       Inf
#> 
#> 
#> $results$AGE
#> $results$AGE$main
#>     df p-value
#> AGE  1 0.75967
#> 
#> $results$AGE$predictor
#> $results$AGE$predictor$term_type
#> [1] "continuous"
#> 
#> $results$AGE$predictor$interact_with
#> [1] "BMRKR1"
#> 
#> 
#> $results$AGE$summary
#> $results$AGE$summary$summary_term
#>            coef         se   p-value df
#> AGE -0.01949014 0.06371161 0.7596717  1
#> 
#> $results$AGE$summary$summary_with_interaction
#> $results$AGE$summary$summary_with_interaction$`5`
#>          or       lcl      ucl
#> 1 0.9959055 0.9399544 1.055187
#> 
#> 
#> 
#> 
#> $results$BMRKR1
#> $results$BMRKR1$main
#>        df p-value
#> BMRKR1  1 0.72933
#> 
#> $results$BMRKR1$predictor
#> $results$BMRKR1$predictor$term_type
#> [1] "continuous"
#> 
#> $results$BMRKR1$predictor$interact_with
#> [1] "AGE"
#> 
#> 
#> $results$BMRKR1$summary
#> $results$BMRKR1$summary$summary_term
#>              coef        se   p-value df
#> BMRKR1 -0.1285854 0.3716118 0.7293271  1
#> 
#> $results$BMRKR1$summary$summary_with_interaction
#> $results$BMRKR1$summary$summary_with_interaction$`34`
#>          or      lcl      ucl
#> 1 0.9763326 0.857576 1.111535
#> 
#> 
#> 
#> 
#> $results$`AGE:BMRKR1`
#> $results$`AGE:BMRKR1`$main
#>            df p-value
#> AGE:BMRKR1  1 0.76919
#> 
#> $results$`AGE:BMRKR1`$predictor
#> $results$`AGE:BMRKR1`$predictor$term_type
#> [1] "interaction"
#> 
#> 
#> $results$`AGE:BMRKR1`$summary
#>                     coef         se   p-value df
#> AGE * BMRKR1 0.003077455 0.01048775 0.7691909  1
#> 
#> 
#>