OR: summary

Summary function for odds ratio estimation.

s_odds_ratio(rsp, col_by, conf_level = 0.95, strat = NULL)

Arguments

rsp
(logical vector)
TRUE if subject is a responder FALSE otherwise

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

conf_level
confidence level of the interval.

strat
Optional factor with one level per stratum and same length as rsp. If `NULL`, then the unstratified analysis is computed.

Value

A data.frame with columns:

level

The level of the `arm` variable which is compared with the reference level.

odds_ratio

The odds ratio estimate.

ci_lower

The lower bound of the confidence interval.

ci_upper

The upper bound of the confidence interval.

The attribute `conf_level` saves the used confidence level.

Details

This function uses either logistic regression for unstratified analyses, or conditional logistic regression for stratified analyses. The Wald confidence interval with the specified confidence level is calculated. Note that for stratified analyses, there is currently no implementation for conditional likelihood confidence intervals, therefore we don't offer the option of a likelihood confidence interval. Note that when `rsp` contains only responders or non-responders, then the result values will be `NA`, because no odds ratio estimation is possible.

Examples

set.seed(1)
rsp <- sample(c(TRUE, FALSE), 100, TRUE)
trt <- sample(c("Placebo", "Treatment", "Combination"), 100, TRUE)
trt <- factor(trt, levels = c("Placebo", "Treatment", "Combination"))
strata_data <- data.frame(
  "f1" = sample(c("a", "b"), 100, TRUE),
  "f2" = sample(c("x", "y", "z"), 100, TRUE),
  stringsAsFactors = TRUE
)

s_odds_ratio(rsp, trt)

#>         level odds_ratio  ci_lower ci_upper
#> 1   Treatment  2.6714286 0.9545455 7.476365
#> 2 Combination  0.6577381 0.2474225 1.748505

s_odds_ratio(rsp, trt, strat = interaction(strata_data))

#>         level odds_ratio  ci_lower ci_upper
#> 1   Treatment  2.7113676 0.9626810 7.636501
#> 2 Combination  0.6516824 0.2440296 1.740322