Summary of difference between two proportions

This function is a wrapper to calculate the difference between two proportions or weighed difference if stratification variables are used. In addition provides various options for types of confidence intervals.

s_proportion_diff(
  x,
  grp,
  strat = NULL,
  conf_level = 0.95,
  diff_ci_method = c("wald", "waldcc", "cmh", "anderson-hauck", "ha", "newcombe")
)

Arguments

x
logical vector. TRUE represents a successful outcome.

grp
a factor vector with 2 levels. Must be the same length as x.

strat
factor with one level per stratum and same length as x. Only required for stratified analysis when method is "cmh".

conf_level
confidence level of the interval.

diff_ci_method
one of ("wald", "waldcc", "cmh", "anderson-hauck" or "ha", "newcombe")
Specifies the method used to construct the confidence interval for the difference in proportion of outcomes.

Value

Named list with analysis summary.

diff

Difference in proportion of successful outcomes between groups.

diff_ci

Confidence interval for diff.

label_ci

Label for confidence interval in diff_ci.

Details

The following options are supported for difference of proportions confidence intervals:

Option wald calls prop.test with correct=FALSE

Option waldcc includes continuity correction for Wald interval and calls prop.test with correct=TRUE

Option cmh derives a CI for CMH-weighted difference of proportions and calls s_adj_proportion_diff.

Options anderson-hauck or ha derives the Anderson-Hauck confidence interval.

Option newcombe derives the Newcombe confidence interval, which is based on the Wilson score confidence interval for a single binomial proportion.

See also

Examples

set.seed(1)
rsp <- sample(c(TRUE, FALSE), 100, TRUE)
trt <- sample(c("Placebo", "Treatment"), 100, TRUE)
trt <- factor(trt, levels = c("Placebo", "Treatment"))
strata_data <- data.frame(
  "f1" = sample(c("a", "b"), 100, TRUE),
  "f2" = sample(c("x", "y", "z"), 100, TRUE),
  stringsAsFactors = TRUE)

s_proportion_diff(rsp, trt, diff_ci_method = "wald")

#> $diff
#> [1] 0.03894018
#> 
#> $diff_ci
#> [1] -0.1572385  0.2351189
#> 
#> $label_ci
#> [1] "95% CI for difference (Wald, without correction)"
#> 

s_proportion_diff(rsp, trt, diff_ci_method = "waldcc")

#> $diff
#> [1] 0.03894018
#> 
#> $diff_ci
#> [1] -0.1773108  0.2551911
#> 
#> $label_ci
#> [1] "95% CI for difference (Wald, with correction)"
#> 

s_proportion_diff(rsp, trt, strat = interaction(strata_data), diff_ci_method = "cmh")

#> $diff
#> [1] 0.04213259
#> 
#> $diff_ci
#> [1] -0.1462339  0.2304991
#> 
#> $label_ci
#> [1] "95% CI for adjusted difference (CMH, without correction)"
#> 

# case without responders
s_proportion_diff(rep(FALSE, length(trt)), trt, diff_ci_method = "wald")

#> $diff
#> [1] 0
#> 
#> $diff_ci
#> [1] NA NA
#> 
#> $label_ci
#> [1] NA
#>