Allows to access nodes recursively, each element in index can either be a number or the name of the node

Ensure that children are not named identically. If cannot be avoided, access by index

# S4 method for node
[[(x, i, j, ..., return_content = TRUE)

Arguments

x
node object

i
index (possibly missing)

j
index (possibly missing)

...
deeper level indexes

return_content
whether to return content of node instead of node

Details

Can use summary to get index to get to certain element

Examples

n2 <- node(name = "D", content = c(1:3), children = list(
  node(name = "A", content = array(c(1:6), dim = c(2,3)), children = list()),
  node(name = "B", content = array(c(1:6), dim = c(2,3)), children = list()),
  node(name = "C", content = array(c(1:6), dim = c(2,3)), children = list())
))
summary(n2)

#>              index            content             dim            hide label          format  
#> ---------------------------------------------------------------------------------------------
#> D                             integer                              FALSE                     
#>   A            1               matrix            2 x 3             FALSE                     
#>   B            2               matrix            2 x 3             FALSE                     
#>   C            3               matrix            2 x 3             FALSE                     

summary(n2[[return_content = FALSE]])

#>              index            content             dim            hide label          format  
#> ---------------------------------------------------------------------------------------------
#> D                             integer                              FALSE                     
#>   A            1               matrix            2 x 3             FALSE                     
#>   B            2               matrix            2 x 3             FALSE                     
#>   C            3               matrix            2 x 3             FALSE                     

n2[[]]

#> [1] 1 2 3

n2[[return_content = TRUE]]

#> [1] 1 2 3

n2[[1]]

#>      [,1] [,2] [,3]
#> [1,]    1    3    5
#> [2,]    2    4    6

n2[["A"]]

#>      [,1] [,2] [,3]
#> [1,]    1    3    5
#> [2,]    2    4    6

summary(n2[["A", return_content = FALSE]])

#>            index            content             dim            hide label          format  
#> -------------------------------------------------------------------------------------------
#> A                            matrix            2 x 3             FALSE                     

n3 <- node(name = "D", content = c(1:12), children = list(
node(name = "A", content = c(1:6), children = list(
  node(name = "A1", content = c(1:3), children = list()),
  node(name = "A2", content = c(4:6), children = list())
)),
node(name = "B", content = c(7:12), children = list(
  node(name = "B1", content = c(7:9), children = list()),
  node(name = "B2", content = c(10:12), children = list())
))
))
cat(displayable(n3))

#> D:
#>   1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12
#>   A:
#>     1, 2, 3, 4, 5, 6
#>     A1:
#>       1, 2, 3
#>     A2:
#>       4, 5, 6
#>   B:
#>     7, 8, 9, 10, 11, 12
#>     B1:
#>       7, 8, 9
#>     B2:
#>       10, 11, 12

cat(displayable(n3[["A"]]))

#> 1, 2, 3, 4, 5, 6

cat(displayable(n3[["A", return_content = FALSE]]))

#> A:
#>   1, 2, 3, 4, 5, 6
#>   A1:
#>     1, 2, 3
#>   A2:
#>     4, 5, 6

cat(displayable(n3[["A", "A1"]]))

#> 1, 2, 3

cat(displayable(n3[["A", "A1", return_content = FALSE]]))

#> A1:
#>   1, 2, 3