Multi-Variable Logistic Regression Table

Layout creating function which summarizes a logistic variable regression for binary outcome with categorical/continuous covariates in model statement. For each covariate category (if categorical) or specified values (if continuous), present degrees of freedom, regression parameter estimate and standard error (SE) relative to reference group or category. Report odds ratios for each covariate category or specified values and corresponding Wald confidence intervals as default but allow user to specify other confidence levels. Report p-value for Wald chi-square test of the null hypothesis that covariate has no effect on response in model containing all specified covariates. Allow option to include one two-way interaction and present similar output for each interaction degree of freedom. Note: For the formula, the variable names need to be standard dataframe column name without special characters.

Usage

summarize_logistic(lyt, conf_level, drop_and_remove_str = "")

Arguments

lyt

(layout)
input layout where analyses will be added to.

conf_level

(proportion)
confidence level of the interval.

drop_and_remove_str

string to be dropped and removed

Examples

library(scda)
library(dplyr)
library(broom)

adrs <- synthetic_cdisc_data("latest")$adrs
adrs_f <- adrs %>%
  filter(PARAMCD == "BESRSPI") %>%
  filter(RACE %in% c("ASIAN", "WHITE", "BLACK OR AFRICAN AMERICAN")) %>%
  mutate(
    Response = case_when(AVALC %in% c("PR", "CR") ~ 1, TRUE ~ 0),
    RACE = factor(RACE),
    SEX = factor(SEX)
  )
formatters::var_labels(adrs_f) <- c(formatters::var_labels(adrs), Response = "Response")
mod1 <- fit_logistic(
  data = adrs_f,
  variables = list(
    response = "Response",
    arm = "ARMCD",
    covariates = c("AGE", "RACE")
  )
)
mod2 <- fit_logistic(
  data = adrs_f,
  variables = list(
    response = "Response",
    arm = "ARMCD",
    covariates = c("AGE", "RACE"),
    interaction = "AGE"
  )
)

df <- tidy(mod1, conf_level = 0.99)
df2 <- tidy(mod2, conf_level = 0.99)

# flagging empty strings with "_"
df <- df_explicit_na(df, na_level = "_")
df2 <- df_explicit_na(df2, na_level = "_")

result1 <- basic_table() %>%
  summarize_logistic(
    conf_level = 0.95,
    drop_and_remove_str = "_"
  ) %>%
  build_table(df = df)
result1
#>                                       Degrees of Freedom   Parameter Estimate   Standard Error   Odds Ratio    Wald 95% CI    p-value
#> —————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> Planned Arm Code                              2                                                                               0.0368 
#>   Reference ARM A, n = 126                                                                                                           
#>   ARM B, n = 121                              1                  -2.155             1.084           0.12      (<0.01, 1.89)   0.0468 
#>   ARM C, n = 126                              1                  -0.127             1.425           0.88      (0.02, 34.55)   0.9287 
#> Age                                                                                                                                  
#>   Age                                         1                  0.067              0.053           1.07      (0.93, 1.23)    0.2068 
#> Race                                          2                                                                               0.6369 
#>   Reference ASIAN, n = 208                                                                                                           
#>   BLACK OR AFRICAN AMERICAN, n = 91           1                  1.027              1.098           2.79      (0.17, 47.24)   0.3498 
#>   WHITE, n = 74                               1                  0.294              0.850           1.34      (0.15, 11.99)   0.7294 

result2 <- basic_table() %>%
  summarize_logistic(
    conf_level = 0.95,
    drop_and_remove_str = "_"
  ) %>%
  build_table(df = df2)
result2
#>                                         Degrees of Freedom   Parameter Estimate   Standard Error   Odds Ratio     Wald 95% CI      p-value
#> ——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> Race                                            2                                                                                  0.6027 
#>   Reference ASIAN, n = 208                                                                                                                
#>   BLACK OR AFRICAN AMERICAN, n = 91             1                  1.079              1.136           2.94       (0.16, 54.88)     0.3423 
#>   WHITE, n = 74                                 1                  0.455              0.900           1.58       (0.16, 16.01)     0.6127 
#> Planned Arm Code                                2                                                                                  0.2890 
#>   Reference ARM A, n = 126                                                                                                                
#>   ARM B, n = 121                                1                  20.235             14.787                                       0.1712 
#>     Age                                                                                                                                   
#>       34                                                                                              0.00      (<0.01, >999.99)          
#>   ARM C, n = 126                                1                  14.713             16.087                                       0.3604 
#>     Age                                                                                                                                   
#>       34                                                                                              0.00      (<0.01, >999.99)          
#> Age                                                                                                                                       
#>   Age                                           1                  0.896              0.629                                        0.1546 
#>     Planned Arm Code                                                                                                                      
#>       ARM A                                                                                           2.45       (0.48, 12.39)            
#>       ARM B                                                                                           1.03        (0.90, 1.18)            
#>       ARM C                                                                                           1.32        (0.72, 2.43)            
#> Interaction of Planned Arm Code * Age           2                                                                                  0.2306 
#>   Reference ARM A, n = 126                                                                                                                
#>   ARM B, n = 121                                1                  -0.868             0.630                                        0.1680 
#>   ARM C, n = 126                                1                  -0.618             0.671                                        0.3577