Summary numeric variables in columns

These functions can be used to produce summary tables for PK datasets.

summary_numeric_in_cols(x, labelstr = "", custom_label = NULL, ...)

summarize_vars_numeric_in_cols(
  lyt,
  var,
  ...,
  .stats = c("n", "mean", "sd", "se", "cv", "geom_cv"),
  .labels = c(n = "n", mean = "Mean", sd = "SD", se = "SE", cv = "CV (%)", geom_cv =
    "CV % Geometric Mean"),
  .indent_mods = NULL,
  col_split = TRUE
)

Arguments

x

(numeric)
vector of numbers we want to analyze.

labelstr

(character)
label of the level of the parent split currently being summarized (must be present as second argument in Content Row Functions).

custom_label

(string or NULL)
if provided and labelstr is empty then this will be used as the row label.

...

additional arguments for the lower level functions.

lyt

(layout)
input layout where analyses will be added to.

var

(string)
single variable name for the primary analysis variable.

.stats

(character)
statistics to select for the table.

.labels

(named character)
labels for the statistics (without indent).

.indent_mods

(named integer)
indent modifiers for the labels.

col_split

(flag)
whether the columns should be split.

Value

A named list of all statistics returned by s_summary.numeric(). See s_summary.numeric() to be aware of all available statistics.

Functions

summary_numeric_in_cols: a wrapper of s_summary.numeric() function that produces a named list of statistics to include as columns.

summarize_vars_numeric_in_cols: Layout creating function which can be used for creating summary tables in columns, primarily used for PK data sets.

See also

Examples

library(scda)
library(dplyr)
ADPC <- scda::synthetic_cdisc_data("latest")$adpc
summary_numeric_in_cols(ADPC$AGE, custom_label = "stats")
#> $n
#>    n 
#> 3582 
#> attr(,"label")
#> [1] "stats"
#> 
#> $mean
#>     mean 
#> 34.87186 
#> attr(,"label")
#> [1] "stats"
#> 
#> $sd
#>       sd 
#> 7.364947 
#> attr(,"label")
#> [1] "stats"
#> 
#> $se
#>        se 
#> 0.1230571 
#> attr(,"label")
#> [1] "stats"
#> 
#> $mean_sd
#>      mean        sd 
#> 34.871859  7.364947 
#> attr(,"label")
#> [1] "stats"
#> 
#> $mean_ci
#> mean_ci_lwr mean_ci_upr 
#>    34.63059    35.11313 
#> attr(,"label")
#> [1] "stats"
#> 
#> $mean_sei
#> mean_sei_lwr mean_sei_upr 
#>     34.74880     34.99492 
#> attr(,"label")
#> [1] "stats"
#> 
#> $mean_sdi
#> mean_sdi_lwr mean_sdi_upr 
#>     27.50691     42.23681 
#> attr(,"label")
#> [1] "stats"
#> 
#> $median
#> median 
#>     34 
#> attr(,"label")
#> [1] "stats"
#> 
#> $mad
#> mad 
#>   0 
#> attr(,"label")
#> [1] "stats"
#> 
#> $median_ci
#> median_ci_lwr median_ci_upr 
#>            34            35 
#> attr(,"conf_level")
#> [1] 0.9532377
#> attr(,"label")
#> [1] "stats"
#> 
#> $quantiles
#> quantile_0.25 quantile_0.75 
#>            30            39 
#> attr(,"label")
#> [1] "stats"
#> 
#> $iqr
#> iqr 
#>   9 
#> attr(,"label")
#> [1] "stats"
#> 
#> $range
#> min max 
#>  20  69 
#> attr(,"label")
#> [1] "stats"
#> 
#> $min
#> min 
#>  20 
#> attr(,"label")
#> [1] "stats"
#> 
#> $max
#> max 
#>  69 
#> attr(,"label")
#> [1] "stats"
#> 
#> $cv
#>       cv 
#> 21.12003 
#> attr(,"label")
#> [1] "stats"
#> 
#> $geom_mean
#> geom_mean 
#>  34.13964 
#> attr(,"label")
#> [1] "stats"
#> 
#> $geom_cv
#>  geom_cv 
#> 20.70285 
#> attr(,"label")
#> [1] "stats"
#> 

lyt <- basic_table() %>%
  split_rows_by(var = "ARM", label_pos = "topleft") %>%
  split_rows_by(var = "SEX", label_pos = "topleft") %>%
  summarize_vars_numeric_in_cols(var = "AGE", col_split = TRUE)
result <- build_table(lyt = lyt, df = ADPC)
result
#> ARM                                                                    
#>   SEX             n     Mean   SD    SE    CV (%)   CV % Geometric Mean
#> ———————————————————————————————————————————————————————————————————————
#> A: Drug X                                                              
#>   F              711    32.8   6.1   0.2    18.5           18.7        
#>   M              495    35.2   6.9   0.3    19.6           19.7        
#> B: Placebo                                                             
#>   F               0      NA    NA    NA      NA             NA         
#>   M               0      NA    NA    NA      NA             NA         
#> C: Combination                                                         
#>   F              1260   35.2   7.3   0.2    20.8           20.4        
#>   M              1116   35.7   8.1   0.2    22.7            22         

# By selecting just some statistics and ad-hoc labels
lyt <- basic_table() %>%
  split_rows_by(var = "ARM", label_pos = "topleft") %>%
  split_rows_by(var = "SEX", label_pos = "topleft") %>%
  summarize_vars_numeric_in_cols(
    var = "AGE",
    .stats = c("n", "cv", "geom_mean", "mean_ci"),
    .labels = c(n = "myN", cv = "myCV", geom_mean = "myGeomMean", mean_ci = "Mean (95%CI)"),
    col_split = TRUE
  )
result <- build_table(lyt = lyt, df = ADPC)
result
#> ARM                                                       
#>   SEX            myN    myCV   myGeomMean    Mean (95%CI) 
#> ——————————————————————————————————————————————————————————
#> A: Drug X                                                 
#>   F              711    18.5      32.2      (32.31, 33.21)
#>   M              495    19.6      34.6      (34.61, 35.83)
#> B: Placebo                                                
#>   F               0      NA       NaN          (NA, NA)   
#>   M               0      NA       NaN          (NA, NA)   
#> C: Combination                                            
#>   F              1260   20.8      34.5       (34.8, 35.6) 
#>   M              1116   22.7      34.8      (35.22, 36.17)

lyt <- basic_table() %>%
  summarize_vars_numeric_in_cols(
    var = "AGE",
    col_split = TRUE,
    custom_label = "some custom label"
  )
result <- build_table(lyt, df = ADPC)
result
#>                      n     Mean   SD    SE    CV (%)   CV % Geometric Mean
#> ——————————————————————————————————————————————————————————————————————————
#> some custom label   3582   34.9   7.4   0.1    21.1           20.7        

# PKPT03
ADPC <- mutate(ADPC, PARAMUNIT = paste0(.data$PARAM, "(", .data$AVALU, ")"))
lyt <- basic_table() %>%
  split_rows_by(var = "PARAMUNIT", split_label = "PK Parameter", label_pos = "topleft") %>%
  summarize_vars_numeric_in_cols(var = "AVAL", col_split = TRUE)
result <- build_table(lyt, df = ADPC)
result
#> PK Parameter            n     Mean    SD    SE    CV (%)   CV % Geometric Mean
#> ——————————————————————————————————————————————————————————————————————————————
#> Plasma Drug X(ug/mL)   2394    8     6.6    0.1    82.7            NA         
#> Plasma Drug Y(ug/mL)   1188    16    13.3   0.4    83.1            NA         

# PKCT01FDS
ADPC <- mutate(ADPC, as.factor(.data$NRELTM1))
lyt <- basic_table() %>%
  split_rows_by(var = "VISIT", split_label = "Visit", label_pos = "topleft") %>%
  split_rows_by(
    var = "NRELTM1",
    split_label = "Norminal time from first dose",
    label_pos = "topleft"
   ) %>%
  summarize_vars_numeric_in_cols(var = "AVAL", col_split = TRUE)
result <- build_table(lyt, df = ADPC)
#> Split var [NRELTM1] was not character or factor. Converting to factor
result
#> Visit                                                                                  
#>   Norminal time from first dose    n    Mean   SD    SE    CV (%)   CV % Geometric Mean
#> ———————————————————————————————————————————————————————————————————————————————————————
#> Day 1                                                                                  
#>   0                               398    0      0     0     NaN             NA         
#>   0.5                             398   17.1   6.5   0.3     38            36.4        
#>   1                               398   21.8   8.2   0.4    37.4           35.7        
#>   1.5                             398   20.9   7.8   0.4    37.1           35.3        
#>   2                               398   17.9   6.7   0.3    37.2           35.4        
#>   3                               398   11.1   4.3   0.2    38.6           36.9        
#>   4                               398   6.2    2.6   0.1    41.3           40.2        
#>   8                               398   0.4    0.3    0     62.6           66.8        
#>   12                              398    0      0     0     94.1           114.2