Summary of PK datasets

These functions can be used to produce summary tables for PK datasets.

s_summary_pk(x, labelstr = "", custom_label = NULL, ...)

summarize_pk_in_cols(
  lyt,
  var,
  ...,
  .stats = c("n", "mean", "sd", "cv", "geom_mean", "geom_cv", "median", "min", "max"),
  .labels = c(n = "n", mean = "Mean", sd = "SD", cv = "CV % Mean", geom_mean =
    "Geometric Mean", geom_cv = "CV % Geometric Mean", median = "Median", min =
    "Minimum", max = "Maximum"),
  .indent_mods = NULL,
  col_split = TRUE
)

Arguments

x
(numeric)
vector of numbers we want to analyze.

labelstr
(character)
label of the level of the parent split currently being summarized (must be present as second argument in Content Row Functions).

custom_label
(string or NULL)
if provided and labelstr is empty then this will be used as the row label.

...
additional arguments for the lower level functions.

lyt
(layout)
input layout where analyses will be added to.

var
(string)
single variable name for the primary analysis variable.

.stats
(character)
statistics to select for the table.

.labels
(named character)
labels for the statistics (without indent).

.indent_mods
(named integer)
indent modifiers for the labels.

col_split
(flag)
whether the columns should be split.

Value

A named list of 9 statistics:

n: count of complete sample size for the group.

mean: group data mean.

sd: standard deviation of the group data.

cv: coefficient of variation of the group.

min: minimum value among the group.

max: maximum value among the group.

median: median of the group data.

geom_mean: geometric mean of the group.

geom_cv: geometric coefficient of variation of the group.

Functions

s_summary_pk: Statistics function that produces a named list of statistics to include as columns in the PK table.

summarize_pk_in_cols: Layout creating function which can be used for creating summary tables for PK data sets.

Examples

library(scda)
ADSL <- scda::synthetic_cdisc_data("latest")$adsl
s_summary_pk(ADSL$AGE, custom_label = "stats")
#> $n
#>   n 
#> 400 
#> attr(,"label")
#> [1] "stats"
#> 
#> $mean
#>   mean 
#> 34.875 
#> attr(,"label")
#> [1] "stats"
#> 
#> $sd
#>     sd 
#> 7.4359 
#> attr(,"label")
#> [1] "stats"
#> 
#> $cv
#>    cv.sd 
#> 21.32158 
#> attr(,"label")
#> [1] "stats"
#> 
#> $min
#> min 
#>  20 
#> attr(,"label")
#> [1] "stats"
#> 
#> $max
#> max 
#>  69 
#> attr(,"label")
#> [1] "stats"
#> 
#> $median
#> median 
#>     34 
#> attr(,"label")
#> [1] "stats"
#> 
#> $geom_mean
#> geom_mean 
#>  34.13023 
#> attr(,"label")
#> [1] "stats"
#> 
#> $geom_cv
#> geom_cv 
#> 20.8869 
#> attr(,"label")
#> [1] "stats"
#> 

lyt <- basic_table() %>%
  split_rows_by(var = "ARM") %>%
  split_rows_by(var = "SEX") %>%
  summarize_pk_in_cols(var = "AGE", col_split = TRUE)
  result <- build_table(lyt, df = ADSL)
  result
#>                  n    Mean     SD    CV % Mean   Geometric Mean   CV % Geometric Mean   Median   Minimum   Maximum
#> ——————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> A: Drug X                                                                                                         
#>   F              79   32.76   6.09     18.6           32.2               18.8             32       21        47   
#>   M              55   35.22   6.97     19.8          34.55               19.9             34       23        50   
#> B: Placebo                                                                                                        
#>   F              82   34.24   6.99     20.4           33.6               19.6             32       23        58   
#>   M              52   37.31   8.9      23.8          36.29               24.1             36       21        62   
#> C: Combination                                                                                                    
#>   F              70   35.2    7.37     20.9          34.48               20.5             35       21        64   
#>   M              62   35.69   8.15     22.8          34.85               22.1             35       20        69   

lyt <- basic_table() %>%
  summarize_pk_in_cols(var = "AGE", col_split = TRUE)
  result <- build_table(lyt, df = ADSL)
  result
#>               n    Mean     SD    CV % Mean   Geometric Mean   CV % Geometric Mean   Median   Minimum   Maximum
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> Statistics   400   34.88   7.44     21.3          34.13               20.9             34       20        69