Tabulate Survival Duration by Subgroup

Tabulate statistics such as median survival time and hazard ratio for population subgroups.

Usage

a_survival_subgroups(
  .formats = list(n = "xx", n_events = "xx", n_tot_events = "xx", median = "xx.x", n_tot
    = "xx", hr = list(format_extreme_values(2L)), ci =
    list(format_extreme_values_ci(2L)), pval = "x.xxxx | (<0.0001)")
)

tabulate_survival_subgroups(
  lyt,
  df,
  vars = c("n_tot_events", "n_events", "median", "hr", "ci"),
  time_unit = NULL
)

Arguments

.formats

(named character or list)
formats for the statistics.

lyt

(layout)
input layout where analyses will be added to.

df

(list)
of data frames containing all analysis variables. List should be created using extract_survival_subgroups().

vars

(character)
the name of statistics to be reported among n_tot_events (total number of events per group), n_events (number of events per group), n_tot (total number of observations per group), n (number of observations per group), median (median survival time), hr (hazard ratio), ci (confidence interval of hazard ratio) and pval (p value of the effect). Note, one of the statistics n_tot and n_tot_events, as well as both hr and ci are required.

time_unit

(string)
label with unit of median survival time. Default NULL skips displaying unit.

data

(data frame)
the dataset containing the variables to summarize.

groups_lists

(named list of list)
optionally contains for each subgroups variable a list, which specifies the new group levels via the names and the levels that belong to it in the character vectors that are elements of the list.

label_all

(string)
label for the total population analysis.

Details

These functions create a layout starting from a data frame which contains the required statistics. Tables typically used as part of forest plot.

Functions

a_survival_subgroups(): Formatted Analysis function used to format the results of extract_survival_subgroups(). Returns is a list of Formatted Analysis functions with one element per statistic.

tabulate_survival_subgroups(): table creating function.

See also

Examples

library(dplyr)
library(forcats)

adtte <- tern_ex_adtte

# Save variable labels before data processing steps.
adtte_labels <- formatters::var_labels(adtte)

adtte_f <- adtte %>%
  filter(
    PARAMCD == "OS",
    ARM %in% c("B: Placebo", "A: Drug X"),
    SEX %in% c("M", "F")
  ) %>%
  mutate(
    # Reorder levels of ARM to display reference arm before treatment arm.
    ARM = droplevels(fct_relevel(ARM, "B: Placebo")),
    SEX = droplevels(SEX),
    AVALU = as.character(AVALU),
    is_event = CNSR == 0
  )
labels <- c(
  "ARM" = adtte_labels[["ARM"]],
  "SEX" = adtte_labels[["SEX"]],
  "AVALU" = adtte_labels[["AVALU"]],
  "is_event" = "Event Flag"
)
formatters::var_labels(adtte_f)[names(labels)] <- labels

df <- extract_survival_subgroups(
  variables = list(
    tte = "AVAL",
    is_event = "is_event",
    arm = "ARM", subgroups = c("SEX", "BMRKR2")
  ),
  data = adtte_f
)
df
#> $survtime
#>           arm  n n_events    median     subgroup    var    var_label row_type
#> 1  B: Placebo 73       57  571.1603 All Patients    ALL All Patients  content
#> 2   A: Drug X 69       41 1458.7844 All Patients    ALL All Patients  content
#> 3  B: Placebo 40       31  722.4495            F    SEX          SEX analysis
#> 4   A: Drug X 38       19 1617.6840            F    SEX          SEX analysis
#> 5  B: Placebo 33       26  551.8749            M    SEX          SEX analysis
#> 6   A: Drug X 31       22  763.4915            M    SEX          SEX analysis
#> 7  B: Placebo 24       22  362.4989          LOW BMRKR2       BMRKR2 analysis
#> 8   A: Drug X 26       15 1218.1336          LOW BMRKR2       BMRKR2 analysis
#> 9  B: Placebo 23       17  506.9159       MEDIUM BMRKR2       BMRKR2 analysis
#> 10  A: Drug X 26       14 1549.4890       MEDIUM BMRKR2       BMRKR2 analysis
#> 11 B: Placebo 26       18  835.0535         HIGH BMRKR2       BMRKR2 analysis
#> 12  A: Drug X 17       12 1438.7642         HIGH BMRKR2       BMRKR2 analysis
#> 
#> $hr
#>   arm n_tot n_tot_events        hr       lcl       ucl conf_level        pval
#> 1       142           98 0.5125892 0.3388862 0.7753272       0.95 0.001293042
#> 2        78           50 0.4776406 0.2636437 0.8653365       0.95 0.012997814
#> 3        64           48 0.5542360 0.3055256 1.0054069       0.95 0.049074335
#> 4        50           37 0.3747258 0.1889873 0.7430097       0.95 0.003579595
#> 5        49           31 0.4131657 0.1981187 0.8616346       0.95 0.015407117
#> 6        43           30 0.7429373 0.3407495 1.6198288       0.95 0.453625856
#>           pval_label     subgroup    var    var_label row_type
#> 1 p-value (log-rank) All Patients    ALL All Patients  content
#> 2 p-value (log-rank)            F    SEX          SEX analysis
#> 3 p-value (log-rank)            M    SEX          SEX analysis
#> 4 p-value (log-rank)          LOW BMRKR2       BMRKR2 analysis
#> 5 p-value (log-rank)       MEDIUM BMRKR2       BMRKR2 analysis
#> 6 p-value (log-rank)         HIGH BMRKR2       BMRKR2 analysis
#> 
# Internal function - a_survival_subgroups
if (FALSE) {
a_survival_subgroups(.formats = list("n" = "xx", "median" = "xx.xx"))
}

## Table with default columns.
basic_table() %>%
  tabulate_survival_subgroups(df, time_unit = adtte_f$AVALU[1])
#> Baseline Risk Factors                        B: Placebo               A: Drug X                                     
#>                         Total Events   Events   Median (DAYS)   Events   Median (DAYS)   Hazard Ratio   95% Wald CI 
#> ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> All Patients                 98          57         571.2         41        1458.8           0.51       (0.34, 0.78)
#> SEX                                                                                                                 
#>   F                          50          31         722.4         19        1617.7           0.48       (0.26, 0.87)
#>   M                          48          26         551.9         22         763.5           0.55       (0.31, 1.01)
#> BMRKR2                                                                                                              
#>   LOW                        37          22         362.5         15        1218.1           0.37       (0.19, 0.74)
#>   MEDIUM                     31          17         506.9         14        1549.5           0.41       (0.20, 0.86)
#>   HIGH                       30          18         835.1         12        1438.8           0.74       (0.34, 1.62)

## Table with a manually chosen set of columns: adding "pval".
basic_table() %>%
  tabulate_survival_subgroups(
    df = df,
    vars = c("n_tot_events", "n_events", "median", "hr", "ci", "pval"),
    time_unit = adtte_f$AVALU[1]
  )
#> Baseline Risk Factors                        B: Placebo               A: Drug X                                                          
#>                         Total Events   Events   Median (DAYS)   Events   Median (DAYS)   Hazard Ratio   95% Wald CI    p-value (log-rank)
#> —————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> All Patients                 98          57         571.2         41        1458.8           0.51       (0.34, 0.78)         0.0013      
#> SEX                                                                                                                                      
#>   F                          50          31         722.4         19        1617.7           0.48       (0.26, 0.87)         0.0130      
#>   M                          48          26         551.9         22         763.5           0.55       (0.31, 1.01)         0.0491      
#> BMRKR2                                                                                                                                   
#>   LOW                        37          22         362.5         15        1218.1           0.37       (0.19, 0.74)         0.0036      
#>   MEDIUM                     31          17         506.9         14        1549.5           0.41       (0.20, 0.86)         0.0154      
#>   HIGH                       30          18         835.1         12        1438.8           0.74       (0.34, 1.62)         0.4536