Abnormality Shift from Baseline Table

Summarize the number of patients at the specified visit with the abnormality (e.g. High, Low), among patients with non-missing value at that visit and a baseline value.

t_abn_shift(
  rri_ana,
  rri_base,
  value_ana,
  value_base,
  abnormal = c("LOW", "HIGH"),
  id,
  col_by,
  row_by,
  col_N = NULL,
  total = NULL,
  table_tree = FALSE
)

Arguments

rri_ana
a vector specifying the Reference Range Indicator (e.g. High/Low/Normal).
Fill blank with NA using sas_na().

rri_base
a vector specifying the Reference Range Indicator (e.g. High/Low/Normal) at baseline
for the corresponding records in rri_ana. Fill blank with NA using sas_na().

value_ana
a vector of numeric measurement.

value_base
a vector of numeric measurement at baseline
for the corresponding records in value_ana.

abnormal
a vector specifying the unique values that represent abnormality in the order of display.

id
a vector of subject identifier with length of rri_ana.

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

row_by
(factor or data.frame)
Defines how data from x is split into sub-tables. Dimensions must match dimensions of x and no missing values are allowed. Multi-level nesting is possible when row_by is a data.frame. Columns should be ordered with the first column specifying the first variable to split by and the last column the specifying the last variable to split by.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

table_tree
(logical value)
If FALSE then a single rtable gets returned, if TRUE a node object gets returned with the elementary rtable objects.

Details

Input data should include all post-baseline records (repeated and unscheduled). Baseline and screening records should be excluded.

Examples

library(dplyr)
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADLB <- radlb(cached = TRUE) %>%
  dplyr::filter(ABLFL != "Y") %>%
  dplyr::filter(AVISIT != "SCREENING") %>%
  dplyr::mutate(AVISIT = droplevels(AVISIT))

ADLB_f <- ADLB %>% dplyr::filter(PARAMCD == "CRP")

tbl1 <- t_abn_shift(
  rri_ana = ADLB_f$ANRIND,
  rri_base = ADLB_f$BNRIND,
  value_ana = ADLB_f$AVAL,
  value_base = ADLB_f$BASE,
  abnormal = c("LOW", "HIGH"),
  id = ADLB_f$USUBJID,
  row_by = ADLB_f[,  "AVISIT"],
  col_by = ADLB_f$ARM,
  col_N = table(ADSL$ARM),
  table_tree = FALSE
)
if (FALSE) {
Viewer(tbl1)
}

tbl2 <- t_abn_shift(
  rri_ana = ADLB$ANRIND,
  rri_base = ADLB$BNRIND,
  value_ana = ADLB$AVAL,
  value_base = ADLB$BASE,
  abnormal = c("LOW", "HIGH"),
  id = ADLB$USUBJID,
  row_by = ADLB[, c("PARAM", "AVISIT")],
  col_by = ADLB$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients",
  table_tree = TRUE
)
if (FALSE) {
Viewer(to_rtable(tbl2))
}