Adverse Events by System Organ Class and Preferred Term

t_ae returns adverse events summary table that corresponds to STREAM template AET02

t_ae(class, term, id, col_by, col_N, total = "All Patients")

Arguments

class
class information as character or factor vector

term
term information as character or factor vector, however factor levels are not repeated by class, only terms with count > 1 are listed per class

id
unique subject identifier. If a particular subject has no adverse event then that information needs to be added to the col_N argument.

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing. See 'Examples'.

col_N
numeric vector with information of the number of patients in the levels of col_by. This is useful if there are patients that have no adverse events can be accounted for with this argument.

total
character string that will be used as a label for a column with pooled total population, default is "All Patients". If the levels of col_by are the only columns of interest then total should be NULL

Examples

# Simple example
library(tibble)
library(dplyr)

ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = factor(paste("ARM", LETTERS[rep(c(1,2), c(3,7))]))
)

ae_lookup <- tribble(
  ~CLASS,         ~TERM,   ~GRADE,
  "cl A",   "trm A_1/2",        1,
  "cl A",   "trm A_2/2",        2,
  "cl B",   "trm B_1/3",        2,
  "cl B",   "trm B_2/3",        3,
  "cl B",   "trm B_3/3",        1,
  "cl C",   "trm C_1/1",        1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2,2,2,3,3,4,4,4,4,5,6,6,7,7)]
  ),
  ae_lookup[c(1,1,2,6,4,2,2,3,4,2,1,5,4,6),]
)

ANL <- left_join(AAE, ASL, by = "USUBJID")

tbl <- t_ae(
  class = ANL$CLASS,
  term =  ANL$TERM,
  id = ANL$USUBJID,
  col_by = ANL$ARM,
  col_N = tapply(ASL$ARM, ASL$ARM, length),
  total = "All Patients"
)

tbl

#> MedDRA System Organ Class                                                    ARM A               ARM B            All Patients
#>   MedDRA Preferred Term                                                      (N=3)               (N=7)               (N=10)   
#> ------------------------------------------------------------------------------------------------------------------------------
#> - Any Adverse Event -
#>   Overall total number of events                                               5                   9                   14     
#>   Overall total number of patients with at least one adverse event         2 (66.67%)          4 (57.14%)           6 (60%)   
#> 
#> cl A
#>   Total number of events                                                       3                   4                   7      
#>   Total number of patients with at least one adverse event                 1 (33.33%)          3 (42.86%)           4 (40%)   
#>   trm A_2/2                                                                1 (33.33%)          2 (28.57%)           3 (30%)   
#>   trm A_1/2                                                                1 (33.33%)          1 (14.29%)           2 (20%)   
#> 
#> cl B
#>   Total number of events                                                       1                   4                   5      
#>   Total number of patients with at least one adverse event                 1 (33.33%)          3 (42.86%)           4 (40%)   
#>   trm B_2/3                                                                1 (33.33%)          2 (28.57%)           3 (30%)   
#>   trm B_1/3                                                                 NA (NA%)           1 (14.29%)           1 (10%)   
#>   trm B_3/3                                                                 NA (NA%)           1 (14.29%)           1 (10%)   
#> 
#> cl C
#>   Total number of events                                                       1                   1                   2      
#>   Total number of patients with at least one adverse event                 1 (33.33%)          1 (14.29%)           2 (20%)   
#>   trm C_1/1                                                                1 (33.33%)          1 (14.29%)           2 (20%)   

library(random.cdisc.data)
library(dplyr)

ASL <- radam("ASL", N = 10)
AAE <- radam("AAE", ADSL = ASL)

ANL <- left_join(AAE, ASL %>% select(USUBJID, STUDYID, ARM), by = c("STUDYID", "USUBJID"))

tbl <- with(ANL,
            t_ae(
              class = AEBODSYS,
              term =  AEDECOD,
              id = USUBJID,
              col_by = factor(ARM),
              total = "All Patients"
            )
)

tbl

#> MedDRA System Organ Class                                                     ARM A               ARM B            All Patients
#>   MedDRA Preferred Term                                                       (N=49)              (N=10)              (N=59)   
#> -------------------------------------------------------------------------------------------------------------------------------
#> - Any Adverse Event -
#>   Overall total number of events                                                49                  10                  59     
#>   Overall total number of patients with at least one adverse event          9 (18.37%)           1 (10%)           10 (16.95%) 
#> 
#> Investigations
#>   Total number of events                                                        12                  3                   15     
#>   Total number of patients with at least one adverse event                  5 (10.2%)            1 (10%)            6 (10.17%) 
#>   Carbon dioxide combining power decreased                                  1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Ehrlichia test                                                            1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Prekallikrein test                                                        1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Chest scan                                                                1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Biopsy heart normal                                                       1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Xanthochromia                                                             1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Bile culture                                                              1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Secretion gastric absent                                                  1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Blood cholinesterase                                                      1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Anti-glomerular basement membrane antibody                                1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Metamyelocyte count                                                       1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Physical thyroid examination                                               NA (NA%)            1 (10%)            1 (1.69%)  
#>   CSF specific gravity                                                       NA (NA%)            1 (10%)            1 (1.69%)  
#>   Fluoride increased                                                         NA (NA%)            1 (10%)            1 (1.69%)  
#>   Blood bromide                                                             1 (2.04%)            NA (NA%)           1 (1.69%)  
#> 
#> Injury, poisoning and procedural complications
#>   Total number of events                                                        3                   2                   5      
#>   Total number of patients with at least one adverse event                  3 (6.12%)            1 (10%)            4 (6.78%)  
#>   Graft ischaemia                                                           1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Cutaneous loxoscelism                                                     1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Skin graft detachment                                                     1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Puncture site pain                                                         NA (NA%)            1 (10%)            1 (1.69%)  
#>   Aqueous humour leakage                                                     NA (NA%)            1 (10%)            1 (1.69%)  
#> 
#> Respiratory, thoracic and mediastinal disorders
#>   Total number of events                                                        4                   1                   5      
#>   Total number of patients with at least one adverse event                  4 (8.16%)            1 (10%)            5 (8.47%)  
#>   Respiratory tract oedema                                                  1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Tracheal ulcer                                                            1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Reactive airways dysfunction syndrome                                     1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Portopulmonary hypertension                                                NA (NA%)            1 (10%)            1 (1.69%)  
#>   Young's syndrome                                                          1 (2.04%)            NA (NA%)           1 (1.69%)  
#> 
#> Nervous system disorders
#>   Total number of events                                                        2                   2                   4      
#>   Total number of patients with at least one adverse event                  2 (4.08%)            1 (10%)            3 (5.08%)  
#>   Intracranial aneurysm                                                     1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Spinal meningioma malignant                                               1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Cerebral toxoplasmosis                                                     NA (NA%)            1 (10%)            1 (1.69%)  
#>   Glycogen storage disease type III                                          NA (NA%)            1 (10%)            1 (1.69%)  
#> 
#> Skin and subcutaneous tissue disorders
#>   Total number of events                                                        4                   NA                  4      
#>   Total number of patients with at least one adverse event                  3 (6.12%)            NA (NA%)           3 (5.08%)  
#>   Administration site eczema                                                1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Erythema elevatum diutinum                                                1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Eyelid rash                                                               1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Eyelid bleeding                                                           1 (2.04%)            NA (NA%)           1 (1.69%)  
#> 
#> Surgical and medical procedures
#>   Total number of events                                                        3                   1                   4      
#>   Total number of patients with at least one adverse event                  2 (4.08%)            1 (10%)            3 (5.08%)  
#>   Fiducial marker placement                                                 1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Therapeutic nerve ablation                                                1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Muscle reattachment                                                       1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Myopia correction                                                          NA (NA%)            1 (10%)            1 (1.69%)  
#> 
#> Gastrointestinal disorders
#>   Total number of events                                                        3                   NA                  3      
#>   Total number of patients with at least one adverse event                  2 (4.08%)            NA (NA%)           2 (3.39%)  
#>   Abdominal wall anomaly                                                    1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Gastroenteritis paracolon bacillus                                        1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Eosinophilic oesophagitis                                                 1 (2.04%)            NA (NA%)           1 (1.69%)  
#> 
#> General disorders and administration site conditions
#>   Total number of events                                                        3                   NA                  3      
#>   Total number of patients with at least one adverse event                  3 (6.12%)            NA (NA%)           3 (5.08%)  
#>   Postoperative hernia                                                      1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Pelvic fibrosis                                                           1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Vessel puncture site inflammation                                         1 (2.04%)            NA (NA%)           1 (1.69%)  
#> 
#> Musculoskeletal and connective tissue disorders
#>   Total number of events                                                        3                   NA                  3      
#>   Total number of patients with at least one adverse event                  3 (6.12%)            NA (NA%)           3 (5.08%)  
#>   Gout                                                                      1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Serotonin syndrome                                                        1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Myopathy endocrine                                                        1 (2.04%)            NA (NA%)           1 (1.69%)  
#> 
#> Vascular disorders
#>   Total number of events                                                        3                   NA                  3      
#>   Total number of patients with at least one adverse event                  2 (4.08%)            NA (NA%)           2 (3.39%)  
#>   Aortic wall hypertrophy                                                   1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Portal fibrosis                                                           1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Stomatitis haemorrhagic                                                   1 (2.04%)            NA (NA%)           1 (1.69%)  
#> 
#> Congenital, familial and genetic disorders
#>   Total number of events                                                        2                   NA                  2      
#>   Total number of patients with at least one adverse event                  2 (4.08%)            NA (NA%)           2 (3.39%)  
#>   Congenital hearing disorder                                               1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Extrarenal pelvis                                                         1 (2.04%)            NA (NA%)           1 (1.69%)  
#> 
#> Immune system disorders
#>   Total number of events                                                        1                   1                   2      
#>   Total number of patients with at least one adverse event                  1 (2.04%)            1 (10%)            2 (3.39%)  
#>   Tongue oedema                                                             1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Implant site urticaria                                                     NA (NA%)            1 (10%)            1 (1.69%)  
#> 
#> Cardiac disorders
#>   Total number of events                                                        1                   NA                  1      
#>   Total number of patients with at least one adverse event                  1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Corrected transposition of great vessels                                  1 (2.04%)            NA (NA%)           1 (1.69%)  
#> 
#> Endocrine disorders
#>   Total number of events                                                        1                   NA                  1      
#>   Total number of patients with at least one adverse event                  1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Growth accelerated                                                        1 (2.04%)            NA (NA%)           1 (1.69%)  
#> 
#> Hepatobiliary disorders
#>   Total number of events                                                        1                   NA                  1      
#>   Total number of patients with at least one adverse event                  1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Hepatitis B reactivation                                                  1 (2.04%)            NA (NA%)           1 (1.69%)  
#> 
#> Infections and infestations
#>   Total number of events                                                        1                   NA                  1      
#>   Total number of patients with at least one adverse event                  1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Infective pericardial effusion                                            1 (2.04%)            NA (NA%)           1 (1.69%)  
#> 
#> Metabolism and nutrition disorders
#>   Total number of events                                                        1                   NA                  1      
#>   Total number of patients with at least one adverse event                  1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Abnormal weight gain                                                      1 (2.04%)            NA (NA%)           1 (1.69%)  
#> 
#> Neoplasms benign, malignant and unspecified (incl cysts and polyps)
#>   Total number of events                                                        1                   NA                  1      
#>   Total number of patients with at least one adverse event                  1 (2.04%)            NA (NA%)           1 (1.69%)  
#>   Malignant nipple neoplasm male                                            1 (2.04%)            NA (NA%)           1 (1.69%)