Summary table for analysis of covariance (ANCOVA).

This is a wrapper around the basic t_el_ancova function with the difference that it has a `row_by` argument. This can e.g. be used to analyze multiple endpoints and/or multiple timepoints within the same response variable.

t_ancova(formula, data, row_by, col_N, conf_level = 0.95, table_tree = FALSE)

Arguments

formula
A formula corresponding to the investigated linear model. The left-hand side must include the dependent variable. The right-hand side must include the group variable name wrapped in the special arm().

data
A data.frame which includes all the variables that are called in formula.

row_by
(factor or data.frame)
Defines how data from `data` is split into sub-tables. Dimensions must match dimensions of `data` and no missing values are allowed. Multi-level nesting is possible when `row_by` is a data.frame. Columns should be ordered with the first column specifying the first variable to split by and the last column specifying the last variable to split by.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

conf_level
confidence level of the interval.

table_tree
(logical value)
If FALSE then a single rtable gets returned, if TRUE a node object gets returned with the elementary rtable objects.

Value

Either rtable or node object (see `table_tree` argument). For every unique combination of levels of `row_by` an ANCOVA table using t_el_ancova will be created. The individual tables are then stacked together.

Examples

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)
ADQS <- radqs(cached = TRUE) %>%
  dplyr::filter(AVISIT == "WEEK 1 DAY 8")

# Multiple endpoints.
ADQS_week1day8 <- dplyr::filter(ADQS, AVISIT == "WEEK 1 DAY 8")
t_ancova(
  formula = CHG ~ BASE + STRATA1 + arm(ARMCD),
  data = ADQS_week1day8,
  row_by = ADQS_week1day8$PARAMCD,
  col_N = table(ADSL$ARMCD)
)

#>                                          ARM A               ARM B               ARM C    
#>                                         (N=134)             (N=134)             (N=132)   
#> ------------------------------------------------------------------------------------------
#> BFIALL
#>   n                                       134                 134                 132     
#>   Adjusted Mean                           4.47                6.33                4.02    
#>   Difference in Adjusted Means                                1.85               -0.46    
#>     95% CI                                                -0.14 - 3.85        -2.45 - 1.54
#>     p-value                                                  0.0679              0.6539   
#> 
#> FATIGI
#>   n                                       134                 134                 132     
#>   Adjusted Mean                           5.42                4.83                4.56    
#>   Difference in Adjusted Means                               -0.59               -0.86    
#>     95% CI                                                -2.58 - 1.41        -2.87 - 1.15
#>     p-value                                                  0.5644              0.4026   
#> 
#> FKSI-FWB
#>   n                                       134                 134                 132     
#>   Adjusted Mean                           4.29                3.51                3.06    
#>   Difference in Adjusted Means                               -0.79               -1.24    
#>     95% CI                                                -2.71 - 1.14        -3.17 - 0.69
#>     p-value                                                  0.4221              0.2088   
#> 
#> FKSI-TSE
#>   n                                       134                 134                 132     
#>   Adjusted Mean                           4.7                 3.84                4.45    
#>   Difference in Adjusted Means                               -0.86               -0.25    
#>     95% CI                                                -2.8 - 1.09          -2.2 - 1.7 
#>     p-value                                                  0.3858              0.8007   
#> 
#> FKSIALL
#>   n                                       134                 134                 132     
#>   Adjusted Mean                           5.03                5.82                6.44    
#>   Difference in Adjusted Means                                0.79                1.42    
#>     95% CI                                                -1.17 - 2.76        -0.56 - 3.39
#>     p-value                                                  0.4288              0.1591   

# Multiple visits (in this data set there is only one visit, though).
ADQS_fksifwb <- dplyr::filter(ADQS, PARAMCD == "FKSI-FWB")
t_ancova(
  formula = CHG ~ BASE + STRATA1 + arm(ARMCD),
  data = ADQS_fksifwb,
  row_by = ADQS_fksifwb$AVISIT,
  col_N = table(ADSL$ARMCD)
)

#>                                          ARM A               ARM B               ARM C    
#>                                         (N=134)             (N=134)             (N=132)   
#> ------------------------------------------------------------------------------------------
#> WEEK 1 DAY 8
#>   n                                       134                 134                 132     
#>   Adjusted Mean                           4.29                3.51                3.06    
#>   Difference in Adjusted Means                               -0.79               -1.24    
#>     95% CI                                                -2.71 - 1.14        -3.17 - 0.69
#>     p-value                                                  0.4221              0.2088   

# Multiple timepoints and visits.
t_ancova(
  formula = CHG ~ BASE + STRATA1 + arm(ARMCD),
  data = ADQS,
  row_by = ADQS[, c("PARAMCD", "AVISIT")],
  col_N = table(ADSL$ARMCD)
)

#>                                            ARM A               ARM B               ARM C    
#>                                           (N=134)             (N=134)             (N=132)   
#> --------------------------------------------------------------------------------------------
#> BFIALL
#>   WEEK 1 DAY 8
#>     n                                       134                 134                 132     
#>     Adjusted Mean                           4.47                6.33                4.02    
#>     Difference in Adjusted Means                                1.85               -0.46    
#>       95% CI                                                -0.14 - 3.85        -2.45 - 1.54
#>       p-value                                                  0.0679              0.6539   
#> 
#> FATIGI
#>   WEEK 1 DAY 8
#>     n                                       134                 134                 132     
#>     Adjusted Mean                           5.42                4.83                4.56    
#>     Difference in Adjusted Means                               -0.59               -0.86    
#>       95% CI                                                -2.58 - 1.41        -2.87 - 1.15
#>       p-value                                                  0.5644              0.4026   
#> 
#> FKSI-FWB
#>   WEEK 1 DAY 8
#>     n                                       134                 134                 132     
#>     Adjusted Mean                           4.29                3.51                3.06    
#>     Difference in Adjusted Means                               -0.79               -1.24    
#>       95% CI                                                -2.71 - 1.14        -3.17 - 0.69
#>       p-value                                                  0.4221              0.2088   
#> 
#> FKSI-TSE
#>   WEEK 1 DAY 8
#>     n                                       134                 134                 132     
#>     Adjusted Mean                           4.7                 3.84                4.45    
#>     Difference in Adjusted Means                               -0.86               -0.25    
#>       95% CI                                                -2.8 - 1.09          -2.2 - 1.7 
#>       p-value                                                  0.3858              0.8007   
#> 
#> FKSIALL
#>   WEEK 1 DAY 8
#>     n                                       134                 134                 132     
#>     Adjusted Mean                           5.03                5.82                6.44    
#>     Difference in Adjusted Means                                0.79                1.42    
#>       95% CI                                                -1.17 - 2.76        -0.56 - 3.39
#>       p-value                                                  0.4288              0.1591