Summary table for binary outcome

This is a convenience wrapper function for creating a summary of common analyses for binary outcome variables. Options include reporting unstratified analysis, stratified analysis or both.

t_binary_outcome(
  rsp,
  col_by,
  strata_data = NULL,
  conf_level = 0.95,
  prop_ci_method = "waldcc",
  unstrat_analysis = control_binary_comparison(),
  strat_analysis = control_binary_comparison(diff_test = "cmh"),
  rsp_multinomial = NULL
)

Arguments

rsp
(logical vector)
TRUE if subject is a responder FALSE otherwise

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

strata_data
(data.frame)
Used for stratification factors. If NULL, no stratified analysis is performed.

conf_level
confidence level of the interval.

prop_ci_method
one of ("waldcc", "wald", "clopper-pearson", "wilson", agresti-coull or jeffreys).
Specifies the method used to construct the confidence interval for proportion of successful outcomes.

unstrat_analysis
list specified by calling control_binary_comparison for controlling the "Unstratified Analysis" section. If NULL, this table section is excluded.

strat_analysis
list specified by calling control_binary_comparison for controlling the "Stratified Analysis" section. Only applicable if stratification factors from strata_data are provided. If NULL, this table section is excluded.

rsp_multinomial
(factor vector) The levels represent ordered response categories. NA values or empty strings are not allowed. If NULL, tabulation by each response category will not be performed.

Value

an rtable object

Details

Table subsections are matched by name, so to remove a subsection set any the corresponding arguments of unstrat_analysis, strat_analysis or rsp_multinomial to NULL.

The display order of response categories in partitioned statistics section inherits the factor level order of partition_analysis. Use factor and its levels argument to include or exclude response categories and arrange display order. If response values contain "NE", "Missing", "NE/Missing", "Not Evaluable (NE)" or "Missing or unevaluable", the confidence interval will not be calculated.

See also

Examples

library(random.cdisc.data)
library(dplyr)
ADRS <- radrs(cached = TRUE) %>%
  dplyr::filter(PARAMCD == "BESRSPI")

# Unstratified analysis only with default options for `unstrat_analysis`.
t_binary_outcome(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD
)

#>                                                                    ARM A                     ARM B                     ARM C       
#>                                                                   (N=134)                   (N=134)                   (N=132)      
#> -----------------------------------------------------------------------------------------------------------------------------------
#> Responders                                                      119 (88.81%)              97 (72.39%)               120 (90.91%)   
#> 95% CI (Wald, with correction)                                 (83.09, 94.52)            (64.45, 80.33)            (85.63, 96.19)  
#> 
#> Unstratified Analysis
#>   Difference in Response Rates (%)                                                           -16.42                     2.1        
#>     95% CI (Wald, with correction)                                                      (-26.43, -6.41)             (-5.9, 10.1)   
#>   p-value (Chi-Squared Test with Schouten Correction)                                        0.0009                    0.6410      
#>   Odds Ratio (95% CI)                                                                  0.33 (0.17 - 0.64)        1.26 (0.57 - 2.81)

# Unstratified + stratified analysis with default options for
# `unstrat_analysis` and `strat_analysis`.
t_binary_outcome(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  strata_data = ADRS[, c("STRATA1", "STRATA2")]
)

#>                                                                    ARM A                     ARM B                     ARM C       
#>                                                                   (N=134)                   (N=134)                   (N=132)      
#> -----------------------------------------------------------------------------------------------------------------------------------
#> Responders                                                      119 (88.81%)              97 (72.39%)               120 (90.91%)   
#> 95% CI (Wald, with correction)                                 (83.09, 94.52)            (64.45, 80.33)            (85.63, 96.19)  
#> 
#> Unstratified Analysis
#>   Difference in Response Rates (%)                                                           -16.42                     2.1        
#>     95% CI (Wald, with correction)                                                      (-26.43, -6.41)             (-5.9, 10.1)   
#>   p-value (Chi-Squared Test with Schouten Correction)                                        0.0009                    0.6410      
#>   Odds Ratio (95% CI)                                                                  0.33 (0.17 - 0.64)        1.26 (0.57 - 2.81)
#> 
#> Stratified Analysis*
#>   Difference in Response Rates (%)                                                           -16.42                     2.1        
#>     95% CI (Wald, with correction)                                                      (-26.43, -6.41)             (-5.9, 10.1)   
#>   p-value (Cochran-Mantel-Haenszel Test)                                                     0.0007                    0.5284      
#>   Odds Ratio (95% CI)                                                                  0.33 (0.17 - 0.64)        1.26 (0.56 - 2.8) 
#> 
#> * Model stratified by STRATA1 and STRATA2

# Stratified analysis only, using CMH weights and corresponding test statistic.
# Exclude Odds Ratio estimates.
t_binary_outcome(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  strata_data = ADRS[, c("STRATA1", "STRATA2")],
  unstrat_analysis = NULL,
  strat_analysis = control_binary_comparison(diff_ci = "cmh", diff_test = "cmh", odds_ratio = FALSE)
 )

#>                                                         ARM A                  ARM B                  ARM C     
#>                                                        (N=134)                (N=134)                (N=132)    
#> ----------------------------------------------------------------------------------------------------------------
#> Responders                                          119 (88.81%)            97 (72.39%)           120 (90.91%)  
#> 95% CI (Wald, with correction)                     (83.09, 94.52)         (64.45, 80.33)         (85.63, 96.19) 
#> 
#> Stratified Analysis*
#>   Weighted Difference in Response Rates (%)                                   -16.54                  2.31      
#>     95% CI (CMH, without correction)                                      (-25.69, -7.39)         (-4.77, 9.4)  
#>   p-value (Cochran-Mantel-Haenszel Test)                                      0.0007                 0.5284     
#> 
#> * Model stratified by STRATA1 and STRATA2

# Exclude any comparisons between arms,
# specify non-default CI method for proportions and CI level,
# and include individual response categories.

# First recode multinomial response with standard labels.
m_rsp <- ADRS$AVALC
levels(m_rsp) <- get_onco_rsp_label()[match(levels(m_rsp), names(get_onco_rsp_label()))]

t_binary_outcome(
  ADRS$AVALC %in% c("CR", "PR"),
  col_by = ADRS$ARMCD,
  prop_ci_method = "clopper-pearson",
  unstrat_analysis = NULL,
  conf_level = 0.90,
  rsp_multinomial = m_rsp
 )

#>                                     ARM A                 ARM B                 ARM C     
#>                                    (N=134)               (N=134)               (N=132)    
#> ------------------------------------------------------------------------------------------
#> Responders                       133 (99.25%)          127 (94.78%)          131 (99.24%) 
#> 90% CI (Clopper-Pearson)        (96.51, 99.96)        (90.41, 97.52)        (96.46, 99.96)
#> 
#> Complete Response (CR)           119 (88.81%)          97 (72.39%)           120 (90.91%) 
#> 90% CI (Clopper-Pearson)        (83.29, 92.97)        (65.33, 78.69)        (85.69, 94.67)
#> 
#> Partial Response (PR)            14 (10.45%)           30 (22.39%)            11 (8.33%)  
#> 90% CI (Clopper-Pearson)        (6.43, 15.85)         (16.6, 29.12)         (4.74, 13.42) 
#> 
#> Stable Disease (SD)               1 (0.75%)             7 (5.22%)             1 (0.76%)   
#> 90% CI (Clopper-Pearson)         (0.04, 3.49)          (2.48, 9.59)          (0.04, 3.54) 
#> 
#> Progressive Disease (PD)              0                     0                     0       
#> 90% CI (Clopper-Pearson)          (0, 2.21)             (0, 2.21)             (0, 2.24)   
#> 
#> Not Evaluable (NE)                    0                     0                     0