Contingency table (stacked per group)

This produces a stacked list of contingency tables, one table per group. The tables report the absolute (n) and relative frequency (%) for every combination of modalities between two factors. For each group, the frequencies (n) add up to the group size (N) and percentages are calculated using N as the denominator.

t_contingency(
  row_grp,
  col,
  row,
  row_grp_lab = NULL,
  col_lab = NULL,
  row_lab = NULL,
  na.strings = "Missing"
)

Arguments

row_grp, col, row
The three factors used for row grouping, column modalities and row modalities, respectively. The three factors are expected to be of equal length.

row_grp_lab, row_lab, col_lab
Optional labels for groups of rows, rows and columns annotation.

na.strings
A character string for transformation of missing data into explicit factor level.

Value

an rtable object

Details

Every group is associated with a table composed of a headline where the group name and size (N) is reported and a contingency table. The table is annotated with the label attributes of row_grp, col and row. If the label attribute is missing, the annotation is left blank, unless labels are manually set via the corresponding *_lab arguments.

The function is expected to be robust to the classes of the provided objects and handling of missing values: non-factors are coerced to factors and missing values are transformed into explicit factor levels.

Examples

# Example 1:
# ==========

n <- 50
g <- sample(LETTERS[1:3], size = n, replace = TRUE)
x <- sample(letters[1:3], size = n, replace = TRUE)
y <- sample(1:3, size = n, replace = TRUE)

t_contingency(g, x, y)

#>                                                            
#>                     a                b                c    
#> -----------------------------------------------------------
#> A (N=17)
#>   1             1 (5.9%)         3 (17.6%)        2 (11.8%)
#>   2             1 (5.9%)         2 (11.8%)        5 (29.4%)
#>   3             1 (5.9%)         1 (5.9%)         1 (5.9%) 
#> 
#> B (N=17)
#>   1             5 (29.4%)        2 (11.8%)        3 (17.6%)
#>   2             1 (5.9%)         2 (11.8%)        1 (5.9%) 
#>   3              0 (0%)          2 (11.8%)        1 (5.9%) 
#> 
#> C (N=16)
#>   1             1 (6.2%)         1 (6.2%)         5 (31.2%)
#>   2             2 (12.5%)        2 (12.5%)        2 (12.5%)
#>   3             2 (12.5%)         0 (0%)          1 (6.2%) 

# Example 2:
# ==========

g <- sample(c(rep(NA, n/5), x), size = n)
x <- sample(c(rep(NA, n/5), y), size = n)
y <- sample(c(rep(NA, n/5), g), size = n)

t_contingency(
  row_grp = g,
  col = x,
  row = y,
  na.strings = "Missing",
  row_grp_lab = "Treatment",
  row_lab = "Levels of Factor A",
  col_lab = "Levels of factor B"
)

#> Treatment                                        Levels of factor B                     
#>   Levels of Factor A            1                2                3             Missing 
#> ----------------------------------------------------------------------------------------
#> a (N=11)
#>   a                         1 (9.1%)         1 (9.1%)          0 (0%)           0 (0%)  
#>   b                         1 (9.1%)          0 (0%)           0 (0%)           0 (0%)  
#>   c                         2 (18.2%)        1 (9.1%)          0 (0%)          1 (9.1%) 
#>   Missing                   1 (9.1%)         3 (27.3%)         0 (0%)           0 (0%)  
#> 
#> b (N=13)
#>   a                          0 (0%)          2 (15.4%)         0 (0%)           0 (0%)  
#>   b                         2 (15.4%)         0 (0%)           0 (0%)          2 (15.4%)
#>   c                         3 (23.1%)         0 (0%)           0 (0%)           0 (0%)  
#>   Missing                   2 (15.4%)        1 (7.7%)         1 (7.7%)          0 (0%)  
#> 
#> c (N=18)
#>   a                         1 (5.6%)         1 (5.6%)          0 (0%)          2 (11.1%)
#>   b                         1 (5.6%)         1 (5.6%)         1 (5.6%)         1 (5.6%) 
#>   c                         2 (11.1%)         0 (0%)          2 (11.1%)        2 (11.1%)
#>   Missing                   2 (11.1%)        2 (11.1%)         0 (0%)           0 (0%)  
#> 
#> Missing (N=8)
#>   a                          0 (0%)          1 (12.5%)         0 (0%)           0 (0%)  
#>   b                         1 (12.5%)        1 (12.5%)        1 (12.5%)        1 (12.5%)
#>   c                          0 (0%)           2 (25%)          0 (0%)           0 (0%)  
#>   Missing                    0 (0%)          1 (12.5%)         0 (0%)           0 (0%)  

# Example 3:
# ==========

attr(g, "label") <- "Arm"
attr(x, "label") <- "Baseline"
attr(y, "label") <- "Post-Baseline"

t_contingency(g, y, x)

#> Arm                                         Post-Baseline                        
#>   Baseline               a                b                c             Missing 
#> ---------------------------------------------------------------------------------
#> a (N=11)
#>   1                  1 (9.1%)         1 (9.1%)         2 (18.2%)        1 (9.1%) 
#>   2                  1 (9.1%)          0 (0%)          1 (9.1%)         3 (27.3%)
#>   3                   0 (0%)           0 (0%)           0 (0%)           0 (0%)  
#>   Missing             0 (0%)           0 (0%)          1 (9.1%)          0 (0%)  
#> 
#> b (N=13)
#>   1                   0 (0%)          2 (15.4%)        3 (23.1%)        2 (15.4%)
#>   2                  2 (15.4%)         0 (0%)           0 (0%)          1 (7.7%) 
#>   3                   0 (0%)           0 (0%)           0 (0%)          1 (7.7%) 
#>   Missing             0 (0%)          2 (15.4%)         0 (0%)           0 (0%)  
#> 
#> c (N=18)
#>   1                  1 (5.6%)         1 (5.6%)         2 (11.1%)        2 (11.1%)
#>   2                  1 (5.6%)         1 (5.6%)          0 (0%)          2 (11.1%)
#>   3                   0 (0%)          1 (5.6%)         2 (11.1%)         0 (0%)  
#>   Missing            2 (11.1%)        1 (5.6%)         2 (11.1%)         0 (0%)  
#> 
#> Missing (N=8)
#>   1                   0 (0%)          1 (12.5%)         0 (0%)           0 (0%)  
#>   2                  1 (12.5%)        1 (12.5%)         2 (25%)         1 (12.5%)
#>   3                   0 (0%)          1 (12.5%)         0 (0%)           0 (0%)  
#>   Missing             0 (0%)          1 (12.5%)         0 (0%)           0 (0%)  

# Example 4:
# ==========

g <- x <- y <- rep(NA, 50)

t_contingency(g, y, x)

#>                                
#>                        Missing 
#> -------------------------------
#> Missing (N=50)
#>   Missing             50 (100%)