Tabulate TRUE counts in a vector

This function is a wrapper for t_summary.logical returning only the counts for TRUE records.

# S3 method for logical
t_count_true(
  x,
  col_by,
  col_N = NULL,
  total = NULL,
  denominator = c("N", "n", "omit"),
  row_name = deparse(substitute(x)),
  ...
)

Arguments

x
a logical vector

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

denominator
(one of ("n", "N", "omit")
Value set to "n" or "N" are for calculating the level associated percentage. With option "N", the reference population from col_N is used as the denominator. With option "n", the number of non-missing records from x is used as the denominator. If "omit" is chosen the percentage is omitted.

row_name
name of row

...
not used arguments

See also

Examples

with(iris, t_count_true(Sepal.Length > 5.5, col_by = Species, total = "All Flowers"))

#>                             setosa           versicolor          virginica         All Flowers
#>                             (N=50)             (N=50)             (N=50)             (N=150)  
#> ----------------------------------------------------------------------------------------------
#> Sepal.Length > 5.5          3 (6%)            39 (78%)           49 (98%)          91 (60.67%)

# compare with t_summary.logical
with(iris, t_summary(Sepal.Length > 5.5, col_by = Species, total = "All Flowers"))

#>                setosa           versicolor          virginica         All Flowers
#>                (N=50)             (N=50)             (N=50)             (N=150)  
#> ---------------------------------------------------------------------------------
#> n                50                 50                 50                 150    
#> TRUE           3 (6%)            39 (78%)           49 (98%)          91 (60.67%)
#> FALSE         47 (94%)           11 (22%)            1 (2%)           59 (39.33%)

# show only counts without percent
with(iris, t_count_true(Sepal.Length > 5.5, col_by = Species, denominator = "omit"))

#>                             setosa          versicolor        virginica 
#>                             (N=50)            (N=50)            (N=50)  
#> ------------------------------------------------------------------------
#> Sepal.Length > 5.5            3                 39                49