Count unique elements per cell in a vector with subgroups

This is a wrapper function for t_el_count_unique useful for tabulating unique values in subgroups.

t_count_unique(
  x,
  col_by,
  subgroups,
  col_N = NULL,
  total = NULL,
  na_rm = TRUE,
  denominator = c("N", "omit"),
  table_tree = FALSE
)

Arguments

x
a vector

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

subgroups
A logical data.frame with rows matching the length of x. Each column represents a subset to be applied to x and col_by prior to tabulating the unique elements per cell. Label attributes are used for the row names if they exist. NAs are cast to FALSE.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

na_rm
(logical value)
Indicating whether NA values should be removed from x prior to counting the unique elements per cell.

denominator
(One of ("N", "omit"))
With option N, the reference population from col_N is used as the denominator. If omit is chosen, the percentage is omitted and only the counts are returned.

table_tree
(logical value)
If FALSE then a single rtable gets returned, if TRUE a node object gets returned with the elementary rtable objects.

Value

an rtable

See also

Examples

# define analysis subgroups
sbs <- data.frame("grp1" = c(TRUE, TRUE, FALSE, FALSE),
                 "grp2" = c(FALSE, TRUE, TRUE, FALSE),
                 "grp3" = c(FALSE, FALSE, FALSE, FALSE),
                 "grp4" = c(NA, FALSE, NA, TRUE))

t_count_unique(x = c(1:4), col_by = factor(c("A", "A", "B", "B")), subgroups = sbs,
  col_N = c(4, 10))

#>                A              B   
#>              (N=4)         (N=10) 
#> ----------------------------------
#> grp1        2 (50%)           0   
#> grp2        1 (25%)        1 (10%)
#> grp3           0              0   
#> grp4           0           1 (10%)

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)
ADAE <- radae(cached = TRUE)

# define analysis subgroups
ANL <- ADAE %>%
 dplyr::transmute(
    SER = AESER == 'Y' ,
    REL = AEREL == 'Y',
    GR5 =  AETOXGR == '5'
  ) %>%
  var_relabel(
    SER = 'Serious AE',
    REL = 'Relasted AE',
    GR5 = 'Grade 5 AE'
  )

# number of unqiue patients with events:
t_count_unique(ADAE$USUBJID, col_by = ADAE$ARM, subgroups = ANL,
  col_N = table(ADSL$ARM), total = "All")

#>                      A: Drug X             B: Placebo          C: Combination             All      
#>                       (N=134)               (N=134)               (N=132)               (N=400)    
#> ---------------------------------------------------------------------------------------------------
#> Serious AE          104 (77.61%)          101 (75.37%)            99 (75%)             304 (76%)   
#> Relasted AE         105 (78.36%)          108 (80.6%)           109 (82.58%)          322 (80.5%)  
#> Grade 5 AE          76 (56.72%)           70 (52.24%)           75 (56.82%)           221 (55.25%) 

# omit percent
tbl <- t_count_unique(ADAE$USUBJID, col_by = ADAE$ARM, subgroups = ANL,
  col_N = table(ADSL$ARM), denominator = "omit")
tbl

#>                      A: Drug X             B: Placebo          C: Combination
#>                       (N=134)               (N=134)               (N=132)    
#> -----------------------------------------------------------------------------
#> Serious AE              104                   101                    99      
#> Relasted AE             105                   108                   109      
#> Grade 5 AE               76                    70                    75      

# add descriptive row
insert_rrow(indent(tbl, 1), rrow("Total number of patients with at least one"))

#>                                                     A: Drug X             B: Placebo          C: Combination
#>                                                      (N=134)               (N=134)               (N=132)    
#> ------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one
#>   Serious AE                                           104                   101                    99      
#>   Relasted AE                                          105                   108                   109      
#>   Grade 5 AE                                            76                    70                    75