Proportional Hazards Regression Model Fit Summary Table (pair-wise for multiple arms)

An rtable format of s_coxph_pairwise results for further annotation on top of Kaplan-Meier grob.

t_coxph_pairwise(
  formula,
  data,
  conf_level = 0.95,
  pval_method = c("log-rank", "wald", "likelihood"),
  ...
)

Arguments

formula
(formula)
Specifies survival model. The arm variable needs to be wrapped in arm. The strata special will only be used for the stratified analysis. If there is not strata specification then the stratified analysis is omitted.

data
(data.frame)
Contains all the variable that are used in formula

conf_level
(numeric value)
level for computation of the confidence intervals. If set to FALSE no confidence intervals are printed

pval_method
(one of ("log-rank", "likelihood", "wald"))
Specifies the method used to calculate the p-value. Default value is "log-rank".

...
(optional)
Other arguments will be passed to coxph

Details

The Cox PH model is evaluated pair-wise (reference to comparison) and coxph is used to get the p-value, calculate the hazard ratio and confidence interval. For example, given an ARM variable has reference group "ARM A" and two comparison groups "ARM B" and "ARM C", t_coxph_pairwise will conduct two Cox PH models for the specified formula, one is for the subset data with only "ARM A" and "ARM B", the other is for the subset data with only "ARM A" and "ARM C".

Examples

library(random.cdisc.data)
ADSL <- cadsl
ADTTE <- cadtte
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")

tbl <- t_coxph_pairwise(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARM) + strata(SEX),
  data = ADTTE_f
)
tbl

#>                              HR                 95% CI of HR          Log-rank p-value
#> --------------------------------------------------------------------------------------
#> B: Placebo                 1.0273             (0.7632, 1.3829)             0.8589     
#> C: Combination             1.2792             (0.9355, 1.7493)             0.1222     

tbl <- t_coxph_pairwise(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + strata(SEX),
  data = ADTTE_f,
  conf_level = 0.8, pval_method = "wald", ties = "exact"
)
tbl

#>                     HR                 80% CI of HR            Wald p-value  
#> -----------------------------------------------------------------------------
#> ARM B             1.0273             (0.8459, 1.2477)             0.8589     
#> ARM C             1.2792             (1.0425, 1.5697)             0.123      

# If the event never happened in selection, a note is returned in the tabulation
tbl <- t_coxph_pairwise(
  formula = Surv(time = AVAL, event = CNSR - CNSR) ~ arm(ARMCD) + strata(SEX),
  data = ADTTE_f,
  conf_level = 0.8, pval_method = "wald", ties = "exact"
)
tbl

#>                        HR                    80% CI of HR               Wald p-value    
#> ----------------------------------------------------------------------------------------
#> *NOTE:                 No estimates as the event was not observed in selection.