Cox proportional hazards regression for standard outputs

Cox models are the most commonly used methods to estimate the magnitude of the effect in survival analysis. It assumes proportional hazards: the ratio of the hazards of between groups (e.g., two arms) is constant over time. This ratio is referred to as the "hazard ratio" (HR) and is one of the most commonly reported metrics to describe the effect size in survival analysis.

t_coxreg(
  formula,
  data,
  simplify = TRUE,
  conf_level = 0.95,
  pval_method = c("wald", "likelihood", "log-rank"),
  increments = NULL,
  ...
)

Arguments

formula
A formula corresponding to the survival model:

the left-hand side must include Surv(time, event) with time and event describing occurrence and censoring (see Surv()).

the right-hand side includes predictors and specials which can be used to specify a pairwise model or a range of candidate univariate models. See details.

data
A data.frame which includes all the variables that are called in formula.

simplify
If TRUE, returns a rtables object instead of a list of rtables object when the variable contained in pairwise has only two levels to compare.

conf_level
The confidence level of the interval.

pval_method
The method used for the estimation of p.values, should be one of wald (default) or likelihood. The log-rank is accepted for models including a single variable, specified through pairwise().

increments
If a quantitative variable is included, it is possible to provide the expected level of estimation for the interaction. If provided, it should be list where each item is a vector giving expected levels and is named after the variable name as it appears in the formula.

...
Optional other arguments passed to the Cox regression:

ties a character string specifying the method for tie handling, one of exact (default), efron, breslow.

see coxph() for additional settings.

Value

Depending on formula, returns an rtable object for every level of pairwise() if included in COXT01; a single rtable otherwise.

Details

Possible model specifications:

Surv(time, event) ~ pairwise(Pred), used to get hazard ratio of each tested level of treatment Pred independently, given in reference to the control treatment (see pairwise()).

Surv(time, event) ~ pairwise(Pred) + univariate(Cov1, Cov2, ...), add candidate covariates separated by comas in special univariate() to test all bivariate combinations of Pred with each Cov. Replacing the symbol + by * will result in the additional estimation of the interaction terms. This is similar to COXT01 standards.

Surv(time, event) ~ Pred + Cov1 + ... estimates the simple effect of multiple Cox regression. This provides the COXT02 standard output. Note that the variables in returned tabulation are matched by position in formula.

Known limits:

the special univariate() requires pairwise() being also specified.

the interaction terms in multiple regression are not estimated.

the third-order interactions are not estimated.

For further information about the Cox Proportional Hazards Model, check "Statistical Analysis of Clinical Trials Data with R", NEST team.

See also

Examples

library(tern)
library(random.cdisc.data)
ADTTE   <- radtte(cached = TRUE)
ADTTE_f <- subset(ADTTE, PARAMCD == "OS") # _f: filtered
ADTTE_f <- within( # nolint
  data = subset(
    ADTTE_f,
    PARAMCD == "OS"
    & SEX %in% c("F", "M")
    & RACE %in% c("ASIAN", "BLACK OR AFRICAN AMERICAN", "WHITE")
  ),
  expr = { # nolint start
    set.seed(1)
    ARMCD <- droplevels(ARMCD)
    ARMCD <- relevel(ARMCD, "ARM B")
    SEX <- droplevels(SEX)
    RACE <- droplevels(RACE)
    X <- rnorm(n = length(ARM))
  }  # nolint end
)

# For annotation on top of Kaplan-Meier grob.
t_coxreg(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ pairwise(ARMCD),
  data = ADTTE_f, conf_level = 0.8, pval_method = "likelihood",
  ties = "breslow"
)

#>                      HR                   80% CI of HR           Likelihood p-value
#> -----------------------------------------------------------------------------------
#> ARM A              0.9656               (0.7889, 1.1819)               0.8242      
#> ARM C              1.324                (1.0855, 1.615)                0.0708      

## COXT01 - Standard output, no interactions
t_coxreg(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ pairwise(ARMCD) +
  univariate(SEX, RACE, AGE),
  data = ADTTE_f
)

#> $`ARM A`
#>                                                                         Treatment effect adjusted for covariate #         
#> Effect/                                                             Hazard                                                
#> Covariate included                                                Ratio for                                               
#> in the model                                    n                 Treatment              95% CI*               p-value*   
#> --------------------------------------------------------------------------------------------------------------------------
#> Treatment ARM A vs control (ARM B)             247                   0.97              (0.71, 1.32)             0.8243    
#> 
#> Covariate:
#>   SEX                                          247                   0.97              (0.71, 1.31)             0.8224    
#>   RACE                                         247                   0.97              (0.71, 1.33)             0.8701    
#>   AGE                                          247                   0.92              (0.68, 1.26)             0.6084    
#> 
#> # Model including treatment, 
#>   covariate, no interaction.
#> * Wald CI/test
#> Ties: Efron
#> 
#> $`ARM C`
#>                                                                         Treatment effect adjusted for covariate #         
#> Effect/                                                             Hazard                                                
#> Covariate included                                                Ratio for                                               
#> in the model                                    n                 Treatment              95% CI*               p-value*   
#> --------------------------------------------------------------------------------------------------------------------------
#> Treatment ARM C vs control (ARM B)             247                   1.32              (0.98, 1.79)             0.0702    
#> 
#> Covariate:
#>   SEX                                          247                   1.4               (1.02, 1.92)             0.0373    
#>   RACE                                         247                   1.32              (0.98, 1.8)              0.0699    
#>   AGE                                          247                   1.33              (0.98, 1.8)              0.0659    
#> 
#> # Model including treatment, 
#>   covariate, no interaction.
#> * Wald CI/test
#> Ties: Efron
#> 

if (FALSE) {
# For comparison
mod10 <- coxph(Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD), data = ADTTE_f)
mod11 <- coxph(Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + SEX, data = ADTTE_f)
mod12 <- coxph(Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + AGE, data = ADTTE_f)
mod13 <- coxph(Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + RACE, data = ADTTE_f)

car::Anova(mod10,  test.statistic = "Wald", type = 'III')
car::Anova(mod11, test.statistic = "Wald", type = 'III')
car::Anova(mod12, test.statistic = "Wald", type = 'III')
car::Anova(mod13, test.statistic = "Wald", type = 'III')

summary(mod13)$coefficients
}

## COXT01 - options: control ties, set confidence interval level,
##   choose a strata, modify pval_method
t_coxreg(
  formula = Surv(time = AVAL, event = 1 - CNSR) ~ pairwise(ARMCD) + strata(SEX) +
    univariate(RACE, AGE),
  data = ADTTE_f, conf_level = 0.80,
  ties = c("exact", "efron", "breslow")[3], pval_method  = c("wald", "likelihood")[2]
)

#> skipping term strata(SEX)

#> Warning in `try_car_anova`: simpleWarning in Anova.coxph(mod, test.statistic = test.statistic, type = "III"): LR tests not supported for models with clusters or strata
#>  Wald tests substituted

#> skipping term strata(SEX)

#> Warning in `try_car_anova`: simpleWarning in Anova.coxph(mod, test.statistic = test.statistic, type = "III"): LR tests not supported for models with clusters or strata
#>  Wald tests substituted

#> skipping term strata(SEX)

#> Warning in `try_car_anova`: simpleWarning in Anova.coxph(mod, test.statistic = test.statistic, type = "III"): LR tests not supported for models with clusters or strata
#>  Wald tests substituted

#> skipping term strata(SEX)

#> Warning in `try_car_anova`: simpleWarning in Anova.coxph(mod, test.statistic = test.statistic, type = "III"): LR tests not supported for models with clusters or strata
#>  Wald tests substituted

#> skipping term strata(SEX)

#> Warning in `try_car_anova`: simpleWarning in Anova.coxph(mod, test.statistic = test.statistic, type = "III"): LR tests not supported for models with clusters or strata
#>  Wald tests substituted

#> skipping term strata(SEX)

#> Warning in `try_car_anova`: simpleWarning in Anova.coxph(mod, test.statistic = test.statistic, type = "III"): LR tests not supported for models with clusters or strata
#>  Wald tests substituted

#> $`ARM A`
#>                                                                         Treatment effect adjusted for covariate #         
#> Effect/                                                             Hazard                                                
#> Covariate included                                                Ratio for                                               
#> in the model                                    n                 Treatment              80% CI*               p-value*   
#> --------------------------------------------------------------------------------------------------------------------------
#> Treatment ARM A vs control (ARM B)             247                   0.97              (0.79, 1.18)             0.8226    
#> 
#> Covariate:
#>   RACE                                         247                   0.97              (0.79, 1.19)             0.8582    
#>   AGE                                          247                   0.91              (0.74, 1.12)             0.5654    
#> 
#> # Model including treatment, 
#>   covariate, no interaction.
#> * Likelihood CI/test
#> Ties: Breslow
#> 
#> $`ARM C`
#>                                                                         Treatment effect adjusted for covariate #         
#> Effect/                                                             Hazard                                                
#> Covariate included                                                Ratio for                                               
#> in the model                                    n                 Treatment              80% CI*               p-value*   
#> --------------------------------------------------------------------------------------------------------------------------
#> Treatment ARM C vs control (ARM B)             247                   1.35              (1.09, 1.66)             0.066     
#> 
#> Covariate:
#>   RACE                                         247                   1.34              (1.09, 1.65)             0.0696    
#>   AGE                                          247                   1.36              (1.1, 1.68)              0.0606    
#> 
#> # Model including treatment, 
#>   covariate, no interaction.
#> * Likelihood CI/test
#> Ties: Breslow
#> 

if (FALSE) {
# For comparison
mod20 <- coxph(Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + strata(SEX), data = ADTTE_f)
mod21 <- coxph(
  Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + strata(SEX) + AGE, data = ADTTE_f
)
mod22 <- coxph(
  Surv(time = AVAL, event = 1 - CNSR) ~ arm(ARMCD) + strata(SEX) + RACE,
  data = ADTTE_f
)

car::Anova(mod20, test.statistic = "Wald", type = 'III')
car::Anova(mod21, test.statistic = "Wald", type = 'III')
car::Anova(mod22, test.statistic = "Wald", type = 'III')

summary(mod22)$coefficients
}

if (FALSE) {
## COXT01 - Addition of the interaction and increments to choose estimated levels
t_coxreg(
    formula =  Surv(time = AVAL, event = 1 - CNSR) ~
      pairwise(ARMCD) * univariate(RACE, SEX, AGE, X),
    data = ADTTE_f, conf_level = 0.74, increments = list(X = c(-1, 0, 2))
)

## COXT01: Other example, other data
library(dplyr)
ADSL <- radsl(cached = TRUE)
ADSL <- ADSL %>% dplyr::filter(SEX %in% c("F", "M"))

ADTTE <- radtte(ADSL, seed = 2) %>%
    dplyr::filter(PARAMCD == "PFS" & ARMCD != "ARM C")
ADTTE$ARMCD <- droplevels(ADTTE$ARMCD)
ADTTE$SEX <- droplevels(ADTTE$SEX)
t_coxreg(
    formula = Surv(time = AVAL, event = 1 - CNSR) ~
      pairwise(ARMCD) * univariate(SEX, BMRKR2, AGE),
    data = ADTTE,
    increments = list(AGE = mean(ADTTE$AGE)),
    conf_level = 0.975)

# check HRs
m2 <- coxph(Surv(time = AVAL, event = 1 - CNSR) ~ ARMCD*BMRKR2, data = ADTTE)
coef(m2)

# For COXT02
t_coxreg(
    formula = Surv(time = AVAL, event = 1 - CNSR) ~ ARMCD + RACE + AGE +
    strata(SEX), data = ADTTE_f,
    conf_level = 0.8, pval_method = "likelihood",  ties = "breslow"
 )
}