Count abnormality shift from baseline for each abnormality level

This is the elementary function used in function t_abn_shift, where the calculation could be repeated
for multiple post-baseline visits and multiple laboratory tests.

t_el_abn_shift(
  rri_ana,
  rri_base,
  value_ana,
  value_base,
  abnormal = c("HIGH", "LOW"),
  id,
  col_by,
  col_N = NULL,
  total = NULL
)

Arguments

rri_ana
a vector specifying the Reference Range Indicator (e.g. High/Low/Normal).
Fill blank with NA using sas_na().

rri_base
a vector specifying the Reference Range Indicator (e.g. High/Low/Normal) at baseline
for the corresponding records in rri_ana. Fill blank with NA using sas_na().

value_ana
a vector of numeric measurement.

value_base
a vector of numeric measurement at baseline
for the corresponding records in value_ana.

abnormal
a vector specifying the unique values that represent abnormality in the order of display.

id
a vector of subject identifier with length of rri_ana.

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

Value

an rtable

Examples

library(dplyr)

ANL <- tibble(
  USUBJID = c("ID1", "ID1", "ID2", "ID2", "ID3", "ID3"),
  ANRIND = c("N", "L", "N", "N", "N", "H"),
  BNRIND = c("N", "N", "L", "L", "", ""),
  AVAL = c(1.5, 0.4, 1.8, 1.9, 2.4, 4.6),
  BASE = c(1.2, 1.2, 0.5, 0.5, NA, NA),
  ARM = factor(rep("ARM A", 6)),
)

t_el_abn_shift(
  rri_ana = sas_na(ANL$ANRIND),
  rri_base = sas_na(ANL$BNRIND),
  value_ana = ANL$AVAL,
  value_base = ANL$BASE,
  abnormal = c("H", "L"),
  id = ANL$USUBJID,
  col_by = ANL$ARM,
  col_N = table(ANL$ARM),
  total = "All Patients"
)

#>                                      ARM A              All Patients  
#>          Baseline Status             (N=6)                  (N=6)     
#> ----------------------------------------------------------------------
#> H             NOT H                   0/2                    0/2      
#>               Total               1/3 (33.3%)            1/3 (33.3%)  
#> L             NOT L               1/1 (100%)             1/1 (100%)   
#>                 L                     0/1                    0/1      
#>               Total               1/3 (33.3%)            1/3 (33.3%)  

t_el_abn_shift(
  rri_ana = sas_na(ANL$ANRIND),
  rri_base = sas_na(ANL$BNRIND),
  value_ana = ANL$AVAL,
  value_base = ANL$BASE,
  abnormal = c("H", "L"),
  id = ANL$USUBJID,
  col_by = by_all("All Patients"),
  col_N = nrow(ANL)
)

#>                                  All Patients  
#>          Baseline Status             (N=6)     
#> -----------------------------------------------
#> H             NOT H                   0/2      
#>               Total               1/3 (33.3%)  
#> L             NOT L               1/1 (100%)   
#>                 L                     0/1      
#>               Total               1/3 (33.3%)