Elementary table for analysis of covariance (ANCOVA).

This function returns an elementary summary table for an ANCOVA analysis, which compares the covariate adjusted means of multiple groups with a reference group. Specifically, the first level of the `arm()` factor variable on the right-hand side of the provided `formula` is taken as the reference group.

t_el_ancova(formula, data, col_N, conf_level = 0.95)

Arguments

formula
A formula corresponding to the investigated linear model. The left-hand side must include the dependent variable. The right-hand side must include the group variable name wrapped in the special arm().

data
A data.frame which includes all the variables that are called in formula.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

conf_level
confidence level of the interval.

Value

an rtable object

See also

Examples

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)
ADQS_filtered <- radqs(cached = TRUE) %>%
  dplyr::filter(PARAMCD == "FKSI-FWB", AVISIT == "WEEK 1 DAY 8")
t_el_ancova(
  formula = CHG ~ BASE + STRATA1 + arm(ARMCD),
  data = ADQS_filtered,
  col_N = table(ADSL$ARMCD)
)

#>                                        ARM A               ARM B               ARM C    
#>                                       (N=134)             (N=134)             (N=132)   
#> ----------------------------------------------------------------------------------------
#> n                                       134                 134                 132     
#> Adjusted Mean                           4.29                3.51                3.06    
#> Difference in Adjusted Means                               -0.79               -1.24    
#>   95% CI                                                -2.71 - 1.14        -3.17 - 0.69
#>   p-value                                                  0.4221              0.2088