Single element for disposition table

Creates a disposition table

t_el_disposition(
  x,
  col_by,
  col_N = NULL,
  total = NULL,
  row.name = NULL,
  subset = NULL,
  show_n = FALSE,
  useNA = c("no", "ifany", "always"),
  drop_levels = NULL,
  denominator = "N"
)

Arguments

x
(factor or logical vector)

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

row.name
(character value)
Only applicable when x is a logical vector. A string to label the row name. Default is "TRUE".

subset
(logical vector)
Contains the same length as x that defines the subset to take from x. NAs are cast to false.

show_n
(logical value) Determines whether the "n" row is displayed. Default is FALSE. Only when x is of type factor

useNA
(logical)
choose whether missing data (NAs) should be displayed as a level.

drop_levels
(logical value)
Whether to drop zero count levels

denominator
(one of ("n", "N", "omit")
Value set to "n" or "N" are for calculating the level associated percentage. With option "N", the reference population from col_N is used as the denominator. With option "n", the number of non-missing records from x is used as the denominator. If "omit" is chosen the percentage is omitted.

Examples

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)

ADSL0 <- ADSL %>%
  mutate(
    COMPSTUD = sample(c('Y','N'),
                      size=nrow(ADSL),
                      replace = TRUE) %>% as.factor,
    STUDONS = sample(c('Alive: On Treatment', 'Alive: In Follow-up', NA),
                     size=nrow(ADSL),
                     replace = TRUE) %>% as.factor,
    STDDRS = sample(c('Death', 'Lost To Follow-Up',
                      'Protocol Violation', 'Withdrawal By Subject',
                      'Other'),
                    size=nrow(ADSL),
                    replace = TRUE) %>% as.factor,
    GOTTRT = ifelse(!is.na(ACTARMCD), 'Y', 'N') %>% as.factor,
    DISTRTFL = sample(c('Y','N'),
                      size=nrow(ADSL),
                      replace = TRUE) %>% as.factor,
    TRTDRS = sample(c('ADVERSE EVENT', 'PROGRESSIVE DISEASE',
                      'PHYSICIAN DECISION', 'LACK OF EFFICACY',
                      'OTHER'),
                    size=nrow(ADSL),
                    replace = TRUE) %>% as.factor,
    STUDONS = case_when(
      COMPSTUD == 'N' ~ STUDONS
    ),
    STDDRS = case_when(
      COMPSTUD == 'N' & is.na(STUDONS) ~ STDDRS
    ) %>% as.factor,
    DISSTDFL = case_when(
      !is.na(STDDRS) ~ "Y"
    ),
    DISTRTFL = case_when(
      GOTTRT == 'Y' ~ DISTRTFL
    ),
    TRTDRS = case_when(
      DISTRTFL == 'Y' ~ TRTDRS
    ),
    DRSCAT = case_when(
      TRTDRS %in% c('ADVERSE EVENT', 'PHYSICIAN DECISION') ~ "Safety",
      !is.na(TRTDRS) ~ "Other"
    )
  ) %>% var_relabel(COMPSTUD = "Complete Study",
                    STUDONS = "On-study Status",
                    DISSTDFL = "Discontinued Study",
                    STDDRS = "Reason for Study \r\nDiscontinuation",
                    GOTTRT = "Received Treatment",
                    DISTRTFL = "Discontinued Treatment",
                    TRTDRS = "Reason for Treatment \r\nDiscontinuation"
  )

library(purrr)
t_el_disposition_indent <- function(..., indent = 0) indent(t_el_disposition(...), indent)
dsp <- partial(
  t_el_disposition_indent,
  col_by = ADSL0$ARM,
  total = "All Patients",
  denominator = "N"
)
# Only using rtables
# old way of doing things with indent variable
rbind(
  dsp(ADSL0$COMPSTUD == "Y", row.name = "Completed study"),
  dsp(ADSL0$STUDONS == "Alive: In Follow-up", row.name = "Alive: In follow-up"),
  dsp(ADSL0$DISSTDFL == "Y",  row.name = "Discontinued study"),
  dsp(ADSL0$STDDRS, indent = 1),
  rrow(),
  rrow("Show example of using n as denominator"),
  indent(t_el_disposition(
    ADSL0$STDDRS,
    col_by = ADSL0$ARM,
    total = "All Patients",
    denominator = "n"
  ), 1),
  rrow(),
  dsp(ADSL0$GOTTRT == "Y", row.name = "Received treatment"),
  dsp(ADSL0$DISTRTFL == "Y", row.name = "Discontinued treatment"),
  dsp(ADSL0$DRSCAT == "Safety", row.name = "Safety", indent = 1),
  dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT %in% "Safety", indent = 2),
  dsp(ADSL0$DRSCAT == "Other", row.name = "Other", indent = 1),
  dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT %in% "Other", indent = 2)
)

#>                                                 A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                                  (N=134)               (N=134)               (N=132)               (N=400)    
#> ------------------------------------------------------------------------------------------------------------------------------
#> Completed study                                72 (53.73%)           79 (58.96%)           68 (51.52%)           219 (54.75%) 
#> Alive: In follow-up                            16 (11.94%)           18 (13.43%)           20 (15.15%)            54 (13.5%)  
#> Discontinued study                             21 (15.67%)           16 (11.94%)           20 (15.15%)           57 (14.25%)  
#>   Death                                         2 (1.49%)             1 (0.75%)             4 (3.03%)             7 (1.75%)   
#>   Lost To Follow-Up                             2 (1.49%)             2 (1.49%)             1 (0.76%)             5 (1.25%)   
#>   Other                                         6 (4.48%)             4 (2.99%)             5 (3.79%)             15 (3.75%)  
#>   Protocol Violation                            7 (5.22%)             5 (3.73%)             3 (2.27%)             15 (3.75%)  
#>   Withdrawal By Subject                         4 (2.99%)             4 (2.99%)              7 (5.3%)             15 (3.75%)  
#> 
#> Show example of using n as denominator
#>   Death                                         2 (9.52%)             1 (6.25%)              4 (20%)              7 (12.28%)  
#>   Lost To Follow-Up                             2 (9.52%)             2 (12.5%)               1 (5%)              5 (8.77%)   
#>   Other                                         6 (28.57%)             4 (25%)               5 (25%)             15 (26.32%)  
#>   Protocol Violation                            7 (33.33%)            5 (31.25%)             3 (15%)             15 (26.32%)  
#>   Withdrawal By Subject                         4 (19.05%)             4 (25%)               7 (35%)             15 (26.32%)  
#> 
#> Received treatment                              134 (100%)            134 (100%)            132 (100%)            400 (100%)  
#> Discontinued treatment                         60 (44.78%)           69 (51.49%)             66 (50%)            195 (48.75%) 
#>   Safety                                        26 (19.4%)           33 (24.63%)           25 (18.94%)             84 (21%)   
#>     ADVERSE EVENT                               12 (8.96%)            13 (9.7%)             12 (9.09%)            37 (9.25%)  
#>     PHYSICIAN DECISION                         14 (10.45%)           20 (14.93%)            13 (9.85%)           47 (11.75%)  
#>   Other                                        34 (25.37%)           36 (26.87%)           41 (31.06%)           111 (27.75%) 
#>     LACK OF EFFICACY                            8 (5.97%)             11 (8.21%)            12 (9.09%)            31 (7.75%)  
#>     OTHER                                       8 (5.97%)             10 (7.46%)           20 (15.15%)            38 (9.5%)   
#>     PROGRESSIVE DISEASE                        18 (13.43%)           15 (11.19%)            9 (6.82%)             42 (10.5%)  

# new way:
treatment_section_rtable <- rbind(
  dsp(ADSL0$GOTTRT == "Y", row.name = "Received treatment"),
  dsp(ADSL0$DISTRTFL == "Y", row.name = "Discontinued treatment"),
  indent(
    rbind(
      dsp(ADSL0$DRSCAT == "Safety", row.name = "Safety"),
      indent(dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT %in% "Safety"), 1),
      dsp(ADSL0$DRSCAT == "Other", row.name = "Other"),
      indent(dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT %in% "Other"), 1)
    ),
    1
  )
)
treatment_section_rtable

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                   (N=134)               (N=134)               (N=132)               (N=400)    
#> ---------------------------------------------------------------------------------------------------------------
#> Received treatment               134 (100%)            134 (100%)            132 (100%)            400 (100%)  
#> Discontinued treatment          60 (44.78%)           69 (51.49%)             66 (50%)            195 (48.75%) 
#>   Safety                         26 (19.4%)           33 (24.63%)           25 (18.94%)             84 (21%)   
#>     ADVERSE EVENT                12 (8.96%)            13 (9.7%)             12 (9.09%)            37 (9.25%)  
#>     PHYSICIAN DECISION          14 (10.45%)           20 (14.93%)            13 (9.85%)           47 (11.75%)  
#>   Other                         34 (25.37%)           36 (26.87%)           41 (31.06%)           111 (27.75%) 
#>     LACK OF EFFICACY             8 (5.97%)             11 (8.21%)            12 (9.09%)            31 (7.75%)  
#>     OTHER                        8 (5.97%)             10 (7.46%)           20 (15.15%)            38 (9.5%)   
#>     PROGRESSIVE DISEASE         18 (13.43%)           15 (11.19%)            9 (6.82%)             42 (10.5%)  

# easier to manipulate for their purposes by other users
treatment_section_tree <- node(
  invisible_node_name("treatment"),
  content = NULL,
  format_data = list(children_gap = 0),
  children = list(
    node(
      invisible_node_name("received_treatment"),
      content = dsp(ADSL0$GOTTRT == "Y", row.name = "Received treatment")
    ),
    node(
      invisible_node_name("discontinuedtreatment"),
      dsp(ADSL0$DISTRTFL == "Y", row.name = "Discontinued treatment"),
      format_data = list(gap_to_children = 0, children_gap = 0, children_indent = 1),
      children = list(
        node(
          invisible_node_name("safety"),
          rbind(
            dsp(ADSL0$DRSCAT == "Safety", row.name = "Safety"),
            indent(dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT %in% "Safety"), 1)
          )
        ),
        node(
          invisible_node_name("other"),
          rbind(
            dsp(ADSL0$DRSCAT == "Other", row.name = "Other"),
            indent(dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT %in% "Other"), 1)
          )
        )
      )
    )
  )
)
to_rtable(treatment_section_tree)

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                   (N=134)               (N=134)               (N=132)               (N=400)    
#> ---------------------------------------------------------------------------------------------------------------
#> Received treatment               134 (100%)            134 (100%)            132 (100%)            400 (100%)  
#> Discontinued treatment          60 (44.78%)           69 (51.49%)             66 (50%)            195 (48.75%) 
#>   Safety                         26 (19.4%)           33 (24.63%)           25 (18.94%)             84 (21%)   
#>     ADVERSE EVENT                12 (8.96%)            13 (9.7%)             12 (9.09%)            37 (9.25%)  
#>     PHYSICIAN DECISION          14 (10.45%)           20 (14.93%)            13 (9.85%)           47 (11.75%)  
#>   Other                         34 (25.37%)           36 (26.87%)           41 (31.06%)           111 (27.75%) 
#>     LACK OF EFFICACY             8 (5.97%)             11 (8.21%)            12 (9.09%)            31 (7.75%)  
#>     OTHER                        8 (5.97%)             10 (7.46%)           20 (15.15%)            38 (9.5%)   
#>     PROGRESSIVE DISEASE         18 (13.43%)           15 (11.19%)            9 (6.82%)             42 (10.5%)  

to_rtable(invisible_node(
  name = "t_el_disp",
  children = list(
    node(
      name = invisible_node_name("Study status"),
      content = rbind(
        dsp(ADSL0$COMPSTUD == "Y", row.name = "Completed study"),
        dsp(ADSL0$STUDONS == "Alive: In Follow-up", row.name = "Alive: In follow-up")
      ),
      children = list(
        node(
          name = invisible_node_name("Reasons for study discontinuation (using capital N)"),
          rbind(
            dsp(ADSL0$DISSTDFL == "Y",  row.name = "Discontinued study"),
            indent(dsp(ADSL0$STDDRS), 1)
          )
        ),
        node(
          name = "Reasons for study discontinuation (using small n)",
          content = t_el_disposition(
            ADSL0$STDDRS,
            col_by = ADSL0$ARM,
            total = "All Patients",
            denominator = "n"
          )
        )
      )
    ),
    treatment_section_tree
  )
))

#>                                                            A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                                             (N=134)               (N=134)               (N=132)               (N=400)    
#> -----------------------------------------------------------------------------------------------------------------------------------------
#> Completed study                                           72 (53.73%)           79 (58.96%)           68 (51.52%)           219 (54.75%) 
#> Alive: In follow-up                                       16 (11.94%)           18 (13.43%)           20 (15.15%)            54 (13.5%)  
#> 
#> Discontinued study                                        21 (15.67%)           16 (11.94%)           20 (15.15%)           57 (14.25%)  
#>   Death                                                    2 (1.49%)             1 (0.75%)             4 (3.03%)             7 (1.75%)   
#>   Lost To Follow-Up                                        2 (1.49%)             2 (1.49%)             1 (0.76%)             5 (1.25%)   
#>   Other                                                    6 (4.48%)             4 (2.99%)             5 (3.79%)             15 (3.75%)  
#>   Protocol Violation                                       7 (5.22%)             5 (3.73%)             3 (2.27%)             15 (3.75%)  
#>   Withdrawal By Subject                                    4 (2.99%)             4 (2.99%)              7 (5.3%)             15 (3.75%)  
#> 
#> Reasons for study discontinuation (using small n)
#>   Death                                                    2 (9.52%)             1 (6.25%)              4 (20%)              7 (12.28%)  
#>   Lost To Follow-Up                                        2 (9.52%)             2 (12.5%)               1 (5%)              5 (8.77%)   
#>   Other                                                    6 (28.57%)             4 (25%)               5 (25%)             15 (26.32%)  
#>   Protocol Violation                                       7 (33.33%)            5 (31.25%)             3 (15%)             15 (26.32%)  
#>   Withdrawal By Subject                                    4 (19.05%)             4 (25%)               7 (35%)             15 (26.32%)  
#> 
#> Received treatment                                         134 (100%)            134 (100%)            132 (100%)            400 (100%)  
#> Discontinued treatment                                    60 (44.78%)           69 (51.49%)             66 (50%)            195 (48.75%) 
#>   Safety                                                   26 (19.4%)           33 (24.63%)           25 (18.94%)             84 (21%)   
#>     ADVERSE EVENT                                          12 (8.96%)            13 (9.7%)             12 (9.09%)            37 (9.25%)  
#>     PHYSICIAN DECISION                                    14 (10.45%)           20 (14.93%)            13 (9.85%)           47 (11.75%)  
#>   Other                                                   34 (25.37%)           36 (26.87%)           41 (31.06%)           111 (27.75%) 
#>     LACK OF EFFICACY                                       8 (5.97%)             11 (8.21%)            12 (9.09%)            31 (7.75%)  
#>     OTHER                                                  8 (5.97%)             10 (7.46%)           20 (15.15%)            38 (9.5%)   
#>     PROGRESSIVE DISEASE                                   18 (13.43%)           15 (11.19%)            9 (6.82%)             42 (10.5%)