Proportions: outcomes by level

Table for proportion of outcomes per factor level

t_el_multinomial_proportion(
  rsp_multinomial,
  col_by,
  conf_level = 0.95,
  prop_ci_method = "waldcc",
  levels_without_ci = NULL
)

Arguments

rsp_multinomial
(factor vector) The levels represent ordered response categories. NA values or empty strings are not allowed.

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

conf_level
confidence level of the interval.

prop_ci_method
one of ("waldcc", "wald", "clopper-pearson", "wilson", agresti-coull or jeffreys).
Specifies the method used to construct the confidence interval for proportion of successful outcomes.

levels_without_ci
(character vector) If any levels of rsp_multinomial match, the corresponding rates are reported without a confidence interval.

Value

return_rtable

See also

Examples

library(random.cdisc.data)
library(dplyr)

ADRS <- radrs(cached = TRUE) %>%
  dplyr::filter(PARAMCD == "BESRSPI")

t_el_multinomial_proportion(ADRS$AVALC, col_by = ADRS$ARMCD)

#>                                           ARM A                 ARM B                 ARM C     
#>                                          (N=134)               (N=134)               (N=132)    
#> ------------------------------------------------------------------------------------------------
#> CR                                     119 (88.81%)          97 (72.39%)           120 (90.91%) 
#> 95% CI (Wald, with correction)        (83.09, 94.52)        (64.45, 80.33)        (85.63, 96.19)
#> 
#> PR                                     14 (10.45%)           30 (22.39%)            11 (8.33%)  
#> 95% CI (Wald, with correction)          (4.9, 16)           (14.96, 29.82)        (3.24, 13.43) 
#> 
#> SD                                      1 (0.75%)             7 (5.22%)             1 (0.76%)   
#> 95% CI (Wald, with correction)          (0, 2.58)            (1.08, 9.36)           (0, 2.62)   
#> 
#> PD                                          0                     0                     0       
#> 95% CI (Wald, with correction)          (0, 0.37)             (0, 0.37)             (0, 0.38)   
#> 
#> NE                                          0                     0                     0       
#> 95% CI (Wald, with correction)          (0, 0.37)             (0, 0.37)             (0, 0.38)   

t_el_multinomial_proportion(ADRS$AVALC, col_by = ADRS$ARMCD, levels_without_ci = "NE")

#>                                           ARM A                 ARM B                 ARM C     
#>                                          (N=134)               (N=134)               (N=132)    
#> ------------------------------------------------------------------------------------------------
#> CR                                     119 (88.81%)          97 (72.39%)           120 (90.91%) 
#> 95% CI (Wald, with correction)        (83.09, 94.52)        (64.45, 80.33)        (85.63, 96.19)
#> 
#> PR                                     14 (10.45%)           30 (22.39%)            11 (8.33%)  
#> 95% CI (Wald, with correction)          (4.9, 16)           (14.96, 29.82)        (3.24, 13.43) 
#> 
#> SD                                      1 (0.75%)             7 (5.22%)             1 (0.76%)   
#> 95% CI (Wald, with correction)          (0, 2.58)            (1.08, 9.36)           (0, 2.62)   
#> 
#> PD                                          0                     0                     0       
#> 95% CI (Wald, with correction)          (0, 0.37)             (0, 0.37)             (0, 0.38)   
#> 
#> NE                                          0                     0                     0       

# apply standard oncology labels, given factor input.
m_rsp <- ADRS$AVALC
levels(m_rsp) <- get_onco_rsp_label()[match(levels(m_rsp), names(get_onco_rsp_label()))]
t_el_multinomial_proportion(m_rsp, col_by = ADRS$ARMCD, levels_without_ci = "Not Evaluable (NE)")

#>                                           ARM A                 ARM B                 ARM C     
#>                                          (N=134)               (N=134)               (N=132)    
#> ------------------------------------------------------------------------------------------------
#> Complete Response (CR)                 119 (88.81%)          97 (72.39%)           120 (90.91%) 
#> 95% CI (Wald, with correction)        (83.09, 94.52)        (64.45, 80.33)        (85.63, 96.19)
#> 
#> Partial Response (PR)                  14 (10.45%)           30 (22.39%)            11 (8.33%)  
#> 95% CI (Wald, with correction)          (4.9, 16)           (14.96, 29.82)        (3.24, 13.43) 
#> 
#> Stable Disease (SD)                     1 (0.75%)             7 (5.22%)             1 (0.76%)   
#> 95% CI (Wald, with correction)          (0, 2.58)            (1.08, 9.36)           (0, 2.62)   
#> 
#> Progressive Disease (PD)                    0                     0                     0       
#> 95% CI (Wald, with correction)          (0, 0.37)             (0, 0.37)             (0, 0.38)   
#> 
#> Not Evaluable (NE)                          0                     0                     0       

t_el_multinomial_proportion(
  ADRS$AVALC,
  col_by = ADRS$ARMCD,
  prop_ci_method = "wilson",
  conf_level = 0.90
)

#>                            ARM A                 ARM B                 ARM C     
#>                           (N=134)               (N=134)               (N=132)    
#> ---------------------------------------------------------------------------------
#> CR                      119 (88.81%)          97 (72.39%)           120 (90.91%) 
#> 90% CI (Wilson)        (83.54, 92.54)        (65.64, 78.25)        (85.93, 94.24)
#> 
#> PR                      14 (10.45%)           30 (22.39%)            11 (8.33%)  
#> 90% CI (Wilson)         (6.86, 15.6)         (17.04, 28.82)        (5.16, 13.18) 
#> 
#> SD                       1 (0.75%)             7 (5.22%)             1 (0.76%)   
#> 90% CI (Wilson)         (0.17, 3.28)          (2.86, 9.36)          (0.17, 3.32) 
#> 
#> PD                           0                     0                     0       
#> 90% CI (Wilson)          (0, 1.98)             (0, 1.98)             (0, 2.01)   
#> 
#> NE                           0                     0                     0       
#> 90% CI (Wilson)          (0, 1.98)             (0, 1.98)             (0, 2.01)   

# single arm
t_el_multinomial_proportion(
  ADRS$AVALC,
  col_by = by_all("All"),
  prop_ci_method = "clopper-pearson"
)

#>                                      All      
#>                                    (N=400)    
#> ----------------------------------------------
#> CR                                336 (84%)   
#> 95% CI (Clopper-Pearson)        (80.03, 87.45)
#> 
#> PR                               55 (13.75%)  
#> 95% CI (Clopper-Pearson)        (10.53, 17.52)
#> 
#> SD                                9 (2.25%)   
#> 95% CI (Clopper-Pearson)         (1.03, 4.23) 
#> 
#> PD                                    0       
#> 95% CI (Clopper-Pearson)          (0, 0.92)   
#> 
#> NE                                    0       
#> 95% CI (Clopper-Pearson)          (0, 0.92)   

# Add total
t_el_multinomial_proportion(
  ADRS$AVALC,
  col_by = by_add_total(ADRS$ARMCD, "All")
)

#>                                           ARM A                 ARM B                 ARM C                  All      
#>                                          (N=134)               (N=134)               (N=132)               (N=400)    
#> ----------------------------------------------------------------------------------------------------------------------
#> CR                                     119 (88.81%)          97 (72.39%)           120 (90.91%)           336 (84%)   
#> 95% CI (Wald, with correction)        (83.09, 94.52)        (64.45, 80.33)        (85.63, 96.19)        (80.28, 87.72)
#> 
#> PR                                     14 (10.45%)           30 (22.39%)            11 (8.33%)           55 (13.75%)  
#> 95% CI (Wald, with correction)          (4.9, 16)           (14.96, 29.82)        (3.24, 13.43)         (10.25, 17.25)
#> 
#> SD                                      1 (0.75%)             7 (5.22%)             1 (0.76%)             9 (2.25%)   
#> 95% CI (Wald, with correction)          (0, 2.58)            (1.08, 9.36)           (0, 2.62)            (0.67, 3.83) 
#> 
#> PD                                          0                     0                     0                     0       
#> 95% CI (Wald, with correction)          (0, 0.37)             (0, 0.37)             (0, 0.38)             (0, 0.12)   
#> 
#> NE                                          0                     0                     0                     0       
#> 95% CI (Wald, with correction)          (0, 0.37)             (0, 0.37)             (0, 0.38)             (0, 0.12)