OR: elementary summary table

Elementary summary table for odds ratio estimation.

t_el_odds_ratio(rsp, col_by, conf_level = 0.95, strata_data = NULL)

Arguments

rsp
(logical vector)
TRUE if subject is a responder FALSE otherwise

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

conf_level
confidence level of the interval.

strata_data
(data.frame) Optional stratification factors. If NULL, the unstratified analysis will be performed.

Value

an rtable object

Details

This function uses either logistic regression for unstratified analyses, or conditional logistic regression for stratified analyses. The Wald confidence interval with the specified confidence level is calculated. Note that for stratified analyses, there is currently no implementation for conditional likelihood confidence intervals, therefore we don't offer the option of a likelihood confidence interval. Note that when `rsp` contains only responders or non-responders, then the result values will be `NA`, because no odds ratio estimation is possible.

See also

Examples

library(random.cdisc.data)
library(dplyr)

ADRS <- radrs(cached = TRUE) %>%
  dplyr::filter(PARAMCD == "BESRSPI")

t_el_odds_ratio(
  rsp = ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  conf_level = 0.9
)

#>                                  ARM A                     ARM B                     ARM C       
#>                                 (N=134)                   (N=134)                   (N=132)      
#> -------------------------------------------------------------------------------------------------
#> Responders                    119 (88.81%)              97 (72.39%)               120 (90.91%)   
#> Odds Ratio (90% CI)                                  0.33 (0.19 - 0.57)        1.26 (0.64 - 2.47)

t_el_odds_ratio(
  rsp = ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  strata_data = ADRS[, c("STRATA1", "STRATA2")]
)

#>                                  ARM A                     ARM B                     ARM C       
#>                                 (N=134)                   (N=134)                   (N=132)      
#> -------------------------------------------------------------------------------------------------
#> Responders                    119 (88.81%)              97 (72.39%)               120 (90.91%)   
#> Odds Ratio (95% CI)                                  0.33 (0.17 - 0.64)        1.26 (0.56 - 2.8)