Proportion: outcomes

Table for proportion of successful outcomes

t_el_proportion(rsp, col_by, conf_level = 0.95, prop_ci_method = "waldcc")

Arguments

rsp
(logical vector)
TRUE if subject is a responder FALSE otherwise

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

conf_level
confidence level of the interval.

prop_ci_method
one of ("waldcc", "wald", "clopper-pearson", "wilson", agresti-coull or jeffreys).
Specifies the method used to construct the confidence interval for proportion of successful outcomes.

Value

an rtable object

Details

The following options are supported for confidence interval for a proportion:

The wald interval follows the usual textbook definition for a single proportion confidence interval using the normal approximation.

The waldcc interval is similar to wald and in addition uses the continuity correction 1/2n.

The clopper-pearson interval calls binom.test. Also referred to as the 'exact' method.

The wilson interval calls prop.test with option correct=FALSE. Also referred to as Wilson score interval.

The agresti-coull interval was created by Alan Agresti and Brent Coull and can be understood (for 95% CI) as adding two successes and two failures to the data, and then using the Wald formula to construct a CI.

The jeffreys interval is an equal-tailed interval based on the non-informative Jeffreys prior for a binomial proportion.

See also

Examples

library(random.cdisc.data)
library(dplyr)

ADRS <- radrs(cached = TRUE) %>%
  dplyr::filter(PARAMCD == "BESRSPI")

t_el_proportion(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD
)

#>                                           ARM A                 ARM B                 ARM C     
#>                                          (N=134)               (N=134)               (N=132)    
#> ------------------------------------------------------------------------------------------------
#> Responders                             119 (88.81%)          97 (72.39%)           120 (90.91%) 
#> 95% CI (Wald, with correction)        (83.09, 94.52)        (64.45, 80.33)        (85.63, 96.19)

t_el_proportion(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  prop_ci_method = "wilson",
  conf_level = 0.90
)

#>                            ARM A                 ARM B                 ARM C     
#>                           (N=134)               (N=134)               (N=132)    
#> ---------------------------------------------------------------------------------
#> Responders              119 (88.81%)          97 (72.39%)           120 (90.91%) 
#> 90% CI (Wilson)        (83.54, 92.54)        (65.64, 78.25)        (85.93, 94.24)

# single arm
t_el_proportion(
  ADRS$AVALC == "CR",
  col_by = by_all("All"),
  prop_ci_method = "clopper-pearson"
)

#>                                      All      
#>                                    (N=400)    
#> ----------------------------------------------
#> Responders                        336 (84%)   
#> 95% CI (Clopper-Pearson)        (80.03, 87.45)

# Add total
t_el_proportion(
  ADRS$AVALC == "CR",
  col_by = by_add_total(ADRS$ARMCD, "All")
)

#>                                           ARM A                 ARM B                 ARM C                  All      
#>                                          (N=134)               (N=134)               (N=132)               (N=400)    
#> ----------------------------------------------------------------------------------------------------------------------
#> Responders                             119 (88.81%)          97 (72.39%)           120 (90.91%)           336 (84%)   
#> 95% CI (Wald, with correction)        (83.09, 94.52)        (64.45, 80.33)        (85.63, 96.19)        (80.28, 87.72)