Events by Highest Grade Table

This function summarizes number of unique subjects by highest grade and events term(s). Events terms can be one level term or two level terms (one higher level and one lower level). An implementation example is to apply t_events_per_term_grade_id on Adverse Event Data to create Adverse Events by Highest NCI CTCAE grade table (AET04, STREAM2.x, STREAM1.17).

t_events_per_term_grade_id(
  terms,
  id,
  grade,
  col_by,
  col_N = NULL,
  total = NULL,
  event_type = "event",
  any_grade = "Any Grade",
  prune_zero_rows = TRUE,
  table_tree = FALSE
)

Arguments

terms
(character or factor vector, or data.frame)
Represents events information. Multi-level nesting is possible when terms is a data.frame and columns should be ordered with the first column specifying the highest level term and the last column the lowest level term.

id
(vector)
contains subject identifier. Length of id must be the same as the length or number of rows of terms.

grade
(vector)
contains grade of adverse event. For factors, it is assumed that intensity corresponds to the order of the factor levels, from lowest to worst level. If that is not the case, you have to preprocess using levels(grade) <- new_levels beforehand. The resulting table is then ordered in the same fashion. To instead display it in descending maximum grade order, you need to apply some postprocessing.

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

col_N
numeric vector with information of the number of patients in the levels of col_by. This is useful if there are patients that have no adverse events can be accounted for with this argument.

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

event_type
(character value) to specify the type of event that is summarized, event by default. Only displayed when terms has 2 columns.

any_grade
string to specify the row name which counts any occurrence, it is named Any Grade by default

prune_zero_rows
logical whether to prune all zero rows

table_tree
(logical value)
If FALSE then a single rtable gets returned, if TRUE a node object gets returned with the elementary rtable objects.

Value

an rtable object.

Details

t_events_per_term_grade_id includes percentages based on the total number of subjects in the column heading (i.e. "N=nnn"). col_N can be explicitly specified to get N for percentage calculation from either events dataset or additional dataset like subject level dataset. See the example.

Multiple occurrences of the same event in an individual id are counted only once when counting number of subjects.

t_events_per_term_grade_id doesn't deal with data with missing grade. Impute or filter missing values with sas_na and explicit_na before using t_events_per_term_grade_id.

t_events_per_term_grade_id orders data by overall frequency across all columns from the most commonly reported higher level term to the least frequent one. Within each group of higher level term, it sorts by decreasing frequency of lower level term. It brakes ties using terms names in alphabetical order.

t_events_per_term_grade_id fills in col_by and grade with 0 value in case there was no events reported for particular col_by and/or grade category.

See also

Examples

library(dplyr)
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADAE <- radae(ADSL, 4L, seed = 2)

t_events_per_term_grade_id(
  terms = ADAE$AEDECOD,
  id = ADAE$USUBJID,
  grade = ADAE$AETOXGR,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM)
)

#> Dictionary-Derived Term                                              A: Drug X                     B: Placebo                   C: Combination     
#>                                Analysis Toxicity Grade                (N=134)                        (N=134)                        (N=132)        
#> ---------------------------------------------------------------------------------------------------------------------------------------------------
#> - Overall -                         - Any Grade -                   110 (82.1%)                    98 (73.1%)                     108 (81.8%)      
#>                                           1                          6 (4.5%)                       11 (8.2%)                      13 (9.8%)       
#>                                           2                         14 (10.4%)                      11 (8.2%)                       4 (3%)         
#>                                           3                         31 (23.1%)                     20 (14.9%)                     26 (19.7%)       
#>                                           4                         18 (13.4%)                     14 (10.4%)                     24 (18.2%)       
#>                                           5                         41 (30.6%)                     42 (31.3%)                     41 (31.1%)       
#> 
#> dcd B.2.1.2.1                       - Any Grade -                   32 (23.9%)                     25 (18.7%)                     25 (18.9%)       
#>                                           3                         32 (23.9%)                     25 (18.7%)                     25 (18.9%)       
#> 
#> dcd C.1.1.1.3                       - Any Grade -                   27 (20.1%)                     23 (17.2%)                     30 (22.7%)       
#>                                           4                         27 (20.1%)                     23 (17.2%)                     30 (22.7%)       
#> 
#> dcd D.1.1.4.2                       - Any Grade -                   24 (17.9%)                     22 (16.4%)                     31 (23.5%)       
#>                                           3                         24 (17.9%)                     22 (16.4%)                     31 (23.5%)       
#> 
#> dcd B.2.2.3.1                       - Any Grade -                   21 (15.7%)                     25 (18.7%)                     26 (19.7%)       
#>                                           1                         21 (15.7%)                     25 (18.7%)                     26 (19.7%)       
#> 
#> dcd C.2.1.2.1                       - Any Grade -                   20 (14.9%)                     28 (20.9%)                     23 (17.4%)       
#>                                           2                         20 (14.9%)                     28 (20.9%)                     23 (17.4%)       
#> 
#> dcd D.1.1.1.1                       - Any Grade -                   22 (16.4%)                     23 (17.2%)                     25 (18.9%)       
#>                                           5                         22 (16.4%)                     23 (17.2%)                     25 (18.9%)       
#> 
#> dcd B.1.1.1.1                       - Any Grade -                   24 (17.9%)                     23 (17.2%)                     21 (15.9%)       
#>                                           5                         24 (17.9%)                     23 (17.2%)                     21 (15.9%)       
#> 
#> dcd A.1.1.1.2                       - Any Grade -                   25 (18.7%)                     20 (14.9%)                     21 (15.9%)       
#>                                           2                         25 (18.7%)                     20 (14.9%)                     21 (15.9%)       
#> 
#> dcd A.1.1.1.1                       - Any Grade -                   21 (15.7%)                     20 (14.9%)                     23 (17.4%)       
#>                                           1                         21 (15.7%)                     20 (14.9%)                     23 (17.4%)       
#> 
#> dcd D.2.1.5.3                       - Any Grade -                   18 (13.4%)                     24 (17.9%)                     22 (16.7%)       
#>                                           1                         18 (13.4%)                     24 (17.9%)                     22 (16.7%)       

# don't prune zero rows
t_events_per_term_grade_id(
  terms = ADAE$AEDECOD,
  id = ADAE$USUBJID,
  grade = ADAE$AETOXGR,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  prune_zero_rows = FALSE
)

#> Dictionary-Derived Term                                              A: Drug X                     B: Placebo                   C: Combination     
#>                                Analysis Toxicity Grade                (N=134)                        (N=134)                        (N=132)        
#> ---------------------------------------------------------------------------------------------------------------------------------------------------
#> - Overall -                         - Any Grade -                   110 (82.1%)                    98 (73.1%)                     108 (81.8%)      
#>                                           1                          6 (4.5%)                       11 (8.2%)                      13 (9.8%)       
#>                                           2                         14 (10.4%)                      11 (8.2%)                       4 (3%)         
#>                                           3                         31 (23.1%)                     20 (14.9%)                     26 (19.7%)       
#>                                           4                         18 (13.4%)                     14 (10.4%)                     24 (18.2%)       
#>                                           5                         41 (30.6%)                     42 (31.3%)                     41 (31.1%)       
#> 
#> dcd B.2.1.2.1                       - Any Grade -                   32 (23.9%)                     25 (18.7%)                     25 (18.9%)       
#>                                           1                              0                              0                              0           
#>                                           2                              0                              0                              0           
#>                                           3                         32 (23.9%)                     25 (18.7%)                     25 (18.9%)       
#>                                           4                              0                              0                              0           
#>                                           5                              0                              0                              0           
#> 
#> dcd C.1.1.1.3                       - Any Grade -                   27 (20.1%)                     23 (17.2%)                     30 (22.7%)       
#>                                           1                              0                              0                              0           
#>                                           2                              0                              0                              0           
#>                                           3                              0                              0                              0           
#>                                           4                         27 (20.1%)                     23 (17.2%)                     30 (22.7%)       
#>                                           5                              0                              0                              0           
#> 
#> dcd D.1.1.4.2                       - Any Grade -                   24 (17.9%)                     22 (16.4%)                     31 (23.5%)       
#>                                           1                              0                              0                              0           
#>                                           2                              0                              0                              0           
#>                                           3                         24 (17.9%)                     22 (16.4%)                     31 (23.5%)       
#>                                           4                              0                              0                              0           
#>                                           5                              0                              0                              0           
#> 
#> dcd B.2.2.3.1                       - Any Grade -                   21 (15.7%)                     25 (18.7%)                     26 (19.7%)       
#>                                           1                         21 (15.7%)                     25 (18.7%)                     26 (19.7%)       
#>                                           2                              0                              0                              0           
#>                                           3                              0                              0                              0           
#>                                           4                              0                              0                              0           
#>                                           5                              0                              0                              0           
#> 
#> dcd C.2.1.2.1                       - Any Grade -                   20 (14.9%)                     28 (20.9%)                     23 (17.4%)       
#>                                           1                              0                              0                              0           
#>                                           2                         20 (14.9%)                     28 (20.9%)                     23 (17.4%)       
#>                                           3                              0                              0                              0           
#>                                           4                              0                              0                              0           
#>                                           5                              0                              0                              0           
#> 
#> dcd D.1.1.1.1                       - Any Grade -                   22 (16.4%)                     23 (17.2%)                     25 (18.9%)       
#>                                           1                              0                              0                              0           
#>                                           2                              0                              0                              0           
#>                                           3                              0                              0                              0           
#>                                           4                              0                              0                              0           
#>                                           5                         22 (16.4%)                     23 (17.2%)                     25 (18.9%)       
#> 
#> dcd B.1.1.1.1                       - Any Grade -                   24 (17.9%)                     23 (17.2%)                     21 (15.9%)       
#>                                           1                              0                              0                              0           
#>                                           2                              0                              0                              0           
#>                                           3                              0                              0                              0           
#>                                           4                              0                              0                              0           
#>                                           5                         24 (17.9%)                     23 (17.2%)                     21 (15.9%)       
#> 
#> dcd A.1.1.1.2                       - Any Grade -                   25 (18.7%)                     20 (14.9%)                     21 (15.9%)       
#>                                           1                              0                              0                              0           
#>                                           2                         25 (18.7%)                     20 (14.9%)                     21 (15.9%)       
#>                                           3                              0                              0                              0           
#>                                           4                              0                              0                              0           
#>                                           5                              0                              0                              0           
#> 
#> dcd A.1.1.1.1                       - Any Grade -                   21 (15.7%)                     20 (14.9%)                     23 (17.4%)       
#>                                           1                         21 (15.7%)                     20 (14.9%)                     23 (17.4%)       
#>                                           2                              0                              0                              0           
#>                                           3                              0                              0                              0           
#>                                           4                              0                              0                              0           
#>                                           5                              0                              0                              0           
#> 
#> dcd D.2.1.5.3                       - Any Grade -                   18 (13.4%)                     24 (17.9%)                     22 (16.7%)       
#>                                           1                         18 (13.4%)                     24 (17.9%)                     22 (16.7%)       
#>                                           2                              0                              0                              0           
#>                                           3                              0                              0                              0           
#>                                           4                              0                              0                              0           
#>                                           5                              0                              0                              0           

# Introducing imperfect data and demo how to preprocess
ADAE$AEDECOD <- as.character(ADAE$AEDECOD)
ADAE$AEBODSYS <- as.character(ADAE$AEBODSYS)
ADAE$AEDECOD[c(1,5)] = ""
ADAE$AEBODSYS[c(2,6)] = " "

ADAE <- ADAE %>%
  mutate(
   grade_category = case_when(
     AETOXGR %in% c(1,2) ~ "Mild",
     AETOXGR %in% c(3,4) ~ "Moderate",
     AETOXGR %in% c(5) ~ "Severe"
   ),
   grade_category = factor(grade_category, levels=c("Mild", "Moderate", "Severe")),
   # Preprocess imperfect data
   AEDECOD = AEDECOD %>% sas_na %>% explicit_na(label = "Not Coded"),
   AEBODSYS = AEBODSYS %>% sas_na %>% explicit_na(label = "Not Coded"),
  ) %>%
  var_relabel(
    AEBODSYS = "MedDRA System Organ Class",
    AEDECOD = "MedDRA Preferred Term",
    AETOXGR = "Analysis Toxicity Grade",
    grade_category = "Insensity")

t_events_per_term_grade_id(
  terms = ADAE$AEDECOD,
  id = ADAE$USUBJID,
  grade = ADAE$AETOXGR,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients"
)

#> MedDRA Preferred Term                                              A: Drug X                     B: Placebo                   C: Combination                  All Patients      
#>                              Analysis Toxicity Grade                (N=134)                        (N=134)                        (N=132)                        (N=400)        
#> --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> - Overall -                       - Any Grade -                   110 (82.1%)                    98 (73.1%)                     108 (81.8%)                     316 (79%)       
#>                                         1                          6 (4.5%)                       11 (8.2%)                      13 (9.8%)                      30 (7.5%)       
#>                                         2                         14 (10.4%)                      11 (8.2%)                       4 (3%)                        29 (7.2%)       
#>                                         3                         31 (23.1%)                     20 (14.9%)                     26 (19.7%)                     77 (19.2%)       
#>                                         4                         18 (13.4%)                     14 (10.4%)                     24 (18.2%)                      56 (14%)        
#>                                         5                         41 (30.6%)                     42 (31.3%)                     41 (31.1%)                      124 (31%)       
#> 
#> dcd B.2.1.2.1                     - Any Grade -                   31 (23.1%)                     25 (18.7%)                     25 (18.9%)                     81 (20.2%)       
#>                                         3                         31 (23.1%)                     25 (18.7%)                     25 (18.9%)                     81 (20.2%)       
#> 
#> dcd C.1.1.1.3                     - Any Grade -                   27 (20.1%)                     23 (17.2%)                     30 (22.7%)                      80 (20%)        
#>                                         4                         27 (20.1%)                     23 (17.2%)                     30 (22.7%)                      80 (20%)        
#> 
#> dcd D.1.1.4.2                     - Any Grade -                   24 (17.9%)                     22 (16.4%)                     31 (23.5%)                     77 (19.2%)       
#>                                         3                         24 (17.9%)                     22 (16.4%)                     31 (23.5%)                     77 (19.2%)       
#> 
#> dcd B.2.2.3.1                     - Any Grade -                   21 (15.7%)                     25 (18.7%)                     26 (19.7%)                      72 (18%)        
#>                                         1                         21 (15.7%)                     25 (18.7%)                     26 (19.7%)                      72 (18%)        
#> 
#> dcd C.2.1.2.1                     - Any Grade -                   20 (14.9%)                     28 (20.9%)                     23 (17.4%)                     71 (17.8%)       
#>                                         2                         20 (14.9%)                     28 (20.9%)                     23 (17.4%)                     71 (17.8%)       
#> 
#> dcd D.1.1.1.1                     - Any Grade -                   22 (16.4%)                     23 (17.2%)                     25 (18.9%)                     70 (17.5%)       
#>                                         5                         22 (16.4%)                     23 (17.2%)                     25 (18.9%)                     70 (17.5%)       
#> 
#> dcd B.1.1.1.1                     - Any Grade -                   24 (17.9%)                     23 (17.2%)                     21 (15.9%)                      68 (17%)        
#>                                         5                         24 (17.9%)                     23 (17.2%)                     21 (15.9%)                      68 (17%)        
#> 
#> dcd A.1.1.1.2                     - Any Grade -                   25 (18.7%)                     20 (14.9%)                     21 (15.9%)                     66 (16.5%)       
#>                                         2                         25 (18.7%)                     20 (14.9%)                     21 (15.9%)                     66 (16.5%)       
#> 
#> dcd A.1.1.1.1                     - Any Grade -                   21 (15.7%)                     20 (14.9%)                     23 (17.4%)                      64 (16%)        
#>                                         1                         21 (15.7%)                     20 (14.9%)                     23 (17.4%)                      64 (16%)        
#> 
#> dcd D.2.1.5.3                     - Any Grade -                   18 (13.4%)                     24 (17.9%)                     22 (16.7%)                      64 (16%)        
#>                                         1                         18 (13.4%)                     24 (17.9%)                     22 (16.7%)                      64 (16%)        
#> 
#> Not Coded                         - Any Grade -                    1 (0.7%)                       1 (0.7%)                           0                          2 (0.5%)        
#>                                         3                          1 (0.7%)                           0                              0                          1 (0.2%)        
#>                                         4                              0                          1 (0.7%)                           0                          1 (0.2%)        

t_events_per_term_grade_id(
  terms = ADAE$AEDECOD,
  id = ADAE$USUBJID,
  grade = ADAE$grade_category,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients"
)

#> MedDRA Preferred Term                                A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                Insensity              (N=134)               (N=134)               (N=132)               (N=400)    
#> -----------------------------------------------------------------------------------------------------------------------------------
#> - Overall -                  - Any Grade -          110 (82.1%)            98 (73.1%)           108 (81.8%)            316 (79%)   
#>                                   Mild               20 (14.9%)            22 (16.4%)            17 (12.9%)            59 (14.8%)  
#>                                 Moderate             49 (36.6%)            34 (25.4%)            50 (37.9%)           133 (33.2%)  
#>                                  Severe              41 (30.6%)            42 (31.3%)            41 (31.1%)            124 (31%)   
#> 
#> dcd B.2.1.2.1                - Any Grade -           31 (23.1%)            25 (18.7%)            25 (18.9%)            81 (20.2%)  
#>                                 Moderate             31 (23.1%)            25 (18.7%)            25 (18.9%)            81 (20.2%)  
#> 
#> dcd C.1.1.1.3                - Any Grade -           27 (20.1%)            23 (17.2%)            30 (22.7%)             80 (20%)   
#>                                 Moderate             27 (20.1%)            23 (17.2%)            30 (22.7%)             80 (20%)   
#> 
#> dcd D.1.1.4.2                - Any Grade -           24 (17.9%)            22 (16.4%)            31 (23.5%)            77 (19.2%)  
#>                                 Moderate             24 (17.9%)            22 (16.4%)            31 (23.5%)            77 (19.2%)  
#> 
#> dcd B.2.2.3.1                - Any Grade -           21 (15.7%)            25 (18.7%)            26 (19.7%)             72 (18%)   
#>                                   Mild               21 (15.7%)            25 (18.7%)            26 (19.7%)             72 (18%)   
#> 
#> dcd C.2.1.2.1                - Any Grade -           20 (14.9%)            28 (20.9%)            23 (17.4%)            71 (17.8%)  
#>                                   Mild               20 (14.9%)            28 (20.9%)            23 (17.4%)            71 (17.8%)  
#> 
#> dcd D.1.1.1.1                - Any Grade -           22 (16.4%)            23 (17.2%)            25 (18.9%)            70 (17.5%)  
#>                                  Severe              22 (16.4%)            23 (17.2%)            25 (18.9%)            70 (17.5%)  
#> 
#> dcd B.1.1.1.1                - Any Grade -           24 (17.9%)            23 (17.2%)            21 (15.9%)             68 (17%)   
#>                                  Severe              24 (17.9%)            23 (17.2%)            21 (15.9%)             68 (17%)   
#> 
#> dcd A.1.1.1.2                - Any Grade -           25 (18.7%)            20 (14.9%)            21 (15.9%)            66 (16.5%)  
#>                                   Mild               25 (18.7%)            20 (14.9%)            21 (15.9%)            66 (16.5%)  
#> 
#> dcd A.1.1.1.1                - Any Grade -           21 (15.7%)            20 (14.9%)            23 (17.4%)             64 (16%)   
#>                                   Mild               21 (15.7%)            20 (14.9%)            23 (17.4%)             64 (16%)   
#> 
#> dcd D.2.1.5.3                - Any Grade -           18 (13.4%)            24 (17.9%)            22 (16.7%)             64 (16%)   
#>                                   Mild               18 (13.4%)            24 (17.9%)            22 (16.7%)             64 (16%)   
#> 
#> Not Coded                    - Any Grade -            1 (0.7%)              1 (0.7%)                 0                  2 (0.5%)   
#>                                 Moderate              1 (0.7%)              1 (0.7%)                 0                  2 (0.5%)   

t_events_per_term_grade_id(
  terms = ADAE %>% select(AEBODSYS, AEDECOD),
  id = ADAE$USUBJID,
  grade = ADAE$AETOXGR,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  event_type = "adverse events"
)

#> MedDRA System Organ Class                                              A: Drug X                     B: Placebo                   C: Combination     
#>   MedDRA Preferred Term          Analysis Toxicity Grade                (N=134)                        (N=134)                        (N=132)        
#> -----------------------------------------------------------------------------------------------------------------------------------------------------
#> - Any adverse events -
#>   - Overall -                         - Any Grade -                   110 (82.1%)                    98 (73.1%)                     108 (81.8%)      
#>                                             1                          6 (4.5%)                       11 (8.2%)                      13 (9.8%)       
#>                                             2                         14 (10.4%)                      11 (8.2%)                       4 (3%)         
#>                                             3                         31 (23.1%)                     20 (14.9%)                     26 (19.7%)       
#>                                             4                         18 (13.4%)                     14 (10.4%)                     24 (18.2%)       
#>                                             5                         41 (30.6%)                     42 (31.3%)                     41 (31.1%)       
#> 
#> cl B.2
#>   - Overall -                         - Any Grade -                   49 (36.6%)                     44 (32.8%)                     48 (36.4%)       
#>                                             1                         17 (12.7%)                     19 (14.2%)                     23 (17.4%)       
#>                                             3                         32 (23.9%)                     25 (18.7%)                     25 (18.9%)       
#> 
#>   dcd B.2.1.2.1                       - Any Grade -                   31 (23.1%)                     25 (18.7%)                     25 (18.9%)       
#>                                             3                         31 (23.1%)                     25 (18.7%)                     25 (18.9%)       
#> 
#>   dcd B.2.2.3.1                       - Any Grade -                   21 (15.7%)                     25 (18.7%)                     26 (19.7%)       
#>                                             1                         21 (15.7%)                     25 (18.7%)                     26 (19.7%)       
#> 
#>   Not Coded                           - Any Grade -                    1 (0.7%)                           0                              0           
#>                                             3                          1 (0.7%)                           0                              0           
#> 
#> cl D.1
#>   - Overall -                         - Any Grade -                   42 (31.3%)                     39 (29.1%)                     53 (40.2%)       
#>                                             3                         21 (15.7%)                     16 (11.9%)                     28 (21.2%)       
#>                                             5                         21 (15.7%)                     23 (17.2%)                     25 (18.9%)       
#> 
#>   dcd D.1.1.4.2                       - Any Grade -                   24 (17.9%)                     22 (16.4%)                     31 (23.5%)       
#>                                             3                         24 (17.9%)                     22 (16.4%)                     31 (23.5%)       
#> 
#>   dcd D.1.1.1.1                       - Any Grade -                   21 (15.7%)                     23 (17.2%)                     25 (18.9%)       
#>                                             5                         21 (15.7%)                     23 (17.2%)                     25 (18.9%)       
#> 
#> cl A.1
#>   - Overall -                         - Any Grade -                   42 (31.3%)                     39 (29.1%)                     39 (29.5%)       
#>                                             1                         17 (12.7%)                     19 (14.2%)                     18 (13.6%)       
#>                                             2                         25 (18.7%)                     20 (14.9%)                     21 (15.9%)       
#> 
#>   dcd A.1.1.1.2                       - Any Grade -                   25 (18.7%)                     20 (14.9%)                     21 (15.9%)       
#>                                             2                         25 (18.7%)                     20 (14.9%)                     21 (15.9%)       
#> 
#>   dcd A.1.1.1.1                       - Any Grade -                   21 (15.7%)                     20 (14.9%)                     23 (17.4%)       
#>                                             1                         21 (15.7%)                     20 (14.9%)                     23 (17.4%)       
#> 
#> cl C.1
#>   - Overall -                         - Any Grade -                   27 (20.1%)                     23 (17.2%)                     30 (22.7%)       
#>                                             4                         27 (20.1%)                     23 (17.2%)                     30 (22.7%)       
#> 
#>   dcd C.1.1.1.3                       - Any Grade -                   27 (20.1%)                     23 (17.2%)                     30 (22.7%)       
#>                                             4                         27 (20.1%)                     23 (17.2%)                     30 (22.7%)       
#> 
#>   Not Coded                           - Any Grade -                        0                          1 (0.7%)                           0           
#>                                             4                              0                          1 (0.7%)                           0           
#> 
#> cl C.2
#>   - Overall -                         - Any Grade -                   20 (14.9%)                     28 (20.9%)                     23 (17.4%)       
#>                                             2                         20 (14.9%)                     28 (20.9%)                     23 (17.4%)       
#> 
#>   dcd C.2.1.2.1                       - Any Grade -                   20 (14.9%)                     28 (20.9%)                     23 (17.4%)       
#>                                             2                         20 (14.9%)                     28 (20.9%)                     23 (17.4%)       
#> 
#> cl B.1
#>   - Overall -                         - Any Grade -                   24 (17.9%)                     22 (16.4%)                     21 (15.9%)       
#>                                             5                         24 (17.9%)                     22 (16.4%)                     21 (15.9%)       
#> 
#>   dcd B.1.1.1.1                       - Any Grade -                   24 (17.9%)                     22 (16.4%)                     21 (15.9%)       
#>                                             5                         24 (17.9%)                     22 (16.4%)                     21 (15.9%)       
#> 
#> cl D.2
#>   - Overall -                         - Any Grade -                   18 (13.4%)                     24 (17.9%)                     22 (16.7%)       
#>                                             1                         18 (13.4%)                     24 (17.9%)                     22 (16.7%)       
#> 
#>   dcd D.2.1.5.3                       - Any Grade -                   18 (13.4%)                     24 (17.9%)                     22 (16.7%)       
#>                                             1                         18 (13.4%)                     24 (17.9%)                     22 (16.7%)       
#> 
#> Not Coded
#>   - Overall -                         - Any Grade -                    1 (0.7%)                       1 (0.7%)                           0           
#>                                             5                          1 (0.7%)                       1 (0.7%)                           0           
#> 
#>   dcd B.1.1.1.1                       - Any Grade -                        0                          1 (0.7%)                           0           
#>                                             5                              0                          1 (0.7%)                           0           
#> 
#>   dcd D.1.1.1.1                       - Any Grade -                    1 (0.7%)                           0                              0           
#>                                             5                          1 (0.7%)                           0                              0           

levels(ADAE$grade_category) <- c("Severe", "Moderate", "Mild")
tbls <- t_events_per_term_grade_id(
  terms = ADAE %>% select(AEBODSYS, AEDECOD),
  id = ADAE$USUBJID,
  grade = ADAE$grade_category,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients",
  event_type = "adverse events",
  any_grade = "Any Severity",
  table_tree = TRUE
)
summary(tbls)

#>                                 index            content             dim            hide label          format  
#> ----------------------------------------------------------------------------------------------------------------
#> - Any adverse events -                            rtable            4 x 5             FALSE               0     
#>   cl B.2                          1               rtable            3 x 5             FALSE                     
#>     dcd B.2.1.2.1                1, 1             rtable            2 x 5              TRUE                     
#>     dcd B.2.2.3.1                1, 2             rtable            2 x 5              TRUE                     
#>     Not Coded                    1, 3             rtable            2 x 5              TRUE                     
#>   cl D.1                          2               rtable            3 x 5             FALSE                     
#>     dcd D.1.1.4.2                2, 1             rtable            2 x 5              TRUE                     
#>     dcd D.1.1.1.1                2, 2             rtable            2 x 5              TRUE                     
#>   cl A.1                          3               rtable            2 x 5             FALSE                     
#>     dcd A.1.1.1.2                3, 1             rtable            2 x 5              TRUE                     
#>     dcd A.1.1.1.1                3, 2             rtable            2 x 5              TRUE                     
#>   cl C.1                          4               rtable            2 x 5             FALSE                     
#>     dcd C.1.1.1.3                4, 1             rtable            2 x 5              TRUE                     
#>     Not Coded                    4, 2             rtable            2 x 5              TRUE                     
#>   cl C.2                          5               rtable            2 x 5             FALSE                     
#>     dcd C.2.1.2.1                5, 1             rtable            2 x 5              TRUE                     
#>   cl B.1                          6               rtable            2 x 5             FALSE                     
#>     dcd B.1.1.1.1                6, 1             rtable            2 x 5              TRUE                     
#>   cl D.2                          7               rtable            2 x 5             FALSE                     
#>     dcd D.2.1.5.3                7, 1             rtable            2 x 5              TRUE                     
#>   Not Coded                       8               rtable            2 x 5             FALSE                     
#>     dcd B.1.1.1.1                8, 1             rtable            2 x 5              TRUE                     
#>     dcd D.1.1.1.1                8, 2             rtable            2 x 5              TRUE                     

to_rtable(tbls)

#> MedDRA System Organ Class                                   A: Drug X               B: Placebo            C: Combination           All Patients  
#>   MedDRA Preferred Term             Insensity                (N=134)                 (N=134)                 (N=132)                 (N=400)     
#> -------------------------------------------------------------------------------------------------------------------------------------------------
#> - Any adverse events -
#>   - Overall -                    - Any Severity -          110 (82.1%)              98 (73.1%)             108 (81.8%)              316 (79%)    
#>                                       Severe                20 (14.9%)              22 (16.4%)              17 (12.9%)              59 (14.8%)   
#>                                      Moderate               49 (36.6%)              34 (25.4%)              50 (37.9%)             133 (33.2%)   
#>                                        Mild                 41 (30.6%)              42 (31.3%)              41 (31.1%)              124 (31%)    
#> 
#> cl B.2
#>   - Overall -                    - Any Severity -           49 (36.6%)              44 (32.8%)              48 (36.4%)             141 (35.2%)   
#>                                       Severe                17 (12.7%)              19 (14.2%)              23 (17.4%)              59 (14.8%)   
#>                                      Moderate               32 (23.9%)              25 (18.7%)              25 (18.9%)              82 (20.5%)   
#> 
#>   dcd B.2.1.2.1                  - Any Severity -           31 (23.1%)              25 (18.7%)              25 (18.9%)              81 (20.2%)   
#>                                      Moderate               31 (23.1%)              25 (18.7%)              25 (18.9%)              81 (20.2%)   
#> 
#>   dcd B.2.2.3.1                  - Any Severity -           21 (15.7%)              25 (18.7%)              26 (19.7%)               72 (18%)    
#>                                       Severe                21 (15.7%)              25 (18.7%)              26 (19.7%)               72 (18%)    
#> 
#>   Not Coded                      - Any Severity -            1 (0.7%)                   0                       0                    1 (0.2%)    
#>                                      Moderate                1 (0.7%)                   0                       0                    1 (0.2%)    
#> 
#> cl D.1
#>   - Overall -                    - Any Severity -           42 (31.3%)              39 (29.1%)              53 (40.2%)             134 (33.5%)   
#>                                      Moderate               21 (15.7%)              16 (11.9%)              28 (21.2%)              65 (16.2%)   
#>                                        Mild                 21 (15.7%)              23 (17.2%)              25 (18.9%)              69 (17.2%)   
#> 
#>   dcd D.1.1.4.2                  - Any Severity -           24 (17.9%)              22 (16.4%)              31 (23.5%)              77 (19.2%)   
#>                                      Moderate               24 (17.9%)              22 (16.4%)              31 (23.5%)              77 (19.2%)   
#> 
#>   dcd D.1.1.1.1                  - Any Severity -           21 (15.7%)              23 (17.2%)              25 (18.9%)              69 (17.2%)   
#>                                        Mild                 21 (15.7%)              23 (17.2%)              25 (18.9%)              69 (17.2%)   
#> 
#> cl A.1
#>   - Overall -                    - Any Severity -           42 (31.3%)              39 (29.1%)              39 (29.5%)              120 (30%)    
#>                                       Severe                42 (31.3%)              39 (29.1%)              39 (29.5%)              120 (30%)    
#> 
#>   dcd A.1.1.1.2                  - Any Severity -           25 (18.7%)              20 (14.9%)              21 (15.9%)              66 (16.5%)   
#>                                       Severe                25 (18.7%)              20 (14.9%)              21 (15.9%)              66 (16.5%)   
#> 
#>   dcd A.1.1.1.1                  - Any Severity -           21 (15.7%)              20 (14.9%)              23 (17.4%)               64 (16%)    
#>                                       Severe                21 (15.7%)              20 (14.9%)              23 (17.4%)               64 (16%)    
#> 
#> cl C.1
#>   - Overall -                    - Any Severity -           27 (20.1%)              23 (17.2%)              30 (22.7%)               80 (20%)    
#>                                      Moderate               27 (20.1%)              23 (17.2%)              30 (22.7%)               80 (20%)    
#> 
#>   dcd C.1.1.1.3                  - Any Severity -           27 (20.1%)              23 (17.2%)              30 (22.7%)               80 (20%)    
#>                                      Moderate               27 (20.1%)              23 (17.2%)              30 (22.7%)               80 (20%)    
#> 
#>   Not Coded                      - Any Severity -               0                    1 (0.7%)                   0                    1 (0.2%)    
#>                                      Moderate                   0                    1 (0.7%)                   0                    1 (0.2%)    
#> 
#> cl C.2
#>   - Overall -                    - Any Severity -           20 (14.9%)              28 (20.9%)              23 (17.4%)              71 (17.8%)   
#>                                       Severe                20 (14.9%)              28 (20.9%)              23 (17.4%)              71 (17.8%)   
#> 
#>   dcd C.2.1.2.1                  - Any Severity -           20 (14.9%)              28 (20.9%)              23 (17.4%)              71 (17.8%)   
#>                                       Severe                20 (14.9%)              28 (20.9%)              23 (17.4%)              71 (17.8%)   
#> 
#> cl B.1
#>   - Overall -                    - Any Severity -           24 (17.9%)              22 (16.4%)              21 (15.9%)              67 (16.8%)   
#>                                        Mild                 24 (17.9%)              22 (16.4%)              21 (15.9%)              67 (16.8%)   
#> 
#>   dcd B.1.1.1.1                  - Any Severity -           24 (17.9%)              22 (16.4%)              21 (15.9%)              67 (16.8%)   
#>                                        Mild                 24 (17.9%)              22 (16.4%)              21 (15.9%)              67 (16.8%)   
#> 
#> cl D.2
#>   - Overall -                    - Any Severity -           18 (13.4%)              24 (17.9%)              22 (16.7%)               64 (16%)    
#>                                       Severe                18 (13.4%)              24 (17.9%)              22 (16.7%)               64 (16%)    
#> 
#>   dcd D.2.1.5.3                  - Any Severity -           18 (13.4%)              24 (17.9%)              22 (16.7%)               64 (16%)    
#>                                       Severe                18 (13.4%)              24 (17.9%)              22 (16.7%)               64 (16%)    
#> 
#> Not Coded
#>   - Overall -                    - Any Severity -            1 (0.7%)                1 (0.7%)                   0                    2 (0.5%)    
#>                                        Mild                  1 (0.7%)                1 (0.7%)                   0                    2 (0.5%)    
#> 
#>   dcd B.1.1.1.1                  - Any Severity -               0                    1 (0.7%)                   0                    1 (0.2%)    
#>                                        Mild                     0                    1 (0.7%)                   0                    1 (0.2%)    
#> 
#>   dcd D.1.1.1.1                  - Any Severity -            1 (0.7%)                   0                       0                    1 (0.2%)    
#>                                        Mild                  1 (0.7%)                   0                       0                    1 (0.2%)