Basic Events Table

This function summarizes number of unique subjects with events and total number of events. It creates basic summary of events and can be used for any events data like Adverse Events, concomitant medication, medical history, etc.

t_events_per_term_id(
  terms,
  id,
  col_by,
  col_N = NULL,
  total = NULL,
  event_type = "event",
  table_tree = FALSE
)

Arguments

terms
(character or factor vector, or data.frame)
Represents events information. Multi-level nesting is possible when terms is a data.frame and columns should be ordered with the first column specifying the highest level term and the last column the lowest level term.

id
(vector)
contains subject identifier. Length of id must be the same as the length or number of rows of terms.

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

col_N
(numeric vector)
Contains information of the number of patients in the levels of col_by. This is useful if there are patients that have no adverse events can be accounted for with this argument.

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

event_type
(character value)
Specifies type of event that is summarized (e.g. adverse event, treatment). Default is "event".

table_tree
(logical value)
If FALSE then a single rtable gets returned, if TRUE a node object gets returned with the elementary rtable objects.

Value

an rtable object.

Details

Implementation examples are to apply t_events_per_term_id on Adverse Event data to create Adverse Events summary table (AET02, STREAM2.x, STREAM1.17 ), or apply on Concomitant Medication data to create Concomitant Treatment summary table (CMT01, STREAM2.x, STREAM1.17).

t_events_per_term_id includes percentages based on the total number of subjects in the column heading (i.e. "N=nnn"). col_N can be explicitly specified to get N for percentage calculation from either events dataset or additional dataset such as subject level dataset. See the example.

Multiple occurrences of the same event in an individual id are counted only once when counting number of subjects.

t_events_per_term_id doesn't deal with data with any non-complete records (has NA), e.g. if any terms are missing. Impute missing values before using t_events_per_term_id.

t_events_per_term_id orders data by overall frequency across all columns from the most commonly reported higher level term to the least frequent one. Within each group of higher level term, it sorts by decreasing frequency of lower level term. It brakes ties using terms names in alphabetical order.

See also

Examples

library(dplyr)

t_events_per_term_id(
 terms = with_label(factor(c("t1", "t1", "t2", "t2", "t2")), "Term"),
 id = c(1, 4, 2, 3, 3),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10),
 total = "All Patients"
)

#>                                                              A                   B                   C              All Patients
#>                                                            (N=2)               (N=4)               (N=10)              (N=16)   
#> --------------------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one event          2 (100%)            1 (25%)             1 (10%)             4 (25%)   
#> Total number of events                                       2                   1                   2                   5      
#> t1                                                        2 (100%)               0                   0               2 (12.5%)  
#> t2                                                           0                1 (25%)             1 (10%)            2 (12.5%)  

if (FALSE) {
# terms can't be empty strings or NA
t_events_per_term_id(
 terms = with_label(factor(c("", "t1", "t2", "t2", "t2")), "Term"),
 id = c(1, 4, 2, 3, 3),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10),
 total = "All Patients"
)

t_events_per_term_id(
 terms = with_label(factor(c(NA, "t1", "t2", "t2", "t2")), "Term"),
 id = c(1, 4, 2, 3, 3),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10),
 total = "All Patients"
)

}

# fill in "" with explicit NA
t_events_per_term_id(
 terms = explicit_na(sas_na(factor(c("", "", "t2", "t1", "t2")))),
 id = c(1, 4, 2, 3, 3),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10),
 total = "All Patients"
)

#>                                                              A                   B                   C              All Patients
#>                                                            (N=2)               (N=4)               (N=10)              (N=16)   
#> --------------------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one event          2 (100%)            1 (25%)             1 (10%)             4 (25%)   
#> Total number of events                                       2                   1                   2                   5      
#> t2                                                           0                1 (25%)             1 (10%)            2 (12.5%)  
#> <Missing>                                                 2 (100%)               0                   0               2 (12.5%)  
#> t1                                                           0                   0                1 (10%)            1 (6.25%)  

library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADAE <- radae(cached = TRUE)

t_events_per_term_id(
  terms = ADAE$AEDECOD,
  id = ADAE$USUBJID,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  event_type = "adverse event"
)

#>                                                                   A: Drug X             B: Placebo          C: Combination
#>                                                                    (N=134)               (N=134)               (N=132)    
#> --------------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event         121 (90.3%)           127 (94.78%)          120 (90.91%) 
#> Total number of adverse events                                       709                   745                   685      
#> dcd D.1.1.4.2                                                    55 (41.04%)           63 (47.01%)           53 (40.15%)  
#> dcd D.2.1.5.3                                                    49 (36.57%)           56 (41.79%)           62 (46.97%)  
#> dcd B.1.1.1.1                                                    51 (38.06%)           53 (39.55%)           60 (45.45%)  
#> dcd C.2.1.2.1                                                    58 (43.28%)           56 (41.79%)           48 (36.36%)  
#> dcd D.1.1.1.1                                                    53 (39.55%)           58 (43.28%)           48 (36.36%)  
#> dcd A.1.1.1.2                                                    53 (39.55%)           57 (42.54%)           48 (36.36%)  
#> dcd B.2.1.2.1                                                    46 (34.33%)           61 (45.52%)           47 (35.61%)  
#> dcd C.1.1.1.3                                                    56 (41.79%)           46 (34.33%)           50 (37.88%)  
#> dcd B.2.2.3.1                                                     54 (40.3%)           48 (35.82%)           49 (37.12%)  
#> dcd A.1.1.1.1                                                    40 (29.85%)           52 (38.81%)           46 (34.85%)  

t_events_per_term_id(
  terms = ADAE$AEDECOD,
  id = ADAE$USUBJID,
  col_by = ADAE$ARM,
  total = "All Patients",
  col_N = table(ADSL$ARM),
  event_type = "adverse event"
)

#>                                                                   A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                                                    (N=134)               (N=134)               (N=132)               (N=400)    
#> ------------------------------------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event         121 (90.3%)           127 (94.78%)          120 (90.91%)           368 (92%)   
#> Total number of adverse events                                       709                   745                   685                   2139     
#> dcd D.1.1.4.2                                                    55 (41.04%)           63 (47.01%)           53 (40.15%)           171 (42.75%) 
#> dcd D.2.1.5.3                                                    49 (36.57%)           56 (41.79%)           62 (46.97%)           167 (41.75%) 
#> dcd B.1.1.1.1                                                    51 (38.06%)           53 (39.55%)           60 (45.45%)            164 (41%)   
#> dcd C.2.1.2.1                                                    58 (43.28%)           56 (41.79%)           48 (36.36%)           162 (40.5%)  
#> dcd D.1.1.1.1                                                    53 (39.55%)           58 (43.28%)           48 (36.36%)           159 (39.75%) 
#> dcd A.1.1.1.2                                                    53 (39.55%)           57 (42.54%)           48 (36.36%)           158 (39.5%)  
#> dcd B.2.1.2.1                                                    46 (34.33%)           61 (45.52%)           47 (35.61%)           154 (38.5%)  
#> dcd C.1.1.1.3                                                    56 (41.79%)           46 (34.33%)           50 (37.88%)            152 (38%)   
#> dcd B.2.2.3.1                                                     54 (40.3%)           48 (35.82%)           49 (37.12%)           151 (37.75%) 
#> dcd A.1.1.1.1                                                    40 (29.85%)           52 (38.81%)           46 (34.85%)           138 (34.5%)  

t_events_per_term_id(
  terms = ADAE[, c("AEBODSYS", "AEDECOD")],
  id = ADAE$USUBJID,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients"
)

#>                                                             A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                                              (N=134)               (N=134)               (N=132)               (N=400)    
#> ------------------------------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one event           121 (90.3%)           127 (94.78%)          120 (90.91%)           368 (92%)   
#> Total number of events                                         709                   745                   685                   2139     
#> 
#> cl D.1
#>   Total number of patients with at least one event         84 (62.69%)           92 (68.66%)           81 (61.36%)           257 (64.25%) 
#>   Total number of events                                       147                   153                   128                   428      
#>   dcd D.1.1.4.2                                            55 (41.04%)           63 (47.01%)           53 (40.15%)           171 (42.75%) 
#>   dcd D.1.1.1.1                                            53 (39.55%)           58 (43.28%)           48 (36.36%)           159 (39.75%) 
#> 
#> cl A.1
#>   Total number of patients with at least one event         75 (55.97%)           86 (64.18%)            73 (55.3%)           234 (58.5%)  
#>   Total number of events                                       127                   146                   124                   397      
#>   dcd A.1.1.1.2                                            53 (39.55%)           57 (42.54%)           48 (36.36%)           158 (39.5%)  
#>   dcd A.1.1.1.1                                            40 (29.85%)           52 (38.81%)           46 (34.85%)           138 (34.5%)  
#> 
#> cl B.2
#>   Total number of patients with at least one event         73 (54.48%)           83 (61.94%)           76 (57.58%)            232 (58%)   
#>   Total number of events                                       135                   152                   147                   434      
#>   dcd B.2.1.2.1                                            46 (34.33%)           61 (45.52%)           47 (35.61%)           154 (38.5%)  
#>   dcd B.2.2.3.1                                             54 (40.3%)           48 (35.82%)           49 (37.12%)           151 (37.75%) 
#> 
#> cl D.2
#>   Total number of patients with at least one event         49 (36.57%)           56 (41.79%)           62 (46.97%)           167 (41.75%) 
#>   Total number of events                                        66                    84                    81                   231      
#>   dcd D.2.1.5.3                                            49 (36.57%)           56 (41.79%)           62 (46.97%)           167 (41.75%) 
#> 
#> cl B.1
#>   Total number of patients with at least one event         51 (38.06%)           53 (39.55%)           60 (45.45%)            164 (41%)   
#>   Total number of events                                        64                    75                    78                   217      
#>   dcd B.1.1.1.1                                            51 (38.06%)           53 (39.55%)           60 (45.45%)            164 (41%)   
#> 
#> cl C.2
#>   Total number of patients with at least one event         58 (43.28%)           56 (41.79%)           48 (36.36%)           162 (40.5%)  
#>   Total number of events                                        91                    73                    61                   225      
#>   dcd C.2.1.2.1                                            58 (43.28%)           56 (41.79%)           48 (36.36%)           162 (40.5%)  
#> 
#> cl C.1
#>   Total number of patients with at least one event         56 (41.79%)           46 (34.33%)           50 (37.88%)            152 (38%)   
#>   Total number of events                                        79                    62                    66                   207      
#>   dcd C.1.1.1.3                                            56 (41.79%)           46 (34.33%)           50 (37.88%)            152 (38%)   

ADSL <- radsl(cached = TRUE)
ADCM <- radcm(cached = TRUE) %>%
  dplyr::filter(ATIREL == "CONCOMITANT")

t_events_per_term_id(
  terms = ADCM[, c("CMCLAS", "CMDECOD")],
  id = ADCM$USUBJID,
  col_by = ADCM$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients",
  event_type = "treatment"
)

#>                                                                 A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                                                  (N=134)               (N=134)               (N=132)               (N=400)    
#> ----------------------------------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one treatment           117 (87.31%)          116 (86.57%)          116 (87.88%)          349 (87.25%) 
#> Total number of treatments                                         415                   414                   460                   1289     
#> 
#> medcl C
#>   Total number of patients with at least one treatment         82 (61.19%)           84 (62.69%)           89 (67.42%)           255 (63.75%) 
#>   Total number of treatments                                       140                   140                   155                   435      
#>   medname C_2/2                                                52 (38.81%)           58 (43.28%)           60 (45.45%)           170 (42.5%)  
#>   medname C_1/2                                                51 (38.06%)           50 (37.31%)           56 (42.42%)           157 (39.25%) 
#> 
#> medcl B
#>   Total number of patients with at least one treatment         83 (61.94%)           74 (55.22%)           88 (66.67%)           245 (61.25%) 
#>   Total number of treatments                                       141                   137                   162                   440      
#>   medname B_1/4                                                52 (38.81%)           57 (42.54%)            59 (44.7%)            168 (42%)   
#>   medname B_4/4                                                50 (37.31%)           45 (33.58%)           55 (41.67%)           150 (37.5%)  
#> 
#> medcl A
#>   Total number of patients with at least one treatment         75 (55.97%)           79 (58.96%)           81 (61.36%)           235 (58.75%) 
#>   Total number of treatments                                       134                   137                   143                   414      
#>   medname A_2/3                                                53 (39.55%)           50 (37.31%)           56 (42.42%)           159 (39.75%) 
#>   medname A_3/3                                                45 (33.58%)            54 (40.3%)           48 (36.36%)           147 (36.75%) 

# Table Tree

summary(t_events_per_term_id(
  terms = ADAE$AEDECOD,
  id = ADAE$USUBJID,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients",
  table_tree = TRUE
))

#>                          index            content             dim            hide label          format  
#> ---------------------------------------------------------------------------------------------------------
#> All                                        rtable            2 x 4              TRUE              0, 0   
#>   dcd D.1.1.4.2            1               rtable            1 x 4              TRUE                     
#>   dcd D.2.1.5.3            2               rtable            1 x 4              TRUE                     
#>   dcd B.1.1.1.1            3               rtable            1 x 4              TRUE                     
#>   dcd C.2.1.2.1            4               rtable            1 x 4              TRUE                     
#>   dcd D.1.1.1.1            5               rtable            1 x 4              TRUE                     
#>   dcd A.1.1.1.2            6               rtable            1 x 4              TRUE                     
#>   dcd B.2.1.2.1            7               rtable            1 x 4              TRUE                     
#>   dcd C.1.1.1.3            8               rtable            1 x 4              TRUE                     
#>   dcd B.2.2.3.1            9               rtable            1 x 4              TRUE                     
#>   dcd A.1.1.1.1            10              rtable            1 x 4              TRUE                     

tbls <- t_events_per_term_id(
 terms = with_label(factor(c("t1", "t1", "t2", "t2", "t2")), "Term"),
 id = c(1, 4, 2, 3, 3),
 col_by = factor(c("A", "A", "B", "C", "C")),
 col_N = c(2, 4, 10),
 table_tree = TRUE
)
summary(tbls)

#>               index            content             dim            hide label          format  
#> ----------------------------------------------------------------------------------------------
#> All                             rtable            2 x 3              TRUE              0, 0   
#>   t1            1               rtable            1 x 3              TRUE                     
#>   t2            2               rtable            1 x 3              TRUE                     

tbls <- t_events_per_term_id(
  terms = ADAE[, c("AEBODSYS", "AEDECOD")],
  id = ADAE$USUBJID,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients",
  table_tree = TRUE
)
summary(tbls)

#>                            index            content             dim            hide label          format  
#> -----------------------------------------------------------------------------------------------------------
#> All                                          rtable            2 x 4              TRUE               0     
#>   cl D.1                     1               rtable            2 x 4             FALSE              0, 0   
#>     dcd D.1.1.4.2           1, 1             rtable            1 x 4              TRUE                     
#>     dcd D.1.1.1.1           1, 2             rtable            1 x 4              TRUE                     
#>   cl A.1                     2               rtable            2 x 4             FALSE              0, 0   
#>     dcd A.1.1.1.2           2, 1             rtable            1 x 4              TRUE                     
#>     dcd A.1.1.1.1           2, 2             rtable            1 x 4              TRUE                     
#>   cl B.2                     3               rtable            2 x 4             FALSE              0, 0   
#>     dcd B.2.1.2.1           3, 1             rtable            1 x 4              TRUE                     
#>     dcd B.2.2.3.1           3, 2             rtable            1 x 4              TRUE                     
#>   cl D.2                     4               rtable            2 x 4             FALSE              0, 0   
#>     dcd D.2.1.5.3           4, 1             rtable            1 x 4              TRUE                     
#>   cl B.1                     5               rtable            2 x 4             FALSE              0, 0   
#>     dcd B.1.1.1.1           5, 1             rtable            1 x 4              TRUE                     
#>   cl C.2                     6               rtable            2 x 4             FALSE              0, 0   
#>     dcd C.2.1.2.1           6, 1             rtable            1 x 4              TRUE                     
#>   cl C.1                     7               rtable            2 x 4             FALSE              0, 0   
#>     dcd C.1.1.1.3           7, 1             rtable            1 x 4              TRUE                     

tbls[[1]]

#>                                                           A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                                            (N=134)               (N=134)               (N=132)               (N=400)    
#> ----------------------------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one event         84 (62.69%)           92 (68.66%)           81 (61.36%)           257 (64.25%) 
#> Total number of events                                       147                   153                   128                   428      

tbls[['cl A.1']]

#>                                                           A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                                            (N=134)               (N=134)               (N=132)               (N=400)    
#> ----------------------------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one event         75 (55.97%)           86 (64.18%)            73 (55.3%)           234 (58.5%)  
#> Total number of events                                       127                   146                   124                   397