Response Forest Plot Table

Response Forest Plot Table

t_forest_rsp(rsp, col_by, group_data = NULL, total = "ALL",
  na.omit.group = TRUE, dense_header = FALSE)

Arguments

rsp
boolean, TRUE if subject is a responder FALSE otherwise

col_by
factor with reference and comparison group information, the first level indicates the reference group

group_data
data frame with one column per grouping variable

total
character with the row name of the analysis run on all data. If NULL analysis is omitted

na.omit.group
boolean, do not list NAs

Value

an rtable object

Details

Logistic regression is used for odds ratio calculation.

Each row in the returned table contains the analysis statistics for a subgroup of data (indicated by the row name). The summary table is consist of the following 9 columns:

1

Total n the total number of subjects included in analysis population

3-4

Summary of responders in reference arm, n and Responders are the total number of patients and the number of responders in reference arm, respectively. Response.Rate is the percentage of responders in reference arm.

5-7

same statistics as for reference arm now for comparison arm

8

Odds Ratio ranges from 0 to infinity, estimated by applying univariate logistic regression. Binary response status (responder or non-responder) is the outcome and arm is the explanatory variable. Odds ratio greater than 1 indicates better performance in comparison arm; odds ratio less than 1 indicates better performance in reference arm.

9

95 % CI The 95 uncertainty around the measure of effect (Odds Ratio). Because only a small sample of the overall population is included in the analysis, by having an upper and lower confidence limit we can infer that the true treatment effect lies in between. If the 95 say that the difference between two arms is not significant at a significance level of 0.05.

Examples

library(random.cdisc.data)

ASL <- radam("ASL")
ARS <- radam("ARS", ADSL = ASL)
ASL$'FAKE Name > -1.3 Flag' <- rep(c('Y', 'N'), 50)
ARS_f <- subset(ARS, PARAMCD == "OVRSPI")
ANL <- merge(ASL, ARS_f)

tbl <- t_forest_rsp(
  rsp = ANL$AVALC %in% c("CR", "PR"),
  col_by = factor(ANL$ARM),
  group_data = ANL[, c("SEX", "RACE", "FAKE Name > -1.3 Flag")]
)

tbl

#>                                                                                                  ARM A                                                          ARM B                                                                   
#> Baseline Risk Factors                                 Total n                 n               Responders          Response.Rate              n               Responders          Response.Rate         Odds Ratio             95% CI    
#> ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                                                     100                  46                   20                  0.43                  54                   18                  0.33                 0.65             (0.29, 1.46) 
#> 
#> SEX
#>   F                                                     56                   24                   12                   0.5                  32                    9                  0.28                 0.39             (0.13, 1.17) 
#>   M                                                     44                   22                    8                  0.36                  22                    9                  0.41                 1.21             (0.36, 4.16) 
#> 
#> RACE
#>   AMERICAN INDIAN OR ALASKA NATIVE                      19                   10                    4                   0.4                   9                    3                  0.33                 0.75             (0.11, 4.94) 
#>   ASIAN                                                 23                    7                    4                  0.57                  16                    7                  0.44                 0.58             (0.09, 3.51) 
#>   BLACK OR AFRICAN AMERICAN                             23                   11                    4                  0.36                  12                    2                  0.17                 0.35             (0.04, 2.32) 
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER             19                   11                    7                  0.64                   8                    2                  0.25                 0.19             (0.02, 1.29) 
#>   WHITE                                                 16                    7                    1                  0.14                   9                    4                  0.44                  4.8             (0.5, 111.69)
#> 
#> FAKE Name > -1.3 Flag
#>   N                                                     50                   23                   11                  0.48                  27                    8                   0.3                 0.46             (0.14, 1.45) 
#>   Y                                                     50                   23                    9                  0.39                  27                   10                  0.37                 0.92             (0.29, 2.91) 

# NOT RUN {
Viewer(tbl)
# }