Time-to-event Forest Plot Table

Time-to-event Forest Plot Table

t_forest_tte(tte, is_event, col_by, group_data = NULL, strata_data = NULL,
  total = "ALL", time_unit = "month", na.omit.group = TRUE,
  dense_header = FALSE)

Arguments

tte
time to event data

is_event
is boolean, TRUE if event, FALSE if tte is censored

col_by
factor with reference and comparison group information, the first level indicates the reference group

group_data
data frame with one column per grouping

Details

Cox PH model is used for hazard ratio calculation

The returned table contains one row per analysis applied on a subset of data (indicated by the row name). The analysis is summarized with the following 9 columns:

1

Total n the total number of subjects used for analysis

2-4

analysis for reference arm, n, Events, Median (moneth)

5-7

same analysis as for reference arm now for comparison arm

8

Hazard Ratio

9

95% Wald CI

Examples

library(random.cdisc.data)

ASL <- radam("ASL")
ATE <- radam("ATE", ADSL = ASL)

ATE_f <- subset(ATE, PARAMCD == "OS")

ANL <- merge(ASL, ATE_f)

tbl <- t_forest_tte(
  tte = ANL$AVAL,
  is_event = ANL$CNSR == 0,
  col_by = factor(ANL$ARM),
  group_data = as.data.frame(lapply(ANL[, c("SEX", "RACE")], as.factor)),
  dense_header = TRUE
)

tbl

#>                                                                                               ARM A                                                       ARM B                                                                
#> Baseline Risk                                         Total                 n                 Events              Median                n                 Events              Median              Hazard               95%     
#> Factors                                                 n                                                        (month)                                                     (month)              Ratio              Wald CI   
#> -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                                                    100                  46                  30                106.82                54                  30                103.38               1.03            (0.61, 1.72)
#> 
#> SEX
#>   F                                                     58                  27                  19                106.82                31                  15                103.38               0.99            (0.49, 1.98)
#>   M                                                     42                  19                  11                85.85                 23                  15                79.99                1.11            (0.5, 2.47) 
#> 
#> RACE
#>   AMERICAN INDIAN OR ALASKA NATIVE                      25                  11                  6                 141.48                14                  8                 86.27                 2              (0.6, 6.69) 
#>   ASIAN                                                 15                  6                   3                 106.82                9                   5                 62.77                2.34            (0.44, 12.4)
#>   BLACK OR AFRICAN AMERICAN                             18                  5                   2                                       13                  7                 120.08               1.45            (0.3, 6.99) 
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER             18                  10                  9                 110.9                 8                   2                                      0.68            (0.12, 3.73)
#>   WHITE                                                 24                  14                  10                34.62                 10                  8                 95.09                0.55            (0.2, 1.54)