Multi-variable logistic regression table

Logistic regression for binary outcome with categorical/continuous covariates in model statement. For each covariate category (if categorical) or specified values (if continuous), present degrees of freedom, regression parameter estimate and standard error (SE) relative to reference group or category. Report odds ratios for each covariate category or specified values and corresponding Wald confidence intervals as default but allow user to specify other confidence levels. Report p-value for Wald chi-square test of the null hypothesis that covariate has no effect on response in model containing all specified covariates. Allow option to include one two-way interaction and present similar output for each interaction degree of freedom. Note: For glm formula, the variable names need to be standard dataframe column name without special characters. The big N is the total number of observations for complete cases.

t_logistic(formula, data, increments = NULL, conf_level = 0.95)

Arguments

formula
a glm formula
The formula object can be all main effect model formula, and a formula with one two-way interaction.

data
a data.frame with all the variables specified in formula.

increments
(named list)
Used to specify numeric values of continuous variables in glm model formula which interact with other variables. This is used to calculate the odds ratio when comparing the other interaction variable effect. For example, for a model with ARM and AGE interaction, increments = list(AGE = c(18, 65)) will enable calculation of odds ratios of comparison ARM vs. reference ARM at AGE = 18 and AGE = 65. If increments = NULL, then default AGE value is ceiling of median.

conf_level
(numeric value)
Confidence level for Wald odds ratio confidence interval.

Examples

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)
ADSL <- ADSL %>% dplyr::filter(SEX %in% c("F", "M"))
ADRS <- radrs(ADSL, seed = 2)

ADRS_f <- subset(ADRS, PARAMCD == "BESRSPI") %>%
  dplyr::mutate(Response = ifelse(AVALC %in% c("PR", "CR"), 1, 0))
var_labels(ADRS_f) <- c(var_labels(ADRS), "Response" = "Tumor Response")

tbl <- t_logistic(
  formula = Response ~ ARM + AGE + SEX,
  data = ADRS_f )

tbl # or Viewer(tbl) for a html view

#>                                                                      Degrees of             Parameter               Standard                 Odds                   Wald                  p-value    
#>                                                                        Freedom                Estimate                Error                  Ratio                  95% CI                           
#> -----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> N = 400
#> 
#> Description of Planned Arm (Reference = A: Drug X, n = 134)               2                                                                                                               0.0758     
#>   B: Placebo, n = 134                                                     1                   -1.508                  0.664                  0.22               (0.06, 0.81)              0.0231     
#>   C: Combination, n = 132                                                 1                   16.812                 1522.2                 >999.99            (0.00, >999.99)            0.9912     
#> 
#> Age                                                                       1                    0.008                  0.035                  1.01               (0.94, 1.08)              0.8174     
#> 
#> Sex (Reference = F, n = 231)
#>   M, n = 169                                                              1                   -0.939                  0.559                  0.39               (0.13, 1.17)              0.0928     

tbl2 <- t_logistic(
  formula = Response ~ ARM + AGE + BMRKR2 + ARM*BMRKR2,
  data = ADRS_f
)

tbl2

#>                                                                                    Degrees of             Parameter               Standard                 Odds                   Wald                  p-value    
#>                                                                                      Freedom                Estimate                Error                  Ratio                  95% CI                           
#> -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> N = 400
#> 
#> Description of Planned Arm(Reference = A: Drug X, n = 134)                              2                                                                                                               0.5745     
#>   B: Placebo, n = 134                                                                   1                   -1.239                  1.177                                                               0.2924     
#>     Categorical Level Biomarker 2
#>       LOW                                                                                                                                                  0.29               (0.03, 2.91)                         
#>       MEDIUM                                                                                                                                                 0               (0.00, >999.99)                       
#>       HIGH                                                                                                                                                  0.7               (0.09, 5.24)                         
#>   C: Combination, n = 132                                                               1                    16.68                2803.178                                                              0.9953     
#>     Categorical Level Biomarker 2
#>       LOW                                                                                                                                                 >999.99            (0.00, >999.99)                       
#>       MEDIUM                                                                                                                                                 1               (0.00, >999.99)                       
#>       HIGH                                                                                                                                                >999.99            (0.00, >999.99)                       
#> 
#> Age                                                                                     1                   -0.005                  0.036                  0.99               (0.93, 1.07)              0.8823     
#> 
#> Categorical Level Biomarker 2(Reference = LOW, n = 135)                                 2                                                                                                               0.8244     
#>   MEDIUM, n = 135                                                                       1                   16.681                2914.584                                                              0.9954     
#>     Description of Planned Arm
#>       A: Drug X                                                                                                                                           >999.99            (0.00, >999.99)                       
#>       B: Placebo                                                                                                                                            0.5               (0.12, 2.05)                         
#>       C: Combination                                                                                                                                         1               (0.00, >999.99)                       
#>   HIGH, n = 130                                                                         1                   -0.772                  1.243                                                               0.5343     
#>     Description of Planned Arm
#>       A: Drug X                                                                                                                                            0.46               (0.04, 5.28)                         
#>       B: Placebo                                                                                                                                           1.11               (0.17, 7.06)                         
#>       C: Combination                                                                                                                                       1.02              (0.00, >999.99)                       
#> 
#> Interaction of Description of Planned Arm * Categorical Level Biomarker 2               4                                                                                                               0.9887     
#>   B: Placebo * MEDIUM , n = 56                                                          1                   -17.377               2914.584                                                              0.9952     
#>   C: Combination * MEDIUM , n = 42                                                      1                   -16.681               4882.229                                                              0.9973     
#>   B: Placebo * HIGH , n = 33                                                            1                    0.878                  1.56                                                                0.5736     
#>   C: Combination * HIGH , n = 50                                                        1                    0.791                3760.779                                                              0.9998     

tbl3 <- t_logistic(
  formula = Response ~ ARM + AGE + BMRKR1 + ARM*BMRKR1,
  data = ADRS_f,
  increments = list("BMRKR1" = c(5, 10)),
  conf_level = 0.9
)
tbl3

#>                                                                                  Degrees of             Parameter               Standard                 Odds                   Wald                  p-value    
#>                                                                                    Freedom                Estimate                Error                  Ratio                  90% CI                           
#> -----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> N = 400
#> 
#> Description of Planned Arm(Reference = A: Drug X, n = 134)                            2                                                                                                               0.5295     
#>   B: Placebo, n = 134                                                                 1                   -1.467                  1.301                                                               0.2595     
#>     Continous Level Biomarker 1
#>       5                                                                                                                                                  0.23               (0.07, 0.73)                         
#>       10                                                                                                                                                 0.23               (0.05, 1.14)                         
#>   C: Combination, n = 132                                                             1                    16.66                2935.437                                                              0.9955     
#>     Continous Level Biomarker 1
#>       5                                                                                                                                                 >999.99            (0.00, >999.99)                       
#>       10                                                                                                                                                >999.99            (0.00, >999.99)                       
#> 
#> Age                                                                                   1                   -0.004                  0.035                    1                (0.94, 1.06)              0.9168     
#> 
#> Continous Level Biomarker 1                                                           1                   -0.022                  0.159                                                                0.888     
#>   Description of Planned Arm
#>     A: Drug X                                                                                                                                            0.98               (0.75, 1.27)                         
#>     B: Placebo                                                                                                                                           0.98               (0.85, 1.13)                         
#>     C: Combination                                                                                                                                         1               (0.00, >999.99)                       
#> 
#> Interaction of Description of Planned Arm * Continous Level Biomarker 1               2                                                                                                                  1       
#>   B: Placebo, n = 134                                                                 1                    0.002                  0.182                                                               0.9931     
#>   C: Combination, n = 132                                                             1                    0.022                 444.496                                                                 1       

tbl4 <- t_logistic(
  formula = Response ~ ARM + AGE + BMRKR1 + AGE*BMRKR1,
  data = ADRS_f,
  increments = list("AGE" = c(20, 65), "BMRKR1" = c(5, 10)),
  conf_level = 0.9
)
tbl4

#>                                                                      Degrees of             Parameter               Standard                 Odds                   Wald                  p-value    
#>                                                                        Freedom                Estimate                Error                  Ratio                  90% CI                           
#> -----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> N = 400
#> 
#> Description of Planned Arm (Reference = A: Drug X, n = 134)               2                                                                                                                0.086     
#>   B: Placebo, n = 134                                                     1                   -1.467                  0.662                  0.23               (0.08, 0.69)              0.0268     
#>   C: Combination, n = 132                                                 1                   16.777                1541.791                >999.99            (0.00, >999.99)            0.9913     
#> 
#> Age                                                                       1                   -0.019                  0.064                                                               0.7597     
#>   Continous Level Biomarker 1
#>     5                                                                                                                                          1                (0.94, 1.06)                         
#>     10                                                                                                                                       1.01               (0.91, 1.12)                         
#> 
#> Continous Level Biomarker 1                                               1                   -0.129                  0.372                                                               0.7293     
#>   Age
#>     20                                                                                                                                       0.94                (0.7, 1.24)                         
#>     65                                                                                                                                       1.07               (0.63, 1.84)                         
#> 
#> Interaction of Age * Continous Level Biomarker 1                          1                    0.003                  0.01                                                                0.7692