Mix model with repeated measurements (MMRM) model

The MMRM table function summarizes MMRM test results by visit and groups. The function produces adjusted lsmeans and standard error, as well as conducts comparisons between groups' adjusted means, where the first level of the group is the reference level.

t_mmrm(
  formula = AVAL ~ arm(ARM) + visit(AVISIT) + ARM * VISIT,
  data,
  id_var = "USUBJID",
  arm_var = "ARM",
  visit_var = "AVISIT",
  col_N,
  mode = c("df.error", "auto", "boot-satterthwaite"),
  conf_level = 0.95,
  weights_emmeans = "proportional",
  corStruct = NULL,
  table_tree = TRUE
)

Arguments

formula
a gls formula.

data
a data.frame with all the variables specified in formula. Records with missing values in any independent variables will be excluded.

id_var
a character describing the variable name used as subject IDs, "USUBJID" by default.

arm_var
a character describing the variable name for arm, "ARM" by default. The arm variable must be factor in data

visit_var
a character describing the variable name for visit, "AVISIT" by default. This variable must be factor in data.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

mode
algorithm for degree of freedom: auto, df.error or boot-satterthwaite.

conf_level
confidence level. Must be number greater than 0 and less than 1.

weights_emmeans
argument from emmeans, "proportional" by default.

corStruct
NULL by default or a string with the name of corClasses.

table_tree
(logical value)
If FALSE then a single rtable gets returned, if TRUE a node object gets returned with the elementary rtable objects.

Value

rtable object or table tree, depending on the the `table_tree` argument

Details

Create rtable of MMRM test results

See also

Examples

library(dplyr)
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADQS <- radqs(cached = TRUE)
ADQS_f <- ADQS %>%
  dplyr::filter(PARAMCD=="FKSI-FWB" & !AVISIT %in% c("BASELINE")) %>%
  droplevels() %>%
  dplyr::mutate(ARM = factor(ARM, levels = c("B: Placebo", "A: Drug X", "C: Combination"))) %>%
  dplyr::mutate(AVISITN = rank(AVISITN) %>% as.factor() %>% as.numeric() %>% as.factor())

t_mmrm(formula = AVAL ~ ARM + AVISIT + STRATA1 + BMRKR2 + ARM*AVISIT,
       data = ADQS_f,
       id_var = "USUBJID",
       arm_var = "ARM",
       visit_var = "AVISIT",
       col_N = table(ADSL$ARM),
       mode = "boot-satterthwaite",
       conf_level = 0.95,
       weights_emmeans = "proportional",
       corStruct = "corSymm",
       table_tree = FALSE
)

#> MMRM methodology in R is different from SAS. Please use as exploratory purpose.

#>                                                                B: Placebo              A: Drug X             C: Combination 
#>                                                                 (N=134)                 (N=134)                 (N=132)     
#> ----------------------------------------------------------------------------------------------------------------------------
#> SCREENING
#>   n                                                               134                     134                     132       
#>   Adjusted Mean (SE)                                         45.190 (0.738)          44.578 (0.737)          44.733 (0.743) 
#>   95% CI                                                    (43.743, 46.637)        (43.133, 46.024)        (43.277, 46.19) 
#> 
#>   Difference in Adjusted Means (SE) (vs. B: Placebo)                                 -0.611 (1.044)          -0.457 (1.048) 
#>   95% CI                                                                            (-2.659, 1.436)         (-2.511, 1.597) 
#>   Relative Reduction (%)                                                                  1.4%                     1%       
#> 
#>   p-value (MMRM)                                                                         0.5581                  0.6628     
#> 
#> WEEK 1 DAY 8
#>   n                                                               134                     134                     132       
#>   Adjusted Mean (SE)                                         54.306 (0.689)          55.032 (0.688)          53.974 (0.693) 
#>   95% CI                                                    (52.956, 55.657)        (53.682, 56.382)        (52.614, 55.333)
#> 
#>   Difference in Adjusted Means (SE) (vs. B: Placebo)                                 0.725 (0.975)           -0.333 (0.978) 
#>   95% CI                                                                            (-1.186, 2.637)         (-2.251, 1.585) 
#>   Relative Reduction (%)                                                                 -1.3%                    0.6%      
#> 
#>   p-value (MMRM)                                                                         0.4568                  0.7338     
#> 
#> WEEK 2 DAY 15
#>   n                                                               134                     134                     132       
#>   Adjusted Mean (SE)                                         59.954 (0.766)          59.804 (0.765)          59.320 (0.771) 
#>   95% CI                                                    (58.452, 61.455)        (58.303, 61.305)        (57.809, 60.832)
#> 
#>   Difference in Adjusted Means (SE) (vs. B: Placebo)                                 -0.150 (1.083)          -0.633 (1.087) 
#>   95% CI                                                                            (-2.274, 1.975)         (-2.765, 1.499) 
#>   Relative Reduction (%)                                                                  0.2%                    1.1%      
#> 
#>   p-value (MMRM)                                                                          0.89                   0.5603     
#> 
#> WEEK 3 DAY 22
#>   n                                                               134                     134                     132       
#>   Adjusted Mean (SE)                                         64.366 (0.870)          66.799 (0.870)          66.600 (0.876) 
#>   95% CI                                                    (62.659, 66.072)        (65.094, 68.505)        (64.882, 68.318)
#> 
#>   Difference in Adjusted Means (SE) (vs. B: Placebo)                                 2.434 (1.231)           2.235 (1.235)  
#>   95% CI                                                                             (0.02, 4.848)          (-0.188, 4.657) 
#>   Relative Reduction (%)                                                                 -3.8%                   -3.5%      
#> 
#>   p-value (MMRM)                                                                         0.0481                  0.0706     
#> 
#> WEEK 4 DAY 29
#>   n                                                               134                     134                     132       
#>   Adjusted Mean (SE)                                         68.921 (0.994)          70.264 (0.994)          70.322 (1.001) 
#>   95% CI                                                    (66.972, 70.87)         (68.316, 72.213)        (68.359, 72.285)
#> 
#>   Difference in Adjusted Means (SE) (vs. B: Placebo)                                 1.343 (1.406)           1.401 (1.411)  
#>   95% CI                                                                            (-1.414, 4.101)         (-1.366, 4.168) 
#>   Relative Reduction (%)                                                                 -1.9%                    -2%       
#> 
#>   p-value (MMRM)                                                                         0.3395                  0.3209     
#> 
#> WEEK 5 DAY 36
#>   n                                                               134                     134                     132       
#>   Adjusted Mean (SE)                                         74.322 (1.057)          75.716 (1.057)          74.515 (1.065) 
#>   95% CI                                                    (72.249, 76.395)        (73.644, 77.789)        (72.427, 76.603)
#> 
#>   Difference in Adjusted Means (SE) (vs. B: Placebo)                                 1.394 (1.495)           0.193 (1.501)  
#>   95% CI                                                                            (-1.538, 4.327)          (-2.75, 3.136) 
#>   Relative Reduction (%)                                                                 -1.9%                   -0.3%      
#> 
#>   p-value (MMRM)                                                                         0.3512                  0.8977