Tabulation

Tabulate the fixed effect estimates of an MMRM fit.

t_mmrm_fixed(object, format = "xx.xxxx")

Arguments

object
the MMRM model result produced by s_mmrm.

format
rtables format for the numbers other than degrees of freedom and p-values (default is "xx.xxxx").

Value

rtable object with the fixed effect estimates, standard errors, degrees of freedom, t-statistics, and p-values.

Examples

library(dplyr)
library(random.cdisc.data)

adsl <- radsl(cached = TRUE)
adqs <- radqs(cached = TRUE)
adqs_f <- adqs %>%
  dplyr::filter(PARAMCD=="FKSI-FWB" & !AVISIT %in% c("BASELINE")) %>%
  droplevels() %>%
  dplyr::mutate(ARM = factor(ARM, levels = c("B: Placebo", "A: Drug X", "C: Combination"))) %>%
  dplyr::mutate(AVISITN = rank(AVISITN) %>% as.factor() %>% as.numeric() %>% as.factor())

mmrm_results <- s_mmrm(
  vars = list(
    response = "AVAL",
    covariates = c("STRATA1", "BMRKR2"),
    id = "USUBJID",
    arm = "ARM",
    visit = "AVISIT"
  ),
  data = adqs_f,
  cor_struct = "random-quadratic",
  optimizer = "nloptwrap_bobyqa"
)
t_mmrm_fixed(mmrm_results)

#>                                               Estimate         Std. Error         t value              df             Pr(>|t|) 
#> -------------------------------------------------------------------------------------------------------------------------------
#> (Intercept)                                   45.7834            0.9268           49.4017             1775            <0.0001  
#> STRATA1B                                      -0.2016            0.515            -0.3915             903              0.6955  
#> STRATA1C                                      -0.3036            0.5078           -0.5978             903              0.5501  
#> BMRKR2MEDIUM                                  -0.4861            0.504            -0.9643             903              0.3351  
#> BMRKR2HIGH                                    -0.7152            0.5077           -1.4088             903              0.1592  
#> ARMA: Drug X                                   -0.642            1.1726           -0.5476             1768             0.5841  
#> ARMC: Combination                             -0.4875            1.1767           -0.4143             1768             0.6787  
#> AVISITWEEK 1 DAY 8                             9.1165            1.1715            7.782              1971            <0.0001  
#> AVISITWEEK 2 DAY 15                           14.7637            1.1733           12.5832             1752            <0.0001  
#> AVISITWEEK 3 DAY 22                           19.1756            1.1655           16.4526             1942            <0.0001  
#> AVISITWEEK 4 DAY 29                            23.731            1.1755           20.1877             1681            <0.0001  
#> AVISITWEEK 5 DAY 36                           29.1319            1.2592           23.1349             890             <0.0001  
#> ARMA: Drug X:AVISITWEEK 1 DAY 8                1.3369            1.6567            0.8069             1971             0.4198  
#> ARMC: Combination:AVISITWEEK 1 DAY 8           0.1241            1.663             0.0746             1971             0.9405  
#> ARMA: Drug X:AVISITWEEK 2 DAY 15               0.4617            1.6593            0.2783             1752             0.7809  
#> ARMC: Combination:AVISITWEEK 2 DAY 15         -0.1766            1.6656            -0.106             1752             0.9156  
#> ARMA: Drug X:AVISITWEEK 3 DAY 22               3.0454            1.6483            1.8476             1942             0.0648  
#> ARMC: Combination:AVISITWEEK 3 DAY 22          2.6914            1.6545            1.6267             1942             0.1040  
#> ARMA: Drug X:AVISITWEEK 4 DAY 29               1.955             1.6624            1.176              1681             0.2398  
#> ARMC: Combination:AVISITWEEK 4 DAY 29          1.8577            1.6687            1.1133             1681             0.2657  
#> ARMA: Drug X:AVISITWEEK 5 DAY 36               2.0059            1.7808            1.1264             890              0.2603  
#> ARMC: Combination:AVISITWEEK 5 DAY 36          0.6498            1.7875            0.3635             890              0.7163