Response Table Caclulation Function

TODO: Description

t_rsp(rsp, col_by, parition_rsp_by = NULL, strata_data = NULL)

Arguments

rsp
boolean, TRUE if subject is a responder FALSE otherwise

col_by
factor with reference and comparison group information, the first level indicates the reference group

parition_rsp_by
factor to partition rsp and run additional analysis

strata_data
data for stratification factors (categorical variables), if NULL no stratified analysis is performed

Value

an rtable object

Details

If no stratification, chi-squared test is performed by default to test for equality of proportions in response rate between reference and comparison arms.

If stratification factors are added, Cochran-Mantel-Haensel test statistics are reported for the test of equal proportions, odds ratio and its corresponding 95

Examples

library(random.cdisc.data)

ASL <- radam("ASL", N = 1000, start_with = list(RACE = c("white", "asian")))
ARS <- radam("ARS", ADSL = ASL)

ASL$ARM <- factor(sample(LETTERS[1:3], nrow(ASL), TRUE))
ASL$RACE <- factor(ASL$RACE)

ANL <- merge(ASL, subset(ARS, PARAMCD == "OVRSPI"))

tbl <- t_rsp(
 rsp = ANL$AVALC %in% c("CR", "PR"),
 col_by = relevel(factor(ANL$ARMCD), "ARM B", "ARM A"),
 parition_rsp_by = droplevels(factor(ANL$AVALC, levels = c("CR", "PR", "SD", "NON CR/PD", "PD", "NE"))),
 strata_data = ANL[, c("SEX", "RACE")]
)

tbl

#>                                                                      ARM B                 ARM A     
#>                                                                     (N=494)               (N=506)    
#> -----------------------------------------------------------------------------------------------------
#> Responders                                                        199 (40.28%)          213 (42.09%) 
#> Non-Responders                                                    295 (59.72%)          293 (57.91%) 
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                      (35.93, 44.76)        (37.75, 46.53)
#> 
#> Difference in Response Rates (x100)                                                         1.81     
#>   95% CI for difference (Wald without correction)                                      (-4.42, 8.16) 
#>   95% CI for difference (Wald with continuity correction)                              (-0.05, 0.08) 
#>   p-value (Cochran-Mantel-Haenszel)                                                    <5 data points
#> 
#> Odds Ratio                                                                             <5 data points
#>   95% CI                                                                               <5 data points
#> 
#> CR                                                                96 (19.43%)           93 (18.38%)  
#>   95% CI (Wald)                                                  (16.03, 23.2)         (15.1, 22.03) 
#> 
#> PR                                                                103 (20.85%)          120 (23.72%) 
#>   95% CI (Wald)                                                  (17.35, 24.7)         (20.07, 27.67)
#> 
#> SD                                                                115 (23.28%)          94 (18.58%)  
#>   95% CI (Wald)                                                  (19.62, 27.26)        (15.28, 22.24)
#> 
#> PD                                                                95 (19.23%)           103 (20.36%) 
#>   95% CI (Wald)                                                  (15.85, 22.99)        (16.93, 24.13)
#> 
#> NE                                                                85 (17.21%)           96 (18.97%)  
#>   95% CI (Wald)                                                  (13.98, 20.83)        (15.65, 22.67)