Summary Table by Visit

This function summarizes test results or change from baseline statistics by visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.

t_summarize_by_visit(data, visit, col_by)

Arguments

data
data frame with numerical variables to be summarized. If the variable has a label attribute then it will be used as the sub-header column name.

visit
factor with visit names ordered by desired display order in the stacked table.

col_by
factor with reference and comparison group information, the first level indicates the reference group

Value

an rtable object

Details

Every variable in data will be mapped to a sub-column under col_by column. Multiple variables will be shown side-by-side with the variable label (if present) or variable name as the sub-column header.

Currently n, Mean (SD), Median, IQR and Min-Max will be derived for each variable from data per each factor in col_by.

Examples

# EXAMPLE 1
ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B")
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA
ANL$PCHG <- ANL$CHG/ANL$AVAL*100

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

t_summarize_by_visit(data = ANL[c("AVAL")], visit = ANL$VISIT, col_by = ANL$ARM)

#>                       ARM A               ARM B    
#>                      (N=1000)            (N=1000)  
#> ---------------------------------------------------
#> visit 1
#>   n                    100                 100     
#>   Mean (SD)        -0.06 (0.91)          0 (0.96)  
#>   Median               0.01                0.09    
#>   IQR              -0.68 - 0.54        -0.73 - 0.66
#>   Min - Max        -2.57 - 1.92        -2.96 - 2.08
#> 
#> visit 2
#>   n                    100                 100     
#>   Mean (SD)        -0.16 (0.95)        -0.01 (1.04)
#>   Median              -0.17               -0.02    
#>   IQR              -0.86 - 0.49        -0.6 - 0.66 
#>   Min - Max        -2.15 - 2.1         -3.4 - 2.28 
#> 
#> visit 3
#>   n                    100                 100     
#>   Mean (SD)        -0.03 (1.02)        -0.17 (0.96)
#>   Median              -0.19               -0.13    
#>   IQR              -0.75 - 0.77        -0.83 - 0.4 
#>   Min - Max        -2.21 - 2.54        -2.65 - 2.72
#> 
#> visit 4
#>   n                    100                 100     
#>   Mean (SD)        -0.05 (0.98)        0.15 (0.98) 
#>   Median               0.03                0.19    
#>   IQR              -0.6 - 0.56         -0.61 - 0.84
#>   Min - Max        -2.71 - 2.8         -1.83 - 3.13
#> 
#> visit 5
#>   n                    100                 100     
#>   Mean (SD)        -0.2 (1.08)         -0.02 (1.04)
#>   Median              -0.25                0.04    
#>   IQR              -0.84 - 0.43        -0.72 - 0.56
#>   Min - Max        -3.37 - 2.71        -2.85 - 2.53
#> 
#> visit 6
#>   n                    100                 100     
#>   Mean (SD)        -0.03 (0.96)        -0.1 (0.89) 
#>   Median              -0.07               -0.22    
#>   IQR              -0.73 - 0.71        -0.67 - 0.54
#>   Min - Max        -2.05 - 2.95        -2.28 - 2.32
#> 
#> visit 7
#>   n                    100                 100     
#>   Mean (SD)         0.01 (1.1)           0 (0.99)  
#>   Median              -0.08               -0.04    
#>   IQR              -0.68 - 0.95        -0.72 - 0.63
#>   Min - Max        -2.73 - 2.74        -2.08 - 2.5 
#> 
#> visit 8
#>   n                    100                 100     
#>   Mean (SD)        -0.07 (1.07)        0.04 (1.03) 
#>   Median               -0.1                0.07    
#>   IQR              -0.82 - 0.61        -0.79 - 0.73
#>   Min - Max        -2.32 - 3.41        -2.06 - 2.83
#> 
#> visit 9
#>   n                    100                 100     
#>   Mean (SD)         -0.03 (1)          0.05 (0.96) 
#>   Median               0.04                0.09    
#>   IQR              -0.69 - 0.73        -0.54 - 0.73
#>   Min - Max        -2.45 - 2.19        -2.36 - 1.94
#> 
#> visit 10
#>   n                    100                 100     
#>   Mean (SD)        -0.13 (1.05)         0.1 (1.02) 
#>   Median               -0.2                0.07    
#>   IQR              -0.83 - 0.56        -0.48 - 0.87
#>   Min - Max        -3.31 - 2.69        -2.98 - 2.16

t_summarize_by_visit(data = ANL[c("PCHG")], visit = ANL$VISIT, col_by = ANL$ARM)

#>                            ARM A                        ARM B        
#>                          (N=1000)                     (N=1000)       
#> ---------------------------------------------------------------------
#> visit 1
#>   n                          0                            0          
#>   Mean (SD)               NE (NE)                      NE (NE)       
#>   Median                    NE                           NE          
#>   IQR                     NE - NE                      NE - NE       
#>   Min - Max               NE - NE                      NE - NE       
#> 
#> visit 2
#>   n                         100                          100         
#>   Mean (SD)           -6.37 (1767.38)            -1032.14 (8130.78)  
#>   Median                  -42.01                       -42.04        
#>   IQR                -331.93 - 245.07             -339.56 - 227.84   
#>   Min - Max        -11019.59 - 10808.31          -80002.49 - 3720.46 
#> 
#> visit 3
#>   n                         100                          100         
#>   Mean (SD)          -223.1 (1456.94)             -111.35 (1668.42)  
#>   Median                   -54.5                       -19.89        
#>   IQR                -345.22 - 183.16             -290.83 - 351.34   
#>   Min - Max         -8534.82 - 8742.15           -9222.45 - 4796.61  
#> 
#> visit 4
#>   n                         100                          100         
#>   Mean (SD)          -158.19 (1435.91)            -36.35 (1763.88)   
#>   Median                   54.01                        35.77        
#>   IQR                -271.09 - 286.63             -198.84 - 344.81   
#>   Min - Max         -11129.59 - 2394.43          -13466.07 - 8259.17 
#> 
#> visit 5
#>   n                         100                          100         
#>   Mean (SD)          -102.95 (1716.71)          -70384.81 (703336.75)
#>   Median                  -106.66                       37.91        
#>   IQR                 -433.03 - 80.27              -301.97 - 279.5   
#>   Min - Max         -11943.23 - 8994.97         -7033410.58 - 3045.87
#> 
#> visit 6
#>   n                         100                          100         
#>   Mean (SD)          627.88 (4774.67)               9.94 (903.18)    
#>   Median                  -30.06                       -108.86       
#>   IQR                -258.75 - 305.86             -334.55 - 302.74   
#>   Min - Max         -4365.76 - 42264.35          -2615.49 - 4374.48  
#> 
#> visit 7
#>   n                         100                          100         
#>   Mean (SD)           13.79 (2379.66)             1012.48 (9200.66)  
#>   Median                   35.34                        30.04        
#>   IQR                 -268.15 - 263.8             -175.96 - 251.78   
#>   Min - Max        -11646.25 - 13857.77          -9874.98 - 90561.26 
#> 
#> visit 8
#>   n                         100                          100         
#>   Mean (SD)          349.64 (1426.08)             171.37 (1869.88)   
#>   Median                   10.91                        55.76        
#>   IQR                 -213.77 - 402.1             -305.49 - 286.96   
#>   Min - Max         -1948.17 - 10885.93          -4730.03 - 15814.89 
#> 
#> visit 9
#>   n                         100                          100         
#>   Mean (SD)           -118 (3824.47)              -202.38 (1347.2)   
#>   Median                   51.46                        3.26         
#>   IQR                -303.93 - 296.99             -284.98 - 324.05   
#>   Min - Max        -26381.14 - 23892.48           -6656.9 - 3835.67  
#> 
#> visit 10
#>   n                         100                          100         
#>   Mean (SD)           -771 (7196.16)              -255.75 (4555.45)  
#>   Median                  -30.95                        17.76        
#>   IQR                -239.03 - 244.45             -265.14 - 301.77   
#>   Min - Max         -71648.44 - 2142.89         -18621.05 - 30222.84 

# Add label to variable instead showing variable name
ANL <- var_relabel(ANL, AVAL = "Value at\nVisit",
                        CHG = "Change from\nBaseline",
                        PCHG = "Percent Change\nfrom Baseline")
t_summarize_by_visit(data = ANL[c("AVAL", "CHG")], visit = ANL$VISIT, col_by = ANL$ARM)

#>                                 ARM A                                   ARM B              
#>                                (N=1000)                                (N=1000)            
#>                      Value at          Change from           Value at          Change from 
#>                       Visit              Baseline             Visit              Baseline  
#> -------------------------------------------------------------------------------------------
#> visit 1
#>   n                    100                  0                  100                  0      
#>   Mean (SD)        -0.06 (0.91)          NE (NE)             0 (0.96)            NE (NE)   
#>   Median               0.01                 NE                 0.09                 NE     
#>   IQR              -0.68 - 0.54          NE - NE           -0.73 - 0.66          NE - NE   
#>   Min - Max        -2.57 - 1.92          NE - NE           -2.96 - 2.08          NE - NE   
#> 
#> visit 2
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.16 (0.95)         1.9 (1.78)         -0.01 (1.04)        1.76 (2.07) 
#>   Median              -0.17                2.18               -0.02                1.73    
#>   IQR              -0.86 - 0.49        0.66 - 3.03         -0.6 - 0.66         0.29 - 3.21 
#>   Min - Max        -2.15 - 2.1         -2.87 - 6.31        -3.4 - 2.28         -3.56 - 5.96
#> 
#> visit 3
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.03 (1.02)        2.04 (1.88)         -0.17 (0.96)        1.84 (1.99) 
#>   Median              -0.19                2.03               -0.13                2.03    
#>   IQR              -0.75 - 0.77        0.66 - 3.26         -0.83 - 0.4          0.42 - 3.1 
#>   Min - Max        -2.21 - 2.54        -2.32 - 7.79        -2.65 - 2.72        -2.79 - 6.59
#> 
#> visit 4
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.05 (0.98)        2.12 (2.17)         0.15 (0.98)          1.8 (1.93) 
#>   Median               0.03                2.08                0.19                1.84    
#>   IQR              -0.6 - 0.56          0.7 - 3.66         -0.61 - 0.84         0.34 - 3.1 
#>   Min - Max        -2.71 - 2.8         -2.27 - 7.36        -1.83 - 3.13        -3.15 - 6.06
#> 
#> visit 5
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.2 (1.08)         1.73 (2.26)         -0.02 (1.04)        1.86 (2.06) 
#>   Median              -0.25                1.64                0.04                1.73    
#>   IQR              -0.84 - 0.43        0.21 - 3.41         -0.72 - 0.56         0.7 - 3.18 
#>   Min - Max        -3.37 - 2.71        -3.16 - 6.62        -2.85 - 2.53        -3.1 - 7.06 
#> 
#> visit 6
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.03 (0.96)        1.74 (1.88)         -0.1 (0.89)         2.11 (2.19) 
#>   Median              -0.07                1.72               -0.22                2.06    
#>   IQR              -0.73 - 0.71         0.6 - 2.94         -0.67 - 0.54         0.5 - 3.42 
#>   Min - Max        -2.05 - 2.95        -4.3 - 6.78         -2.28 - 2.32        -3.61 - 8.71
#> 
#> visit 7
#>   n                    100                 100                 100                 100     
#>   Mean (SD)         0.01 (1.1)          1.9 (2.04)           0 (0.99)          1.69 (2.16) 
#>   Median              -0.08                1.9                -0.04                1.53    
#>   IQR              -0.68 - 0.95        0.72 - 3.25         -0.72 - 0.63        0.17 - 3.18 
#>   Min - Max        -2.73 - 2.74        -4.15 - 8.1         -2.08 - 2.5         -2.81 - 8.08
#> 
#> visit 8
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.07 (1.07)        2.04 (2.17)         0.04 (1.03)         2.49 (1.92) 
#>   Median               -0.1                1.67                0.07                2.59    
#>   IQR              -0.82 - 0.61        0.68 - 3.49         -0.79 - 0.73         1.19 - 3.9 
#>   Min - Max        -2.32 - 3.41        -3.7 - 7.59         -2.06 - 2.83        -2.51 - 6.95
#> 
#> visit 9
#>   n                    100                 100                 100                 100     
#>   Mean (SD)         -0.03 (1)          1.92 (2.22)         0.05 (0.96)         1.99 (1.88) 
#>   Median               0.04                2.01                0.09                1.74    
#>   IQR              -0.69 - 0.73        0.42 - 3.62         -0.54 - 0.73         0.8 - 3.31 
#>   Min - Max        -2.45 - 2.19        -3.2 - 8.35         -2.36 - 1.94        -3.06 - 6.12
#> 
#> visit 10
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.13 (1.05)        1.81 (2.08)          0.1 (1.02)         2.11 (1.79) 
#>   Median               -0.2                1.55                0.07                1.83    
#>   IQR              -0.83 - 0.56        0.46 - 3.27         -0.48 - 0.87          1 - 3.29  
#>   Min - Max        -3.31 - 2.69        -3.27 - 6.3         -2.98 - 2.16        -2.58 - 8.42

# EXAMPLE 2
library(random.cdisc.data)
library(dplyr)

AQS <- radam("AQS") %>%
   filter(PARAMCD == "BFIALL") %>%
   mutate(
     ARM = factor(sample(c("ARM A", "ARM B", "ARM C"), nrow(.), replace = TRUE)),
     AVISIT = factor(AVISIT, levels = unique(AVISIT))
   )

tbl <- t_summarize_by_visit(data = AQS[c("AVAL", "CHG")], visit = AQS$AVISIT, col_by = AQS$ARM)
tbl

#>                                  ARM A                                     ARM B                                     ARM C               
#>                                 (N=219)                                   (N=231)                                   (N=273)              
#>                        AVAL                  CHG                 AVAL                  CHG                 AVAL                  CHG     
#> -----------------------------------------------------------------------------------------------------------------------------------------
#> WEEK 1
#>   n                     24                   24                   36                   36                   40                   40      
#>   Mean (SD)         0.06 (0.87)             0 (0)            -0.02 (1.05)             0 (0)             0.05 (1.03)             0 (0)    
#>   Median               -0.02                  0                  0.14                   0                  0.07                   0      
#>   IQR               -0.3 - 0.44             0 - 0            -0.63 - 0.81             0 - 0            -0.53 - 0.55             0 - 0    
#>   Min - Max         -1.7 - 1.69             0 - 0            -2.52 - 1.91             0 - 0            -2.34 - 2.09             0 - 0    
#> 
#> WEEK 2
#>   n                     35                   35                   27                   27                   38                   38      
#>   Mean (SD)        -0.23 (0.89)          -0.1 (1.54)          0.45 (0.9)           0.4 (1.25)           0.09 (1.16)         -0.06 (1.55) 
#>   Median               -0.4                 -0.38                0.42                  0.3                 0.06                 -0.24    
#>   IQR               -0.8 - 0.12         -1.16 - 1.32         -0.17 - 0.86         -0.47 - 1.39         -0.58 - 0.84         -1.03 - 0.71 
#>   Min - Max         -1.8 - 1.63         -2.88 - 3.02         -0.89 - 2.62         -1.96 - 2.41          -2.91 - 2.2         -4.22 - 3.45 
#> 
#> WEEK 3
#>   n                     26                   26                   31                   31                   38                   38      
#>   Mean (SD)        -0.39 (0.94)         -0.66 (1.31)          0.19 (1.03)          0.37 (1.42)         -0.27 (1.01)         -0.23 (1.22) 
#>   Median               -0.64                -0.87                 0.2                 0.42                 -0.35                -0.11    
#>   IQR                -0.98 - 0          -1.51 - 0.19         -0.47 - 0.89         -0.25 - 0.77         -0.96 - 0.51         -0.83 - 0.55 
#>   Min - Max        -1.83 - 2.14         -2.81 - 2.62         -1.57 - 2.73         -2.88 - 3.33         -2.43 - 2.34          -2.8 - 2.3  
#> 
#> WEEK 4
#>   n                     29                   29                   34                   34                   25                   25      
#>   Mean (SD)         0.17 (0.9)           0.2 (1.29)           0.02 (0.75)         -0.13 (1.17)          0.01 (0.97)          0.18 (1.33) 
#>   Median               -0.04                0.27                 -0.06                -0.26                  0                   0.4     
#>   IQR              -0.41 - 0.79         -0.33 - 0.61         -0.43 - 0.57          -0.9 - 0.83         -0.48 - 0.67          -0.7 - 0.81 
#>   Min - Max        -1.17 - 2.25         -2.63 - 3.19         -1.59 - 1.56         -2.38 - 2.91         -1.79 - 2.46         -2.35 - 2.22 
#> 
#> WEEK 5
#>   n                     26                   26                   23                   23                   30                   30      
#>   Mean (SD)        -0.22 (1.08)         -0.27 (1.29)           -0.41 (1)          -0.26 (1.18)          0.12 (1.12)          0.2 (1.33)  
#>   Median               -0.1                   0                  -0.66                -0.33                0.14                 0.18     
#>   IQR              -0.71 - 0.59         -1.11 - 0.53         -1.06 - 0.36         -1.08 - 0.53         -0.98 - 0.95         -0.46 - 0.96 
#>   Min - Max         -2.44 - 1.8         -2.93 - 1.92         -1.88 - 1.97         -2.33 - 1.89         -1.57 - 2.78         -2.49 - 2.87 
#> 
#> WEEK 6
#>   n                     20                   20                   20                   20                   28                   28      
#>   Mean (SD)        -0.22 (1.02)         -0.26 (1.51)           -0.16 (1)          -0.25 (1.31)         -0.04 (1.14)         -0.02 (1.35) 
#>   Median               -0.53                -0.07                0.01                 -0.05                0.09                 0.34     
#>   IQR              -0.64 - 0.61          -1.9 - 0.79         -0.72 - 0.58         -0.67 - 0.29         -0.83 - 0.98          -1.03 - 0.8 
#>   Min - Max         -2.36 - 1.6         -2.47 - 2.22          -2.8 - 1.13          -4.1 - 1.88         -2.23 - 1.58         -2.71 - 1.94 
#> 
#> WEEK 7
#>   n                     16                   16                   19                   19                   24                   24      
#>   Mean (SD)         0.12 (1.06)          0.06 (1.24)          0.02 (1.08)          0.11 (1.74)          0.04 (0.95)         -0.19 (1.04) 
#>   Median               0.22                 0.17                 -0.29                -0.32                0.02                 -0.11    
#>   IQR              -0.84 - 0.97         -0.62 - 1.14          -0.6 - 0.69         -1.01 - 0.61         -0.52 - 0.44          -0.69 - 0.5 
#>   Min - Max        -1.58 - 1.99         -2.68 - 1.78         -2.26 - 2.31          -2.3 - 3.84         -2.23 - 2.01         -2.81 - 1.65 
#> 
#> WEEK 8
#>   n                     14                   14                   16                   16                   22                   22      
#>   Mean (SD)          -0.17 (1)          -0.61 (1.43)          0.35 (1.13)          0.45 (1.16)          0.05 (1.11)          0.02 (1.65) 
#>   Median               -0.32                -1.02                0.51                 0.38                 0.12                 0.16     
#>   IQR              -0.83 - 0.18         -1.83 - 0.26         -0.46 - 0.73         -0.01 - 0.78         -0.49 - 0.42         -1.06 - 0.84 
#>   Min - Max         -1.6 - 2.02         -2.27 - 1.99         -1.63 - 2.83          -1.3 - 2.78         -2.12 - 3.11         -3.98 - 3.29 
#> 
#> WEEK 9
#>   n                     15                   15                   10                   10                   17                   17      
#>   Mean (SD)          0 (0.91)            0.32 (1.37)          0.25 (1.16)          0.24 (1.63)          0.18 (1.05)         -0.11 (1.15) 
#>   Median               -0.18                -0.02                0.32                 0.52                  0.3                 0.08     
#>   IQR              -0.45 - 0.53         -0.45 - 0.78         -0.66 - 1.32          -1.23 - 1.6           0 - 0.68           -0.46 - 0.61 
#>   Min - Max        -1.73 - 1.51         -2.05 - 3.47          -1.37 - 1.7         -2.19 - 2.24         -2.67 - 1.63         -2.03 - 1.82 
#> 
#> WEEK 10
#>   n                      9                    9                    9                    9                    7                    7      
#>   Mean (SD)         0.48 (1.05)          0.36 (1.33)          0.26 (1.4)           0.65 (1.59)         -0.22 (1.04)         -0.84 (1.42) 
#>   Median               0.09                 0.18                  0.4                  0.7                 -0.55                -0.62    
#>   IQR              -0.16 - 1.03         -0.32 - 1.34         -0.85 - 1.07          0.51 - 1.95         -0.97 - 0.04         -1.54 - -0.36
#>   Min - Max        -0.94 - 2.16          -1.65 - 2.4         -1.47 - 2.82          -2.36 - 2.4         -1.01 - 1.91         -2.91 - 1.44 
#> 
#> WEEK 11
#>   n                      5                    5                    6                    6                    4                    4      
#>   Mean (SD)         0.4 (0.48)           0.29 (1.24)          0.62 (0.65)          0.56 (0.44)         -0.42 (0.61)         -0.59 (0.83) 
#>   Median               0.27                  0.6                 0.83                 0.61                 -0.25                -0.71    
#>   IQR              -0.02 - 0.85          -0.5 - 0.9           0.21 - 1.07          0.32 - 0.74         -0.59 - -0.07        -1.17 - -0.13
#>   Min - Max        -0.06 - 0.97         -1.36 - 1.81         -0.37 - 1.26         -0.07 - 1.19          -1.28 - 0.1         -1.38 - 0.46 
# NOT RUN {
Viewer(tbl)
# }