Table of Variable Summaries

Similiar as the demographic table in STREAM

t_summarize_variables(data, col_by, total = NULL, drop_levels = TRUE,
  useNA_factors = c("no", "ifany", "always"))

Arguments

data
data frame with variables to be summarized as described in the details section. If the variable has a label attribute then it will be used for the row name.

col_by
a vector of length nrow(x) that defines which levels in col_by define a column. If data should not be split into columns use the no_by function.

total
if not NULL then it must be a string and an addition column will be added with the overall summaries

drop_levels
boolean, drop zero-count levels for factors

useNA_factors
forwarded to useNA of rtabulate.factor

Details

Every variable in data will be mapped to a summary table of that variable and the tables for all variables will be stacked.

Currently we distinguish the summary by the variable type

numneric

n, Mean (SD), Median, Min-Max will be derived for each factor in col_by

non-numeric

will be converted to a factor (if it isn't already) and the number subjects in a cell and the percentage within the col_by level will be calculated

Examples

library(random.cdisc.data)

ASL <- radam("ASL")

# control the label
ASL <- var_relabel(ASL, BAGE = "Baseline Age of patient")

# control categorical order
ASL$SEX <- relevel(ASL$SEX, "M", "F")

t_summarize_variables(ASL[, c("SEX", "BAGE")], col_by = ASL$ARM, total = "All Patients")

#>                                   ARM A               ARM B            All Patients
#>                                   (N=59)              (N=41)             (N=100)   
#> -----------------------------------------------------------------------------------
#> 
#>   n                                 59                  41                 100     
#>   M                            24 (40.68%)         20 (48.78%)           44 (44%)  
#>   F                            35 (59.32%)         21 (51.22%)           56 (56%)  
#> 
#> Baseline Age of patient
#>   n                                 59                  41                 100     
#>   Mean (SD)                    40.9 (14.92)        38.9 (14.73)        40.08 (14.8)
#>   Median                            41                  40                 40.5    
#>   Min - Max                      13 - 84             14 - 74             13 - 84   

t_summarize_variables(ASL[, c("SEX", "BAGE")], col_by = ASL$ARM, total = "All Patients",
                      useNA_factors = 'always')

#>                                   ARM A               ARM B            All Patients
#>                                   (N=59)              (N=41)             (N=100)   
#> -----------------------------------------------------------------------------------
#> 
#>   n                                 59                  41                 100     
#>   M                            24 (40.68%)         20 (48.78%)           44 (44%)  
#>   F                            35 (59.32%)         21 (51.22%)           56 (56%)  
#>   NA                              0 (0%)              0 (0%)              0 (0%)   
#> 
#> Baseline Age of patient
#>   n                                 59                  41                 100     
#>   Mean (SD)                    40.9 (14.92)        38.9 (14.73)        40.08 (14.8)
#>   Median                            41                  40                 40.5    
#>   Min - Max                      13 - 84             14 - 74             13 - 84   

ASL$SEX[1:10] <- NA

t_summarize_variables(ASL[, c("SEX", "BAGE")], col_by = ASL$ARM, total = "All Patients",
                      useNA_factors = 'ifany')

#>                                   ARM A               ARM B            All Patients
#>                                   (N=59)              (N=41)             (N=100)   
#> -----------------------------------------------------------------------------------
#> 
#>   n                                 54                  36                  90     
#>   M                            24 (40.68%)         16 (39.02%)           40 (40%)  
#>   F                            30 (50.85%)         20 (48.78%)           50 (50%)  
#>   NA                            5 (8.47%)           5 (12.2%)            10 (10%)  
#> 
#> Baseline Age of patient
#>   n                                 59                  41                 100     
#>   Mean (SD)                    40.9 (14.92)        38.9 (14.73)        40.08 (14.8)
#>   Median                            41                  40                 40.5    
#>   Min - Max                      13 - 84             14 - 74             13 - 84   

# with iris data

t_summarize_variables(iris[, -5], col_by  = iris$Species)

#>                       setosa           versicolor          virginica 
#>                       (N=50)             (N=50)             (N=50)   
#> ---------------------------------------------------------------------
#> Sepal.Length
#>   n                     50                 50                 50     
#>   Mean (SD)         5.01 (0.35)        5.94 (0.52)        6.59 (0.64)
#>   Median                 5                 5.9                6.5    
#>   Min - Max          4.3 - 5.8           4.9 - 7           4.9 - 7.9 
#> 
#> Sepal.Width
#>   n                     50                 50                 50     
#>   Mean (SD)         3.43 (0.38)        2.77 (0.31)        2.97 (0.32)
#>   Median                3.4                2.8                 3     
#>   Min - Max          2.3 - 4.4           2 - 3.4           2.2 - 3.8 
#> 
#> Petal.Length
#>   n                     50                 50                 50     
#>   Mean (SD)         1.46 (0.17)        4.26 (0.47)        5.55 (0.55)
#>   Median                1.5               4.35               5.55    
#>   Min - Max           1 - 1.9            3 - 5.1           4.5 - 6.9 
#> 
#> Petal.Width
#>   n                     50                 50                 50     
#>   Mean (SD)         0.25 (0.11)        1.33 (0.2)         2.03 (0.27)
#>   Median                0.2                1.3                 2     
#>   Min - Max          0.1 - 0.6           1 - 1.8           1.4 - 2.5 

t_summarize_variables(iris, col_by  = no_by("All Species"))

#>                     All Species
#>                       (N=150)  
#> -------------------------------
#> Sepal.Length
#>   n                     150    
#>   Mean (SD)         5.84 (0.83)
#>   Median                5.8    
#>   Min - Max          4.3 - 7.9 
#> 
#> Sepal.Width
#>   n                     150    
#>   Mean (SD)         3.06 (0.44)
#>   Median                 3     
#>   Min - Max           2 - 4.4  
#> 
#> Petal.Length
#>   n                     150    
#>   Mean (SD)         3.76 (1.77)
#>   Median               4.35    
#>   Min - Max           1 - 6.9  
#> 
#> Petal.Width
#>   n                     150    
#>   Mean (SD)         1.2 (0.76) 
#>   Median                1.3    
#>   Min - Max          0.1 - 2.5 
#> 
#> Species
#>   n                     150    
#>   setosa            50 (33.33%)
#>   versicolor        50 (33.33%)
#>   virginica         50 (33.33%)