Summary

Summarize an Object for Different Groups

t_summary(x, col_by, col_N = NULL, total = NULL, ...)

Arguments

x
an object to dispatch on, see methods("t_summary")

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

...
(optional)
additional arguments passed on to methods

See also

Examples

t_summary(iris$Sepal.Length, iris$Species)

#>                    setosa           versicolor          virginica 
#>                    (N=50)             (N=50)             (N=50)   
#> ------------------------------------------------------------------
#> n                    50                 50                 50     
#> Mean (SD)        5.01 (0.35)        5.94 (0.52)        6.59 (0.64)
#> Median                5                 5.9                6.5    
#> Min - Max         4.3 - 5.8           4.9 - 7           4.9 - 7.9 

library(random.cdisc.data)
library(dplyr)
ADSL <- radsl(cached = TRUE)

t_summary(x = ADSL$AGE, col_by = ADSL$ARMCD)

#>                     ARM A               ARM B               ARM C    
#>                    (N=134)             (N=134)             (N=132)   
#> ---------------------------------------------------------------------
#> n                    134                 134                 132     
#> Mean (SD)        33.77 (6.55)        35.43 (7.9)         35.43 (7.72)
#> Median                33                  35                  35     
#> Min - Max          21 - 50             21 - 62             20 - 69   

t_summary(x = ADSL[, c("AGE", "SEX", "RACE")], col_by = ADSL$ARMCD)

#>                                                       ARM A               ARM B               ARM C    
#>                                                      (N=134)             (N=134)             (N=132)   
#> -------------------------------------------------------------------------------------------------------
#> Age
#>   n                                                    134                 134                 132     
#>   Mean (SD)                                        33.77 (6.55)        35.43 (7.9)         35.43 (7.72)
#>   Median                                                33                  35                  35     
#>   Min - Max                                          21 - 50             21 - 62             20 - 69   
#> 
#> Sex
#>   n                                                    134                 134                 132     
#>   F                                                79 (58.96%)         82 (61.19%)         70 (53.03%) 
#>   M                                                55 (41.04%)         52 (38.81%)         62 (46.97%) 
#> 
#> Race
#>   n                                                    134                 134                 132     
#>   ASIAN                                            68 (50.75%)           67 (50%)           73 (55.3%) 
#>   BLACK OR AFRICAN AMERICAN                        31 (23.13%)          28 (20.9%)         32 (24.24%) 
#>   WHITE                                            27 (20.15%)          26 (19.4%)         21 (15.91%) 
#>   AMERICAN INDIAN OR ALASKA NATIVE                  8 (5.97%)           11 (8.21%)          6 (4.55%)  
#>   MULTIPLE                                              0               1 (0.75%)               0      
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER             0               1 (0.75%)               0      
#>   OTHER                                                 0                   0                   0      
#>   UNKNOWN                                               0                   0                   0      

t_summary(
  x = ADSL[, c("AGE", "SEX", "RACE")],
  col_by = ADSL$ARMCD,
  total = "All Patients"
)

#>                                                       ARM A               ARM B               ARM C            All Patients
#>                                                      (N=134)             (N=134)             (N=132)             (N=400)   
#> ---------------------------------------------------------------------------------------------------------------------------
#> Age
#>   n                                                    134                 134                 132                 400     
#>   Mean (SD)                                        33.77 (6.55)        35.43 (7.9)         35.43 (7.72)        34.88 (7.44)
#>   Median                                                33                  35                  35                  34     
#>   Min - Max                                          21 - 50             21 - 62             20 - 69             20 - 69   
#> 
#> Sex
#>   n                                                    134                 134                 132                 400     
#>   F                                                79 (58.96%)         82 (61.19%)         70 (53.03%)         231 (57.75%)
#>   M                                                55 (41.04%)         52 (38.81%)         62 (46.97%)         169 (42.25%)
#> 
#> Race
#>   n                                                    134                 134                 132                 400     
#>   ASIAN                                            68 (50.75%)           67 (50%)           73 (55.3%)          208 (52%)  
#>   BLACK OR AFRICAN AMERICAN                        31 (23.13%)          28 (20.9%)         32 (24.24%)         91 (22.75%) 
#>   WHITE                                            27 (20.15%)          26 (19.4%)         21 (15.91%)          74 (18.5%) 
#>   AMERICAN INDIAN OR ALASKA NATIVE                  8 (5.97%)           11 (8.21%)          6 (4.55%)           25 (6.25%) 
#>   MULTIPLE                                              0               1 (0.75%)               0               1 (0.25%)  
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER             0               1 (0.75%)               0               1 (0.25%)  
#>   OTHER                                                 0                   0                   0                   0      
#>   UNKNOWN                                               0                   0                   0                   0      

with(
  ADSL,
  t_summary(AGE > 65, ARMCD)
)

#>                 ARM A               ARM B               ARM C    
#>                (N=134)             (N=134)             (N=132)   
#> -----------------------------------------------------------------
#> n                134                 134                 132     
#> TRUE              0                   0               1 (0.76%)  
#> FALSE         134 (100%)          134 (100%)         131 (99.24%)

# examples with hierarchical header
ADSL <- radsl(cached = TRUE)
t_summary(
  x = ADSL[, c("SEX", "AGE")],
  col_by = ADSL$ARM
)

#>                      A: Drug X             B: Placebo          C: Combination
#>                       (N=134)               (N=134)               (N=132)    
#> -----------------------------------------------------------------------------
#> Sex
#>   n                     134                   134                   132      
#>   F                 79 (58.96%)           82 (61.19%)           70 (53.03%)  
#>   M                 55 (41.04%)           52 (38.81%)           62 (46.97%)  
#> 
#> Age
#>   n                     134                   134                   132      
#>   Mean (SD)         33.77 (6.55)          35.43 (7.9)           35.43 (7.72) 
#>   Median                 33                    35                    35      
#>   Min - Max           21 - 50               21 - 62               20 - 69    

ADSL_f <- ADSL %>%
  dplyr::filter(SEX %in% c("M", "F")) %>%
  mutate(SEX = droplevels(SEX))
t_summary(
  x = ADSL_f$SEX,
  col_by = by_hierarchical(ADSL_f$ARM, ADSL_f$SEX, sep = "-")
)

#>                  A: Drug X                         B: Placebo                      C: Combination      
#>              F                M                F                M                F                M    
#>           (N=79)           (N=55)           (N=82)           (N=52)           (N=70)           (N=62)  
#> -------------------------------------------------------------------------------------------------------
#> n           79               55               82               52               70               62    
#> F        79 (100%)            0            82 (100%)            0            70 (100%)            0    
#> M            0            55 (100%)            0            52 (100%)            0            62 (100%)

t_summary(
  x = ADSL_f[, c("SEX", "AGE")],
  col_by = by_hierarchical(ADSL_f$ARM, ADSL_f$SEX, sep = "-")
)

#>                               A: Drug X                               B: Placebo                            C: Combination         
#>                         F                   M                   F                   M                   F                   M      
#>                       (N=79)              (N=55)              (N=82)              (N=52)              (N=70)              (N=62)   
#> -----------------------------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                     79                  55                  82                  52                  70                  62     
#>   F                 79 (100%)               0               82 (100%)               0               70 (100%)               0      
#>   M                     0               55 (100%)               0               52 (100%)               0               62 (100%)  
#> 
#> Age
#>   n                     79                  55                  82                  52                  70                  62     
#>   Mean (SD)        32.76 (6.09)        35.22 (6.97)        34.24 (6.99)        37.31 (8.9)         35.2 (7.37)         35.69 (8.15)
#>   Median                32                  34                  32                  36                  35                  35     
#>   Min - Max          21 - 47             23 - 50             23 - 58             21 - 62             21 - 64             20 - 69   

ADSL_f <- ADSL %>%
  dplyr::filter(SEX %in% c("M", "F"), RACE %in% c("ASIAN", "WHITE")) %>%
  mutate(
    SEX = droplevels(SEX),
    RACE = droplevels(as.factor(RACE))
  )
t_summary(
  x = ADSL_f[, c("SEX", "RACE")],
  col_by = by_hierarchical(ADSL_f$SEX, ADSL_f$RACE)
)

#>                             F                                   M              
#>                  ASIAN             WHITE             ASIAN             WHITE   
#>                 (N=120)            (N=46)            (N=88)            (N=28)  
#> -------------------------------------------------------------------------------
#> SEX
#>   n               120                46                88                28    
#>   F            120 (100%)        46 (100%)             0                 0     
#>   M                0                 0             88 (100%)         28 (100%) 
#> 
#> RACE
#>   n               120                46                88                28    
#>   ASIAN        120 (100%)            0             88 (100%)             0     
#>   WHITE            0             46 (100%)             0             28 (100%) 

t_summary(
  x = ADSL_f[, c("SEX", "RACE")],
  col_by = by_hierarchical(ADSL_f$RACE, ADSL_f$SEX)
)

#>                           ASIAN                               WHITE            
#>                    F                 M                 F                 M     
#>                 (N=120)            (N=88)            (N=46)            (N=28)  
#> -------------------------------------------------------------------------------
#> SEX
#>   n               120                88                46                28    
#>   F            120 (100%)            0             46 (100%)             0     
#>   M                0             88 (100%)             0             28 (100%) 
#> 
#> RACE
#>   n               120                88                46                28    
#>   ASIAN        120 (100%)        88 (100%)             0                 0     
#>   WHITE            0                 0             46 (100%)         28 (100%) 

t_summary(
  x = ADSL_f[, c("ITTFL", "BMRKR1", "BEP01FL")],
  col_by = by_hierarchical(ADSL_f$RACE, ADSL_f$SEX)
)

#>                                                 ASIAN                                   WHITE              
#>                                         F                   M                   F                   M      
#>                                      (N=120)              (N=88)              (N=46)              (N=28)   
#> -----------------------------------------------------------------------------------------------------------
#> ITTFL
#>   n                                    120                  88                  46                  28     
#>   Y                                 120 (100%)          88 (100%)           46 (100%)           28 (100%)  
#> 
#> Continous Level Biomarker 1
#>   n                                    120                  88                  46                  28     
#>   Mean (SD)                        5.37 (3.27)         5.77 (3.19)         6.46 (4.48)           5.02 (3)  
#>   Median                               4.75                5.07                5.54                4.4     
#>   Min - Max                        0.46 - 15.11        0.41 - 15.07        0.99 - 21.39        1.07 - 12.87
#> 
#> BEP01FL
#>   n                                    120                  88                  46                  28     
#>   Y                                67 (55.83%)         43 (48.86%)           23 (50%)          11 (39.29%) 
#>   N                                53 (44.17%)         45 (51.14%)           23 (50%)          17 (60.71%) 

# add total
t_summary(
  x = ADSL_f[, c("ITTFL", "BMRKR1", "BEP01FL")],
  col_by = by_hierarchical(ADSL_f$RACE, ADSL_f$SEX),
  total = "All"
)

#>                                                 ASIAN                                   WHITE                          All     
#>                                         F                   M                   F                   M                          
#>                                      (N=120)              (N=88)              (N=46)              (N=28)             (N=282)   
#> -------------------------------------------------------------------------------------------------------------------------------
#> ITTFL
#>   n                                    120                  88                  46                  28                 282     
#>   Y                                 120 (100%)          88 (100%)           46 (100%)           28 (100%)           282 (100%) 
#> 
#> Continous Level Biomarker 1
#>   n                                    120                  88                  46                  28                 282     
#>   Mean (SD)                        5.37 (3.27)         5.77 (3.19)         6.46 (4.48)           5.02 (3)          5.63 (3.45) 
#>   Median                               4.75                5.07                5.54                4.4                 4.84    
#>   Min - Max                        0.46 - 15.11        0.41 - 15.07        0.99 - 21.39        1.07 - 12.87        0.41 - 21.39
#> 
#> BEP01FL
#>   n                                    120                  88                  46                  28                 282     
#>   Y                                67 (55.83%)         43 (48.86%)           23 (50%)          11 (39.29%)         144 (51.06%)
#>   N                                53 (44.17%)         45 (51.14%)           23 (50%)          17 (60.71%)         138 (48.94%)