Summarize Date Data

Tabulate the range of dates.

# S3 method for Date
t_summary(x, col_by, col_N = NULL, total = NULL, ...)

Arguments

x
(Date) object

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

...
(optional)
additional arguments passed on to methods.

q_type an integer between 1 and 9 selecting one of the nine quantile algorithms as detailed in quantile(), type argument

conf_level the confidence level between 0 and 1 (default: 0.95)

See also

Examples

today <- Sys.Date()
tenweeks <- seq(today, length.out=10, by="1 week")
t_summary(
  x = tenweeks,
  col_by = by_all("all"),
  col_N = length(tenweeks)
)

#>                                 all           
#>                                (N=10)         
#> ----------------------------------------------
#> range of dates        2020-08-20 to 2020-10-22

t_summary(
  x = tenweeks,
  col_by = factor(LETTERS[c(1,1,1,2,2,3,3,3,4,4)])
)

#>                                  A                               B                               C                               D            
#>                                (N=3)                           (N=2)                           (N=3)                           (N=2)          
#> ----------------------------------------------------------------------------------------------------------------------------------------------
#> range of dates        2020-08-20 to 2020-09-03        2020-09-10 to 2020-09-17        2020-09-24 to 2020-10-08        2020-10-15 to 2020-10-22

t_summary(
  x = tenweeks,
  col_by = factor(LETTERS[c(1,1,1,2,2,3,3,3,4,4)]),
  total = "All"
)

#>                                  A                               B                               C                               D                              All           
#>                                (N=3)                           (N=2)                           (N=3)                           (N=2)                           (N=10)         
#> ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> range of dates        2020-08-20 to 2020-09-03        2020-09-10 to 2020-09-17        2020-09-24 to 2020-10-08        2020-10-15 to 2020-10-22        2020-08-20 to 2020-10-22