Variables Summary Table

Similar as the demographic table in STREAM

# S3 method for data.frame
t_summary(x, col_by, col_N = NULL, total = NULL, ..., table_tree = FALSE)

Arguments

x
(data.frame)
Contains variables to be summarized as described in the details section. If the variable has a label attribute then it will be used for the row name.

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

...
(optional)
additional arguments passed on to methods.

q_type an integer between 1 and 9 selecting one of the nine quantile algorithms as detailed in quantile(), type argument

conf_level the confidence level between 0 and 1 (default: 0.95)

table_tree
(logical value)
If FALSE then a single rtable gets returned, if TRUE a node object gets returned with the elementary rtable objects.

Details

You can use the label attribute of the columns in the data frame to modify the node value / displayed value, e.g. AGE -> "Baseline Age of patient"

Every variable in x will be mapped to a summary table using t_summary and then be stacked.

See also

Examples

library(random.cdisc.data)
library(dplyr)

# with CDISC like data
ADSL <- radsl(cached = TRUE)
t_summary(
  x = ADSL[, c("SEX", "AGE")],
  col_by = ADSL$ARM
)

#>                      A: Drug X             B: Placebo          C: Combination
#>                       (N=134)               (N=134)               (N=132)    
#> -----------------------------------------------------------------------------
#> Sex
#>   n                     134                   134                   132      
#>   F                 79 (58.96%)           82 (61.19%)           70 (53.03%)  
#>   M                 55 (41.04%)           52 (38.81%)           62 (46.97%)  
#> 
#> Age
#>   n                     134                   134                   132      
#>   Mean (SD)         33.77 (6.55)          35.43 (7.9)           35.43 (7.72) 
#>   Median                 33                    35                    35      
#>   Min - Max           21 - 50               21 - 62               20 - 69    

t_summary(
  x = ADSL[, c("SEX", "AGE")],
  col_by = ADSL$ARM,
  total = "All Patients"
)

#>                      A: Drug X             B: Placebo          C: Combination         All Patients 
#>                       (N=134)               (N=134)               (N=132)               (N=400)    
#> ---------------------------------------------------------------------------------------------------
#> Sex
#>   n                     134                   134                   132                   400      
#>   F                 79 (58.96%)           82 (61.19%)           70 (53.03%)           231 (57.75%) 
#>   M                 55 (41.04%)           52 (38.81%)           62 (46.97%)           169 (42.25%) 
#> 
#> Age
#>   n                     134                   134                   132                   400      
#>   Mean (SD)         33.77 (6.55)          35.43 (7.9)           35.43 (7.72)          34.88 (7.44) 
#>   Median                 33                    35                    35                    34      
#>   Min - Max           21 - 50               21 - 62               20 - 69               20 - 69    

# or
t_summary(
  x = ADSL[, c("SEX", "AGE")],
  col_by = ADSL$ARM %>% by_add_total("All Patients")
)

#>                      A: Drug X             B: Placebo          C: Combination         All Patients 
#>                       (N=134)               (N=134)               (N=132)               (N=400)    
#> ---------------------------------------------------------------------------------------------------
#> Sex
#>   n                     134                   134                   132                   400      
#>   F                 79 (58.96%)           82 (61.19%)           70 (53.03%)           231 (57.75%) 
#>   M                 55 (41.04%)           52 (38.81%)           62 (46.97%)           169 (42.25%) 
#> 
#> Age
#>   n                     134                   134                   132                   400      
#>   Mean (SD)         33.77 (6.55)          35.43 (7.9)           35.43 (7.72)          34.88 (7.44) 
#>   Median                 33                    35                    35                    34      
#>   Min - Max           21 - 50               21 - 62               20 - 69               20 - 69    

# Special Variants
t_summary(
  x = ADSL[, c("SEX", "AGE")],
  col_by = by_all("All"),
  col_N = nrow(ADSL)
)

#>                        All     
#>                      (N=400)   
#> -------------------------------
#> Sex
#>   n                    400     
#>   F                231 (57.75%)
#>   M                169 (42.25%)
#> 
#> Age
#>   n                    400     
#>   Mean (SD)        34.88 (7.44)
#>   Median                34     
#>   Min - Max          20 - 69   

# control the label
ADSL <- var_relabel(ADSL, AGE = "Baseline Age of patient")

# control categorical order
ADSL$SEX <- relevel(ADSL$SEX, "M", "F")

t_summary(
  x = ADSL[, c("SEX", "AGE")],
  col_by = ADSL$ARM,
  total = "All Patients"
)

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                   (N=134)               (N=134)               (N=132)               (N=400)    
#> ---------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                                 134                   134                   132                   400      
#>   M                             55 (41.04%)           52 (38.81%)           62 (46.97%)           169 (42.25%) 
#>   F                             79 (58.96%)           82 (61.19%)           70 (53.03%)           231 (57.75%) 
#> 
#> Baseline Age of patient
#>   n                                 134                   134                   132                   400      
#>   Mean (SD)                     33.77 (6.55)          35.43 (7.9)           35.43 (7.72)          34.88 (7.44) 
#>   Median                             33                    35                    35                    34      
#>   Min - Max                       21 - 50               21 - 62               20 - 69               20 - 69    

t_summary(
  x = ADSL[, c("SEX", "AGE")],
  col_by = ADSL$ARM,
  total = "All Patients",
  useNA = "always"
)

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                   (N=134)               (N=134)               (N=132)               (N=400)    
#> ---------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                                 134                   134                   132                   400      
#>   M                             55 (41.04%)           52 (38.81%)           62 (46.97%)           169 (42.25%) 
#>   F                             79 (58.96%)           82 (61.19%)           70 (53.03%)           231 (57.75%) 
#>   <NA>                               0                     0                     0                     0       
#> 
#> Baseline Age of patient
#>   n                                 134                   134                   132                   400      
#>   Mean (SD)                     33.77 (6.55)          35.43 (7.9)           35.43 (7.72)          34.88 (7.44) 
#>   Median                             33                    35                    35                    34      
#>   Min - Max                       21 - 50               21 - 62               20 - 69               20 - 69    

# use different col_by

## Biomarker Evaluable Population
t_summary(
  x = ADSL[, c("SEX", "AGE")],
  col_by = ADSL$ARM %>%
    by_compare_subset(
      ADSL$BEP01FL == "Y",
      label_all = "ALL",
      label_subset = "BEP"
    )
)

#> Description of Planned Arm                   A: Drug X                               B: Placebo                            C: Combination         
#>                                       ALL                 BEP                 ALL                 BEP                 ALL                 BEP     
#>                                     (N=134)              (N=68)             (N=134)              (N=63)             (N=132)              (N=66)   
#> --------------------------------------------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                                   134                  68                 134                  63                 132                  66     
#>   M                               55 (41.04%)         24 (35.29%)         52 (38.81%)         21 (33.33%)         62 (46.97%)         26 (39.39%) 
#>   F                               79 (58.96%)         44 (64.71%)         82 (61.19%)         42 (66.67%)         70 (53.03%)         40 (60.61%) 
#> 
#> Baseline Age of patient
#>   n                                   134                  68                 134                  63                 132                  66     
#>   Mean (SD)                       33.77 (6.55)        32.94 (6.57)        35.43 (7.9)         35.73 (8.29)        35.43 (7.72)        35.09 (8.12)
#>   Median                               33                  32                  35                  35                  35                  34     
#>   Min - Max                         21 - 50             21 - 50             21 - 62             23 - 62             20 - 69             21 - 69   

# With Missing Data
ADSL$SEX[1:10] <- NA
t_summary(
  x = ADSL[, c("SEX", "AGE")],
  col_by = ADSL$ARM,
  total = "All Patients",
  useNA = 'ifany'
)

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                   (N=134)               (N=134)               (N=132)               (N=400)    
#> ---------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                                 132                   130                   128                   390      
#>   M                              54 (40.3%)           50 (37.31%)           60 (45.45%)            164 (41%)   
#>   F                             78 (58.21%)            80 (59.7%)           68 (51.52%)           226 (56.5%)  
#>   <NA>                           2 (1.49%)             4 (2.99%)             4 (3.03%)             10 (2.5%)   
#> 
#> Baseline Age of patient
#>   n                                 134                   134                   132                   400      
#>   Mean (SD)                     33.77 (6.55)          35.43 (7.9)           35.43 (7.72)          34.88 (7.44) 
#>   Median                             33                    35                    35                    34      
#>   Min - Max                       21 - 50               21 - 62               20 - 69               20 - 69    

t_summary(
  x = ADSL[, c("SEX", "AGE")],
  col_by = ADSL$ARM,
  total = "All Patients",
  denominator = "N",
  useNA = "ifany"
)

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                   (N=134)               (N=134)               (N=132)               (N=400)    
#> ---------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                                 132                   130                   128                   390      
#>   M                              54 (40.3%)           50 (37.31%)           60 (45.45%)            164 (41%)   
#>   F                             78 (58.21%)            80 (59.7%)           68 (51.52%)           226 (56.5%)  
#>   <NA>                           2 (1.49%)             4 (2.99%)             4 (3.03%)             10 (2.5%)   
#> 
#> Baseline Age of patient
#>   n                                 134                   134                   132                   400      
#>   Mean (SD)                     33.77 (6.55)          35.43 (7.9)           35.43 (7.72)          34.88 (7.44) 
#>   Median                             33                    35                    35                    34      
#>   Min - Max                       21 - 50               21 - 62               20 - 69               20 - 69    

# Table tree
tbls <- t_summary(
  ADSL[, c("SEX", "AGE")],
  col_by = ADSL$ARM,
  table_tree = TRUE
)
summary(tbls)

#>                                    index            content             dim            hide label          format  
#> -------------------------------------------------------------------------------------------------------------------
#> root                                                  NULL                                TRUE                     
#>   SEX                                1               rtable            4 x 3             FALSE                     
#>   Baseline Age of patient            2               rtable            4 x 3             FALSE                     

to_rtable(tbls)

#>                                  A: Drug X             B: Placebo          C: Combination
#>                                   (N=134)               (N=134)               (N=132)    
#> -----------------------------------------------------------------------------------------
#> SEX
#>   n                                 132                   130                   128      
#>   M                              54 (40.3%)           50 (37.31%)           60 (45.45%)  
#>   F                             78 (58.21%)            80 (59.7%)           68 (51.52%)  
#>   <NA>                           2 (1.49%)             4 (2.99%)             4 (3.03%)   
#> 
#> Baseline Age of patient
#>   n                                 134                   134                   132      
#>   Mean (SD)                     33.77 (6.55)          35.43 (7.9)           35.43 (7.72) 
#>   Median                             33                    35                    35      
#>   Min - Max                       21 - 50               21 - 62               20 - 69