Return an rtable with a no Method Message

If there is no explicit method for an object an rtable with one row with an appropriate message is returned.

# S3 method for default
t_summary(x, col_by, col_N = NULL, total = NULL, ...)

Arguments

x
(data.frame)
Contains variables to be summarized as described in the details section. If the variable has a label attribute then it will be used for the row name.

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

...
(not used with this object)
not used arguments

Value

rtable

Examples

t_summary(
  x = structure(1:5, class = "aaa"),
  col_by = factor(LETTERS[c(1,2,1,1,2)])
)

#>                                             A            B  
#>                                           (N=3)        (N=2)
#> ------------------------------------------------------------
#> no t_summary method for class: aaa          -            -