Summarize Categorical Data

Tabulate the number of non-missing observations, number of observations per level and percentage.

# S3 method for factor
t_summary(x, col_by, col_N = table(col_by),
  total = NULL, useNA = c("no", "ifany", "always"),
  denominator = c("n", "N"), drop_levels = FALSE, ...)

Arguments

x
factor variable

col_by
a factor of length nrow(x) that defines which levels in col_by define a column. If data should not be split into columns use the no_by function.

col_N
The column total for each group that is displayed in the table header with (N=xx).

total
character string that will be used as a label for a column with pooled total population. If the levels of col_by are the only columns of interest then total should be NULL.

useNA
choose whether missing data (NAs) should be displayed as a level.

denominator
either n or N for calculating the level associated percentage. With option N, the reference population from col_N is used as the denominator. With option n, the number of non-missing records from x is used as the denominator.

drop_levels
boolean whether to drop zero count levels

...
not used arguments

Value

an rtable object

See also

Examples

# with iris data
t_summary(iris$Species, iris$Species)

#>                     setosa          versicolor        virginica 
#>                     (N=50)            (N=50)            (N=50)  
#> ----------------------------------------------------------------
#> n                     50                50                50    
#> setosa            50 (100%)             -                 -     
#> versicolor            -             50 (100%)             -     
#> virginica             -                 -             50 (100%) 

library(random.cdisc.data)
ADSL <- radsl(N = 100, seed = 1)

t_summary(ADSL$SEX, ADSL$ARM, total = "All")

#>                           A: Drug X             B: Placebo          C: Combination             All      
#>                             (N=23)                (N=42)                (N=35)               (N=100)    
#> --------------------------------------------------------------------------------------------------------
#> n                             23                    42                    35                   100      
#> F                        13 (56.52%)             21 (50%)            19 (54.29%)             53 (53%)   
#> M                         9 (39.13%)             21 (50%)            15 (42.86%)             45 (45%)   
#> U                         1 (4.35%)                 -                 1 (2.86%)               2 (2%)    
#> UNDIFFERENTIATED              -                     -                     -                     -       

t_summary(ADSL$SEX, ADSL$ARM, useNA = "always")

#>                           A: Drug X             B: Placebo          C: Combination
#>                             (N=23)                (N=42)                (N=35)    
#> ----------------------------------------------------------------------------------
#> n                             23                    42                    35      
#> F                        13 (56.52%)             21 (50%)            19 (54.29%)  
#> M                         9 (39.13%)             21 (50%)            15 (42.86%)  
#> U                         1 (4.35%)                 -                 1 (2.86%)   
#> UNDIFFERENTIATED              -                     -                     -       
#> <NA>                          0                     0                     0       

ADSL$SEX[1:10] <- NA
t_summary(ADSL$SEX, ADSL$ARM, denominator = "N", useNA = "ifany", total = "All")

#>                           A: Drug X             B: Placebo          C: Combination             All      
#>                             (N=23)                (N=42)                (N=35)               (N=100)    
#> --------------------------------------------------------------------------------------------------------
#> n                             23                    36                    31                    90      
#> F                        13 (56.52%)           20 (47.62%)           17 (48.57%)             50 (50%)   
#> M                         9 (39.13%)            16 (38.1%)           13 (37.14%)             38 (38%)   
#> U                         1 (4.35%)                 -                 1 (2.86%)               2 (2%)    
#> UNDIFFERENTIATED              -                     -                     -                     -       
#> <NA>                          0                     6                     4                     10      

t_summary(ADSL$SEX, ADSL$ARM, denominator = "n", useNA = "no", total = "All")

#>                           A: Drug X             B: Placebo          C: Combination             All      
#>                             (N=23)                (N=42)                (N=35)               (N=100)    
#> --------------------------------------------------------------------------------------------------------
#> n                             23                    36                    31                    90      
#> F                        13 (56.52%)           20 (55.56%)           17 (54.84%)           50 (55.56%)  
#> M                         9 (39.13%)           16 (44.44%)           13 (41.94%)           38 (42.22%)  
#> U                         1 (4.35%)                 -                 1 (3.23%)             2 (2.22%)   
#> UNDIFFERENTIATED              -                     -                     -                     -