Summarize Boolean Data

Boolean data will be converted to factor

# S3 method for logical
t_summary(x, col_by, col_N = table(col_by),
  total = NULL, row_name_true = "TRUE", row_name_false = "FALSE",
  ...)

Arguments

x
a logical vector

col_by
factor with reference and comparison group information, the first level indicates the reference group.

col_N
a vecor with the reference population per level of col_by. This often by default

total
character string that will be used as a label for a column with total population. If the levels of col_by are the only columns of interest then total should be NULL.

row_name_true
character string with row.name for TRUE summary

row_name_false
character string with row.name for FALSE summary

...
arguments passed on to t_summary.factor

See also

Examples

t_summary(
 x = c(TRUE,FALSE,NA,TRUE,FALSE,FALSE,FALSE,TRUE),
 col_by = factor(LETTERS[c(1,1,1,2,2,3,3,2)])
)

#>                  A                 B                 C     
#>                (N=3)             (N=3)             (N=2)   
#> -----------------------------------------------------------
#> n                2                 3                 2     
#> TRUE          1 (50%)          2 (66.67%)            -     
#> FALSE         1 (50%)          1 (33.33%)         2 (100%) 

library(random.cdisc.data)
ADSL <- cadsl

ADSL$AGE[1:10] <- NA

with(ADSL, t_summary(AGE > 65, ARM, useNA = "ifany"))

#>                A: Drug X             B: Placebo          C: Combination
#>                 (N=142)               (N=133)               (N=125)    
#> -----------------------------------------------------------------------
#> n                 139                   129                   122      
#> TRUE           12 (8.63%)            10 (7.75%)           18 (14.75%)  
#> FALSE         127 (91.37%)          119 (92.25%)          104 (85.25%) 
#> <NA>               3                     4                     3       

with(ADSL, t_summary(AGE > 65, ARM, denominator = "N", total = "All", useNA = "ifany",
                     row.name.TRUE = "Baseline Age > 65", row.name.FALSE = "Baseline Age <= 65"))

#>                A: Drug X             B: Placebo          C: Combination             All      
#>                 (N=142)               (N=133)               (N=125)               (N=400)    
#> ---------------------------------------------------------------------------------------------
#> n                 139                   129                   122                   390      
#> TRUE           12 (8.45%)            10 (7.52%)            18 (14.4%)             40 (10%)   
#> FALSE         127 (89.44%)          119 (89.47%)          104 (83.2%)           350 (87.5%)  
#> <NA>               3                     4                     3                     10