Summarize Numeric Variables

Tabulate the number of non-missing observations, mean, standard error, confidence interval, median, and range for different groups.

# S3 method for numeric
t_summary(
  x,
  col_by,
  col_N = NULL,
  total = NULL,
  f_numeric = c("count_n", "mean_sd", "median", "range"),
  ...
)

Arguments

x
(numeric) variable

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

f_numeric
a combination of the analysis functions to be evaluated "count_n", "mean_sd", "ci", "median", "q1_q3", "range", "se", as well as the functions which are wrapped by patient_numeric_fcns (f_numeric = patient_numeric_fcns()) for summarizing patient data as for t_summarize_by_visit

...
(optional)
additional arguments passed on to methods.

q_type an integer between 1 and 9 selecting one of the nine quantile algorithms as detailed in quantile(), type argument

conf_level the confidence level between 0 and 1 (default: 0.95)

Value

an rtable object

Details

The analysis function "ci" uses a two-sided confidence interval based on the t-test statistic, see ttest_ci_one_arm for details. By default, the confidence level is 95 with the optional `conf_level` parameter.

See also

Examples

# with iris data
t_summary(x = iris$Sepal.Length, col_by = iris$Species)

#>                    setosa           versicolor          virginica 
#>                    (N=50)             (N=50)             (N=50)   
#> ------------------------------------------------------------------
#> n                    50                 50                 50     
#> Mean (SD)        5.01 (0.35)        5.94 (0.52)        6.59 (0.64)
#> Median                5                 5.9                6.5    
#> Min - Max         4.3 - 5.8           4.9 - 7           4.9 - 7.9 

library(random.cdisc.data)
ADSL <- cadsl

t_summary(x = ADSL$AGE, col_by = ADSL$ARM)

#>                    A: Drug X             B: Placebo          C: Combination
#>                     (N=134)               (N=134)               (N=132)    
#> ---------------------------------------------------------------------------
#> n                     134                   134                   132      
#> Mean (SD)         33.77 (6.55)          35.43 (7.9)           35.43 (7.72) 
#> Median                 33                    35                    35      
#> Min - Max           21 - 50               21 - 62               20 - 69    

t_summary(
  x = ADSL$AGE,
  col_by = ADSL$ARM,
  f_numeric = c("count_n", "mean_sd", "se", "ci")
)

#>                    A: Drug X             B: Placebo          C: Combination
#>                     (N=134)               (N=134)               (N=132)    
#> ---------------------------------------------------------------------------
#> n                     134                   134                   132      
#> Mean (SD)         33.77 (6.55)          35.43 (7.9)           35.43 (7.72) 
#> SE                    0.57                  0.68                  0.67     
#> 95% CI           32.65 - 34.89         34.08 - 36.78          34.1 - 36.76 

t_summary(
  x = ADSL$AGE,
  col_by = col_by_to_matrix(ADSL$ARM),
  total = "All"
)

#>                    A: Drug X             B: Placebo          C: Combination             All      
#>                     (N=134)               (N=134)               (N=132)               (N=400)    
#> -------------------------------------------------------------------------------------------------
#> n                     134                   134                   132                   400      
#> Mean (SD)         33.77 (6.55)          35.43 (7.9)           35.43 (7.72)          34.88 (7.44) 
#> Median                 33                    35                    35                    34      
#> Min - Max           21 - 50               21 - 62               20 - 69               20 - 69    

t_summary(
  x = ADSL$AGE,
  col_by = ADSL$ARM,
  total = "All"
)

#>                    A: Drug X             B: Placebo          C: Combination             All      
#>                     (N=134)               (N=134)               (N=132)               (N=400)    
#> -------------------------------------------------------------------------------------------------
#> n                     134                   134                   132                   400      
#> Mean (SD)         33.77 (6.55)          35.43 (7.9)           35.43 (7.72)          34.88 (7.44) 
#> Median                 33                    35                    35                    34      
#> Min - Max           21 - 50               21 - 62               20 - 69               20 - 69    

ADSL$AGE[1:10] <- NA
t_summary(
  x = ADSL$AGE,
  col_by = by_all("All"),
  col_N = nrow(ADSL)
)

#>                      All     
#>                    (N=400)   
#> -----------------------------
#> n                    390     
#> Mean (SD)        34.93 (7.43)
#> Median                34     
#> Min - Max          20 - 69