Summarize an Object for Different Groups with by Variable

Summarize an Object for Different Groups with by Variable

t_summary_by(x, by, col_by, col_N, total = NULL, ...)

Arguments

x
vector

by
a factor of length nrow(x) with no missing values. The levels of by define the summary sub-groups in the table.

col_by
a factor of length nrow(x) with no missing values. The levels of col_by define the columns in the table.

col_N
a table object with the reference population used for the header of the table. See examples below.

total
character string that will be used as a label for a column with pooled total population. If the levels of col_by are the only columns of interest then total should be NULL.

...
arguments passed on to methods

Details

For every level of the variable by a summary table using t_summary will be created. The individual tables are then stacked together.

See also

Examples

library(dplyr)
library(random.cdisc.data)

ADSL <- cadsl

t_summary_by(
 x = ADSL$SEX,
 by = ADSL$COUNTRY,
 col_by = ADSL$ARMCD,
 col_N = table(ADSL$ARMCD),
 total = "All Patients",
 drop_levels = TRUE
)

#> Country                      ARM A               ARM B               ARM C            All Patients
#>   Sex                       (N=142)             (N=133)             (N=125)             (N=400)   
#> --------------------------------------------------------------------------------------------------
#> CHN
#>   n                            70                  73                  76                 219     
#>   F                       34 (48.57%)         38 (52.05%)         44 (57.89%)         116 (52.97%)
#>   M                         35 (50%)          34 (46.58%)         28 (36.84%)         97 (44.29%) 
#>   U                            -               1 (1.37%)           4 (5.26%)           5 (2.28%)  
#>   UNDIFFERENTIATED         1 (1.43%)               -                   -               1 (0.46%)  
#> 
#> USA
#>   n                            18                  13                  9                   40     
#>   F                        5 (27.78%)          6 (46.15%)          2 (22.22%)          13 (32.5%) 
#>   M                       12 (66.67%)          6 (46.15%)          7 (77.78%)          25 (62.5%) 
#>   U                        1 (5.56%)           1 (7.69%)               -                 2 (5%)   
#> 
#> BRA
#>   n                            11                  9                   10                  30     
#>   F                        5 (45.45%)          4 (44.44%)           4 (40%)           13 (43.33%) 
#>   M                        6 (54.55%)          5 (55.56%)           6 (60%)           17 (56.67%) 
#> 
#> PAK
#>   n                            13                  11                  7                   31     
#>   F                        6 (46.15%)          5 (45.45%)          4 (57.14%)         15 (48.39%) 
#>   M                        6 (46.15%)          6 (54.55%)          3 (42.86%)         15 (48.39%) 
#>   UNDIFFERENTIATED         1 (7.69%)               -                   -               1 (3.23%)  
#> 
#> NGA
#>   n                            11                  7                   8                   26     
#>   F                        6 (54.55%)          2 (28.57%)           2 (25%)           10 (38.46%) 
#>   M                        5 (45.45%)          4 (57.14%)           6 (75%)           15 (57.69%) 
#>   U                            -               1 (14.29%)              -               1 (3.85%)  
#> 
#> RUS
#>   n                            6                   8                   5                   19     
#>   F                        4 (66.67%)           4 (50%)             5 (100%)          13 (68.42%) 
#>   M                        1 (16.67%)           4 (50%)                -               5 (26.32%) 
#>   U                        1 (16.67%)              -                   -               1 (5.26%)  
#> 
#> JPN
#>   n                            5                   6                   7                   18     
#>   F                         3 (60%)            5 (83.33%)          5 (71.43%)         13 (72.22%) 
#>   M                         1 (20%)            1 (16.67%)          2 (28.57%)          4 (22.22%) 
#>   UNDIFFERENTIATED          1 (20%)                -                   -               1 (5.56%)  
#> 
#> GBR
#>   n                            3                   3                   3                   9      
#>   F                        2 (66.67%)          2 (66.67%)          2 (66.67%)          6 (66.67%) 
#>   M                        1 (33.33%)          1 (33.33%)          1 (33.33%)          3 (33.33%) 
#> 
#> CAN
#>   n                            5                   3                   0                   8      
#>   F                         4 (80%)            1 (33.33%)              -               5 (62.5%)  
#>   M                         1 (20%)            2 (66.67%)              -               3 (37.5%)  
#> 
#> CHE
#>   n                            0                   0                   0                   0      

ADSL$SEX[1:5] <- NA

t_summary_by(
 x = ADSL$SEX,
 by = ADSL$COUNTRY,
 col_by = ADSL$ARMCD,
 col_N = table(ADSL$ARMCD),
 total = "All Patients",
 drop_levels = TRUE,
 useNA = "ifany"
)

#> Country                      ARM A               ARM B               ARM C            All Patients
#>   Sex                       (N=142)             (N=133)             (N=125)             (N=400)   
#> --------------------------------------------------------------------------------------------------
#> CHN
#>   n                            69                  73                  74                 216     
#>   F                       34 (49.28%)         38 (52.05%)         42 (56.76%)         114 (52.78%)
#>   M                       34 (49.28%)         34 (46.58%)         28 (37.84%)         96 (44.44%) 
#>   U                            -               1 (1.37%)           4 (5.41%)           5 (2.31%)  
#>   UNDIFFERENTIATED         1 (1.45%)               -                   -               1 (0.46%)  
#>   <NA>                         1                   0                   2                   3      
#> 
#> USA
#>   n                            18                  13                  8                   39     
#>   F                        5 (27.78%)          6 (46.15%)          1 (12.5%)          12 (30.77%) 
#>   M                       12 (66.67%)          6 (46.15%)          7 (87.5%)           25 (64.1%) 
#>   U                        1 (5.56%)           1 (7.69%)               -               2 (5.13%)  
#>   <NA>                         0                   0                   1                   1      
#> 
#> BRA
#>   n                            11                  9                   10                  30     
#>   F                        5 (45.45%)          4 (44.44%)           4 (40%)           13 (43.33%) 
#>   M                        6 (54.55%)          5 (55.56%)           6 (60%)           17 (56.67%) 
#> 
#> PAK
#>   n                            13                  11                  7                   31     
#>   F                        6 (46.15%)          5 (45.45%)          4 (57.14%)         15 (48.39%) 
#>   M                        6 (46.15%)          6 (54.55%)          3 (42.86%)         15 (48.39%) 
#>   UNDIFFERENTIATED         1 (7.69%)               -                   -               1 (3.23%)  
#> 
#> NGA
#>   n                            11                  7                   8                   26     
#>   F                        6 (54.55%)          2 (28.57%)           2 (25%)           10 (38.46%) 
#>   M                        5 (45.45%)          4 (57.14%)           6 (75%)           15 (57.69%) 
#>   U                            -               1 (14.29%)              -               1 (3.85%)  
#> 
#> RUS
#>   n                            6                   7                   5                   18     
#>   F                        4 (66.67%)          4 (57.14%)           5 (100%)          13 (72.22%) 
#>   M                        1 (16.67%)          3 (42.86%)              -               4 (22.22%) 
#>   U                        1 (16.67%)              -                   -               1 (5.56%)  
#>   <NA>                         0                   1                   0                   1      
#> 
#> JPN
#>   n                            5                   6                   7                   18     
#>   F                         3 (60%)            5 (83.33%)          5 (71.43%)         13 (72.22%) 
#>   M                         1 (20%)            1 (16.67%)          2 (28.57%)          4 (22.22%) 
#>   UNDIFFERENTIATED          1 (20%)                -                   -               1 (5.56%)  
#> 
#> GBR
#>   n                            3                   3                   3                   9      
#>   F                        2 (66.67%)          2 (66.67%)          2 (66.67%)          6 (66.67%) 
#>   M                        1 (33.33%)          1 (33.33%)          1 (33.33%)          3 (33.33%) 
#> 
#> CAN
#>   n                            5                   3                   0                   8      
#>   F                         4 (80%)            1 (33.33%)              -               5 (62.5%)  
#>   M                         1 (20%)            2 (66.67%)              -               3 (37.5%)  
#> 
#> CHE
#>   n                            0                   0                   0                   0      

ADSL <- ADSL %>% select(STUDYID, USUBJID, ARMCD)

ADQS <- cadqs
ADQS_f <- ADQS %>%
  dplyr::filter(PARAMCD=="BFIALL")

t_summary_by(
 x = ADQS_f$AVAL,
 by = ADQS_f$AVISIT,
 col_by = ADQS_f$ARMCD,
 col_N = table(ADSL$ARMCD),
 total = "All Patients"
)

#> ADQS_f$AVISIT               ARM A                ARM B                ARM C            All Patients 
#>   Analysis Value           (N=142)              (N=133)              (N=125)              (N=400)   
#> ----------------------------------------------------------------------------------------------------
#> SCREENING
#>   n                          142                  133                  125                  400     
#>   Mean (SD)              50.3 (8.04)          50.47 (7.5)         50.14 (7.75)          50.3 (7.76) 
#>   Median                    49.71                49.89                49.57                49.79    
#>   Min - Max             26.89 - 71.19        32.28 - 69.98        29.26 - 67.65        26.89 - 71.19
#> 
#> BASELINE
#>   n                          142                  133                  125                  400     
#>   Mean (SD)             50.47 (8.42)          49.83 (8.3)         49.61 (8.09)         49.99 (8.27) 
#>   Median                    50.11                49.14                48.98                49.4     
#>   Min - Max             26.04 - 71.47        28.48 - 68.28        25.82 - 69.16        25.82 - 71.47
#> 
#> WEEK 1 DAY 8
#>   n                          142                  133                  125                  400     
#>   Mean (SD)             50.34 (8.26)         50.22 (7.97)         49.11 (7.75)           49.92 (8)  
#>   Median                    50.59                50.21                50.27                50.27    
#>   Min - Max             28.07 - 70.52        20.63 - 69.08        29.02 - 67.7         20.63 - 70.52
#> 
#> WEEK 2 DAY 15
#>   n                          142                  133                  125                  400     
#>   Mean (SD)             50.13 (8.32)         49.78 (8.29)         49.67 (9.37)         49.87 (8.63) 
#>   Median                    49.67                48.69                49.6                 49.47    
#>   Min - Max             33.89 - 78.99        32.03 - 74.24        25.52 - 72.29        25.52 - 78.99
#> 
#> WEEK 3 DAY 22
#>   n                          142                  133                  125                  400     
#>   Mean (SD)             49.84 (8.39)         50.16 (7.83)         49.29 (8.69)         49.78 (8.29) 
#>   Median                    49.85                49.72                50.08                49.79    
#>   Min - Max             32.16 - 77.02        30.81 - 67.65          22.4 - 71          22.4 - 77.02 
#> 
#> WEEK 4 DAY 29
#>   n                          142                  133                  125                  400     
#>   Mean (SD)             50.18 (7.24)         49.72 (7.62)          50.28 (7.6)         50.06 (7.46) 
#>   Median                    50.24                49.34                50.42                49.68    
#>   Min - Max             27.84 - 72.97        31.57 - 68.77        27.72 - 71.02        27.72 - 72.97
#> 
#> WEEK 5 DAY 36
#>   n                          142                  133                  125                  400     
#>   Mean (SD)              50.36 (7.4)         49.46 (7.52)         49.52 (8.33)          49.8 (7.73) 
#>   Median                    50.43                49.21                48.37                49.27    
#>   Min - Max             34.41 - 69.3         32.71 - 71.36        28.75 - 68.45        28.75 - 71.36

ADQS_f$AVALCAT1 <- factor(ifelse(ADQS_f$AVAL >= 0, "Positive", "Negative"),
  levels = c("Positive", "Negative"))
ADQS_f <- var_relabel(ADQS_f, AVALCAT1 = "Result" )

t_summary_by(
 x = ADQS_f$AVALCAT1,
 by = ADQS_f$AVISIT,
 col_by = ADQS_f$ARMCD,
 col_N = table(ADSL$ARMCD),
 total = "All Patients"
)

#> ADQS_f$AVISIT           ARM A               ARM B               ARM C            All Patients
#>   Result               (N=142)             (N=133)             (N=125)             (N=400)   
#> ---------------------------------------------------------------------------------------------
#> SCREENING
#>   n                      142                 133                 125                 400     
#>   Positive            142 (100%)          133 (100%)          125 (100%)          400 (100%) 
#>   Negative                -                   -                   -                   -      
#> 
#> BASELINE
#>   n                      142                 133                 125                 400     
#>   Positive            142 (100%)          133 (100%)          125 (100%)          400 (100%) 
#>   Negative                -                   -                   -                   -      
#> 
#> WEEK 1 DAY 8
#>   n                      142                 133                 125                 400     
#>   Positive            142 (100%)          133 (100%)          125 (100%)          400 (100%) 
#>   Negative                -                   -                   -                   -      
#> 
#> WEEK 2 DAY 15
#>   n                      142                 133                 125                 400     
#>   Positive            142 (100%)          133 (100%)          125 (100%)          400 (100%) 
#>   Negative                -                   -                   -                   -      
#> 
#> WEEK 3 DAY 22
#>   n                      142                 133                 125                 400     
#>   Positive            142 (100%)          133 (100%)          125 (100%)          400 (100%) 
#>   Negative                -                   -                   -                   -      
#> 
#> WEEK 4 DAY 29
#>   n                      142                 133                 125                 400     
#>   Positive            142 (100%)          133 (100%)          125 (100%)          400 (100%) 
#>   Negative                -                   -                   -                   -      
#> 
#> WEEK 5 DAY 36
#>   n                      142                 133                 125                 400     
#>   Positive            142 (100%)          133 (100%)          125 (100%)          400 (100%) 
#>   Negative                -                   -                   -                   -