Summarize the TRUE Counts

Note that for this function the default denominator = "N" and not n as for the other t_summary functions.

t_summary_true(
  x,
  col_by,
  col_N = NULL,
  total = NULL,
  row_name = deparse(substitute(x)),
  denominator = c("N", "n", "omit")
)

Arguments

x
factor variable

col_by
logical factor or data.frame with reference and comparison group information, the first level indicates the reference group. See the functions factor_to_matrix_by and add_total_by.

col_N
a vecor with the reference population per level of col_by. This often by default get_N(col_by). Names of the list are ignored, they are added in the order they appear.

total
string of column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

row_name
name of row

denominator
either n, N or omit. n and N are for calculating the level associated percentage. With option N, the reference population from col_N is used as the denominator. With option n, the number of non-missing records from x is used as the denominator. If omit is chosen the percentage is omitted.

Examples

with(iris, t_summary_true(Sepal.Length > 5.5, col_by = Species, total = "All Flowers"))

#>                             setosa           versicolor          virginica         All Flowers
#>                             (N=50)             (N=50)             (N=50)             (N=150)  
#> ----------------------------------------------------------------------------------------------
#> Sepal.Length > 5.5          3 (6%)            39 (78%)           49 (98%)          91 (60.67%)