Tabulation of MMRM Results

These functions can be used to produce tables from a fitted MMRM produced with fit_mmrm().

# S3 method for mmrm
as.rtable(x, type = c("fixed", "cov", "diagnostic"), ...)

h_mmrm_fixed(x, format = "xx.xxxx")

h_mmrm_cov(x, format = "xx.xxxx")

h_mmrm_diagnostic(x, format = "xx.xxxx")

# S3 method for mmrm
tidy(x)

s_mmrm_lsmeans(df, .in_ref_col, show_relative = c("reduction", "increase"))

s_mmrm_lsmeans_single(df)

summarize_lsmeans(
  lyt,
  arms = TRUE,
  ...,
  table_names = "lsmeans_summary",
  .stats = NULL,
  .formats = NULL,
  .indent_mods = NULL,
  .labels = NULL
)

Arguments

x

(mmrm)
the original MMRM fit object.

type

(string)
type of table which should be returned.

...

additional argument format for controlling the numeric format.

format

(string)
format for the numbers in the table.

df

(data frame)
data set containing all analysis variables.

.in_ref_col

(logical)
TRUE when working with the reference level, FALSE otherwise.

show_relative

should the "reduction" (control - treatment, default) or the "increase" (treatment - control) be shown for the relative change from baseline?

lyt

(layout)
input layout where analyses will be added to.

arms

(flag)
should treatment variable be considered when using summarize_lsmeans layout generating function.

table_names

(character)
this can be customized in case that the same vars are analyzed multiple times, to avoid warnings from rtables.

.stats

(character)
statistics to select for the table.

.formats

(named character or list)
formats for the statistics.

.indent_mods

(named integer)
indent modifiers for the labels.

.labels

(named character)
labels for the statistics (without indent).

Value

as.rtable.mmrm() returns the fixed effects, covariance estimate or diagnostic statistics tables.

Functions

as.rtable.mmrm: Deprecated by tern.mmrm::tabulate_mmrm: Produce simple MMRM tables via the generic as.rtable().

h_mmrm_fixed: Deprecated by tern.mmrm::tabulate_mmrm: Helper function to produce fixed effects table.

h_mmrm_cov: Deprecated by tern.mmrm::tabulate_mmrm: Helper function to produce covariance matrix table.

h_mmrm_diagnostic: Deprecated by tern.mmrm::tabulate_mmrm: Helper function to produce diagnostic statistics table.

tidy.mmrm: Deprecated by tern.mmrm::tabulate_mmrm: Helper method (for broom::tidy()) to prepare a data frame from an mmrm object containing the LS means and contrasts.

s_mmrm_lsmeans: Deprecated by tern.mmrm::tabulate_mmrm: Statistics function which is extracting estimates from a tidied LS means data frame.

s_mmrm_lsmeans_single: Deprecated by tern.mmrm::tabulate_mmrm: Statistics function which is extracting estimates from a tidied LS means data frame when ARM is not considered in the model.

summarize_lsmeans: Deprecated by tern.mmrm::tabulate_mmrm: Analyze function for tabulating LS means estimates from tidied mmrm results.

Examples

if (FALSE) {
library(dplyr)
library(lme4)

dat <- sleepstudy %>%
  mutate(
    group = factor(rep(c("A", "B"), length = nrow(sleepstudy))),
    days_grouped = cut(
      Days,
      breaks = stats::quantile(Days, probs = seq(0, 1, length = 5)),
      include.lowest = TRUE
    ),
    Subject = case_when(
      group == "A" ~ as.character(Subject),
      TRUE ~ as.character(as.numeric(as.character(Subject)) + 50)
    )
  ) %>%
  distinct_at(.vars = c("Subject", "days_grouped", "group"), .keep_all = TRUE)

result <- fit_mmrm(
  vars = list(
    response = "Reaction",
    covariates = c(),
    id = "Subject",
    arm = "group",
    visit = "days_grouped"
  ),
  data = dat,
  cor_struct = "compound-symmetry",
  parallel = TRUE
)
as.rtable(result, type = "cov", format = "xx.x")

result_no_arm <- fit_mmrm(
  vars = list(
    response = "Reaction",
    covariates = c(),
    id = "Subject",
    visit = "days_grouped"
  ),
  data = dat,
  cor_struct = "compound-symmetry",
  parallel = TRUE
)
as.rtable(result_no_arm, type = "cov", format = "xx.x")
}
if (FALSE) {
library(broom)
df <- tidy(result)
df_no_arm <- tidy(result_no_arm)
}
if (FALSE) {
s_mmrm_lsmeans(df[8, ], .in_ref_col = FALSE)
}
if (FALSE) {
s_mmrm_lsmeans_single(df_no_arm[4, ])
}
if (FALSE) {
dat_adsl <- dat %>%
  select(Subject, group) %>%
  unique()

basic_table() %>%
  split_cols_by("group", ref_group = result$ref_level) %>%
  add_colcounts() %>%
  split_rows_by("days_grouped") %>%
  summarize_lsmeans(show_relative = "increase") %>%
  build_table(
    df = broom::tidy(result),
    alt_counts_df = dat_adsl
  )

basic_table() %>%
  split_rows_by("days_grouped") %>%
  summarize_lsmeans(arms = FALSE) %>%
  build_table(
    df = broom::tidy(result_no_arm),
    alt_counts_df = dat_adsl
  )
}