Calculate statistics Pairwise with reference level

Calculate statistics Pairwise with reference level

tabulate_pairwise(x, col_by, FUN, row.name, ...)

Arguments

x
a vector

col_by
a factor first level is taken as reference level

FUN
a function with two arguments, first argument is the subset of x and the second argument is a factor with two levels from col_by

...
arguments passed to rrow

Value

an rtable

Examples

tabulate_pairwise(
  x = iris$Sepal.Length,
  col_by = iris$Species,
  FUN = function(xi, col_by_i) diff(tapply(xi, col_by_i, mean)),
  row.name = "diff mean"
)

#>                    setosa          versicolor        virginica 
#> ---------------------------------------------------------------
#> diff mean                             0.93             1.582