Time to Event Table

This is the description of the function

time_to_event_table(time_to_event, event, arm, earliest_contributing_event,
  strata_data, time_points, time_unit = "month", ties = "exact")

Arguments

time_to_event
time to event data

is_event
boolean with TRUE if event and FALSE if censored

Value

a named list with one element per row where each element contains the data for the row

Examples

library(random.cdisc.data)
library(forcats)

ASL <- radam("ASL")
ATE <- merge(ASL, radam("ATE", ADSL = ASL))

ATE_f <- subset(ATE, PARAMCD == "OS")

tbl <- time_to_event_table(
  time_to_event = ATE_f$AVAL,
  event = ATE_f$CNSR == 0,
  earliest_contributing_event = ATE_f$EVNTDESC,
  arm = factor(ATE_f$ARM),
  strata_data = ATE_f %>% select(SEX, RACE),
  time_points = 6,
  time_unit = "month"
)

tbl

#>                                                                          ARM A                   ARM B      
#>                                                                          (N=56)                  (N=44)     
#> ------------------------------------------------------------------------------------------------------------
#> Patients with event (%)                                                 28 (50%)                22 (50%)    
#>   Earliest contributing event
#>     Alive                                                                  8                       4        
#>     Death                                                                  6                       6        
#>     Disease Progression                                                    12                      5        
#>     Disease Progression or Death Censored for Missing Visits               5                       8        
#>     First Response                                                         10                      11       
#>     Last tumor assessment                                                  9                       3        
#>     Randomization                                                          6                       7        
#> Patients without event (%)                                              28 (50%)                22 (50%)    
#> 
#> Time to Event (months)
#>   Median                                                                107.464                  91.863     
#>     95% CI                                                           (66.8, 191.2)           (77.4, 159.7)  
#>   25% and 75%−ile                                                     48.8, 222.7             52.5, 166.3   
#>   Range                                                                1 to 442.4             4.7 to 279.6  
#> 
#> Unstratified Analysis
#>   p-value (log-rank)                                                                             0.799      
#> 
#>   Hazard Ratio                                                                                   1.0764     
#>     95% CI                                                                                  (0.6106, 1.8975)
#> 
#> Stratified Analysis
#>   p-value (log-rank)                                                                             0.6819     
#> 
#>   Hazard Ratio                                                                                   1.1503     
#>     95% CI                                                                                  (0.5887, 2.2478)
#> 
#> Time Point Analysis
#>   6 months
#>     Patients remaining at risk                                             52                      43       
#>     Event Free Rate (%)                                                  96.4%                   97.73%     
#>       95% CI                                                          (91.49, 100)            (93.32, 100)  
#>     Difference in Event Free Rate                                                                 1.33      
#>       95% CI                                                                                 (-5.26, 7.92)  
#>     p-value (Z-test)                                                                             0.6921     

# NOT RUN {
library(atezo.data)
library(dplyr)
library(forcats)

ASL <- asl(com.roche.cdt30019.go29436.re)
ATE <- ate(com.roche.cdt30019.go29436.re)

tte_tbl_stream <- get_time_to_event_table(com.roche.cdt30019.go29436.re)
Viewer(tte_tbl_stream)

ATE_f <- ATE %>%
 filter(ITTFL == "Y", PARAMCD == "OS") %>%
 mutate(ARM_ANL = fct_relevel(ARM1, "DUMMY C", "DUMMY B", "DUMMY A"))

tbl <- time_to_event_table(
  time_to_event = ATE_f$AVAL,
  event = ATE_f$CNSR == 0,
  earliest_contributing_event = ATE_f$EVNTDESC,
  arm = ATE_f$ARM_ANL,
  strata_data = ATE_f %>% select(SEX, MLIVER, TCICLVL2),
  time_points = 6,
  time_unit = "month"
)

Viewer(tbl)

Viewer(tbl, tte_tbl_stream)

compare_rtables(tbl, tte_tbl_stream, comp.attr = FALSE)

tbl[13,1]
tte_tbl_stream[13,1]
# }