Set Label Attributes of All Variables in a data.frame

Variable labels can be stored as a label attribute for each variable. This functions sets all non-missing variable labels in a data.frame

var_labels(x) <- value

Arguments

x
a data.frame object

value
new variable labels, NA removes the variable label

Value

modifies the variable labels of x

Examples

x <- iris
var_labels(x)

#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width      Species 
#>           NA           NA           NA           NA           NA 

var_labels(x) <- paste("label for", names(iris))
var_labels(x)

#>             Sepal.Length              Sepal.Width             Petal.Length 
#> "label for Sepal.Length"  "label for Sepal.Width" "label for Petal.Length" 
#>              Petal.Width                  Species 
#>  "label for Petal.Width"      "label for Species" 

# NOT RUN {
View(x) # in RStudio data viewer labels are displayed
# }