Patient Counts with Abnormal Range Values

Primary analysis variable .var indicates the abnormal range result (character or factor) and additional analysis variables are id (character or factor) and baseline (character or factor). For each direction specified in abnormal (e.g. high or low) count patients in the numerator and denominator as follows:

num

the number of patients with this abnormality recorded while on treatment.

denom

the number of patients with at least one post-baseline assessment.

Note, the denominator includes patients that might have other abnormal levels at baseline, and patients with missing baseline. Note, optionally patients with this abnormality at baseline can be excluded from numerator and denominator.

s_count_abnormal(
  df,
  .var,
  abnormal,
  variables = list(id = "USUBJID", baseline = "BNRIND"),
  exclude_base_abn = FALSE
)

a_count_abnormal(
  df,
  .var,
  abnormal,
  variables = list(id = "USUBJID", baseline = "BNRIND"),
  exclude_base_abn = FALSE
)

count_abnormal(
  lyt,
  var,
  abnormal,
  ...,
  table_names = abnormal,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)

Arguments

df
(data frame)
data set containing all analysis variables.

.var
(string)
single variable name that is passed by rtables when requested by a statistics function.

abnormal
(named string or character)
identifying the abnormal range level(s) in var.

variables
(named list of string)
list of additional analysis variables.

exclude_base_abn
(flag)
whether to exclude subjects with baseline abnormality from numerator and denominator.

lyt
(layout)
input layout where analyses will be added to.

var
(string)
single variable name for the primary analysis variable.

...
additional arguments for the lower level functions.

table_names
(character)
this can be customized in case that the same vars are analyzed multiple times, to avoid warnings from rtables.

.stats
(character)
statistics to select for the table.

.formats
(named character or list)
formats for the statistics.

.labels
(named character)
labels for the statistics (without indent).

.indent_mods
(named integer)
indent modifiers for the labels.

Value

s_count_abnormal() returns the statistic fraction which is a vector with num and denom counts of patients.

a_count_abnormal() returns the corresponding list with formatted rtables::CellValue().

count_abnormal() can be used with multiple abnormal levels and modifies the layout.

Details

Note that df should be filtered to include only post-baseline records.

Functions

s_count_abnormal: Statistics function which counts patients with abnormal range values for a single abnormal level.

a_count_abnormal: Formatted Analysis function which can be further customized by calling rtables::make_afun() on it. It is used as afun in rtables::analyze().

count_abnormal: Layout creating function which can be used for creating tables, which can take statistics function arguments and additional format arguments (see below). Note that it only works with a single variable but multiple abnormal levels.

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

df <- data.frame(
  USUBJID = as.character(c(1, 1, 2, 2)),
  ANRIND = factor(c("NORMAL", "LOW", "HIGH", "HIGH")),
  BNRIND = factor(c("NORMAL", "NORMAL", "HIGH", "HIGH")),
  ONTRTFL = c("", "Y", "", "Y"),
  stringsAsFactors = FALSE
)

# Select only post-baseline records.
df <- df %>%
  filter(ONTRTFL == "Y")

# For abnormal level "HIGH" we get the following counts.
s_count_abnormal(df, .var = "ANRIND", abnormal = c(high = "HIGH"))

#> $fraction
#>   num denom 
#>     1     2 
#> attr(,"label")
#> [1] "high"
#> 

# Optionally exclude patients with abnormality at baseline.
s_count_abnormal(df, .var = "ANRIND", abnormal = c(high = "HIGH"), exclude_base_abn = TRUE)

#> $fraction
#>   num denom 
#>     0     1 
#> attr(,"label")
#> [1] "high"
#> 

# Use the Formatted Analysis function for `analyze()`.
a_count_abnormal(df, .var = "ANRIND", abnormal = c(low = "LOW"))

#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>   row_name formatted_cell indent_mod row_label
#> 1 fraction      1/2 (50%)          0       low

# Layout creating function.
basic_table() %>%
  count_abnormal(var = "ANRIND", abnormal = c(high = "HIGH", low = "LOW")) %>%
  build_table(df)

#>         all obs 
#> ----------------
#> high   1/2 (50%)
#> low    1/2 (50%)

# Passing of statistics function and formatting arguments.
df2 <- data.frame(
  ID = as.character(c(1, 1, 2, 2)),
  RANGE = factor(c("NORMAL", "LOW", "HIGH", "HIGH")),
  BL_RANGE = factor(c("NORMAL", "NORMAL", "HIGH", "HIGH")),
  ONTRTFL = c("", "Y", "", "Y"),
  stringsAsFactors = FALSE
)

# Select only post-baseline records.
df2 <- df2 %>%
  filter(ONTRTFL == "Y")

basic_table() %>%
  count_abnormal(
    var = "RANGE",
    abnormal = c(low = "LOW", high = "HIGH"),
    variables = list(id = "ID", baseline = "BL_RANGE")
  ) %>%
  build_table(df2)

#>         all obs 
#> ----------------
#> low    1/2 (50%)
#> high   1/2 (50%)