Count patients with marked laboratory abnormalities

Primary analysis variable .var indicates whether single, replicated or last marked laboratory abnormality was observed (factor). Additional analysis variables are id (character or factor) and direction indicating the direction of the abnormality (factor). Denominator is number of patients with at least one valid measurement during treatment (post-baseline), and patients in the numerator are considered as follows:

For Single, not last and Last or replicated: Numerator is number of patients with Single, not last and Last or replicated levels, respectively.

For Any: Numerator is the number of patients with either single or replicated marked abnormalities.

s_count_abnormal_by_marked(
  df,
  .var = "AVALCAT1",
  abnormal = c(Low = "LOW LOW"),
  variables = list(id = "USUBJID", direction = "ANRIND")
)

a_count_abnormal_by_marked(
  df,
  .var = "AVALCAT1",
  abnormal = c(Low = "LOW LOW"),
  variables = list(id = "USUBJID", direction = "ANRIND")
)

count_abnormal_by_marked(
  lyt,
  var,
  abnormal,
  variables,
  ...,
  table_names = abnormal,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)

Arguments

df
(data frame)
data set containing all analysis variables.

.var
(string)
single variable name that is passed by rtables when requested by a statistics function.

abnormal
(named string or character)
identifying the abnormal direction in var.

variables
(named list of string)
list of additional analysis variables.

lyt
(layout)
input layout where analyses will be added to.

var
(string)
single variable name for the primary analysis variable.

...
additional arguments for the lower level functions.

table_names
(character)
this can be customized in case that the same vars are analyzed multiple times, to avoid warnings from rtables.

.stats
(character)
statistics to select for the table.

.formats
(named character or list)
formats for the statistics.

.labels
(named character)
labels for the statistics (without indent).

.indent_mods
(named integer)
indent modifiers for the labels.

Value

s_count_abnormal_by_marked() the single statistic count_fraction with Single, not last, Last or replicated and Any results.

a_count_abnormal_by_marked() returns the corresponding list with formatted rtables::CellValue().

Details

Note that Single, not last and Last or replicated levels are mutually exclusive. If a patient has abnormalities that meet both the Single, not last and Last or replicated criteria, then the patient will be counted only under the Last or replicated category.

Functions

s_count_abnormal_by_marked: Statistics function which returns the counts and fractions of patients with Single, not last, Last or replicated and Any marked laboratory abnormalities for a single abnormal level.

a_count_abnormal_by_marked: Formatted Analysis function which can be further customized by calling rtables::make_afun() on it. It is used as afun in rtables::analyze().

count_abnormal_by_marked: Layout creating function which can be used for creating tables, which can take statistics function arguments and additional format arguments (see below).

Examples

library(dplyr)

df <- data.frame(
  USUBJID = as.character(c(1, 1, 1, 1, 1, 2, 2, 2, 2, 2)),
  ARMCD = factor(c(rep("ARM A", 5),rep("ARM B", 5))),
  ANRIND = factor(c("NORMAL", "LOW", "HIGH", "HIGH HIGH","LOW LOW",
  "LOW", "LOW", "HIGH HIGH", "LOW", "HIGH HIGH")),
  ONTRTFL = c("", "Y", "Y", "Y", "Y", "Y", "Y", "Y", "Y", "Y"),
  PARAMCD = factor(rep("CRP", 10)),
  AVALCAT1 = factor(c("", "", "", "SINGLE", "REPLICATED", "", "", "LAST", "", "SINGLE")),
  stringsAsFactors = FALSE
)

# Select only post-baseline records.
df <- df %>% filter(ONTRTFL == "Y")
 s_count_abnormal_by_marked(df = df,
 .var = "AVALCAT1",
 abnormal = c(High = "HIGH HIGH"),
 variables = list(id = "USUBJID", direction = "ANRIND")
)

#> $count_fraction
#> $count_fraction$`Single, not last`
#> [1] 1.0 0.5
#> 
#> $count_fraction$`Last or replicated`
#> [1] 1.0 0.5
#> 
#> $count_fraction$`Any Abnormality`
#> [1] 2 1
#> 
#> 

 s_count_abnormal_by_marked(df = df,
 .var = "AVALCAT1",
 abnormal = c(Low = "LOW LOW"),
 variables = list(id = "USUBJID", direction = "ANRIND")
)

#> $count_fraction
#> $count_fraction$`Single, not last`
#> [1] 0 0
#> 
#> $count_fraction$`Last or replicated`
#> [1] 1.0 0.5
#> 
#> $count_fraction$`Any Abnormality`
#> [1] 1.0 0.5
#> 
#> 

# Use the Formatted Analysis function for `analyze()`. We need to ungroup `count_fraction` first
# so that the rtables formatting function `format_count_fraction()` can be applied correctly.
afun <- make_afun(a_count_abnormal_by_marked, .ungroup_stats = "count_fraction")
afun(
  df = df,
  .var = "AVALCAT1",
  abnormal = c(High = "HIGH HIGH"),
  variables = list(id = "USUBJID", direction = "ANRIND")
)

#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>             row_name formatted_cell indent_mod          row_label
#> 1   Single, not last        1 (50%)          0   Single, not last
#> 2 Last or replicated        1 (50%)          0 Last or replicated
#> 3    Any Abnormality       2 (100%)          0    Any Abnormality

basic_table() %>%
split_cols_by("ARMCD") %>%
summarize_num_patients(var = "USUBJID",
                       .stats = "unique_count") %>%
count_abnormal_by_marked(var = "AVALCAT1" ,
                         abnormal = c(Low = "LOW LOW", High = "HIGH HIGH"),
                         variables = list(id = "USUBJID", direction = "ANRIND")) %>%
                         build_table(df = df)

#>                           ARM A      ARM B  
#> --------------------------------------------
#>  (n)                        1          1    
#>   Low                                       
#>     Single, not last        0          0    
#>     Last or replicated   1 (100%)      0    
#>     Any Abnormality      1 (100%)      0    
#>   High                                      
#>     Single, not last     1 (100%)      0    
#>     Last or replicated      0       1 (100%)
#>     Any Abnormality      1 (100%)   1 (100%)