Patient Counts with the Most Extreme Post-baseline Toxicity Grade per Direction of Abnormality

Primary analysis variable .var indicates the toxicity grade (numeric), and additional analysis variables are id (character or factor) and worst_grade_flag (logical). For each direction specified in abnormal (e.g. high or low) we condition on the worst grade flag and count patients in the denominator as number of patients with at least one valid measurement during treatment, and patients in the numerator as follows:

1 to 4: Numerator is number of patients with worst grades 1-4 respectively;

Any: Numerator is number of patients with at least one abnormality, which means grade is different from 0.

s_count_abnormal_by_worst_grade(
  df,
  .var = "ATOXGR",
  abnormal = c("low", "high"),
  variables = list(id = "USUBJID", worst_grade_flag = "WGRLOFL")
)

a_count_abnormal_by_worst_grade(
  df,
  .var = "ATOXGR",
  abnormal = c("low", "high"),
  variables = list(id = "USUBJID", worst_grade_flag = "WGRLOFL")
)

count_abnormal_by_worst_grade(
  lyt,
  var,
  abnormal,
  variables,
  ...,
  table_names = abnormal,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)

Arguments

df
(data frame)
data set containing all analysis variables.

.var
(string)
single variable name that is passed by rtables when requested by a statistics function.

abnormal
(character)
identifying the abnormality direction.

variables
(named list of string)
list of additional analysis variables.

lyt
(layout)
input layout where analyses will be added to.

var
(string)
single variable name for the primary analysis variable.

...
additional arguments for the lower level functions.

table_names
(character)
this can be customized in case that the same vars are analyzed multiple times, to avoid warnings from rtables.

.stats
(character)
statistics to select for the table.

.formats
(named character or list)
formats for the statistics.

.labels
(named character)
labels for the statistics (without indent).

.indent_mods
(named integer)
indent modifiers for the labels.

Value

s_count_abnormal_by_worst_grade() the single statistic count_fraction with grade 1 to 4 and "Any" results.

a_count_abnormal_by_worst_grade() returns the corresponding list with formatted rtables::CellValue().

Details

Note that df should be filtered to include only post-baseline records.

Functions

s_count_abnormal_by_worst_grade: Statistics function which counts patients with worst grade for a single abnormal level, consisting of counts and percentages of patients with worst grade 1 to 4, and Any non-zero grade.

a_count_abnormal_by_worst_grade: Formatted Analysis function which can be further customized by calling rtables::make_afun() on it. It is used as afun in rtables::analyze().

count_abnormal_by_worst_grade: Layout creating function which can be used for creating tables, which can take statistics function arguments and additional format arguments (see below).

Examples

library(scda)
library(dplyr)

adlb <- synthetic_cdisc_data("latest")$adlb
adlb_f <- adlb %>%
  filter(!AVISIT %in% c("SCREENING", "BASELINE")) %>%
  mutate(
    ATOXGR = as.numeric(as.character(ATOXGR)),
    WGRLOFL = case_when(WGRLOFL == "Y" ~ TRUE, TRUE ~ FALSE),
    WGRHIFL = case_when(WGRHIFL == "Y" ~ TRUE, TRUE ~ FALSE)
  )

s_count_abnormal_by_worst_grade(
  df = adlb_f %>% filter(ARMCD == "ARM A" & PARAMCD == "CRP"),
  .var = "ATOXGR",
  abnormal = "low",
  variables = list(id = "USUBJID", worst_grade_flag = "WGRLOFL")
)

#> $count_fraction
#> $count_fraction$`1`
#> [1] 16.000000  0.119403
#> 
#> $count_fraction$`2`
#> [1] 21.0000000  0.1567164
#> 
#> $count_fraction$`3`
#> [1] 12.00000000  0.08955224
#> 
#> $count_fraction$`4`
#> [1] 11.00000000  0.08208955
#> 
#> $count_fraction$Any
#> [1] 60.0000000  0.4477612
#> 
#> 

s_count_abnormal_by_worst_grade(
  df = adlb_f %>% filter(ARMCD == "ARM A" & PARAMCD == "CRP"),
  .var = "ATOXGR",
  abnormal = "high",
  variables = list(id = "USUBJID", worst_grade_flag = "WGRHIFL")
)

#> $count_fraction
#> $count_fraction$`1`
#> [1] 17.0000000  0.1268657
#> 
#> $count_fraction$`2`
#> [1] 15.0000000  0.1119403
#> 
#> $count_fraction$`3`
#> [1] 16.000000  0.119403
#> 
#> $count_fraction$`4`
#> [1] 12.00000000  0.08955224
#> 
#> $count_fraction$Any
#> [1] 60.0000000  0.4477612
#> 
#> 

# Use the Formatted Analysis function for `analyze()`. We need to ungroup `count_fraction` first
# so that the rtables formatting function `format_count_fraction()` can be applied correctly.
afun <- make_afun(a_count_abnormal_by_worst_grade, .ungroup_stats = "count_fraction")
afun(
  df = adlb_f %>% filter(ARMCD == "ARM A" & PARAMCD == "CRP"),
  .var = "ATOXGR",
  abnormal = "high",
  variables = list(id = "USUBJID", worst_grade_flag = "WGRHIFL")
)

#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>   row_name formatted_cell indent_mod row_label
#> 1        1     17 (12.7%)          0         1
#> 2        2     15 (11.2%)          0         2
#> 3        3     16 (11.9%)          0         3
#> 4        4        12 (9%)          0         4
#> 5      Any     60 (44.8%)          0       Any

basic_table() %>%
  count_abnormal_by_worst_grade(
    var = "ATOXGR",
    abnormal = c(Low = "low", High = "high"),
    variables = list(id = "USUBJID", worst_grade_flag = c(Low = "WGRLOFL", High = "WGRHIFL"))
  ) %>%
  build_table(df = adlb_f %>% filter(ARMCD == "ARM A" & PARAMCD == "CRP"))

#>          all obs  
#> ------------------
#> Low               
#>   1     16 (11.9%)
#>   2     21 (15.7%)
#>   3      12 (9%)  
#>   4     11 (8.2%) 
#>   Any   60 (44.8%)
#> High              
#>   1     17 (12.7%)
#>   2     15 (11.2%)
#>   3     16 (11.9%)
#>   4      12 (9%)  
#>   Any   60 (44.8%)

basic_table() %>%
  split_cols_by("ARMCD") %>%
  split_rows_by("PARAMCD") %>%
  count_abnormal_by_worst_grade(
    var = "ATOXGR",
    abnormal = c(Low = "low", High = "high"),
    variables = list(id = "USUBJID", worst_grade_flag = c(Low = "WGRLOFL", High = "WGRHIFL"))
  ) %>%
  build_table(df = adlb_f)

#>             ARM A        ARM B        ARM C   
#> ----------------------------------------------
#> ALT                                           
#>   Low                                         
#>     1     14 (10.4%)   15 (11.2%)   10 (7.6%) 
#>     2     13 (9.7%)    18 (13.4%)   11 (8.3%) 
#>     3     20 (14.9%)    12 (9%)     10 (7.6%) 
#>     4      7 (5.2%)      8 (6%)     10 (7.6%) 
#>     Any   54 (40.3%)   53 (39.6%)   41 (31.1%)
#>   High                                        
#>     1     18 (13.4%)   16 (11.9%)   18 (13.6%)
#>     2     15 (11.2%)   13 (9.7%)    18 (13.6%)
#>     3     10 (7.5%)    16 (11.9%)   13 (9.8%) 
#>     4      12 (9%)     11 (8.2%)    13 (9.8%) 
#>     Any    55 (41%)    56 (41.8%)    62 (47%) 
#> CRP                                           
#>   Low                                         
#>     1     16 (11.9%)   19 (14.2%)   15 (11.4%)
#>     2     21 (15.7%)   13 (9.7%)    16 (12.1%)
#>     3      12 (9%)      9 (6.7%)    18 (13.6%)
#>     4     11 (8.2%)     7 (5.2%)     5 (3.8%) 
#>     Any   60 (44.8%)   48 (35.8%)   54 (40.9%)
#>   High                                        
#>     1     17 (12.7%)   15 (11.2%)    9 (6.8%) 
#>     2     15 (11.2%)   16 (11.9%)   13 (9.8%) 
#>     3     16 (11.9%)    12 (9%)     14 (10.6%)
#>     4      12 (9%)      12 (9%)     10 (7.6%) 
#>     Any   60 (44.8%)    55 (41%)    46 (34.8%)
#> IGA                                           
#>   Low                                         
#>     1     17 (12.7%)   23 (17.2%)   17 (12.9%)
#>     2      12 (9%)     15 (11.2%)   14 (10.6%)
#>     3     15 (11.2%)   11 (8.2%)    14 (10.6%)
#>     4     10 (7.5%)    16 (11.9%)   10 (7.6%) 
#>     Any   54 (40.3%)   65 (48.5%)   55 (41.7%)
#>   High                                        
#>     1     25 (18.7%)   14 (10.4%)   13 (9.8%) 
#>     2     14 (10.4%)   20 (14.9%)   16 (12.1%)
#>     3      12 (9%)     13 (9.7%)    17 (12.9%)
#>     4     11 (8.2%)    13 (9.7%)      4 (3%)  
#>     Any   62 (46.3%)   60 (44.8%)   50 (37.9%)