Standard Arguments

The documentation to this function lists all the arguments in tern that are used repeatedly to express an analysis.

Arguments

...
additional arguments for the lower level functions.

.df_row
(data frame)
data frame across all of the columns for the given row split.

.in_ref_col
(logical)
TRUE when working with the reference level, FALSE otherwise.

.N_col
(count)
row-wise N (row group count) for the group of observations being analyzed (i.e. with no column-based subsetting) that is passed by rtables.

.N_row
(count)
column-wise N (column count) for the full column that is passed by rtables.

.ref_group
(data frame or vector)
the data corresponding to the reference group.

.stats
(character)
statistics to select for the table.

.indent_mods
(named integer)
indent modifiers for the labels.

.formats
(named character or list)
formats for the statistics.

.labels
(named character)
labels for the statistics (without indent).

.var
(string)
single variable name that is passed by rtables when requested by a statistics function.

col_by
(factor)
defining column groups.

conf_level
(proportion)
confidence level of the interval.

data
(data.frame)
the dataset containing the variables to summarize.

df
(data frame)
data set containing all analysis variables.

draw
(flag)
whether the plot should be drawn.

drop
(flag)
should non appearing occurrence levels be dropped from the resulting table. Note that in that case the remaining occurrence levels in the table are sorted alphabetically.

id
(string)
subject variable name.

is_event
(logical)
TRUE if event, FALSE if time to event is censored.

labelstr
(character)
label of the level of the parent split currently being summarized (must be present as second argument in Content Row Functions).

lyt
(layout)
input layout where analyses will be added to.

na.rm
(flag)
whether NA values should be removed from x prior to analysis.

na_rm
(flag)
(deprecated) whether NA values should be removed from x prior to analysis.

na_level
(string)
used to replace all NA or empty values in factors.

newpage
(flag)
whether the plot should be drawn on a new page. Only considered if draw = TRUE is used.

prune_zero_rows
(flag)
whether to prune all zero rows.

rsp
(logical)
whether each subject is a responder or not.

show_labels
label visibility: one of "default", "visible" and "hidden".

table_names
(character)
this can be customized in case that the same vars are analyzed multiple times, to avoid warnings from rtables.

tte
(numeric)
contains time-to-event duration values.

var_labels
character for label.

variables
(named list of string)
list of additional analysis variables.

vars
(character)
variable names for the primary analysis variable to be iterated over.

var
(string)
single variable name for the primary analysis variable.

x
(numeric)
vector of numbers we want to analyze.

Details

Although this function just returns NULL it has two uses, for the tern users it provides a documentation of arguments that are commonly and consistently used in the framework. For the developer it adds a single reference point to import the roxygen argument description with: @inheritParams argument_convention