Additional Assertions for assert_that

We provide additional assertion functions which can be used together with assertthat::assert_that().

is_character_or_factor(x)

is_nonnegative_count(x)

is_variables(x)

is_df_with_variables(df, variables)

is_df_with_factors(df, variables)

is_equal_length(...)

is_proportion(x, include_boundaries = FALSE)

is_proportion_vector(x, include_boundaries = FALSE)

is_valid_factor(x)

is_factor_no_na(x)

is_valid_character(x)

all_elements_in_ref(x, ref)

has_tabletree_colnames(x, col_names)

Arguments

x
object to test

df
supposed data frame to test

variables
supposed variables list to test

...
a collection of objects to test.

include_boundaries
(logical)
whether to include boundaries when testing for proportions.

ref
(vector)
where matches from x are sought for all_elements_in_ref.

col_names
(character)
column names which should be present in the table.

Value

flag whether the assertion holds (TRUE or FALSE). When used inside assertthat::assert_that() produces a meaningful error message.

Functions

is_character_or_factor: Check whether x is a character or factor vector.

is_nonnegative_count: Check whether x is a nonnegative count.

is_variables: Check whether x is a valid list of variable names.

is_df_with_variables: Check whether df is a data frame with the analysis variables.

is_df_with_factors: Check whether df is a data frame where the analysis variables are all factors.

is_equal_length: Check that objects provided are of same length.

is_proportion: Check whether x is a proportion: number between 0 and 1.

is_proportion_vector: Check whether x is a vector of proportions (numbers between 0 and 1).

is_valid_factor: Check whether x is a valid factor (has levels and no empty string levels).

is_factor_no_na: Check whether NA has been handled when x is a factor.

is_valid_character: Check whether x is a valid character (has no NAs and no empty strings).

all_elements_in_ref: Check whether all elements of x are in a reference vector.

has_tabletree_colnames: Check whether rtables object x has the specified column names.

Examples

# Check whether `x` is a character or factor vector.
is_character_or_factor(-1)

#> [1] FALSE

is_character_or_factor(c("a", "b"))

#> [1] TRUE

is_character_or_factor(factor(c("a", "b")))

#> [1] TRUE

# Check whether `x` is a nonnegative count.
is_nonnegative_count(-1)

#> [1] FALSE

is_nonnegative_count(0L)

#> [1] TRUE

is_nonnegative_count(10L)

#> [1] TRUE

is_nonnegative_count(0)

#> [1] TRUE

is_nonnegative_count(10)

#> [1] TRUE

# Check whether `x` is a valid list of variable names.
is_variables(list(val = "a"))

#> [1] TRUE

is_variables(list(val = c("a", "b")))

#> [1] TRUE

is_variables(list(1, 2))

#> [1] FALSE

is_variables(list("bla"))

#> [1] FALSE

# Check whether `df` contains the analysis `variables`.
is_df_with_variables(df = data.frame(a = 5, b = 3), variables = list(val = "a"))

#> [1] TRUE

is_df_with_variables(df = data.frame(a = 5, b = 3), variables = list(val = c("a", "b")))

#> [1] TRUE

# Check whether `df` contains all factor analysis `variables`.
is_df_with_factors(
  df = data.frame(a = factor("A", levels = c("A", "B")), b = 3),
  variables = list(val = "a")
)

#> [1] TRUE

#' # Check whether `x` is a valid list of variable names.
a <- 1
b <- NULL
c <- c(1, "car")
d <- 5
is_equal_length(a, b, c, d)

#> [1] FALSE

# Check whether `x` is between 0 and 1.
is_proportion(x = 0.3)

#> [1] TRUE

is_proportion(x = 1.3)

#> [1] FALSE

is_proportion(x = 0, include_boundaries = TRUE)

#> [1] TRUE

# Check whether `x` is a vector of numbers between 0 and 1.
is_proportion_vector(c(0.3, 0.1))

#> [1] TRUE

is_proportion_vector(c(1.3, 0.1))

#> [1] FALSE

is_proportion_vector(0, include_boundaries = TRUE)

#> [1] TRUE

# Check whether `x` is a valid factor.
is_valid_factor(-1)

#> [1] FALSE

is_valid_factor(factor(c("a", "b")))

#> [1] TRUE

is_valid_factor(factor(c("a", "")))

#> [1] FALSE

is_valid_factor(factor())

#> [1] FALSE

# Check whether `NA` has been handled when `x` is a factor.
is_factor_no_na(factor(c("a", NA)))

#> [1] FALSE

is_factor_no_na(factor(c("a", "<Missing>")))

#> [1] TRUE

# Check whether `x` is a valid character vector
is_valid_character(-1)

#> [1] FALSE

is_valid_character(c("a", "b"))

#> [1] TRUE

is_valid_character(c("a", ""))

#> [1] FALSE

is_valid_character("")

#> [1] FALSE

# Check whether all elements of `x` are in a reference vector.
all_elements_in_ref(c("a", "b"), c("a", "b", "c"))

#> [1] TRUE

all_elements_in_ref(c("a", "d"), c("a", "b", "c"))

#> [1] FALSE

all_elements_in_ref(c(1:3), c(1:5))

#> [1] TRUE