Wrap data generated from chgfbl_data into rtable format, to create a test result and change from baseline by visit summary table

Wrap data generated from chgfbl_data into rtable format, to create a test result and change from baseline by visit summary table

chgfbl_table(data)

Arguments

data
data generated by chgfbl_data function

Examples

library(tern)
library(random.cdisc.data)

ASL <- radam("ASL")

ANL <- merge(ASL, radam("AQS", ADSL = ASL)) %>%
   filter(PARAMCD == "FATIGI")

df <- chgfbl_data(data = ANL)
chgfbl_table(df)

#>                        ARM A                 ARM A                 ARM B                 ARM B     
#>                    Value at Visit         Change from          Value at Visit         Change from  
#>                        (N=41)               Baseline               (N=59)               Baseline   
#>                                              (N=41)                                      (N=59)    
#> ---------------------------------------------------------------------------------------------------
#> WEEK 1
#>   n                      41                                          59                            
#>   Mean (SD)          -0.1 (0.8)                                   -0.1 (1)                         
#>   Median                -0.2                                        -0.1                           
#>   IQR               -0.76 - 0.41                                -0.62 - 0.56                       
#>   Min - Max         -1.82 - 1.59                                -2.16 - 1.91                       
#> 
#> WEEK 2
#>   n                      41                    41                    59                    59      
#>   Mean (SD)           0 (1.2)               0 (1.4)              -0.2 (1.1)             0 (1.6)    
#>   Median                -0.1                  -0.1                  -0.1                  -0.2     
#>   IQR               -0.87 - 0.58          -0.9 - 1.06           -0.97 - 0.59           -0.95 - 1   
#>   Min - Max         -1.71 - 3.59          -3.21 - 2.61          -2.45 - 2.95          -3.91 - 3.68 
#> 
#> WEEK 3
#>   n                      38                    38                    48                    48      
#>   Mean (SD)          -0.4 (1.1)            -0.5 (1.6)            -0.3 (0.8)            -0.1 (1.6)  
#>   Median                -0.3                  -0.7                  -0.1                  -0.1     
#>   IQR               -1.26 - 0.32          -1.7 - 0.73           -0.59 - 0.19          -1.23 - 0.82 
#>   Min - Max         -2.21 - 2.45          -2.69 - 4.14          -2.32 - 1.19            -3.6 - 3   
#> 
#> WEEK 4
#>   n                      34                    34                    38                    38      
#>   Mean (SD)           -0.1 (1)              0 (1.3)               -0.1 (1)             0.1 (1.4)   
#>   Median                -0.2                  0.4                   -0.1                  0.2      
#>   IQR               -0.59 - 0.4           -1.1 - 0.83           -0.65 - 0.46          -0.66 - 0.85 
#>   Min - Max         -2.39 - 2.13          -2.7 - 2.44           -2.18 - 1.95          -2.56 - 3.34 
#> 
#> WEEK 5
#>   n                      25                    25                    37                    37      
#>   Mean (SD)          -0.1 (0.9)             0 (1.4)              0.2 (1.1)             0.3 (1.4)   
#>   Median                 0                    -0.3                  0.1                   0.4      
#>   IQR               -0.58 - 0.45          -0.64 - 0.82          -0.48 - 0.82          -0.61 - 1.46 
#>   Min - Max         -2.41 - 1.51          -3.27 - 2.83          -1.97 - 1.87          -2.72 - 3.35 
#> 
#> WEEK 6
#>   n                      24                    24                    32                    32      
#>   Mean (SD)           0.2 (1)              0.3 (1.2)             -0.3 (1.1)             0 (1.7)    
#>   Median                0.3                   0.4                   -0.2                  -0.1     
#>   IQR               -0.34 - 0.78          -0.38 - 1.07          -0.96 - 0.22          -1.09 - 0.98 
#>   Min - Max         -1.59 - 1.96          -2.24 - 2.46          -2.34 - 3.45          -3.23 - 4.25 
#> 
#> WEEK 7
#>   n                      19                    19                    29                    29      
#>   Mean (SD)          -0.4 (1.2)            -0.4 (1.4)             0.2 (1)              0.4 (1.6)   
#>   Median                -0.4                  -0.6                   0                    0.7      
#>   IQR               -1.42 - 0.55          -1.34 - 0.87          -0.71 - 0.97          -0.69 - 1.22 
#>   Min - Max         -2.43 - 2.1           -2.88 - 1.94          -1.63 - 1.96          -3.05 - 3.72 
#> 
#> WEEK 8
#>   n                      17                    17                    24                    24      
#>   Mean (SD)           -0.1 (1)              0 (1.6)              0.1 (0.8)             0.4 (1.6)   
#>   Median                -0.2                  0.1                   0.3                   0.5      
#>   IQR               -0.79 - 0.3           -0.82 - 0.71          -0.36 - 0.58          -0.71 - 1.18 
#>   Min - Max         -1.88 - 1.85          -2.87 - 2.28          -1.78 - 1.35          -2.54 - 3.75 
#> 
#> WEEK 9
#>   n                      13                    13                    20                    20      
#>   Mean (SD)          -0.3 (0.7)            -0.2 (1.2)             -0.4 (1)             -0.3 (1.4)  
#>   Median                -0.1                  -0.3                  -0.6                  -0.2     
#>   IQR               -0.62 - 0.02          -1.25 - 0.73          -0.91 - 0.02          -1.38 - 0.66 
#>   Min - Max         -1.94 - 0.49          -1.67 - 1.8           -2.16 - 1.74          -2.78 - 1.97 
#> 
#> WEEK 10
#>   n                      5                     5                     14                    14      
#>   Mean (SD)          -0.4 (1.5)            -0.5 (1.6)             -0.4 (1)             -0.1 (1.8)  
#>   Median                0.3                   -0.4                  -0.5                  -0.5     
#>   IQR               -1.46 - 0.72          -1.78 - 0.62          -0.93 - 0.18          -0.84 - 0.94 
#>   Min - Max         -2.52 - 0.94          -2.45 - 1.53          -2.31 - 1.44          -2.48 - 4.44 
#> 
#> WEEK 11
#>   n                      3                     3                     7                     7       
#>   Mean (SD)           -0.5 (1)             0.1 (0.8)             -0.3 (0.9)            -0.6 (1.3)  
#>   Median                -0.9                  -0.1                  -0.6                  -1.2     
#>   IQR              -1.07 - -0.08          -0.29 - 0.44          -0.98 - 0.33         -1.25 - -0.21 
#>   Min - Max         -1.26 - 0.71          -0.45 - 1.02          -1.16 - 0.99          -2.03 - 1.95 
#>