Compare Variables Between Groups

We use the new S3 generic function s_compare() to implement comparisons for different x objects. This is used as Statistics Function in combination with the new Analyze Function compare_vars().

s_compare(x, .ref_group, .in_ref_col, ...)

# S3 method for numeric
s_compare(x, .ref_group, .in_ref_col, ...)

# S3 method for factor
s_compare(
  x,
  .ref_group,
  .in_ref_col,
  denom = "n",
  na.rm = TRUE,
  na_level = "<Missing>",
  ...
)

# S3 method for character
s_compare(
  x,
  .ref_group,
  .in_ref_col,
  denom = "n",
  na.rm = TRUE,
  na_level = "<Missing>",
  .var,
  ...
)

# S3 method for logical
s_compare(x, .ref_group, .in_ref_col, na.rm = TRUE, denom = "n", ...)

a_compare(x, .ref_group, .in_ref_col, ..., .var)

# S3 method for numeric
a_compare(x, .ref_group, .in_ref_col, ...)

# S3 method for factor
a_compare(
  x,
  .ref_group,
  .in_ref_col,
  denom = "n",
  na.rm = TRUE,
  na_level = "<Missing>",
  ...
)

# S3 method for character
a_compare(
  x,
  .ref_group,
  .in_ref_col,
  denom = "n",
  na.rm = TRUE,
  na_level = "<Missing>",
  .var,
  ...
)

# S3 method for logical
a_compare(x, .ref_group, .in_ref_col, na.rm = TRUE, denom = "n", ...)

create_afun_compare(
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)

compare_vars(
  lyt,
  vars,
  var_labels = vars,
  nested = TRUE,
  ...,
  show_labels = "default",
  table_names = vars,
  .stats = c("n", "mean_sd", "count_fraction", "pval"),
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)

Arguments

x
(numeric)
vector of numbers we want to analyze.

.ref_group
(data frame or vector)
the data corresponding to the reference group.

.in_ref_col
(logical)
TRUE when working with the reference level, FALSE otherwise.

...
arguments passed to s_compare().

denom
(string)
choice of denominator for factor proportions, can only be n (number of values in this row and column intersection).

na.rm
(flag)
whether NA values should be removed from x prior to analysis.

na_level
(string)
used to replace all NA or empty values in factors.

.var
(string)
single variable name that is passed by rtables when requested by a statistics function.

.stats
(character)
statistics to select for the table.

.formats
(named character or list)
formats for the statistics.

.labels
(named character)
labels for the statistics (without indent).

.indent_mods
(named integer)
indent modifiers for the labels.

lyt
(layout)
input layout where analyses will be added to.

vars
(character)
variable names for the primary analysis variable to be iterated over.

var_labels
character for label.

nested
boolean, Add this as a new top-level split (defining a new subtable directly under root). Defaults to FALSE

show_labels
label visibility: one of "default", "visible" and "hidden".

table_names
(character)
this can be customized in case that the same vars are analyzed multiple times, to avoid warnings from rtables.

Value

If x is of class numeric, returns a list with named items:

all items from s_summary.numeric().

pval: the p-value.

If x is of class factor or converted from character, returns a list with named items:

all items from s_summary.factor().

pval: the p-value.

If x is of class logical, returns a list with named items:

all items from s_summary.logical().

pval: the p-value.

Functions

s_compare: s_compare is a S3 generic function to produce an object description and comparison versus the reference column in the form of p-values.

s_compare.numeric: Method for numeric class. This uses the standard t-test to calculate the p-value.

s_compare.factor: Method for factor class. This uses the chi-squared test to calculate the p-value. Note that the denom for factor proportions can only be n here since the usage is for comparing proportions between columns. Therefore a row-based proportion would not make sense. Also proportion based on N_col would be difficult since for the chi-squared test statistic we use the counts. Therefore missing values should be accounted for explicitly as factor levels.

s_compare.character: Method for character class. This makes an automatic conversion to factor (with a warning) and then forwards to the method for factors.

s_compare.logical: Method for logical class. A chi-squared test is used. If missing values are not removed, then they are counted as FALSE.

a_compare: S3 generic Formatted Analysis function to produce an object description and comparison versus the reference column in the form of p-values. It is used as afun in rtables::analyze().

a_compare.numeric: Formatted Analysis function method for numeric.

a_compare.factor: Formatted Analysis function method for factor.

a_compare.character: Formatted Analysis function method for character.

a_compare.logical: Formatted Analysis function method for logical.

create_afun_compare: Constructor function which creates a combined Formatted Analysis function for use in layout creating functions compare_vars().

compare_vars: Analyze Function to add a comparison of variables to rtables pipelines. The column split needs to have the reference group defined via ref_group so that the comparison is well defined. The ellipsis (...) conveys arguments to s_compare(). When factor variables contains NA, it is expected that NA have been conveyed to na_level appropriately beforehand with df_explicit_na().

Note

Automatic conversion of character to factor does not guarantee that the table can be generated correctly. In particular for sparse tables this very likely can fail. It is therefore better to always preprocess the dataset such that factors are manually created from character variables before passing the dataset to rtables::build_table().

Since a_compare() is generic and we want customization of the formatting arguments via rtables::make_afun(), we need to create another temporary generic function, with corresponding customized methods. Then in order for the methods to be found, we need to wrap them in a combined afun. Since this is required by two layout creating functions (and possibly others in the future), we provide a constructor that does this: create_afun_compare().

Formatting arguments

These additional formatting arguments can be passed to the layout creating function:

.stats

(character)
names of the statistics to use

.indent_mods

(integer)
named vector of indent modifiers for the labels

.formats

(character or list)
named vector of formats for the statistics

.labels

(character)
named vector of labels for the statistics (without indent)

See also

Examples

# `s_compare.numeric`

## Usual case where both this and the reference group vector have more than 1 value.
s_compare(rnorm(10, 5, 1), .ref_group = rnorm(5, -5, 1), .in_ref_col = FALSE)

#> $n
#> [1] 10
#> 
#> $mean_sd
#>      mean        sd 
#> 5.1158855 0.7143188 
#> 
#> $median
#> [1] 5.082333
#> 
#> $range
#> [1] 4.112619 6.585859
#> 
#> $mean_ci
#> [1] 4.604893 5.626878
#> attr(,"label")
#> [1] "Mean 95% CI"
#> 
#> $median_ci
#> [1] 4.522784 5.878768
#> attr(,"label")
#> [1] "Median 95% CI"
#> 
#> $pval
#> [1] 2.220933e-12
#> 

## If one group has not more than 1 value, then p-value is not calculated.
s_compare(rnorm(10, 5, 1), .ref_group = 1, .in_ref_col = FALSE)

#> $n
#> [1] 10
#> 
#> $mean_sd
#>      mean        sd 
#> 4.8311931 0.4813606 
#> 
#> $median
#> [1] 4.749932
#> 
#> $range
#> [1] 4.153552 5.981895
#> 
#> $mean_ci
#> [1] 4.486849 5.175538
#> attr(,"label")
#> [1] "Mean 95% CI"
#> 
#> $median_ci
#> [1] 4.537948 5.073046
#> attr(,"label")
#> [1] "Median 95% CI"
#> 
#> $pval
#> character(0)
#> 

## Empty numeric does not fail, it returns NA-filled items and no p-value.
s_compare(numeric(), .ref_group = numeric(), .in_ref_col = FALSE)

#> $n
#> [1] 0
#> 
#> $mean_sd
#> mean   sd 
#>   NA   NA 
#> 
#> $median
#> [1] NA
#> 
#> $range
#> [1] NA NA
#> 
#> $mean_ci
#> [1] NA NA
#> attr(,"label")
#> [1] "Mean 95% CI"
#> 
#> $median_ci
#> [1] NA NA
#> attr(,"label")
#> [1] "Median 95% CI"
#> 
#> $pval
#> character(0)
#> 

# `s_compare.factor`

## Basic usage:
x <- factor(c("a", "a", "b", "c", "a"))
y <- factor(c("a", "b", "c"))
s_compare(x = x, .ref_group = y, .in_ref_col = FALSE)

#> $n
#> [1] 5
#> 
#> $count
#> $count$a
#> [1] 3
#> 
#> $count$b
#> [1] 1
#> 
#> $count$c
#> [1] 1
#> 
#> 
#> $count_fraction
#> $count_fraction$a
#> [1] 3.0 0.6
#> 
#> $count_fraction$b
#> [1] 1.0 0.2
#> 
#> $count_fraction$c
#> [1] 1.0 0.2
#> 
#> 
#> $pval
#> [1] 0.7659283
#> 

## Management of NA values.
x <- explicit_na(factor(c("a", "a", "b", "c", "a", NA, NA)))
y <- explicit_na(factor(c("a", "b", "c", NA)))
s_compare(x = x, .ref_group = y, .in_ref_col = FALSE, na.rm = TRUE)

#> $n
#> [1] 5
#> 
#> $count
#> $count$a
#> [1] 3
#> 
#> $count$b
#> [1] 1
#> 
#> $count$c
#> [1] 1
#> 
#> 
#> $count_fraction
#> $count_fraction$a
#> [1] 3.0 0.6
#> 
#> $count_fraction$b
#> [1] 1.0 0.2
#> 
#> $count_fraction$c
#> [1] 1.0 0.2
#> 
#> 
#> $pval
#> [1] 0.7659283
#> 

s_compare(x = x, .ref_group = y, .in_ref_col = FALSE, na.rm = FALSE)

#> $n
#> [1] 7
#> 
#> $count
#> $count$a
#> [1] 3
#> 
#> $count$b
#> [1] 1
#> 
#> $count$c
#> [1] 1
#> 
#> $count$`<Missing>`
#> [1] 2
#> 
#> 
#> $count_fraction
#> $count_fraction$a
#> [1] 3.0000000 0.4285714
#> 
#> $count_fraction$b
#> [1] 1.0000000 0.1428571
#> 
#> $count_fraction$c
#> [1] 1.0000000 0.1428571
#> 
#> $count_fraction$`<Missing>`
#> [1] 2.0000000 0.2857143
#> 
#> 
#> $pval
#> [1] 0.9063036
#> 

# `s_compare.character`

## Basic usage:
x <- c("a", "a", "b", "c", "a")
y <- c("a", "b", "c")
s_compare(x, .ref_group = y, .in_ref_col = FALSE, .var = "x")

#> Warning: automatically converting character variable x to factor, better manually convert to factor to avoid failures

#> Warning: automatically converting character variable x to factor, better manually convert to factor to avoid failures

#> $n
#> [1] 5
#> 
#> $count
#> $count$a
#> [1] 3
#> 
#> $count$b
#> [1] 1
#> 
#> $count$c
#> [1] 1
#> 
#> 
#> $count_fraction
#> $count_fraction$a
#> [1] 3.0 0.6
#> 
#> $count_fraction$b
#> [1] 1.0 0.2
#> 
#> $count_fraction$c
#> [1] 1.0 0.2
#> 
#> 
#> $pval
#> [1] 0.7659283
#> 

## Note that missing values handling can make a large difference:
x <- c("a", "a", "b", "c", "a", NA)
y <- c("a", "b", "c", rep(NA, 20))
s_compare(x, .ref_group = y, .in_ref_col = FALSE, .var = "x")

#> Warning: automatically converting character variable x to factor, better manually convert to factor to avoid failures

#> Warning: automatically converting character variable x to factor, better manually convert to factor to avoid failures

#> $n
#> [1] 5
#> 
#> $count
#> $count$a
#> [1] 3
#> 
#> $count$b
#> [1] 1
#> 
#> $count$c
#> [1] 1
#> 
#> 
#> $count_fraction
#> $count_fraction$a
#> [1] 3.0 0.6
#> 
#> $count_fraction$b
#> [1] 1.0 0.2
#> 
#> $count_fraction$c
#> [1] 1.0 0.2
#> 
#> 
#> $pval
#> [1] 0.7659283
#> 

s_compare(x, .ref_group = y, .in_ref_col = FALSE, .var = "x", na.rm = FALSE)

#> Warning: automatically converting character variable x to factor, better manually convert to factor to avoid failures

#> Warning: automatically converting character variable x to factor, better manually convert to factor to avoid failures

#> $n
#> [1] 6
#> 
#> $count
#> $count$a
#> [1] 3
#> 
#> $count$b
#> [1] 1
#> 
#> $count$c
#> [1] 1
#> 
#> $count$`<Missing>`
#> [1] 1
#> 
#> 
#> $count_fraction
#> $count_fraction$a
#> [1] 3.0 0.5
#> 
#> $count_fraction$b
#> [1] 1.0000000 0.1666667
#> 
#> $count_fraction$c
#> [1] 1.0000000 0.1666667
#> 
#> $count_fraction$`<Missing>`
#> [1] 1.0000000 0.1666667
#> 
#> 
#> $pval
#> [1] 0.005768471
#> 

# `s_compare.logical`

## Basic usage:
x <- c(TRUE, FALSE, TRUE, TRUE)
y <- c(FALSE, FALSE, TRUE)
s_compare(x, .ref_group = y, .in_ref_col = FALSE)

#> $n
#> [1] 4
#> 
#> $count
#> [1] 3
#> 
#> $count_fraction
#> [1] 3.00 0.75
#> 
#> $pval
#> [1] 0.2702894
#> 

## Management of NA values.
x <- c(NA, TRUE, FALSE)
y <- c(NA, NA, NA, NA, FALSE)
s_compare(x, .ref_group = y, .in_ref_col = FALSE, na.rm = TRUE)

#> $n
#> [1] 2
#> 
#> $count
#> [1] 1
#> 
#> $count_fraction
#> [1] 1.0 0.5
#> 
#> $pval
#> [1] 0.3864762
#> 

s_compare(x, .ref_group = y, .in_ref_col = FALSE, na.rm = FALSE)

#> $n
#> [1] 3
#> 
#> $count
#> [1] 1
#> 
#> $count_fraction
#> [1] 1.0000000 0.3333333
#> 
#> $pval
#> [1] 0.1675463
#> 

# `a_compare.numeric`
a_compare(
  rnorm(10, 5, 1),
  .ref_group = rnorm(20, -5, 1),
  .in_ref_col = FALSE,
  .var = "bla"
)

#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>    row_name formatted_cell indent_mod        row_label
#> 1         n             10          0                n
#> 2   mean_sd      4.9 (1.1)          0        Mean (SD)
#> 3    median            4.7          0           Median
#> 4     range      3.7 - 6.9          0        Min - Max
#> 5   mean_ci   (4.12, 5.63)          0      Mean 95% CI
#> 6 median_ci   (3.86, 6.23)          0    Median 95% CI
#> 7      pval        <0.0001          0 p-value (t-test)

# `a_compare.factor`
# We need to ungroup `count` and `count_fraction` first so that the rtables formatting
# functions can be applied correctly.
afun <- make_afun(
  getS3method("a_compare", "factor"),
  .ungroup_stats = c("count", "count_fraction")
)
x <- factor(c("a", "a", "b", "c", "a"))
y <- factor(c("a", "a", "b", "c"))
afun(x, .ref_group = y, .in_ref_col = FALSE)

#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>   row_name formatted_cell indent_mod                  row_label
#> 1        n              5          0                          n
#> 2        a              3          0                          a
#> 3        b              1          0                          b
#> 4        c              1          0                          c
#> 5        a        3 (60%)          0                          a
#> 6        b        1 (20%)          0                          b
#> 7        c        1 (20%)          0                          c
#> 8     pval         0.9560          0 p-value (chi-squared test)

# `a_compare.character`
afun <- make_afun(
  getS3method("a_compare", "character"),
  .ungroup_stats = c("count", "count_fraction")
)
x <- c("A", "B", "A", "C")
y <- c("B", "A", "C")
afun(x, .ref_group = y, .in_ref_col = FALSE, .var = "x")

#> Warning: automatically converting character variable x to factor, better manually convert to factor to avoid failures

#> Warning: automatically converting character variable x to factor, better manually convert to factor to avoid failures

#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>   row_name formatted_cell indent_mod                  row_label
#> 1        n              4          0                          n
#> 2        A              2          0                          A
#> 3        B              1          0                          B
#> 4        C              1          0                          C
#> 5        A        2 (50%)          0                          A
#> 6        B        1 (25%)          0                          B
#> 7        C        1 (25%)          0                          C
#> 8     pval         0.9074          0 p-value (chi-squared test)

# `a_compare.logical`
afun <- make_afun(
  getS3method("a_compare", "logical")
)
x <- c(TRUE, FALSE, FALSE, TRUE, TRUE)
y <- c(TRUE, FALSE)
afun(x, .ref_group = y, .in_ref_col = FALSE)

#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>         row_name formatted_cell indent_mod                  row_label
#> 1              n              5          0                          n
#> 2          count              3          0                      count
#> 3 count_fraction        3 (60%)          0             count_fraction
#> 4           pval         0.8091          0 p-value (chi-squared test)

# `create_afun_compare()` to create combined `afun`

afun <- create_afun_compare(
  .stats = c("n", "count_fraction", "mean_sd", "pval"),
  .indent_mods = c(pval = 1L)
)

lyt <- basic_table() %>%
  split_cols_by("ARMCD", ref_group = "ARM A") %>%
  analyze(
    "AGE",
    afun = afun,
    show_labels = "visible"
  )
build_table(lyt, df = ex_adsl)

#>                          ARM A        ARM B        ARM C   
#> -----------------------------------------------------------
#> AGE                                                        
#>   n                       134          134          132    
#>   Mean (SD)            33.8 (6.6)   35.4 (7.9)   35.4 (7.7)
#>     p-value (t-test)                  0.0616       0.0595  

lyt <- basic_table() %>%
  split_cols_by("ARMCD", ref_group = "ARM A") %>%
  analyze(
    "SEX",
    afun = afun,
    show_labels = "visible"
  )
build_table(lyt, df = ex_adsl)

#>                                    ARM A        ARM B        ARM C   
#> ---------------------------------------------------------------------
#> SEX                                                                  
#>   n                                 134          134          132    
#>   F                               79 (59%)    77 (57.5%)    66 (50%) 
#>   M                              51 (38.1%)    55 (41%)    60 (45.5%)
#>   U                               3 (2.2%)     2 (1.5%)      4 (3%)  
#>   UNDIFFERENTIATED                1 (0.7%)        0         2 (1.5%) 
#>     p-value (chi-squared test)                  0.7110       0.5018  

# `compare_vars()` in rtables pipelines

## Default output within a `rtables` pipeline.
lyt <- basic_table() %>%
  split_cols_by("ARMCD", ref_group = "ARM B") %>%
  compare_vars(c("AGE", "SEX"))
build_table(lyt, ex_adsl)

#>                                  ARM B        ARM A        ARM C   
#> -------------------------------------------------------------------
#> AGE                                                                
#>   n                               134          134          132    
#>   Mean (SD)                    35.4 (7.9)   33.8 (6.6)   35.4 (7.7)
#>   p-value (t-test)                            0.0616       0.9992  
#> SEX                                                                
#>   n                               134          134          132    
#>   F                            77 (57.5%)    79 (59%)     66 (50%) 
#>   M                             55 (41%)    51 (38.1%)   60 (45.5%)
#>   U                             2 (1.5%)     3 (2.2%)      4 (3%)  
#>   UNDIFFERENTIATED                 0         1 (0.7%)     2 (1.5%) 
#>   p-value (chi-squared test)                  0.7110       0.2939  

## Select and format statistics output.
lyt <- basic_table() %>%
  split_cols_by("ARMCD", ref_group = "ARM C") %>%
  compare_vars(
    vars = "AGE",
    .stats = c("mean_sd", "pval"),
    .formats = c(mean_sd = "xx.x, xx.x"),
    .labels = c(mean_sd = "Mean, SD")
  )
build_table(lyt, df = ex_adsl)

#>                      ARM C       ARM A       ARM B  
#> ----------------------------------------------------
#> Mean, SD           35.4, 7.7   33.8, 6.6   35.4, 7.9
#> p-value (t-test)                0.0595      0.9992