Estimation of Proportions per Level of Factor

Estimate the proportion along with confidence interval of a proportion regarding the level of a factor.

s_length_proportion(x, .N_col, ...)

a_length_proportion(x, .N_col, ...)

estimate_multinomial_response(
  lyt,
  var,
  ...,
  show_labels = "hidden",
  table_names = var,
  .stats = "prop_ci",
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)

Arguments

x
(numeric)
vector of numbers we want to analyze.

.N_col
(count)
row-wise N (row group count) for the group of observations being analyzed (i.e. with no column-based subsetting) that is passed by rtables.

...
additional arguments for the lower level functions.

lyt
(layout)
input layout where analyses will be added to.

var
(string)
single variable name for the primary analysis variable.

show_labels
label visibility: one of "default", "visible" and "hidden".

table_names
(character)
this can be customized in case that the same vars are analyzed multiple times, to avoid warnings from rtables.

.stats
(character)
statistics to select for the table.

.formats
(named character or list)
formats for the statistics.

.labels
(named character)
labels for the statistics (without indent).

.indent_mods
(named integer)
indent modifiers for the labels.

Value

See s_proportion() for statistics and additional possible arguments.

Functions

s_length_proportion: Statistics function which takes the length of the input x and takes that as the number of successes, and the column number .N_col as the total number, and feeds that into s_proportion().

a_length_proportion: Formatted Analysis function which can be further customized by calling rtables::make_afun() on it. It is used as afun in rtables::analyze().

estimate_multinomial_response: Analyze Function which adds the multinomial proportion analysis to the input layout. Note that additional formatting arguments can be used here.

Examples

s_length_proportion(rep("CR", 10), .N_col = 100)

#> $n_prop
#> [1] 10.0  0.1
#> attr(,"label")
#> [1] "Responders"
#> 
#> $prop_ci
#> [1]  3.620108 16.379892
#> attr(,"label")
#> [1] "95% CI (Wald, with correction)"
#> 

s_length_proportion(factor(character(0)), .N_col = 100)

#> $n_prop
#> [1] 0 0
#> attr(,"label")
#> [1] "Responders"
#> 
#> $prop_ci
#> [1] 0.0 0.5
#> attr(,"label")
#> [1] "95% CI (Wald, with correction)"
#> 

a_length_proportion(rep("CR", 10), .N_col = 100)

#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>   row_name formatted_cell indent_mod                      row_label
#> 1   n_prop       10 (10%)          0                     Responders
#> 2  prop_ci  (3.62, 16.38)          0 95% CI (Wald, with correction)

a_length_proportion(factor(character(0)), .N_col = 100)

#> Error in sfunargs[[var]] <- get(var): replacement has length zero

library(dplyr)
# Use of the layout creating function.
dta_test <- data.frame(
  USUBJID = paste0("S", 1:12),
  ARM     = factor(rep(LETTERS[1:3], each = 4)),
  AVAL    = c(A = c(1, 1, 1, 1), B = c(0, 0, 1, 1), C = c(0, 0, 0, 0))
) %>% mutate(
   AVALC = factor(AVAL, levels = c(0, 1),
                 labels = c("Complete Response (CR)", "Partial Response (PR)"))
)

lyt <- basic_table() %>%
  split_cols_by("ARM") %>%
  estimate_multinomial_response(var = "AVALC")

tbl <- build_table(lyt, dta_test)

#> Error in sfunargs[[var]] <- get(var): replacement has length zero

html <- as_html(tbl)

#> Error in as_html(tbl): is(x, "VTableTree") is not TRUE

html

#> Error in eval(expr, envir, enclos): object 'html' not found

if (FALSE) {
Viewer(html)
}