Diagnostic Plots for MMRM

This function produces diagnostic plots.

g_mmrm_diagnostic(
  object,
  type = c("fit-residual", "q-q-residual"),
  z_threshold = NULL
)

Arguments

object

(mmrm)
model result produced by fit_mmrm.

type

(string)
specifying the type of diagnostic plot to be produced:

fit-residual

A fitted vs residuals plot, grouped by visits. This allows to see if there is remaining structure in the residuals that might be captured by adding additional covariates to the model.

q-q-residual

A Q-Q plot for the residuals (i.e. sorted standardized residuals vs. normal quantiles), grouped by visits. Observations with an absolute standardized residual above z_threshold will be labeled.

z_threshold

(numeric)
optional number indicating the normal quantile threshold for the Q-Q plot.

Value

a ggplot2 plot

Details

Here we use marginal fitted values and residuals. That is, only the fixed effects are used to estimate fitted values, and the difference of those fitted values vs. the observed data are the residuals. This is in line with the MMRM philosophy where random effects are only used to model the marginal residual distribution.

See also

Examples

if (FALSE) {
library(dplyr)
library(scda)

adsl <- synthetic_cdisc_data("latest")$adsl
adqs <- synthetic_cdisc_data("latest")$adqs
adqs_f <- adqs %>%
  filter(PARAMCD == "FKSI-FWB" & !AVISIT %in% c("BASELINE")) %>%
  droplevels() %>%
  mutate(ARM = factor(ARM, levels = c("B: Placebo", "A: Drug X", "C: Combination"))) %>%
  mutate(AVISITN = rank(AVISITN) %>% as.factor() %>% as.numeric() %>% as.factor())

mmrm_results <- fit_mmrm(
  vars = list(
    response = "AVAL",
    covariates = c("STRATA1", "BMRKR2"),
    id = "USUBJID",
    arm = "ARM",
    visit = "AVISIT"
  ),
  data = adqs_f,
  cor_struct = "random-quadratic",
  weights_emmeans = "equal"
)
g_mmrm_diagnostic(mmrm_results)
g_mmrm_diagnostic(mmrm_results, type = "q-q-residual")
}