Kaplan-Meier Plot

Kaplan-Meier Plot

Arguments

df
(data frame)
data set containing all analysis variables.

variables
(named list) of variable names. Details are:

tte: variable indicating time-to-event duration values (numeric).

is_event: event variable (logical)
TRUE if event, FALSE if time to event is censored.

arm: the treatment group variable (factor).

strat: (character or NULL) variable names indicating stratification factors.

control_surv
a (list) of parameters for comparison details, specified by using
the helper function control_surv_timepoint. Some possible parameter options are:

conf_level: (proportion)
confidence level of the interval for survival rate.

conf_type: (string)
"plain" (default), "log", "log-log" for confidence interval type,
see more in survival::survfit(). Note that the option "none" is no longer supported.

data
(data.frame)
survival data as pre-processed by h_data_plot.

xticks
(numeric, number, or NULL)
numeric vector of ticks or single number with spacing between ticks on the x axis. If NULL (default), labeling::extended() is used to determine an optimal tick position on the x axis.

yval
(string)
value of y-axis. Should be either Survival (default) or Failure probability.

censor_show
(flag)
whether to show censored.

xlab
(string)
label of x-axis.

ylab
(string)
label of y-axis.

title
(string)
title for plot.

col
(character)
lines colors. Length of a vector should be equal to number of strata from survival::survfit().

lty
(numeric)
line type. Length of a vector should be equal to number of strata from survival::survfit().

lwd
(numeric)
line width. Length of a vector should be equal to number of strata from survival::survfit().

pch
(numeric, string)
value or character of points symbol to indicate censored cases.

size
(numeric)
size of censored point, a class of unit.

max_time
(numeric)
maximum value to show on X axis. Only data values less than or up to to this threshold value will be plotted.(NULL for default)

font_size
(number)
font size to be used.

ci_ribbon
(flag)
draw the confidence interval around the Kaplan-Meier curve.

ggtheme
(theme)
a graphical theme as provided by ggplot2 to control outlook of the Kaplan-Meier curve.

annot_at_risk
(flag)
compute and add the annotation table reporting the number of patient at risk matching the main grid of the Kaplan-Meier curve.

annot_surv_med
(flag)
compute and add the annotation table on the Kaplan-Meier curve estimating the median survival time per group.

annot_coxph
(flag)
add the annotation table from a survival::coxph() model.

control_coxph_pw
(list)
parameters for comparison details, specified by using
the helper function control_coxph(). Some possible parameter options are:

pval_method: (string)
p-value method for testing hazard ratio = 1. Default method is "log-rank", can also be set to "wald" or "likelihood".

ties: (string)
specifying the method for tie handling. Default is "efron", can also be set to "breslow" or "exact". See more in survival::coxph()

conf_level: (proportion)
confidence level of the interval for HR.

position_coxph
numeric
x and y positions for plotting survival::coxph() model.

position_surv_med
numeric
x and y positions for plotting annotation table estimating median survival time per group