List of Adverse Events Terms Tables

lt_ae_max_grade_class_term returns a nested list of adverse events tables by unique id (t_count_unique).

lt_ae_class_term(terms, id, col_by, col_N, total = "All Patients")

Arguments

terms
term information as character or factor dataframe, however factor levels are not repeated by class, only terms with count > 1 are listed per class. Currently terms can only be a dataframe with 2 columns. var_relabel is used as the character string used as a label in the column header for each term.

id
unique subject identifier. If a particular subject has no adverse event then that information needs to be added to the col_N argument.

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing. See 'Examples'.

col_N
numeric vector with information of the number of patients in the levels of col_by. This is useful if there are patients that have no adverse events can be accounted for with this argument.

total
character string that will be used as a label for a column with pooled total population, default is "All Patients". If the levels of col_by are the only columns of interest then total should be NULL

Examples

library(dplyr)
library(random.cdisc.data)

ASL <- rasl(10, seed = 1)
AAE <- raae(ASL, 4, seed = 2)

ANL <- left_join(AAE, ASL %>% select(USUBJID, STUDYID, ARM), by = c("USUBJID", "STUDYID"))

l_tbls <- tern:::lt_ae_class_term(
  terms = ANL[, c("AEBODSYS", "AEDECOD")],
  id = ANL$USUBJID,
  col_by = ANL$ARM,
  col_N = table(ASL$ARM),
  total = "All Patients"
)
recursive_stack_rtables(l_tbls)

#>                                                                             A: Drug X             B: Placebo          C: Combination         All Patients 
#> label                                                                         (N=1)                 (N=4)                 (N=5)                 (N=10)    
#> ----------------------------------------------------------------------------------------------------------------------------------------------------------
#> - Any Adverse Event -
#>   Overall total number of events                                                1                     5                     10                    16      
#>   Overall total number of patients with at least one adverse event           1 (100%)              2 (50%)               4 (80%)               7 (70%)    
#> 
#> cl A
#>   Total number of events                                                        0                     2                     2                     4       
#>   Total number of patients with at least one adverse event                    0 (0%)               2 (50%)               2 (40%)               4 (40%)    
#>   trm A_1/2                                                                   0 (0%)                0 (0%)               2 (40%)               2 (20%)    
#>   trm A_2/2                                                                   0 (0%)               2 (50%)                0 (0%)               2 (20%)    
#> 
#> cl B
#>   Total number of events                                                        1                     2                     4                     7       
#>   Total number of patients with at least one adverse event                   1 (100%)              2 (50%)               3 (60%)               6 (60%)    
#>   trm B_1/3                                                                  1 (100%)               0 (0%)               3 (60%)               4 (40%)    
#>   trm B_2/3                                                                   0 (0%)               1 (25%)               1 (20%)               2 (20%)    
#>   trm B_3/3                                                                   0 (0%)               1 (25%)                0 (0%)               1 (10%)    
#> 
#> cl C
#>   Total number of events                                                        0                     1                     4                     5       
#>   Total number of patients with at least one adverse event                    0 (0%)               1 (25%)               3 (60%)               4 (40%)    
#>   trm C_1/1                                                                   0 (0%)               1 (25%)               3 (60%)               4 (40%)