Summary of logistic regression with one two-way interaction term

Summary of logistic regression with one two-way interaction term

s_logistic_interaction(glm_model, conf_level = 0.95, increments = NULL)

Arguments

glm_model
(glm model object)
The model object can be all main effect model, and a model with one two-way interaction.

conf_level
(numeric value)
Confidence level for Wald odds ratio confidence interval.

increments
(named list)
Used to specify numeric values of continuous variables in glm_model which interact with other variables. This is used to calculate the odds ratio when comparing the other interaction variable effect. For example, for a model with ARM and AGE interaction, increments = list(AGE = c(18, 65)) will enable calculation of odds ratios of comparison ARM vs. reference ARM at AGE = 18 and AGE = 65. If increments = NULL, then default AGE value is ceiling of median.

Examples

library(random.cdisc.data)
library(dplyr)
library(purrr)

#> 
#> Attaching package: ‘purrr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     is_null

ADSL <- radsl(cached = TRUE)
ADSL <- ADSL %>% mutate(
 SEX = as.character(SEX),
 SEX = case_when(!SEX %in% c("F", "M") ~ "U",
               TRUE ~ SEX))
ADRS <- radrs(ADSL, seed = 2)
ADRS_f <- subset(ADRS, PARAMCD == "BESRSPI") %>%
  mutate(Response = case_when(AVALC %in% c("PR", "CR") ~ 1,
                              TRUE ~ 0))
glm_model <- glm(
 formula = Response ~ ARM + AGE + SEX + ARM*SEX,
 data = ADRS_f,
 family = "binomial")
s_logistic_interaction(glm_model, conf_level = 0.90)

#> $N
#> [1] 400
#> 
#> $results
#> $results$ARM
#> $results$ARM$main
#>     df    p-value
#> ARM  2 1.1105e-06
#> 
#> $results$ARM$predictor
#> $results$ARM$predictor$term_type
#> [1] "categorical"
#> 
#> $results$ARM$predictor$interact_with
#> [1] "SEX"
#> 
#> $results$ARM$predictor$counts_by_level
#> 
#>      A: Drug X     B: Placebo C: Combination 
#>            134            134            132 
#> 
#> $results$ARM$predictor$term_ref_level
#> [1] "A: Drug X"
#> 
#> $results$ARM$predictor$term_comp_level
#> [1] "B: Placebo"     "C: Combination"
#> 
#> 
#> $results$ARM$summary
#> $results$ARM$summary$`B: Placebo`
#> $results$ARM$summary$`B: Placebo`$summary_comp_level
#>                    coef        se      p-value df
#> ARMB: Placebo -2.106488 0.5202531 5.144427e-05  1
#> 
#> $results$ARM$summary$`B: Placebo`$summary_with_interaction
#> $results$ARM$summary$`B: Placebo`$summary_with_interaction$F
#>          or        lcl       ucl
#> 1 0.1216646 0.05170361 0.2862908
#> 
#> $results$ARM$summary$`B: Placebo`$summary_with_interaction$M
#>          or       lcl       ucl
#> 1 0.2540055 0.1017433 0.6341329
#> 
#> $results$ARM$summary$`B: Placebo`$summary_with_interaction$U
#>             or lcl ucl
#> 1 1.745212e-07   0 Inf
#> 
#> 
#> 
#> $results$ARM$summary$`C: Combination`
#> $results$ARM$summary$`C: Combination`$summary_comp_level
#>                        coef        se   p-value df
#> ARMC: Combination 0.4202822 0.7530905 0.5767921  1
#> 
#> $results$ARM$summary$`C: Combination`$summary_with_interaction
#> $results$ARM$summary$`C: Combination`$summary_with_interaction$F
#>         or       lcl      ucl
#> 1 1.522391 0.4411186 5.254085
#> 
#> $results$ARM$summary$`C: Combination`$summary_with_interaction$M
#>         or       lcl      ucl
#> 1 2.072335 0.5957528 7.208651
#> 
#> $results$ARM$summary$`C: Combination`$summary_with_interaction$U
#>            or lcl ucl
#> 1 1.06759e-06   0 Inf
#> 
#> 
#> 
#> 
#> 
#> $results$AGE
#> $results$AGE$main
#>     df p-value
#> AGE  1 0.18072
#> 
#> $results$AGE$predictor
#> $results$AGE$predictor$term_type
#> [1] "continuous"
#> 
#> $results$AGE$predictor$term_ref_level
#> NULL
#> 
#> $results$AGE$predictor$counts_by_level
#> NULL
#> 
#> 
#> $results$AGE$summary
#>           coef         se   p-value        or df       lcl      ucl
#> AGE -0.0260952 0.01949523 0.1807196 0.9742423  1 0.9434971 1.005989
#> 
#> 
#> $results$SEX
#> $results$SEX$main
#>     df p-value
#> SEX  2 0.83386
#> 
#> $results$SEX$predictor
#> $results$SEX$predictor$term_type
#> [1] "categorical"
#> 
#> $results$SEX$predictor$interact_with
#> [1] "ARM"
#> 
#> $results$SEX$predictor$counts_by_level
#> 
#>   F   M   U 
#> 222 166  12 
#> 
#> $results$SEX$predictor$term_ref_level
#> [1] "F"
#> 
#> $results$SEX$predictor$term_comp_level
#> [1] "M" "U"
#> 
#> 
#> $results$SEX$summary
#> $results$SEX$summary$M
#> $results$SEX$summary$M$summary_comp_level
#>            coef       se   p-value df
#> SEXM -0.3995479 0.663093 0.5468069  1
#> 
#> $results$SEX$summary$M$summary_with_interaction
#> $results$SEX$summary$M$summary_with_interaction$`A: Drug X`
#>          or       lcl      ucl
#> 1 0.6706232 0.2253189 1.995996
#> 
#> $results$SEX$summary$M$summary_with_interaction$`B: Placebo`
#>         or       lcl      ucl
#> 1 1.400095 0.7437404 2.635687
#> 
#> $results$SEX$summary$M$summary_with_interaction$`C: Combination`
#>          or       lcl     ucl
#> 1 0.9128772 0.2301136 3.62145
#> 
#> 
#> 
#> $results$SEX$summary$U
#> $results$SEX$summary$U$summary_comp_level
#>         coef       se   p-value df
#> SEXU 12.8097 727.3559 0.9859489  1
#> 
#> $results$SEX$summary$U$summary_with_interaction
#> $results$SEX$summary$U$summary_with_interaction$`A: Drug X`
#>         or lcl ucl
#> 1 365748.7   0 Inf
#> 
#> $results$SEX$summary$U$summary_with_interaction$`B: Placebo`
#>          or        lcl      ucl
#> 1 0.5246465 0.04931486 5.581562
#> 
#> $results$SEX$summary$U$summary_with_interaction$`C: Combination`
#>          or        lcl      ucl
#> 1 0.2564844 0.03298045 1.994643
#> 
#> 
#> 
#> 
#> 
#> $results$`ARM:SEX`
#> $results$`ARM:SEX`$main
#>         df p-value
#> ARM:SEX  4 0.89563
#> 
#> $results$`ARM:SEX`$predictor
#> $results$`ARM:SEX`$predictor$term_type
#> [1] "interaction"
#> 
#> 
#> $results$`ARM:SEX`$summary
#>                                    coef          se   p-value df
#> B: Placebo * M , n = 55       0.7360881   0.7609973 0.3334102  1
#> C: Combination * M , n = 60   0.3083940   1.0678940 0.7727443  1
#> B: Placebo * U , n = 2      -13.4547322 727.3573030 0.9852415  1
#> C: Combination * U , n = 6  -14.1703892 727.3569595 0.9844566  1
#> 
#> 
#> 

glm_model2 <- glm(
 formula = Response ~ ARM + AGE + ARM*AGE,
 data = ADRS_f,
 family = "binomial")
s_logistic_interaction(glm_model2, conf_level = 0.90)

#> $N
#> [1] 400
#> 
#> $results
#> $results$ARM
#> $results$ARM$main
#>     df  p-value
#> ARM  2 0.080769
#> 
#> $results$ARM$predictor
#> $results$ARM$predictor$term_type
#> [1] "categorical"
#> 
#> $results$ARM$predictor$interact_with
#> [1] "AGE"
#> 
#> $results$ARM$predictor$counts_by_level
#> 
#>      A: Drug X     B: Placebo C: Combination 
#>            134            134            132 
#> 
#> $results$ARM$predictor$term_ref_level
#> [1] "A: Drug X"
#> 
#> $results$ARM$predictor$term_comp_level
#> [1] "B: Placebo"     "C: Combination"
#> 
#> 
#> $results$ARM$summary
#> $results$ARM$summary$`B: Placebo`
#> $results$ARM$summary$`B: Placebo`$summary_comp_level
#>                    coef       se    p-value df
#> ARMB: Placebo -3.990531 2.012629 0.04739583  1
#> 
#> $results$ARM$summary$`B: Placebo`$summary_with_interaction
#> $results$ARM$summary$`B: Placebo`$summary_with_interaction$`34`
#>          or        lcl       ucl
#> 1 0.1448359 0.07502125 0.2796198
#> 
#> 
#> 
#> $results$ARM$summary$`C: Combination`
#> $results$ARM$summary$`C: Combination`$summary_comp_level
#>                        coef       se   p-value df
#> ARMC: Combination -1.311594 2.521592 0.6029625  1
#> 
#> $results$ARM$summary$`C: Combination`$summary_with_interaction
#> $results$ARM$summary$`C: Combination`$summary_with_interaction$`34`
#>         or       lcl      ucl
#> 1 1.398218 0.5731557 3.410966
#> 
#> 
#> 
#> 
#> 
#> $results$AGE
#> $results$AGE$main
#>     df p-value
#> AGE  1 0.13437
#> 
#> $results$AGE$predictor
#> $results$AGE$predictor$term_type
#> [1] "continuous"
#> 
#> $results$AGE$predictor$interact_with
#> [1] "ARM"
#> 
#> 
#> $results$AGE$summary
#> $results$AGE$summary$summary_term
#>            coef         se   p-value df
#> AGE -0.07381638 0.04930643 0.1343687  1
#> 
#> $results$AGE$summary$summary_with_interaction
#> $results$AGE$summary$summary_with_interaction$`A: Drug X`
#>          or       lcl      ucl
#> 1 0.9288422 0.8564852 1.007312
#> 
#> $results$AGE$summary$summary_with_interaction$`B: Placebo`
#>          or       lcl      ucl
#> 1 0.9868119 0.9503108 1.024715
#> 
#> $results$AGE$summary$summary_with_interaction$`C: Combination`
#>          or       lcl      ucl
#> 1 0.9749381 0.9041373 1.051283
#> 
#> 
#> 
#> 
#> $results$`ARM:AGE`
#> $results$`ARM:AGE`$main
#>         df p-value
#> ARM:AGE  2  0.5378
#> 
#> $results$`ARM:AGE`$predictor
#> $results$`ARM:AGE`$predictor$term_type
#> [1] "interaction"
#> 
#> 
#> $results$`ARM:AGE`$summary
#>                               coef         se   p-value df
#> B: Placebo, n = 134     0.06054051 0.05437075 0.2655042  1
#> C: Combination, n = 132 0.04843508 0.06732026 0.4718498  1
#> 
#> 
#> 

glm_model3 <- glm(
 formula = Response ~ ARM + AGE + BMRKR1 + BMRKR1*AGE,
 data = ADRS_f,
 family = "binomial")
s_logistic_interaction(glm_model3 , conf_level = 0.90)

#> $N
#> [1] 400
#> 
#> $results
#> $results$ARM
#> $results$ARM$main
#>     df    p-value
#> ARM  2 6.9797e-10
#> 
#> $results$ARM$predictor
#> $results$ARM$predictor$term_type
#> [1] "categorical"
#> 
#> $results$ARM$predictor$term_ref_level
#> [1] "A: Drug X"
#> 
#> $results$ARM$predictor$counts_by_level
#> 
#>      A: Drug X     B: Placebo C: Combination 
#>            134            134            132 
#> 
#> 
#> $results$ARM$summary
#>                      coef        se      p-value        or df        lcl
#> B: Placebo     -1.8252017 0.3778217 1.359484e-06 0.1611851  1 0.08658202
#> C: Combination  0.4319742 0.5116987 3.985597e-01 1.5402953  1 0.66385248
#>                      ucl
#> B: Placebo     0.3000698
#> C: Combination 3.5738507
#> 
#> 
#> $results$AGE
#> $results$AGE$main
#>     df p-value
#> AGE  1 0.71208
#> 
#> $results$AGE$predictor
#> $results$AGE$predictor$term_type
#> [1] "continuous"
#> 
#> $results$AGE$predictor$interact_with
#> [1] "BMRKR1"
#> 
#> 
#> $results$AGE$summary
#> $results$AGE$summary$summary_term
#>           coef         se   p-value df
#> AGE 0.01337012 0.03622711 0.7120801  1
#> 
#> $results$AGE$summary$summary_with_interaction
#> $results$AGE$summary$summary_with_interaction$`5`
#>          or       lcl      ucl
#> 1 0.9752905 0.9447992 1.006766
#> 
#> 
#> 
#> 
#> $results$BMRKR1
#> $results$BMRKR1$main
#>        df p-value
#> BMRKR1  1 0.25831
#> 
#> $results$BMRKR1$predictor
#> $results$BMRKR1$predictor$term_type
#> [1] "continuous"
#> 
#> $results$BMRKR1$predictor$interact_with
#> [1] "AGE"
#> 
#> 
#> $results$BMRKR1$summary
#> $results$BMRKR1$summary$summary_term
#>            coef       se   p-value df
#> BMRKR1 0.256421 0.226844 0.2583142  1
#> 
#> $results$BMRKR1$summary$summary_with_interaction
#> $results$BMRKR1$summary$summary_with_interaction$`34`
#>          or       lcl      ucl
#> 1 0.9953797 0.9240681 1.072195
#> 
#> 
#> 
#> 
#> $results$`AGE:BMRKR1`
#> $results$`AGE:BMRKR1`$main
#>            df p-value
#> AGE:BMRKR1  1 0.21754
#> 
#> $results$`AGE:BMRKR1`$predictor
#> $results$`AGE:BMRKR1`$predictor$term_type
#> [1] "interaction"
#> 
#> 
#> $results$`AGE:BMRKR1`$summary
#>                      coef          se   p-value df
#> AGE * BMRKR1 -0.007677998 0.006226607 0.2175404  1
#> 
#> 
#>