Summarize the Change from Baseline or Absolute Baseline Values

The primary analysis variable .var indicates the numerical change from baseline results, and additional required secondary analysis variables are value and baseline_flag. Depending on the baseline flag, either the absolute baseline values (at baseline) or the change from baseline values (post-baseline) are then summarized.

s_change_from_baseline(df, .var, variables, na.rm = TRUE, ...)

a_change_from_baseline(df, .var, variables, na.rm = TRUE, ...)

summarize_change(
  lyt,
  vars,
  ...,
  table_names = vars,
  .stats = c("n", "mean_sd", "median", "range"),
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)

Arguments

df

(data frame)
data set containing all analysis variables.

.var

(string)
single variable name that is passed by rtables when requested by a statistics function.

variables

(named list of string)
list of additional analysis variables.

na.rm

(flag)
whether NA values should be removed from x prior to analysis.

...

additional arguments for the lower level functions.

lyt

(layout)
input layout where analyses will be added to.

vars

(character)
variable names for the primary analysis variable to be iterated over.

table_names

(character)
this can be customized in case that the same vars are analyzed multiple times, to avoid warnings from rtables.

.stats

(character)
statistics to select for the table.

.formats

(named character or list)
formats for the statistics.

.labels

(named character)
labels for the statistics (without indent).

.indent_mods

(named integer)
indent modifiers for the labels.

Value

See s_summary.numeric() for the return values.

Functions

s_change_from_baseline: Statistics Function that summarizes baseline or post-baseline visits.

a_change_from_baseline: Formatted Analysis function which can be further customized by calling rtables::make_afun() on it. It is used as afun in rtables::analyze().

summarize_change: Analyze Function for change from baseline analysis. To be used after a split on visits in the layout, such that each data subset only contains either baseline or post-baseline data. Allows additional formatting options.

Note

The data in df must be either all be from baseline or post-baseline visits. Otherwise an error will be thrown.

Examples

df <- data.frame(
  chg = c(1, 2, 3),
  is_bl = c(TRUE, TRUE, TRUE),
  val = c(4, 5, 6)
)
s_change_from_baseline(
  df,
  .var = "chg",
  variables = list(value = "val", baseline_flag = "is_bl")
)
#> $n
#> n 
#> 3 
#> 
#> $mean
#> mean 
#>    5 
#> 
#> $sd
#> sd 
#>  1 
#> 
#> $se
#>        se 
#> 0.5773503 
#> 
#> $mean_sd
#> mean   sd 
#>    5    1 
#> 
#> $mean_ci
#> mean_ci_lwr mean_ci_upr 
#>    2.515862    7.484138 
#> attr(,"label")
#> [1] "Mean 95% CI"
#> 
#> $mean_sei
#> mean_sei_lwr mean_sei_upr 
#>      4.42265      5.57735 
#> attr(,"label")
#> [1] "Mean -/+ 1xSE"
#> 
#> $mean_sdi
#> mean_sdi_lwr mean_sdi_upr 
#>            4            6 
#> attr(,"label")
#> [1] "Mean -/+ 1xSD"
#> 
#> $median
#> median 
#>      5 
#> 
#> $mad
#> mad 
#>   0 
#> 
#> $median_ci
#> median_ci_lwr median_ci_upr 
#>            NA            NA 
#> attr(,"conf_level")
#> [1] NA
#> attr(,"label")
#> [1] "Median 95% CI"
#> 
#> $quantiles
#> quantile_0.25 quantile_0.75 
#>             4             6 
#> attr(,"label")
#> [1] "25% and 75%-ile"
#> 
#> $iqr
#> iqr 
#>   2 
#> 
#> $range
#> min max 
#>   4   6 
#> 
#> $min
#> min 
#>   4 
#> 
#> $max
#> max 
#>   6 
#> 
#> $cv
#> cv 
#> 20 
#> 
#> $geom_mean
#> geom_mean 
#>  4.932424 
#> 
#> $geom_cv
#>  geom_cv 
#> 20.51865 
#> 
a_change_from_baseline(
  df,
  .var = "chg",
  variables = list(value = "val", baseline_flag = "is_bl")
)
#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>     row_name                     formatted_cell indent_mod       row_label
#> 1          n                                  3          0               n
#> 2       mean                                  5          0            mean
#> 3         sd                                  1          0              sd
#> 4         se                  0.577350269189626          0              se
#> 5    mean_sd                              5 (1)          0       Mean (SD)
#> 6    mean_ci                       (2.52, 7.48)          0     Mean 95% CI
#> 7   mean_sei 4.42264973081037, 5.57735026918963          0   Mean -/+ 1xSE
#> 8   mean_sdi                               4, 6          0   Mean -/+ 1xSD
#> 9     median                                  5          0          Median
#> 10       mad                                  0          0             mad
#> 11 median_ci                           (NA, NA)          0   Median 95% CI
#> 12 quantiles                               4, 6          0 25% and 75%-ile
#> 13       iqr                                  2          0             iqr
#> 14     range                              4 - 6          0       Min - Max
#> 15       min                                  4          0             min
#> 16       max                                  6          0             max
#> 17        cv                                 20          0              cv
#> 18 geom_mean                   4.93242414866094          0       geom_mean
#> 19   geom_cv                   20.5186527600433          0         geom_cv

# `summarize_change()`

## Fabricated dataset.
library(dplyr)

dta_test <- data.frame(
  USUBJID = rep(1:6, each = 3),
  AVISIT = rep(paste0("V", 1:3), 6),
  ARM = rep(LETTERS[1:3], rep(6, 3)),
  AVAL = c(9:1, rep(NA, 9))
) %>%
  mutate(ABLFLL = AVISIT == "V1") %>%
  group_by(USUBJID) %>%
  mutate(
    BLVAL = AVAL[ABLFLL],
    CHG = AVAL - BLVAL
  ) %>%
  ungroup()

results <- basic_table() %>%
  split_cols_by("ARM") %>%
  split_rows_by("AVISIT") %>%
  summarize_change("CHG", variables = list(value = "AVAL", baseline_flag = "ABLFLL")) %>%
  build_table(dta_test)
if (FALSE) {
Viewer(results)
}