Pairwise CoxPH model

Summarize p-value, HR and CIs from stratified or unstratified CoxPH model.

s_coxph_pairwise(
  df,
  .ref_group,
  .in_ref_col,
  .var,
  is_event,
  strat = NULL,
  control = control_coxph()
)

a_coxph_pairwise(
  df,
  .ref_group,
  .in_ref_col,
  .var,
  is_event,
  strat = NULL,
  control = control_coxph()
)

coxph_pairwise(
  lyt,
  vars,
  ...,
  var_labels = "CoxPH",
  show_labels = "visible",
  table_names = vars,
  .stats = NULL,
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)

Arguments

df
(data frame)
data set containing all analysis variables.

.ref_group
(data frame or vector)
the data corresponding to the reference group.

.in_ref_col
(logical)
TRUE when working with the reference level, FALSE otherwise.

.var
(string)
single variable name that is passed by rtables when requested by a statistics function.

is_event
(logical)
TRUE if event, FALSE if time to event is censored.

strat
(character or NULL) variable names indicating stratification factors.

control
(list)
parameters for comparison details, specified by using
the helper function control_coxph(). Some possible parameter options are:

pval_method: (string)
p-value method for testing hazard ratio = 1. Default method is "log-rank", can also be set to "wald" or "likelihood".

ties: (string)
specifying the method for tie handling. Default is "efron", can also be set to "breslow" or "exact". See more in survival::coxph()

conf_level: (proportion)
confidence level of the interval for HR.

lyt
(layout)
input layout where analyses will be added to.

vars
(character)
variable names for the primary analysis variable to be iterated over.

...
additional arguments for the lower level functions.

var_labels
character for label.

show_labels
label visibility: one of "default", "visible" and "hidden".

table_names
(character)
this can be customized in case that the same vars are analyzed multiple times, to avoid warnings from rtables.

.stats
(character)
statistics to select for the table.

.formats
(named character or list)
formats for the statistics.

.labels
(named character)
labels for the statistics (without indent).

.indent_mods
(named integer)
indent modifiers for the labels.

Value

The statistics are:

pvalue : p-value to test HR = 1.

hr : hazard ratio.

hr_ci : confidence interval for hazard ratio.

Functions

s_coxph_pairwise: Statistics Function which analyzes HR, CIs of HR and p-value with coxph model.

a_coxph_pairwise: Formatted Analysis function which can be further customized by calling rtables::make_afun() on it. It is used as afun in rtables::analyze().

coxph_pairwise: Analyze Function which adds the pairwise coxph analysis to the input layout. Note that additional formatting arguments can be used here.

Examples

library(random.cdisc.data)
library(dplyr)

ADTTE <- radtte(cached = TRUE)
ADTTE_f <- ADTTE %>%
  filter(PARAMCD == "OS") %>%
  mutate(is_event = CNSR == 0)
df <- ADTTE_f %>%
  filter(ARMCD == "ARM A")
df_ref_group <- ADTTE_f %>%
  filter(ARMCD == "ARM B")

s_coxph_pairwise(df, df_ref_group, .in_ref_col = FALSE, .var = "AVAL", is_event = "is_event")

#> $pvalue
#> [1] 0.03340293
#> attr(,"label")
#> [1] "p-value (log-rank)"
#> 
#> $hr
#> [1] 0.7173651
#> attr(,"label")
#> [1] "Hazard Ratio"
#> 
#> $hr_ci
#> [1] 0.5275231 0.9755262
#> attr(,"label")
#> [1] "95% CI"
#> 

a_coxph_pairwise(df, df_ref_group, .in_ref_col = FALSE, .var = "AVAL", is_event = "is_event")

#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>   row_name formatted_cell indent_mod          row_label
#> 1   pvalue         0.0334          0 p-value (log-rank)
#> 2       hr           0.72          0       Hazard Ratio
#> 3    hr_ci   (0.53, 0.98)          0             95% CI

split_cols_by(lyt = NULL, var = "ARMCD", ref_group = "ARM A") %>%
  add_colcounts() %>%
  coxph_pairwise(
    vars = "AVAL",
    is_event = "is_event",
    var_labels = "Unstratified Analysis"
  ) %>%
  build_table(df = ADTTE_f)

#>                          ARM A       ARM B         ARM C   
#>                         (N=134)     (N=134)       (N=132)  
#> -----------------------------------------------------------
#> Unstratified Analysis                                      
#>   p-value (log-rank)                0.0334        <0.0001  
#>   Hazard Ratio                       1.39          2.75    
#>     95% CI                        (1.03, 1.9)   (2.05, 3.7)

split_cols_by(lyt = NULL, var = "ARMCD", ref_group = "ARM A") %>%
  add_colcounts() %>%
  coxph_pairwise(
    vars = "AVAL",
    is_event = "is_event",
    var_labels = "Stratified Analysis",
    strat = "SEX",
    control = control_coxph(pval_method = "wald")
  ) %>%
  build_table(df = ADTTE_f)

#>                        ARM A      ARM B        ARM C    
#>                       (N=134)    (N=134)      (N=132)   
#> --------------------------------------------------------
#> Stratified Analysis                                     
#>   p-value (wald)                 0.0487       <0.0001   
#>   Hazard Ratio                    1.36          2.73    
#>     95% CI                      (1, 1.86)   (2.02, 3.69)