Survival Time Analysis

Summarize median survival time and CIs, percentiles of survival times,
survival time range of censored/event patients.

Usage

s_surv_time(df, .var, is_event, control = control_surv_time())

a_surv_time(df, .var, is_event, control = control_surv_time())

surv_time(
  lyt,
  vars,
  ...,
  var_labels = "Time to Event",
  table_names = vars,
  .stats = c("median", "median_ci", "quantiles", "range_censor", "range_event"),
  .formats = NULL,
  .labels = NULL,
  .indent_mods = NULL
)

Arguments

df

(data frame)
data set containing all analysis variables.

.var

(string)
single variable name that is passed by rtables when requested by a statistics function.

is_event

(logical)
TRUE if event, FALSE if time to event is censored.

control

a (list) of parameters for comparison details, specified by using
the helper function control_surv_time. Some possible parameter options are:

conf_level: (proportion)
confidence level of the interval for survival time.

conf_type: (string)
"plain" (default), "log", "log-log" for confidence interval type,
see more in survival::survfit(). Note option, "none" is not supported.

quantiles: numeric vector of length two to specify the quantiles of survival time.

lyt

(layout)
input layout where analyses will be added to.

vars

(character)
variable names for the primary analysis variable to be iterated over.

...

additional arguments for the lower level functions.

var_labels

character for label.

table_names

(character)
this can be customized in case that the same vars are analyzed multiple times, to avoid warnings from rtables.

.stats

(character)
statistics to select for the table.

.formats

(named character or list)
formats for the statistics.

.labels

(named character)
labels for the statistics (without indent).

.indent_mods

(named integer)
indent modifiers for the labels.

Value

The statistics are:

median : median survival time.

median_ci : confidence interval for median time.

quantiles : survival time for two specified quantiles.

range_censor : survival time range for censored observations.

range_event : survival time range for observations with events.

range : survival time range for all observations.

Functions

s_surv_time: Statistics Function which analyzes survival times. range_censor and range_event.

a_surv_time: Formatted Analysis function which can be further customized by calling rtables::make_afun() on it. It is used as afun in rtables::analyze().

surv_time: Analyze Function which adds the survival times analysis to the input layout. Note that additional formatting arguments can be used here.

Examples

library(scda)
library(dplyr)

ADTTE <- synthetic_cdisc_data("latest")$adtte
ADTTE_f <- ADTTE %>%
  filter(PARAMCD == "OS") %>%
  mutate(
    AVAL = day2month(AVAL),
    is_event = CNSR == 0
  )
df <- ADTTE_f %>% filter(ARMCD == "ARM A")
s_surv_time(df, .var = "AVAL", is_event = "is_event")
#> $median
#> [1] 41.41242
#> attr(,"label")
#> [1] "Median"
#> 
#> $median_ci
#> [1] 27.65076 54.73122
#> attr(,"label")
#> [1] "95% CI"
#> 
#> $quantiles
#> [1] 15.40087 75.24796
#> attr(,"label")
#> [1] "25% and 75%-ile"
#> 
#> $range_censor
#> [1]   0.3958452 154.7000110
#> attr(,"label")
#> [1] "Range (censored)"
#> 
#> $range_event
#> [1]   0.2605094 116.3984412
#> attr(,"label")
#> [1] "Range (event)"
#> 
#> $range
#> [1]   0.2605094 154.7000110
#> attr(,"label")
#> [1] "Range"
#> 
a_surv_time(df, .var = "AVAL", is_event = "is_event")
#> RowsVerticalSection (in_rows) object print method:
#> ----------------------------
#>       row_name formatted_cell indent_mod        row_label
#> 1       median           41.4          0           Median
#> 2    median_ci   (27.7, 54.7)          1           95% CI
#> 3    quantiles     15.4, 75.2          0  25% and 75%-ile
#> 4 range_censor   0.4 to 154.7          0 Range (censored)
#> 5  range_event   0.3 to 116.4          0    Range (event)
#> 6        range   0.3 to 154.7          0            Range
basic_table() %>%
  split_cols_by(var = "ARMCD") %>%
  add_colcounts() %>%
  surv_time(
    vars = "AVAL",
    var_labels = "Survival Time (Months)",
    is_event = "is_event",
    control = control_surv_time(conf_level = 0.9, conf_type = "log-log")
  ) %>%
  build_table(df = ADTTE_f)
#>                             ARM A          ARM B          ARM C    
#>                            (N=134)        (N=134)        (N=132)   
#> ———————————————————————————————————————————————————————————————————
#> Survival Time (Months)                                             
#>   Median                     41.4           27.5           11.1    
#>   90% CI                 (27.9, 54.1)   (18.8, 29.9)   (9.7, 15.1) 
#>   25% and 75%-ile         15.4, 75.2     9.5, 54.9      5.3, 25.2  
#>   Range (censored)       0.4 to 154.7   0.9 to 91.0    0.3 to 49.4 
#>   Range (event)          0.3 to 116.4   0.0 to 122.4   0.1 to 101.6