Summary table for binary outcome

This is a convenience wrapper function for creating a summary of common analyses for binary outcome variables. Options include reporting unstratified analysis, stratified analysis or both.

t_binary_outcome(
  rsp,
  col_by,
  strata_data = NULL,
  conf_level = 0.95,
  prop_ci_method = c("waldcc", "wald", "clopper-pearson", "wilson"),
  unstrat_analysis = c(diff_ci_method = "waldcc", diff_test = "chisq"),
  strat_analysis = c(diff_ci_method = "waldcc", diff_test = "cmh")
)

Arguments

rsp
(logical vector)
TRUE if subject is a responder FALSE otherwise

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

strata_data
(data.frame)
Used for stratification factors. If NULL, no stratified analysis is performed.

conf_level
confidence level of the interval.

prop_ci_method
one of ("waldcc", "wald", "clopper-pearson", "wilson").
Specifies the method used to construct the confidence interval for proportion of successful outcomes.

unstrat_analysis
named character vector specifying sections to include under the heading "Unstratified Analysis". If NULL, the table section is excluded. The named slot options are:

diff_ci_method The method used for the confidence interval for difference of proportions. Default method is "waldcc", Wald with continuity correction.

diff_test Test for a difference between two proportions. Default method is "chisq", Chi-squared test.

strat_analysis
named character vector specifying sections to include under the heading "Stratified Analysis". Only applicable if stratification factors from strata_data are provided. The named slot options are:

diff_ci_method The method used for the confidence interval for difference of proportions. Default method is "waldcc", Wald with continuity correction.

diff_test Test for a difference between two proportions. Default method is "cmh", Cochran-Mantel-Haenszel test.

Value

an rtable object

Details

Table sections are matched by name, so to remove a section specify a subset of the named slots in strat_analysis or strat_analysis.

The following options are supported for confidence interval for a proportion:

The wald interval follows the usual textbook definition for a single proportion confidence interval using the normal approximation.

The waldcc interval is similar to wald and in addition uses the continuity correction 1/2n.

The clopper-pearson interval calls binom.test. Also referred to as the 'exact' method.

The wilson interval calls prop.test with option correct=FALSE. Also referred to as Wilson score interval.

The following options are supported for difference of proportions confidence intervals:

Option "wald" calls prop.test with correct=FALSE

Option "waldcc" includes continuity correction for Wald interval and calls prop.test with correct=TRUE

Option "cmh" derives a CI for CMH-weighted difference of proportions and calls s_adj_proportion_diff

The following options are supported for testing difference of proportions:

Option chisq performs Chi-Squared test. Internally calls prop.test.

Option cmh performs stratified Cochran-Mantel-Haenszel test. Internally calls mantelhaen.test.

See also

Examples

library(random.cdisc.data)
library(dplyr)
ADRS <- radrs(cached = TRUE) %>%
  dplyr::filter(PARAMCD == "BESRSPI")

# unstratified analysis only with default options for unstrat_analysis
t_binary_outcome(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD
)

#>                                                              ARM A                 ARM B                 ARM C     
#>                                                             (N=134)               (N=134)               (N=132)    
#> -------------------------------------------------------------------------------------------------------------------
#> Responders                                                78 (58.21%)           55 (41.04%)           97 (73.48%)  
#> 95% CI for Response Rates (Wald, with correction)        (49.48, 66.93)        (32.34, 49.75)        (65.58, 81.39)
#> 
#> Unstratified Analysis
#>   Difference in Response Rates (%)                                                 -17.16                15.28     
#>     95% CI (Wald with correction)                                              (-29.7, -4.62)        (3.28, 27.27) 
#>   p-value (Chi-squared Test)                                                       0.005                 0.0086    

# unstratified + stratified analysis with default options for
# unstrat_analysis and strat_analysis
t_binary_outcome(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  strata_data = ADRS[, c("STRATA1", "STRATA2")]
)

#>                                                              ARM A                 ARM B                 ARM C     
#>                                                             (N=134)               (N=134)               (N=132)    
#> -------------------------------------------------------------------------------------------------------------------
#> Responders                                                78 (58.21%)           55 (41.04%)           97 (73.48%)  
#> 95% CI for Response Rates (Wald, with correction)        (49.48, 66.93)        (32.34, 49.75)        (65.58, 81.39)
#> 
#> Unstratified Analysis
#>   Difference in Response Rates (%)                                                 -17.16                15.28     
#>     95% CI (Wald with correction)                                              (-29.7, -4.62)        (3.28, 27.27) 
#>   p-value (Chi-squared Test)                                                       0.005                 0.0086    
#> 
#> Stratified Analysis*
#>   Difference in Response Rates (%)                                                 -17.16                15.28     
#>     95% CI (Wald with correction)                                              (-29.7, -4.62)        (3.28, 27.27) 
#>   p-value (CMH Test)                                                               0.004                 0.0119    
#> 
#> * Model stratified by STRATA1 and STRATA2

# stratified analysis only
# Show difference with CMH weights and p-value from CMH test
t_binary_outcome(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  strata_data = ADRS[, c("STRATA1", "STRATA2")],
  unstrat_analysis = NULL,
  strat_analysis = c("diff_ci_method" = "cmh", "diff_test" = "cmh")
 )

#>                                                              ARM A                 ARM B                 ARM C     
#>                                                             (N=134)               (N=134)               (N=132)    
#> -------------------------------------------------------------------------------------------------------------------
#> Responders                                                78 (58.21%)           55 (41.04%)           97 (73.48%)  
#> 95% CI for Response Rates (Wald, with correction)        (49.48, 66.93)        (32.34, 49.75)        (65.58, 81.39)
#> 
#> Stratified Analysis*
#>   Weighted Difference in Response Rates (%)                                        -17.51                14.85     
#>     95% CI (CMH, without correction)                                            (-29, -6.02)          (3.7, 26.01) 
#>   p-value (CMH Test)                                                               0.004                 0.0119    
#> 
#> * Model stratified by STRATA1 and STRATA2

# exclude comparison between arms
# specify CI method for proportions and change CI level
t_binary_outcome(
  ADRS$AVALC == "CR",
  col_by = ADRS$ARMCD,
  prop_ci_method = "clopper-pearson",
  unstrat_analysis = NULL,
  conf_level = 0.90
 )

#>                                                        ARM A                 ARM B                 ARM C     
#>                                                       (N=134)               (N=134)               (N=132)    
#> -------------------------------------------------------------------------------------------------------------
#> Responders                                          78 (58.21%)           55 (41.04%)           97 (73.48%)  
#> 90% CI for Response Rates (Clopper-Pearson)        (50.74, 65.4)         (33.89, 48.51)        (66.42, 79.73)