Single element for disposition table

Single element for disposition table

t_el_disposition(
  x = x,
  col_by,
  col_N = NULL,
  total = NULL,
  row.name = NULL,
  subset = NULL,
  show_n = FALSE,
  useNA = c("no", "ifany", "always"),
  drop_levels = NULL,
  denominator = "N"
)

Arguments

x
A factor or logical vector

col_by
logical factor or data.frame with reference and comparison group information, the first level indicates the reference group. See the functions factor_to_matrix_by and add_total_by.

col_N
a vecor with the reference population per level of col_by. This often by default get_N(col_by). Names of the list are ignored, they are added in the order they appear.

total
string of column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

row.name
Only applicable when x is a logical vector. A string to label the row name. Default is "TRUE".

subset
A logical vector with the same length as x that defines the subset. Applies to x and subset.

show_n
Logic value to determine whether the "n" row is displayed. Default is FALSE. Only when x is of type factor

useNA
choose whether missing data (NAs) should be displayed as a level.

drop_levels
boolean whether to drop zero count levels

denominator
either n, N or omit. n and N are for calculating the level associated percentage. With option N, the reference population from col_N is used as the denominator. With option n, the number of non-missing records from x is used as the denominator. If omit is chosen the percentage is omitted.

Examples

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)

ADSL0 <- ADSL %>%
  mutate(
    COMPSTUD = sample(c('Y','N'),
                      size=nrow(ADSL),
                      replace = TRUE) %>% as.factor,
    STUDONS = sample(c('Alive: On Treatment', 'Alive: In Follow-up', NA),
                     size=nrow(ADSL),
                     replace = TRUE) %>% as.factor,
    STDDRS = sample(c('Death', 'Lost To Follow-Up',
                      'Protocol Violation', 'Withdrawal By Subject',
                      'Other'),
                    size=nrow(ADSL),
                    replace = TRUE) %>% as.factor,
    GOTTRT = ifelse(!is.na(ACTARMCD), 'Y', 'N') %>% as.factor,
    DISTRTFL = sample(c('Y','N'),
                      size=nrow(ADSL),
                      replace = TRUE) %>% as.factor,
    TRTDRS = sample(c('ADVERSE EVENT', 'PROGRESSIVE DISEASE',
                      'PHYSICIAN DECISION', 'LACK OF EFFICACY',
                      'OTHER'),
                    size=nrow(ADSL),
                    replace = TRUE) %>% as.factor,
    STUDONS = case_when(
      COMPSTUD == 'N' ~ STUDONS
    ),
    STDDRS = case_when(
      COMPSTUD == 'N' & is.na(STUDONS) ~ STDDRS
    ) %>% as.factor,
    DISSTDFL = case_when(
      !is.na(STDDRS) ~ "Y"
    ),
    DISTRTFL = case_when(
      GOTTRT == 'Y' ~ DISTRTFL
    ),
    TRTDRS = case_when(
      DISTRTFL == 'Y' ~ TRTDRS
    ),
    DRSCAT = case_when(
      TRTDRS %in% c('ADVERSE EVENT', 'PHYSICIAN DECISION') ~ "Safety",
      !is.na(TRTDRS) ~ "Other"
    )
  ) %>% var_relabel(COMPSTUD = "Complete Study",
                    STUDONS = "On-study Status",
                    DISSTDFL = "Discontinued Study",
                    STDDRS = "Reason for Study \r\nDiscontinuation",
                    GOTTRT = "Received Treatment",
                    DISTRTFL = "Discontinued Treatment",
                    TRTDRS = "Reason for Treatment \r\nDiscontinuation"
  )

library(purrr)

#> 
#> Attaching package: ‘purrr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     is_null

t_el_disposition_indent <- function(..., indent = 0) indent(t_el_disposition(...), indent)
dsp <- partial(
  t_el_disposition_indent,
  col_by = ADSL0$ARM,
  total = "All Patients",
  denominator = "N"
)
# Only using rtables
# old way of doing things with indent variable
rbind(
  dsp(ADSL0$COMPSTUD == "Y", row.name = "Completed study"),
  dsp(ADSL0$STUDONS == "Alive: In Follow-up", row.name = "Alive: In follow-up"),
  dsp(ADSL0$DISSTDFL == "Y",  row.name = "Discontinued study"),
  dsp(ADSL0$STDDRS, indent = 1),
  rrow(),
  rrow("Show example of using n as denominator"),
  indent(t_el_disposition(
    ADSL0$STDDRS,
    col_by = ADSL0$ARM,
    total = "All Patients",
    denominator = "n"
  ), 1),
  rrow(),
  dsp(ADSL0$GOTTRT == "Y", row.name = "Received treatment"),
  dsp(ADSL0$DISTRTFL == "Y", row.name = "Discontinued treatment"),
  dsp(ADSL0$DRSCAT == "Safety", row.name = "Safety", indent = 1),
  dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT %in% "Safety", indent = 2),
  dsp(ADSL0$DRSCAT == "Other", row.name = "Other", indent = 1),
  dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT %in% "Other", indent = 2)
)

#>                                                 A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                                  (N=134)               (N=134)               (N=132)               (N=400)    
#> ------------------------------------------------------------------------------------------------------------------------------
#> Completed study                                73 (54.48%)           77 (57.46%)           75 (56.82%)           225 (56.25%) 
#> Alive: In follow-up                            23 (17.16%)           21 (15.67%)            12 (9.09%)             56 (14%)   
#> Discontinued study                             20 (14.93%)           18 (13.43%)            26 (19.7%)             64 (16%)   
#>   Death                                         8 (5.97%)             4 (2.99%)             5 (3.79%)             17 (4.25%)  
#>   Lost To Follow-Up                             1 (0.75%)                 0                 3 (2.27%)               4 (1%)    
#>   Other                                         5 (3.73%)             7 (5.22%)             6 (4.55%)             18 (4.5%)   
#>   Protocol Violation                            3 (2.24%)             2 (1.49%)             8 (6.06%)             13 (3.25%)  
#>   Withdrawal By Subject                         3 (2.24%)             5 (3.73%)             4 (3.03%)              12 (3%)    
#> 
#> Show example of using n as denominator
#>   Death                                          8 (40%)              4 (22.22%)            5 (19.23%)           17 (26.56%)  
#>   Lost To Follow-Up                               1 (5%)                  0                 3 (11.54%)            4 (6.25%)   
#>   Other                                          5 (25%)              7 (38.89%)            6 (23.08%)           18 (28.12%)  
#>   Protocol Violation                             3 (15%)              2 (11.11%)            8 (30.77%)           13 (20.31%)  
#>   Withdrawal By Subject                          3 (15%)              5 (27.78%)            4 (15.38%)           12 (18.75%)  
#> 
#> Received treatment                              134 (100%)            134 (100%)            132 (100%)            400 (100%)  
#> Discontinued treatment                         68 (50.75%)           61 (45.52%)           65 (49.24%)           194 (48.5%)  
#>   Safety                                       35 (26.12%)           21 (15.67%)           17 (12.88%)           73 (18.25%)  
#>     ADVERSE EVENT                              21 (15.67%)            10 (7.46%)            11 (8.33%)            42 (10.5%)  
#>     PHYSICIAN DECISION                         14 (10.45%)            11 (8.21%)            6 (4.55%)             31 (7.75%)  
#>   Other                                        33 (24.63%)           40 (29.85%)           48 (36.36%)           121 (30.25%) 
#>     LACK OF EFFICACY                            13 (9.7%)             10 (7.46%)           14 (10.61%)            37 (9.25%)  
#>     OTHER                                       12 (8.96%)           20 (14.93%)           21 (15.91%)           53 (13.25%)  
#>     PROGRESSIVE DISEASE                         8 (5.97%)             10 (7.46%)            13 (9.85%)            31 (7.75%)  

# new way:
treatment_section_rtable <- rbind(
  dsp(ADSL0$GOTTRT == "Y", row.name = "Received treatment"),
  dsp(ADSL0$DISTRTFL == "Y", row.name = "Discontinued treatment"),
  indent(
    rbind(
      dsp(ADSL0$DRSCAT == "Safety", row.name = "Safety"),
      indent(dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT %in% "Safety"), 1),
      dsp(ADSL0$DRSCAT == "Other", row.name = "Other"),
      indent(dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT %in% "Other"), 1)
    ),
    1
  )
)
treatment_section_rtable

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                   (N=134)               (N=134)               (N=132)               (N=400)    
#> ---------------------------------------------------------------------------------------------------------------
#> Received treatment               134 (100%)            134 (100%)            132 (100%)            400 (100%)  
#> Discontinued treatment          68 (50.75%)           61 (45.52%)           65 (49.24%)           194 (48.5%)  
#>   Safety                        35 (26.12%)           21 (15.67%)           17 (12.88%)           73 (18.25%)  
#>     ADVERSE EVENT               21 (15.67%)            10 (7.46%)            11 (8.33%)            42 (10.5%)  
#>     PHYSICIAN DECISION          14 (10.45%)            11 (8.21%)            6 (4.55%)             31 (7.75%)  
#>   Other                         33 (24.63%)           40 (29.85%)           48 (36.36%)           121 (30.25%) 
#>     LACK OF EFFICACY             13 (9.7%)             10 (7.46%)           14 (10.61%)            37 (9.25%)  
#>     OTHER                        12 (8.96%)           20 (14.93%)           21 (15.91%)           53 (13.25%)  
#>     PROGRESSIVE DISEASE          8 (5.97%)             10 (7.46%)            13 (9.85%)            31 (7.75%)  

# easier to manipulate for their purposes by other users
treatment_section_tree <- node(
  invisible_node_name("treatment"),
  content = NULL,
  format_data = list(children_gap = 0),
  children = list(
    node(
      invisible_node_name("received_treatment"),
      content = dsp(ADSL0$GOTTRT == "Y", row.name = "Received treatment")
    ),
    node(
      invisible_node_name("discontinuedtreatment"),
      dsp(ADSL0$DISTRTFL == "Y", row.name = "Discontinued treatment"),
      format_data = list(gap_to_children = 0, children_gap = 0, children_indent = 1),
      children = list(
        node(
          invisible_node_name("safety"),
          rbind(
            dsp(ADSL0$DRSCAT == "Safety", row.name = "Safety"),
            indent(dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT %in% "Safety"), 1)
          )
        ),
        node(
          invisible_node_name("other"),
          rbind(
            dsp(ADSL0$DRSCAT == "Other", row.name = "Other"),
            indent(dsp(ADSL0$TRTDRS, subset = ADSL0$DRSCAT %in% "Other"), 1)
          )
        )
      )
    )
  )
)
to_rtable(treatment_section_tree)

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                   (N=134)               (N=134)               (N=132)               (N=400)    
#> ---------------------------------------------------------------------------------------------------------------
#> Received treatment               134 (100%)            134 (100%)            132 (100%)            400 (100%)  
#> Discontinued treatment          68 (50.75%)           61 (45.52%)           65 (49.24%)           194 (48.5%)  
#>   Safety                        35 (26.12%)           21 (15.67%)           17 (12.88%)           73 (18.25%)  
#>     ADVERSE EVENT               21 (15.67%)            10 (7.46%)            11 (8.33%)            42 (10.5%)  
#>     PHYSICIAN DECISION          14 (10.45%)            11 (8.21%)            6 (4.55%)             31 (7.75%)  
#>   Other                         33 (24.63%)           40 (29.85%)           48 (36.36%)           121 (30.25%) 
#>     LACK OF EFFICACY             13 (9.7%)             10 (7.46%)           14 (10.61%)            37 (9.25%)  
#>     OTHER                        12 (8.96%)           20 (14.93%)           21 (15.91%)           53 (13.25%)  
#>     PROGRESSIVE DISEASE          8 (5.97%)             10 (7.46%)            13 (9.85%)            31 (7.75%)  

to_rtable(invisible_node(
  name = "t_el_disp",
  children = list(
    node(
      name = invisible_node_name("Study status"),
      content = rbind(
        dsp(ADSL0$COMPSTUD == "Y", row.name = "Completed study"),
        dsp(ADSL0$STUDONS == "Alive: In Follow-up", row.name = "Alive: In follow-up")
      ),
      children = list(
        node(
          name = invisible_node_name("Reasons for study discontinuation (using capital N)"),
          rbind(
            dsp(ADSL0$DISSTDFL == "Y",  row.name = "Discontinued study"),
            indent(dsp(ADSL0$STDDRS), 1)
          )
        ),
        node(
          name = "Reasons for study discontinuation (using small n)",
          content = t_el_disposition(
            ADSL0$STDDRS,
            col_by = ADSL0$ARM,
            total = "All Patients",
            denominator = "n"
          )
        )
      )
    ),
    treatment_section_tree
  )
))

#>                                                            A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                                             (N=134)               (N=134)               (N=132)               (N=400)    
#> -----------------------------------------------------------------------------------------------------------------------------------------
#> Completed study                                           73 (54.48%)           77 (57.46%)           75 (56.82%)           225 (56.25%) 
#> Alive: In follow-up                                       23 (17.16%)           21 (15.67%)            12 (9.09%)             56 (14%)   
#> 
#> Discontinued study                                        20 (14.93%)           18 (13.43%)            26 (19.7%)             64 (16%)   
#>   Death                                                    8 (5.97%)             4 (2.99%)             5 (3.79%)             17 (4.25%)  
#>   Lost To Follow-Up                                        1 (0.75%)                 0                 3 (2.27%)               4 (1%)    
#>   Other                                                    5 (3.73%)             7 (5.22%)             6 (4.55%)             18 (4.5%)   
#>   Protocol Violation                                       3 (2.24%)             2 (1.49%)             8 (6.06%)             13 (3.25%)  
#>   Withdrawal By Subject                                    3 (2.24%)             5 (3.73%)             4 (3.03%)              12 (3%)    
#> 
#> Reasons for study discontinuation (using small n)
#>   Death                                                     8 (40%)              4 (22.22%)            5 (19.23%)           17 (26.56%)  
#>   Lost To Follow-Up                                          1 (5%)                  0                 3 (11.54%)            4 (6.25%)   
#>   Other                                                     5 (25%)              7 (38.89%)            6 (23.08%)           18 (28.12%)  
#>   Protocol Violation                                        3 (15%)              2 (11.11%)            8 (30.77%)           13 (20.31%)  
#>   Withdrawal By Subject                                     3 (15%)              5 (27.78%)            4 (15.38%)           12 (18.75%)  
#> 
#> Received treatment                                         134 (100%)            134 (100%)            132 (100%)            400 (100%)  
#> Discontinued treatment                                    68 (50.75%)           61 (45.52%)           65 (49.24%)           194 (48.5%)  
#>   Safety                                                  35 (26.12%)           21 (15.67%)           17 (12.88%)           73 (18.25%)  
#>     ADVERSE EVENT                                         21 (15.67%)            10 (7.46%)            11 (8.33%)            42 (10.5%)  
#>     PHYSICIAN DECISION                                    14 (10.45%)            11 (8.21%)            6 (4.55%)             31 (7.75%)  
#>   Other                                                   33 (24.63%)           40 (29.85%)           48 (36.36%)           121 (30.25%) 
#>     LACK OF EFFICACY                                       13 (9.7%)             10 (7.46%)           14 (10.61%)            37 (9.25%)  
#>     OTHER                                                  12 (8.96%)           20 (14.93%)           21 (15.91%)           53 (13.25%)  
#>     PROGRESSIVE DISEASE                                    8 (5.97%)             10 (7.46%)            13 (9.85%)            31 (7.75%)