Events by Highest Grade Table

This function summarizes number of unique subjects by highest grade and events term(s). Events terms can be one level term or two level terms (one higher level and one lower level). An implementation example is to apply t_events_per_term_grade_id on Adverse Event Data to create Adverse Events by Highest NCI CTCAE grade table (AET04, STREAM2.x, STREAM1.17).

t_events_per_term_grade_id(terms, id, grade, col_by, col_N,
  total = "All Patients", grade_levels = 1:5)

Arguments

terms
character or factor vector, or dataframe to represent events information; Currently terms can only be a vector or dataframe with 1 or 2 columns. For terms with 2 columns, 1st column should represent higher level term and 2nd column should be lower level term. var_relabel is used as the character string used as a label in the column header for each term.

id
vector of subject identifier. Length of id must be the same as the length or number of rows of terms.

grade
grade of adverse event as numeric. var_relabel is used as the character string used as a label in the column header for each grade.

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing. See 'Examples'.

col_N
numeric vector with information of the number of patients in the levels of col_by. This is useful if there are patients that have no adverse events can be accounted for with this argument.

total
character string that will be used as a label for a column with pooled total population, default is "All Patients". If the levels of col_by are the only columns of interest then total should be NULL

grade_levels
numeric, ordered values of possible of grades in a form of x:y, default is 1:5.

Value

an rtable object.

Details

t_events_per_term_grade_id includes percentages based on the total number of subjects in the column heading (i.e. "N=nnn"). col_N can be explicitly specified to get N for percentage calculation from either events dataset or additional dataset like subject level dataset. See the example.

Multiple events within a subject of the same term (if terms is one level) or lower level term (if terms is two levels) are counted once using the greatest intensity reported.

t_events_per_term_grade_id doesn't deal with data with any non-complete records (has NA's), e.g. if any terms are missing. Impute missing values before using t_events_per_term_grade_id.

t_events_per_term_grade_id orders data by "All Patients" column from the most commonly reported higher level term to the least frequent one. Within each group of higher level term, it sorts by decreasing frequency of lower level term. It brakes ties using terms names in alphabetical order.

t_events_per_term_grade_id fills in col_by and grade with 0 value in case there was no events reported for particular col_by and/or grade category. Use grade_levels to modify the range of existing grades. If data does not have any records with grade 5 and the intent is to show only grades 1-4 rows then use grade_levels = 1:4.

See also

Examples

library(dplyr)
library(random.cdisc.data)

ADSL <- radsl(10, seed = 1)
ADAE <- radae(ADSL, 4, seed = 2)

t_events_per_term_grade_id(
  terms = ADAE$AEDECOD,
  id = ADAE$USUBJID,
  grade = ADAE$AETOXGR,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients",
  grade_levels = 1:5
)

#>                                                                            A: Drug X                         B: Placebo                       C: Combination                      All Patients        
#> Dictionary-Derived Term        (N=Standard Toxicity Grade)                   (N=4)                              (N=5)                              (N=1)                             (N=10)           
#> ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> - Overall -                           - Any Grade -                         3 (75%)                            3 (60%)                           1 (100%)                            7 (70%)          
#>                                             1                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             2                               1 (25%)                            0 (0%)                             0 (0%)                             1 (10%)          
#>                                             3                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             4                               1 (25%)                            0 (0%)                            1 (100%)                            2 (20%)          
#>                                             5                               1 (25%)                            3 (60%)                            0 (0%)                             4 (40%)          
#> 
#> trm B_1/3                             - Any Grade -                         1 (25%)                            3 (60%)                            0 (0%)                             4 (40%)          
#>                                             1                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             2                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             3                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             4                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             5                               1 (25%)                            3 (60%)                            0 (0%)                             4 (40%)          
#> 
#> trm C_1/1                             - Any Grade -                         1 (25%)                            2 (40%)                           1 (100%)                            4 (40%)          
#>                                             1                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             2                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             3                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             4                               1 (25%)                            2 (40%)                           1 (100%)                            4 (40%)          
#>                                             5                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#> 
#> trm A_1/2                             - Any Grade -                         0 (0%)                             2 (40%)                            0 (0%)                             2 (20%)          
#>                                             1                               0 (0%)                             2 (40%)                            0 (0%)                             2 (20%)          
#>                                             2                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             3                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             4                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             5                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#> 
#> trm A_2/2                             - Any Grade -                         1 (25%)                            0 (0%)                            1 (100%)                            2 (20%)          
#>                                             1                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             2                               1 (25%)                            0 (0%)                            1 (100%)                            2 (20%)          
#>                                             3                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             4                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             5                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#> 
#> trm B_2/3                             - Any Grade -                         0 (0%)                             1 (20%)                           1 (100%)                            2 (20%)          
#>                                             1                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             2                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             3                               0 (0%)                             1 (20%)                           1 (100%)                            2 (20%)          
#>                                             4                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             5                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#> 
#> trm B_3/3                             - Any Grade -                         1 (25%)                            0 (0%)                             0 (0%)                             1 (10%)          
#>                                             1                               1 (25%)                            0 (0%)                             0 (0%)                             1 (10%)          
#>                                             2                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             3                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             4                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           
#>                                             5                               0 (0%)                             0 (0%)                             0 (0%)                             0 (0%)           

t_events_per_term_grade_id(
  terms = ADAE %>% select(AEBODSYS, AEDECOD),
  id = ADAE$USUBJID,
  grade = ADAE$AETOXGR,
  col_by = ADAE$ARM,
  col_N = table(ADSL$ARM),
  total = "All Patients",
  grade_levels = 1:5
)

#> Body System or Organ Class                                              A: Drug X                     B: Placebo                   C: Combination                  All Patients      
#>   Dictionary-Derived Term         Standard Toxicity Grade                 (N=4)                          (N=5)                          (N=1)                         (N=10)         
#> -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> - Any adverse events -
#>   - Overall -                          - Any Grade -                     3 (75%)                        3 (60%)                       1 (100%)                        7 (70%)        
#>                                              1                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              2                           1 (25%)                        0 (0%)                         0 (0%)                         1 (10%)        
#>                                              3                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              4                           1 (25%)                        0 (0%)                        1 (100%)                        2 (20%)        
#>                                              5                           1 (25%)                        3 (60%)                        0 (0%)                         4 (40%)        
#> 
#> cl B
#>   - Overall -                          - Any Grade -                     2 (50%)                        3 (60%)                       1 (100%)                        6 (60%)        
#>                                              1                           1 (25%)                        0 (0%)                         0 (0%)                         1 (10%)        
#>                                              2                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              3                           0 (0%)                         0 (0%)                        1 (100%)                        1 (10%)        
#>                                              4                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              5                           1 (25%)                        3 (60%)                        0 (0%)                         4 (40%)        
#> 
#>   trm B_1/3                            - Any Grade -                     1 (25%)                        3 (60%)                        0 (0%)                         4 (40%)        
#>                                              1                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              2                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              3                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              4                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              5                           1 (25%)                        3 (60%)                        0 (0%)                         4 (40%)        
#> 
#>   trm B_2/3                            - Any Grade -                     0 (0%)                         1 (20%)                       1 (100%)                        2 (20%)        
#>                                              1                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              2                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              3                           0 (0%)                         1 (20%)                       1 (100%)                        2 (20%)        
#>                                              4                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              5                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#> 
#>   trm B_3/3                            - Any Grade -                     1 (25%)                        0 (0%)                         0 (0%)                         1 (10%)        
#>                                              1                           1 (25%)                        0 (0%)                         0 (0%)                         1 (10%)        
#>                                              2                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              3                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              4                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              5                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#> 
#> cl A
#>   - Overall -                          - Any Grade -                     1 (25%)                        2 (40%)                       1 (100%)                        4 (40%)        
#>                                              1                           0 (0%)                         2 (40%)                        0 (0%)                         2 (20%)        
#>                                              2                           1 (25%)                        0 (0%)                        1 (100%)                        2 (20%)        
#>                                              3                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              4                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              5                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#> 
#>   trm A_1/2                            - Any Grade -                     0 (0%)                         2 (40%)                        0 (0%)                         2 (20%)        
#>                                              1                           0 (0%)                         2 (40%)                        0 (0%)                         2 (20%)        
#>                                              2                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              3                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              4                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              5                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#> 
#>   trm A_2/2                            - Any Grade -                     1 (25%)                        0 (0%)                        1 (100%)                        2 (20%)        
#>                                              1                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              2                           1 (25%)                        0 (0%)                        1 (100%)                        2 (20%)        
#>                                              3                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              4                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              5                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#> 
#> cl C
#>   - Overall -                          - Any Grade -                     1 (25%)                        2 (40%)                       1 (100%)                        4 (40%)        
#>                                              1                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              2                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              3                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              4                           1 (25%)                        2 (40%)                       1 (100%)                        4 (40%)        
#>                                              5                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#> 
#>   trm C_1/1                            - Any Grade -                     1 (25%)                        2 (40%)                       1 (100%)                        4 (40%)        
#>                                              1                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              2                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              3                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)         
#>                                              4                           1 (25%)                        2 (40%)                       1 (100%)                        4 (40%)        
#>                                              5                           0 (0%)                         0 (0%)                         0 (0%)                         0 (0%)