Adverse Events Table by Highest NCI CTCAE Grade

t_events_per_term_grade_id returns adverse events sorted by highest NCI (National Cancer Institute) CTCAE (common terminology criteria for adverse avents) grade.

t_events_per_term_grade_id(terms, id, grade, col_by, col_N,
  total = "All Patients", grade_levels = 1:5)

Arguments

terms
term information as character or factor vector or dataframe, however factor levels are not repeated by class, only terms with count > 1 are listed per class. Currently terms can only be a vector or dataframe with 1/2 column. var_relabel is used as the character string used as a label in the column header for each term.

id
unique subject identifier. If a particular subject has no adverse event then that information needs to be added to the col_N argument.

grade
grade of adverse event as numeric. var_relabel is used as the character string used as a label in the column header for each grade.

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing. See 'Examples'.

col_N
numeric vector with information of the number of patients in the levels of col_by. This is useful if there are patients that have no adverse events can be accounted for with this argument.

total
character string that will be used as a label for a column with pooled total population, default is "All Patients". If the levels of col_by are the only columns of interest then total should be NULL

grade_levels
numeric, ordered values of possible of grades in a form of x:y, default is 1:5.

Value

rtable object

Details

t_events_per_term_grade_id counts patients according to adverse events (AEs) of greatest intensity for system organ class (SOC) and overall rows and includes percentages based on the total number of patients in the column heading (i.e. "N=nnn"). If the intention is to use patients number from subject level dataset as N for percentage calculation then adeverse events dataset should be left joined to subject level dataset and the col_by variable should be dropped from adverse events dataset, see the example. Otherwise, N will be derived using adverse events dataset. At the preferred term (PT) level, multiple events within a patient of the same PT are counted once using the greatest intensity reported.

t_events_per_term_grade_id removes any non-complete records, e.g. if class or term are missing. If the intent is to preserve such records, then impute missing values before using t_events_per_term_grade_id.

t_events_per_term_grade_id orders data by "All Patients" column from the most commonly reported SOC to the least frequent one. Within SOC, it sorts by decreasing frequency of PT. It brakes ties using SOC/PT names in alphabetical order.

t_events_per_term_grade_id fills in col_by and grade with 0 value in case there was no AEs reported for particular col_by and/or grade category. Use grade_levels to modify the range of existing grades. If data does not have any records with grade 5 and the intent is to show only grades 1-4 rows then use grade_levels = 1:4.

This is an equivalent of the STREAM output %stream_t_summary(templates = aet04) (http://bioportal.roche.com/stream_doc/2_05/um/report_outputs_aet04.html)

This is an equivalent of the STREAM 1.17 output %stream_t_events_bygrade(templates = aet04) (<https://rochewiki.roche.com/confluence/pages/viewpage.action?pageId=294027501>)

Examples

library(dplyr)
library(random.cdisc.data)

ASL <- rasl(10, seed = 1)
AAE <- raae(ASL, 4, seed = 2)

ANL <- left_join(AAE, ASL %>% select(USUBJID, STUDYID, ARM), by = c("USUBJID", "STUDYID")) %>%
  var_relabel(
    AEBODSYS = 'MedDRA System Organ Class',
    AEDECOD = 'MedDRA Preferred Term',
    AETOXGR = 'GRADE'
)

t_events_per_term_grade_id(
  terms = ANL$AEDECOD,
  id = ANL$USUBJID,
  grade = ANL$AETOXGR,
  col_by = ANL$ARM,
  col_N = table(ASL$ARM),
  total = "All Patients",
  grade_levels = 1:5
)

#>                                                      A: Drug X             B: Placebo          C: Combination         All Patients 
#> MedDRA Preferred Term                                  (N=1)                 (N=4)                 (N=5)                 (N=10)    
#> -----------------------------------------------------------------------------------------------------------------------------------
#> - Overall -                  - Any Grade -            1 (100%)              2 (50%)               4 (80%)               7 (70%)    
#>                                    1                     0                     0                     0                     0       
#>                                    2                     0                  1 (25%)                  0                  1 (10%)    
#>                                    3                     0                     0                     0                     0       
#>                                    4                     0                  1 (25%)               1 (20%)               2 (20%)    
#>                                    5                  1 (100%)                 0                  3 (60%)               4 (40%)    
#> 
#> trm B_1/3                    - Any Grade -            1 (100%)                 0                  3 (60%)               4 (40%)    
#>                                    1                     0                     0                     0                     0       
#>                                    2                     0                     0                     0                     0       
#>                                    3                     0                     0                     0                     0       
#>                                    4                     0                     0                     0                     0       
#>                                    5                  1 (100%)                 0                  3 (60%)               4 (40%)    
#> 
#> trm C_1/1                    - Any Grade -               0                  1 (25%)               3 (60%)               4 (40%)    
#>                                    1                     0                     0                     0                     0       
#>                                    2                     0                     0                     0                     0       
#>                                    3                     0                     0                     0                     0       
#>                                    4                     0                  1 (25%)               3 (60%)               4 (40%)    
#>                                    5                     0                     0                     0                     0       
#> 
#> trm A_1/2                    - Any Grade -               0                     0                  2 (40%)               2 (20%)    
#>                                    1                     0                     0                  2 (40%)               2 (20%)    
#>                                    2                     0                     0                     0                     0       
#>                                    3                     0                     0                     0                     0       
#>                                    4                     0                     0                     0                     0       
#>                                    5                     0                     0                     0                     0       
#> 
#> trm A_2/2                    - Any Grade -               0                  2 (50%)                  0                  2 (20%)    
#>                                    1                     0                     0                     0                     0       
#>                                    2                     0                  2 (50%)                  0                  2 (20%)    
#>                                    3                     0                     0                     0                     0       
#>                                    4                     0                     0                     0                     0       
#>                                    5                     0                     0                     0                     0       
#> 
#> trm B_2/3                    - Any Grade -               0                  1 (25%)               1 (20%)               2 (20%)    
#>                                    1                     0                     0                     0                     0       
#>                                    2                     0                     0                     0                     0       
#>                                    3                     0                  1 (25%)               1 (20%)               2 (20%)    
#>                                    4                     0                     0                     0                     0       
#>                                    5                     0                     0                     0                     0       
#> 
#> trm B_3/3                    - Any Grade -               0                  1 (25%)                  0                  1 (10%)    
#>                                    1                     0                  1 (25%)                  0                  1 (10%)    
#>                                    2                     0                     0                     0                     0       
#>                                    3                     0                     0                     0                     0       
#>                                    4                     0                     0                     0                     0       
#>                                    5                     0                     0                     0                     0       

t_events_per_term_grade_id(
  terms = ANL %>% select(AEBODSYS, AEDECOD),
  id = ANL$USUBJID,
  grade = ANL$AETOXGR,
  col_by = ANL$ARM,
  col_N = table(ASL$ARM),
  total = "All Patients",
  grade_levels = 1:5
)

#> MedDRA System Organ Class                                A: Drug X             B: Placebo          C: Combination         All Patients 
#>   MedDRA Preferred Term              GRADE                 (N=1)                 (N=4)                 (N=5)                 (N=10)    
#> ---------------------------------------------------------------------------------------------------------------------------------------
#> - Any adverse events -
#>   - Overall -                    - Any Grade -            1 (100%)              2 (50%)               4 (80%)               7 (70%)    
#>                                        1                     0                     0                     0                     0       
#>                                        2                     0                  1 (25%)                  0                  1 (10%)    
#>                                        3                     0                     0                     0                     0       
#>                                        4                     0                  1 (25%)               1 (20%)               2 (20%)    
#>                                        5                  1 (100%)                 0                  3 (60%)               4 (40%)    
#> 
#> cl B
#>   - Overall -                    - Any Grade -            1 (100%)              2 (50%)               3 (60%)               6 (60%)    
#>                                        1                     0                  1 (25%)                  0                  1 (10%)    
#>                                        2                     0                     0                     0                     0       
#>                                        3                     0                  1 (25%)                  0                  1 (10%)    
#>                                        4                     0                     0                     0                     0       
#>                                        5                  1 (100%)                 0                  3 (60%)               4 (40%)    
#> 
#>   trm B_1/3                      - Any Grade -            1 (100%)                 0                  3 (60%)               4 (40%)    
#>                                        1                     0                     0                     0                     0       
#>                                        2                     0                     0                     0                     0       
#>                                        3                     0                     0                     0                     0       
#>                                        4                     0                     0                     0                     0       
#>                                        5                  1 (100%)                 0                  3 (60%)               4 (40%)    
#> 
#>   trm B_2/3                      - Any Grade -               0                  1 (25%)               1 (20%)               2 (20%)    
#>                                        1                     0                     0                     0                     0       
#>                                        2                     0                     0                     0                     0       
#>                                        3                     0                  1 (25%)               1 (20%)               2 (20%)    
#>                                        4                     0                     0                     0                     0       
#>                                        5                     0                     0                     0                     0       
#> 
#>   trm B_3/3                      - Any Grade -               0                  1 (25%)                  0                  1 (10%)    
#>                                        1                     0                  1 (25%)                  0                  1 (10%)    
#>                                        2                     0                     0                     0                     0       
#>                                        3                     0                     0                     0                     0       
#>                                        4                     0                     0                     0                     0       
#>                                        5                     0                     0                     0                     0       
#> 
#> cl A
#>   - Overall -                    - Any Grade -               0                  2 (50%)               2 (40%)               4 (40%)    
#>                                        1                     0                     0                  2 (40%)               2 (20%)    
#>                                        2                     0                  2 (50%)                  0                  2 (20%)    
#>                                        3                     0                     0                     0                     0       
#>                                        4                     0                     0                     0                     0       
#>                                        5                     0                     0                     0                     0       
#> 
#>   trm A_1/2                      - Any Grade -               0                     0                  2 (40%)               2 (20%)    
#>                                        1                     0                     0                  2 (40%)               2 (20%)    
#>                                        2                     0                     0                     0                     0       
#>                                        3                     0                     0                     0                     0       
#>                                        4                     0                     0                     0                     0       
#>                                        5                     0                     0                     0                     0       
#> 
#>   trm A_2/2                      - Any Grade -               0                  2 (50%)                  0                  2 (20%)    
#>                                        1                     0                     0                     0                     0       
#>                                        2                     0                  2 (50%)                  0                  2 (20%)    
#>                                        3                     0                     0                     0                     0       
#>                                        4                     0                     0                     0                     0       
#>                                        5                     0                     0                     0                     0       
#> 
#> cl C
#>   - Overall -                    - Any Grade -               0                  1 (25%)               3 (60%)               4 (40%)    
#>                                        1                     0                     0                     0                     0       
#>                                        2                     0                     0                     0                     0       
#>                                        3                     0                     0                     0                     0       
#>                                        4                     0                  1 (25%)               3 (60%)               4 (40%)    
#>                                        5                     0                     0                     0                     0       
#> 
#>   trm C_1/1                      - Any Grade -               0                  1 (25%)               3 (60%)               4 (40%)    
#>                                        1                     0                     0                     0                     0       
#>                                        2                     0                     0                     0                     0       
#>                                        3                     0                     0                     0                     0       
#>                                        4                     0                  1 (25%)               3 (60%)               4 (40%)    
#>                                        5                     0                     0                     0                     0