Adverse Events by System Organ Class and/or Preferred Term

t_events_per_term_id returns adverse events summary table that corresponds to STREAM template AET02

t_events_per_term_id(terms, id, col_by, col_N, total = "All Patients")

Arguments

terms
term information as character or factor vector or dataframe, however factor levels are not repeated by class, only terms with count > 1 are listed per class. Currently terms can only be a vector or dataframe with 1/2 column. var_relabel is used as the character string used as a label in the column header for each term.

id
unique subject identifier. If a particular subject has no adverse event then that information needs to be added to the col_N argument.

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing. See 'Examples'.

col_N
numeric vector with information of the number of patients in the levels of col_by. This is useful if there are patients that have no adverse events can be accounted for with this argument.

total
character string that will be used as a label for a column with pooled total population, default is "All Patients". If the levels of col_by are the only columns of interest then total should be NULL

Value

rtable object

Details

This is an equivalent of the STREAM output %stream_t_summary(templates = aet02) (http://bioportal.roche.com/stream_doc/2_05/um/report_outputs_aet02.html)

This is an equivalent of the STREAM 1.17 output %stream_t_events_basic(templates = aet02) (https://rochewiki.roche.com/confluence/pages/viewpage.action?pageId=294027342)

Examples

library(dplyr)
library(random.cdisc.data)

ASL <- rasl(10, seed = 1)
AAE <- raae(ASL, 4, seed = 2)

ANL <- left_join(AAE, ASL %>% select(USUBJID, STUDYID, ARM), by = c("USUBJID", "STUDYID")) %>%
  var_relabel(
    AEBODSYS = 'MedDRA System Organ Class',
    AEDECOD = 'MedDRA Preferred Term',
    AETOXGR = 'GRADE'
)

t_events_per_term_id(
  terms = ANL$AEDECOD,
  id = ANL$USUBJID,
  col_by = ANL$ARM,
  col_N = tapply(ASL$ARM, ASL$ARM, length),
  total = NULL
)

#> Error in (function (header, ..., format = NULL) {    is_rcell_format(format, stop_otherwise = TRUE)    if (!is(header, "rheader"))         header <- rheader(header)    body <- list(...)    if (!are(body, "rrow"))         stop("not all arguments in ... are of class rrow")    ncol_header <- vapply(header, ncell, numeric(1))    ncol_body <- vapply(body, ncell, numeric(1))    nrow <- length(body)    ncol <- max(c(ncol_header, ncol_body, 0))    if (ncol < 1)         stop("table needs at least one 1 columns")    if (!all(ncol_header %in% c(0, ncol)))         stop(paste("not all header rows have", ncol, "columns"))    if (!all(ncol_body %in% c(0, ncol)))         stop(paste("not all body rows have", ncol, "columns"))    body_formatted <- lapply(body, propagate_format_to_rcells,         format = format)    structure(body_formatted, header = header, ncol = ncol, nrow = nrow,         class = "rtable")})(header = structure(list(structure(list(structure("A: Drug X", colspan = 1, class = "rcell", format = "xx"),     structure("B: Placebo", colspan = 1, class = "rcell", format = "xx"),     structure("C: Combination", colspan = 1, class = "rcell", format = "xx")), row.name = "", indent = 0, class = "rrow"),     structure(list(`A: Drug X` = structure(1L, format = "(N=xx)", colspan = 1, class = "rcell"),         `B: Placebo` = structure(4L, format = "(N=xx)", colspan = 1, class = "rcell"),         `C: Combination` = structure(5L, format = "(N=xx)", colspan = 1, class = "rcell")), row.name = "", indent = 0, class = "rrow")), nrow = 2L, ncol = 3, class = "rheader"),     format = NULL, NULL, NULL, structure(list(`A: Drug X` = structure(c(1,     1), format = "xx.xx (xx.xx%)", colspan = 1, class = "rcell"),         `B: Placebo` = structure(c(2, 0.5), format = "xx.xx (xx.xx%)", colspan = 1, class = "rcell"),         `C: Combination` = structure(c(4, 0.8), format = "xx.xx (xx.xx%)", colspan = 1, class = "rcell")), row.name = "Total number of patients with at least one adverse event", indent = 0, class = "rrow")): not all arguments in ... are of class rrow

t_events_per_term_id(
  terms = ANL[, c("AEBODSYS", "AEDECOD")],
  id = ANL$USUBJID,
  col_by = ANL$ARM,
  col_N = tapply(ASL$ARM, ASL$ARM, length),
  total = "All Patients"
)

#> MedDRA System Organ Class                                                   A: Drug X             B: Placebo          C: Combination         All Patients 
#>   MedDRA Preferred Term                                                       (N=1)                 (N=4)                 (N=5)                 (N=10)    
#> ----------------------------------------------------------------------------------------------------------------------------------------------------------
#> - Any Adverse Event -
#>   Overall total number of events                                                1                     5                     10                    16      
#>   Overall total number of patients with at least one adverse event           1 (100%)              2 (50%)               4 (80%)               7 (70%)    
#> 
#> cl B
#>   Total number of events                                                        1                     2                     4                     7       
#>   Total number of patients with at least one adverse event                   1 (100%)              2 (50%)               3 (60%)               6 (60%)    
#>   trm B_1/3                                                                  1 (100%)               0 (0%)               3 (60%)               4 (40%)    
#>   trm B_2/3                                                                   0 (0%)               1 (25%)               1 (20%)               2 (20%)    
#>   trm B_3/3                                                                   0 (0%)               1 (25%)                0 (0%)               1 (10%)    
#> 
#> cl C
#>   Total number of events                                                        NA                    1                     4                     5       
#>   Total number of patients with at least one adverse event                    0 (0%)               1 (25%)               3 (60%)               4 (40%)    
#> 
#> cl A
#>   Total number of events                                                        NA                    2                     2                     4       
#>   Total number of patients with at least one adverse event                    0 (0%)               2 (50%)               2 (40%)               4 (40%)    
#>   trm A_2/2                                                                   0 (0%)               2 (50%)                0 (0%)               2 (20%)