Adverse Events by System Organ Class and Preferred Term

t_events_per_term_id returns adverse events summary table that corresponds to STREAM template AET02

t_events_per_term_id(terms, id, col_by, col_N, total = "All Patients")

Arguments

terms
term information as character or factor dataframe, however factor levels are not repeated by class, only terms with count > 1 are listed per class. Currently terms can only be a dataframe with 2 columns. var_relabel is used as the character string used as a label in the column header for each term.

id
unique subject identifier. If a particular subject has no adverse event then that information needs to be added to the col_N argument.

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing. See 'Examples'.

col_N
numeric vector with information of the number of patients in the levels of col_by. This is useful if there are patients that have no adverse events can be accounted for with this argument.

total
character string that will be used as a label for a column with pooled total population, default is "All Patients". If the levels of col_by are the only columns of interest then total should be NULL

Value

rtable object

Details

This is an equivalent of the STREAM output %stream_t_summary(templates = aet02) (http://bioportal.roche.com/stream_doc/2_05/um/report_outputs_aet02.html)

This is an equivalent of the STREAM 1.17 output %stream_t_events_basic(templates = aet02) (https://rochewiki.roche.com/confluence/pages/viewpage.action?pageId=294027342)

Examples

# Simple example
library(tibble)
library(dplyr)

ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = factor(paste("ARM", LETTERS[rep(c(1,2), c(3,7))]))
)

ae_lookup <- tribble(
  ~CLASS,         ~TERM,   ~GRADE,
  "cl A",   "trm A_1/2",        1,
  "cl A",   "trm A_2/2",        2,
  "cl B",   "trm B_1/3",        2,
  "cl B",   "trm B_2/3",        3,
  "cl B",   "trm B_3/3",        1,
  "cl C",   "trm C_1/1",        1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2,2,2,3,3,4,4,4,4,5,6,6,7,7)]
  ),
  ae_lookup[c(1,1,2,6,4,2,2,3,4,2,1,5,4,6),]
)

ANL <- left_join(AAE, ASL, by = "USUBJID")

terms <- data.frame(class = ANL$CLASS, term = ANL$TERM, stringsAsFactors = FALSE)
var_labels(terms) <- c('MedDRA System Organ Class', 'MedDRA Preferred Term')

tbl <- t_events_per_term_id(
  terms = terms,
  id = ANL$USUBJID,
  col_by = ANL$ARM,
  col_N = tapply(ASL$ARM, ASL$ARM, length),
  total = "All Patients"
)

tbl

#> MedDRA System Organ Class                                                    ARM A               ARM B            All Patients
#>   MedDRA Preferred Term                                                      (N=3)               (N=7)               (N=10)   
#> ------------------------------------------------------------------------------------------------------------------------------
#> - Any Adverse Event -
#>   Overall total number of events                                               5                   9                   14     
#>   Overall total number of patients with at least one adverse event         2 (66.67%)          4 (57.14%)           6 (60%)   
#> 
#> cl A
#>   Total number of events                                                       3                   4                   7      
#>   Total number of patients with at least one adverse event                 1 (33.33%)          3 (42.86%)           4 (40%)   
#>   trm A_2/2                                                                1 (33.33%)          2 (28.57%)           3 (30%)   
#>   trm A_1/2                                                                1 (33.33%)          1 (14.29%)           2 (20%)   
#> 
#> cl B
#>   Total number of events                                                       1                   4                   5      
#>   Total number of patients with at least one adverse event                 1 (33.33%)          3 (42.86%)           4 (40%)   
#>   trm B_2/3                                                                1 (33.33%)          2 (28.57%)           3 (30%)   
#>   trm B_1/3                                                                  0 (0%)            1 (14.29%)           1 (10%)   
#>   trm B_3/3                                                                  0 (0%)            1 (14.29%)           1 (10%)   
#> 
#> cl C
#>   Total number of events                                                       1                   1                   2      
#>   Total number of patients with at least one adverse event                 1 (33.33%)          1 (14.29%)           2 (20%)   
#>   trm C_1/1                                                                1 (33.33%)          1 (14.29%)           2 (20%)   

# Example using dummy data
library(random.cdisc.data)
library(dplyr)

ASL <- radam("ASL", N = 10)
AAE <- radam("AAE", ADSL = ASL)

ANL <- left_join(AAE, ASL %>% select(USUBJID, STUDYID, ARM), by = c("STUDYID", "USUBJID"))

terms <- data.frame(class = ANL$AEBODSYS, term = ANL$AEDECOD, stringsAsFactors = FALSE)
var_labels(terms) <- c('MedDRA System Organ Class', 'MedDRA Preferred Term')

tbl <- t_events_per_term_id(terms = terms,
            id = ANL$USUBJID,
            col_by = factor(ANL$ARM),
            total = "All Patients"
            )

tbl

#> MedDRA System Organ Class                                                     ARM A               ARM B            All Patients
#>   MedDRA Preferred Term                                                       (N=40)              (N=23)              (N=63)   
#> -------------------------------------------------------------------------------------------------------------------------------
#> - Any Adverse Event -
#>   Overall total number of events                                                40                  23                  63     
#>   Overall total number of patients with at least one adverse event           6 (15%)            4 (17.39%)         10 (15.87%) 
#> 
#> Investigations
#>   Total number of events                                                        6                   7                   13     
#>   Total number of patients with at least one adverse event                  5 (12.5%)           4 (17.39%)          9 (14.29%) 
#>   Hepcidin test                                                               0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Peripheral nervous system function test abnormal                           1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Matrix metalloproteinase-3                                                  0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Urine cotinine test negative                                                0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Model for end stage liver disease score                                     0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Renal scan                                                                 1 (2.5%)             0 (0%)            1 (1.59%)  
#>   CSF neutrophil count decreased                                             1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Cytomegalovirus test                                                        0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Iron binding capacity unsaturated increased                                1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Visual evoked potentials abnormal                                          1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Blood lead abnormal                                                        1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Troponin T normal                                                           0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Carbon monoxide diffusing capacity                                          0 (0%)            1 (4.35%)           1 (1.59%)  
#> 
#> Infections and infestations
#>   Total number of events                                                        4                   2                   6      
#>   Total number of patients with at least one adverse event                   4 (10%)            1 (4.35%)           5 (7.94%)  
#>   Prostate infection                                                         1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Cholecystitis infective                                                    1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Malignant syphilis                                                         1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Pyelonephritis acute                                                       1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Ear tuberculosis                                                            0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Milker's nodules                                                            0 (0%)            1 (4.35%)           1 (1.59%)  
#> 
#> General disorders and administration site conditions
#>   Total number of events                                                        3                   2                   5      
#>   Total number of patients with at least one adverse event                   3 (7.5%)            2 (8.7%)           5 (7.94%)  
#>   Administration site photosensitivity reaction                               0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Deficiency of the interleukin-1 receptor antagonist                        1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Drug interaction                                                           1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Vaccination site cyst                                                      1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Infusion site abscess sterile                                               0 (0%)            1 (4.35%)           1 (1.59%)  
#> 
#> Gastrointestinal disorders
#>   Total number of events                                                        3                   1                   4      
#>   Total number of patients with at least one adverse event                   3 (7.5%)           1 (4.35%)           4 (6.35%)  
#>   Defaecation disorder                                                       1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Tongue disorder                                                            1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Mouth haemorrhage                                                          1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Peritoneal lesion                                                           0 (0%)            1 (4.35%)           1 (1.59%)  
#> 
#> Injury, poisoning and procedural complications
#>   Total number of events                                                        2                   2                   4      
#>   Total number of patients with at least one adverse event                   1 (2.5%)           1 (4.35%)           2 (3.17%)  
#>   Radiation neuropathy                                                       1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Catheter site abscess                                                      1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Application site scar                                                       0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Application site photosensitivity reaction                                  0 (0%)            1 (4.35%)           1 (1.59%)  
#> 
#> Neoplasms benign, malignant and unspecified (incl cysts and polyps)
#>   Total number of events                                                        3                   1                   4      
#>   Total number of patients with at least one adverse event                   3 (7.5%)           1 (4.35%)           4 (6.35%)  
#>   Malignant neoplasm of spermatic cord                                       1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Benign laryngeal neoplasm                                                   0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Testicular teratoma benign                                                 1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Chondromatosis                                                             1 (2.5%)             0 (0%)            1 (1.59%)  
#> 
#> Nervous system disorders
#>   Total number of events                                                        3                   NA                  3      
#>   Total number of patients with at least one adverse event                   3 (7.5%)             0 (0%)            3 (4.76%)  
#>   Idioglossia                                                                1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Parkinsonian crisis                                                        1 (2.5%)             0 (0%)            1 (1.59%)  
#> 
#> Psychiatric disorders
#>   Total number of events                                                        2                   1                   3      
#>   Total number of patients with at least one adverse event                    2 (5%)            1 (4.35%)           3 (4.76%)  
#>   Change in sustained attention                                               0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Thermophobia                                                               1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Delayed sleep phase                                                        1 (2.5%)             0 (0%)            1 (1.59%)  
#> 
#> Surgical and medical procedures
#>   Total number of events                                                        1                   2                   3      
#>   Total number of patients with at least one adverse event                   1 (2.5%)            2 (8.7%)           3 (4.76%)  
#>   Low-level laser therapy                                                     0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Intra-uterine contraceptive device                                          0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Adrenocortical steroid therapy                                             1 (2.5%)             0 (0%)            1 (1.59%)  
#> 
#> Blood and lymphatic system disorders
#>   Total number of events                                                        2                   NA                  2      
#>   Total number of patients with at least one adverse event                    2 (5%)              0 (0%)            2 (3.17%)  
#>   Malignant histiocytosis                                                    1 (2.5%)             0 (0%)            1 (1.59%)  
#> 
#> Cardiac disorders
#>   Total number of events                                                        2                   NA                  2      
#>   Total number of patients with at least one adverse event                    2 (5%)              0 (0%)            2 (3.17%)  
#>   Foetal heart rate disorder                                                 1 (2.5%)             0 (0%)            1 (1.59%)  
#> 
#> Immune system disorders
#>   Total number of events                                                        NA                  2                   2      
#>   Total number of patients with at least one adverse event                    0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Reaction to preservatives                                                   0 (0%)            1 (4.35%)           1 (1.59%)  
#> 
#> Musculoskeletal and connective tissue disorders
#>   Total number of events                                                        1                   1                   2      
#>   Total number of patients with at least one adverse event                   1 (2.5%)           1 (4.35%)           2 (3.17%)  
#>   Extraskeletal ossification                                                  0 (0%)            1 (4.35%)           1 (1.59%)  
#>   Perichondritis                                                             1 (2.5%)             0 (0%)            1 (1.59%)  
#> 
#> Reproductive system and breast disorders
#>   Total number of events                                                        2                   NA                  2      
#>   Total number of patients with at least one adverse event                    2 (5%)              0 (0%)            2 (3.17%)  
#>   Vaginal cancer stage I                                                     1 (2.5%)             0 (0%)            1 (1.59%)  
#> 
#> Social circumstances
#>   Total number of events                                                        1                   1                   2      
#>   Total number of patients with at least one adverse event                   1 (2.5%)           1 (4.35%)           2 (3.17%)  
#>   Circumcised                                                                1 (2.5%)             0 (0%)            1 (1.59%)  
#>   Refusal of treatment by patient                                             0 (0%)            1 (4.35%)           1 (1.59%)  
#> 
#> Vascular disorders
#>   Total number of events                                                        2                   NA                  2      
#>   Total number of patients with at least one adverse event                    2 (5%)              0 (0%)            2 (3.17%)  
#>   Pulmonary contusion                                                        1 (2.5%)             0 (0%)            1 (1.59%)  
#> 
#> Congenital, familial and genetic disorders
#>   Total number of events                                                        NA                  1                   1      
#>   Total number of patients with at least one adverse event                    0 (0%)            1 (4.35%)           1 (1.59%)  
#> 
#> Endocrine disorders
#>   Total number of events                                                        1                   NA                  1      
#>   Total number of patients with at least one adverse event                   1 (2.5%)             0 (0%)            1 (1.59%)  
#> 
#> Respiratory, thoracic and mediastinal disorders
#>   Total number of events                                                        1                   NA                  1      
#>   Total number of patients with at least one adverse event                   1 (2.5%)             0 (0%)            1 (1.59%)  
#> 
#> Skin and subcutaneous tissue disorders
#>   Total number of events                                                        1                   NA                  1      
#>   Total number of patients with at least one adverse event                   1 (2.5%)             0 (0%)            1 (1.59%)