Basic Events Table

This function summarizes number of unique subjects with events and total number of events. It creates basic summary of events and can be used for any events data like Adverse Events, concomitant medication, medical history, etc. Implementation examples are to apply t_events_per_term_id on Adverse Event data to create Adverse Events summary table (AET02, STREAM2.x, STREAM1.17 ), or apply on Concomitatant Medication data to create Concomitant Treatment summary table (CMT01, STREAM2.x, STREAM1.17).

t_events_per_term_id(terms, id, col_by, col_N, total = "All Patients",
  event_type = "event")

Arguments

terms
character or factor vector, or dataframe to represent events information; Currently terms can only be a vector or dataframe with 1 or 2 columns. For terms with 2 columns, 1st column should represent higher level term and 2nd column should be lower level term. var_relabel is used as the character string used as a label in the column header for each term.

id
vector of subject identifier. Length of id must be the same as the length or number of rows of terms.

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing. See 'Examples'.

col_N
numeric vector with information of the number of patients in the levels of col_by. This is useful if there are patients that have no adverse events can be accounted for with this argument.

total
character string that will be used as a label for a column with pooled total population, default is "All Patients". If the levels of col_by are the only columns of interest then total should be NULL

event_type
type of event that is summarized (e.g. adverse event, treatment). Default is "event".

Value

an rtable object.

Details

t_events_per_term_id includes percentages based on the total number of subjects in the column heading (i.e. "N=nnn"). col_N can be explicitly specified to get N for percentage calculation from either events dataset or additional dataset such as subject level dataset. See the example.

Multiple events within a subject of the same term (if terms is one level) or lower level term (if terms is two levels) are counted once when counting number of subjects.

t_events_per_term_id doesn't deal with data with any non-complete records (has NA's), e.g. if any terms are missing. Impute missing values before using t_events_per_term_id.

t_events_per_term_id orders data by "All Patients" column from the most commonly reported higher level term to the least frequent one. Within each group of higher level term, it sorts by decreasing frequency of lower level term. It brakes ties using terms names in alphabetical order.

Examples

library(dplyr)
library(random.cdisc.data)

ASL <- rasl(10, seed = 1)
AAE <- raae(ASL, 4, seed = 2)

ANL <- left_join(AAE, ASL %>% select(USUBJID, STUDYID, ARM), by = c("USUBJID", "STUDYID")) %>%
  var_relabel(
    AEBODSYS = 'MedDRA System Organ Class',
    AEDECOD = 'MedDRA Preferred Term',
    AETOXGR = 'GRADE'
)

t_events_per_term_id(
  terms = ANL$AEDECOD,
  id = ANL$USUBJID,
  col_by = ANL$ARM,
  col_N = table(ASL$ARM),
  total = NULL,
  event_type = "adverse event"
)

#>                                                                   A: Drug X             B: Placebo          C: Combination
#> MedDRA Preferred Term                                               (N=1)                 (N=4)                 (N=5)     
#> --------------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event           1 (100%)              2 (50%)               4 (80%)    
#> Total number of adverse events                                        1                     5                     10      
#> trm B_1/3                                                          1 (100%)               0 (0%)               3 (60%)    
#> trm C_1/1                                                           0 (0%)               1 (25%)               3 (60%)    
#> trm A_1/2                                                           0 (0%)                0 (0%)               2 (40%)    
#> trm A_2/2                                                           0 (0%)               2 (50%)                0 (0%)    
#> trm B_2/3                                                           0 (0%)               1 (25%)               1 (20%)    
#> trm B_3/3                                                           0 (0%)               1 (25%)                0 (0%)    

t_events_per_term_id(
  terms = ANL[, c("AEBODSYS", "AEDECOD")],
  id = ANL$USUBJID,
  col_by = ANL$ARM,
  col_N = table(ASL$ARM),
  total = "All Patients"
)

#> MedDRA System Organ Class                                   A: Drug X             B: Placebo          C: Combination         All Patients 
#>   MedDRA Preferred Term                                       (N=1)                 (N=4)                 (N=5)                 (N=10)    
#> ------------------------------------------------------------------------------------------------------------------------------------------
#> - Any event -
#>   Total number of patients with at least one event           1 (100%)              2 (50%)               4 (80%)               7 (70%)    
#>   Total number of events                                        1                     5                     10                    16      
#> 
#> cl B
#>   Total number of patients with at least one event           1 (100%)              2 (50%)               3 (60%)               6 (60%)    
#>   Total number of events                                        1                     2                     4                     7       
#>   trm B_1/3                                                  1 (100%)               0 (0%)               3 (60%)               4 (40%)    
#>   trm B_2/3                                                   0 (0%)               1 (25%)               1 (20%)               2 (20%)    
#>   trm B_3/3                                                   0 (0%)               1 (25%)                0 (0%)               1 (10%)    
#> 
#> cl A
#>   Total number of patients with at least one event            0 (0%)               2 (50%)               2 (40%)               4 (40%)    
#>   Total number of events                                        0                     2                     2                     4       
#>   trm A_1/2                                                   0 (0%)                0 (0%)               2 (40%)               2 (20%)    
#>   trm A_2/2                                                   0 (0%)               2 (50%)                0 (0%)               2 (20%)    
#> 
#> cl C
#>   Total number of patients with at least one event            0 (0%)               1 (25%)               3 (60%)               4 (40%)    
#>   Total number of events                                        0                     1                     4                     5       
#>   trm C_1/1                                                   0 (0%)               1 (25%)               3 (60%)               4 (40%)    

ACM <- racm(ASL, 5, seed = 4)
ANL <- left_join(ACM, ASL %>% select(USUBJID, STUDYID, ARM), by = c("USUBJID", "STUDYID")) %>%
       filter(CMTIREL == "CONCOMITANT") %>%
  var_relabel(
    CMCLAS = 'Medication Class',
    CMDECOD = 'Standardized Medication Name'
)
t_events_per_term_id(
  terms = ANL[, c("CMCLAS", "CMDECOD")],
  id = ANL$USUBJID,
  col_by = ANL$ARM,
  col_N = table(ASL$ARM),
  total = "All Patients",
  event_type = "treatment"
)

#> Medication Class                                                A: Drug X             B: Placebo          C: Combination         All Patients 
#>   Standardized Medication Name                                    (N=1)                 (N=4)                 (N=5)                 (N=10)    
#> ----------------------------------------------------------------------------------------------------------------------------------------------
#> - Any treatment -
#>   Total number of patients with at least one treatment           1 (100%)              4 (100%)              4 (80%)               9 (90%)    
#>   Total number of treatments                                        2                     9                     7                     18      
#> 
#> medcl C
#>   Total number of patients with at least one treatment           1 (100%)              3 (75%)               3 (60%)               7 (70%)    
#>   Total number of treatments                                        2                     4                     3                     9       
#>   medname C_2/2                                                   0 (0%)               3 (75%)               3 (60%)               6 (60%)    
#>   medname C_1/2                                                  1 (100%)               0 (0%)                0 (0%)               1 (10%)    
#> 
#> medcl A
#>   Total number of patients with at least one treatment            0 (0%)               3 (75%)               1 (20%)               4 (40%)    
#>   Total number of treatments                                        0                     3                     2                     5       
#>   medname A_3/3                                                   0 (0%)               3 (75%)               1 (20%)               4 (40%)    
#> 
#> medcl B
#>   Total number of patients with at least one treatment            0 (0%)               2 (50%)               1 (20%)               3 (30%)    
#>   Total number of treatments                                        0                     2                     2                     4       
#>   medname B_4/4                                                   0 (0%)               1 (25%)               1 (20%)               2 (20%)    
#>   medname B_1/4                                                   0 (0%)               1 (25%)                0 (0%)               1 (10%)