Response Forest Plot Table

Response Forest Plot Table

t_forest_rsp(rsp, col_by, group_data = NULL, total = "ALL",
  na.omit.group = TRUE, dense_header = FALSE)

Arguments

rsp
boolean, TRUE if subject is a responder FALSE otherwise

col_by
factor with reference and comparison group information, the first level indicates the reference group

group_data
data frame with one column per grouping variable

total
character with the row name of the analysis run on all data. If NULL analysis is omitted

na.omit.group
boolean, do not list NAs

Value

an rtable object

Details

Logistic model is used for odds ratio calculation.

The returned table contains one row per analysis applied on a subset of data (indicated by the row name). The analysis is summarized with the following 9 columns:

1

Total n the total number of subjects included in analysis population

3-4

Summary of responders in reference arm, n and Responders are the total number of patients and the number of responders in reference arm, respectively. Response.Rate is the percentage of responders in reference arm.

5-7

same statistics as for reference arm now for comparison arm

8

Odds Ratio ranges from 0 to infinity, estimated by applying univariate logistic regression with binary response status (responder or non-responder) as the the outcome and arm variable as the explanatory variable. Odds ratio greater than 1 indicates better performance in comparison arm; odds ratio less than 1 indicates better performance in reference arm.

9

95 % CI The 95 around the measure of effect (Odds Ratio). Because only a small sample of the overall population is included in the analysis, by having an upper and lower confidence limit we can infer that the true treatment effect lies in between. If the 95 includes 1, then we say that the difference between two arms is not significant at a significance level of 0.05.

Examples

library(random.cdisc.data)

ASL <- radam("ASL")
ARS <- radam("ARS", ADSL = ASL)

ARS_f <- subset(ARS, PARAMCD == "OVRSPI")
ANL <- merge(ASL, ARS_f)

tbl <- t_forest_rsp(
  rsp = ANL$AVALC %in% c("CR", "PR"),
  col_by = factor(ANL$ARM),
  group_data = ANL[, c("SEX", "RACE")]
)

tbl

#>                                                                                                  ARM A                                                          ARM B                                                                   
#> Baseline Risk Factors                                 Total n                 n               Responders          Response.Rate              n               Responders          Response.Rate         Odds Ratio             95% CI    
#> ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                                                     100                  54                   22                  0.41                  46                   22                  0.48                 1.33              (0.6, 2.96) 
#> 
#> SEX
#>   F                                                     54                   27                   13                  0.48                  27                   14                  0.52                 1.16              (0.4, 3.41) 
#>   M                                                     46                   27                    9                  0.33                  19                    8                  0.42                 1.45             (0.43, 4.96) 
#> 
#> RACE
#>   AMERICAN INDIAN OR ALASKA NATIVE                      18                    5                    0                    0                   13                    6                  0.46              99127537.56            (0, NA)   
#>   ASIAN                                                 25                   16                    8                   0.5                   9                    6                  0.67                   2              (0.38, 12.26)
#>   BLACK OR AFRICAN AMERICAN                             18                   12                    6                   0.5                   6                    2                  0.33                  0.5             (0.05, 3.66) 
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER             22                   13                    5                  0.38                   9                    4                  0.44                 1.28             (0.22, 7.42) 
#>   WHITE                                                 17                    8                    3                  0.38                   9                    4                  0.44                 1.33             (0.19, 10.07)

# NOT RUN {
Viewer(tbl)
# }