Response Forest Plot Table

Response Forest Plot Table

t_forest_rsp(rsp, col_by, group_data = NULL, total = "ALL",
  na.omit.group = TRUE)

Arguments

rsp
boolean, TRUE if subject is a responder FALSE otherwise

col_by
factor with reference and comparison group information, the first level indicates the reference group

group_data
data frame with one column per grouping variable

total
character with the row name of the analysis run on all data. If NULL analysis is omitted

na.omit.group
boolean, do not list NAs

Value

an rtable object

Details

Logistic model is used for odds ratio calculation.

The returned table contains one row per analysis applied on a subset of data (indicated by the row name). The analysis is summarized with the following 8 columns:

1

Total n the total number of subjects used for analysis

1-3

analysis for reference arm, n, Responders, Response.Rate

4-6

same analysis as for reference arm now for comparison arm

7

Odds Ratio

8

95 % CI

Examples

library(random.cdisc.data)

ASL <- radam("ASL")
ARS <- radam("ARS", ADSL = ASL)

ARS_f <- subset(ARS, PARAMCD == "OVRSPI")
ANL <- merge(ASL, ARS_f)

tbl <- t_forest_rsp(
  rsp = ANL$AVALC %in% c("CR", "PR"),
  col_by = ANL$ARM,
  group_data = ANL[, c("SEX", "RACE")]
)

tbl

#>                                                                                                  ARM A                                                          ARM B                                                                   
#> Baseline Risk Factors                                 Total n                 n               Responders          Response.Rate              n               Responders          Response.Rate         Odds Ratio             95% CI    
#> ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                                                     100                  42                   14                  0.33                  58                   19                  0.33                 0.97             (0.42, 2.29) 
#> 
#> SEX
#>   F                                                     44                   17                    5                  0.29                  27                    6                  0.22                 0.69             (0.17, 2.83) 
#>   M                                                     56                   25                    9                  0.36                  31                   13                  0.42                 1.28             (0.44, 3.88) 
#> 
#> RACE
#>   AMERICAN INDIAN OR ALASKA NATIVE                      20                    8                    3                  0.38                  12                    4                  0.33                 0.83             (0.13, 5.77) 
#>   ASIAN                                                 18                    9                    6                  0.67                   9                    1                  0.11                 0.06               (0, 0.57)  
#>   BLACK OR AFRICAN AMERICAN                             20                    7                    2                  0.29                  13                    5                  0.38                 1.56             (0.23, 13.99)
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER             24                   12                    2                  0.17                  12                    4                  0.33                  2.5             (0.38, 21.6) 
#>   WHITE                                                 18                    6                    1                  0.17                  12                    5                  0.42                 3.57             (0.4, 80.15)