Response Table as used for Forest Plot

The Response forest plot table summarizes response data by groups. The function returns sample sizes and responder counts and response rates for each analysis arm, as well as a odds ratio and the corresponding confidence interval from a univariate logistic model.

t_forest_rsp(
  rsp,
  col_by,
  row_by_list = NULL,
  total = "ALL",
  strata_data = NULL,
  conf_level = 0.95,
  dense_header = FALSE,
  table_tree = FALSE
)

Arguments

rsp
(logical vector)
TRUE if subject is a responder FALSE otherwise

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

row_by_list
(list or data.frame) of factors
Contains one factor variable to calculate the t_el_forest_tte

total
(character value)
Total row added. If NULL then no total row is added.

strata_data
(character, factor or data.frame)
Used for stratification factors If NULL, no stratified analysis is performed.

conf_level
confidence level of the interval.

dense_header
(logical value) Display the table headers in multiple rows.

table_tree
(logical value)
If FALSE then a single rtable gets returned, if TRUE a node object gets returned with the elementary rtable objects.

Value

an rtable object

Details

Logistic regression is used for odds ratio calculation.

Each row in the returned table contains the analysis statistics for a subgroup of data (indicated by the row name). The summary table is consist of the following 9 columns:

1

Total n the total number of subjects included in analysis population

3-4

Summary of responders in reference arm, n and responders are the total number of patients and the number of responders in reference arm, respectively. Response.Rate is the percentage of responders in reference arm.

5-7

same statistics as for reference arm now for comparison arm

8

Odds Ratio ranges from 0 to infinity, estimated by applying tern::odds_ratio. If stratification factors are added, Cochran-Mantel-Haensel test is performed instead. Statistics from the stratification-adjusted test will be reported for the p-value of test of equal proportions, odds ratio and its corresponding 95% confidence interval. Odds ratio greater than 1 indicates better performance in comparison arm; odds ratio less than 1 indicates better performance in reference arm.

9

Confidence Interval The confidence interval indicates the level of uncertainty around the measure of effect (Odds Ratio). Because only a small sample of the overall population is included in the analysis, by having an upper and lower confidence limit we can infer that the true treatment effect lies in between. By default, the 95 If it includes 1, then we say that the difference between two arms is not significant at a significance level of 0.05.

See also

Examples

library(random.cdisc.data)
library(dplyr)
library(purrr)

ADSL <- radsl(cached = TRUE)
ADSL$'FAKE Name > -1.3 Flag' <- rep(c('Y', 'N'), length.out = nrow(ADSL))
attr(ADSL$'FAKE Name > -1.3 Flag', "label") <- "dummy"

ADRS <- radrs(ADSL, seed = 2)
ADRS_f <- subset(ADRS, PARAMCD == "BESRSPI") %>%
  dplyr::filter(ARM != 'C: Combination') %>%
  mutate(ARM = droplevels(ARM))
var_labels(ADRS_f) <- var_labels(ADRS)

tbl <- t_forest_rsp(
  rsp = ADRS_f$AVALC %in% c("CR", "PR"),
  col_by = as_factor_keep_attributes(ADRS_f$ARM),
  row_by_list = ADRS_f[, c("SEX", "RACE", "FAKE Name > -1.3 Flag")] %>%
    map(as_factor_keep_attributes),
  conf_level = 0.9
)

tbl

#>                                                                                                A: Drug X                                                     B: Placebo                                                                 
#> Baseline Risk Factors                                 Total n                 n               Responders          Response.Rate              n               Responders          Response.Rate         Odds Ratio             90% CI    
#> ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                                                     268                  134                  131                 0.98                  134                  122                 0.91                 0.23             (0.08, 0.69) 
#> 
#> Sex
#>   F                                                     161                  79                   77                  0.97                  82                   78                  0.95                 0.51             (0.12, 2.16) 
#>   M                                                     107                  55                   54                  0.98                  52                   44                  0.85                  0.1              (0.02, 0.6) 
#> 
#> Race
#>   ASIAN                                                 135                  68                   67                  0.99                  67                   61                  0.91                 0.15             (0.03, 0.92) 
#>   BLACK OR AFRICAN AMERICAN                             59                   31                   29                  0.94                  28                   26                  0.93                  0.9             (0.16, 4.93) 
#>   WHITE                                                 53                   27                   27                    1                   26                   24                  0.92                   0                (0, NaN)   
#>   AMERICAN INDIAN OR ALASKA NATIVE                      19                    8                    8                    1                   11                   10                  0.91                   0                (0, NaN)   
#>   MULTIPLE                                               1                    0                    0                                         1                    1                    1                                     (NA, NA)   
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER              1                    0                    0                                         1                    0                    0                                     (NA, NA)   
#>   OTHER                                                  0                    0                    0                                         0                    0                                                          (NA, NA)   
#>   UNKNOWN                                                0                    0                    0                                         0                    0                                                          (NA, NA)   
#> 
#> dummy
#>   N                                                     132                  72                   69                  0.96                  60                   54                   0.9                 0.39              (0.12, 1.3) 
#>   Y                                                     136                  62                   62                    1                   74                   68                  0.92                   0                (0, NaN)   

if (FALSE) {
Viewer(tbl)
}

tbl2 <- t_forest_rsp(
  rsp = ADRS_f$AVALC %in% c("CR", "PR"),
  col_by = as_factor_keep_attributes(ADRS_f$ARM),
  row_by_list = ADRS_f[, c("BMRKR2")] %>%
    map(as_factor_keep_attributes),
  strata_data = ADRS_f[ , "STRATA1"]
)

if (FALSE) {
Viewer(tbl2)
}

# table tree
tbls <- t_forest_rsp(
  rsp = ADRS_f$AVALC %in% c("CR", "PR"),
  col_by = ADRS_f$ARM,
  row_by_list = ADRS_f[, c("SEX", "RACE", "FAKE Name > -1.3 Flag")] %>%
    map(as_factor_keep_attributes),
  table_tree = TRUE
)
summary(tbls)

#>                                                        index            content             dim            hide label          format  
#> ---------------------------------------------------------------------------------------------------------------------------------------
#> root                                                                      NULL                                TRUE               1     
#>   ALL                                                    1               rtable            1 x 9              TRUE                     
#>   Sex                                                    2                NULL                               FALSE               0     
#>     F                                                   2, 1             rtable            1 x 9              TRUE               0     
#>     M                                                   2, 2             rtable            1 x 9              TRUE               0     
#>   Race                                                   3                NULL                               FALSE               0     
#>     ASIAN                                               3, 1             rtable            1 x 9              TRUE               0     
#>     BLACK OR AFRICAN AMERICAN                           3, 2             rtable            1 x 9              TRUE               0     
#>     WHITE                                               3, 3             rtable            1 x 9              TRUE               0     
#>     AMERICAN INDIAN OR ALASKA NATIVE                    3, 4             rtable            1 x 9              TRUE               0     
#>     MULTIPLE                                            3, 5             rtable            1 x 9              TRUE               0     
#>     NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER           3, 6             rtable            1 x 9              TRUE               0     
#>     OTHER                                               3, 7             rtable            1 x 9              TRUE               0     
#>     UNKNOWN                                             3, 8             rtable            1 x 9              TRUE               0     
#>   dummy                                                  4                NULL                               FALSE               0     
#>     N                                                   4, 1             rtable            1 x 9              TRUE               0     
#>     Y                                                   4, 2             rtable            1 x 9              TRUE               0