Response Table as used for Forest Plot

The Response forest plot table summarizes response data by groups. The function returns sample sizes and responder counts and response rates for each analysis arm, as well as a odds ratio and the corresponding 95% confidence interval from a univariate logistic model.

t_forest_rsp(rsp, col_by, group_data = NULL, total = "ALL",
  na.omit.group = TRUE, dense_header = FALSE)

Arguments

rsp
is a boolean vector. If TRUE, observation is a response, otherwise, FALSE.

col_by
factor with reference and comparison group information, the first level indicates the reference group

group_data
data frame with one variable per grouping

total
character with the row name of the analysis run on all data. If NULL analysis is omitted.

na.omit.group
is boolean. Default is TRUE, do not display NA's as a category.

dense_header
Display the table headers in mulitple rows. Default is FALSE.

Value

an rtable object

Details

Logistic regression is used for odds ratio calculation.

Each row in the returned table contains the analysis statistics for a subgroup of data (indicated by the row name). The summary table is consist of the following 9 columns:

1

Total n the total number of subjects included in analysis population

3-4

Summary of responders in reference arm, n and Responders are the total number of patients and the number of responders in reference arm, respectively. Response.Rate is the percentage of responders in reference arm.

5-7

same statistics as for reference arm now for comparison arm

8

Odds Ratio ranges from 0 to infinity, estimated by applying univariate logistic regression. Binary response status (responder or non-responder) is the outcome and arm is the explanatory variable. Odds ratio greater than 1 indicates better performance in comparison arm; odds ratio less than 1 indicates better performance in reference arm.

9

95 % CI The 95 uncertainty around the measure of effect (Odds Ratio). Because only a small sample of the overall population is included in the analysis, by having an upper and lower confidence limit we can infer that the true treatment effect lies in between. If the 95 say that the difference between two arms is not significant at a significance level of 0.05.

See also

Examples

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(seed = 1)
ADSL$'FAKE Name > -1.3 Flag' <- rep(c('Y', 'N'), 50)

ADRS <- radrs(ADSL, seed = 2)
ADRS_f <- subset(ADRS, PARAMCD == "BESRSPI") %>%
  filter(ARM != 'C: Combination') %>%
  mutate(ARM = droplevels(ARM))

tbl <- t_forest_rsp(
  rsp = ADRS_f$AVALC %in% c("CR", "PR"),
  col_by = factor(ADRS_f$ARM),
  group_data = ADRS_f[, c("SEX", "RACE", "FAKE Name > -1.3 Flag")]
)

tbl

#>                                                                                                A: Drug X                                                     B: Placebo                                                                 
#> Baseline Risk Factors                                 Total n                 n               Responders          Response.Rate              n               Responders          Response.Rate         Odds Ratio             95% CI    
#> ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                                                     275                  142                  137                 0.96                  133                  87                  0.65                 0.07             (0.02, 0.17) 
#> 
#> SEX
#>   F                                                     136                  69                   66                  0.96                  67                   44                  0.66                 0.09             (0.02, 0.27) 
#>   M                                                     131                  68                   66                  0.97                  63                   40                  0.63                 0.05             (0.01, 0.19) 
#>   U                                                      5                    2                    2                    1                    3                    3                    1                    1                (NA, NA)   
#>   UNDIFFERENTIATED                                       3                    3                    3                    1                    0                    0                   NaN                                    (NA, NA)   
#> 
#> RACE
#>   ASIAN                                                 166                  92                   89                  0.97                  74                   51                  0.69                 0.07             (0.02, 0.23) 
#>   BLACK OR AFRICAN AMERICAN                             56                   26                   26                    1                   30                   18                   0.6                   0              (NA, >999.9) 
#>   WHITE                                                 37                   18                   16                  0.89                  19                   11                  0.58                 0.17             (0.02, 0.84) 
#>   AMERICAN INDIAN OR ALASKA NATIVE                      12                    4                    4                    1                    8                    5                  0.62                   0              (NA, >999.9) 
#>   MULTIPLE                                               1                    0                    0                   NaN                   1                    1                    1                                     (NA, NA)   
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER              2                    2                    2                    1                    0                    0                   NaN                                    (NA, NA)   
#>   OTHER                                                  0                    0                    0                   NaN                   0                    0                   NaN                                    (NA, NA)   
#>   UNKNOWN                                                1                    0                    0                   NaN                   1                    1                    1                                     (NA, NA)   
#> 
#> FAKE Name > -1.3 Flag
#>   N                                                     137                  69                   67                  0.97                  68                   42                  0.62                 0.05             (0.01, 0.17) 
#>   Y                                                     138                  73                   70                  0.96                  65                   45                  0.69                  0.1              (0.02, 0.3) 

# NOT RUN {
Viewer(tbl)
# }