Response Table as used for Forest Plot

The Response forest plot table summarizes response data by groups. The function returns sample sizes and responder counts and response rates for each analysis arm, as well as a odds ratio and the corresponding 95% confidence interval from a univariate logistic model.

t_forest_rsp(rsp, col_by, group_data = NULL, total = "ALL",
  na.omit.group = TRUE, dense_header = FALSE)

Arguments

rsp
is a boolean vector. If TRUE, observation is a response, otherwise, FALSE.

col_by
factor with reference and comparison group information, the first level indicates the reference group

group_data
data frame with one variable per grouping

total
character with the row name of the analysis run on all data. If NULL analysis is omitted.

na.omit.group
is boolean. Default is TRUE, do not display NA's as a category.

dense_header
Display the table headers in mulitple rows. Default is FALSE.

Value

an rtable object

Details

Logistic regression is used for odds ratio calculation.

Each row in the returned table contains the analysis statistics for a subgroup of data (indicated by the row name). The summary table is consist of the following 9 columns:

1

Total n the total number of subjects included in analysis population

3-4

Summary of responders in reference arm, n and Responders are the total number of patients and the number of responders in reference arm, respectively. Response.Rate is the percentage of responders in reference arm.

5-7

same statistics as for reference arm now for comparison arm

8

Odds Ratio ranges from 0 to infinity, estimated by applying univariate logistic regression. Binary response status (responder or non-responder) is the outcome and arm is the explanatory variable. Odds ratio greater than 1 indicates better performance in comparison arm; odds ratio less than 1 indicates better performance in reference arm.

9

95 % CI The 95 uncertainty around the measure of effect (Odds Ratio). Because only a small sample of the overall population is included in the analysis, by having an upper and lower confidence limit we can infer that the true treatment effect lies in between. If the 95 say that the difference between two arms is not significant at a significance level of 0.05.

Examples

library(random.cdisc.data)

ADSL <- radsl()
ADSL$'FAKE Name > -1.3 Flag' <- rep(c('Y', 'N'), 50)

ADRS <- radrs(ADSL)
ADRS_f <- subset(ADRS, PARAMCD == "BESRSPI") %>%
  filter(ARM != 'C: Combination')

tbl <- t_forest_rsp(
  rsp = ADRS_f$AVALC %in% c("CR", "PR"),
  col_by = factor(ADRS_f$ARM),
  group_data = ADRS_f[, c("SEX", "RACE", "FAKE Name > -1.3 Flag")]
)

tbl

#>                                                                                                A: Drug X                                                     B: Placebo                                                                 
#> Baseline Risk Factors                                 Total n                 n               Responders          Response.Rate              n               Responders          Response.Rate         Odds Ratio             95% CI    
#> ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                                                     270                  132                  124                 0.94                  138                  108                 0.78                 0.23              (0.1, 0.51) 
#> 
#> SEX
#>   F                                                     117                  65                   59                  0.91                  52                   37                  0.71                 0.25             (0.08, 0.68) 
#>   M                                                     149                  66                   64                  0.97                  83                   68                  0.82                 0.14             (0.02, 0.53) 
#>   U                                                      3                    1                    1                    1                    2                    2                    1                    1                (NA, NA)   
#>   UNDIFFERENTIATED                                       1                    0                    0                   NaN                   1                    1                    1                                     (NA, NA)   
#> 
#> RACE
#>   ASIAN                                                 148                  70                   67                  0.96                  78                   63                  0.81                 0.19              (0.04, 0.6) 
#>   BLACK OR AFRICAN AMERICAN                             55                   25                   23                  0.92                  30                   23                  0.77                 0.29             (0.04, 1.33) 
#>   WHITE                                                 51                   32                   30                  0.94                  19                   15                  0.79                 0.25             (0.03, 1.43) 
#>   AMERICAN INDIAN OR ALASKA NATIVE                      13                    5                    4                   0.8                   8                    5                  0.62                 0.42              (0.02, 4.9) 
#>   MULTIPLE                                               2                    0                    0                   NaN                   2                    1                   0.5                                    (NA, NA)   
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER              0                    0                    0                   NaN                   0                    0                   NaN                                    (NA, NA)   
#>   OTHER                                                  0                    0                    0                   NaN                   0                    0                   NaN                                    (NA, NA)   
#>   UNKNOWN                                                1                    0                    0                   NaN                   1                    1                    1                                     (NA, NA)   
#> 
#> FAKE Name > -1.3 Flag
#>   N                                                     134                  62                   61                  0.98                  72                   52                  0.72                 0.04               (0, 0.22)  
#>   Y                                                     136                  70                   63                   0.9                  66                   56                  0.85                 0.62             (0.21, 1.73) 

# NOT RUN {
Viewer(tbl)
# }