Time-to-event Table as used for Forest Plot

The time-to-event forest plot table summarizes time-to-event data by groups. The function returns event counts and median survival time for each analysis arm, as well as a hazard ratio and the corresponding confidence interval from a Cox proportional hazard model.

t_forest_tte(
  tte,
  is_event,
  col_by,
  row_by_list = NULL,
  total = "ALL",
  strata_data = NULL,
  ties = "exact",
  time_unit = "month",
  conf_int = 0.95,
  dense_header = FALSE,
  table_tree = FALSE
)

Arguments

tte
(numeric vector)
Contains time to event data, currently no NA allowed

is_event
(logical value)
TRUE if event, FALSE if tte is censored

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

row_by_list
(list or data.frame) of factors
Contains one factor variable to calculate the t_el_forest_tte

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

strata_data
(character, factor or data.frame)
Used for stratification factors If NULL, no stratified analysis is performed.

ties
(character value) the method used for tie handling in coxph.

time_unit
(character value) The unit of median survival time. Default is "months".

conf_int
(numeric value) confidence level of the interval

dense_header
(logical value) Display the table headers in multiple rows.

table_tree
(logical value)
If FALSE then a single rtable gets returned, if TRUE a node object gets returned with the elementary rtable objects.

Details

Cox proportional hazard model is used for hazard ratio calculation

The returned table contains one row per analysis within a subgroup of data (indicated by the row name). The analysis is summarized with the following 9 columns:

1

Total n the total number of subjects included in analysis population

2-4

Survival statistics for reference arm, n and Events are the total number of patients and the number of events in reference arm, respectively. Median (month) is the survival time estimated by Kaplan-Meier method. Time unit can be modified per study needs.

5-7

same analysis as for reference arm now for comparison arm

8

Hazard Ratio ranges from 0 to infinity. The hazard ratio is an estimate of the ratio of the hazard rate in the comparison group versus that in the reference group. Univariate Cox proportional hazard model is applied to obtain the estimated hazard ratio. Hazard ratio > 1 implies better treatment effect in reference arm, and hazard ratio < 1 when comparison arm is better.

9

Wald CI The confidence interval indicates the level of uncertainty around the measure of effect (Hazard Ratio). Because only a small sample of the overall population is included in the analysis, by having an upper and lower confidence limit we can infer that the true treatment effect lies in between. By default, the 95 interval is calculated. If it includes 1, then we say that the difference between two arms is not significant at a significance level of 0.05.

See also

Examples

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)

ADTTE <- radtte(cached = TRUE)
ADTTE_f <- ADTTE %>%
  dplyr::filter(PARAMCD == "OS", ARMCD %in% c("ARM B", "ARM A")) %>%
  mutate(ARMCD = droplevels(ARMCD))

tbl <- t_forest_tte(
  tte = ADTTE_f$AVAL,
  is_event = ADTTE_f$CNSR == 0,
  col_by = ADTTE_f$ARMCD,
  row_by_list = droplevels(ADTTE_f[, c("SEX", "RACE")]),
  ties = "exact",
  conf_int = 0.9,
  dense_header = TRUE
)

#> Warning: Ran out of iterations and did not converge

tbl2 <- t_forest_tte(
  tte = ADTTE_f$AVAL,
  is_event = ADTTE_f$CNSR == 0,
  col_by = ADTTE_f$ARMCD,
  row_by_list = droplevels(ADTTE_f[,  "BMRKR2" ]),
  strata_data = ADTTE_f[ , c("STRATA1", "STRATA2")],
  ties = "exact",
  dense_header = TRUE
)

tbl2

#>                                                                 ARM A                                                       ARM B                                                                
#> Baseline Risk           Total                 n                 Events              Median                n                 Events              Median              Hazard               95%     
#> Factors                   n                                                        (month)                                                     (month)              Ratio*             Wald CI   
#> -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                      268                 134                  82                96.86                134                 100                92.31                1.1             (0.81, 1.48)
#> 
#> BMRKR2
#>   LOW                     95                  50                  30                96.86                 45                  32                90.66                0.95            (0.56, 1.62)
#>   MEDIUM                  93                  37                  23                90.68                 56                  44                71.71                0.97            (0.55, 1.7) 
#>   HIGH                    80                  47                  29                102.47                33                  24                100.84               1.19            (0.67, 2.1) 

if (FALSE) {
Viewer(tbl)
}

# table tree
tbls <- t_forest_tte(
  tte = ADTTE_f$AVAL,
  is_event = ADTTE_f$CNSR == 0,
  col_by = ADTTE_f$ARMCD,
  row_by_list = droplevels(ADTTE_f[, c("SEX", "RACE")]),
  ties = "exact",
  dense_header = TRUE,
  table_tree = TRUE
)

#> Warning: Ran out of iterations and did not converge

summary(tbls)

#>                                                        index            content             dim            hide label          format  
#> ---------------------------------------------------------------------------------------------------------------------------------------
#> root                                                                      NULL                                TRUE               1     
#>   ALL                                                    1               rtable            1 x 9              TRUE                     
#>   SEX                                                    2                NULL                               FALSE               0     
#>     F                                                   2, 1             rtable            1 x 9              TRUE               0     
#>     M                                                   2, 2             rtable            1 x 9              TRUE               0     
#>     U                                                   2, 3             rtable            1 x 9              TRUE               0     
#>     UNDIFFERENTIATED                                    2, 4             rtable            1 x 9              TRUE               0     
#>   RACE                                                   3                NULL                               FALSE               0     
#>     ASIAN                                               3, 1             rtable            1 x 9              TRUE               0     
#>     BLACK OR AFRICAN AMERICAN                           3, 2             rtable            1 x 9              TRUE               0     
#>     WHITE                                               3, 3             rtable            1 x 9              TRUE               0     
#>     AMERICAN INDIAN OR ALASKA NATIVE                    3, 4             rtable            1 x 9              TRUE               0     
#>     MULTIPLE                                            3, 5             rtable            1 x 9              TRUE               0     
#>     NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER           3, 6             rtable            1 x 9              TRUE               0     

to_rtable(tbls)

#>                                                                                                    ARM A                                                             ARM B                                                                       
#> Baseline Risk                                          Total                   n                   Events                Median                  n                   Events                Median                Hazard                 95%      
#> Factors                                                  n                                                              (month)                                                           (month)                Ratio                Wald CI    
#> -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                                                     268                   134                    82                  96.86                  134                   100                  92.31                  1.03              (0.77, 1.38) 
#> 
#> SEX
#>   F                                                     156                    79                    46                  101.43                  77                    61                  93.54                  1.11              (0.75, 1.63) 
#>   M                                                     106                    51                    34                  85.56                   55                    38                  91.98                  0.9               (0.56, 1.45) 
#>   U                                                      5                     3                     1                   124.94                  2                     1                   60.65                 >999.9            (0.00, >999.9)
#>   UNDIFFERENTIATED                                       1                     1                     1                   74.04                   0                     0                                                              (NA, NA)   
#> 
#> RACE
#>   ASIAN                                                 135                    68                    43                  100.86                  67                    49                  95.63                  0.96              (0.63, 1.46) 
#>   BLACK OR AFRICAN AMERICAN                              59                    31                    17                  125.29                  28                    21                  94.17                  1.26              (0.65, 2.41) 
#>   WHITE                                                  53                    27                    17                  54.09                   26                    19                  70.76                  0.87              (0.44, 1.72) 
#>   AMERICAN INDIAN OR ALASKA NATIVE                       19                    8                     5                   66.66                   11                    9                   101.28                 0.64              (0.2, 2.04)  
#>   MULTIPLE                                               1                     0                     0                                           1                     1                   90.66                                      (NA, NA)   
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER              1                     0                     0                                           1                     1                   113.43                                     (NA, NA)