Time-to-event Table as used for Forest Plot

The time-to-event forest plot table summarizes time-to-event data by groups. The function returns event counts and median survival time for each analysis arm, as well as a hazard ratio and the corresponding 95% confidence interval from a Cox proportional hazard model.

t_forest_tte(
  tte,
  is_event,
  col_by,
  row_by_list = NULL,
  total = "ALL",
  strata_data = NULL,
  ties = "exact",
  time_unit = "month",
  dense_header = FALSE,
  table_tree = FALSE
)

Arguments

tte
vector with numeric time to event data, currently no NA allowed

is_event
is boolean, TRUE if event, FALSE if tte is censored

col_by
logical factor or data.frame with reference and comparison group information, the first level indicates the reference group. See the functions factor_to_matrix_by and add_total_by.

row_by_list
list or data.frame with one factor variable to calculate the t_el_forest_tte

total
string of column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

strata_data
currently not supported

ties
the method used for tie handling in coxph.

time_unit
The unit of median survival time. Default is months.

dense_header
Display the table headers in multiple rows.

table_tree
logical, if FALSE then a single rtable gets returned, if TRUE a node object gets returned with the elementary rtable objects.

Details

Cox proportional hazard model is used for hazard ratio calculation

The returned table contains one row per analysis within a subgroup of data (indicated by the row name). The analysis is summarized with the following 9 columns:

1

Total n the total number of subjects included in analysis population

2-4

Survival statistics for reference arm, n and Events are the total number of patients and the number of events in reference arm, respectively. Median (month) is the survival time estimated by Kaplan-Meier method. Time unit can be modified per study needs.

5-7

same analysis as for reference arm now for comparison arm

8

Hazard Ratio ranges from 0 to infinity. The hazard ratio is an estimate of the ratio of the hazard rate in the comparison group versus that in the reference group. Univariate Cox proportional hazard model is applied to obtain the estimated hazard ratio. Hazard ratio > 1 implies better treatment effect in reference arm, and hazard ratio < 1 when comparison arm is better.

9

95% Wald CI The 95 around the measure of effect (Hazard Ratio). Because only a small sample of the overall population is included in the analysis, by having an upper and lower confidence limit we can infer that the true treatment effect lies in between. If the 95 includes 1, then we say that the difference between two arms is not significant at a significance level of 0.05.

See also

Examples

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(seed = 1)

ADTTE <- radtte(ADSL, seed = 2)
ADTTE_f <- ADTTE %>%
  dplyr::filter(PARAMCD == "OS", ARMCD %in% c("ARM B", "ARM A")) %>%
  mutate(ARMCD = droplevels(ARMCD))

tbl <- t_forest_tte(
  tte = ADTTE_f$AVAL,
  is_event = ADTTE_f$CNSR == 0,
  col_by = ADTTE_f$ARMCD,
  row_by_list = droplevels(ADTTE_f[, c("SEX", "RACE")]),
  ties = "exact",
  dense_header = TRUE
)

tbl

#>                                                                                         ARM A                                                          ARM B                                                                   
#> Baseline Risk                                 Total                  n                 Events               Median                  n                 Events               Median               Hazard                 95%     
#> Factors                                         n                                                           (month)                                                        (month)               Ratio               Wald CI   
#> -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                                            274                  133                  86                  86.71                 141                  120                 84.59                1.06              (0.8, 1.4)  
#> 
#> SEX
#>   F                                            147                  75                   49                  86.71                 72                   62                  93.15                1.03              (0.7, 1.5)  
#>   M                                            119                  54                   35                  82.8                  65                   54                  72.34                0.97             (0.63, 1.51) 
#>   U                                             5                    3                    1                                         2                    2                 222.76                2.12             (0.13, 35.36)
#>   UNDIFFERENTIATED                              3                    1                    1                  93.61                  2                    2                  70.78                0.71             (0.04, 11.79)
#> 
#> RACE
#>   ASIAN                                        156                  70                   40                  103.8                 86                   73                  78.96                1.38             (0.94, 2.04) 
#>   BLACK OR AFRICAN AMERICAN                    56                   28                   21                  57.72                 28                   21                 131.43                0.63             (0.34, 1.18) 
#>   WHITE                                        48                   26                   16                  63.05                 22                   21                  91.45                1.03             (0.52, 2.01) 
#>   AMERICAN INDIAN OR ALASKA NATIVE             12                    7                    7                  94.67                  5                    5                  66.63                0.95             (0.27, 3.27) 
#>   MULTIPLE                                      1                    1                    1                  0.69                   0                    0                                                          (NA, NA)   
#>   OTHER                                         1                    1                    1                  32.66                  0                    0                                                          (NA, NA)   

if (FALSE) {
Viewer(tbl)
}

# table tree
tbls <- t_forest_tte(
  tte = ADTTE_f$AVAL,
  is_event = ADTTE_f$CNSR == 0,
  col_by = ADTTE_f$ARMCD,
  row_by_list = droplevels(ADTTE_f[, c("SEX", "RACE")]),
  ties = "exact",
  dense_header = TRUE,
  table_tree = TRUE
)
summary(tbls)

#>                                               index            content             dim            hide label          format  
#> ------------------------------------------------------------------------------------------------------------------------------
#> root                                                             NULL                                TRUE               1     
#>   ALL                                           1               rtable            1 x 9              TRUE                     
#>   SEX                                           2                NULL                               FALSE               0     
#>     F                                          2, 1             rtable            1 x 9              TRUE               0     
#>     M                                          2, 2             rtable            1 x 9              TRUE               0     
#>     U                                          2, 3             rtable            1 x 9              TRUE               0     
#>     UNDIFFERENTIATED                           2, 4             rtable            1 x 9              TRUE               0     
#>   RACE                                          3                NULL                               FALSE               0     
#>     ASIAN                                      3, 1             rtable            1 x 9              TRUE               0     
#>     BLACK OR AFRICAN AMERICAN                  3, 2             rtable            1 x 9              TRUE               0     
#>     WHITE                                      3, 3             rtable            1 x 9              TRUE               0     
#>     AMERICAN INDIAN OR ALASKA NATIVE           3, 4             rtable            1 x 9              TRUE               0     
#>     MULTIPLE                                   3, 5             rtable            1 x 9              TRUE               0     
#>     OTHER                                      3, 6             rtable            1 x 9              TRUE               0     

to_rtable(tbls)

#>                                                                                         ARM A                                                          ARM B                                                                   
#> Baseline Risk                                 Total                  n                 Events               Median                  n                 Events               Median               Hazard                 95%     
#> Factors                                         n                                                           (month)                                                        (month)               Ratio               Wald CI   
#> -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                                            274                  133                  86                  86.71                 141                  120                 84.59                1.06              (0.8, 1.4)  
#> 
#> SEX
#>   F                                            147                  75                   49                  86.71                 72                   62                  93.15                1.03              (0.7, 1.5)  
#>   M                                            119                  54                   35                  82.8                  65                   54                  72.34                0.97             (0.63, 1.51) 
#>   U                                             5                    3                    1                                         2                    2                 222.76                2.12             (0.13, 35.36)
#>   UNDIFFERENTIATED                              3                    1                    1                  93.61                  2                    2                  70.78                0.71             (0.04, 11.79)
#> 
#> RACE
#>   ASIAN                                        156                  70                   40                  103.8                 86                   73                  78.96                1.38             (0.94, 2.04) 
#>   BLACK OR AFRICAN AMERICAN                    56                   28                   21                  57.72                 28                   21                 131.43                0.63             (0.34, 1.18) 
#>   WHITE                                        48                   26                   16                  63.05                 22                   21                  91.45                1.03             (0.52, 2.01) 
#>   AMERICAN INDIAN OR ALASKA NATIVE             12                    7                    7                  94.67                  5                    5                  66.63                0.95             (0.27, 3.27) 
#>   MULTIPLE                                      1                    1                    1                  0.69                   0                    0                                                          (NA, NA)   
#>   OTHER                                         1                    1                    1                  32.66                  0                    0                                                          (NA, NA)