Time-to-event Table as used for Forest Plot

The time-to-event forest plot table summarizes time-to-event data by groups. The function returns event counts and median survival time for each analysis arm, as well as a hazard ratio and the corresponding 95% confidence interval from a Cox proportional hazard model.

t_forest_tte(tte, is_event, col_by, group_data = NULL,
  strata_data = NULL, total = "ALL", time_unit = "month",
  ties = "exact", na_omit_group = TRUE, dense_header = FALSE)

Arguments

tte
vector with numeric time to event data, currently no NA allowed

is_event
is boolean, TRUE if event, FALSE if tte is censored

col_by
factor with reference and comparison group information, the first level indicates the reference group.

group_data
data frame with one variable per grouping

strata_data
data frame with stratification variables

total
character string that will be used as a label for a column with total population. If the levels of col_by are the only columns of interest then total should be NULL.

time_unit
The unit of median survival time. Default is months.

ties
the method used for tie handling in coxph.

na_omit_group
is boolean. Default is TRUE, do not display NA as a category.

dense_header
Display the table headers in multiple rows. Default is FALSE.

Details

Cox proportional hazard model is used for hazard ratio calculation

The returned table contains one row per analysis within a subgroup of data (indicated by the row name). The analysis is summarized with the following 9 columns:

1

Total n the total number of subjects included in analysis population

2-4

Survival statistics for reference arm, n and Events are the total number of patients and the number of events in reference arm, respectively. Median (month) is the survival time estimated by Kaplan-Meier method. Time unit can be modified per study needs.

5-7

same analysis as for reference arm now for comparison arm

8

Hazard Ratio ranges from 0 to infinity. The hazard ratio is an estimate of the ratio of the hazard rate in the comparison group versus that in the reference group. Univariate Cox proportional hazard model is applied to obtain the estimated hazard ratio. Hazard ratio > 1 implies better treatment effect in reference arm, and hazard ratio < 1 when comparison arm is better.

9

95% Wald CI The 95 around the measure of effect (Hazard Ratio). Because only a small sample of the overall population is included in the analysis, by having an upper and lower confidence limit we can infer that the true treatment effect lies in between. If the 95 includes 1, then we say that the difference between two arms is not significant at a significance level of 0.05.

See also

Examples

library(random.cdisc.data)
library(dplyr)

ADSL <- cadsl

ADTTE <- cadtte
ADTTE_f <- ADTTE %>%
  dplyr::filter(PARAMCD == "OS", ARMCD %in% c("ARM B", "ARM A")) %>%
  mutate(ARMCD = droplevels(ARMCD))

tbl <- t_forest_tte(
  tte = ADTTE_f$AVAL,
  is_event = ADTTE_f$CNSR == 0,
  col_by = ADTTE_f$ARMCD,
  group_data = droplevels(ADTTE_f[, c("SEX", "RACE")]),
  ties = "exact",
  dense_header = TRUE
)

tbl

#>                                                                                               ARM A                                                       ARM B                                                                
#> Baseline Risk                                         Total                 n                 Events              Median                n                 Events              Median              Hazard               95%     
#> Factors                                                 n                                                        (month)                                                     (month)              Ratio              Wald CI   
#> -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                                                    275                 142                  88                88.67                133                 105                 77.4                1.14            (0.86, 1.51)
#> 
#> SEX
#>   F                                                    136                  69                  44                88.67                 67                  52                79.99                1.1             (0.74, 1.66)
#>   M                                                    131                  68                  40                85.69                 63                  50                 77.4                1.15            (0.75, 1.74)
#>   U                                                     5                   2                   2                 19.75                 3                   3                 54.23                0.34            (0.03, 3.85)
#>   UNDIFFERENTIATED                                      3                   3                   2                 163.15                0                   0                                                        (NA, NA)  
#> 
#> RACE
#>   ASIAN                                                166                  92                  53                111.43                74                  61                78.79                1.43            (0.98, 2.07)
#>   BLACK OR AFRICAN AMERICAN                             56                  26                  16                65.27                 30                  23                 77.4                 1              (0.52, 1.94)
#>   WHITE                                                 37                  18                  13                73.27                 19                  12                104.76               0.61            (0.27, 1.39)
#>   AMERICAN INDIAN OR ALASKA NATIVE                      12                  4                   4                 12.32                 8                   7                 53.85                0.25            (0.05, 1.14)
#>   MULTIPLE                                              1                   0                   0                                       1                   1                  9.2                                   (NA, NA)  
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER             2                   2                   2                 90.81                 0                   0                                                        (NA, NA)  
#>   UNKNOWN                                               1                   0                   0                                       1                   1                  8.16                                  (NA, NA)  

# NOT RUN {
Viewer(tbl)
# }