Time-to-event Table as used for Forest Plot

The time-to-event forest plot table summarizes time-to-event data by groups. The function returns event counts and median survival time for each analysis arm, as well as a hazard ratio and the corresponding 95% confidence interval from a Cox proportional hazard model.

t_forest_tte(tte, is_event, col_by, group_data = NULL,
  strata_data = NULL, total = "ALL", time_unit = "month",
  ties = "exact", na.omit.group = TRUE, dense_header = FALSE)

Arguments

tte
a vector of time to event data

is_event
is boolean, TRUE if event, FALSE if tte is censored

col_by
factor with reference and comparison group information, the first level indicates the reference group

group_data
data frame with one variable per grouping

strata_data
data frame with stratification variables

total
character with the row name of the analysis run on all data. If NULL analysis is omitted.

time_unit
The unit of median survival time. Default is months.

ties
the method used for tie handling in coxph.

na.omit.group
is boolean. Default is TRUE, do not display NA's as a category.

dense_header
Display the table headers in mulitple rows. Default is FALSE.

Details

Cox propotionl hazard model is used for hazard ratio calculation

The returned table contains one row per analysis within a subgroup of data (indicated by the row name). The analysis is summarized with the following 9 columns:

1

Total n the total number of subjects included in analysis population

2-4

Survival statistics for reference arm, n and Events are the total number of patients and the number of events in reference arm, respectively. Median (month) is the survival time estimated by Kaplan-Meier method. Time unit can be modified per study needs.

5-7

same analysis as for reference arm now for comparison arm

8

Hazard Ratio ranges from 0 to infinity. The hazard ratio is an estimate of the ratio of the hazard rate in the comparison group versus that in the reference group. Univariate Cox proportional hazard model is applied to obtain the estimated hazard ratio. Hazard ratio > 1 implies better treatment effect in reference arm, and hazard ratio < 1 when comparion arm is better.

9

95% Wald CI The 95 around the measure of effect (Hazard Ratio). Because only a small sample of the overall population is included in the analysis, by having an upper and lower confidence limit we can infer that the true treatment effect lies in between. If the 95 includes 1, then we say that the difference between two arms is not significant at a significance level of 0.05.

See also

Examples

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(seed = 1)

ADTTE <- radtte(ADSL, seed = 2)
ADTTE_f <- filter(ADTTE, PARAMCD == "OS", ARMCD %in% c("ARM B", "ARM A")) %>%
  mutate(ARMCD = droplevels(ARMCD))

tbl <- t_forest_tte(
  tte = ADTTE_f$AVAL,
  is_event = ADTTE_f$CNSR == 0,
  col_by = ADTTE_f$ARMCD,
  group_data = droplevels(ADTTE_f[, c("SEX", "RACE")]),
  ties = "exact",
  dense_header = TRUE
)

tbl

#>                                                                                                  ARM A                                                          ARM B                                                                   
#> Baseline Risk                                          Total                  n                 Events               Median                  n                 Events               Median               Hazard                 95%     
#> Factors                                                  n                                                           (month)                                                        (month)               Ratio               Wald CI   
#> ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                                                     275                  142                  97                  79.52                 133                  112                 72.34                1.18              (0.9, 1.55) 
#> 
#> SEX
#>   F                                                     136                  69                   50                  84.28                 67                   58                  72.72                1.23              (0.83, 1.8) 
#>   M                                                     131                  68                   45                  60.7                  63                   52                  72.34                1.22             (0.82, 1.83) 
#>   U                                                      5                    2                    1                  71.55                  3                    2                  36.17                2.27             (0.19, 26.54)
#>   UNDIFFERENTIATED                                       3                    3                    1                                         0                    0                                                          (NA, NA)   
#> 
#> RACE
#>   ASIAN                                                 166                  92                   64                  70.33                 74                   63                  72.72                1.08             (0.76, 1.53) 
#>   BLACK OR AFRICAN AMERICAN                             56                   26                   17                  84.28                 30                   26                  86.22                1.43             (0.75, 2.71) 
#>   WHITE                                                 37                   18                   12                 112.19                 19                   15                  55.73                1.33              (0.61, 2.9) 
#>   AMERICAN INDIAN OR ALASKA NATIVE                      12                    4                    2                  98.66                  8                    6                  96.77                0.82             (0.15, 4.48) 
#>   MULTIPLE                                               1                    0                    0                                         1                    1                  35.03                                   (NA, NA)   
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER              2                    2                    2                 112.47                  0                    0                                                          (NA, NA)   
#>   UNKNOWN                                                1                    0                    0                                         1                    1                  70.21                                   (NA, NA)   

# NOT RUN {
Viewer(tbl)
# }