Time-to-event Table as used for Forest Plot

The time-to-event forest plot table summarizes time-to-event data by groups. The function returns event counts and median survival time for each analysis arm, as well as a hazard ratio and the corresponding 95% confidence interval from a Cox proportional hazard model.

t_forest_tte(tte, is_event, col_by, group_data = NULL,
  strata_data = NULL, total = "ALL", time_unit = "month",
  ties = "exact", na.omit.group = TRUE, dense_header = FALSE)

Arguments

tte
a vector of time to event data

is_event
is boolean, TRUE if event, FALSE if tte is censored

col_by
factor with reference and comparison group information, the first level indicates the reference group

group_data
data frame with one variable per grouping

strata_data
data frame with stratification variables

total
character with the row name of the analysis run on all data. If NULL analysis is omitted.

time_unit
The unit of median survival time. Default is months.

ties
the method used for tie handling in coxph.

na.omit.group
is boolean. Default is TRUE, do not display NA's as a category.

dense_header
Display the table headers in mulitple rows. Default is FALSE.

Details

Cox propotionl hazard model is used for hazard ratio calculation

The returned table contains one row per analysis within a subgroup of data (indicated by the row name). The analysis is summarized with the following 9 columns:

1

Total n the total number of subjects included in analysis population

2-4

Survival statistics for reference arm, n and Events are the total number of patients and the number of events in reference arm, respectively. Median (month) is the survival time estimated by Kaplan-Meier method. Time unit can be modified per study needs.

5-7

same analysis as for reference arm now for comparison arm

8

Hazard Ratio ranges from 0 to infinity. The hazard ratio is an estimate of the ratio of the hazard rate in the comparison group versus that in the reference group. Univariate Cox proportional hazard model is applied to obtain the estimated hazard ratio. Hazard ratio > 1 implies better treatment effect in reference arm, and hazard ratio < 1 when comparion arm is better.

9

95% Wald CI The 95 around the measure of effect (Hazard Ratio). Because only a small sample of the overall population is included in the analysis, by having an upper and lower confidence limit we can infer that the true treatment effect lies in between. If the 95 includes 1, then we say that the difference between two arms is not significant at a significance level of 0.05.

Examples

library(random.cdisc.data)

ADSL <- radsl()

ADTTE <- radtte(ADSL)
ADTTE_f <- subset(ADTTE, PARAMCD == "OS" & ARMCD %in% c("ARM B", "ARM A"))

tbl <- t_forest_tte(
  tte = ADTTE_f$AVAL,
  is_event = ADTTE_f$CNSR == 0,
  col_by = factor(ADTTE_f$ARM),
  group_data = as.data.frame(lapply(ADTTE_f[, c("SEX", "RACE")], as.factor)),
  ties = "exact",
  dense_header = TRUE
)

tbl

#>                                                                                             A: Drug X                                                   B: Placebo                                                             
#> Baseline Risk                                         Total                 n                 Events              Median                n                 Events              Median              Hazard               95%     
#> Factors                                                 n                                                        (month)                                                     (month)              Ratio              Wald CI   
#> -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                                                    266                 143                  87                111.99               123                 103                82.46                1.05            (0.78, 1.4) 
#> 
#> SEX
#>   F                                                    119                  66                  43                83.54                 53                  47                69.01                0.87            (0.56, 1.35)
#>   M                                                    141                  74                  44                115.11                67                  54                84.84                1.11            (0.74, 1.66)
#>   U                                                     3                   1                   0                                       2                   1                 13.98                                  (NA, NA)  
#>   UNDIFFERENTIATED                                      3                   2                   0                                       1                   1                 82.88                                  (NA, NA)  
#> 
#> RACE
#>   ASIAN                                                138                  75                  48                115.11                63                  53                 85.6                1.09            (0.73, 1.63)
#>   BLACK OR AFRICAN AMERICAN                             68                  40                  22                120.61                28                  24                78.51                1.09            (0.59, 2.02)
#>   WHITE                                                 39                  20                  14                53.56                 19                  15                69.01                0.87            (0.42, 1.8) 
#>   AMERICAN INDIAN OR ALASKA NATIVE                      14                  7                   3                 152.63                7                   6                 39.72                2.06            (0.49, 8.75)
#>   MULTIPLE                                              1                   0                   0                                       1                   0                                                        (NA, NA)  
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER             4                   1                   0                                       3                   3                 184.47                                 (NA, NA)  
#>   OTHER                                                 1                   0                   0                                       1                   1                 123.54                                 (NA, NA)  
#>   UNKNOWN                                               1                   0                   0                                       1                   1                 55.45                                  (NA, NA)  

# NOT RUN {
Viewer(tbl)
# }