Time-to-event Table as used for Forest Plot

The time-to-event forest plot table summarizes time-to-event data by groups. The function returns event counts and median survival time for each analysis arm, as well as a hazard ratio and the corresponding 95% confidence interval from a Cox proportional hazard model.

t_forest_tte(tte, is_event, col_by, group_data = NULL, strata_data = NULL,
  total = "ALL", time_unit = "month", ties = "exact",
  na.omit.group = TRUE, dense_header = FALSE)

Arguments

tte
a vector of time to event data

is_event
is boolean, TRUE if event, FALSE if tte is censored

col_by
factor with reference and comparison group information, the first level indicates the reference group

group_data
data frame with one variable per grouping

strata_data
data frame with stratification variables

total
character with the row name of the analysis run on all data. If NULL analysis is omitted.

time_unit
The unit of median survival time. Default is months.

ties
the method used for tie handling in coxph.

na.omit.group
is boolean. Default is TRUE, do not display NA's as a category.

dense_header
Display the table headers in mulitple rows. Default is FALSE.

Details

Cox propotionl hazard model is used for hazard ratio calculation

The returned table contains one row per analysis within a subgroup of data (indicated by the row name). The analysis is summarized with the following 9 columns:

1

Total n the total number of subjects included in analysis population

2-4

Survival statistics for reference arm, n and Events are the total number of patients and the number of events in reference arm, respectively. Median (month) is the survival time estimated by Kaplan-Meier method. Time unit can be modified per study needs.

5-7

same analysis as for reference arm now for comparison arm

8

Hazard Ratio ranges from 0 to infinity. The hazard ratio is an estimate of the ratio of the hazard rate in the comparison group versus that in the reference group. Univariate Cox proportional hazard model is applied to obtain the estimated hazard ratio. Hazard ratio > 1 implies better treatment effect in reference arm, and hazard ratio < 1 when comparion arm is better.

9

95% Wald CI The 95 around the measure of effect (Hazard Ratio). Because only a small sample of the overall population is included in the analysis, by having an upper and lower confidence limit we can infer that the true treatment effect lies in between. If the 95 includes 1, then we say that the difference between two arms is not significant at a significance level of 0.05.

Examples

library(random.cdisc.data)

ASL <- radam("ASL")
ATE <- radam("ATE", ADSL = ASL)

ATE_f <- subset(ATE, PARAMCD == "OS")

ANL <- merge(ASL, ATE_f)

tbl <- t_forest_tte(
  tte = ANL$AVAL,
  is_event = ANL$CNSR == 0,
  col_by = factor(ANL$ARM),
  group_data = as.data.frame(lapply(ANL[, c("SEX", "RACE")], as.factor)),
  ties = "exact",
  dense_header = TRUE
)

tbl

#>                                                                                                  ARM A                                                          ARM B                                                                   
#> Baseline Risk                                          Total                  n                 Events               Median                  n                 Events               Median               Hazard                 95%     
#> Factors                                                  n                                                           (month)                                                        (month)               Ratio               Wald CI   
#> ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> ALL                                                     100                  53                   31                  83.32                 47                   28                  89.74                1.25              (0.74, 2.1) 
#> 
#> SEX
#>   F                                                     50                   29                   18                  74.84                 21                   12                 101.35                1.08             (0.51, 2.28) 
#>   M                                                     50                   24                   13                 181.86                 26                   16                  74.18                1.49              (0.7, 3.16) 
#> 
#> RACE
#>   AMERICAN INDIAN OR ALASKA NATIVE                      19                   13                   12                  46.3                   6                    3                 119.69                0.44             (0.12, 1.59) 
#>   ASIAN                                                 23                    6                    3                  73.22                 17                   12                  74.18                0.56             (0.14, 2.29) 
#>   BLACK OR AFRICAN AMERICAN                             17                   11                    9                  79.04                  6                    2                                       0.52             (0.11, 2.44) 
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER             18                    8                    2                                        10                    5                 101.35                1.58              (0.3, 8.29) 
#>   WHITE                                                 23                   15                    5                  311.2                  8                    6                  105.1                 5.1             (1.38, 18.89)

# NOT RUN {
Viewer(tbl)
# }