Response Table Caclulation Function

Response Table Caclulation Function

t_rsp(rsp, col_by, parition_rsp_by = NULL, strata_data = NULL)

Arguments

rsp
boolean, TRUE if subject is a responder FALSE otherwise

col_by
factor with reference and comparison group information, the first level indicates the reference group

parition_rsp_by
factor to partition rsp and run additional analysis

strata_data
data for stratification factors (categorical variables), if NULL no stratified analysis is performed

Value

an rtable object

Details

We use the epiR package to compute 2-by-2 contingency table statistics and analysis with stratification factors.

If no stratification, chi-squared test is performed by default to test for equality of proportions in response rate between reference and comparison arms.

If stratification factors are added, Cochran-Mantel-Haensel test statistics are reported for the test of equal proportions, odds ratio and its corresponding 95

Examples

library(random.cdisc.data)

ASL <- radam("ASL", start_with = list(RACE = c("white", "asian")))
ARS <- radam("ARS", ADSL = ASL)

ASL$RACE <- factor(ASL$RACE)

ANL <- merge(ASL, subset(ARS, PARAMCD == "OVRSPI"))

tbl <- t_rsp(
 rsp = ANL$AVALC %in% c("CR", "PR"),
 col_by = relevel(factor(ANL$ARMCD), "ARM B", "ARM A"),
 parition_rsp_by = relevel(factor(ANL$AVALC), "CR", "PR", "SD", "NON CR/PD", "PD"),
 strata_data = ANL[, c("SEX", "RACE")]
)

tbl

#>                                                                       ARM B                  ARM A     
#>                                                                      (N=45)                 (N=55)     
#> -------------------------------------------------------------------------------------------------------
#> Responders                                                          18 (40%)              19 (34.55%)  
#> Non-Responders                                                      27 (60%)              36 (65.45%)  
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                       (25.7, 55.67)         (22.24, 48.58) 
#> 
#> Difference in Response Rates (x100)                                                          -5.45     
#>   95% CI for difference (Wald without correction)                                       (-26.01, 14.42)
#>   95% CI for difference (Wald with continuity correction)                                (-0.28, 0.17) 
#>   p-value (Cochran-Mantel-Haenszel)                                                         0.5443     
#> 
#> Odds Ratio                                                                                   0.78      
#>   95% CI                                                                                (0.344, 1.753) 
#> 
#> CR                                                                 11 (24.44%)            7 (12.73%)   
#>   95% CI (Wald)                                                  (12.88, 39.54)          (5.27, 24.48) 
#> 
#> NE                                                                 11 (24.44%)            10 (18.18%)  
#>   95% CI (Wald)                                                  (12.88, 39.54)          (9.08, 30.9)  
#> 
#> PD                                                                 8 (17.78%)             15 (27.27%)  
#>   95% CI (Wald)                                                    (8, 32.05)           (16.14, 40.96) 
#> 
#> PR                                                                 7 (15.56%)             12 (21.82%)  
#>   95% CI (Wald)                                                   (6.49, 29.46)         (11.81, 35.01) 
#> 
#> SD                                                                 8 (17.78%)              11 (20%)    
#>   95% CI (Wald)                                                    (8, 32.05)           (10.43, 32.97)