Response Table Caclulation Function

Response Table Caclulation Function

t_rsp(rsp, col_by, parition_rsp_by = NULL, strata_data = NULL)

Arguments

rsp
boolean, TRUE if subject is a responder FALSE otherwise

col_by
factor with reference and comparison group information, the first level indicates the reference group

parition_rsp_by
factor to partition rsp and run additional analysis

strata_data
data for stratification factors (categorical variables), if NULL no stratified analysis is performed

Value

an rtable object

Details

We use the epiR package to compute 2-by-2 contingency table statistics and analysis with stratification factors.

If no stratification, chi-squared test is performed by default to test for equality of proportions in response rate between reference and comparison arms.

If stratification factors are added, Cochran-Mantel-Haensel test statistics are reported for the test of equal proportions, odds ratio and its corresponding 95

Examples

library(random.cdisc.data)

ASL <- radam("ASL")
ARS <- radam("ARS", ADSL = ASL)

ANL <- merge(ASL, subset(ARS, PARAMCD == "OVRSPI"))

tbl <- t_rsp(
 rsp = ANL$AVALC %in% c("CR", "PR"),
 col_by = relevel(factor(ANL$ARMCD), "ARM B", "ARM A"),
 parition_rsp_by = factor(ANL$AVALC, levels =  c("CR", "PR", "SD", "NON CR/PD", "PD")),
 strata_data = NULL
)

tbl

#>                                                                      ARM B                 ARM A     
#> -----------------------------------------------------------------------------------------------------
#> Responders                                                        16 (35.56%)           23 (41.82%)  
#> Non-Responders                                                    29 (64.44%)           32 (58.18%)  
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                      (21.87, 51.22)        (28.65, 55.89)
#> 
#> Difference in Response Rates                                                                0.06     
#>   95% CI for difference (Wald without correction)                                        add logic   
#>   95% CI for difference (Wald with continuity correction)                                add logic   
#>   p-value (Chi-squared)                                                                  add logic   
#> 
#> odds ratio table
#> 
#> CR                                                                 10 (100%)             14 (100%)   
#>   95% CI (Wald)                                                    to be done            to be done  
#> 
#> PR                                                                  6 (100%)              9 (100%)   
#>   95% CI (Wald)                                                    to be done            to be done  
#> 
#> SD                                                                   0 (0%)                0 (0%)    
#>   95% CI (Wald)                                                    to be done            to be done  
#> 
#> NON CR/PD                                                           0 (NaN%)              0 (NaN%)   
#>   95% CI (Wald)                                                    to be done            to be done  
#> 
#> PD                                                                   0 (0%)                0 (0%)    
#>   95% CI (Wald)                                                    to be done            to be done