Response Table

The response table function summarizes response data by groups. The function produces frequency counts and rates for responders and each response categories, as well as conducts comparisons between groups' response rates and odds ratio.

t_rsp(
  rsp,
  col_by,
  partition_rsp_by = NULL,
  strata_data = NULL,
  table_tree = FALSE
)

Arguments

rsp
(logical vector)
TRUE if subject is a responder FALSE otherwise

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

partition_rsp_by
(factor vector) Contains one or more response categories, generate additional statistics partitioned by each response categories. If NULL, tabulation by each response categories will not be performed.

strata_data
(character, factor or data.frame)
Used for stratification factors If NULL, no stratified analysis is performed. See details for further explanation.

table_tree
(logical value)
If FALSE then a single rtable gets returned, if TRUE a node object gets returned with the elementary rtable objects.

Value

an rtable object

Details

For the test of difference in response rates, Wald confidence interval with and without continuity correction are both reported. If no stratification factors are specified, Chi-squared test is performed by default to test for equality of proportions in response rate between reference and comparison groups. If stratification factors are added, Cochran-Mantel-Haensel test is performed instead. Statistics from the stratification-adjusted test will be reported for the p-value of test of equal proportions, odds ratio and its corresponding 95% confidence interval.

The display order of response categories in partitioned statistics section inherits the factor level order of partition_rsp_by. Use factor and its levels argument to include or exclude response categories and arrange display order. If response values contains missing or "Non Evaluable (NE)", 95% confidence interval will not be calculated.

See also

Examples

set.seed(1)
t_rsp(rsp = sample(c(TRUE, FALSE), 200, TRUE), col_by = factor(rep(c("A", "B"), each = 100)))

#>                                                                         A                      B       
#> -------------------------------------------------------------------------------------------------------
#> Responders                                                          49 (49%)               53 (53%)    
#> Non-Responders                                                      51 (51%)               47 (47%)    
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                       (38.86, 59.2)         (42.76, 63.06) 
#> 
#> Difference in Response Rates (%)                                                               4       
#>   95% CI for difference (Wald without correction)                                       (-9.85, 17.85) 
#>   95% CI for difference (Wald with continuity correction)                               (-10.85, 18.85)
#>   p-value (Chi-squared)                                                                     0.5715     
#> 
#> Odds Ratio                                                                                   1.17      
#>   95% CI                                                                                 (0.67, 2.04)  

library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)

ADRS <- radrs(cached = TRUE)
ADRS_f <- subset(ADRS, PARAMCD == "BESRSPI")

# Example 1 - ARM B as reference
#    "NON CR/PD" response category dropped from partition section since no observations
#     model with no stratifiaction factors, Chi-square test is performed
tbl <- t_rsp(
  rsp = ADRS_f$AVALC %in% c("CR", "PR"),
  col_by = relevel(factor(ADRS_f$ARMCD), "ARM B"),
  partition_rsp_by = droplevels(factor(
    ADRS_f$AVALC, levels = c("CR", "PR", "SD", "NON CR/PD", "PD", "NE")
  ))
)

#> Warning: Chi-squared approximation may be incorrect

#> Warning: Chi-squared approximation may be incorrect

#> Warning: Chi-squared approximation may be incorrect

#> Warning: Chi-squared approximation may be incorrect

#> Warning: Chi-squared approximation may be incorrect

#> Warning: Chi-squared approximation may be incorrect

tbl

#>                                                                      ARM B                 ARM A                 ARM C     
#> ---------------------------------------------------------------------------------------------------------------------------
#> Responders                                                        127 (94.78%)          133 (99.25%)          131 (99.24%) 
#> Non-Responders                                                     7 (5.22%)             1 (0.75%)             1 (0.76%)   
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                      (89.53, 97.87)        (95.91, 99.98)        (95.85, 99.98)
#> 
#> Difference in Response Rates (%)                                                            4.48                  4.47     
#>   95% CI for difference (Wald without correction)                                       (0.44, 8.52)          (0.42, 8.51) 
#>   95% CI for difference (Wald with continuity correction)                              (-0.31, 9.26)         (-0.33, 9.27) 
#>   p-value (Chi-squared)                                                                    0.0313                0.033     
#> 
#> Odds Ratio                                                                                  7.33                  7.22     
#>   95% CI                                                                               (0.89, 60.43)         (0.88, 59.52) 
#> 
#> Complete Response (CR)                                            97 (72.39%)           119 (88.81%)          120 (90.91%) 
#>   95% CI                                                          (64, 79.76)          (82.21, 93.6)         (84.66, 95.21)
#> 
#> Partial Response (PR)                                             30 (22.39%)           14 (10.45%)            11 (8.33%)  
#>   95% CI                                                         (15.64, 30.39)        (5.83, 16.91)         (4.23, 14.42) 
#> 
#> Stable Disease (SD)                                                7 (5.22%)             1 (0.75%)             1 (0.76%)   
#>   95% CI                                                         (2.13, 10.47)          (0.02, 4.09)          (0.02, 4.15) 

if (FALSE) {
Viewer(tbl)
}

if (FALSE) {
# Example 2 - ARM B as reference, ARM C displayed before ARM A
#    "NON CR/PD" response category displayed in partition section, "NE" responses
#     are not displayed model with two stratifiaction factors, CMH test performed
tbl2 <- t_rsp(
 rsp = ADRS_f$AVALC %in% c("CR", "PR"),
 col_by = factor(ADRS_f$ARMCD, c("ARM B", "ARM C", "ARM A")),
 partition_rsp_by = factor(ADRS_f$AVALC, levels = c("CR", "PR", "SD", "NON CR/PD", "PD")),
 strata_data = ADRS_f[c("RACE")]
)

tbl2
}

if (FALSE) {
Viewer(tbl2)
}

# Example 3 - when all observations are non-responders
ADRS <- data.frame(
  rsp = FALSE,
  arm = rep(c("A", "B"), each = 200),
  stringsAsFactors = FALSE
)

t_rsp(rsp = ADRS$rsp, col_by = factor(ADRS$arm))

#>                                                                      A                 B     
#> ---------------------------------------------------------------------------------------------
#> Responders                                                         0 (0%)            0 (0%)  
#> Non-Responders                                                   200 (100%)        200 (100%)
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                      (0, 1.83)         (0, 1.83) 
#> 
#> Difference in Response Rates (%)                                                       0     
#>   95% CI for difference (Wald without correction)                                      -     
#>   95% CI for difference (Wald with continuity correction)                              -     
#>   p-value (Chi-squared)                                                                -     
#> 
#> Odds Ratio                                                                             -     
#>   95% CI                                                                               -     

# table_tree
tbls <- t_rsp(rsp = ADRS$rsp, col_by = factor(ADRS$arm), table_tree = TRUE)
summary(tbls)

#>                             index            content             dim            hide label          format  
#> ------------------------------------------------------------------------------------------------------------
#> root                                           NULL                                TRUE                     
#>   Responders                  1               rtable            2 x 2              TRUE                     
#>   Clopper Pearson             2               rtable            1 x 2              TRUE                     
#>   Difference Tests            3               rtable            4 x 2              TRUE                     
#>   Odds Ratio                  4               rtable            2 x 2              TRUE                     

# Example 4 - Single Arm study
t_rsp(
   rsp = ADRS_f$AVALC %in% c("CR", "PR"),
   col_by = factor(rep("Single ARM", nrow(ADRS_f))),
   partition_rsp_by = droplevels(factor(
     ADRS_f$AVALC, levels = c("CR", "PR", "MR", "SD", "PD")
  ))
)

#>                                                      Single ARM  
#> -----------------------------------------------------------------
#> Responders                                          391 (97.75%) 
#> Non-Responders                                       9 (2.25%)   
#> 
#> 95% CI for Response Rates (Clopper-Pearson)        (95.77, 98.97)
#> 
#> Complete Response (CR)                               336 (84%)   
#>   95% CI                                           (80.03, 87.45)
#> 
#> Partial Response (PR)                               55 (13.75%)  
#>   95% CI                                           (10.53, 17.52)
#> 
#> Stable Disease (SD)                                  9 (2.25%)   
#>   95% CI                                            (1.03, 4.23) 

# Example 5 - Stratified

t_rsp(
rsp = ADRS_f$AVALC %in% c("CR", "PR"),
col_by = relevel(factor(ADRS_f$ARMCD), "ARM B"),
strata_data = ADSL[,c("STRATA2", "STRATA1")]
)

#> Warning: Chi-squared approximation may be incorrect

#> Warning: Chi-squared approximation may be incorrect

#> Warning: Chi-squared approximation may be incorrect

#> Warning: Chi-squared approximation may be incorrect

#>                                                                      ARM B                 ARM A                 ARM C     
#> ---------------------------------------------------------------------------------------------------------------------------
#> Responders                                                        127 (94.78%)          133 (99.25%)          131 (99.24%) 
#> Non-Responders                                                     7 (5.22%)             1 (0.75%)             1 (0.76%)   
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                      (89.53, 97.87)        (95.91, 99.98)        (95.85, 99.98)
#> 
#> Difference in Response Rates (%)                                                            4.48                  4.47     
#>   95% CI for difference (Wald without correction)                                       (0.44, 8.52)          (0.42, 8.51) 
#>   95% CI for difference (Wald with continuity correction)                              (-0.31, 9.26)         (-0.33, 9.27) 
#>   p-value (Cochran-Mantel-Haenszel)*                                                       0.0458                0.0277    
#> 
#> Odds Ratio*                                                                                 5.96                  8.98     
#>   95% CI                                                                               (0.77, 46.36)         (0.96, 84.28) 
#> 
#> * Model stratified by STRATA2 and STRATA1