Response Table

The response table function summarizes response data by groups. The function produces frequency counts and rates for responders and each response categories, as well as conducts comparisons between groups' response rates and odds ratio.

t_rsp(rsp, col_by, partition_rsp_by = NULL, strata_data = NULL)

Arguments

rsp
boolean, TRUE if subject is a responder FALSE otherwise

col_by
factor with reference and comparison group information, the first level indicates the reference group.

partition_rsp_by
factor with one or more response categories, generate additional statistics partitioned by each response categories. If NULL, tabulation by each response categories will not be performed.

strata_data
data for stratification factors (categorical variables). If NULL, no stratified analysis is performed. See details for further explanation.

Value

an rtable object

Details

For the test of difference in response rates, Wald confidence interval with and without continuity correction are both reported. If no stratification factors are specified, Chi-squared test is performed by default to test for equality of proportions in response rate between reference and comparison groups. If stratification factors are added, Cochran-Mantel-Haensel test is performed instead. Statistics from the stratification-adjusted test will be reported for the p-value of test of equal proportions, odds ratio and its corresponding 95% confidence interval.

The display order of response categories in partitioned statistics section inherits the factor level order of partition_rsp_by. Use factor and its levels argument to include or exclude response categories and arrange display order. If response values contains missing or "Non Evaluable (NE)", 95% confidence interval will not be calculated.

See also

Examples

t_rsp(rsp = sample(c(TRUE, FALSE), 200, TRUE), col_by = factor(rep(c("A", "B"), each = 100)))

#>                                                                        A                     B       
#>                                                                     (N=100)               (N=100)    
#> -----------------------------------------------------------------------------------------------------
#> Responders                                                          45 (45%)              59 (59%)   
#> Non-Responders                                                      55 (55%)              41 (41%)   
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                      (35.03, 55.27)        (48.71, 68.74)
#> 
#> Difference in Response Rates (%)                                                             14      
#>   95% CI for difference (Wald without correction)                                      (0.29, 27.71) 
#>   95% CI for difference (Wald with continuity correction)                              (-0.71, 28.71)
#>   p-value (Chi-squared)                                                                    0.0475    
#> 
#> Odds Ratio                                                                                  1.76     
#>   95% CI                                                                                 (1, 3.08)   

library(random.cdisc.data)
ADSL <- cadsl

ADRS <- cadrs
ADRS_f <- subset(ADRS, PARAMCD == "BESRSPI")

# Example 1 - ARM B as reference
#    "NON CR/PD" response category dropped from partition section since no observations
#     model with no stratifiaction factors, Chi-square test is performed
tbl <- t_rsp(
  rsp = ADRS_f$AVALC %in% c("CR", "PR"),
  col_by = relevel(factor(ADRS_f$ARMCD), "ARM B"),
  partition_rsp_by = droplevels(factor(
    ADRS_f$AVALC, levels = c("CR", "PR", "SD", "NON CR/PD", "PD", "NE")
  ))
)

tbl

#>                                                                      ARM B                 ARM A                 ARM C     
#>                                                                     (N=133)               (N=142)               (N=125)    
#> ---------------------------------------------------------------------------------------------------------------------------
#> Responders                                                        87 (65.41%)           136 (95.77%)          124 (99.2%)  
#> Non-Responders                                                    46 (34.59%)            6 (4.23%)              1 (0.8%)   
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                      (56.68, 73.44)        (91.03, 98.43)        (95.62, 99.98)
#> 
#> Difference in Response Rates (%)                                                           30.36                 33.79     
#>   95% CI for difference (Wald without correction)                                      (21.63, 39.1)         (25.55, 42.02)
#>   95% CI for difference (Wald with continuity correction)                              (20.9, 39.82)         (24.78, 42.8) 
#>   p-value (Chi-squared)                                                                      0                     0       
#> 
#> Odds Ratio                                                                                 11.98                 65.56     
#>   95% CI                                                                               (4.91, 29.25)         (8.87, 484.46)
#> 
#> Complete Response (CR)                                            65 (48.87%)           107 (75.35%)          107 (85.6%)  
#>   95% CI                                                         (40.11, 57.68)        (67.42, 82.19)        (78.2, 91.24) 
#> 
#> Partial Response (PR)                                             22 (16.54%)           29 (20.42%)            17 (13.6%)  
#>   95% CI                                                         (10.67, 23.97)         (14.12, 28)          (8.13, 20.88) 
#> 
#> Stable Disease (SD)                                               46 (34.59%)            6 (4.23%)              1 (0.8%)   
#>   95% CI                                                         (26.56, 43.32)         (1.57, 8.97)          (0.02, 4.38) 

# NOT RUN {
Viewer(tbl)
# }

# Example 2 - ARM B as reference, ARM C displayed before ARM A
#    "NON CR/PD" response category displayed in partition section, "NE" responses
#     are not displayed model with two stratifiaction factors, CMH test performed
tbl2 <- t_rsp(
 rsp = ADRS_f$AVALC %in% c("CR", "PR"),
 col_by = factor(ADRS_f$ARMCD, c("ARM B", "ARM C", "ARM A")),
 partition_rsp_by = factor(ADRS_f$AVALC, levels = c("CR", "PR", "SD", "NON CR/PD", "PD")),
 strata_data = ADRS_f[c("RACE")]
)

tbl2

#>                                                                      ARM B                 ARM C                 ARM A     
#>                                                                     (N=133)               (N=125)               (N=142)    
#> ---------------------------------------------------------------------------------------------------------------------------
#> Responders                                                        87 (65.41%)           124 (99.2%)           136 (95.77%) 
#> Non-Responders                                                    46 (34.59%)             1 (0.8%)             6 (4.23%)   
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                      (56.68, 73.44)        (95.62, 99.98)        (91.03, 98.43)
#> 
#> Difference in Response Rates (%)                                                           33.79                 30.36     
#>   95% CI for difference (Wald without correction)                                      (25.55, 42.02)        (21.63, 39.1) 
#>   95% CI for difference (Wald with continuity correction)                              (24.78, 42.8)         (20.9, 39.82) 
#>   p-value (Cochran-Mantel-Haenszel)*                                                   <5 data points        <5 data points
#> 
#> Odds Ratio*                                                                            <5 data points        <5 data points
#>   95% CI                                                                               <5 data points        <5 data points
#> 
#> Complete Response (CR)                                            65 (48.87%)           107 (85.6%)           107 (75.35%) 
#>   95% CI                                                         (40.11, 57.68)        (78.2, 91.24)         (67.42, 82.19)
#> 
#> Partial Response (PR)                                             22 (16.54%)            17 (13.6%)           29 (20.42%)  
#>   95% CI                                                         (10.67, 23.97)        (8.13, 20.88)          (14.12, 28)  
#> 
#> Stable Disease (SD)                                               46 (34.59%)             1 (0.8%)             6 (4.23%)   
#>   95% CI                                                         (26.56, 43.32)         (0.02, 4.38)          (1.57, 8.97) 
#> 
#> Non-CR or Non-PD (NON CR/PD)                                           0                     0                     0       
#>   95% CI                                                               -                     -                     -       
#> 
#> Progressive Disease (PD)                                               0                     0                     0       
#>   95% CI                                                               -                     -                     -       
#> 
#> * Model stratified by RACE

# NOT RUN {
Viewer(tbl2)
# }

# Example 3 - when all observations are non-responders
ADRS <- data.frame(
rsp = FALSE,
arm = rep(c("A", "B"), each = 200),
stringsAsFactors = FALSE
)

t_rsp(rsp = ADRS$rsp, col_by = factor(ADRS$arm))

#>                                                                      A                 B     
#>                                                                   (N=200)           (N=200)  
#> ---------------------------------------------------------------------------------------------
#> Responders                                                         0 (0%)            0 (0%)  
#> Non-Responders                                                   200 (100%)        200 (100%)
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                      (0, 1.83)         (0, 1.83) 
#> 
#> Difference in Response Rates (%)                                                       0     
#>   95% CI for difference (Wald without correction)                                      -     
#>   95% CI for difference (Wald with continuity correction)                              -     
#>   p-value (Chi-squared)                                                                -     
#> 
#> Odds Ratio                                                                             -     
#>   95% CI                                                                               -