Response Table

The response table function summarizes response data by groups. The function produces frequency counts and rates for responders and each response categories, as well as conducts comparisons between groups' response rates and odds ratio.

t_rsp(rsp, col_by, partition_rsp_by = NULL, strata_data = NULL)

Arguments

rsp
boolean, TRUE if subject is a responder FALSE otherwise

col_by
factor with reference and comparison group information, the first level indicates the reference group

partition_rsp_by
factor with one or more response categories, generate additional statistics partitioned by each response categories. If NULL, tabulation by each response categories will not be performed.

strata_data
data for stratification factors (categorical variables). If NULL, no stratified analysis is performed. See details for further explanation.

Value

an rtable object

Details

For the test of difference in response rates, Wald confidence interval with and without continuity correction are both reported. If no stratification factors are specified, Chi-squared test is performed by default to test for equality of proportions in response rate between reference and comparison groups. If stratification factors are added, Cochran-Mantel-Haensel test is performed instead. Statistics from the stratificaiton-adjusted test will be reported for the p-value of test of equal proportions, odds ratio and its corresponding 95% confidence interval.

The display order of response categories in partitioned statistics section inherits the factor level order of partition_rsp_by. Use factor and its levels argument to include or exclude response categories and arrange display order. If response values contains missing or "Not Evaluable (NE)", 95% confidence interval will not be calculated.

Examples

t_rsp(rsp = sample(c(TRUE, FALSE), 200, TRUE), col_by = factor(rep(c("A", "B"), each = 100)))

#> Error in check_col_by(col_by, min_num_levels = 2): argument "col_N" is missing, with no default

library(random.cdisc.data)
ADSL <- radsl(N = 1000)

ADRS <- radrs(ADSL)
ADRS_f <- subset(ADRS, PARAMCD == "BESRSPI")

# Example 1 - ARM B as reference
#    "NON CR/PD" response category dropped from partition section since no observations
#     model with no stratifiaction factors, Chi-square test is performed
tbl <- t_rsp(
 rsp = ADRS_f$AVALC %in% c("CR", "PR"),
 col_by = relevel(factor(ADRS_f$ARMCD), "ARM B"),
 partition_rsp_by = droplevels(factor(
   ADRS_f$AVALC, levels = c("CR", "PR", "SD", "NON CR/PD", "PD", "NE")
 ))
)

#> Error in check_col_by(col_by, min_num_levels = 2): argument "col_N" is missing, with no default

tbl

#> function (src, ...) 
#> {
#>     UseMethod("tbl")
#> }
#> <bytecode: 0x000000000ac73b58>
#> <environment: namespace:dplyr>
# NOT RUN {
Viewer(tbl)
# }

# Example 2 - ARM B as reference, ARM C displayed before ARM A
#    "NON CR/PD" response category displayed in partition section, "NE" responses
#     are not displayed model with two stratifiaction factors, CMH test performed
tbl2 <- t_rsp(
 rsp = ADRS_f$AVALC %in% c("CR", "PR"),
 col_by = factor(ADRS_f$ARMCD, c("ARM B", "ARM C", "ARM A")),
 partition_rsp_by = factor(ADRS_f$AVALC, levels = c("CR", "PR", "SD", "NON CR/PD", "PD")),
 strata_data = ADRS_f[c("RACE")]
)

#> Error in check_col_by(col_by, min_num_levels = 2): argument "col_N" is missing, with no default

tbl2

#> Error in eval(expr, envir, enclos): object 'tbl2' not found
# NOT RUN {
Viewer(tbl2)
# }

# Example 3 - when all observations are non-responders
ADRS <- data.frame(
rsp = FALSE,
arm = rep(c("A", "B"), each = 200)
)

t_rsp(rsp = ADRS$rsp, col_by = factor(ADRS$arm))

#> Error in check_col_by(col_by, min_num_levels = 2): argument "col_N" is missing, with no default