Response Table

The response table function summarizes response data by groups. The function produces frequency counts and rates for responders and each response categories, as well as conducts comparisons between groups' response rates and odds ratio.

t_rsp(rsp, col_by, partition_rsp_by = NULL, strata_data = NULL)

Arguments

rsp
boolean, TRUE if subject is a responder FALSE otherwise

col_by
factor with reference and comparison group information, the first level indicates the reference group

partition_rsp_by
factor with one or more response categories, generate additional statistics partitioned by each response categories. If NULL, tabulation by each response categories will not be performed.

strata_data
data for stratification factors (categorical variables). If NULL, no stratified analysis is performed. See details for further explanation.

Value

an rtable object

Details

For the test of difference in response rates, Wald confidence interval with and without continuity correction are both reported. If no stratification factors are specified, Chi-squared test is performed by default to test for equality of proportions in response rate between reference and comparison groups. If stratification factors are added, Cochran-Mantel-Haensel test is performed instead. Statistics from the stratificaiton-adjusted test will be reported for the p-value of test of equal proportions, odds ratio and its corresponding 95% confidence interval.

The display order of response categories in partitioned statistics section inherits the factor level order of partition_rsp_by. Use factor and its levels argument to include or exclude response categories and arrange display order. If response values contains missing or "Not Evaluable (NE)", 95% confidence interval will not be calculated.

Examples

t_rsp(rsp = sample(c(TRUE, FALSE), 200, TRUE), col_by = factor(rep(c("A", "B"), each = 100)))

#>                                                                         A                      B       
#>                                                                      (N=100)                (N=100)    
#> -------------------------------------------------------------------------------------------------------
#> Responders                                                          56 (56%)               40 (40%)    
#> Non-Responders                                                      44 (44%)               60 (60%)    
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                      (45.72, 65.92)         (30.33, 50.28) 
#> 
#> Difference in Response Rates (%)                                                              -16      
#>   95% CI for difference (Wald without correction)                                       (-29.67, -2.33)
#>   95% CI for difference (Wald with continuity correction)                               (-30.67, -1.33)
#>   p-value (Chi-squared)                                                                     0.0235     
#> 
#> Odds Ratio                                                                                   0.52      
#>   95% CI                                                                                  (0.3, 0.92)  

library(random.cdisc.data)

ASL <- radam("ASL", arm_choices = c("ARM A", "ARM B", "ARM C"), N = 1000,
   start_with = list(RACE = c("white", "asian")))
ARS <- radam("ARS", ADSL = ASL)
ANL <- merge(ASL, subset(ARS, PARAMCD == "OVRSPI"))

# Example 1 - ARM B as reference
#    "NON CR/PD" response category dropped from partition section since no observations
#     model with no stratifiaction factors, Chi-square test is performed
tbl <- t_rsp(
 rsp = ANL$AVALC %in% c("CR", "PR"),
 col_by = relevel(factor(ANL$ARMCD), "ARM B"),
 partition_rsp_by = droplevels(factor(
   ANL$AVALC, levels = c("CR", "PR", "SD", "NON CR/PD", "PD", "NE")
 ))
)
tbl

#>                                                                      ARM B                 ARM A                 ARM C     
#>                                                                     (N=349)               (N=321)               (N=330)    
#> ---------------------------------------------------------------------------------------------------------------------------
#> Responders                                                        144 (41.26%)          128 (39.88%)          121 (36.67%) 
#> Non-Responders                                                    205 (58.74%)          193 (60.12%)          209 (63.33%) 
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                      (36.05, 46.63)        (34.48, 45.46)        (31.46, 42.12)
#> 
#> Difference in Response Rates (%)                                                           -1.39                 -4.59     
#>   95% CI for difference (Wald without correction)                                      (-8.83, 6.06)         (-11.92, 2.73)
#>   95% CI for difference (Wald with continuity correction)                              (-9.13, 6.35)         (-12.22, 3.03)
#>   p-value (Chi-squared)                                                                    0.7153                 0.22     
#> 
#> Odds Ratio                                                                                  0.94                  0.82     
#>   95% CI                                                                                (0.69, 1.29)          (0.61, 1.12) 
#> 
#> Complete Response (CR)                                            75 (21.49%)           62 (19.31%)           51 (15.45%)  
#>   95% CI                                                         (17.3, 26.18)         (15.14, 24.06)        (11.73, 19.81)
#> 
#> Partial Response (PR)                                             69 (19.77%)           66 (20.56%)           70 (21.21%)  
#>   95% CI                                                         (15.72, 24.34)        (16.27, 25.4)         (16.93, 26.02)
#> 
#> Stable Disease (SD)                                               64 (18.34%)           64 (19.94%)           74 (22.42%)  
#>   95% CI                                                         (14.42, 22.8)         (15.71, 24.73)        (18.04, 27.31)
#> 
#> Progressive Disease (PD)                                          69 (19.77%)           62 (19.31%)           61 (18.48%)  
#>   95% CI                                                         (15.72, 24.34)        (15.14, 24.06)        (14.44, 23.1) 
#> 
#> Not Evaluable (NE)                                                72 (20.63%)           67 (20.87%)           74 (22.42%)  
#> 
# NOT RUN {
Viewer(tbl)
# }

# Example 2 - ARM B as reference, ARM C displayed before ARM A
#    "NON CR/PD" response category displayed in partition section, "NE" responses
#     are not displayed model with two stratifiaction factors, CMH test performed
tbl2 <- t_rsp(
 rsp = ANL$AVALC %in% c("CR", "PR"),
 col_by = factor(ANL$ARMCD, c("ARM B", "ARM C", "ARM A")),
 partition_rsp_by = factor(ANL$AVALC, levels = c("CR", "PR", "SD", "NON CR/PD", "PD")),
 strata_data = ANL[c("RACE")]
)
tbl2

#>                                                                      ARM B                 ARM C                 ARM A     
#>                                                                     (N=349)               (N=330)               (N=321)    
#> ---------------------------------------------------------------------------------------------------------------------------
#> Responders                                                        144 (41.26%)          121 (36.67%)          128 (39.88%) 
#> Non-Responders                                                    205 (58.74%)          209 (63.33%)          193 (60.12%) 
#> 
#> 95% CI for Response Rates (Clopper-Pearson)                      (36.05, 46.63)        (31.46, 42.12)        (34.48, 45.46)
#> 
#> Difference in Response Rates (%)                                                           -4.59                 -1.39     
#>   95% CI for difference (Wald without correction)                                      (-11.92, 2.73)        (-8.83, 6.06) 
#>   95% CI for difference (Wald with continuity correction)                              (-12.22, 3.03)        (-9.13, 6.35) 
#>   p-value (Cochran-Mantel-Haenszel)*                                                       0.2224                0.7283    
#> 
#> Odds Ratio*                                                                                 0.83                  0.95     
#>   95% CI                                                                                (0.61, 1.12)          (0.69, 1.29) 
#> 
#> Complete Response (CR)                                            75 (21.49%)           51 (15.45%)           62 (19.31%)  
#>   95% CI                                                         (17.3, 26.18)         (11.73, 19.81)        (15.14, 24.06)
#> 
#> Partial Response (PR)                                             69 (19.77%)           70 (21.21%)           66 (20.56%)  
#>   95% CI                                                         (15.72, 24.34)        (16.93, 26.02)        (16.27, 25.4) 
#> 
#> Stable Disease (SD)                                               64 (18.34%)           74 (22.42%)           64 (19.94%)  
#>   95% CI                                                         (14.42, 22.8)         (18.04, 27.31)        (15.71, 24.73)
#> 
#> Non-CR or Non-PD (NON CR/PD)                                           0                     0                     0       
#>   95% CI                                                               -                     -                     -       
#> 
#> Progressive Disease (PD)                                          69 (19.77%)           61 (18.48%)           62 (19.31%)  
#>   95% CI                                                         (15.72, 24.34)        (14.44, 23.1)         (15.14, 24.06)
#> 
#> * Model stratified by RACE
# NOT RUN {
Viewer(tbl2)
# }