Summary Table by Visit

This function summarizes test results or change from baseline statistics by visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.

t_summarize_by_visit(data, visit, id, col_by, col_N = NULL)

Arguments

data
(data.frame)
with numerical variables to be summarized. If the variable has a label attribute then it will be used as the sub-header column name.

visit
(factor vector)
with visit names ordered by desired display order in the stacked table.

id
(vector)
unique subject identifier variable.

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

Value

an rtable object

Details

Every variable in data will be mapped to a sub-column under col_by column. Multiple variables will be shown side-by-side with the variable label (if present) or variable name as the sub-column header.

Currently n, Mean (SD), Median, IQR and Min-Max will be derived for each variable from data per each factor in col_by.

Examples

library(random.cdisc.data)
library(dplyr)
ADSL <- radsl(cached = TRUE)
ADVS <- radvs(cached = TRUE)

tbl1 <- t_summarize_by_visit(
  data = ADVS[c("AVAL")],
  visit = ADVS$AVISIT,
  col_by = ADVS$ARM,
  id = ADVS$USUBJID,
  col_N = table(ADSL$ARM)
)
if (FALSE) {
Viewer(tbl1)
}

tbl2 <- t_summarize_by_visit(
  data = ADVS[c("PCHG")],
  visit = ADVS$AVISIT,
  col_by = ADVS$ARM,
  id = ADVS$USUBJID,
  col_N = table(ADSL$ARM)
)
if (FALSE) {
Viewer(tbl2)
}

tbl3 <- t_summarize_by_visit(
  data = ADVS[c("AVAL", "CHG")],
  visit = ADVS$AVISIT,
  col_by = ADVS$ARM,
  id = ADVS$USUBJID,
  col_N = table(ADSL$ARM)
)
if (FALSE) {
Viewer(tbl3)
}

ADSL <- cadsl
ADQS <- cadqs %>%
  dplyr::filter(PARAMCD == "BFIALL")

tbl <- t_summarize_by_visit(
  data = ADQS[c("AVAL", "CHG")],
  visit = ADQS$AVISIT,
  col_by = ADQS$ARM,
  id = ADQS$USUBJID,
  col_N = table(ADSL$ARM)
)

if (FALSE) {
Viewer(tbl)
}