Summary Table by Visit

This function summarizes test results or change from baseline statistics by visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.

t_summarize_by_visit(data, visit, col_by)

Arguments

data
data frame with numerical variables to be summarized. If the variable has a label attribute then it will be used as the sub-header column name.

visit
factor with visit names ordered by desired display order in the stacked table.

col_by
factor with reference and comparison group information, the first level indicates the reference group

Value

an rtable object

Details

Every variable in data will be mapped to a sub-column under col_by column. Multiple variables will be shown side-by-side with the variable label (if present) or variable name as the sub-column header.

Currently n, Mean (SD), Median, IQR and Min-Max will be derived for each variable from data per each factor in col_by.

Examples

# EXAMPLE 1
ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B")
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA
ANL$PCHG <- ANL$CHG/ANL$AVAL*100

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

t_summarize_by_visit(data = ANL[c("AVAL")], visit = ANL$VISIT, col_by = ANL$ARM)

#>                       ARM A               ARM B    
#>                      (N=1000)            (N=1000)  
#> ---------------------------------------------------
#> visit 1
#>   n                    100                 100     
#>   Mean (SD)        0.01 (1.07)         -0.05 (1.01)
#>   Median              -0.15               -0.18    
#>   IQR              -0.69 - 0.66        -0.74 - 0.75
#>   Min - Max        -2.59 - 3.21        -2.43 - 2.37
#> 
#> visit 2
#>   n                    100                 100     
#>   Mean (SD)        0.02 (1.12)         -0.2 (0.99) 
#>   Median               0.17               -0.23    
#>   IQR              -0.72 - 0.86        -0.77 - 0.4 
#>   Min - Max        -3.34 - 2.18        -3.68 - 1.92
#> 
#> visit 3
#>   n                    100                 100     
#>   Mean (SD)        -0.16 (1.13)        -0.02 (0.95)
#>   Median              -0.09               -0.05    
#>   IQR              -0.83 - 0.64        -0.64 - 0.69
#>   Min - Max        -2.89 - 2.94        -2.52 - 2.08
#> 
#> visit 4
#>   n                    100                 100     
#>   Mean (SD)        0.29 (1.01)         -0.13 (0.99)
#>   Median               0.3                -0.21    
#>   IQR              -0.39 - 0.91        -0.84 - 0.49
#>   Min - Max        -2.52 - 2.83         -2 - 2.69  
#> 
#> visit 5
#>   n                    100                 100     
#>   Mean (SD)          0.13 (1)          0.15 (1.01) 
#>   Median               0.03                0.13    
#>   IQR              -0.42 - 0.72        -0.48 - 0.88
#>   Min - Max         -2.4 - 2.7         -3.08 - 2.64
#> 
#> visit 6
#>   n                    100                 100     
#>   Mean (SD)        0.07 (0.98)         -0.01 (1.03)
#>   Median               0.19                0.08    
#>   IQR              -0.67 - 0.77        -0.76 - 0.61
#>   Min - Max        -3.15 - 2.19        -2.25 - 3.58
#> 
#> visit 7
#>   n                    100                 100     
#>   Mean (SD)        0.02 (0.96)         0.08 (0.95) 
#>   Median               0.06                 0      
#>   IQR              -0.53 - 0.66        -0.42 - 0.82
#>   Min - Max        -2.18 - 2.34        -2.34 - 2.21
#> 
#> visit 8
#>   n                    100                 100     
#>   Mean (SD)        0.12 (1.03)         0.12 (0.92) 
#>   Median               0.12                0.15    
#>   IQR              -0.55 - 0.77        -0.44 - 0.68
#>   Min - Max        -2.26 - 2.78        -2.56 - 2.16
#> 
#> visit 9
#>   n                    100                 100     
#>   Mean (SD)          0.12 (1)          0.02 (1.04) 
#>   Median               0.09                0.16    
#>   IQR              -0.48 - 0.86        -0.7 - 0.68 
#>   Min - Max        -2.57 - 2.44        -2.88 - 2.68
#> 
#> visit 10
#>   n                    100                 100     
#>   Mean (SD)        0.14 (1.11)         -0.09 (0.97)
#>   Median               0.17               -0.13    
#>   IQR              -0.59 - 0.89        -0.74 - 0.51
#>   Min - Max        -2.89 - 3.07        -1.91 - 2.36

t_summarize_by_visit(data = ANL[c("PCHG")], visit = ANL$VISIT, col_by = ANL$ARM)

#>                            ARM A                        ARM B        
#>                          (N=1000)                     (N=1000)       
#> ---------------------------------------------------------------------
#> visit 1
#>   n                          0                            0          
#>   Mean (SD)               NE (NE)                      NE (NE)       
#>   Median                    NE                           NE          
#>   IQR                     NE - NE                      NE - NE       
#>   Min - Max               NE - NE                      NE - NE       
#> 
#> visit 2
#>   n                         100                          100         
#>   Mean (SD)          155.27 (5335.29)             -188.16 (1786.27)  
#>   Median                  -51.63                       -76.04        
#>   IQR                 -283.9 - 300.59             -430.15 - 273.16   
#>   Min - Max        -21277.16 - 45644.35          -10714.67 - 3473.94 
#> 
#> visit 3
#>   n                         100                          100         
#>   Mean (SD)          -376.31 (3668.3)             -384.11 (8430.71)  
#>   Median                  -55.17                       -58.62        
#>   IQR                -332.63 - 142.48              -294.76 - 382.9   
#>   Min - Max         -29502.4 - 13190.1          -76549.33 - 28851.82 
#> 
#> visit 4
#>   n                         100                          100         
#>   Mean (SD)            -67 (2425.73)              -91.44 (2634.17)   
#>   Median                   57.98                       -40.88        
#>   IQR                -196.27 - 291.66             -306.46 - 315.32   
#>   Min - Max         -13335.03 - 14031.6         -16364.85 - 15785.52 
#> 
#> visit 5
#>   n                         100                          100         
#>   Mean (SD)          -562.44 (3124.33)            -13.38 (1506.82)   
#>   Median                   -3.4                         48.15        
#>   IQR                -306.88 - 207.91              -285.09 - 312.4   
#>   Min - Max          -22152 - 6609.97            -5330.26 - 7374.05  
#> 
#> visit 6
#>   n                         100                          100         
#>   Mean (SD)          -69.85 (1997.23)              55.01 (1376.06)   
#>   Median                   11.07                        13.46        
#>   IQR                -268.34 - 300.64             -244.62 - 346.58   
#>   Min - Max         -7199.63 - 13798.33          -6582.87 - 9010.39  
#> 
#> visit 7
#>   n                         100                          100         
#>   Mean (SD)         -2004.47 (16667.03)           -321.86 (5694.65)  
#>   Median                   66.99                       -21.47        
#>   IQR                -208.36 - 477.49             -478.46 - 375.24   
#>   Min - Max        -148995.39 - 11529.61        -45901.19 - 25414.87 
#> 
#> visit 8
#>   n                         100                          100         
#>   Mean (SD)         -6979.92 (66107.94)           -72.91 (1509.11)   
#>   Median                  -40.68                        57.43        
#>   IQR                -315.38 - 261.25              -265.7 - 276.58   
#>   Min - Max        -660547.93 - 14170.13         -10637.17 - 5258.32 
#> 
#> visit 9
#>   n                         100                          100         
#>   Mean (SD)          -335.48 (1994.29)            -182.52 (2030.06)  
#>   Median                  -43.44                        27.48        
#>   IQR                -283.98 - 196.64             -243.34 - 250.95   
#>   Min - Max         -15708.34 - 2956.24          -15354.92 - 3680.6  
#> 
#> visit 10
#>   n                         100                          100         
#>   Mean (SD)          -117.04 (1145.03)             -30.42 (1314.9)   
#>   Median                  -18.76                       -24.87        
#>   IQR                -280.51 - 257.86             -294.33 - 293.38   
#>   Min - Max         -6821.16 - 5547.81           -8208.33 - 3921.06  

# Add label to variable instead showing variable name
ANL <- var_relabel(ANL, AVAL = "Value at\nVisit",
                        CHG = "Change from\nBaseline",
                        PCHG = "Percent Change\nfrom Baseline")
t_summarize_by_visit(data = ANL[c("AVAL", "CHG")], visit = ANL$VISIT, col_by = ANL$ARM)

#>                                 ARM A                                   ARM B              
#>                                (N=1000)                                (N=1000)            
#>                      Value at          Change from           Value at          Change from 
#>                       Visit              Baseline             Visit              Baseline  
#> -------------------------------------------------------------------------------------------
#> visit 1
#>   n                    100                  0                  100                  0      
#>   Mean (SD)        0.01 (1.07)           NE (NE)           -0.05 (1.01)          NE (NE)   
#>   Median              -0.15                 NE                -0.18                 NE     
#>   IQR              -0.69 - 0.66          NE - NE           -0.74 - 0.75          NE - NE   
#>   Min - Max        -2.59 - 3.21          NE - NE           -2.43 - 2.37          NE - NE   
#> 
#> visit 2
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.02 (1.12)         2.18 (2.08)         -0.2 (0.99)         2.24 (1.92) 
#>   Median               0.17                2.08               -0.23                2.3     
#>   IQR              -0.72 - 0.86        0.81 - 3.67         -0.77 - 0.4         0.96 - 3.46 
#>   Min - Max        -3.34 - 2.18        -3.16 - 8.15        -3.68 - 1.92         -3 - 6.92  
#> 
#> visit 3
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.16 (1.13)        1.74 (2.01)         -0.02 (0.95)        2.09 (1.85) 
#>   Median              -0.09                1.61               -0.05                2.31    
#>   IQR              -0.83 - 0.64        0.56 - 3.13         -0.64 - 0.69        1.03 - 3.38 
#>   Min - Max        -2.89 - 2.94        -2.81 - 7.6         -2.52 - 2.08        -2.57 - 5.7 
#> 
#> visit 4
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.29 (1.01)         1.98 (1.67)         -0.13 (0.99)        2.13 (1.79) 
#>   Median               0.3                 1.94               -0.21                2.18    
#>   IQR              -0.39 - 0.91        0.66 - 3.23         -0.84 - 0.49        0.98 - 3.32 
#>   Min - Max        -2.52 - 2.83        -1.26 - 7.93         -2 - 2.69          -3.73 - 5.8 
#> 
#> visit 5
#>   n                    100                 100                 100                 100     
#>   Mean (SD)          0.13 (1)          1.73 (2.05)         0.15 (1.01)         2.01 (2.04) 
#>   Median               0.03                1.62                0.13                1.93    
#>   IQR              -0.42 - 0.72        0.22 - 3.35         -0.48 - 0.88         0.72 - 3.6 
#>   Min - Max         -2.4 - 2.7         -3.3 - 5.79         -3.08 - 2.64        -3.09 - 8.46
#> 
#> visit 6
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.07 (0.98)         2.09 (2.02)         -0.01 (1.03)        1.97 (1.73) 
#>   Median               0.19                1.92                0.08                1.84    
#>   IQR              -0.67 - 0.77        0.73 - 3.34         -0.76 - 0.61         0.9 - 2.98 
#>   Min - Max        -3.15 - 2.19        -4.26 - 7.19        -2.25 - 3.58        -2.23 - 6.39
#> 
#> visit 7
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.02 (0.96)         2.04 (2.11)         0.08 (0.95)         2.25 (2.19) 
#>   Median               0.06                2.19                 0                  2.18    
#>   IQR              -0.53 - 0.66        0.58 - 3.42         -0.42 - 0.82        0.96 - 3.44 
#>   Min - Max        -2.18 - 2.34        -3.72 - 6.79        -2.34 - 2.21        -3.3 - 7.35 
#> 
#> visit 8
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.12 (1.03)         1.87 (1.89)         0.12 (0.92)         1.91 (1.87) 
#>   Median               0.12                1.8                 0.15                1.73    
#>   IQR              -0.55 - 0.77        0.95 - 3.03         -0.44 - 0.68         0.8 - 3.04 
#>   Min - Max        -2.26 - 2.78        -3.08 - 7.45        -2.56 - 2.16        -2.49 - 6.52
#> 
#> visit 9
#>   n                    100                 100                 100                 100     
#>   Mean (SD)          0.12 (1)          1.69 (1.94)         0.02 (1.04)         2.11 (2.11) 
#>   Median               0.09                1.42                0.16                2.08    
#>   IQR              -0.48 - 0.86        0.61 - 2.85         -0.7 - 0.68         0.84 - 3.69 
#>   Min - Max        -2.57 - 2.44        -3.16 - 7.18        -2.88 - 2.68        -2.7 - 7.01 
#> 
#> visit 10
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.14 (1.11)         2.15 (1.94)         -0.09 (0.97)        2.16 (1.65) 
#>   Median               0.17                2.24               -0.13                2.01    
#>   IQR              -0.59 - 0.89        1.08 - 3.32         -0.74 - 0.51         1.1 - 3.21 
#>   Min - Max        -2.89 - 3.07        -2.29 - 6.64        -1.91 - 2.36        -3.46 - 6.26

# EXAMPLE 2
library(random.cdisc.data)
library(dplyr)

AQS <- radam("AQS") %>%
   filter(PARAMCD == "BFIALL") %>%
   mutate(
     ARM = factor(sample(c("ARM A", "ARM B", "ARM C"), nrow(.), replace = TRUE)),
     AVISIT = factor(AVISIT, levels = unique(AVISIT))
   )

tbl <- t_summarize_by_visit(data = AQS[c("AVAL", "CHG")], visit = AQS$AVISIT, col_by = AQS$ARM)
tbl

#>                                  ARM A                                     ARM B                                     ARM C               
#>                                 (N=213)                                   (N=226)                                   (N=250)              
#>                        AVAL                  CHG                 AVAL                  CHG                 AVAL                  CHG     
#> -----------------------------------------------------------------------------------------------------------------------------------------
#> WEEK 1
#>   n                     36                   36                   34                   34                   30                   30      
#>   Mean (SD)        -0.06 (1.05)             0 (0)             0.03 (0.76)             0 (0)             0.06 (1.12)             0 (0)    
#>   Median               -0.11                  0                  0.24                   0                  -0.19                  0      
#>   IQR              -0.73 - 0.75             0 - 0            -0.45 - 0.51             0 - 0            -0.63 - 0.96             0 - 0    
#>   Min - Max         -1.7 - 1.94             0 - 0            -1.52 - 1.43             0 - 0            -2.61 - 2.82             0 - 0    
#> 
#> WEEK 2
#>   n                     34                   34                   31                   31                   35                   35      
#>   Mean (SD)        -0.17 (0.94)          0.04 (1.27)          0.06 (0.85)         -0.12 (0.86)          0.11 (1.12)          0.05 (1.94) 
#>   Median               -0.21                -0.01                0.01                 -0.13                 0.1                 0.16     
#>   IQR              -0.76 - 0.15         -0.97 - 0.78         -0.54 - 0.61         -0.72 - 0.58         -0.43 - 0.81         -1.46 - 1.74 
#>   Min - Max        -2.34 - 1.91          -2.7 - 3.45          -1.6 - 1.8          -1.85 - 1.39         -2.99 - 2.62         -4.42 - 3.25 
#> 
#> WEEK 3
#>   n                     24                   24                   39                   39                   30                   30      
#>   Mean (SD)          0.08 (1)            0.08 (1.37)          0.03 (1.01)          0.08 (1.47)         -0.44 (0.69)          -0.45 (1.1) 
#>   Median               0.14                 0.16                 0.05                 0.03                 -0.66                -0.42    
#>   IQR              -0.55 - 0.68         -0.88 - 1.02         -0.59 - 0.82         -0.96 - 0.83         -0.94 - -0.08         -1.15 - 0.3 
#>   Min - Max        -1.78 - 2.73         -2.21 - 3.33         -2.14 - 1.99          -3.4 - 3.16         -1.39 - 1.28         -2.68 - 1.82 
#> 
#> WEEK 4
#>   n                     23                   23                   26                   26                   35                   35      
#>   Mean (SD)        -0.12 (1.12)         -0.01 (1.52)         -0.46 (0.78)         -0.47 (1.13)          0.11 (0.92)          0.11 (1.31) 
#>   Median               -0.22                 0.4                 -0.4                 -0.35                  0                  -0.13    
#>   IQR              -0.94 - 0.61          -1.41 - 0.9         -0.85 - -0.07        -1.16 - 0.29         -0.35 - 0.66         -0.75 - 1.11 
#>   Min - Max        -2.18 - 2.46         -2.68 - 3.19         -1.96 - 1.36         -2.63 - 2.09         -1.88 - 2.38         -2.86 - 2.19 
#> 
#> WEEK 5
#>   n                     26                   26                   21                   21                   31                   31      
#>   Mean (SD)         0.13 (1.25)          0.26 (1.2)           -0.16 (1.1)         -0.01 (1.38)          0.14 (0.99)          0.1 (1.35)  
#>   Median               0.26                 0.39                 0.05                 0.32                 0.35                 0.14     
#>   IQR              -0.58 - 1.29         -0.47 - 1.04         -0.71 - 0.45         -0.37 - 0.89         -0.57 - 0.83         -0.97 - 1.05 
#>   Min - Max        -2.44 - 2.11         -2.49 - 2.48         -2.63 - 1.97         -4.05 - 1.77         -1.98 - 1.83         -2.22 - 4.11 
#> 
#> WEEK 6
#>   n                     17                   17                   23                   23                   22                   22      
#>   Mean (SD)        -0.08 (0.94)           0 (1.52)             -0.23 (1)          -0.28 (1.37)          0.01 (0.95)         -0.04 (1.39) 
#>   Median               -0.36                0.01                 -0.52                -0.04                0.02                 0.07     
#>   IQR               -0.79 - 0.6         -0.61 - 0.97         -0.88 - 0.69         -1.31 - 0.65         -0.71 - 0.54         -1.35 - 0.75 
#>   Min - Max        -1.57 - 1.63          -2.55 - 2.4         -2.36 - 1.17         -2.46 - 1.88         -1.42 - 2.63         -2.47 - 2.79 
#> 
#> WEEK 7
#>   n                     17                   17                   14                   14                   22                   22      
#>   Mean (SD)        -0.32 (1.06)         -0.38 (1.46)          0.18 (0.96)          0.06 (1.25)          0.11 (0.86)          0.34 (1.21) 
#>   Median               -0.56                -0.6                 0.13                 0.17                 0.27                 0.42     
#>   IQR              -0.93 - 0.05          -1.2 - 0.24         -0.43 - 0.82          -0.9 - 0.73         -0.28 - 0.52         -0.55 - 1.13 
#>   Min - Max        -1.69 - 2.31         -2.68 - 3.84         -1.46 - 2.01         -1.72 - 2.64         -1.61 - 1.47          -1.8 - 2.88 
#> 
#> WEEK 8
#>   n                     15                   15                   15                   15                   15                   15      
#>   Mean (SD)        -0.14 (0.97)         -0.11 (1.74)          0.22 (0.99)          0.28 (1.15)         -0.56 (1.13)         -0.11 (1.44) 
#>   Median               0.09                 -0.22                -0.07                 0.4                 0.11                 0.43     
#>   IQR              -0.67 - 0.48         -1.24 - 0.95         -0.41 - 0.83         -0.52 - 1.16         -1.28 - 0.19         -1.14 - 0.69 
#>   Min - Max        -2.04 - 1.57         -3.98 - 2.74         -0.91 - 2.02         -1.97 - 1.99         -2.71 - 0.55          -2.9 - 2.06 
#> 
#> WEEK 9
#>   n                      9                    9                   14                   14                   14                   14      
#>   Mean (SD)         -0.2 (0.81)         -0.65 (1.41)          0.24 (0.82)          0.27 (1.09)          0.55 (0.97)          1.2 (0.84)  
#>   Median               -0.36                -0.77                0.15                 0.35                 0.46                 1.36     
#>   IQR              -0.83 - 0.21         -1.62 - -0.24         -0.2 - 0.63           -0.37 - 1           0.17 - 1.22          0.62 - 1.69 
#>   Min - Max        -1.35 - 1.14         -1.84 - 2.65         -1.03 - 1.68           -1.94 - 2          -1.73 - 1.95         -0.21 - 2.54 
#> 
#> WEEK 10
#>   n                      5                    5                    7                    7                   11                   11      
#>   Mean (SD)         0.37 (0.95)          0.13 (1.13)          0.15 (0.69)          0.02 (1.75)          0.4 (0.87)           0.79 (1.28) 
#>   Median                 0                  0.15                 0.09                 0.51                  0.4                  0.7     
#>   IQR              -0.09 - 0.52         -0.36 - 0.24         -0.04 - 0.51         -0.88 - 0.92         -0.12 - 0.62          0.17 - 1.92 
#>   Min - Max        -0.51 - 1.93         -1.24 - 1.87           -1 - 1.06          -2.91 - 2.43         -0.85 - 2.16          -1.68 - 2.4 
#> 
#> WEEK 11
#>   n                      7                    7                    2                    2                    5                    5      
#>   Mean (SD)        -0.13 (0.77)          0.43 (1.38)          1.43 (0.82)          1.94 (1.05)          0.35 (0.59)          0.6 (0.82)  
#>   Median               -0.13                0.25                 1.43                 1.94                 0.27                 0.55     
#>   IQR              -0.38 - 0.03          0.05 - 0.84          1.14 - 1.72          1.57 - 2.31          0.21 - 0.81          0.03 - 0.9  
#>   Min - Max        -1.34 - 1.26         -1.82 - 2.76          0.85 - 2.01          1.19 - 2.68         -0.52 - 0.97         -0.28 - 1.81 
# NOT RUN {
Viewer(tbl)
# }