Summary Table by Visit

This function summarizes test results or change from baseline statistics by visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.

t_summarize_by_visit(data, visit, id, col_by)

Arguments

data
data frame with numerical variables to be summarized. If the variable has a label attribute then it will be used as the sub-header column name.

visit
factor with visit names ordered by desired display order in the stacked table.

id
unique subject identifier variable.

col_by
factor with reference and comparison group information, the first level indicates the reference group

Value

an rtable object

Details

Every variable in data will be mapped to a sub-column under col_by column. Multiple variables will be shown side-by-side with the variable label (if present) or variable name as the sub-column header.

Currently n, Mean (SD), Median, IQR and Min-Max will be derived for each variable from data per each factor in col_by.

Examples

# EXAMPLE 1
ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B")
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA
ANL$PCHG <- ANL$CHG/ANL$AVAL*100

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

t_summarize_by_visit(data = ANL[c("AVAL")], visit = ANL$VISIT, col_by = ANL$ARM, id = ANL$USUBJID)

#>                       ARM A               ARM B    
#>                      (N=100)             (N=100)   
#> ---------------------------------------------------
#> visit 1
#>   n                    100                 100     
#>   Mean (SD)        0.07 (1.05)         -0.15 (1.06)
#>   Median               0.09               -0.17    
#>   IQR              -0.36 - 0.62        -0.71 - 0.57
#>   Min - Max        -2.61 - 2.76        -2.76 - 2.75
#> 
#> visit 2
#>   n                    100                 100     
#>   Mean (SD)        0.11 (1.01)         -0.21 (0.99)
#>   Median               0.13                -0.3    
#>   IQR              -0.61 - 0.86        -0.8 - 0.38 
#>   Min - Max        -1.9 - 2.56         -2.28 - 2.72
#> 
#> visit 3
#>   n                    100                 100     
#>   Mean (SD)        -0.03 (1.05)        -0.09 (0.99)
#>   Median               0.01               -0.02    
#>   IQR              -0.83 - 0.68        -0.8 - 0.54 
#>   Min - Max        -2.45 - 2.65        -1.95 - 3.11
#> 
#> visit 4
#>   n                    100                 100     
#>   Mean (SD)        0.19 (0.99)         0.08 (0.94) 
#>   Median               0.06                0.01    
#>   IQR               -0.42 - 1          -0.53 - 0.83
#>   Min - Max        -1.98 - 2.13        -1.7 - 2.59 
#> 
#> visit 5
#>   n                    100                 100     
#>   Mean (SD)        -0.18 (1.03)        0.06 (0.89) 
#>   Median              -0.27                0.04    
#>   IQR              -0.85 - 0.63        -0.54 - 0.73
#>   Min - Max        -2.94 - 2.1         -2.95 - 1.67
#> 
#> visit 6
#>   n                    100                 100     
#>   Mean (SD)        -0.05 (0.97)        0.01 (0.94) 
#>   Median               -0.1               -0.05    
#>   IQR              -0.69 - 0.61        -0.62 - 0.55
#>   Min - Max        -2.81 - 2.65        -1.92 - 2.05
#> 
#> visit 7
#>   n                    100                 100     
#>   Mean (SD)        -0.16 (0.99)        0.11 (0.96) 
#>   Median              -0.17                0.06    
#>   IQR              -0.88 - 0.54        -0.42 - 0.82
#>   Min - Max        -2.45 - 2.48        -2.37 - 2.21
#> 
#> visit 8
#>   n                    100                 100     
#>   Mean (SD)        0.06 (0.94)         0.11 (0.97) 
#>   Median               0.13                0.03    
#>   IQR              -0.53 - 0.61        -0.52 - 0.68
#>   Min - Max        -2.39 - 2.27        -2.42 - 2.84
#> 
#> visit 9
#>   n                    100                 100     
#>   Mean (SD)        -0.1 (0.98)         -0.04 (0.97)
#>   Median              -0.21               -0.04    
#>   IQR              -0.8 - 0.52         -0.76 - 0.61
#>   Min - Max        -2.09 - 2.69        -2.27 - 2.32
#> 
#> visit 10
#>   n                    100                 100     
#>   Mean (SD)        0.14 (0.95)         -0.05 (0.94)
#>   Median               0.2                -0.14    
#>   IQR              -0.49 - 0.81        -0.69 - 0.65
#>   Min - Max        -2.36 - 2.22        -2.07 - 2.53

t_summarize_by_visit(data = ANL[c("PCHG")], visit = ANL$VISIT, col_by = ANL$ARM, id = ANL$USUBJID)

#>                            ARM A                        ARM B        
#>                           (N=100)                      (N=100)       
#> ---------------------------------------------------------------------
#> visit 1
#>   n                          0                            0          
#>   Mean (SD)               NE (NE)                      NE (NE)       
#>   Median                    NE                           NE          
#>   IQR                     NE - NE                      NE - NE       
#>   Min - Max               NE - NE                      NE - NE       
#> 
#> visit 2
#>   n                         100                          100         
#>   Mean (SD)          104.55 (1110.09)             -65.26 (2323.81)   
#>   Median                   24.31                       -123.59       
#>   IQR                -266.79 - 307.52             -413.92 - 111.09   
#>   Min - Max         -3158.04 - 7864.09           -8643.82 - 18492.48 
#> 
#> visit 3
#>   n                         100                          100         
#>   Mean (SD)         1179.44 (12950.24)            286.38 (2289.98)   
#>   Median                  -13.51                        10.37        
#>   IQR                 -265.62 - 302.6             -266.82 - 320.14   
#>   Min - Max        -28769.67 - 123730.51         -9116.82 - 13231.49 
#> 
#> visit 4
#>   n                         100                          100         
#>   Mean (SD)          338.44 (3852.73)             -408.05 (2464.12)  
#>   Median                   25.97                       -43.64        
#>   IQR                -259.19 - 300.69             -416.77 - 205.69   
#>   Min - Max         -5229.93 - 31180.41          -15997.59 - 8340.77 
#> 
#> visit 5
#>   n                         100                          100         
#>   Mean (SD)          956.51 (16159.33)             -80.36 (3941.8)   
#>   Median                  -72.66                        6.78         
#>   IQR                -273.67 - 234.98             -264.28 - 261.07   
#>   Min - Max        -47988.13 - 153499.88         -33871.8 - 12582.63 
#> 
#> visit 6
#>   n                         100                          100         
#>   Mean (SD)           -189.7 (1164.2)            1227.99 (13130.65)  
#>   Median                  -78.72                        -26.1        
#>   IQR                 -385.1 - 327.23              -275.2 - 313.47   
#>   Min - Max         -5663.16 - 3885.12           -14216.6 - 129522.2 
#> 
#> visit 7
#>   n                         100                          100         
#>   Mean (SD)          331.08 (3726.15)             580.92 (5670.46)   
#>   Median                  -116.35                       38.37        
#>   IQR                -414.56 - 197.48             -269.36 - 370.67   
#>   Min - Max         -4345.85 - 31771.89         -22646.99 - 41965.14 
#> 
#> visit 8
#>   n                         100                          100         
#>   Mean (SD)          -192.91 (2230.89)             172.91 (3404.7)   
#>   Median                   50.48                        46.97        
#>   IQR                 -233.8 - 310.46             -304.21 - 282.66   
#>   Min - Max         -18422.09 - 4794.23         -10203.02 - 26380.63 
#> 
#> visit 9
#>   n                         100                          100         
#>   Mean (SD)          169.67 (1718.59)             -168.25 (2771.41)  
#>   Median                  -30.21                        18.13        
#>   IQR                -323.53 - 287.18             -292.19 - 431.32   
#>   Min - Max         -2896.47 - 13820.34         -22408.05 - 10466.93 
#> 
#> visit 10
#>   n                         100                          100         
#>   Mean (SD)          700.82 (5594.59)             1390.29 (7184.03)  
#>   Median                  138.45                       -22.45        
#>   IQR                -203.97 - 684.36             -276.97 - 276.87   
#>   Min - Max        -24200.16 - 43555.02          -5675.29 - 53523.41 

# Add label to variable instead showing variable name
ANL <- var_relabel(ANL, AVAL = "Value at\nVisit",
                        CHG = "Change from\nBaseline",
                        PCHG = "Percent Change\nfrom Baseline")

t_summarize_by_visit(
  data = ANL[c("AVAL", "CHG")],
  visit = ANL$VISIT,
  col_by = ANL$ARM,
  id = ANL$USUBJID
)

#>                                 ARM A                                   ARM B              
#>                                (N=100)                                 (N=100)             
#>                      Value at          Change from           Value at          Change from 
#>                       Visit              Baseline             Visit              Baseline  
#> -------------------------------------------------------------------------------------------
#> visit 1
#>   n                    100                  0                  100                  0      
#>   Mean (SD)        0.07 (1.05)           NE (NE)           -0.15 (1.06)          NE (NE)   
#>   Median               0.09                 NE                -0.17                 NE     
#>   IQR              -0.36 - 0.62          NE - NE           -0.71 - 0.57          NE - NE   
#>   Min - Max        -2.61 - 2.76          NE - NE           -2.76 - 2.75          NE - NE   
#> 
#> visit 2
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.11 (1.01)          1.9 (2.08)         -0.21 (0.99)        2.15 (1.84) 
#>   Median               0.13                1.81                -0.3                2.24    
#>   IQR              -0.61 - 0.86         0.6 - 3.37         -0.8 - 0.38         0.95 - 3.28 
#>   Min - Max        -1.9 - 2.56         -3.64 - 8.02        -2.28 - 2.72        -2.33 - 6.14
#> 
#> visit 3
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.03 (1.05)         2.11 (2.2)         -0.09 (0.99)        1.88 (1.79) 
#>   Median               0.01                2.18               -0.02                2.03    
#>   IQR              -0.83 - 0.68        0.39 - 3.69         -0.8 - 0.54         0.93 - 2.89 
#>   Min - Max        -2.45 - 2.65        -2.95 - 6.95        -1.95 - 3.11        -2.84 - 6.5 
#> 
#> visit 4
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.19 (0.99)         1.92 (1.75)         0.08 (0.94)         2.19 (2.06) 
#>   Median               0.06                1.96                0.01                2.16    
#>   IQR               -0.42 - 1          0.76 - 3.15         -0.53 - 0.83        0.66 - 3.43 
#>   Min - Max        -1.98 - 2.13        -1.73 - 5.37        -1.7 - 2.59         -2.45 - 9.17
#> 
#> visit 5
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.18 (1.03)        2.12 (2.04)         0.06 (0.89)         1.42 (1.91) 
#>   Median              -0.27                2.21                0.04                1.51    
#>   IQR              -0.85 - 0.63        0.76 - 3.55         -0.54 - 0.73        0.24 - 2.63 
#>   Min - Max        -2.94 - 2.1         -2.5 - 8.19         -2.95 - 1.67        -3.25 - 6.94
#> 
#> visit 6
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.05 (0.97)        2.07 (1.94)         0.01 (0.94)         1.78 (1.85) 
#>   Median               -0.1                2.14               -0.05                1.93    
#>   IQR              -0.69 - 0.61         0.74 - 3.5         -0.62 - 0.55        1.03 - 2.91 
#>   Min - Max        -2.81 - 2.65        -2.59 - 7.23        -1.92 - 2.05        -3.81 - 6.2 
#> 
#> visit 7
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.16 (0.99)        2.07 (1.97)         0.11 (0.96)          2.1 (1.74) 
#>   Median              -0.17                2.03                0.06                2.19    
#>   IQR              -0.88 - 0.54        0.93 - 3.41         -0.42 - 0.82        0.78 - 3.31 
#>   Min - Max        -2.45 - 2.48        -2.6 - 6.63         -2.37 - 2.21        -2.14 - 6.47
#> 
#> visit 8
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.06 (0.94)         1.86 (2.08)         0.11 (0.97)         1.93 (2.02) 
#>   Median               0.13                1.8                 0.03                1.88    
#>   IQR              -0.53 - 0.61        0.31 - 3.35         -0.52 - 0.68        0.66 - 3.23 
#>   Min - Max        -2.39 - 2.27        -3.06 - 8.14        -2.42 - 2.84        -3.57 - 7.01
#> 
#> visit 9
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.1 (0.98)          2.14 (1.9)         -0.04 (0.97)        2.06 (2.25) 
#>   Median              -0.21                2.07               -0.04                2.22    
#>   IQR              -0.8 - 0.52          1.21 - 3.2         -0.76 - 0.61         0.7 - 3.71 
#>   Min - Max        -2.09 - 2.69        -2.6 - 7.19         -2.27 - 2.32        -4.61 - 7.27
#> 
#> visit 10
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.14 (0.95)          2.5 (2.12)         -0.05 (0.94)        1.79 (2.15) 
#>   Median               0.2                 2.61               -0.14                1.86    
#>   IQR              -0.49 - 0.81        1.16 - 3.84         -0.69 - 0.65        -0.14 - 3.43
#>   Min - Max        -2.36 - 2.22        -2.34 - 6.65        -2.07 - 2.53        -3.92 - 7.57

# EXAMPLE 2
library(random.cdisc.data)
library(dplyr)

AQS <- radam("AQS") %>%
   filter(PARAMCD == "BFIALL") %>%
   mutate(
     ARM = factor(sample(c("ARM A", "ARM B", "ARM C"), nrow(.), replace = TRUE)),
     AVISIT = factor(AVISIT, levels = unique(AVISIT))
   )

tbl <- t_summarize_by_visit(
  data = AQS[c("AVAL", "CHG")],
  visit = AQS$AVISIT,
  col_by = AQS$ARM,
  id = AQS$USUBJID
)
tbl

#>                                 ARM A                                   ARM B                                   ARM C              
#>                                 (N=89)                                  (N=89)                                  (N=93)             
#>                        AVAL                CHG                 AVAL                CHG                 AVAL                CHG     
#> -----------------------------------------------------------------------------------------------------------------------------------
#> WEEK 1
#>   n                     31                  31                  38                  38                  31                  31     
#>   Mean (SD)        -0.05 (0.9)            0 (0)            -0.22 (1.06)           0 (0)            0.09 (1.06)            0 (0)    
#>   Median              -0.28                 0                 -0.36                 0                 -0.01                 0      
#>   IQR              -0.58 - 0.25           0 - 0            -0.89 - 0.42           0 - 0            -0.45 - 0.54           0 - 0    
#>   Min - Max        -1.92 - 1.9            0 - 0            -2.62 - 1.82           0 - 0            -1.63 - 2.59           0 - 0    
#> 
#> WEEK 2
#>   n                     31                  31                  31                  31                  38                  38     
#>   Mean (SD)         0.14 (0.9)         0.57 (1.26)         0.06 (1.07)         -0.08 (1.67)        -0.17 (1.08)        -0.22 (1.24)
#>   Median               0.14                0.66                0.16               -0.26               -0.11                0.01    
#>   IQR              -0.24 - 0.92        -0.02 - 1.52        -0.45 - 0.82        -1.29 - 1.35        -1.06 - 0.52        -1.07 - 0.75
#>   Min - Max        -1.96 - 1.42        -2.04 - 2.94        -2.54 - 1.61        -4.23 - 2.48        -2.38 - 1.94        -2.9 - 1.92 
#> 
#> WEEK 3
#>   n                     24                  24                  31                  31                  34                  34     
#>   Mean (SD)        0.08 (1.09)         0.26 (1.45)         -0.14 (0.92)        -0.03 (1.53)         0.05 (1.1)          0.16 (1.3) 
#>   Median               0.03                0.12               -0.25               -0.18                0.04                0.33    
#>   IQR              -0.63 - 0.97        -0.76 - 1.43        -0.83 - 0.54        -0.92 - 1.1         -0.57 - 0.96        -0.92 - 1.01
#>   Min - Max        -2.44 - 1.8         -2.4 - 3.43         -1.73 - 1.61        -3.41 - 3.08        -2.21 - 2.16        -2.04 - 2.81
#> 
#> WEEK 4
#>   n                     25                  25                  29                  29                  27                  27     
#>   Mean (SD)         0.19 (0.7)         0.18 (1.49)         0.27 (0.71)         0.27 (0.95)         -0.11 (0.83)        0.26 (1.34) 
#>   Median               0.04                0.46                0.22                0.39               -0.18                0.4     
#>   IQR              -0.36 - 0.81        -0.38 - 1.45        -0.3 - 0.83         -0.35 - 0.94        -0.47 - 0.4         -0.25 - 1.12
#>   Min - Max        -1.02 - 1.7         -2.71 - 2.31        -1.28 - 1.75        -1.36 - 2.18        -2.14 - 1.29        -4.05 - 1.93
#> 
#> WEEK 5
#>   n                     25                  25                  22                  22                  23                  23     
#>   Mean (SD)          0 (1.04)          0.17 (1.21)         0.13 (1.09)         0.15 (1.37)         0.06 (0.97)         0.17 (1.24) 
#>   Median              -0.03                0.26                0.27                0.09               -0.22                0.24    
#>   IQR              -0.66 - 0.63        -0.54 - 1.02        -0.57 - 0.69        -0.59 - 1.19        -0.74 - 0.83        -0.65 - 1.2 
#>   Min - Max        -2.09 - 1.68        -2.44 - 2.16        -1.89 - 2.29        -2.84 - 2.38        -1.46 - 2.14        -2.98 - 2.13
#> 
#> WEEK 6
#>   n                     20                  20                  28                  28                  15                  15     
#>   Mean (SD)        0.09 (1.15)         0.09 (1.67)         -0.16 (0.92)        0.11 (1.27)          0.1 (1.02)         0.21 (1.41) 
#>   Median              -0.02                0.29                -0.1                0.23                0.41                0.55    
#>   IQR              -0.3 - 0.84         -1.03 - 1.28        -0.76 - 0.57        -0.99 - 0.91        -0.78 - 0.84        -0.66 - 1.16
#>   Min - Max        -3.22 - 2.18        -3.52 - 3.13        -1.86 - 1.68        -2.04 - 2.58        -1.66 - 1.82        -3.05 - 2.27
#> 
#> WEEK 7
#>   n                     20                  20                  19                  19                  20                  20     
#>   Mean (SD)         0.24 (0.7)         0.41 (0.96)         0.57 (0.96)         0.83 (0.94)         -0.43 (0.92)        -0.35 (1.52)
#>   Median               0.22                0.53                0.62                0.65                -0.4               -0.47    
#>   IQR              -0.07 - 0.83        -0.29 - 0.92        0.19 - 1.19          0.3 - 1.59         -1.14 - 0.1         -1.12 - 0.63
#>   Min - Max        -1.13 - 1.32         -2 - 1.98          -1.92 - 2.04        -0.81 - 2.11        -2.05 - 1.65        -3.1 - 2.68 
#> 
#> WEEK 8
#>   n                     18                  18                  19                  19                  14                  14     
#>   Mean (SD)        -0.49 (1.03)        -0.44 (1.55)        -0.1 (0.74)         -0.03 (1.2)          0.1 (0.75)         0.53 (0.98) 
#>   Median              -0.35               -0.15               -0.15                0.27                0.3                 0.7     
#>   IQR              -1.41 - 0.23        -1.63 - 0.83        -0.55 - 0.63        -1.22 - 1.1         -0.17 - 0.57        -0.16 - 1.07
#>   Min - Max        -2.24 - 1.37        -3.35 - 2.41        -1.8 - 0.85         -1.67 - 1.57        -1.54 - 1.09        -1.2 - 2.12 
#> 
#> WEEK 9
#>   n                     12                  12                  13                  13                  7                   7      
#>   Mean (SD)         0.4 (0.76)         0.78 (1.03)         -0.24 (0.95)        -0.24 (1.28)        0.01 (1.46)         0.26 (1.38) 
#>   Median               0.38                0.88               -0.19                0.01                0.11                0.56    
#>   IQR              -0.14 - 0.82        0.36 - 1.48         -0.93 - 0.36        -1.16 - 0.45        -0.76 - 1.1         -0.6 - 1.24 
#>   Min - Max        -0.77 - 1.91        -1.58 - 2.04        -1.74 - 1.74        -2.27 - 2.43        -2.37 - 1.69        -1.92 - 1.92
#> 
#> WEEK 10
#>   n                     3                   3                   5                   5                   10                  10     
#>   Mean (SD)         0.05 (0.2)         0.63 (0.11)         0.04 (0.82)         0.22 (1.49)         -0.44 (1.33)        -0.12 (1.25)
#>   Median               0.01                0.66               -0.16                0.34               -0.64               -0.36    
#>   IQR              -0.05 - 0.14        0.59 - 0.69         -0.7 - 0.66         -1.06 - 0.97        -1.49 - 0.37        -1.1 - 0.74 
#>   Min - Max        -0.12 - 0.27        0.51 - 0.72         -0.71 - 1.12        -1.39 - 2.24        -2.38 - 1.86        -1.78 - 1.92
#> 
#> WEEK 11
#>   n                     4                   4                   1                   1                   3                   3      
#>   Mean (SD)        0.52 (0.66)          1.04 (0.8)          -0.6 (NA)           0.05 (NA)           0.19 (0.3)         0.48 (1.05) 
#>   Median               0.33                1.17                -0.6                0.05                0.09                0.99    
#>   IQR              0.22 - 0.63         0.57 - 1.63         -0.6 - -0.6         0.05 - 0.05         0.02 - 0.31         0.13 - 1.08 
#>   Min - Max        -0.05 - 1.47        0.03 - 1.78         -0.6 - -0.6         0.05 - 0.05         -0.04 - 0.53        -0.73 - 1.16
# NOT RUN {
Viewer(tbl)
# }