Summary Table by Visit

This function summarizes test results or change from baseline statistics by visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.

t_summarize_by_visit(data, visit, id, col_by)

Arguments

data
data frame with numerical variables to be summarized. If the variable has a label attribute then it will be used as the sub-header column name.

visit
factor with visit names ordered by desired display order in the stacked table.

id
unique subject identifier variable.

col_by
factor with reference and comparison group information, the first level indicates the reference group

Value

an rtable object

Details

Every variable in data will be mapped to a sub-column under col_by column. Multiple variables will be shown side-by-side with the variable label (if present) or variable name as the sub-column header.

Currently n, Mean (SD), Median, IQR and Min-Max will be derived for each variable from data per each factor in col_by.

Examples

# EXAMPLE 1
ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B")
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA
ANL$PCHG <- ANL$CHG/ANL$AVAL*100

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

t_summarize_by_visit(data = ANL[c("AVAL")], visit = ANL$VISIT, col_by = ANL$ARM, id = ANL$USUBJID)

#>                       ARM A               ARM B    
#>                      (N=100)             (N=100)   
#> ---------------------------------------------------
#> visit 1
#>   n                    100                 100     
#>   Mean (SD)        -0.08 (1.07)        0.04 (1.04) 
#>   Median              -0.12                0.09    
#>   IQR              -0.82 - 0.61        -0.79 - 0.73
#>   Min - Max        -2.32 - 3.41        -2.06 - 2.83
#> 
#> visit 2
#>   n                    100                 100     
#>   Mean (SD)         -0.02 (1)          0.06 (0.96) 
#>   Median               0.04                0.09    
#>   IQR              -0.69 - 0.73        -0.49 - 0.73
#>   Min - Max        -2.45 - 2.19        -2.36 - 1.94
#> 
#> visit 3
#>   n                    100                 100     
#>   Mean (SD)        -0.14 (1.04)         0.1 (1.03) 
#>   Median               -0.2                0.07    
#>   IQR              -0.83 - 0.56        -0.48 - 0.87
#>   Min - Max        -3.31 - 2.69        -2.98 - 2.16
#> 
#> visit 4
#>   n                    100                 100     
#>   Mean (SD)          0 (0.96)          -0.18 (1.01)
#>   Median               0.09               -0.12    
#>   IQR              -0.73 - 0.68        -0.74 - 0.42
#>   Min - Max        -2.96 - 2.08        -2.76 - 2.24
#> 
#> visit 5
#>   n                    100                 100     
#>   Mean (SD)          0 (1.04)          -0.06 (0.89)
#>   Median               0.02                0.08    
#>   IQR              -0.59 - 0.66        -0.67 - 0.52
#>   Min - Max        -3.4 - 2.28         -2.43 - 2.16
#> 
#> visit 6
#>   n                    100                 100     
#>   Mean (SD)        -0.16 (0.96)        0.03 (0.94) 
#>   Median              -0.13                0.07    
#>   IQR              -0.81 - 0.4         -0.67 - 0.63
#>   Min - Max        -2.65 - 2.72        -2.16 - 2.89
#> 
#> visit 7
#>   n                    100                 100     
#>   Mean (SD)        0.16 (0.97)         0.05 (1.09) 
#>   Median               0.19                0.03    
#>   IQR              -0.61 - 0.84        -0.65 - 0.83
#>   Min - Max        -1.83 - 3.13        -2.14 - 2.68
#> 
#> visit 8
#>   n                    100                 100     
#>   Mean (SD)        -0.04 (1.05)        -0.15 (1.12)
#>   Median               0.04               -0.18    
#>   IQR              -0.76 - 0.56        -0.9 - 0.68 
#>   Min - Max        -2.85 - 2.53        -2.58 - 2.31
#> 
#> visit 9
#>   n                    100                 100     
#>   Mean (SD)        -0.1 (0.89)         -0.12 (0.95)
#>   Median               -0.2               -0.14    
#>   IQR              -0.67 - 0.54        -0.7 - 0.47 
#>   Min - Max        -2.28 - 2.32        -3.15 - 2.39
#> 
#> visit 10
#>   n                    100                 100     
#>   Mean (SD)          0 (0.99)          -0.04 (1.03)
#>   Median              -0.05               -0.01    
#>   IQR              -0.72 - 0.63        -0.64 - 0.63
#>   Min - Max        -2.08 - 2.5         -3.07 - 3.05

t_summarize_by_visit(data = ANL[c("PCHG")], visit = ANL$VISIT, col_by = ANL$ARM, id = ANL$USUBJID)

#>                            ARM A                        ARM B        
#>                           (N=100)                      (N=100)       
#> ---------------------------------------------------------------------
#> visit 1
#>   n                          0                            0          
#>   Mean (SD)               NE (NE)                      NE (NE)       
#>   Median                    NE                           NE          
#>   IQR                     NE - NE                      NE - NE       
#>   Min - Max               NE - NE                      NE - NE       
#> 
#> visit 2
#>   n                         100                          100         
#>   Mean (SD)          -354.2 (2957.79)              -196.72 (1346)    
#>   Median                   51.46                        3.26         
#>   IQR                 -303.93 - 276.9             -279.47 - 324.05   
#>   Min - Max         -26381.14 - 5108.99           -6656.9 - 3835.67  
#> 
#> visit 3
#>   n                         100                          100         
#>   Mean (SD)          -532.97 (7606.86)            -230.4 (4545.38)   
#>   Median                  -30.95                        17.76        
#>   IQR                -239.03 - 255.41             -265.14 - 301.77   
#>   Min - Max        -71648.44 - 23892.48         -18621.05 - 30222.84 
#> 
#> visit 4
#>   n                         100                          100         
#>   Mean (SD)           124.78 (2166.2)            -2167.97 (18165.27) 
#>   Median                   22.25                       -91.04        
#>   IQR                -256.12 - 347.41             -491.82 - 268.12   
#>   Min - Max         -7224.39 - 16987.01         -181084.38 - 2826.88 
#> 
#> visit 5
#>   n                         100                          100         
#>   Mean (SD)          -1028.9 (8131.06)            205.99 (1422.62)   
#>   Median                  -42.04                        58.25        
#>   IQR                -330.48 - 227.84             -204.44 - 450.47   
#>   Min - Max         -80002.49 - 3720.46          -9303.19 - 5644.46  
#> 
#> visit 6
#>   n                         100                          100         
#>   Mean (SD)          -115.85 (1668.83)            256.84 (1518.49)   
#>   Median                  -19.89                        28.71        
#>   IQR                -309.38 - 351.34             -209.15 - 293.85   
#>   Min - Max         -9222.45 - 4796.61           -5533.03 - 9595.66  
#> 
#> visit 7
#>   n                         100                          100         
#>   Mean (SD)           -36.5 (1763.88)             289.85 (2192.48)   
#>   Median                   35.77                        -8.2         
#>   IQR                -198.84 - 344.81             -364.33 - 238.74   
#>   Min - Max         -13466.07 - 8259.17          -2868.87 - 14741.33 
#> 
#> visit 8
#>   n                         100                          100         
#>   Mean (SD)        -70371.59 (703338.08)           48.16 (1394.4)    
#>   Median                   37.91                       -61.53        
#>   IQR                 -293.49 - 279.5             -338.15 - 160.02   
#>   Min - Max        -7033410.58 - 3045.87         -3781.46 - 9583.14  
#> 
#> visit 9
#>   n                         100                          100         
#>   Mean (SD)            3.15 (910.83)               484.8 (4207.21)   
#>   Median                  -108.86                      -16.88        
#>   IQR                -334.55 - 302.74              -271.04 - 427.5   
#>   Min - Max         -2615.49 - 4374.48           -16790.8 - 29284.59 
#> 
#> visit 10
#>   n                         100                          100         
#>   Mean (SD)          1004.93 (9201.73)             94.24 (2009.01)   
#>   Median                   16.03                        62.04        
#>   IQR                -186.37 - 251.78             -312.65 - 370.31   
#>   Min - Max         -9874.98 - 90561.26          -7759.87 - 12104.97 

# Add label to variable instead showing variable name
ANL <- var_relabel(ANL, AVAL = "Value at\nVisit",
                        CHG = "Change from\nBaseline",
                        PCHG = "Percent Change\nfrom Baseline")

t_summarize_by_visit(
  data = ANL[c("AVAL", "CHG")],
  visit = ANL$VISIT,
  col_by = ANL$ARM,
  id = ANL$USUBJID
)

#>                                 ARM A                                   ARM B              
#>                                (N=100)                                 (N=100)             
#>                      Value at          Change from           Value at          Change from 
#>                       Visit              Baseline             Visit              Baseline  
#> -------------------------------------------------------------------------------------------
#> visit 1
#>   n                    100                  0                  100                  0      
#>   Mean (SD)        -0.08 (1.07)          NE (NE)           0.04 (1.04)           NE (NE)   
#>   Median              -0.12                 NE                 0.09                 NE     
#>   IQR              -0.82 - 0.61          NE - NE           -0.79 - 0.73          NE - NE   
#>   Min - Max        -2.32 - 3.41          NE - NE           -2.06 - 2.83          NE - NE   
#> 
#> visit 2
#>   n                    100                 100                 100                 100     
#>   Mean (SD)         -0.02 (1)          1.85 (2.13)         0.06 (0.96)         1.95 (1.87) 
#>   Median               0.04                1.96                0.09                1.73    
#>   IQR              -0.69 - 0.73        0.42 - 3.59         -0.49 - 0.73        0.72 - 3.19 
#>   Min - Max        -2.45 - 2.19        -3.2 - 6.78         -2.36 - 1.94        -3.06 - 6.12
#> 
#> visit 3
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.14 (1.04)        1.88 (2.18)          0.1 (1.03)          2.12 (1.8) 
#>   Median               -0.2                1.73                0.07                1.83    
#>   IQR              -0.83 - 0.56        0.46 - 3.36         -0.48 - 0.87          1 - 3.29  
#>   Min - Max        -3.31 - 2.69        -3.27 - 8.35        -2.98 - 2.16        -2.58 - 8.42
#> 
#> visit 4
#>   n                    100                 100                 100                 100     
#>   Mean (SD)          0 (0.96)          2.06 (2.02)         -0.18 (1.01)         2.01 (1.9) 
#>   Median               0.09                2.25               -0.12                2.02    
#>   IQR              -0.73 - 0.68         0.9 - 3.19         -0.74 - 0.42        0.86 - 2.99 
#>   Min - Max        -2.96 - 2.08        -3.86 - 7.11        -2.76 - 2.24        -3.06 - 8.17
#> 
#> visit 5
#>   n                    100                 100                 100                 100     
#>   Mean (SD)          0 (1.04)          1.71 (2.08)         -0.06 (0.89)        1.86 (2.01) 
#>   Median               0.02                1.71                0.08                1.7     
#>   IQR              -0.59 - 0.66         0.16 - 3.2         -0.67 - 0.52         0.39 - 3.3 
#>   Min - Max        -3.4 - 2.28         -3.56 - 5.96        -2.43 - 2.16        -2.51 - 7.85
#> 
#> visit 6
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.16 (0.96)          1.87 (2)          0.03 (0.94)         1.74 (2.11) 
#>   Median              -0.13                2.08                0.07                1.83    
#>   IQR              -0.81 - 0.4         0.47 - 3.32         -0.67 - 0.63        0.14 - 3.16 
#>   Min - Max        -2.65 - 2.72        -2.79 - 6.59        -2.16 - 2.89        -3.71 - 6.59
#> 
#> visit 7
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.16 (0.97)          1.8 (1.93)         0.05 (1.09)         1.81 (2.17) 
#>   Median               0.19                1.84                0.03                 2      
#>   IQR              -0.61 - 0.84         0.37 - 3.1         -0.65 - 0.83        -0.01 - 3.51
#>   Min - Max        -1.83 - 3.13        -3.15 - 6.06        -2.14 - 2.68        -4.27 - 6.43
#> 
#> visit 8
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.04 (1.05)        1.86 (2.06)         -0.15 (1.12)        1.95 (1.89) 
#>   Median               0.04                1.73               -0.18                1.79    
#>   IQR              -0.76 - 0.56        0.64 - 3.18         -0.9 - 0.68         0.71 - 3.27 
#>   Min - Max        -2.85 - 2.53        -3.1 - 7.06         -2.58 - 2.31        -3.1 - 6.22 
#> 
#> visit 9
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.1 (0.89)         2.09 (2.19)         -0.12 (0.95)         2.1 (2.1)  
#>   Median               -0.2                2.05               -0.14                2.1     
#>   IQR              -0.67 - 0.54         0.5 - 3.42         -0.7 - 0.47         0.63 - 3.54 
#>   Min - Max        -2.28 - 2.32        -3.61 - 8.71        -3.15 - 2.39        -2.86 - 7.64
#> 
#> visit 10
#>   n                    100                 100                 100                 100     
#>   Mean (SD)          0 (0.99)          1.71 (2.15)         -0.04 (1.03)        2.14 (1.86) 
#>   Median              -0.05                1.58               -0.01                2.08    
#>   IQR              -0.72 - 0.63        0.17 - 3.18         -0.64 - 0.63        1.02 - 3.54 
#>   Min - Max        -2.08 - 2.5         -2.81 - 8.08        -3.07 - 3.05        -2.53 - 6.16

# EXAMPLE 2
library(random.cdisc.data)
library(dplyr)

AQS <- radam("AQS") %>%
   filter(PARAMCD == "BFIALL") %>%
   mutate(
     ARM = factor(sample(c("ARM A", "ARM B", "ARM C"), nrow(.), replace = TRUE)),
     AVISIT = factor(AVISIT, levels = unique(AVISIT))
   )

tbl <- t_summarize_by_visit(
  data = AQS[c("AVAL", "CHG")],
  visit = AQS$AVISIT,
  col_by = AQS$ARM,
  id = AQS$USUBJID
)
tbl

#>                                  ARM A                                     ARM B                                     ARM C               
#>                                  (N=89)                                    (N=82)                                    (N=91)              
#>                        AVAL                  CHG                 AVAL                  CHG                 AVAL                  CHG     
#> -----------------------------------------------------------------------------------------------------------------------------------------
#> WEEK 1
#>   n                     36                   36                   28                   28                   36                   36      
#>   Mean (SD)        -0.34 (1.03)             0 (0)             0.03 (0.85)             0 (0)             0.2 (0.83)              0 (0)    
#>   Median               -0.31                  0                  -0.02                  0                  0.39                   0      
#>   IQR              -1.04 - 0.51             0 - 0            -0.38 - 0.63             0 - 0            -0.49 - 0.81             0 - 0    
#>   Min - Max        -3.25 - 1.91             0 - 0            -2.37 - 1.94             0 - 0             -1.7 - 1.45             0 - 0    
#> 
#> WEEK 2
#>   n                     32                   32                   27                   27                   41                   41      
#>   Mean (SD)         0.01 (0.9)           0.09 (1.33)          0.11 (1.01)          0.01 (1.5)          -0.06 (0.93)          0.04 (1.45) 
#>   Median               -0.03                  0                  -0.02                -0.13                0.05                 -0.11    
#>   IQR               -0.5 - 0.46         -0.98 - 1.12         -0.46 - 0.82         -0.98 - 1.09          -0.84 - 0.5           -1 - 0.79  
#>   Min - Max        -1.81 - 2.62         -1.89 - 2.77         -2.09 - 1.68         -2.88 - 3.17          -2.1 - 1.94         -2.57 - 3.45 
#> 
#> WEEK 3
#>   n                     26                   26                   25                   25                   42                   42      
#>   Mean (SD)         0.01 (1.12)         -0.15 (1.32)          -0.05 (1.1)          -0.1 (1.61)          0.25 (1.07)          0.38 (1.64) 
#>   Median               0.02                 -0.05                -0.31                0.02                  0.2                 0.46     
#>   IQR              -0.79 - 0.96         -0.84 - 0.53         -0.87 - 0.56         -1.11 - 0.42         -0.52 - 0.95         -0.59 - 1.32 
#>   Min - Max        -1.97 - 2.14          -3.4 - 2.62         -1.39 - 2.73         -2.81 - 3.33         -2.65 - 2.49         -3.26 - 4.41 
#> 
#> WEEK 4
#>   n                     28                   28                   26                   26                   32                   32      
#>   Mean (SD)        -0.02 (0.78)          0.06 (1.12)         -0.05 (1.15)          0.01 (1.55)          0.08 (1.01)          0.02 (1.5)  
#>   Median               -0.05                0.27                 -0.15                -0.19                0.15                 0.27     
#>   IQR              -0.49 - 0.25         -0.61 - 0.71         -0.89 - 0.78         -0.74 - 0.75         -0.39 - 0.75         -1.18 - 1.05 
#>   Min - Max        -1.33 - 2.38         -2.63 - 1.92         -1.81 - 2.46         -2.26 - 4.61         -2.05 - 1.84         -2.86 - 3.19 
#> 
#> WEEK 5
#>   n                     25                   25                   26                   26                   26                   26      
#>   Mean (SD)         0.06 (1.03)          0.31 (1.29)          -0.34 (1.1)         -0.14 (1.34)           0.09 (1)           -0.17 (1.41) 
#>   Median               0.35                 0.79                 -0.44                -0.09                0.22                 0.12     
#>   IQR              -0.63 - 0.59         -0.37 - 1.21         -1.03 - 0.49         -0.94 - 0.75          -0.34 - 0.8         -0.66 - 0.43 
#>   Min - Max        -2.15 - 1.97         -2.93 - 2.62         -2.55 - 1.93         -3.57 - 2.94         -2.63 - 2.03         -4.05 - 2.48 
#> 
#> WEEK 6
#>   n                     20                   20                   21                   21                   21                   21      
#>   Mean (SD)        -0.19 (1.12)          0.13 (1.45)          -0.2 (1.15)         -0.43 (1.56)         -0.23 (0.78)         -0.44 (0.99) 
#>   Median               -0.2                 0.08                 -0.61                -0.32                -0.36                -0.29    
#>   IQR              -1.01 - 0.83          -0.57 - 1.3         -1.07 - 0.74          -1.97 - 0.9          -0.6 - 0.18           -1 - 0.15  
#>   Min - Max        -2.36 - 1.42          -2.46 - 2.3         -1.75 - 2.08         -2.55 - 1.88          -1.51 - 1.4         -2.47 - 1.48 
#> 
#> WEEK 7
#>   n                     13                   13                   21                   21                   21                   21      
#>   Mean (SD)        -0.25 (1.12)         -0.37 (1.18)         -0.01 (0.79)         -0.03 (1.23)          0.34 (1.05)          0.29 (1.5)  
#>   Median               -0.08                -0.91                0.07                 0.26                  0.7                 0.49     
#>   IQR               -0.91 - 0.2         -1.17 - 0.25         -0.47 - 0.39          -0.7 - 0.46         -0.52 - 0.97         -0.76 - 1.21 
#>   Min - Max        -1.93 - 2.01         -1.76 - 2.34         -1.79 - 1.32         -2.47 - 2.64         -1.46 - 2.31         -2.68 - 3.84 
#> 
#> WEEK 8
#>   n                     19                   19                   12                   12                   16                   16      
#>   Mean (SD)          0.03 (1)            0.18 (1.12)         -0.45 (0.77)         -0.47 (1.88)          0.01 (0.93)         -0.05 (1.14) 
#>   Median               0.17                 0.43                 -0.42                -0.22                -0.09                0.16     
#>   IQR               -0.51 - 0.6         -0.53 - 0.84         -0.88 - 0.03         -1.72 - 0.88          -0.48 - 0.6         -0.76 - 0.72 
#>   Min - Max        -2.12 - 2.02         -2.12 - 1.99         -2.04 - 0.65         -3.98 - 1.99          -1.61 - 1.8         -2.16 - 1.53 
#> 
#> WEEK 9
#>   n                     11                   11                   15                   15                   12                   12      
#>   Mean (SD)         0.39 (0.77)          0.92 (1.02)          0.25 (1.01)         -0.03 (1.26)         -0.08 (0.99)          -0.33 (1.5) 
#>   Median               0.43                 1.15                 0.24                 -0.02                -0.06                -0.17    
#>   IQR               0.17 - 0.76          0.58 - 1.5          -0.51 - 0.94         -0.58 - 0.55         -0.71 - 0.51          -0.98 - 0.2 
#>   Min - Max        -1.35 - 1.63           -1.62 - 2          -1.73 - 1.95         -2.19 - 2.54         -1.89 - 1.75         -3.31 - 2.37 
#> 
#> WEEK 10
#>   n                      8                    8                    9                    9                    6                    6      
#>   Mean (SD)         0.09 (0.36)          0.45 (0.88)           0.61 (1)            0.25 (1.29)         -0.49 (0.56)         -0.99 (1.48) 
#>   Median               0.17                 0.37                 0.16                 -0.32                -0.68                -0.76    
#>   IQR              -0.16 - 0.41          0.03 - 0.84         -0.09 - 1.38          -0.62 - 0.9         -0.91 - -0.19        -2.13 - 0.16 
#>   Min - Max        -0.55 - 0.44         -0.81 - 1.95         -0.51 - 2.16          -1.24 - 2.4           -1 - 0.41          -2.91 - 0.59 
#> 
#> WEEK 11
#>   n                      7                    7                    6                    6                    0                    0      
#>   Mean (SD)         0.25 (1.24)          0.76 (1.31)          0.01 (0.84)          -0.36 (0.9)            NE (NE)              NE (NE)   
#>   Median               0.27                  0.9                 0.08                 -0.29                 NE                   NE      
#>   IQR              -0.25 - 0.91          -0.02 - 1.5         -0.31 - 0.68         -0.69 - 0.33            NE - NE              NE - NE   
#>   Min - Max        -1.88 - 2.01         -1.21 - 2.68         -1.39 - 0.85         -1.87 - 0.55            NE - NE              NE - NE   
# NOT RUN {
Viewer(tbl)
# }