Summary Table by Visit

This function summarizes test results or change from baseline statistics by visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.

t_summarize_by_visit(data, visit, id, col_by, col_N)

Arguments

data
data frame with numerical variables to be summarized. If the variable has a label attribute then it will be used as the sub-header column name.

visit
factor with visit names ordered by desired display order in the stacked table.

id
unique subject identifier variable.

col_by
factor with reference and comparison group information, the first level indicates the reference group

col_N
a table object with the reference population used for the header of the table.

Value

an rtable object

Details

Every variable in data will be mapped to a sub-column under col_by column. Multiple variables will be shown side-by-side with the variable label (if present) or variable name as the sub-column header.

Currently n, Mean (SD), Median, IQR and Min-Max will be derived for each variable from data per each factor in col_by.

Examples

# EXAMPLE 1
ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B")
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA
ANL$PCHG <- ANL$CHG/ANL$AVAL*100

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

ASL <- unique(ANL[, c("USUBJID","ARM")])

ADSL <- radsl()
ADVS <- radvs(ADSL)

t_summarize_by_visit(data = ADVS[c("AVAL")], visit = ADVS$AVISIT, col_by = ADVS$ARM,
  id = ADVS$USUBJID, col_N = table(ADSL$ARM))

#>                        A: Drug X             B: Placebo          C: Combination
#>                         (N=127)               (N=134)               (N=139)    
#> -------------------------------------------------------------------------------
#> SCREENING
#>   n                       762                   804                   834      
#>   Mean (SD)           50.22 (7.48)          49.56 (7.99)          49.53 (7.85) 
#>   Median                 50.29                 49.62                 49.63     
#>   IQR                45.23 - 55.12         43.87 - 55.34         44.49 - 55.05 
#>   Min - Max          28.07 - 75.64         27.34 - 73.41         18.11 - 73.16 
#> 
#> BASELINE
#>   n                       762                   804                   834      
#>   Mean (SD)           50.33 (7.97)          50.39 (7.93)          50.63 (7.79) 
#>   Median                 50.66                  50.2                 50.88     
#>   IQR                45.38 - 55.67          45.1 - 56.11         45.44 - 55.67 
#>   Min - Max          23.41 - 72.07         22.38 - 75.33          24.12 - 75.3 
#> 
#> WEEK 1 DAY 8
#>   n                       762                   804                   834      
#>   Mean (SD)           50.25 (8.01)          49.64 (7.77)          50.13 (7.91) 
#>   Median                 49.95                 49.55                 50.14     
#>   IQR                44.95 - 55.94         44.31 - 54.78         44.94 - 55.21 
#>   Min - Max           30.7 - 74.75         28.68 - 70.98          19.8 - 74.25 
#> 
#> WEEK 2 DAY 15
#>   n                       762                   804                   834      
#>   Mean (SD)           50.19 (7.8)           49.98 (8.06)          50.11 (8.09) 
#>   Median                 50.33                 49.89                 50.27     
#>   IQR                44.86 - 55.77          44.69 - 55.4         44.83 - 55.53 
#>   Min - Max          25.01 - 74.93         23.38 - 75.24         23.56 - 72.55 
#> 
#> WEEK 3 DAY 22
#>   n                       762                   804                   834      
#>   Mean (SD)           49.98 (7.95)          50.44 (7.75)          50.25 (7.91) 
#>   Median                 50.34                 50.29                 50.29     
#>   IQR                44.95 - 55.73         44.91 - 55.74         44.48 - 55.82 
#>   Min - Max          26.78 - 71.22         27.04 - 72.21          25.73 - 73.9 
#> 
#> WEEK 4 DAY 29
#>   n                       762                   804                   834      
#>   Mean (SD)           50.18 (8.05)          49.54 (7.85)          50.22 (8.03) 
#>   Median                 49.83                 49.53                 50.56     
#>   IQR                45.13 - 55.61         44.26 - 54.56         44.44 - 55.48 
#>   Min - Max          25.85 - 70.05         24.21 - 75.69         28.13 - 77.23 
#> 
#> WEEK 5 DAY 36
#>   n                       762                   804                   834      
#>   Mean (SD)           49.81 (8.06)          49.45 (7.7)            49.86 (8)   
#>   Median                 49.74                 49.36                 49.77     
#>   IQR                44.59 - 54.92         44.61 - 54.52          44.2 - 55.29 
#>   Min - Max          25.12 - 76.84         26.02 - 71.57          28.7 - 76.65 

t_summarize_by_visit(data = ADVS[c("PCHG")], visit = ADVS$AVISIT, col_by = ADVS$ARM,
  id = ADVS$USUBJID, col_N = table(ADSL$ARM))

#>                         A: Drug X             B: Placebo           C: Combination 
#>                          (N=127)                (N=134)                (N=139)    
#> ----------------------------------------------------------------------------------
#> SCREENING
#>   n                         0                      0                      0       
#>   Mean (SD)              NE (NE)                NE (NE)                NE (NE)    
#>   Median                   NE                     NE                     NE       
#>   IQR                    NE - NE                NE - NE                NE - NE    
#>   Min - Max              NE - NE                NE - NE                NE - NE    
#> 
#> BASELINE
#>   n                        762                    804                    834      
#>   Mean (SD)               0 (0)                  0 (0)                  0 (0)     
#>   Median                    0                      0                      0       
#>   IQR                     0 - 0                  0 - 0                  0 - 0     
#>   Min - Max               0 - 0                  0 - 0                  0 - 0     
#> 
#> WEEK 1 DAY 8
#>   n                        762                    804                    834      
#>   Mean (SD)           2.61 (24.74)           0.91 (22.44)           1.55 (23.56)  
#>   Median                  -0.56                  -1.79                  -1.42     
#>   IQR                -13.83 - 16.56         -15.43 - 14.76         -14.48 - 14.39 
#>   Min - Max          -55.84 - 130.23        -50.25 - 94.57         -73.16 - 132.17
#> 
#> WEEK 2 DAY 15
#>   n                        762                    804                    834      
#>   Mean (SD)           2.56 (24.41)            1.67 (23.4)           1.54 (23.81)  
#>   Median                  0.51                   -0.34                  -1.36     
#>   IQR                -13.52 - 15.27         -14.96 - 16.07         -15.63 - 16.05 
#>   Min - Max          -53.36 - 127.41        -50.18 - 175.78        -60.25 - 130.42
#> 
#> WEEK 3 DAY 22
#>   n                        762                    804                    834      
#>   Mean (SD)           2.02 (24.24)           2.83 (23.99)            1.84 (23.9)  
#>   Median                  -0.9                   0.17                   -0.66     
#>   IQR                -14.79 - 15.68         -15.07 - 16.79         -13.97 - 15.08 
#>   Min - Max          -57.93 - 151.36        -50.46 - 113.15        -54.87 - 155.9 
#> 
#> WEEK 4 DAY 29
#>   n                        762                    804                    834      
#>   Mean (SD)           2.67 (25.55)           0.91 (23.76)           1.58 (22.99)  
#>   Median                  -0.99                  -1.54                  -0.79     
#>   IQR                 -15.06 - 16.3         -16.17 - 14.21         -14.84 - 15.62 
#>   Min - Max          -56.01 - 153.96        -56.96 - 181.74        -52.36 - 153.21
#> 
#> WEEK 5 DAY 36
#>   n                        762                    804                    834      
#>   Mean (SD)           1.61 (23.84)           0.65 (23.06)           1.01 (23.48)  
#>   Median                  -0.44                  -1.89                  -2.12     
#>   IQR                -13.77 - 14.09         -15.13 - 13.33         -15.04 - 14.95 
#>   Min - Max          -52.26 - 112.66        -49.31 - 161.86        -56.71 - 121.88

# DO NOT THINK WE NEED THIS BLOCK AS LABELS ALREADY AVAILABLE
# ADDED THIS NOTE GIVEN FINDING OF ATTRIBUTES BEING REMOVED
#Add label to variable instead showing variable name
#ANL <- var_relabel(ANL, AVAL = "Value at\nVisit",
 #                       CHG = "Change from\nBaseline",
  #                      PCHG = "Percent Change\nfrom Baseline")

t_summarize_by_visit(
  data = ADVS[c("AVAL", "CHG")],
  visit = ADVS$AVISIT,
  col_by = ADVS$ARM,
  id = ADVS$USUBJID,
  col_N = table(ADSL$ARM)
)

#>                                                  A: Drug X                                                                 B: Placebo                                                              C: Combination                          
#>                                                   (N=127)                                                                   (N=134)                                                                   (N=139)                              
#>                             Analysis Value                Absolute Change from Baseline               Analysis Value                Absolute Change from Baseline               Analysis Value                Absolute Change from Baseline
#> -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> SCREENING
#>   n                               762                                   0                                   804                                   0                                   834                                   0              
#>   Mean (SD)                  50.22 (7.48)                            NE (NE)                           49.56 (7.99)                            NE (NE)                           49.53 (7.85)                            NE (NE)           
#>   Median                         50.29                                 NE                                  49.62                                 NE                                  49.63                                 NE              
#>   IQR                        45.23 - 55.12                           NE - NE                           43.87 - 55.34                           NE - NE                           44.49 - 55.05                           NE - NE           
#>   Min - Max                  28.07 - 75.64                           NE - NE                           27.34 - 73.41                           NE - NE                           18.11 - 73.16                           NE - NE           
#> 
#> BASELINE
#>   n                               762                                  762                                  804                                  804                                  834                                  834             
#>   Mean (SD)                  50.33 (7.97)                             0 (0)                            50.39 (7.93)                             0 (0)                            50.63 (7.79)                             0 (0)            
#>   Median                         50.66                                  0                                  50.2                                   0                                  50.88                                  0              
#>   IQR                        45.38 - 55.67                            0 - 0                            45.1 - 56.11                             0 - 0                            45.44 - 55.67                            0 - 0            
#>   Min - Max                  23.41 - 72.07                            0 - 0                            22.38 - 75.33                            0 - 0                            24.12 - 75.3                             0 - 0            
#> 
#> WEEK 1 DAY 8
#>   n                               762                                  762                                  804                                  804                                  834                                  834             
#>   Mean (SD)                  50.25 (8.01)                         -0.08 (11.11)                        49.64 (7.77)                         -0.75 (10.61)                        50.13 (7.91)                         -0.5 (11.04)         
#>   Median                         49.95                                -0.27                                49.55                                -0.91                                50.14                                -0.72            
#>   IQR                        44.95 - 55.94                         -7.7 - 7.47                         44.31 - 54.78                         -8.2 - 6.84                         44.94 - 55.21                        -7.95 - 6.82         
#>   Min - Max                  30.7 - 74.75                        -39.97 - 33.85                        28.68 - 70.98                       -33.36 - 29.89                        19.8 - 74.25                         -53.97 - 36.5        
#> 
#> WEEK 2 DAY 15
#>   n                               762                                  762                                  804                                  804                                  834                                  834             
#>   Mean (SD)                   50.19 (7.8)                         -0.14 (11.23)                        49.98 (8.06)                         -0.41 (10.97)                        50.11 (8.09)                         -0.52 (11.28)        
#>   Median                         50.33                                0.26                                 49.89                                -0.18                                50.27                                -0.67            
#>   IQR                        44.86 - 55.77                        -7.38 - 7.26                         44.69 - 55.4                          -7.91 - 7.3                         44.83 - 55.53                        -8.17 - 7.45         
#>   Min - Max                  25.01 - 74.93                       -32.16 - 37.28                        23.38 - 75.24                        -31.3 - 39.33                        23.56 - 72.55                       -36.74 - 37.18        
#> 
#> WEEK 3 DAY 22
#>   n                               762                                  762                                  804                                  804                                  834                                  834             
#>   Mean (SD)                  49.98 (7.95)                         -0.35 (11.1)                         50.44 (7.75)                         0.05 (11.17)                         50.25 (7.91)                         -0.38 (11.17)        
#>   Median                         50.34                                -0.48                                50.29                                0.08                                 50.29                                -0.31            
#>   IQR                        44.95 - 55.73                        -7.95 - 7.23                         44.91 - 55.74                        -8.21 - 7.83                         44.48 - 55.82                        -7.68 - 7.05         
#>   Min - Max                  26.78 - 71.22                       -36.88 - 35.44                        27.04 - 72.21                       -30.45 - 33.75                        25.73 - 73.9                        -38.88 - 37.61        
#> 
#> WEEK 4 DAY 29
#>   n                               762                                  762                                  804                                  804                                  834                                  834             
#>   Mean (SD)                  50.18 (8.05)                         -0.15 (11.6)                         49.54 (7.85)                         -0.85 (11.02)                        50.22 (8.03)                         -0.41 (10.81)        
#>   Median                         49.83                                -0.52                                49.53                                -0.77                                50.56                                -0.38            
#>   IQR                        45.13 - 55.61                        -8.05 - 7.69                         44.26 - 54.56                        -8.74 - 6.62                         44.44 - 55.48                        -8.12 - 7.42         
#>   Min - Max                  25.85 - 70.05                       -32.92 - 36.05                        24.21 - 75.69                       -32.03 - 40.67                        28.13 - 77.23                       -34.82 - 36.96        
#> 
#> WEEK 5 DAY 36
#>   n                               762                                  762                                  804                                  804                                  834                                  834             
#>   Mean (SD)                  49.81 (8.06)                         -0.52 (11.04)                         49.45 (7.7)                         -0.94 (10.77)                          49.86 (8)                          -0.77 (11.26)        
#>   Median                         49.74                                -0.23                                49.36                                -0.98                                49.77                                -1.01            
#>   IQR                        44.59 - 54.92                         -7.16 - 6.6                         44.61 - 54.52                        -8.52 - 6.27                         44.2 - 55.29                          -8.1 - 6.98         
#>   Min - Max                  25.12 - 76.84                       -36.48 - 30.43                        26.02 - 71.57                       -34.21 - 36.22                        28.7 - 76.65                         -38.47 - 37.4        

# EXAMPLE 2
library(random.cdisc.data)
library(dplyr)

ADSL <- radsl()
ADQS <- radqs(ADSL) %>%
   filter(PARAMCD == "BFIALL")

tbl <- t_summarize_by_visit(
  data = ADQS[c("AVAL", "CHG")],
  visit = ADQS$AVISIT,
  col_by = ADQS$ARM,
  id = ADQS$USUBJID,
  col_N = table(ADSL$ARM)
)
tbl

#>                                                  A: Drug X                                                                 B: Placebo                                                              C: Combination                          
#>                                                   (N=125)                                                                   (N=136)                                                                   (N=139)                              
#>                             Analysis Value                Absolute Change from Baseline               Analysis Value                Absolute Change from Baseline               Analysis Value                Absolute Change from Baseline
#> -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> SCREENING
#>   n                               125                                   0                                   136                                   0                                   139                                   0              
#>   Mean (SD)                  49.99 (8.54)                            NE (NE)                           49.42 (8.32)                            NE (NE)                            50.48 (8.6)                            NE (NE)           
#>   Median                         48.77                                 NE                                  48.86                                 NE                                  50.65                                 NE              
#>   IQR                        44.3 - 54.89                            NE - NE                           42.78 - 55.67                           NE - NE                            44 - 56.29                             NE - NE           
#>   Min - Max                  32.72 - 78.89                           NE - NE                           29.71 - 72.95                           NE - NE                           27.95 - 71.84                           NE - NE           
#> 
#> BASELINE
#>   n                               125                                  125                                  136                                  136                                  139                                  139             
#>   Mean (SD)                  49.97 (8.05)                             0 (0)                            49.34 (7.21)                             0 (0)                            50.29 (7.93)                             0 (0)            
#>   Median                         50.2                                   0                                  49.79                                  0                                  50.48                                  0              
#>   IQR                        44.21 - 54.91                            0 - 0                            44.03 - 53.77                            0 - 0                            45.93 - 54.23                            0 - 0            
#>   Min - Max                  31.17 - 71.51                            0 - 0                            31.43 - 67.16                            0 - 0                            25.33 - 75.25                            0 - 0            
#> 
#> WEEK 1 DAY 8
#>   n                               125                                  125                                  136                                  136                                  139                                  139             
#>   Mean (SD)                  50.28 (7.56)                         0.31 (10.82)                         51.24 (8.02)                         1.91 (10.58)                         50.65 (8.06)                         0.36 (11.49)         
#>   Median                         51.17                                -1.15                                51.29                                1.64                                 50.4                                 -0.27            
#>   IQR                        44.39 - 54.79                        -6.86 - 8.45                         45.75 - 56.43                        -4.26 - 8.87                         45.33 - 56.01                        -7.01 - 8.07         
#>   Min - Max                  35.89 - 69.5                        -22.34 - 29.36                        27.19 - 69.36                        -21.5 - 26.33                        27.93 - 77.04                       -32.39 - 27.45        
#> 
#> WEEK 2 DAY 15
#>   n                               125                                  125                                  136                                  136                                  139                                  139             
#>   Mean (SD)                  51.38 (7.98)                         1.41 (11.04)                         50.37 (8.58)                         1.04 (11.62)                         49.84 (7.71)                         -0.46 (10.42)        
#>   Median                         51.99                                1.97                                 51.66                                2.46                                 50.25                                -0.06            
#>   IQR                        46.32 - 56.54                        -5.27 - 8.05                         44.25 - 56.67                        -6.16 - 9.31                         44.91 - 55.74                        -8.01 - 5.87         
#>   Min - Max                  28.31 - 71.77                        -29.8 - 27.73                        22.77 - 69.37                       -35.29 - 24.44                        28.18 - 63.96                       -25.97 - 31.32        
#> 
#> WEEK 3 DAY 22
#>   n                               125                                  125                                  136                                  136                                  139                                  139             
#>   Mean (SD)                  50.96 (7.62)                         0.99 (10.95)                          50.2 (8.39)                         0.87 (10.14)                         50.04 (8.08)                         -0.25 (10.78)        
#>   Median                         51.11                                0.12                                 49.8                                  0.6                                 50.28                                0.06             
#>   IQR                        46.03 - 56.13                          -7 - 8.4                           44.81 - 56.46                         -5.2 - 7.74                         44.98 - 55.1                         -7.03 - 6.44         
#>   Min - Max                  29.61 - 71.08                       -29.95 - 26.97                        26.2 - 73.51                        -24.08 - 21.76                        26.06 - 69.62                       -28.29 - 27.28        
#> 
#> WEEK 4 DAY 29
#>   n                               125                                  125                                  136                                  136                                  139                                  139             
#>   Mean (SD)                   50.3 (7.92)                         0.33 (11.46)                          49.09 (7.7)                         -0.24 (9.82)                         51.03 (8.26)                         0.73 (11.77)         
#>   Median                         50.23                                0.38                                 49.37                                -0.59                                51.36                                1.35             
#>   IQR                        46.3 - 55.26                         -8.15 - 7.34                         43.45 - 54.23                        -6.82 - 5.79                         46.05 - 55.99                        -7.93 - 10.53        
#>   Min - Max                  28.09 - 76.32                       -24.79 - 29.84                        26.63 - 66.9                         -24.28 - 26.2                        31.39 - 70.92                       -26.52 - 25.93        
#> 
#> WEEK 5 DAY 36
#>   n                               125                                  125                                  136                                  136                                  139                                  139             
#>   Mean (SD)                   50.15 (7.7)                         0.17 (11.03)                         49.71 (8.22)                         0.38 (10.89)                         49.29 (7.84)                         -1.01 (10.85)        
#>   Median                         49.86                                0.35                                 49.45                                0.88                                 48.98                                -0.55            
#>   IQR                        45.68 - 56.47                        -6.11 - 8.02                         44.66 - 55.63                        -6.82 - 7.68                         43.32 - 55.22                        -9.01 - 6.36         
#>   Min - Max                  29.41 - 65.29                        -38.5 - 22.66                        27.28 - 70.88                       -29.61 - 28.49                        26.22 - 65.73                       -31.62 - 29.57        
# NOT RUN {
Viewer(tbl)
# }