Summary Table by Visit

This function summarizes test results or change from baseline statistics by visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.

t_summarize_by_visit(data, visit, id, col_by)

Arguments

data
data frame with numerical variables to be summarized. If the variable has a label attribute then it will be used as the sub-header column name.

visit
factor with visit names ordered by desired display order in the stacked table.

id
unique subject identifier variable.

col_by
factor with reference and comparison group information, the first level indicates the reference group

Value

an rtable object

Details

Every variable in data will be mapped to a sub-column under col_by column. Multiple variables will be shown side-by-side with the variable label (if present) or variable name as the sub-column header.

Currently n, Mean (SD), Median, IQR and Min-Max will be derived for each variable from data per each factor in col_by.

Examples

# EXAMPLE 1
ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B")
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA
ANL$PCHG <- ANL$CHG/ANL$AVAL*100

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

t_summarize_by_visit(data = ANL[c("AVAL")], visit = ANL$VISIT, col_by = ANL$ARM, id = ANL$USUBJID)

#>                       ARM A               ARM B    
#>                      (N=100)             (N=100)   
#> ---------------------------------------------------
#> visit 1
#>   n                    100                 100     
#>   Mean (SD)        -0.07 (1.12)        0.06 (1.04) 
#>   Median                0                  0.01    
#>   IQR              -0.93 - 0.69        -0.74 - 0.73
#>   Min - Max        -2.47 - 3.41        -2.06 - 2.83
#> 
#> visit 2
#>   n                    100                 100     
#>   Mean (SD)        -0.09 (1.01)        0.03 (0.97) 
#>   Median              -0.12                0.07    
#>   IQR              -0.8 - 0.63         -0.74 - 0.73
#>   Min - Max        -2.45 - 2.19        -2.36 - 1.94
#> 
#> visit 3
#>   n                    100                 100     
#>   Mean (SD)        -0.11 (1.06)        0.13 (0.98) 
#>   Median               -0.2                0.09    
#>   IQR              -0.83 - 0.55        -0.44 - 0.84
#>   Min - Max        -3.31 - 2.69        -2.98 - 2.16
#> 
#> visit 4
#>   n                    100                 100     
#>   Mean (SD)        0.05 (0.94)         -0.15 (1.02)
#>   Median               0.13                -0.1    
#>   IQR              -0.54 - 0.66        -0.74 - 0.49
#>   Min - Max        -2.96 - 2.08        -2.76 - 2.24
#> 
#> visit 5
#>   n                    100                 100     
#>   Mean (SD)        -0.02 (1.01)        -0.08 (0.94)
#>   Median              -0.08                0.06    
#>   IQR              -0.63 - 0.63        -0.72 - 0.5 
#>   Min - Max        -2.23 - 2.28        -2.58 - 2.16
#> 
#> visit 6
#>   n                    100                 100     
#>   Mean (SD)         -0.14 (1)          0.03 (0.89) 
#>   Median              -0.13                0.03    
#>   IQR              -0.81 - 0.46        -0.58 - 0.59
#>   Min - Max        -3.4 - 2.72         -2.16 - 2.89
#> 
#> visit 7
#>   n                    100                 100     
#>   Mean (SD)         0.1 (0.95)         0.09 (1.07) 
#>   Median               0.18                0.03    
#>   IQR              -0.56 - 0.77        -0.65 - 0.89
#>   Min - Max        -2.16 - 2.26        -2.08 - 2.68
#> 
#> visit 8
#>   n                    100                 100     
#>   Mean (SD)        -0.07 (1.08)        -0.27 (1.1) 
#>   Median              -0.07               -0.28    
#>   IQR              -0.79 - 0.52        -0.99 - 0.65
#>   Min - Max        -2.85 - 3.13        -2.58 - 2.31
#> 
#> visit 9
#>   n                    100                 100     
#>   Mean (SD)        -0.05 (0.87)        -0.03 (0.98)
#>   Median              -0.11                0.02    
#>   IQR              -0.63 - 0.57        -0.65 - 0.5 
#>   Min - Max        -2.28 - 2.19        -3.15 - 2.39
#> 
#> visit 10
#>   n                    100                 100     
#>   Mean (SD)        -0.02 (1.01)        -0.05 (1.02)
#>   Median              -0.07                 0      
#>   IQR              -0.83 - 0.7         -0.64 - 0.62
#>   Min - Max        -2.08 - 2.5         -3.07 - 3.05

t_summarize_by_visit(data = ANL[c("PCHG")], visit = ANL$VISIT, col_by = ANL$ARM, id = ANL$USUBJID)

#>                            ARM A                        ARM B        
#>                           (N=100)                      (N=100)       
#> ---------------------------------------------------------------------
#> visit 1
#>   n                          0                            0          
#>   Mean (SD)               NE (NE)                      NE (NE)       
#>   Median                    NE                           NE          
#>   IQR                     NE - NE                      NE - NE       
#>   Min - Max               NE - NE                      NE - NE       
#> 
#> visit 2
#>   n                         100                          100         
#>   Mean (SD)          -288.31 (2863.3)             -197.68 (1239.35)  
#>   Median                   51.46                        -7.37        
#>   IQR                -303.93 - 296.99             -283.61 - 348.84   
#>   Min - Max         -26381.14 - 3237.73           -6656.9 - 2285.27  
#> 
#> visit 3
#>   n                         100                          100         
#>   Mean (SD)          124.52 (2617.79)             -271.94 (4553.28)  
#>   Median                  -61.03                        75.81        
#>   IQR                 -239.6 - 238.05             -326.07 - 329.75   
#>   Min - Max         -5182.95 - 23892.48         -18621.05 - 30222.84 
#> 
#> visit 4
#>   n                         100                          100         
#>   Mean (SD)          -499.63 (7471.85)           -2070.36 (18176.53) 
#>   Median                   81.03                       -76.46        
#>   IQR                -226.02 - 468.95              -328.2 - 268.12   
#>   Min - Max        -71648.44 - 16987.01         -181084.38 - 5069.25 
#> 
#> visit 5
#>   n                         100                          100         
#>   Mean (SD)          -1140.65 (8150.1)            206.27 (1178.12)   
#>   Median                  -77.97                        6.26         
#>   IQR                -330.48 - 202.54              -215.6 - 439.63   
#>   Min - Max         -80002.49 - 3720.46          -3385.45 - 5644.46  
#> 
#> visit 6
#>   n                         100                          100         
#>   Mean (SD)          -87.26 (1640.54)             161.84 (1574.52)   
#>   Median                  -19.12                        28.71        
#>   IQR                -356.29 - 421.49             -259.81 - 293.85   
#>   Min - Max         -9222.45 - 4796.61           -9303.19 - 9595.66  
#> 
#> visit 7
#>   n                         100                          100         
#>   Mean (SD)          -142.98 (1579.23)            291.19 (2353.87)   
#>   Median                   35.77                       -16.57        
#>   IQR                -191.74 - 344.81             -356.66 - 244.42   
#>   Min - Max         -13466.07 - 1910.46          -5533.03 - 14741.33 
#> 
#> visit 8
#>   n                         100                          100         
#>   Mean (SD)        -70341.91 (703341.53)          -23.08 (1018.85)   
#>   Median                  -34.93                       -49.66        
#>   IQR                 -321.72 - 227.2             -332.58 - 131.74   
#>   Min - Max        -7033410.58 - 8259.17         -3781.46 - 5075.53  
#> 
#> visit 9
#>   n                         100                          100         
#>   Mean (SD)           72.59 (907.27)              570.73 (4309.02)   
#>   Median                  -35.19                       -38.71        
#>   IQR                -263.28 - 375.32              -325.9 - 427.5    
#>   Min - Max         -2615.49 - 4374.48           -16790.8 - 29284.59 
#> 
#> visit 10
#>   n                         100                          100         
#>   Mean (SD)          1002.82 (9191.85)             45.26 (1532.88)   
#>   Median                   -0.01                       111.33        
#>   IQR                -217.21 - 233.69             -250.75 - 370.31   
#>   Min - Max         -9874.98 - 90561.26          -7759.87 - 10236.52 

# Add label to variable instead showing variable name
ANL <- var_relabel(ANL, AVAL = "Value at\nVisit",
                        CHG = "Change from\nBaseline",
                        PCHG = "Percent Change\nfrom Baseline")

t_summarize_by_visit(
  data = ANL[c("AVAL", "CHG")],
  visit = ANL$VISIT,
  col_by = ANL$ARM,
  id = ANL$USUBJID
)

#>                                 ARM A                                   ARM B              
#>                                (N=100)                                 (N=100)             
#>                      Value at          Change from           Value at          Change from 
#>                       Visit              Baseline             Visit              Baseline  
#> -------------------------------------------------------------------------------------------
#> visit 1
#>   n                    100                  0                  100                  0      
#>   Mean (SD)        -0.07 (1.12)          NE (NE)           0.06 (1.04)           NE (NE)   
#>   Median                0                   NE                 0.01                 NE     
#>   IQR              -0.93 - 0.69          NE - NE           -0.74 - 0.73          NE - NE   
#>   Min - Max        -2.47 - 3.41          NE - NE           -2.06 - 2.83          NE - NE   
#> 
#> visit 2
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.09 (1.01)        2.01 (2.11)         0.03 (0.97)          2.1 (1.83) 
#>   Median              -0.12                2.05                0.07                2.16    
#>   IQR              -0.8 - 0.63         0.64 - 3.66         -0.74 - 0.73        0.94 - 3.31 
#>   Min - Max        -2.45 - 2.19        -3.2 - 6.78         -2.36 - 1.94        -3.06 - 6.12
#> 
#> visit 3
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.11 (1.06)        1.74 (2.13)         0.13 (0.98)         2.12 (1.66) 
#>   Median               -0.2                1.55                0.09                1.84    
#>   IQR              -0.83 - 0.55        0.44 - 3.24         -0.44 - 0.84          1 - 3.29  
#>   Min - Max        -3.31 - 2.69        -3.27 - 8.35        -2.98 - 2.16        -1.62 - 5.7 
#> 
#> visit 4
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.05 (0.94)         2.01 (2.08)         -0.15 (1.02)         1.9 (2.02) 
#>   Median               0.13                2.15                -0.1                1.85    
#>   IQR              -0.54 - 0.66        0.83 - 3.23         -0.74 - 0.49        0.74 - 2.96 
#>   Min - Max        -2.96 - 2.08        -3.86 - 7.11        -2.76 - 2.24        -3.06 - 8.42
#> 
#> visit 5
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.02 (1.01)        1.73 (2.12)         -0.08 (0.94)        1.93 (1.92) 
#>   Median              -0.08                1.71                0.06                1.94    
#>   IQR              -0.63 - 0.63        0.14 - 3.21         -0.72 - 0.5         0.49 - 3.26 
#>   Min - Max        -2.23 - 2.28        -3.56 - 6.03        -2.58 - 2.16        -2.51 - 7.85
#> 
#> visit 6
#>   n                    100                 100                 100                 100     
#>   Mean (SD)         -0.14 (1)           2.15 (1.9)         0.03 (0.89)         1.69 (2.22) 
#>   Median              -0.13                2.18                0.03                1.58    
#>   IQR              -0.81 - 0.46         0.77 - 3.4         -0.58 - 0.59        0.07 - 3.33 
#>   Min - Max        -3.4 - 2.72         -1.96 - 6.59        -2.16 - 2.89        -3.71 - 6.72
#> 
#> visit 7
#>   n                    100                 100                 100                 100     
#>   Mean (SD)         0.1 (0.95)         1.56 (1.91)         0.09 (1.07)         2.03 (2.11) 
#>   Median               0.18                1.69                0.03                2.25    
#>   IQR              -0.56 - 0.77        0.02 - 2.79         -0.65 - 0.89        0.31 - 3.54 
#>   Min - Max        -2.16 - 2.26        -3.15 - 6.06        -2.08 - 2.68        -4.27 - 6.43
#> 
#> visit 8
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.07 (1.08)        1.89 (2.04)         -0.27 (1.1)         1.82 (1.85) 
#>   Median              -0.07                2.14               -0.28                1.48    
#>   IQR              -0.79 - 0.52        0.62 - 3.28         -0.99 - 0.65        0.68 - 3.19 
#>   Min - Max        -2.85 - 3.13        -2.89 - 6.29        -2.58 - 2.31        -3.1 - 6.22 
#> 
#> visit 9
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.05 (0.87)        2.07 (2.22)         -0.03 (0.98)        2.09 (2.17) 
#>   Median              -0.11                1.84                0.02                2.13    
#>   IQR              -0.63 - 0.57         0.58 - 3.4         -0.65 - 0.5         0.47 - 3.52 
#>   Min - Max        -2.28 - 2.19        -3.1 - 8.71         -3.15 - 2.39        -2.86 - 7.64
#> 
#> visit 10
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.02 (1.01)        1.77 (2.19)         -0.05 (1.02)        2.11 (1.83) 
#>   Median              -0.07                1.68                 0                  2.06    
#>   IQR              -0.83 - 0.7         0.22 - 3.18         -0.64 - 0.62        1.12 - 3.53 
#>   Min - Max        -2.08 - 2.5         -3.61 - 8.08        -3.07 - 3.05        -2.53 - 6.16

# EXAMPLE 2
library(random.cdisc.data)
library(dplyr)

AQS <- radam("AQS") %>%
   filter(PARAMCD == "BFIALL") %>%
   mutate(
     ARM = factor(sample(c("ARM A", "ARM B", "ARM C"), nrow(.), replace = TRUE)),
     AVISIT = factor(AVISIT, levels = unique(AVISIT))
   )

tbl <- t_summarize_by_visit(
  data = AQS[c("AVAL", "CHG")],
  visit = AQS$AVISIT,
  col_by = AQS$ARM,
  id = AQS$USUBJID
)
tbl

#>                                  ARM A                                     ARM B                                     ARM C               
#>                                  (N=82)                                    (N=86)                                    (N=86)              
#>                        AVAL                  CHG                 AVAL                  CHG                 AVAL                  CHG     
#> -----------------------------------------------------------------------------------------------------------------------------------------
#> WEEK 1
#>   n                     31                   31                   31                   31                   38                   38      
#>   Mean (SD)         0.1 (0.81)              0 (0)             0.02 (0.78)             0 (0)             0.38 (0.87)             0 (0)    
#>   Median               -0.12                  0                  0.07                   0                   0.5                   0      
#>   IQR              -0.49 - 0.56             0 - 0             -0.4 - 0.57             0 - 0            -0.23 - 0.79             0 - 0    
#>   Min - Max        -1.01 - 2.55             0 - 0            -1.56 - 1.84             0 - 0            -1.75 - 2.16             0 - 0    
#> 
#> WEEK 2
#>   n                     35                   35                   33                   33                   32                   32      
#>   Mean (SD)         0.06 (1.06)         -0.16 (1.55)          0.12 (0.99)          -0.1 (1.47)         -0.04 (0.89)         -0.14 (1.04) 
#>   Median               -0.04                -0.12                0.07                 0.11                 0.02                   0      
#>   IQR              -0.65 - 0.71         -1.46 - 1.02         -0.55 - 0.77         -1.06 - 0.93         -0.75 - 0.64         -0.79 - 0.63 
#>   Min - Max         -1.97 - 2.5         -2.55 - 4.06         -1.62 - 2.73          -3.2 - 2.57         -1.88 - 1.63         -2.14 - 1.67 
#> 
#> WEEK 3
#>   n                     29                   29                   31                   31                   28                   28      
#>   Mean (SD)         0.22 (1.17)         -0.18 (1.31)          0.03 (1.2)          -0.05 (1.42)         -0.12 (1.19)          -0.2 (1.24) 
#>   Median               0.06                 -0.36                0.11                  0.1                 -0.44                -0.23    
#>   IQR              -0.49 - 1.01         -1.04 - 0.73         -0.79 - 0.85         -0.87 - 0.93         -1.09 - 0.78          -1.15 - 0.9 
#>   Min - Max        -3.01 - 2.82         -3.21 - 2.57         -2.61 - 2.24         -2.77 - 2.57         -2.18 - 2.06         -2.74 - 2.06 
#> 
#> WEEK 4
#>   n                     23                   23                   22                   22                   34                   34      
#>   Mean (SD)         0.18 (0.84)          0.06 (1.22)         -0.11 (1.07)         -0.45 (1.47)          0.34 (0.95)          0.19 (1.28) 
#>   Median               0.15                   0                  0.09                 -0.43                0.12                 0.24     
#>   IQR               -0.4 - 0.66         -0.66 - 0.82          -0.6 - 0.47          -1.28 - 0.4          -0.3 - 1.03         -0.57 - 0.89 
#>   Min - Max        -1.22 - 2.16          -2.6 - 2.23         -2.54 - 2.14         -3.92 - 2.72         -1.82 - 3.01         -2.46 - 3.33 
#> 
#> WEEK 5
#>   n                     22                   22                   20                   20                   25                   25      
#>   Mean (SD)        -0.08 (0.93)         -0.43 (1.12)         -0.13 (0.85)         -0.16 (0.84)         -0.11 (0.96)         -0.31 (1.39) 
#>   Median               -0.05                -0.49                0.07                 -0.2                 -0.07                -0.65    
#>   IQR               -0.7 - 0.66         -1.38 - 0.18         -0.56 - 0.43         -0.81 - 0.58         -0.87 - 0.59          -1.25 - 0.6 
#>   Min - Max        -2.37 - 1.59         -2.06 - 1.92         -2.05 - 1.25         -1.46 - 1.04         -1.79 - 1.63         -1.98 - 3.39 
#> 
#> WEEK 6
#>   n                     24                   24                   16                   16                   14                   14      
#>   Mean (SD)        -0.03 (0.89)         -0.29 (1.39)         -0.22 (0.84)          -0.1 (1.18)         -0.41 (1.02)          -0.89 (1.5) 
#>   Median               0.23                   0                  -0.09                -0.42                -0.45                -0.57    
#>   IQR              -0.39 - 0.43         -0.57 - 0.48         -0.57 - 0.24          -0.9 - 0.77          -0.78 - 0.4         -2.19 - 0.09 
#>   Min - Max        -1.98 - 1.84         -3.65 - 1.92         -1.99 - 1.14         -1.93 - 2.08         -2.29 - 1.43         -3.31 - 1.44 
#> 
#> WEEK 7
#>   n                     12                   12                   17                   17                   14                   14      
#>   Mean (SD)        -0.09 (0.91)         -0.26 (1.21)          0.04 (0.99)          0.04 (1.18)         -0.25 (0.85)          -0.67 (1.1) 
#>   Median               0.08                 -0.61                -0.1                 -0.35                -0.33                -0.73    
#>   IQR              -0.82 - 0.52          -0.95 - 0.2          -0.87 - 0.7         -0.82 - 0.53         -0.98 - 0.59         -1.48 - 0.15 
#>   Min - Max        -1.41 - 1.37           -2 - 2.67          -1.23 - 1.56         -1.46 - 2.22         -1.43 - 0.78         -2.28 - 1.41 
#> 
#> WEEK 8
#>   n                     11                   11                   10                   10                   10                   10      
#>   Mean (SD)         0.09 (1.08)         -0.01 (1.79)           -0.48 (1)          -0.36 (1.22)          0.04 (0.78)         -0.49 (1.01) 
#>   Median               -0.18                -0.26                -0.74                0.23                 0.05                 -0.07    
#>   IQR              -0.69 - 0.49         -1.04 - 0.54         -1.16 - 0.11          -1.06 - 0.4         -0.52 - 0.41          -1.4 - 0.18 
#>   Min - Max        -1.08 - 2.18         -2.33 - 3.53         -1.59 - 1.64          -2.97 - 0.9         -1.02 - 1.37         -2.03 - 0.94 
#> 
#> WEEK 9
#>   n                      5                    5                    9                    9                   11                   11      
#>   Mean (SD)         0.06 (0.56)          0.01 (0.57)          0.58 (1.03)          0.58 (1.37)          0.06 (0.81)         -0.06 (0.92) 
#>   Median               0.21                 0.37                 0.98                 1.43                 0.26                 -0.34    
#>   IQR               0.02 - 0.35          -0.4 - 0.42          0.14 - 1.1          -0.37 - 1.63         -0.57 - 0.51         -0.63 - 0.69 
#>   Min - Max        -0.86 - 0.59         -0.79 - 0.44         -1.59 - 1.94          -1.8 - 1.84         -1.41 - 1.35         -1.26 - 1.36 
#> 
#> WEEK 10
#>   n                      7                    7                    5                    5                    5                    5      
#>   Mean (SD)          0.4 (0.7)           0.29 (0.38)         -1.08 (0.95)         -0.93 (1.72)          1.16 (1.14)          0.78 (0.89) 
#>   Median               0.57                 0.11                 -0.86                -1.07                1.72                  0.7     
#>   IQR              -0.24 - 0.83          0.08 - 0.58         -0.9 - -0.51         -1.86 - 0.18          0.85 - 1.95          0.07 - 0.98 
#>   Min - Max         -0.4 - 1.41         -0.24 - 0.81         -2.73 - -0.4         -3.15 - 1.24         -0.71 - 1.99         -0.01 - 2.18 
#> 
#> WEEK 11
#>   n                      0                    0                    5                    5                    5                    5      
#>   Mean (SD)           NE (NE)              NE (NE)            0.82 (1.33)          0.27 (1.8)          -0.68 (0.82)         -0.88 (0.66) 
#>   Median                NE                   NE                   0.5                 0.07                 -0.79                -0.84    
#>   IQR                 NE - NE              NE - NE           -0.32 - 1.74         -1.06 - 1.53          -1 - -0.26          -1.19 - -0.63
#>   Min - Max           NE - NE              NE - NE           -0.43 - 2.61         -1.77 - 2.59         -1.78 - 0.41         -1.75 - 0.03 
# NOT RUN {
Viewer(tbl)
# }