Summary Table by Visit

This function summarizes test results or change from baseline statistics by visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.

t_summarize_by_visit(data, visit, id, col_by, col_N)

Arguments

data
data frame with numerical variables to be summarized. If the variable has a label attribute then it will be used as the sub-header column name.

visit
factor with visit names ordered by desired display order in the stacked table.

id
unique subject identifier variable.

col_by
factor with reference and comparison group information, the first level indicates the reference group

col_N
a table object with the reference population used for the header of the table.

Value

an rtable object

Details

Every variable in data will be mapped to a sub-column under col_by column. Multiple variables will be shown side-by-side with the variable label (if present) or variable name as the sub-column header.

Currently n, Mean (SD), Median, IQR and Min-Max will be derived for each variable from data per each factor in col_by.

Examples

# EXAMPLE 1
ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B")
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA
ANL$PCHG <- ANL$CHG/ANL$AVAL*100

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

ASL <- unique(ANL[, c("USUBJID","ARM")])

ADSL <- radsl()
ADVS <- radvs(ADSL)

t_summarize_by_visit(data = ADVS[c("AVAL")], visit = ADVS$AVISIT, col_by = ADVS$ARM,
  id = ADVS$USUBJID, col_N = table(ADSL$ARM))

#>                        A: Drug X             B: Placebo          C: Combination
#>                         (N=141)               (N=120)               (N=139)    
#> -------------------------------------------------------------------------------
#> SCREENING
#>   n                       846                   720                   834      
#>   Mean (SD)           50.02 (7.9)           50.48 (8.25)          50.13 (7.8)  
#>   Median                 49.93                 50.06                  50.3     
#>   IQR                 44.49 - 55.5         45.52 - 56.16          44.85 - 55.3 
#>   Min - Max          29.06 - 74.75          19.8 - 74.16         28.41 - 74.25 
#> 
#> BASELINE
#>   n                       846                   720                   834      
#>   Mean (SD)           49.75 (7.88)          49.93 (7.94)          50.24 (7.84) 
#>   Median                 49.58                 50.41                 49.99     
#>   IQR                 44.2 - 55.21         44.84 - 55.44         45.05 - 55.68 
#>   Min - Max          28.93 - 74.93         23.38 - 73.36         28.99 - 75.24 
#> 
#> WEEK 1 DAY 8
#>   n                       846                   720                   834      
#>   Mean (SD)           50.11 (7.93)          49.8 (7.97)           50.48 (7.65) 
#>   Median                 50.34                 50.15                 50.32     
#>   IQR                 44.7 - 55.71         44.21 - 55.69         45.45 - 55.72 
#>   Min - Max          25.01 - 70.99         26.78 - 72.21         27.67 - 70.66 
#> 
#> WEEK 2 DAY 15
#>   n                       846                   720                   834      
#>   Mean (SD)           49.96 (7.96)          50.08 (8.02)          49.89 (7.75) 
#>   Median                 49.99                 49.92                 49.98     
#>   IQR                44.22 - 55.41         44.95 - 55.41         44.41 - 55.03 
#>   Min - Max          25.78 - 71.57         24.21 - 74.91         25.73 - 75.69 
#> 
#> WEEK 3 DAY 22
#>   n                       846                   720                   834      
#>   Mean (SD)           50.17 (8.2)           50.14 (7.97)          49.63 (7.91) 
#>   Median                 50.04                 50.14                 49.76     
#>   IQR                44.61 - 55.65         44.79 - 55.37         44.62 - 54.59 
#>   Min - Max          25.12 - 76.84         24.86 - 77.23          27.7 - 76.65 
#> 
#> WEEK 4 DAY 29
#>   n                       846                   720                   834      
#>   Mean (SD)           49.82 (7.39)          49.64 (8.02)          49.79 (8.26) 
#>   Median                 49.67                 49.49                  49.4     
#>   IQR                44.78 - 54.89         44.34 - 54.94         44.24 - 55.55 
#>   Min - Max          26.92 - 71.47           29.34 - 75           26.5 - 75.44 
#> 
#> WEEK 5 DAY 36
#>   n                       846                   720                   834      
#>   Mean (SD)           50.63 (8.26)          50.51 (7.7)           50.08 (8.15) 
#>   Median                 50.49                 50.53                 50.16     
#>   IQR                44.92 - 56.15          45.43 - 55.7         44.45 - 55.58 
#>   Min - Max          25.59 - 79.57          26.64 - 73.9         19.27 - 73.13 

t_summarize_by_visit(data = ADVS[c("PCHG")], visit = ADVS$AVISIT, col_by = ADVS$ARM,
  id = ADVS$USUBJID, col_N = table(ADSL$ARM))

#>                         A: Drug X             B: Placebo           C: Combination 
#>                          (N=141)                (N=120)                (N=139)    
#> ----------------------------------------------------------------------------------
#> SCREENING
#>   n                         0                      0                      0       
#>   Mean (SD)              NE (NE)                NE (NE)                NE (NE)    
#>   Median                   NE                     NE                     NE       
#>   IQR                    NE - NE                NE - NE                NE - NE    
#>   Min - Max              NE - NE                NE - NE                NE - NE    
#> 
#> BASELINE
#>   n                        846                    720                    834      
#>   Mean (SD)               0 (0)                  0 (0)                  0 (0)     
#>   Median                    0                      0                      0       
#>   IQR                     0 - 0                  0 - 0                  0 - 0     
#>   Min - Max               0 - 0                  0 - 0                  0 - 0     
#> 
#> WEEK 1 DAY 8
#>   n                        846                    720                    834      
#>   Mean (SD)            3.5 (24.27)           2.71 (25.37)           2.99 (22.88)  
#>   Median                  1.42                    0.4                   -0.75     
#>   IQR                 -14.32 - 18.2         -14.53 - 16.74         -13.58 - 15.99 
#>   Min - Max          -59.82 - 111.59        -53.36 - 145.5         -48.04 - 97.68 
#> 
#> WEEK 2 DAY 15
#>   n                        846                    720                    834      
#>   Mean (SD)           2.92 (23.23)           3.15 (25.05)           1.77 (22.55)  
#>   Median                  1.28                   -0.44                  0.08      
#>   IQR                -13.26 - 17.05         -14.09 - 16.57         -14.46 - 16.15 
#>   Min - Max          -46.04 - 113.54        -52.37 - 147.91         -53.85 - 91.5 
#> 
#> WEEK 3 DAY 22
#>   n                        846                    720                    834      
#>   Mean (SD)           3.41 (23.58)           3.35 (24.92)           1.44 (23.92)  
#>   Median                  0.35                   0.61                   -1.75     
#>   IQR                -13.55 - 17.99         -15.04 - 17.91         -14.86 - 15.18 
#>   Min - Max           -56.5 - 80.01         -55.69 - 107.03        -52.19 - 92.51 
#> 
#> WEEK 4 DAY 29
#>   n                        846                    720                    834      
#>   Mean (SD)           2.68 (22.48)           2.13 (24.19)           1.62 (23.61)  
#>   Median                  1.35                   -0.89                  -1.02     
#>   IQR                -13.92 - 15.93          -13.8 - 15.26         -15.23 - 16.38 
#>   Min - Max          -52.21 - 93.26         -48.81 - 106.17        -52.84 - 94.64 
#> 
#> WEEK 5 DAY 36
#>   n                        846                    720                    834      
#>   Mean (SD)            4.53 (24.9)           4.17 (25.72)           2.19 (23.63)  
#>   Median                  2.11                   0.76                   -0.31     
#>   IQR                -12.66 - 17.88         -12.53 - 15.51         -13.66 - 15.41 
#>   Min - Max          -52.95 - 119.37        -50.57 - 143.95        -55.79 - 102.43

# DO NOT THINK WE NEED THIS BLOCK AS LABELS ALREADY AVAILABLE
# ADDED THIS NOTE GIVEN FINDING OF ATTRIBUTES BEING REMOVED
#Add label to variable instead showing variable name
#ANL <- var_relabel(ANL, AVAL = "Value at\nVisit",
 #                       CHG = "Change from\nBaseline",
  #                      PCHG = "Percent Change\nfrom Baseline")

t_summarize_by_visit(
  data = ADVS[c("AVAL", "CHG")],
  visit = ADVS$AVISIT,
  col_by = ADVS$ARM,
  id = ADVS$USUBJID,
  col_N = table(ADSL$ARM)
)

#>                                                  A: Drug X                                                                 B: Placebo                                                              C: Combination                          
#>                                                   (N=141)                                                                   (N=120)                                                                   (N=139)                              
#>                             Analysis Value                Absolute Change from Baseline               Analysis Value                Absolute Change from Baseline               Analysis Value                Absolute Change from Baseline
#> -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> SCREENING
#>   n                               846                                   0                                   720                                   0                                   834                                   0              
#>   Mean (SD)                   50.02 (7.9)                            NE (NE)                           50.48 (8.25)                            NE (NE)                            50.13 (7.8)                            NE (NE)           
#>   Median                         49.93                                 NE                                  50.06                                 NE                                  50.3                                  NE              
#>   IQR                        44.49 - 55.5                            NE - NE                           45.52 - 56.16                           NE - NE                           44.85 - 55.3                            NE - NE           
#>   Min - Max                  29.06 - 74.75                           NE - NE                           19.8 - 74.16                            NE - NE                           28.41 - 74.25                           NE - NE           
#> 
#> BASELINE
#>   n                               846                                  846                                  720                                  720                                  834                                  834             
#>   Mean (SD)                  49.75 (7.88)                             0 (0)                            49.93 (7.94)                             0 (0)                            50.24 (7.84)                             0 (0)            
#>   Median                         49.58                                  0                                  50.41                                  0                                  49.99                                  0              
#>   IQR                        44.2 - 55.21                             0 - 0                            44.84 - 55.44                            0 - 0                            45.05 - 55.68                            0 - 0            
#>   Min - Max                  28.93 - 74.93                            0 - 0                            23.38 - 73.36                            0 - 0                            28.99 - 75.24                            0 - 0            
#> 
#> WEEK 1 DAY 8
#>   n                               846                                  846                                  720                                  720                                  834                                  834             
#>   Mean (SD)                  50.11 (7.93)                         0.36 (11.45)                          49.8 (7.97)                         -0.12 (11.44)                        50.48 (7.65)                         0.24 (10.77)         
#>   Median                         50.34                                0.72                                 50.15                                0.21                                 50.32                                -0.33            
#>   IQR                        44.7 - 55.71                         -7.41 - 8.25                         44.21 - 55.69                         -7.8 - 7.7                          45.45 - 55.72                        -7.26 - 7.62         
#>   Min - Max                  25.01 - 70.99                       -40.26 - 37.44                        26.78 - 72.21                        -34.4 - 36.65                        27.67 - 70.66                       -36.12 - 29.92        
#> 
#> WEEK 2 DAY 15
#>   n                               846                                  846                                  720                                  720                                  834                                  834             
#>   Mean (SD)                  49.96 (7.96)                         0.21 (10.91)                         50.08 (8.02)                         0.15 (11.12)                         49.89 (7.75)                         -0.35 (10.83)        
#>   Median                         49.99                                0.65                                 49.92                                -0.21                                49.98                                0.04             
#>   IQR                        44.22 - 55.41                        -6.97 - 7.84                         44.95 - 55.41                        -7.47 - 7.79                         44.41 - 55.03                         -7.66 - 7.5         
#>   Min - Max                  25.78 - 71.57                       -30.63 - 37.03                        24.21 - 74.91                       -30.24 - 34.85                        25.73 - 75.69                        -31.52 - 31.2        
#> 
#> WEEK 3 DAY 22
#>   n                               846                                  846                                  720                                  720                                  834                                  834             
#>   Mean (SD)                   50.17 (8.2)                         0.42 (11.18)                         50.14 (7.97)                         0.21 (11.42)                         49.63 (7.91)                         -0.61 (11.37)        
#>   Median                         50.04                                0.18                                 50.14                                 0.3                                 49.76                                -0.92            
#>   IQR                        44.61 - 55.65                        -6.99 - 8.08                         44.79 - 55.37                        -8.09 - 8.08                         44.62 - 54.59                        -8.02 - 6.87         
#>   Min - Max                  25.12 - 76.84                        -32.63 - 30.5                        24.86 - 77.23                        -33.68 - 33.3                        27.7 - 76.65                        -32.05 - 36.83        
#> 
#> WEEK 4 DAY 29
#>   n                               846                                  846                                  720                                  720                                  834                                  834             
#>   Mean (SD)                  49.82 (7.39)                         0.07 (10.68)                         49.64 (8.02)                         -0.29 (11.05)                        49.79 (8.26)                         -0.45 (11.27)        
#>   Median                         49.67                                0.66                                 49.49                                -0.45                                49.4                                 -0.54            
#>   IQR                        44.78 - 54.89                         -7.4 - 7.26                         44.34 - 54.94                        -7.15 - 7.24                         44.24 - 55.55                        -8.18 - 7.63         
#>   Min - Max                  26.92 - 71.47                       -31.54 - 34.49                         29.34 - 75                         -31.52 - 33.76                        26.5 - 75.44                        -35.89 - 36.68        
#> 
#> WEEK 5 DAY 36
#>   n                               846                                  846                                  720                                  720                                  834                                  834             
#>   Mean (SD)                  50.63 (8.26)                         0.88 (11.59)                          50.51 (7.7)                         0.58 (11.13)                         50.08 (8.15)                         -0.16 (11.11)        
#>   Median                         50.49                                1.02                                 50.53                                0.41                                 50.16                                -0.17            
#>   IQR                        44.92 - 56.15                        -6.84 - 8.29                         45.43 - 55.7                         -6.92 - 7.56                         44.45 - 55.58                        -7.15 - 7.25         
#>   Min - Max                  25.59 - 79.57                       -32.46 - 43.03                        26.64 - 73.9                        -31.77 - 35.83                        19.27 - 73.13                       -31.97 - 32.69        

# EXAMPLE 2
library(random.cdisc.data)
library(dplyr)

ADSL <- radsl()
ADQS <- radqs(ADSL) %>%
   filter(PARAMCD == "BFIALL")

tbl <- t_summarize_by_visit(
  data = ADQS[c("AVAL", "CHG")],
  visit = ADQS$AVISIT,
  col_by = ADQS$ARM,
  id = ADQS$USUBJID,
  col_N = table(ADSL$ARM)
)
tbl

#>                                                  A: Drug X                                                                 B: Placebo                                                              C: Combination                          
#>                                                   (N=146)                                                                   (N=119)                                                                   (N=135)                              
#>                             Analysis Value                Absolute Change from Baseline               Analysis Value                Absolute Change from Baseline               Analysis Value                Absolute Change from Baseline
#> -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> SCREENING
#>   n                               146                                   0                                   119                                   0                                   135                                   0              
#>   Mean (SD)                  48.64 (8.72)                            NE (NE)                           50.76 (7.99)                            NE (NE)                            49.45 (7.5)                            NE (NE)           
#>   Median                         48.98                                 NE                                  51.63                                 NE                                  49.89                                 NE              
#>   IQR                        43.34 - 54.64                           NE - NE                           45.52 - 56.57                           NE - NE                           44.35 - 54.1                            NE - NE           
#>   Min - Max                  27.19 - 71.24                           NE - NE                           28.61 - 65.37                           NE - NE                           24.6 - 73.36                            NE - NE           
#> 
#> BASELINE
#>   n                               146                                  146                                  119                                  119                                  135                                  135             
#>   Mean (SD)                  50.68 (7.23)                             0 (0)                             49.9 (8.19)                             0 (0)                            50.56 (8.18)                             0 (0)            
#>   Median                         51.57                                  0                                  49.97                                  0                                  50.33                                  0              
#>   IQR                        45.99 - 55.73                            0 - 0                            44.81 - 55.55                            0 - 0                            44.58 - 56.23                            0 - 0            
#>   Min - Max                  24.61 - 64.98                            0 - 0                            28.31 - 69.5                             0 - 0                            28.18 - 73.85                            0 - 0            
#> 
#> WEEK 1 DAY 8
#>   n                               146                                  146                                  119                                  119                                  135                                  135             
#>   Mean (SD)                  49.72 (8.12)                         -0.96 (10.62)                        49.22 (6.81)                         -0.68 (10.52)                        51.78 (7.17)                         1.22 (11.15)         
#>   Median                         49.39                                -1.09                                48.95                                -1.86                                51.81                                1.13             
#>   IQR                        45.39 - 55.14                         -6.55 - 5.2                         44.76 - 52.93                         -7.58 - 7.4                         47.01 - 56.43                        -6.49 - 9.07         
#>   Min - Max                  25.66 - 71.08                       -28.28 - 23.63                        27.12 - 67.15                        -31.2 - 26.64                        35.76 - 69.38                       -26.88 - 35.72        
#> 
#> WEEK 2 DAY 15
#>   n                               146                                  146                                  119                                  119                                  135                                  135             
#>   Mean (SD)                  49.21 (8.04)                         -1.47 (10.5)                         51.08 (7.68)                          1.19 (11.6)                         49.63 (7.61)                         -0.93 (11.73)        
#>   Median                         49.51                                -2.07                                50.23                                0.32                                 49.99                                0.09             
#>   IQR                        43.56 - 53.21                        -7.08 - 4.44                         45.94 - 55.86                        -5.94 - 8.03                         43.93 - 54.92                        -8.37 - 6.57         
#>   Min - Max                  31.04 - 71.46                       -33.83 - 26.96                        30.75 - 70.05                       -24.94 - 36.93                        30.67 - 68.94                       -39.92 - 31.22        
#> 
#> WEEK 3 DAY 22
#>   n                               146                                  146                                  119                                  119                                  135                                  135             
#>   Mean (SD)                   49.9 (8.29)                          -0.78 (9.5)                         49.02 (8.05)                         -0.88 (11.92)                        50.11 (7.71)                         -0.45 (10.54)        
#>   Median                         50.11                                -1.2                                 49.56                                -1.54                                50.04                                -1.22            
#>   IQR                        44.8 - 55.46                         -7.11 - 5.64                         42.85 - 54.55                         -7.29 - 6.2                         44.67 - 54.99                        -7.55 - 8.22         
#>   Min - Max                  29.24 - 74.32                       -22.08 - 29.83                        27.28 - 71.34                       -33.38 - 32.25                        28.02 - 71.83                       -30.97 - 24.72        
#> 
#> WEEK 4 DAY 29
#>   n                               146                                  146                                  119                                  119                                  135                                  135             
#>   Mean (SD)                  49.48 (7.64)                         -1.2 (10.32)                          50.2 (8.51)                         0.31 (11.63)                         50.96 (8.31)                         0.41 (11.98)         
#>   Median                         49.31                                -2.99                                50.26                                1.34                                 50.78                                1.07             
#>   IQR                        44.05 - 54.84                        -7.95 - 5.48                         44.65 - 55.12                        -7.05 - 7.54                         46.26 - 56.95                        -6.72 - 9.89         
#>   Min - Max                  31.17 - 71.57                       -23.55 - 33.62                        31.66 - 72.59                       -27.92 - 29.48                        26.44 - 69.67                       -33.59 - 27.62        
#> 
#> WEEK 5 DAY 36
#>   n                               146                                  146                                  119                                  119                                  135                                  135             
#>   Mean (SD)                  48.38 (7.85)                         -2.29 (10.61)                        50.58 (8.38)                         0.69 (12.26)                          49.91 (8.6)                         -0.65 (11.78)        
#>   Median                         47.74                                -2.52                                49.78                                0.98                                 49.69                                -1.42            
#>   IQR                        43.11 - 53.35                         -9.8 - 4.51                         45.1 - 56.62                         -6.91 - 8.12                         43.86 - 55.51                        -7.76 - 6.93         
#>   Min - Max                  28.1 - 67.76                        -25.44 - 24.23                        32.07 - 69.01                       -28.12 - 32.84                        21.22 - 70.91                       -43.57 - 31.45        
# NOT RUN {
Viewer(tbl)
# }