Summary Table by Visit

This function summarizes test results or change from baseline statistics by visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.

t_summarize_by_visit(data, visit, id, col_by)

Arguments

data
data frame with numerical variables to be summarized. If the variable has a label attribute then it will be used as the sub-header column name.

visit
factor with visit names ordered by desired display order in the stacked table.

id
unique subject identifier variable.

col_by
factor with reference and comparison group information, the first level indicates the reference group

Value

an rtable object

Details

Every variable in data will be mapped to a sub-column under col_by column. Multiple variables will be shown side-by-side with the variable label (if present) or variable name as the sub-column header.

Currently n, Mean (SD), Median, IQR and Min-Max will be derived for each variable from data per each factor in col_by.

Examples

# EXAMPLE 1
ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B")
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA
ANL$PCHG <- ANL$CHG/ANL$AVAL*100

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

t_summarize_by_visit(data = ANL[c("AVAL")], visit = ANL$VISIT, col_by = ANL$ARM, id = ANL$USUBJID)

#>                       ARM A               ARM B    
#>                      (N=100)             (N=100)   
#> ---------------------------------------------------
#> visit 1
#>   n                    100                 100     
#>   Mean (SD)        -0.18 (1.11)         -0.13 (1)  
#>   Median              -0.11               -0.21    
#>   IQR              -0.89 - 0.55        -0.85 - 0.58
#>   Min - Max        -2.66 - 2.89         -2 - 2.29  
#> 
#> visit 2
#>   n                    100                 100     
#>   Mean (SD)        -0.04 (1.1)         -0.04 (0.99)
#>   Median               0.04               -0.01    
#>   IQR              -0.77 - 0.64        -0.72 - 0.55
#>   Min - Max        -3.16 - 2.63        -2.18 - 2.36
#> 
#> visit 3
#>   n                    100                 100     
#>   Mean (SD)        0.09 (0.93)           0.09 (1)  
#>   Median               0.13                0.12    
#>   IQR              -0.68 - 0.82        -0.6 - 0.76 
#>   Min - Max        -2.06 - 2.03          -2 - 3.5  
#> 
#> visit 4
#>   n                    100                 100     
#>   Mean (SD)        -0.06 (1.12)        0.09 (0.98) 
#>   Median               0.01                0.18    
#>   IQR              -0.84 - 0.78        -0.61 - 0.72
#>   Min - Max        -3.29 - 2.26        -2.27 - 2.54
#> 
#> visit 5
#>   n                    100                 100     
#>   Mean (SD)        -0.12 (1.06)        0.04 (0.98) 
#>   Median              -0.08                0.19    
#>   IQR              -0.81 - 0.55        -0.68 - 0.65
#>   Min - Max        -2.82 - 3.06        -2.14 - 2.26
#> 
#> visit 6
#>   n                    100                 100     
#>   Mean (SD)         0.17 (0.9)         -0.11 (0.86)
#>   Median               0.06               -0.13    
#>   IQR              -0.37 - 0.88        -0.71 - 0.6 
#>   Min - Max        -2.08 - 2.24        -2.81 - 2.13
#> 
#> visit 7
#>   n                    100                 100     
#>   Mean (SD)        0.04 (1.07)         -0.14 (1.09)
#>   Median               0.05                -0.1    
#>   IQR              -0.56 - 0.74        -0.81 - 0.75
#>   Min - Max        -3.95 - 3.13        -3.11 - 1.79
#> 
#> visit 8
#>   n                    100                 100     
#>   Mean (SD)        0.05 (0.88)         -0.1 (1.05) 
#>   Median               0.17               -0.17    
#>   IQR              -0.55 - 0.68        -0.83 - 0.43
#>   Min - Max        -2.07 - 2.2         -2.61 - 2.82
#> 
#> visit 9
#>   n                    100                 100     
#>   Mean (SD)        0.09 (0.98)         -0.08 (1.13)
#>   Median               0.22               -0.02    
#>   IQR              -0.65 - 0.7         -0.66 - 0.59
#>   Min - Max        -2.26 - 3.2         -3.31 - 3.35
#> 
#> visit 10
#>   n                    100                 100     
#>   Mean (SD)        0.04 (0.96)         -0.03 (0.97)
#>   Median              -0.06               -0.07    
#>   IQR              -0.6 - 0.66         -0.62 - 0.67
#>   Min - Max        -2.87 - 2.65        -2.67 - 2.18

t_summarize_by_visit(data = ANL[c("PCHG")], visit = ANL$VISIT, col_by = ANL$ARM, id = ANL$USUBJID)

#>                            ARM A                        ARM B        
#>                           (N=100)                      (N=100)       
#> ---------------------------------------------------------------------
#> visit 1
#>   n                          0                            0          
#>   Mean (SD)               NE (NE)                      NE (NE)       
#>   Median                    NE                           NE          
#>   IQR                     NE - NE                      NE - NE       
#>   Min - Max               NE - NE                      NE - NE       
#> 
#> visit 2
#>   n                         100                          100         
#>   Mean (SD)          -367.82 (3165.63)             31.7 (2532.71)    
#>   Median                   26.53                        7.15         
#>   IQR                -333.69 - 332.79             -266.57 - 383.37   
#>   Min - Max         -26626.03 - 9345.49         -17002.24 - 14081.52 
#> 
#> visit 3
#>   n                         100                          100         
#>   Mean (SD)          -225.15 (2405.56)           2620.62 (36542.85)  
#>   Median                   0.63                         22.42        
#>   IQR                -197.74 - 319.54               -379 - 379.82    
#>   Min - Max         -17229.81 - 5050.93         -67901.37 - 357235.37
#> 
#> visit 4
#>   n                         100                          100         
#>   Mean (SD)         1974.61 (11370.06)            350.02 (2879.66)   
#>   Median                   45.83                        69.32        
#>   IQR                -190.57 - 376.95              -198.9 - 329.54   
#>   Min - Max        -12340.88 - 95408.16          -5343.38 - 25768.91 
#> 
#> visit 5
#>   n                         100                          100         
#>   Mean (SD)          434.45 (3173.26)             -18.56 (1373.77)   
#>   Median                  -29.17                        23.69        
#>   IQR                -339.66 - 338.67              -220.1 - 351.83   
#>   Min - Max         -4623.6 - 28349.32           -6252.62 - 8566.68  
#> 
#> visit 6
#>   n                         100                          100         
#>   Mean (SD)         2313.95 (29708.01)            210.37 (3818.08)   
#>   Median                   51.86                       -31.16        
#>   IQR                 -407.1 - 382.67             -311.44 - 274.55   
#>   Min - Max        -46694.57 - 291342.14        -21613.76 - 26264.25 
#> 
#> visit 7
#>   n                         100                          100         
#>   Mean (SD)          -500.35 (3913.52)            -108.75 (2168.19)  
#>   Median                  -21.35                       -43.48        
#>   IQR                -287.18 - 254.34             -406.36 - 269.21   
#>   Min - Max         -37777.48 - 2290.04          -13565.14 - 7641.83 
#> 
#> visit 8
#>   n                         100                          100         
#>   Mean (SD)          610.03 (4476.35)             -1304 (13410.35)   
#>   Median                   46.44                       -21.33        
#>   IQR                -249.62 - 540.17              -281.4 - 306.98   
#>   Min - Max        -17006.71 - 36628.44         -133646.25 - 5379.84 
#> 
#> visit 9
#>   n                         100                          100         
#>   Mean (SD)          -61.35 (1101.78)              223.08 (2366.1)   
#>   Median                   33.52                       -15.89        
#>   IQR                 -201.9 - 275.11             -182.02 - 325.01   
#>   Min - Max         -7701.66 - 2477.16           -6231.34 - 15506.37 
#> 
#> visit 10
#>   n                         100                          100         
#>   Mean (SD)          -149.81 (1008.52)            -89.37 (2361.55)   
#>   Median                  -43.63                        -0.12        
#>   IQR                 -396.48 - 252.6             -390.71 - 237.43   
#>   Min - Max         -4515.09 - 2768.42          -12254.38 - 12171.38 

# Add label to variable instead showing variable name
ANL <- var_relabel(ANL, AVAL = "Value at\nVisit",
                        CHG = "Change from\nBaseline",
                        PCHG = "Percent Change\nfrom Baseline")

t_summarize_by_visit(
  data = ANL[c("AVAL", "CHG")],
  visit = ANL$VISIT,
  col_by = ANL$ARM,
  id = ANL$USUBJID
)

#>                                 ARM A                                   ARM B              
#>                                (N=100)                                 (N=100)             
#>                      Value at          Change from           Value at          Change from 
#>                       Visit              Baseline             Visit              Baseline  
#> -------------------------------------------------------------------------------------------
#> visit 1
#>   n                    100                  0                  100                  0      
#>   Mean (SD)        -0.18 (1.11)          NE (NE)            -0.13 (1)            NE (NE)   
#>   Median              -0.11                 NE                -0.21                 NE     
#>   IQR              -0.89 - 0.55          NE - NE           -0.85 - 0.58          NE - NE   
#>   Min - Max        -2.66 - 2.89          NE - NE            -2 - 2.29            NE - NE   
#> 
#> visit 2
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.04 (1.1)         1.94 (2.12)         -0.04 (0.99)        2.08 (2.18) 
#>   Median               0.04                1.72               -0.01                2.56    
#>   IQR              -0.77 - 0.64        0.59 - 3.17         -0.72 - 0.55        0.37 - 3.68 
#>   Min - Max        -3.16 - 2.63        -3.94 - 8.72        -2.18 - 2.36        -2.98 - 5.78
#> 
#> visit 3
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.09 (0.93)         1.76 (2.13)           0.09 (1)           2.59 (1.9) 
#>   Median               0.13                1.48                0.12                2.72    
#>   IQR              -0.68 - 0.82         0.38 - 3.5         -0.6 - 0.76         1.38 - 3.67 
#>   Min - Max        -2.06 - 2.03        -3.45 - 6.51          -2 - 3.5          -3.05 - 6.86
#> 
#> visit 4
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.06 (1.12)        1.77 (2.15)         0.09 (0.98)         2.16 (1.81) 
#>   Median               0.01                1.66                0.18                2.08    
#>   IQR              -0.84 - 0.78        0.17 - 3.57         -0.61 - 0.72         1.3 - 3.23 
#>   Min - Max        -3.29 - 2.26        -3.29 - 6.8         -2.27 - 2.54        -3.05 - 6.65
#> 
#> visit 5
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.12 (1.06)        2.11 (1.98)         0.04 (0.98)          2.14 (1.7) 
#>   Median              -0.08                2.05                0.19                1.78    
#>   IQR              -0.81 - 0.55        0.88 - 3.34         -0.68 - 0.65        1.14 - 3.29 
#>   Min - Max        -2.82 - 3.06         -2 - 7.68          -2.14 - 2.26        -3.68 - 5.94
#> 
#> visit 6
#>   n                    100                 100                 100                 100     
#>   Mean (SD)         0.17 (0.9)         2.15 (2.05)         -0.11 (0.86)        1.79 (1.84) 
#>   Median               0.06                2.26               -0.13                1.88    
#>   IQR              -0.37 - 0.88        0.77 - 3.52         -0.71 - 0.6         0.63 - 3.03 
#>   Min - Max        -2.08 - 2.24        -2.28 - 7.29        -2.81 - 2.13        -3.22 - 6.02
#> 
#> visit 7
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.04 (1.07)         1.81 (2.17)         -0.14 (1.09)         2.64 (1.9) 
#>   Median               0.05                1.85                -0.1                2.57    
#>   IQR              -0.56 - 0.74        0.33 - 3.14         -0.81 - 0.75        1.06 - 4.02 
#>   Min - Max        -3.95 - 3.13        -3.86 - 7.03        -3.11 - 1.79        -2.58 - 6.69
#> 
#> visit 8
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.05 (0.88)           2 (2.08)          -0.1 (1.05)          1.9 (2.08) 
#>   Median               0.17                2.16               -0.17                1.86    
#>   IQR              -0.55 - 0.68        0.61 - 3.36         -0.83 - 0.43         0.91 - 3.1 
#>   Min - Max        -2.07 - 2.2         -4.51 - 6.62        -2.61 - 2.82        -5.1 - 8.39 
#> 
#> visit 9
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.09 (0.98)          1.67 (1.9)         -0.08 (1.13)        1.93 (1.95) 
#>   Median               0.22                1.6                -0.02                1.87    
#>   IQR              -0.65 - 0.7         0.23 - 2.87         -0.66 - 0.59        0.71 - 3.32 
#>   Min - Max        -2.26 - 3.2         -2.38 - 6.13        -3.31 - 3.35        -3.63 - 7.2 
#> 
#> visit 10
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.04 (0.96)         2.19 (1.93)         -0.03 (0.97)        1.93 (2.15) 
#>   Median              -0.06                2.18               -0.07                1.79    
#>   IQR              -0.6 - 0.66         0.66 - 3.41         -0.62 - 0.67        0.74 - 3.45 
#>   Min - Max        -2.87 - 2.65        -2.55 - 7.53        -2.67 - 2.18        -3.44 - 7.19

# EXAMPLE 2
library(random.cdisc.data)
library(dplyr)

AQS <- radam("AQS") %>%
   filter(PARAMCD == "BFIALL") %>%
   mutate(
     ARM = factor(sample(c("ARM A", "ARM B", "ARM C"), nrow(.), replace = TRUE)),
     AVISIT = factor(AVISIT, levels = unique(AVISIT))
   )

tbl <- t_summarize_by_visit(
  data = AQS[c("AVAL", "CHG")],
  visit = AQS$AVISIT,
  col_by = AQS$ARM,
  id = AQS$USUBJID
)
tbl

#>                                  ARM A                                     ARM B                                     ARM C               
#>                                  (N=87)                                    (N=91)                                    (N=88)              
#>                        AVAL                  CHG                 AVAL                  CHG                 AVAL                  CHG     
#> -----------------------------------------------------------------------------------------------------------------------------------------
#> WEEK 1
#>   n                     35                   35                   40                   40                   25                   25      
#>   Mean (SD)         0.07 (0.96)             0 (0)             0.12 (0.96)             0 (0)             0.03 (1.17)             0 (0)    
#>   Median               -0.08                  0                    0                    0                  -0.07                  0      
#>   IQR              -0.64 - 0.73             0 - 0            -0.48 - 0.75             0 - 0            -0.73 - 0.99             0 - 0    
#>   Min - Max          -2 - 1.8               0 - 0            -2.38 - 2.43             0 - 0             -2.7 - 2.06             0 - 0    
#> 
#> WEEK 2
#>   n                     41                   41                   25                   25                   34                   34      
#>   Mean (SD)         0.11 (0.94)          0.09 (1.37)         -0.28 (1.02)         -0.48 (1.28)         -0.13 (1.06)          -0.2 (1.42) 
#>   Median               0.16                 0.23                 -0.27                -0.51                -0.03                -0.06    
#>   IQR              -0.66 - 0.84         -0.81 - 1.12         -1.09 - 0.22         -1.22 - 0.44         -0.75 - 0.48          -0.7 - 0.43 
#>   Min - Max        -2.03 - 1.71         -3.15 - 2.59         -2.37 - 1.62         -3.51 - 1.55          -2.4 - 1.96         -3.91 - 2.52 
#> 
#> WEEK 3
#>   n                     31                   31                   26                   26                   37                   37      
#>   Mean (SD)         -0.19 (0.8)         -0.23 (1.39)         -0.03 (0.98)         -0.11 (1.28)          0.05 (0.76)         -0.17 (1.15) 
#>   Median               -0.16                -0.43                -0.17                -0.23                -0.03                -0.22    
#>   IQR              -0.89 - 0.25         -1.19 - 0.45         -0.72 - 0.68         -0.59 - 0.56         -0.39 - 0.62         -0.59 - 0.48 
#>   Min - Max        -1.37 - 1.48         -2.38 - 2.93         -1.74 - 2.42         -2.64 - 2.15         -1.88 - 1.45          -2.55 - 2.6 
#> 
#> WEEK 4
#>   n                     35                   35                   27                   27                   20                   20      
#>   Mean (SD)         0.31 (0.96)          0.17 (1.56)         -0.27 (1.19)          -0.08 (1.4)          0.28 (1.13)         -0.18 (1.32) 
#>   Median               0.41                 -0.09                -0.44                0.09                 0.33                 -0.25    
#>   IQR              -0.48 - 1.03          -0.9 - 1.17         -1.03 - 0.54         -0.86 - 0.93         -0.42 - 1.38         -0.93 - 0.91 
#>   Min - Max         -1.6 - 2.37         -2.21 - 4.75         -2.51 - 2.03         -3.37 - 2.68         -1.56 - 1.89         -3.11 - 2.08 
#> 
#> WEEK 5
#>   n                     25                   25                   24                   24                   22                   22      
#>   Mean (SD)         0.13 (0.94)          0.27 (1.35)         -0.34 (0.75)         -0.43 (1.17)         -0.11 (1.06)         -0.48 (1.51) 
#>   Median               0.29                 0.09                 -0.44                -0.46                -0.01                -0.54    
#>   IQR              -0.46 - 0.72         -0.88 - 1.32         -0.85 - 0.17          -1.3 - 0.31         -0.91 - 0.41         -1.57 - 0.73 
#>   Min - Max         -2.4 - 2.01         -1.86 - 2.99         -1.47 - 1.09         -2.31 - 2.42          -1.77 - 2.2         -3.79 - 1.73 
#> 
#> WEEK 6
#>   n                     13                   13                   19                   19                   25                   25      
#>   Mean (SD)         0.31 (0.97)          0.35 (1.71)          0.3 (0.92)           0.14 (1.01)          0.22 (0.95)          0.05 (1.57) 
#>   Median               0.03                 -0.26                0.15                 0.07                 0.24                 0.02     
#>   IQR               -0.22 - 1.3          -0.9 - 1.21         -0.24 - 0.92         -0.28 - 0.57         -0.47 - 0.78         -0.61 - 1.04 
#>   Min - Max         -1.2 - 1.76         -2.32 - 4.03         -1.32 - 1.91           -2 - 2.35          -1.73 - 2.18         -2.86 - 3.33 
#> 
#> WEEK 7
#>   n                     16                   16                   19                   19                   11                   11      
#>   Mean (SD)        -0.21 (0.98)         -0.26 (1.77)          0.16 (0.78)          0.2 (1.49)          -0.64 (0.98)         -1.06 (1.36) 
#>   Median               -0.21                -0.32                0.05                 -0.16                -0.9                 -1.14    
#>   IQR              -0.94 - 0.52         -1.55 - 1.12         -0.35 - 0.68         -0.52 - 1.02         -1.38 - 0.07         -1.93 - -0.56
#>   Min - Max        -1.67 - 1.18          -3.62 - 2.6         -1.11 - 1.91         -2.17 - 3.24         -2.33 - 0.74         -3.28 - 2.01 
#> 
#> WEEK 8
#>   n                     19                   19                   13                   13                   10                   10      
#>   Mean (SD)        -0.08 (1.14)          0.17 (2.03)         -0.46 (1.19)         -0.59 (1.17)          0.19 (1.09)          -0.2 (1.09) 
#>   Median               -0.31                0.06                 -0.08                -0.67                0.18                 -0.12    
#>   IQR              -0.89 - 0.24         -1.29 - 1.71         -1.31 - 0.15          -1.02 - 0.1         -0.68 - 1.13         -0.67 - 0.62 
#>   Min - Max        -1.66 - 2.18         -3.05 - 4.87         -2.85 - 1.46         -2.65 - 1.37         -1.45 - 1.61         -2.26 - 1.38 
#> 
#> WEEK 9
#>   n                     11                   11                   11                   11                   12                   12      
#>   Mean (SD)         0.13 (1.14)          0.07 (1.85)         -0.67 (1.38)          -0.9 (1.48)          0.48 (1.38)          0.37 (1.47) 
#>   Median               -0.08                0.29                 -0.26                -0.39                1.11                 0.61     
#>   IQR               -0.3 - 0.53         -0.99 - 1.39          -1.09 - 0.2          -1.7 - 0.35          -0.7 - 1.5          -0.42 - 1.18 
#>   Min - Max        -2.04 - 2.25         -3.84 - 2.67         -3.33 - 1.03         -4.03 - 0.46         -1.84 - 1.84           -2.01 - 3  
#> 
#> WEEK 10
#>   n                      5                    5                   10                   10                   10                   10      
#>   Mean (SD)         0.63 (0.66)          0.82 (1.09)         -0.07 (1.04)         -0.11 (1.56)         -0.09 (0.99)         -0.17 (1.33) 
#>   Median               0.51                 0.83                 -0.1                 -0.16                -0.1                 0.21     
#>   IQR               0.43 - 0.79          0.71 - 1.03         -0.55 - 0.73           -1 - 0.71          -0.86 - 0.58         -0.84 - 0.75 
#>   Min - Max         -0.2 - 1.61         -0.76 - 2.29         -1.58 - 1.59         -2.39 - 2.26          -1.6 - 1.34         -2.35 - 1.44 
#> 
#> WEEK 11
#>   n                      4                    4                    5                    5                    5                    5      
#>   Mean (SD)          0 (1.27)            0.25 (1.04)         -0.14 (0.83)         -0.71 (1.03)          0.98 (0.82)          1.26 (0.89) 
#>   Median               0.32                 0.41                 0.22                 -0.81                1.16                  1.8     
#>   IQR              -0.52 - 0.84         -0.33 - 0.99         -0.23 - 0.33         -1.04 - -0.24         0.32 - 1.63          0.56 - 1.83 
#>   Min - Max        -1.74 - 1.11         -1.06 - 1.22         -1.54 - 0.52         -2.13 - 0.66         -0.04 - 1.86          0.07 - 2.06 
# NOT RUN {
Viewer(tbl)
# }