Summary Table by Visit

This function summarizes test results or change from baseline statistics by visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.

t_summarize_by_visit(data, visit, id, col_by)

Arguments

data
data frame with numerical variables to be summarized. If the variable has a label attribute then it will be used as the sub-header column name.

visit
factor with visit names ordered by desired display order in the stacked table.

id
unique subject identifier variable.

col_by
factor with reference and comparison group information, the first level indicates the reference group

Value

an rtable object

Details

Every variable in data will be mapped to a sub-column under col_by column. Multiple variables will be shown side-by-side with the variable label (if present) or variable name as the sub-column header.

Currently n, Mean (SD), Median, IQR and Min-Max will be derived for each variable from data per each factor in col_by.

Examples

# EXAMPLE 1
ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B")
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA
ANL$PCHG <- ANL$CHG/ANL$AVAL*100

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

t_summarize_by_visit(data = ANL[c("AVAL")], visit = ANL$VISIT, col_by = ANL$ARM, id = ANL$USUBJID)

#>                       ARM A               ARM B    
#>                      (N=100)             (N=100)   
#> ---------------------------------------------------
#> visit 1
#>   n                    100                 100     
#>   Mean (SD)         -0.16 (1)          -0.05 (1.11)
#>   Median              -0.01               -0.11    
#>   IQR              -0.79 - 0.42        -0.92 - 0.78
#>   Min - Max        -2.7 - 2.54         -2.9 - 2.16 
#> 
#> visit 2
#>   n                    100                 100     
#>   Mean (SD)        -0.05 (1.04)        -0.19 (1.04)
#>   Median              -0.04                0.01    
#>   IQR              -0.74 - 0.72        -0.94 - 0.6 
#>   Min - Max        -3.08 - 2.14        -3.42 - 1.79
#> 
#> visit 3
#>   n                    100                 100     
#>   Mean (SD)        0.04 (1.12)         -0.28 (1.03)
#>   Median               0.08               -0.31    
#>   IQR              -0.84 - 0.93        -0.89 - 0.45
#>   Min - Max        -2.44 - 2.53        -2.87 - 2.16
#> 
#> visit 4
#>   n                    100                 100     
#>   Mean (SD)        0.08 (0.97)         -0.05 (1.08)
#>   Median               0.1                -0.15    
#>   IQR              -0.44 - 0.77        -0.82 - 0.74
#>   Min - Max        -2.6 - 2.12         -1.75 - 2.79
#> 
#> visit 5
#>   n                    100                 100     
#>   Mean (SD)        0.01 (1.01)         -0.11 (0.98)
#>   Median               0.12               -0.21    
#>   IQR              -0.57 - 0.71         -0.7 - 0.6 
#>   Min - Max        -3.47 - 1.89         -2.6 - 2.4 
#> 
#> visit 6
#>   n                    100                 100     
#>   Mean (SD)        -0.09 (1.09)          0.01 (1)  
#>   Median              -0.08               -0.09    
#>   IQR              -0.7 - 0.61         -0.66 - 0.72
#>   Min - Max        -2.71 - 2.29        -2.45 - 2.37
#> 
#> visit 7
#>   n                    100                 100     
#>   Mean (SD)         0.13 (0.9)         0.14 (0.93) 
#>   Median               0.16                0.14    
#>   IQR              -0.65 - 0.81        -0.4 - 0.81 
#>   Min - Max        -2.31 - 1.91        -2.53 - 2.7 
#> 
#> visit 8
#>   n                    100                 100     
#>   Mean (SD)        -0.21 (1.05)        0.03 (0.84) 
#>   Median              -0.26                0.11    
#>   IQR              -1.03 - 0.42        -0.53 - 0.53
#>   Min - Max        -2.86 - 2.48         -2 - 2.03  
#> 
#> visit 9
#>   n                    100                 100     
#>   Mean (SD)        -0.01 (0.9)         -0.1 (1.06) 
#>   Median              -0.02               -0.09    
#>   IQR              -0.58 - 0.62        -0.84 - 0.59
#>   Min - Max        -2.03 - 2.25        -3.13 - 2.11
#> 
#> visit 10
#>   n                    100                 100     
#>   Mean (SD)         0.1 (1.06)         0.07 (0.99) 
#>   Median              -0.03                0.05    
#>   IQR              -0.53 - 0.87        -0.69 - 0.71
#>   Min - Max        -2.69 - 2.14        -2.9 - 2.49 

t_summarize_by_visit(data = ANL[c("PCHG")], visit = ANL$VISIT, col_by = ANL$ARM, id = ANL$USUBJID)

#>                            ARM A                        ARM B        
#>                           (N=100)                      (N=100)       
#> ---------------------------------------------------------------------
#> visit 1
#>   n                          0                            0          
#>   Mean (SD)               NE (NE)                      NE (NE)       
#>   Median                    NE                           NE          
#>   IQR                     NE - NE                      NE - NE       
#>   Min - Max               NE - NE                      NE - NE       
#> 
#> visit 2
#>   n                         100                          100         
#>   Mean (SD)         -1211.67 (15040.74)          -2442.06 (19976.99) 
#>   Median                  -89.35                        -4.33        
#>   IQR                -478.78 - 344.12              -186.4 - 356.75   
#>   Min - Max        -135175.53 - 57031.4         -182877.37 - 18200.83
#> 
#> visit 3
#>   n                         100                          100         
#>   Mean (SD)           -71.01 (1209.1)             -132.68 (3484.56)  
#>   Median                   -4.87                       -14.42        
#>   IQR                -215.22 - 227.69             -288.36 - 248.31   
#>   Min - Max         -7127.14 - 3064.59          -31745.76 - 10279.67 
#> 
#> visit 4
#>   n                         100                          100         
#>   Mean (SD)          -74.01 (1635.72)             -594.37 (6341.46)  
#>   Median                   68.08                       -84.64        
#>   IQR                -252.26 - 444.13             -385.79 - 224.45   
#>   Min - Max         -10499.12 - 4952.07          -62276.81 - 5758.61 
#> 
#> visit 5
#>   n                         100                          100         
#>   Mean (SD)          241.75 (1604.63)             -252.96 (1223.63)  
#>   Median                  152.88                       -77.44        
#>   IQR                -234.81 - 489.24             -362.38 - 168.17   
#>   Min - Max         -7409.44 - 10389.91          -5004.43 - 5913.57  
#> 
#> visit 6
#>   n                         100                          100         
#>   Mean (SD)          496.94 (4903.92)              15.95 (1022.52)   
#>   Median                  -43.09                        -8.43        
#>   IQR                 -271.84 - 247.3             -356.63 - 258.35   
#>   Min - Max         -7900.22 - 46802.95          -2234.59 - 5971.12  
#> 
#> visit 7
#>   n                         100                          100         
#>   Mean (SD)          422.12 (3190.62)             -352.14 (2579.79)  
#>   Median                   81.85                        53.51        
#>   IQR                -214.91 - 384.93             -316.44 - 297.71   
#>   Min - Max        -10295.72 - 26189.42          -16816.18 - 4159.24 
#> 
#> visit 8
#>   n                         100                          100         
#>   Mean (SD)          141.94 (3470.31)             -216.65 (5898.12)  
#>   Median                  -18.87                        -3.23        
#>   IQR                -251.37 - 278.67              -299.25 - 385.6   
#>   Min - Max         -20616.54 - 26194.4         -50220.12 - 18314.34 
#> 
#> visit 9
#>   n                         100                          100         
#>   Mean (SD)           175.55 (1562.3)              -2.96 (1502.78)   
#>   Median                   35.19                       -64.93        
#>   IQR                 -264.14 - 236.6             -266.18 - 210.57   
#>   Min - Max         -3715.07 - 8488.52           -6389.39 - 7159.99  
#> 
#> visit 10
#>   n                         100                          100         
#>   Mean (SD)          -276.56 (1993.73)            -149.44 (7689.76)  
#>   Median                   3.17                         13.62        
#>   IQR                 -290.7 - 217.7              -266.53 - 376.25   
#>   Min - Max         -15084.76 - 3027.89         -53332.42 - 30157.02 

# Add label to variable instead showing variable name
ANL <- var_relabel(ANL, AVAL = "Value at\nVisit",
                        CHG = "Change from\nBaseline",
                        PCHG = "Percent Change\nfrom Baseline")

t_summarize_by_visit(
  data = ANL[c("AVAL", "CHG")],
  visit = ANL$VISIT,
  col_by = ANL$ARM,
  id = ANL$USUBJID
)

#>                                 ARM A                                   ARM B              
#>                                (N=100)                                 (N=100)             
#>                      Value at          Change from           Value at          Change from 
#>                       Visit              Baseline             Visit              Baseline  
#> -------------------------------------------------------------------------------------------
#> visit 1
#>   n                    100                  0                  100                  0      
#>   Mean (SD)         -0.16 (1)            NE (NE)           -0.05 (1.11)          NE (NE)   
#>   Median              -0.01                 NE                -0.11                 NE     
#>   IQR              -0.79 - 0.42          NE - NE           -0.92 - 0.78          NE - NE   
#>   Min - Max        -2.7 - 2.54           NE - NE           -2.9 - 2.16           NE - NE   
#> 
#> visit 2
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.05 (1.04)        2.35 (2.08)         -0.19 (1.04)         1.9 (2.05) 
#>   Median              -0.04                2.33                0.01                1.72    
#>   IQR              -0.74 - 0.72        1.18 - 3.73         -0.94 - 0.6          0.5 - 3.35 
#>   Min - Max        -3.08 - 2.14        -3.36 - 7.03        -3.42 - 1.79        -2.51 - 6.5 
#> 
#> visit 3
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.04 (1.12)         1.99 (1.79)         -0.28 (1.03)         1.89 (2.1) 
#>   Median               0.08                1.95               -0.31                1.91    
#>   IQR              -0.84 - 0.93        0.97 - 3.05         -0.89 - 0.45        0.25 - 3.55 
#>   Min - Max        -2.44 - 2.53        -4.72 - 6.44        -2.87 - 2.16        -3.52 - 8.67
#> 
#> visit 4
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.08 (0.97)         1.78 (2.12)         -0.05 (1.08)        2.28 (2.07) 
#>   Median               0.1                 1.91               -0.15                2.25    
#>   IQR              -0.44 - 0.77        0.47 - 3.23         -0.82 - 0.74        0.81 - 3.81 
#>   Min - Max        -2.6 - 2.12         -3.02 - 8.49        -1.75 - 2.79        -2.66 - 6.74
#> 
#> visit 5
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.01 (1.01)          2.2 (1.85)         -0.11 (0.98)         1.97 (1.9) 
#>   Median               0.12                2.19               -0.21                1.86    
#>   IQR              -0.57 - 0.71         0.9 - 3.47          -0.7 - 0.6          0.7 - 3.4  
#>   Min - Max        -3.47 - 1.89        -1.89 - 6.86         -2.6 - 2.4         -2.45 - 6.23
#> 
#> visit 6
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.09 (1.09)        1.84 (2.14)           0.01 (1)          2.07 (2.12) 
#>   Median              -0.08                1.81               -0.09                2.3     
#>   IQR              -0.7 - 0.61          0.4 - 3.19         -0.66 - 0.72         0.68 - 3.2 
#>   Min - Max        -2.71 - 2.29        -3.43 - 7.19        -2.45 - 2.37        -3.35 - 7.01
#> 
#> visit 7
#>   n                    100                 100                 100                 100     
#>   Mean (SD)         0.13 (0.9)         1.63 (2.01)         0.14 (0.93)         1.89 (2.05) 
#>   Median               0.16                1.82                0.14                2.09    
#>   IQR              -0.65 - 0.81        0.41 - 3.01         -0.4 - 0.81         0.54 - 3.38 
#>   Min - Max        -2.31 - 1.91        -5.08 - 6.45        -2.53 - 2.7         -3.12 - 5.9 
#> 
#> visit 8
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.21 (1.05)        2.13 (1.87)         0.03 (0.84)         2.08 (1.97) 
#>   Median              -0.26                2.03                0.11                1.98    
#>   IQR              -1.03 - 0.42         0.8 - 3.47         -0.53 - 0.53         0.5 - 3.3  
#>   Min - Max        -2.86 - 2.48        -1.85 - 7.49         -2 - 2.03          -1.62 - 7.32
#> 
#> visit 9
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.01 (0.9)         1.67 (1.99)         -0.1 (1.06)          1.75 (2.2) 
#>   Median              -0.02                1.42               -0.09                1.95    
#>   IQR              -0.58 - 0.62        0.22 - 2.89         -0.84 - 0.59        0.34 - 3.58 
#>   Min - Max        -2.03 - 2.25        -2.59 - 8.03        -3.13 - 2.11        -3.16 - 6.41
#> 
#> visit 10
#>   n                    100                 100                 100                 100     
#>   Mean (SD)         0.1 (1.06)         2.06 (2.05)         0.07 (0.99)         1.69 (2.04) 
#>   Median              -0.03                1.77                0.05                1.9     
#>   IQR              -0.53 - 0.87         0.27 - 3.5         -0.69 - 0.71        0.17 - 2.87 
#>   Min - Max        -2.69 - 2.14        -2.14 - 7.58        -2.9 - 2.49         -3.59 - 6.84

# EXAMPLE 2
library(random.cdisc.data)
library(dplyr)

AQS <- radam("AQS") %>%
   filter(PARAMCD == "BFIALL") %>%
   mutate(
     ARM = factor(sample(c("ARM A", "ARM B", "ARM C"), nrow(.), replace = TRUE)),
     AVISIT = factor(AVISIT, levels = unique(AVISIT))
   )

tbl <- t_summarize_by_visit(
  data = AQS[c("AVAL", "CHG")],
  visit = AQS$AVISIT,
  col_by = AQS$ARM,
  id = AQS$USUBJID
)
tbl

#>                                  ARM A                                     ARM B                                     ARM C               
#>                                  (N=87)                                    (N=93)                                    (N=85)              
#>                        AVAL                  CHG                 AVAL                  CHG                 AVAL                  CHG     
#> -----------------------------------------------------------------------------------------------------------------------------------------
#> WEEK 1
#>   n                     34                   34                   32                   32                   34                   34      
#>   Mean (SD)          -0.03 (1)              0 (0)            -0.18 (1.15)             0 (0)             0.15 (1.16)             0 (0)    
#>   Median               -0.17                  0                  -0.19                  0                  0.05                   0      
#>   IQR              -0.83 - 0.83             0 - 0            -0.87 - 0.17             0 - 0            -0.46 - 0.93             0 - 0    
#>   Min - Max        -2.13 - 2.14             0 - 0            -2.41 - 2.67             0 - 0            -2.59 - 3.14             0 - 0    
#> 
#> WEEK 2
#>   n                     27                   27                   43                   43                   30                   30      
#>   Mean (SD)        -0.01 (1.16)         -0.14 (1.52)          0.27 (1.13)          0.11 (1.74)         -0.19 (0.85)          0.21 (1.32) 
#>   Median               0.05                 -0.14                0.03                 0.25                 -0.14                0.16     
#>   IQR              -0.56 - 0.57          -0.94 - 0.7         -0.42 - 1.28         -0.99 - 1.35         -0.71 - 0.36         -0.99 - 1.12 
#>   Min - Max        -2.67 - 2.63          -3.58 - 3.7         -1.75 - 2.96         -4.36 - 3.85         -2.12 - 2.26         -1.68 - 3.15 
#> 
#> WEEK 3
#>   n                     23                   23                   30                   30                   38                   38      
#>   Mean (SD)         0.11 (1.14)         -0.13 (1.46)         -0.43 (1.08)         -0.52 (1.41)           0 (1.12)            0.23 (1.49) 
#>   Median               0.13                 0.01                 -0.43                -0.59                  0                  0.56     
#>   IQR              -0.61 - 0.73         -1.17 - 1.19         -1.01 - 0.54         -1.15 - 0.14         -0.57 - 0.72         -0.87 - 1.23 
#>   Min - Max        -2.72 - 2.05         -3.64 - 2.16         -3.54 - 1.22          -3.3 - 3.3          -2.44 - 2.64         -3.38 - 3.08 
#> 
#> WEEK 4
#>   n                     29                   29                   29                   29                   23                   23      
#>   Mean (SD)         -0.3 (0.82)         -0.57 (1.63)          0.22 (0.88)          0.49 (1.5)          -0.08 (1.18)          0.14 (1.37) 
#>   Median               -0.3                 -0.53                0.37                 0.82                 -0.03                -0.18    
#>   IQR               -0.99 - 0.4         -1.03 - 0.37         -0.05 - 0.69          -0.4 - 1.29          -0.7 - 0.8          -0.54 - 1.01 
#>   Min - Max        -2.01 - 1.17         -3.66 - 2.32         -2.47 - 1.43         -3.42 - 2.79         -2.61 - 1.76         -2.62 - 3.06 
#> 
#> WEEK 5
#>   n                     31                   31                   21                   21                   21                   21      
#>   Mean (SD)         0.07 (0.85)         -0.06 (1.35)         -0.04 (0.97)         -0.11 (1.45)           -0.34 (1)           0.15 (0.99) 
#>   Median                 0                  -0.16                -0.04                0.31                 -0.22                0.22     
#>   IQR              -0.63 - 0.64         -0.98 - 1.03         -0.53 - 0.73           -0.82 - 1          -0.87 - -0.04        -0.21 - 0.51 
#>   Min - Max        -1.51 - 2.37         -4.15 - 2.53         -2.21 - 1.26         -3.13 - 1.77         -2.23 - 1.59         -1.92 - 2.87 
#> 
#> WEEK 6
#>   n                     22                   22                   18                   18                   16                   16      
#>   Mean (SD)         0.12 (1.01)          0.47 (1.74)          0.25 (1.03)          0.27 (1.77)         -0.46 (1.01)         -0.39 (0.91) 
#>   Median               0.11                 0.82                 0.23                 0.15                 -0.53                -0.54    
#>   IQR               -0.53 - 0.9         -1.02 - 1.49         -0.57 - 0.95         -0.76 - 1.24         -0.89 - -0.08        -1.22 - 0.33 
#>   Min - Max        -1.68 - 1.73         -2.82 - 4.29         -1.75 - 1.99         -3.55 - 3.21          -2.5 - 1.88         -1.62 - 1.04 
#> 
#> WEEK 7
#>   n                     13                   13                   16                   16                   20                   20      
#>   Mean (SD)        -0.02 (0.95)          0.44 (1.31)          0.1 (0.85)           0.37 (1.29)          -0.01 (0.8)         -0.21 (1.23) 
#>   Median               -0.13                0.12                 0.02                 0.21                 -0.15                -0.2     
#>   IQR              -0.55 - 0.42         -0.36 - 1.58         -0.37 - 0.61         -0.22 - 1.05         -0.47 - 0.59         -0.93 - 0.42 
#>   Min - Max        -1.28 - 2.43         -1.43 - 2.46         -1.32 - 1.62         -2.76 - 2.66         -1.84 - 1.24         -2.39 - 2.42 
#> 
#> WEEK 8
#>   n                     15                   15                   11                   11                   11                   11      
#>   Mean (SD)         0.19 (1.02)          0.11 (1.51)          0.18 (1.02)         -0.21 (1.52)          0.47 (1.09)          0.93 (1.36) 
#>   Median               0.23                 0.36                 0.23                 -0.24                0.48                 0.91     
#>   IQR              -0.56 - 0.94         -0.68 - 0.95         -0.32 - 1.04          -1.25 - 0.6         -0.21 - 1.26         -0.16 - 1.69 
#>   Min - Max        -1.72 - 2.13         -2.91 - 2.62          -1.72 - 1.5         -2.76 - 2.29          -1.4 - 2.08         -0.78 - 3.07 
#> 
#> WEEK 9
#>   n                      6                    6                    9                    9                    8                    8      
#>   Mean (SD)         0.79 (1.3)           0.62 (0.99)         -0.14 (0.57)           0.04 (1)            0.36 (1.09)         -0.11 (2.11) 
#>   Median               0.62                 0.52                 -0.09                0.03                 0.33                 0.09     
#>   IQR               0.02 - 1.8          -0.16 - 1.36         -0.31 - 0.17         -0.12 - 0.12         -0.15 - 1.08         -1.44 - 1.52 
#>   Min - Max        -0.93 - 2.38         -0.49 - 1.92          -1.4 - 0.51         -2.12 - 1.33         -1.37 - 1.71         -3.38 - 2.63 
#> 
#> WEEK 10
#>   n                      9                    9                    3                    3                    2                    2      
#>   Mean (SD)        -0.64 (1.35)         -0.37 (2.02)          0.76 (2.15)         -0.48 (2.25)         -0.32 (0.02)         -0.35 (0.25) 
#>   Median               -0.92                -0.97                0.18                 -0.89                -0.32                -0.35    
#>   IQR              -1.55 - 0.57         -1.41 - 0.26         -0.43 - 1.66         -1.69 - 0.53         -0.33 - -0.31        -0.44 - -0.27
#>   Min - Max        -2.27 - 1.64         -2.66 - 4.23         -1.03 - 3.15         -2.49 - 1.95         -0.34 - -0.31        -0.53 - -0.18
#> 
#> WEEK 11
#>   n                      5                    5                    3                    3                    1                    1      
#>   Mean (SD)         -0.31 (0.8)         -0.44 (1.41)         -0.61 (1.25)         -0.74 (1.73)           0.72 (NA)            -0.1 (NA)  
#>   Median               0.02                 -0.17                -1.32                -1.64                0.72                 -0.1     
#>   IQR               -0.77 - 0.1         -1.68 - 0.17         -1.33 - -0.24        -1.73 - -0.19         0.72 - 0.72          -0.1 - -0.1 
#>   Min - Max        -1.46 - 0.56         -1.97 - 1.46         -1.34 - 0.84         -1.83 - 1.25          0.72 - 0.72          -0.1 - -0.1 
# NOT RUN {
Viewer(tbl)
# }