Summary Table by Visit

This function summarizes test results or change from baseline statistics by visit. Corresponds to STREAM table templates EGT01, VST01 and LBT01.

t_summarize_by_visit(data, visit, id, col_by)

Arguments

data
data frame with numerical variables to be summarized. If the variable has a label attribute then it will be used as the sub-header column name.

visit
factor with visit names ordered by desired display order in the stacked table.

id
unique subject identifier variable.

col_by
factor with reference and comparison group information, the first level indicates the reference group

Value

an rtable object

Details

Every variable in data will be mapped to a sub-column under col_by column. Multiple variables will be shown side-by-side with the variable label (if present) or variable name as the sub-column header.

Currently n, Mean (SD), Median, IQR and Min-Max will be derived for each variable from data per each factor in col_by.

Examples

# EXAMPLE 1
ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B")
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA
ANL$PCHG <- ANL$CHG/ANL$AVAL*100

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

t_summarize_by_visit(data = ANL[c("AVAL")], visit = ANL$VISIT, col_by = ANL$ARM, id = ANL$USUBJID)

#>                       ARM A               ARM B    
#>                      (N=100)             (N=100)   
#> ---------------------------------------------------
#> visit 1
#>   n                    100                 100     
#>   Mean (SD)        -0.01 (0.98)        -0.09 (0.88)
#>   Median              -0.07               -0.13    
#>   IQR              -0.77 - 0.76        -0.67 - 0.56
#>   Min - Max        -2.05 - 2.95        -2.28 - 2.32
#> 
#> visit 2
#>   n                    100                 100     
#>   Mean (SD)         0.02 (1.1)         -0.04 (0.99)
#>   Median              -0.04               -0.09    
#>   IQR              -0.68 - 0.96        -0.79 - 0.59
#>   Min - Max        -2.73 - 2.74        -2.08 - 2.5 
#> 
#> visit 3
#>   n                    100                 100     
#>   Mean (SD)        -0.07 (1.04)        0.11 (1.05) 
#>   Median                0                  0.14    
#>   IQR              -0.8 - 0.61         -0.76 - 0.78
#>   Min - Max        -2.32 - 3.41        -2.06 - 2.83
#> 
#> visit 4
#>   n                    100                 100     
#>   Mean (SD)        -0.18 (1.11)        0.04 (0.95) 
#>   Median              -0.17                0.09    
#>   IQR              -0.92 - 0.63        -0.53 - 0.73
#>   Min - Max        -3.31 - 2.19        -2.36 - 1.91
#> 
#> visit 5
#>   n                    100                 100     
#>   Mean (SD)        -0.03 (0.93)          0.09 (1)  
#>   Median              -0.09                0.04    
#>   IQR              -0.53 - 0.56        -0.39 - 0.77
#>   Min - Max        -2.05 - 2.69        -2.98 - 2.16
#> 
#> visit 6
#>   n                    100                 100     
#>   Mean (SD)        -0.01 (0.99)        -0.22 (1.01)
#>   Median               0.09               -0.21    
#>   IQR              -0.73 - 0.68        -0.94 - 0.42
#>   Min - Max        -2.96 - 2.08        -2.76 - 2.24
#> 
#> visit 7
#>   n                    100                 100     
#>   Mean (SD)        -0.02 (1.05)          0 (0.89)  
#>   Median               0.02                0.11    
#>   IQR              -0.59 - 0.65        -0.64 - 0.55
#>   Min - Max        -3.4 - 2.28         -2.43 - 2.16
#> 
#> visit 8
#>   n                    100                 100     
#>   Mean (SD)        -0.14 (0.94)        0.01 (0.97) 
#>   Median               -0.1               -0.07    
#>   IQR              -0.83 - 0.4         -0.69 - 0.59
#>   Min - Max        -2.16 - 2.72        -2.16 - 2.89
#> 
#> visit 9
#>   n                    100                 100     
#>   Mean (SD)        0.13 (1.04)         0.01 (1.06) 
#>   Median               0.17                0.01    
#>   IQR              -0.61 - 0.89        -0.66 - 0.78
#>   Min - Max        -2.85 - 3.13        -2.14 - 2.68
#> 
#> visit 10
#>   n                    100                 100     
#>   Mean (SD)          0 (0.99)          -0.1 (1.14) 
#>   Median               0.07               -0.13    
#>   IQR              -0.69 - 0.52        -0.88 - 0.79
#>   Min - Max        -2.36 - 2.53        -2.58 - 2.31

t_summarize_by_visit(data = ANL[c("PCHG")], visit = ANL$VISIT, col_by = ANL$ARM, id = ANL$USUBJID)

#>                            ARM A                        ARM B        
#>                           (N=100)                      (N=100)       
#> ---------------------------------------------------------------------
#> visit 1
#>   n                          0                            0          
#>   Mean (SD)               NE (NE)                      NE (NE)       
#>   Median                    NE                           NE          
#>   IQR                     NE - NE                      NE - NE       
#>   Min - Max               NE - NE                      NE - NE       
#> 
#> visit 2
#>   n                         100                          100         
#>   Mean (SD)           90.49 (2418.01)             1247.62 (9245.59)  
#>   Median                   40.66                        30.04        
#>   IQR                 -232.8 - 310.84             -163.06 - 267.24   
#>   Min - Max        -11646.25 - 13857.77          -3470.62 - 90561.26 
#> 
#> visit 3
#>   n                         100                          100         
#>   Mean (SD)          316.61 (1393.27)              30.18 (1037.47)   
#>   Median                   17.55                        79.66        
#>   IQR                -217.62 - 352.72             -280.92 - 303.67   
#>   Min - Max         -1948.17 - 10885.93          -4730.03 - 4468.58  
#> 
#> visit 4
#>   n                         100                          100         
#>   Mean (SD)          -192.2 (3811.93)             -169.35 (1307.28)  
#>   Median                  -54.57                        -1.06        
#>   IQR                 -348.4 - 265.29             -279.47 - 301.76   
#>   Min - Max        -26381.14 - 23892.48           -6656.9 - 3835.67  
#> 
#> visit 5
#>   n                         100                          100         
#>   Mean (SD)          -770.3 (7207.99)             -448.78 (4690.99)  
#>   Median                   -2.82                        5.95         
#>   IQR                -239.03 - 306.34             -427.14 - 308.96   
#>   Min - Max         -71648.44 - 2142.89         -18621.05 - 30222.84 
#> 
#> visit 6
#>   n                         100                          100         
#>   Mean (SD)           37.85 (2325.08)            -1940.71 (18151.26) 
#>   Median                   -29.1                       -102.58       
#>   IQR                -256.12 - 330.15             -335.66 - 211.19   
#>   Min - Max         -9475.8 - 16987.01          -181084.38 - 2826.88 
#> 
#> visit 7
#>   n                         100                          100         
#>   Mean (SD)           -877 (8100.13)              218.82 (1415.17)   
#>   Median                  -11.42                        87.44        
#>   IQR                -339.56 - 338.23              -227.06 - 489.6   
#>   Min - Max         -80002.49 - 3720.46          -9303.19 - 5644.46  
#> 
#> visit 8
#>   n                         100                          100         
#>   Mean (SD)          -247.47 (2117.54)            312.16 (2123.05)   
#>   Median                  -24.16                        -1.02        
#>   IQR                 -264.93 - 344.7             -231.68 - 228.18   
#>   Min - Max         -13466.07 - 4796.61          -5533.03 - 14741.33 
#> 
#> visit 9
#>   n                         100                          100         
#>   Mean (SD)          -14.78 (1389.48)              259.69 (1746.5)   
#>   Median                   27.11                        -0.78        
#>   IQR                -193.71 - 304.85             -350.55 - 285.88   
#>   Min - Max         -5792.35 - 8259.17           -2741.38 - 11941.88 
#> 
#> visit 10
#>   n                         100                          100         
#>   Mean (SD)        -70278.13 (703347.08)          -36.81 (1202.69)   
#>   Median                   48.8                        -61.53        
#>   IQR                -301.97 - 307.19             -326.25 - 135.23   
#>   Min - Max        -7033410.58 - 2244.02         -3781.46 - 9583.14  

# Add label to variable instead showing variable name
ANL <- var_relabel(ANL, AVAL = "Value at\nVisit",
                        CHG = "Change from\nBaseline",
                        PCHG = "Percent Change\nfrom Baseline")
t_summarize_by_visit(
  data = ANL[c("AVAL", "CHG")],
  visit = ANL$VISIT,
  col_by = ANL$ARM,
  id = ANL$USUBJID
)

#>                                 ARM A                                   ARM B              
#>                                (N=100)                                 (N=100)             
#>                      Value at          Change from           Value at          Change from 
#>                       Visit              Baseline             Visit              Baseline  
#> -------------------------------------------------------------------------------------------
#> visit 1
#>   n                    100                  0                  100                  0      
#>   Mean (SD)        -0.01 (0.98)          NE (NE)           -0.09 (0.88)          NE (NE)   
#>   Median              -0.07                 NE                -0.13                 NE     
#>   IQR              -0.77 - 0.76          NE - NE           -0.67 - 0.56          NE - NE   
#>   Min - Max        -2.05 - 2.95          NE - NE           -2.28 - 2.32          NE - NE   
#> 
#> visit 2
#>   n                    100                 100                 100                 100     
#>   Mean (SD)         0.02 (1.1)         1.83 (2.04)         -0.04 (0.99)        1.74 (2.18) 
#>   Median              -0.04                1.9                -0.09                1.53    
#>   IQR              -0.68 - 0.96        0.72 - 3.26         -0.79 - 0.59         0.22 - 3.1 
#>   Min - Max        -2.73 - 2.74        -4.15 - 7.02        -2.08 - 2.5         -2.19 - 8.08
#> 
#> visit 3
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.07 (1.04)        2.03 (2.05)         0.11 (1.05)         2.42 (1.85) 
#>   Median                0                  1.67                0.14                2.56    
#>   IQR              -0.8 - 0.61         0.66 - 2.92         -0.76 - 0.78        1.15 - 3.95 
#>   Min - Max        -2.32 - 3.41        -2.46 - 7.59        -2.06 - 2.83        -2.51 - 5.84
#> 
#> visit 4
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.18 (1.11)         2.1 (2.28)         0.04 (0.95)         1.93 (1.84) 
#>   Median              -0.17                2.27                0.09                1.74    
#>   IQR              -0.92 - 0.63        0.47 - 3.89         -0.53 - 0.73        0.72 - 3.16 
#>   Min - Max        -3.31 - 2.19        -3.2 - 8.35         -2.36 - 1.91        -3.06 - 6.12
#> 
#> visit 5
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.03 (0.93)        1.68 (2.13)           0.09 (1)          2.16 (1.92) 
#>   Median              -0.09                1.37                0.04                1.85    
#>   IQR              -0.53 - 0.56        0.44 - 3.14         -0.39 - 0.77          1 - 3.3   
#>   Min - Max        -2.05 - 2.69        -3.27 - 7.11        -2.98 - 2.16        -2.58 - 8.42
#> 
#> visit 6
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.01 (0.99)        2.07 (1.96)         -0.22 (1.01)        1.75 (1.79) 
#>   Median               0.09                2.31               -0.21                1.86    
#>   IQR              -0.73 - 0.68        0.87 - 3.22         -0.94 - 0.42        0.82 - 2.88 
#>   Min - Max        -2.96 - 2.08        -3.86 - 6.03        -2.76 - 2.24        -3.06 - 6.18
#> 
#> visit 7
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.02 (1.05)        1.74 (2.13)           0 (0.89)            2.04 (2)  
#>   Median               0.02                1.71                0.11                2.02    
#>   IQR              -0.59 - 0.65        0.29 - 3.25         -0.64 - 0.55        0.68 - 3.43 
#>   Min - Max        -3.4 - 2.28         -3.56 - 5.96        -2.43 - 2.16        -2.51 - 7.85
#> 
#> visit 8
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        -0.14 (0.94)        1.79 (1.93)         0.01 (0.97)         1.72 (2.09) 
#>   Median               -0.1                1.92               -0.07                1.76    
#>   IQR              -0.83 - 0.4         0.36 - 2.76         -0.69 - 0.59        0.14 - 2.88 
#>   Min - Max        -2.16 - 2.72        -2.79 - 6.59        -2.16 - 2.89        -3.71 - 6.59
#> 
#> visit 9
#>   n                    100                 100                 100                 100     
#>   Mean (SD)        0.13 (1.04)          1.84 (1.9)         0.01 (1.06)         1.87 (2.23) 
#>   Median               0.17                1.96                0.01                2.14    
#>   IQR              -0.61 - 0.89        0.45 - 3.12         -0.66 - 0.78        0.04 - 3.51 
#>   Min - Max        -2.85 - 3.13        -3.15 - 6.06        -2.14 - 2.68        -4.27 - 6.43
#> 
#> visit 10
#>   n                    100                 100                 100                 100     
#>   Mean (SD)          0 (0.99)          1.83 (2.11)         -0.1 (1.14)         1.82 (1.87) 
#>   Median               0.07                1.71               -0.13                1.61    
#>   IQR              -0.69 - 0.52        0.64 - 3.22         -0.88 - 0.79        0.67 - 2.99 
#>   Min - Max        -2.36 - 2.53        -3.1 - 7.06         -2.58 - 2.31        -2.35 - 6.22

# EXAMPLE 2
library(random.cdisc.data)
library(dplyr)

AQS <- radam("AQS") %>%
   filter(PARAMCD == "BFIALL") %>%
   mutate(
     ARM = factor(sample(c("ARM A", "ARM B", "ARM C"), nrow(.), replace = TRUE)),
     AVISIT = factor(AVISIT, levels = unique(AVISIT))
   )

tbl <- t_summarize_by_visit(
  data = AQS[c("AVAL", "CHG")],
  visit = AQS$AVISIT,
  col_by = AQS$ARM,
  id = AQS$USUBJID
)
tbl

#>                                  ARM A                                     ARM B                                     ARM C               
#>                                  (N=89)                                    (N=87)                                    (N=90)              
#>                        AVAL                  CHG                 AVAL                  CHG                 AVAL                  CHG     
#> -----------------------------------------------------------------------------------------------------------------------------------------
#> WEEK 1
#>   n                     28                   28                   31                   31                   41                   41      
#>   Mean (SD)         0.04 (0.76)             0 (0)            -0.04 (1.15)             0 (0)            -0.07 (1.08)             0 (0)    
#>   Median               -0.02                  0                  0.09                   0                  0.07                   0      
#>   IQR              -0.45 - 0.79             0 - 0            -0.78 - 0.59             0 - 0            -0.61 - 0.76             0 - 0    
#>   Min - Max        -1.51 - 1.21             0 - 0            -2.61 - 2.82             0 - 0            -3.25 - 1.94             0 - 0    
#> 
#> WEEK 2
#>   n                     37                   37                   29                   29                   34                   34      
#>   Mean (SD)        -0.21 (0.95)         -0.21 (1.42)          0.06 (1.12)           0 (1.44)            0.2 (0.77)           0.34 (1.53) 
#>   Median               -0.36                -0.32                -0.02                -0.1                 0.12                 0.31     
#>   IQR              -0.87 - 0.38         -1.23 - 0.58          -0.44 - 0.7          -0.7 - 0.82         -0.16 - 0.76         -0.85 - 1.13 
#>   Min - Max         -1.8 - 1.88         -2.88 - 3.02         -2.34 - 2.62          -2.7 - 2.58          -1.6 - 1.91         -2.72 - 3.45 
#> 
#> WEEK 3
#>   n                     26                   26                   39                   39                   28                   28      
#>   Mean (SD)         -0.23 (0.9)          0.1 (1.51)            -0.14 (1)          -0.17 (1.48)          0.02 (1.07)         -0.15 (1.37) 
#>   Median               -0.43                -0.01                -0.01                0.17                 -0.23                -0.27    
#>   IQR              -1.06 - 0.57         -1.04 - 0.84         -0.76 - 0.67         -0.88 - 0.66         -0.72 - 0.62         -0.94 - 0.27 
#>   Min - Max        -1.39 - 1.29         -2.38 - 4.41         -2.65 - 1.68          -3.4 - 3.07         -1.78 - 2.73         -2.21 - 3.33 
#> 
#> WEEK 4
#>   n                     29                   29                   26                   26                   32                   32      
#>   Mean (SD)        -0.05 (1.07)          -0.1 (1.61)          0.2 (0.81)           0.4 (1.27)          -0.17 (0.99)          -0.19 (1.3) 
#>   Median               -0.04                -0.08                 0.1                 0.52                 -0.15                -0.23    
#>   IQR               -0.5 - 0.42          -1.5 - 1.15         -0.33 - 0.77         -0.28 - 1.01         -0.92 - 0.34         -1.11 - 0.52 
#>   Min - Max        -2.05 - 2.38         -2.63 - 3.19         -1.18 - 1.75         -1.84 - 4.61         -1.88 - 2.46         -2.86 - 2.19 
#> 
#> WEEK 5
#>   n                     23                   23                   25                   25                   31                   31      
#>   Mean (SD)         0.07 (1.03)          0.21 (1.02)         -0.25 (0.79)         -0.09 (1.22)          0.08 (1.16)          0.07 (1.6)  
#>   Median               0.19                 0.32                 -0.14                -0.17                0.23                 0.14     
#>   IQR              -0.55 - 0.64           -0.28 - 1            -1 - 0.34          -0.95 - 0.79         -0.73 - 0.89          -0.8 - 0.91 
#>   Min - Max        -2.44 - 1.97         -2.37 - 1.93          -1.88 - 1.3         -2.54 - 2.62         -2.63 - 2.03         -4.05 - 4.11 
#> 
#> WEEK 6
#>   n                     22                   22                   22                   22                   19                   19      
#>   Mean (SD)        -0.03 (0.97)          0.12 (1.28)          0.03 (1.01)         -0.13 (1.27)         -0.52 (0.84)         -0.62 (1.44) 
#>   Median               0.07                 0.12                 0.12                 0.03                 -0.52                -0.61    
#>   IQR              -0.92 - 0.84         -0.54 - 0.89          -0.65 - 0.6         -0.96 - 0.78          -1 - -0.33          -1.84 - 0.26 
#>   Min - Max        -1.44 - 1.63          -2.27 - 2.4         -1.95 - 2.08         -2.46 - 1.78         -2.36 - 1.34         -2.55 - 2.08 
#> 
#> WEEK 7
#>   n                     19                   19                   13                   13                   22                   22      
#>   Mean (SD)         0.39 (1.01)          0.44 (1.38)         -0.09 (0.84)         -0.04 (1.26)          0.06 (0.99)          0.01 (1.33) 
#>   Median               0.29                 0.28                 0.05                 -0.17                  0                  0.12     
#>   IQR              -0.15 - 1.29          -0.71 - 1.4         -0.52 - 0.26         -0.81 - 0.49         -0.72 - 0.63         -0.73 - 0.47 
#>   Min - Max        -1.46 - 2.01         -2.47 - 2.88         -1.69 - 1.23         -1.89 - 2.34         -1.58 - 2.31         -2.68 - 3.84 
#> 
#> WEEK 8
#>   n                     19                   19                    8                    8                   19                   19      
#>   Mean (SD)        -0.27 (1.19)         -0.01 (1.62)          0.06 (1.02)           0.3 (1.6)           0.09 (0.76)          0.04 (1.12) 
#>   Median               -0.07                0.48                 0.02                 0.69                 0.15                 0.27     
#>   IQR               -0.91 - 0.5           -1.31 - 1          -0.27 - 0.23         -1.26 - 1.72         -0.51 - 0.62         -0.47 - 0.59 
#>   Min - Max         -2.37 - 1.8         -3.98 - 1.99         -1.63 - 2.02         -1.97 - 1.99         -1.29 - 1.41         -2.71 - 2.06 
#> 
#> WEEK 9
#>   n                     17                   17                    9                    9                   12                   12      
#>   Mean (SD)         0.05 (0.63)         -0.06 (1.13)          0.01 (1.34)          0.05 (1.65)          0.64 (0.91)          0.95 (1.34) 
#>   Median               0.13                 -0.12                0.48                 0.21                 0.84                 1.07     
#>   IQR               -0.2 - 0.43         -0.65 - 0.61         -1.35 - 0.86         -0.02 - 0.87         -0.03 - 1.24          0.31 - 1.87 
#>   Min - Max        -0.99 - 1.25         -2.19 - 1.65         -1.89 - 1.75         -3.31 - 2.37         -1.03 - 1.95         -1.94 - 2.65 
#> 
#> WEEK 10
#>   n                      7                    7                   10                   10                    6                    6      
#>   Mean (SD)         0.04 (0.47)         -0.04 (1.36)          0.24 (0.62)          0.64 (1.1)           0.43 (1.35)         -0.09 (1.99) 
#>   Median               0.16                 -0.36                0.25                 0.63                 0.01                 -0.33    
#>   IQR              -0.04 - 0.35         -0.52 - 0.75         -0.14 - 0.41          0.26 - 0.85         -0.54 - 1.58         -1.13 - 1.44 
#>   Min - Max        -0.93 - 0.44         -2.28 - 1.88         -0.85 - 1.38         -1.68 - 2.43           -1 - 2.16           -2.91 - 2.4 
#> 
#> WEEK 11
#>   n                      5                    5                    5                    5                    3                    3      
#>   Mean (SD)         0.09 (0.93)          0.22 (1.45)          0.91 (0.92)          1.32 (1.35)          -0.56 (0.7)          0.12 (0.17) 
#>   Median               0.27                 0.46                 0.97                  0.9                 -0.37                0.18     
#>   IQR              -0.13 - 0.85          -0.5 - 1.19          0.81 - 1.26          0.55 - 2.68         -0.85 - -0.18         0.05 - 0.21 
#>   Min - Max        -1.39 - 0.85         -1.87 - 1.81         -0.52 - 2.01         -0.28 - 2.76         -1.34 - 0.01         -0.07 - 0.25 
# NOT RUN {
Viewer(tbl)
# }