Summarize Variables

Summarize Variables

Similiar as the demographic table in STREAM. This funtion is deprecated, please use t_summary instead.

t_summarize_variables(data, col_by, total = NULL, drop_levels = TRUE,
  useNA_factors = c("no", "ifany", "always"))

t_summarize_variables(data, col_by, total = NULL, drop_levels = TRUE,
  useNA_factors = c("no", "ifany", "always"))

Arguments

data
data frame with variables to be summarized as described in the details section. If the variable has a label attribute then it will be used for the row name.

col_by
a factor of length nrow(x) that defines which levels in col_by define a column. If data should not be split into columns use the no_by function.

total
if not NULL then it must be a string and an addition column will be added with the overall summaries

drop_levels
boolean, drop zero-count levels for factors

useNA_factors
forwarded to useNA of t_summary.factor

Details

Every variable in data will be mapped to a summary table of that variable and the tables for all variables will be stacked.

Note that this function is deprecated, please use t_summary instead.

See also

Examples

library(random.cdisc.data)

ASL <- radam("ASL")

# control the label
ASL <- var_relabel(ASL, BAGE = "Baseline Age of patient")

# control categorical order
ASL$SEX <- relevel(ASL$SEX, "M", "F")

t_summarize_variables(ASL[, c("SEX", "BAGE")], col_by = ASL$ARM, total = "All Patients")

#> Warning: t_summarize_variables is deprecated, please use t_summary instead.

#>                                    ARM A                ARM B    
#>                                   (N=53)               (N=47)    
#> -----------------------------------------------------------------
#> SEX
#>   n                                 53                   47      
#>   M                             30 (56.6%)           29 (61.7%)  
#>   F                             22 (41.51%)          18 (38.3%)  
#>   U                              1 (1.89%)             0 (0%)    
#> 
#> Baseline Age of patient
#>   n                                 53                   47      
#>   Mean (SD)                    37.51 (13.47)        43.72 (15.32)
#>   Median                            36                   42      
#>   Min - Max                       11 - 64              16 - 81   

t_summarize_variables(ASL[, c("SEX", "BAGE")], col_by = ASL$ARM, total = "All Patients",
                      useNA_factors = 'always')

#> Warning: t_summarize_variables is deprecated, please use t_summary instead.

#>                                    ARM A                ARM B    
#>                                   (N=53)               (N=47)    
#> -----------------------------------------------------------------
#> SEX
#>   n                                 53                   47      
#>   M                             30 (56.6%)           29 (61.7%)  
#>   F                             22 (41.51%)          18 (38.3%)  
#>   U                              1 (1.89%)             0 (0%)    
#> 
#> Baseline Age of patient
#>   n                                 53                   47      
#>   Mean (SD)                    37.51 (13.47)        43.72 (15.32)
#>   Median                            36                   42      
#>   Min - Max                       11 - 64              16 - 81