Summarize an Object for Different Groups

Summarize an Object for Different Groups

t_summary(x, col_by, col_N = NULL, total = NULL, ...)

Arguments

x
an object to dispatch on, see methods("t_summary")

col_by
(logical, factor or data.frame)
Contains reference and comparison group information, the first level indicates the reference group. See the functions col_by_to_matrix and by_add_total.

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

total
(character value)
Denotes column name of an added total column using by_add_total to col_by and col_N_add_total to col_N. If NULL no total column is added.

...
(optional)
additional arguments passed on to methods

See also

Examples

t_summary(iris$Sepal.Length, iris$Species)

#>                    setosa           versicolor          virginica 
#>                    (N=50)             (N=50)             (N=50)   
#> ------------------------------------------------------------------
#> n                    50                 50                 50     
#> Mean (SD)        5.01 (0.35)        5.94 (0.52)        6.59 (0.64)
#> Median                5                 5.9                6.5    
#> Min - Max         4.3 - 5.8           4.9 - 7           4.9 - 7.9 

library(random.cdisc.data)
library(dplyr)
ADSL <- radsl(cached = TRUE)

t_summary(x = ADSL$AGE, col_by = ADSL$ARMCD)

#>                     ARM A               ARM B               ARM C    
#>                    (N=134)             (N=134)             (N=132)   
#> ---------------------------------------------------------------------
#> n                    134                 134                 132     
#> Mean (SD)        33.77 (6.55)        35.43 (7.9)         35.43 (7.72)
#> Median                33                  35                  35     
#> Min - Max          21 - 50             21 - 62             20 - 69   

t_summary(x = ADSL[, c("AGE", "SEX", "RACE")], col_by = ADSL$ARMCD)

#>                                                       ARM A               ARM B               ARM C    
#>                                                      (N=134)             (N=134)             (N=132)   
#> -------------------------------------------------------------------------------------------------------
#> Age
#>   n                                                    134                 134                 132     
#>   Mean (SD)                                        33.77 (6.55)        35.43 (7.9)         35.43 (7.72)
#>   Median                                                33                  35                  35     
#>   Min - Max                                          21 - 50             21 - 62             20 - 69   
#> 
#> Sex
#>   n                                                    134                 134                 132     
#>   F                                                79 (58.96%)         77 (57.46%)           66 (50%)  
#>   M                                                51 (38.06%)         55 (41.04%)         60 (45.45%) 
#>   U                                                 3 (2.24%)           2 (1.49%)           4 (3.03%)  
#>   UNDIFFERENTIATED                                  1 (0.75%)               0               2 (1.52%)  
#> 
#> Race
#>   n                                                    134                 134                 132     
#>   ASIAN                                            68 (50.75%)           67 (50%)           73 (55.3%) 
#>   BLACK OR AFRICAN AMERICAN                        31 (23.13%)          28 (20.9%)         32 (24.24%) 
#>   WHITE                                            27 (20.15%)          26 (19.4%)         21 (15.91%) 
#>   AMERICAN INDIAN OR ALASKA NATIVE                  8 (5.97%)           11 (8.21%)          6 (4.55%)  
#>   MULTIPLE                                              0               1 (0.75%)               0      
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER             0               1 (0.75%)               0      
#>   OTHER                                                 0                   0                   0      
#>   UNKNOWN                                               0                   0                   0      

t_summary(
  x = ADSL[, c("AGE", "SEX", "RACE")],
  col_by = ADSL$ARMCD,
  total = "All Patients"
)

#>                                                       ARM A               ARM B               ARM C            All Patients
#>                                                      (N=134)             (N=134)             (N=132)             (N=400)   
#> ---------------------------------------------------------------------------------------------------------------------------
#> Age
#>   n                                                    134                 134                 132                 400     
#>   Mean (SD)                                        33.77 (6.55)        35.43 (7.9)         35.43 (7.72)        34.88 (7.44)
#>   Median                                                33                  35                  35                  34     
#>   Min - Max                                          21 - 50             21 - 62             20 - 69             20 - 69   
#> 
#> Sex
#>   n                                                    134                 134                 132                 400     
#>   F                                                79 (58.96%)         77 (57.46%)           66 (50%)          222 (55.5%) 
#>   M                                                51 (38.06%)         55 (41.04%)         60 (45.45%)         166 (41.5%) 
#>   U                                                 3 (2.24%)           2 (1.49%)           4 (3.03%)           9 (2.25%)  
#>   UNDIFFERENTIATED                                  1 (0.75%)               0               2 (1.52%)           3 (0.75%)  
#> 
#> Race
#>   n                                                    134                 134                 132                 400     
#>   ASIAN                                            68 (50.75%)           67 (50%)           73 (55.3%)          208 (52%)  
#>   BLACK OR AFRICAN AMERICAN                        31 (23.13%)          28 (20.9%)         32 (24.24%)         91 (22.75%) 
#>   WHITE                                            27 (20.15%)          26 (19.4%)         21 (15.91%)          74 (18.5%) 
#>   AMERICAN INDIAN OR ALASKA NATIVE                  8 (5.97%)           11 (8.21%)          6 (4.55%)           25 (6.25%) 
#>   MULTIPLE                                              0               1 (0.75%)               0               1 (0.25%)  
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER             0               1 (0.75%)               0               1 (0.25%)  
#>   OTHER                                                 0                   0                   0                   0      
#>   UNKNOWN                                               0                   0                   0                   0      

with(
  ADSL,
  t_summary(AGE > 65, ARMCD)
)

#>                 ARM A               ARM B               ARM C    
#>                (N=134)             (N=134)             (N=132)   
#> -----------------------------------------------------------------
#> n                134                 134                 132     
#> TRUE              0                   0               1 (0.76%)  
#> FALSE         134 (100%)          134 (100%)         131 (99.24%)

# examples with hierarchical header
ADSL <- radsl(cached = TRUE)
t_summary(
  x = ADSL[, c("SEX", "AGE")],
  col_by = ADSL$ARM
)

#>                             A: Drug X             B: Placebo          C: Combination
#>                              (N=134)               (N=134)               (N=132)    
#> ------------------------------------------------------------------------------------
#> Sex
#>   n                            134                   134                   132      
#>   F                        79 (58.96%)           77 (57.46%)             66 (50%)   
#>   M                        51 (38.06%)           55 (41.04%)           60 (45.45%)  
#>   U                         3 (2.24%)             2 (1.49%)             4 (3.03%)   
#>   UNDIFFERENTIATED          1 (0.75%)                 0                 2 (1.52%)   
#> 
#> Age
#>   n                            134                   134                   132      
#>   Mean (SD)                33.77 (6.55)          35.43 (7.9)           35.43 (7.72) 
#>   Median                        33                    35                    35      
#>   Min - Max                  21 - 50               21 - 62               20 - 69    

ADSL_f <- ADSL %>%
  dplyr::filter(SEX %in% c("M", "F")) %>%
  mutate(SEX = droplevels(SEX))
t_summary(
  x = ADSL_f$SEX,
  col_by = by_hierarchical(ADSL_f$ARM, ADSL_f$SEX, sep = "-")
)

#>                  A: Drug X                         B: Placebo                      C: Combination      
#>              F                M                F                M                F                M    
#>           (N=79)           (N=51)           (N=77)           (N=55)           (N=66)           (N=60)  
#> -------------------------------------------------------------------------------------------------------
#> n           79               51               77               55               66               60    
#> F        79 (100%)            0            77 (100%)            0            66 (100%)            0    
#> M            0            51 (100%)            0            55 (100%)            0            60 (100%)

t_summary(
  x = ADSL_f[, c("SEX", "AGE")],
  col_by = by_hierarchical(ADSL_f$ARM, ADSL_f$SEX, sep = "-")
)

#>                               A: Drug X                               B: Placebo                            C: Combination         
#>                         F                   M                   F                   M                   F                   M      
#>                       (N=79)              (N=51)              (N=77)              (N=55)              (N=66)              (N=60)   
#> -----------------------------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                     79                  51                  77                  55                  66                  60     
#>   F                 79 (100%)               0               77 (100%)               0               66 (100%)               0      
#>   M                     0               51 (100%)               0               55 (100%)               0               60 (100%)  
#> 
#> Age
#>   n                     79                  51                  77                  55                  66                  60     
#>   Mean (SD)        32.76 (6.09)        35.57 (7.08)        34.12 (7.06)        37.44 (8.69)        35.2 (7.43)         35.38 (8.24)
#>   Median                32                  37                  32                  37                  35                 33.5    
#>   Min - Max          21 - 47             23 - 50             23 - 58             21 - 62             21 - 64             20 - 69   

ADSL_f <- ADSL %>%
  dplyr::filter(SEX %in% c("M", "F"), RACE %in% c("ASIAN", "WHITE")) %>%
  mutate(
    SEX = droplevels(SEX),
    RACE = droplevels(as.factor(RACE))
  )
t_summary(
  x = ADSL_f[, c("SEX", "RACE")],
  col_by = by_hierarchical(ADSL_f$SEX, ADSL_f$RACE)
)

#>                             F                                   M              
#>                  ASIAN             WHITE             ASIAN             WHITE   
#>                 (N=116)            (N=45)            (N=87)            (N=27)  
#> -------------------------------------------------------------------------------
#> SEX
#>   n               116                45                87                27    
#>   F            116 (100%)        45 (100%)             0                 0     
#>   M                0                 0             87 (100%)         27 (100%) 
#> 
#> RACE
#>   n               116                45                87                27    
#>   ASIAN        116 (100%)            0             87 (100%)             0     
#>   WHITE            0             45 (100%)             0             27 (100%) 

t_summary(
  x = ADSL_f[, c("SEX", "RACE")],
  col_by = by_hierarchical(ADSL_f$RACE, ADSL_f$SEX)
)

#>                           ASIAN                               WHITE            
#>                    F                 M                 F                 M     
#>                 (N=116)            (N=87)            (N=45)            (N=27)  
#> -------------------------------------------------------------------------------
#> SEX
#>   n               116                87                45                27    
#>   F            116 (100%)            0             45 (100%)             0     
#>   M                0             87 (100%)             0             27 (100%) 
#> 
#> RACE
#>   n               116                87                45                27    
#>   ASIAN        116 (100%)        87 (100%)             0                 0     
#>   WHITE            0                 0             45 (100%)         27 (100%) 

t_summary(
  x = ADSL_f[, c("ITTFL", "BMRKR1", "BEP01FL")],
  col_by = by_hierarchical(ADSL_f$RACE, ADSL_f$SEX)
)

#>                                                 ASIAN                                   WHITE              
#>                                         F                   M                   F                   M      
#>                                      (N=116)              (N=87)              (N=45)              (N=27)   
#> -----------------------------------------------------------------------------------------------------------
#> ITTFL
#>   n                                    116                  87                  45                  27     
#>   Y                                 116 (100%)          87 (100%)           45 (100%)           27 (100%)  
#> 
#> Continous Level Biomarker 1
#>   n                                    116                  87                  45                  27     
#>   Mean (SD)                        5.38 (3.28)          5.6 (3.17)         6.41 (4.52)         5.24 (3.11) 
#>   Median                               4.75                4.85                5.49                4.61    
#>   Min - Max                        0.46 - 15.11        0.41 - 15.07        0.99 - 21.39        1.07 - 12.87
#> 
#> BEP01FL
#>   n                                    116                  87                  45                  27     
#>   Y                                65 (56.03%)         43 (49.43%)         23 (51.11%)         11 (40.74%) 
#>   N                                51 (43.97%)         44 (50.57%)         22 (48.89%)         16 (59.26%) 

# add total
t_summary(
  x = ADSL_f[, c("ITTFL", "BMRKR1", "BEP01FL")],
  col_by = by_hierarchical(ADSL_f$RACE, ADSL_f$SEX),
  total = "All"
)

#>                                                 ASIAN                                   WHITE                          All     
#>                                         F                   M                   F                   M                          
#>                                      (N=116)              (N=87)              (N=45)              (N=27)             (N=275)   
#> -------------------------------------------------------------------------------------------------------------------------------
#> ITTFL
#>   n                                    116                  87                  45                  27                 275     
#>   Y                                 116 (100%)          87 (100%)           45 (100%)           27 (100%)           275 (100%) 
#> 
#> Continous Level Biomarker 1
#>   n                                    116                  87                  45                  27                 275     
#>   Mean (SD)                        5.38 (3.28)          5.6 (3.17)         6.41 (4.52)         5.24 (3.11)          5.6 (3.46) 
#>   Median                               4.75                4.85                5.49                4.61                4.82    
#>   Min - Max                        0.46 - 15.11        0.41 - 15.07        0.99 - 21.39        1.07 - 12.87        0.41 - 21.39
#> 
#> BEP01FL
#>   n                                    116                  87                  45                  27                 275     
#>   Y                                65 (56.03%)         43 (49.43%)         23 (51.11%)         11 (40.74%)         142 (51.64%)
#>   N                                51 (43.97%)         44 (50.57%)         22 (48.89%)         16 (59.26%)         133 (48.36%)