Summarize an Object for Different Groups

Summarize an Object for Different Groups

t_summary(x, col_by, col_N = table(col_by), ...)

Arguments

x
an object to dispatch on

col_by
a factor of length nrow(x) that defines which levels in col_by define a column. If data should not be split into columns use the no_by function.

col_N
The column total for each group that is displayed in the table header with (N=xx).

...
arguments passed on to methods

See also

Examples

t_summary(iris$Sepal.Length, iris$Species)

#>                    setosa           versicolor          virginica 
#>                    (N=50)             (N=50)             (N=50)   
#> ------------------------------------------------------------------
#> n                    50                 50                 50     
#> Mean (SD)        5.01 (0.35)        5.94 (0.52)        6.59 (0.64)
#> Median                5                 5.9                6.5    
#> Min - Max         4.3 - 5.8           4.9 - 7           4.9 - 7.9 

library(random.cdisc.data)
ADSL <- radsl(N = 100, seed = 1)

t_summary(ADSL$AGE, ADSL$ARMCD)

#>                      ARM A                ARM B                ARM C    
#>                     (N=23)               (N=42)               (N=35)    
#> ------------------------------------------------------------------------
#> n                     23                   42                   35      
#> Mean (SD)        38.52 (17.76)        37.93 (15.14)        43.54 (19.63)
#> Median                34                   36                   39      
#> Min - Max           20 - 86              20 - 81              20 - 84   

t_summary(ADSL[, c("AGE", "SEX", "RACE")], ADSL$ARMCD)

#>                                                        ARM A                ARM B                ARM C    
#>                                                       (N=23)               (N=42)               (N=35)    
#> ----------------------------------------------------------------------------------------------------------
#> Age
#>   n                                                     23                   42                   35      
#>   Mean (SD)                                        38.52 (17.76)        37.93 (15.14)        43.54 (19.63)
#>   Median                                                34                   36                   39      
#>   Min - Max                                           20 - 86              20 - 81              20 - 84   
#> 
#> Sex
#>   n                                                     23                   42                   35      
#>   F                                                 13 (56.52%)           21 (50%)            19 (54.29%) 
#>   M                                                 9 (39.13%)            21 (50%)            15 (42.86%) 
#>   U                                                  1 (4.35%)                -                1 (2.86%)  
#>   UNDIFFERENTIATED                                       -                    -                    -      
#> 
#> Race
#>   n                                                     23                   42                   35      
#>   ASIAN                                             15 (65.22%)          25 (59.52%)          15 (42.86%) 
#>   BLACK OR AFRICAN AMERICAN                         3 (13.04%)           8 (19.05%)           10 (28.57%) 
#>   WHITE                                             5 (21.74%)           7 (16.67%)             7 (20%)   
#>   AMERICAN INDIAN OR ALASKA NATIVE                       -                2 (4.76%)            3 (8.57%)  
#>   MULTIPLE                                               -                    -                    -      
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER              -                    -                    -      
#>   OTHER                                                  -                    -                    -      
#>   UNKNOWN                                                -                    -                    -      

with(ADSL, t_summary(AGE > 65, ARMCD))

#>                 ARM A              ARM B              ARM C   
#>                (N=23)             (N=42)             (N=35)   
#> --------------------------------------------------------------
#> n                23                 42                 35     
#> TRUE         3 (13.04%)          3 (7.14%)         5 (14.29%) 
#> FALSE        20 (86.96%)        39 (92.86%)        30 (85.71%)