Variables Summary Table

Similar as the demographic table in STREAM

# S3 method for data.frame
t_summary(x, col_by, col_N = table(col_by),
  total = NULL, ...)

Arguments

x
data frame with variables to be summarized as described in the details section. If the variable has a label attribute then it will be used for the row name.

col_by
factor with reference and comparison group information, the first level indicates the reference group.

col_N
a vecor with the reference population per level of col_by. This often by default

total
character string that will be used as a label for a column with total population. If the levels of col_by are the only columns of interest then total should be NULL.

...
arguments passed on to methods

Details

Every variable in x will be mapped to a summary table using t_summary and then be stacked.

See also

Examples

# with iris data
t_summary(iris$Sepal.Length, iris$Species)

#>                    setosa           versicolor          virginica 
#>                    (N=50)             (N=50)             (N=50)   
#> ------------------------------------------------------------------
#> n                    50                 50                 50     
#> Mean (SD)        5.01 (0.35)        5.94 (0.52)        6.59 (0.64)
#> Median                5                 5.9                6.5    
#> Min - Max         4.3 - 5.8           4.9 - 7           4.9 - 7.9 

library(random.cdisc.data)
library(dplyr)
ADSL <- cadsl

t_summary(ADSL[, c("SEX", "AGE")], col_by  = no_by("All"), col_N = nrow(ADSL))

#>                               All    
#>                             (N=400)  
#> -------------------------------------
#> Sex
#>   n                           400    
#>   F                        204 (51%) 
#>   M                        184 (46%) 
#>   U                        9 (2.25%) 
#>   UNDIFFERENTIATED         3 (0.75%) 
#> 
#> Age
#>   n                           400    
#>   Mean (SD)               40.4 (18.2)
#>   Median                      37     
#>   Min - Max                20 - 116  

# control the label
ADSL <- var_relabel(ADSL, AGE = "Baseline Age of patient")

# control categorical order
ADSL$SEX <- relevel(ADSL$SEX, "M", "F")

t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients")

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                   (N=142)               (N=133)               (N=125)               (N=400)    
#> ---------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                                 142                   133                   125                   400      
#>   M                             68 (47.89%)           63 (47.37%)            53 (42.4%)            184 (46%)   
#>   F                             69 (48.59%)           67 (50.38%)            68 (54.4%)            204 (51%)   
#>   U                              2 (1.41%)             3 (2.26%)              4 (3.2%)             9 (2.25%)   
#>   UNDIFFERENTIATED               3 (2.11%)                 -                     -                 3 (0.75%)   
#> 
#> Baseline Age of patient
#>   n                                 142                   133                   125                   400      
#>   Mean (SD)                    40.34 (17.42)          39.82 (17.7)          41.1 (19.67)          40.4 (18.2)  
#>   Median                             39                    37                    37                    37      
#>   Min - Max                       20 - 93               20 - 116              20 - 88               20 - 116   

t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients", useNA = 'always')

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                   (N=142)               (N=133)               (N=125)               (N=400)    
#> ---------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                                 142                   133                   125                   400      
#>   M                             68 (47.89%)           63 (47.37%)            53 (42.4%)            184 (46%)   
#>   F                             69 (48.59%)           67 (50.38%)            68 (54.4%)            204 (51%)   
#>   U                              2 (1.41%)             3 (2.26%)              4 (3.2%)             9 (2.25%)   
#>   UNDIFFERENTIATED               3 (2.11%)                 -                     -                 3 (0.75%)   
#>   <NA>                               0                     0                     0                     0       
#> 
#> Baseline Age of patient
#>   n                                 142                   133                   125                   400      
#>   Mean (SD)                    40.34 (17.42)          39.82 (17.7)          41.1 (19.67)          40.4 (18.2)  
#>   Median                             39                    37                    37                    37      
#>   Min - Max                       20 - 93               20 - 116              20 - 88               20 - 116   

ADSL$SEX[1:10] <- NA
t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients", useNA = 'ifany')

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                   (N=142)               (N=133)               (N=125)               (N=400)    
#> ---------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                                 139                   129                   122                   390      
#>   M                              67 (48.2%)           62 (48.06%)           53 (43.44%)           182 (46.67%) 
#>   F                              67 (48.2%)           64 (49.61%)           65 (53.28%)           196 (50.26%) 
#>   U                              2 (1.44%)             3 (2.33%)             4 (3.28%)             9 (2.31%)   
#>   UNDIFFERENTIATED               3 (2.16%)                 -                     -                 3 (0.77%)   
#>   <NA>                               3                     4                     3                     10      
#> 
#> Baseline Age of patient
#>   n                                 142                   133                   125                   400      
#>   Mean (SD)                    40.34 (17.42)          39.82 (17.7)          41.1 (19.67)          40.4 (18.2)  
#>   Median                             39                    37                    37                    37      
#>   Min - Max                       20 - 93               20 - 116              20 - 88               20 - 116   

t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients", denominator = "N",
          useNA = 'ifany')

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                   (N=142)               (N=133)               (N=125)               (N=400)    
#> ---------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                                 139                   129                   122                   390      
#>   M                             67 (47.18%)           62 (46.62%)            53 (42.4%)           182 (45.5%)  
#>   F                             67 (47.18%)           64 (48.12%)             65 (52%)             196 (49%)   
#>   U                              2 (1.41%)             3 (2.26%)              4 (3.2%)             9 (2.25%)   
#>   UNDIFFERENTIATED               3 (2.11%)                 -                     -                 3 (0.75%)   
#>   <NA>                               3                     4                     3                     10      
#> 
#> Baseline Age of patient
#>   n                                 142                   133                   125                   400      
#>   Mean (SD)                    40.34 (17.42)          39.82 (17.7)          41.1 (19.67)          40.4 (18.2)  
#>   Median                             39                    37                    37                    37      
#>   Min - Max                       20 - 93               20 - 116              20 - 88               20 - 116   

ADSL_AGE65 <- ADSL %>%
  dplyr::filter(AGE > 65)
t_summary(ADSL_AGE65[, c("AGE", "SEX")], ADSL_AGE65$ARM, total = "All Patients",
          col_N = table(ADSL$ARM), denominator = "N", drop_levels = TRUE)

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                   (N=142)               (N=133)               (N=125)               (N=400)    
#> ---------------------------------------------------------------------------------------------------------------
#> Baseline Age of patient
#>   n                                  12                    11                    19                    42      
#>   Mean (SD)                     75.5 (7.65)          78.27 (14.28)          73.68 (5.46)          75.4 (9.06)  
#>   Median                            73.5                   76                    73                   73.5     
#>   Min - Max                       68 - 93               66 - 116              67 - 88               66 - 116   
#> 
#> SEX
#>   n                                  12                    10                    18                    40      
#>   M                              8 (5.63%)             4 (3.01%)              7 (5.6%)             19 (4.75%)  
#>   F                              3 (2.11%)             6 (4.51%)             11 (8.8%)              20 (5%)    
#>   UNDIFFERENTIATED                1 (0.7%)                 -                     -                 1 (0.25%)