Variables Summary Table

Similiar as the demographic table in STREAM

# S3 method for data.frame
t_summary(x, col_by, col_N = table(col_by),
  total = NULL, ...)

Arguments

x
data frame with variables to be summarized as described in the details section. If the variable has a label attribute then it will be used for the row name.

col_by
a factor of length nrow(x) that defines which levels in col_by define a column. If data should not be split into columns use the no_by function.

col_N
The column total for each group that is displayed in the table header with (N=xx).

total
if not NULL then it must be a string and an addition column will be added with the overall summaries

...
arguments passed on to methods

Details

Every variable in x will be mapped to a summary table using t_summary and then be stacked.

See also

Examples

# with iris data
t_summary(iris$Sepal.Length, iris$Species)

#>                    setosa           versicolor          virginica 
#>                    (N=50)             (N=50)             (N=50)   
#> ------------------------------------------------------------------
#> n                    50                 50                 50     
#> Mean (SD)        5.01 (0.35)        5.94 (0.52)        6.59 (0.64)
#> Median                5                 5.9                6.5    
#> Min - Max         4.3 - 5.8           4.9 - 7           4.9 - 7.9 

library(random.cdisc.data)
library(dplyr)
ADSL <- radsl(N = 100, seed = 1)

t_summary(ADSL[, c("SEX", "AGE")], col_by  = no_by("All"), col_N = nrow(ADSL))

#>                                All     
#>                              (N=100)   
#> ---------------------------------------
#> Sex
#>   n                            100     
#>   F                         53 (53%)   
#>   M                         45 (45%)   
#>   U                          2 (2%)    
#>   UNDIFFERENTIATED              -      
#> 
#> Age
#>   n                            100     
#>   Mean (SD)               40.03 (17.44)
#>   Median                       36      
#>   Min - Max                  20 - 86   

# control the label
ADSL <- var_relabel(ADSL, AGE = "Baseline Age of patient")

# control categorical order
ADSL$SEX <- relevel(ADSL$SEX, "M", "F")

t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients")

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                    (N=23)                (N=42)                (N=35)               (N=100)    
#> ---------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                                  23                    42                    35                   100      
#>   M                              9 (39.13%)             21 (50%)            15 (42.86%)             45 (45%)   
#>   F                             13 (56.52%)             21 (50%)            19 (54.29%)             53 (53%)   
#>   U                              1 (4.35%)                 -                 1 (2.86%)               2 (2%)    
#>   UNDIFFERENTIATED                   -                     -                     -                     -       
#> 
#> Baseline Age of patient
#>   n                                  23                    42                    35                   100      
#>   Mean (SD)                    38.52 (17.76)         37.93 (15.14)         43.54 (19.63)         40.03 (17.44) 
#>   Median                             34                    36                    39                    36      
#>   Min - Max                       20 - 86               20 - 81               20 - 84               20 - 86    

t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients",
         useNA = 'always')

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                    (N=23)                (N=42)                (N=35)               (N=100)    
#> ---------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                                  23                    42                    35                   100      
#>   M                              9 (39.13%)             21 (50%)            15 (42.86%)             45 (45%)   
#>   F                             13 (56.52%)             21 (50%)            19 (54.29%)             53 (53%)   
#>   U                              1 (4.35%)                 -                 1 (2.86%)               2 (2%)    
#>   UNDIFFERENTIATED                   -                     -                     -                     -       
#>   <NA>                               0                     0                     0                     0       
#> 
#> Baseline Age of patient
#>   n                                  23                    42                    35                   100      
#>   Mean (SD)                    38.52 (17.76)         37.93 (15.14)         43.54 (19.63)         40.03 (17.44) 
#>   Median                             34                    36                    39                    36      
#>   Min - Max                       20 - 86               20 - 81               20 - 84               20 - 86    

ADSL$SEX[1:10] <- NA
t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients",
         useNA = 'ifany')

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                    (N=23)                (N=42)                (N=35)               (N=100)    
#> ---------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                                  23                    36                    31                    90      
#>   M                              9 (39.13%)           16 (44.44%)           13 (41.94%)           38 (42.22%)  
#>   F                             13 (56.52%)           20 (55.56%)           17 (54.84%)           50 (55.56%)  
#>   U                              1 (4.35%)                 -                 1 (3.23%)             2 (2.22%)   
#>   UNDIFFERENTIATED                   -                     -                     -                     -       
#>   <NA>                               0                     6                     4                     10      
#> 
#> Baseline Age of patient
#>   n                                  23                    42                    35                   100      
#>   Mean (SD)                    38.52 (17.76)         37.93 (15.14)         43.54 (19.63)         40.03 (17.44) 
#>   Median                             34                    36                    39                    36      
#>   Min - Max                       20 - 86               20 - 81               20 - 84               20 - 86    

t_summary(ADSL[, c("SEX", "AGE")], col_by = ADSL$ARM, total = "All Patients",
         denominator = "N", useNA = 'ifany')

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                    (N=23)                (N=42)                (N=35)               (N=100)    
#> ---------------------------------------------------------------------------------------------------------------
#> SEX
#>   n                                  23                    36                    31                    90      
#>   M                              9 (39.13%)            16 (38.1%)           13 (37.14%)             38 (38%)   
#>   F                             13 (56.52%)           20 (47.62%)           17 (48.57%)             50 (50%)   
#>   U                              1 (4.35%)                 -                 1 (2.86%)               2 (2%)    
#>   UNDIFFERENTIATED                   -                     -                     -                     -       
#>   <NA>                               0                     6                     4                     10      
#> 
#> Baseline Age of patient
#>   n                                  23                    42                    35                   100      
#>   Mean (SD)                    38.52 (17.76)         37.93 (15.14)         43.54 (19.63)         40.03 (17.44) 
#>   Median                             34                    36                    39                    36      
#>   Min - Max                       20 - 86               20 - 81               20 - 84               20 - 86    

ADSL_AGE65 <- ADSL %>% filter(AGE > 65)
t_summary(ADSL_AGE65[, c("AGE", "SEX")], ADSL_AGE65$ARM, total = "All Patients",
         col_N = table(ADSL$ARM), denominator = "N", drop_levels = TRUE)

#>                                  A: Drug X             B: Placebo          C: Combination         All Patients 
#>                                    (N=23)                (N=42)                (N=35)               (N=100)    
#> ---------------------------------------------------------------------------------------------------------------
#> Baseline Age of patient
#>   n                                  3                     3                     5                     11      
#>   Mean (SD)                    74.33 (10.12)           74 (6.56)            76.6 (6.88)           75.27 (7.04) 
#>   Median                             69                    73                    77                    75      
#>   Min - Max                       68 - 86               68 - 81               66 - 84               66 - 86    
#> 
#> SEX
#>   n                                  3                     2                     4                     9       
#>   M                              1 (4.35%)                 -                 2 (5.71%)               3 (3%)    
#>   F                               2 (8.7%)             2 (4.76%)             2 (5.71%)               6 (6%)