Summarize Categorical Data

Tabulate the number on non-missing observations, number of per level and percentage

# S3 method for factor
t_summary(x, col_by, useNA = c("no", "ifany", "always"),
  denominator = c("n", "N"), drop_levels = FALSE, ...)

Arguments

x
numeric variable

col_by
a factor of length nrow(x) that defines which levels in col_by define a column. If data should not be split into columns use the no_by function.

useNA
choose whether missing data (NAs) should be displayed as a level

denominator
either n or N for calculating the level associated percentage.

drop_levels
boolean whether to drio zero count levels

...
not used arguments

Value

an rtable object

Examples

t_summary(iris$Species, iris$Species)

#>                     setosa          versicolor        virginica 
#>                     (N=50)            (N=50)            (N=50)  
#> ----------------------------------------------------------------
#> n                     50                50                50    
#> setosa            50 (100%)           0 (0%)            0 (0%)  
#> versicolor          0 (0%)          50 (100%)           0 (0%)  
#> virginica           0 (0%)            0 (0%)          50 (100%) 

x <-  factor(c("a", "b", "a", "a",  NA,  NA,  NA, "b", "b"))
cb <- factor(c("E", "E", "F", "F", "E", "E", "F", "F", "F"))

t_summary(x, cb)

#>             E              F   
#>           (N=4)          (N=5) 
#> -------------------------------
#> n           2              4   
#> a        1 (50%)        2 (50%)
#> b        1 (50%)        2 (50%)

t_summary(x, cb, useNA = "always")

#>                E              F   
#>              (N=4)          (N=5) 
#> ----------------------------------
#> n              2              4   
#> a           1 (50%)        2 (50%)
#> b           1 (50%)        2 (50%)
#> <NA>           2              1   

t_summary(x, cb, denominator = "N")

#>             E              F   
#>           (N=4)          (N=5) 
#> -------------------------------
#> n           2              4   
#> a        1 (25%)        2 (40%)
#> b        1 (25%)        2 (40%)

t_summary(x, cb, useNA = "always",  denominator = "N")

#>                E              F   
#>              (N=4)          (N=5) 
#> ----------------------------------
#> n              2              4   
#> a           1 (25%)        2 (40%)
#> b           1 (25%)        2 (40%)
#> <NA>           2              1   

t_summary(x, cb, useNA = "always",  denominator = "n")

#>                E              F   
#>              (N=4)          (N=5) 
#> ----------------------------------
#> n              2              4   
#> a           1 (50%)        2 (50%)
#> b           1 (50%)        2 (50%)
#> <NA>           2              1