Summarize an Object for Different Groups with by Variable

Summarize an Object for Different Groups with by Variable

t_summary_by(x, by, col_by, col_N, total = NULL, ...)

Arguments

x
vector

by
a factor of length nrow(x) with no missing values. The levels of by define the summary sub-groups in the table.

col_by
a factor of length nrow(x) with no missing values. The levels of col_by define the columns in the table.

col_N
a table object with the reference population used for the header of the table. See examples below.

total
character string that will be used as a label for a column with pooled total population. If the levels of col_by are the only columns of interest then total should be NULL.

...
arguments passed on to methods

Details

For every level of the variable by a summary table using t_summary will be created. The individual tables are then stacked together.

See also

Examples

library(dplyr)
library(random.cdisc.data)

ADSL <- radsl(N = 30, seed = 1)

t_summary_by(
 x = ADSL$SEX,
 by = ADSL$COUNTRY,
 col_by = ADSL$ARMCD,
 col_N = table(ADSL$ARMCD),
 total = "All Patients",
 drop_levels = TRUE
)

#> Country           ARM A               ARM B               ARM C            All Patients
#>   Sex             (N=7)               (N=14)              (N=9)               (N=30)   
#> ---------------------------------------------------------------------------------------
#> CHN
#>   n                 2                   7                   7                   16     
#>   F                 -               3 (42.86%)          2 (28.57%)          5 (31.25%) 
#>   M              2 (100%)           3 (42.86%)          4 (57.14%)          9 (56.25%) 
#>   U                 -               1 (14.29%)          1 (14.29%)          2 (12.5%)  
#> 
#> USA
#>   n                 1                   0                   0                   1      
#>   F              1 (100%)               -                   -                1 (100%)  
#> 
#> BRA
#>   n                 2                   1                   1                   4      
#>   F              2 (100%)            1 (100%)            1 (100%)            4 (100%)  
#> 
#> PAK
#>   n                 1                   1                   1                   3      
#>   M              1 (100%)            1 (100%)            1 (100%)            3 (100%)  
#> 
#> NGA
#>   n                 0                   0                   0                   0      
#> 
#> RUS
#>   n                 1                   2                   0                   3      
#>   F              1 (100%)            2 (100%)               -                3 (100%)  
#> 
#> JPN
#>   n                 0                   2                   0                   2      
#>   M                 -                2 (100%)               -                2 (100%)  
#> 
#> GBR
#>   n                 0                   0                   0                   0      
#> 
#> CAN
#>   n                 0                   1                   0                   1      
#>   M                 -                1 (100%)               -                1 (100%)  
#> 
#> CHE
#>   n                 0                   0                   0                   0      

ADSL$SEX[1:5] <- NA

t_summary_by(
 x = ADSL$SEX,
 by = ADSL$COUNTRY,
 col_by = ADSL$ARMCD,
 col_N = table(ADSL$ARMCD),
 total = "All Patients",
 drop_levels = TRUE,
 useNA = "ifany"
)

#> Country           ARM A               ARM B               ARM C            All Patients
#>   Sex             (N=7)               (N=14)              (N=9)               (N=30)   
#> ---------------------------------------------------------------------------------------
#> CHN
#>   n                 2                   4                   7                   13     
#>   F                 -                2 (50%)            2 (28.57%)          4 (30.77%) 
#>   M              2 (100%)            2 (50%)            4 (57.14%)          8 (61.54%) 
#>   U                 -                   -               1 (14.29%)          1 (7.69%)  
#>   <NA>              0                   3                   0                   3      
#> 
#> USA
#>   n                 0                   0                   0                   0      
#>   <NA>              1                   NA                  NA                  1      
#> 
#> BRA
#>   n                 2                   1                   1                   4      
#>   F              2 (100%)            1 (100%)            1 (100%)            4 (100%)  
#> 
#> PAK
#>   n                 1                   1                   1                   3      
#>   M              1 (100%)            1 (100%)            1 (100%)            3 (100%)  
#> 
#> NGA
#>   n                 0                   0                   0                   0      
#> 
#> RUS
#>   n                 1                   1                   0                   2      
#>   F              1 (100%)            1 (100%)               -                2 (100%)  
#>   <NA>              0                   1                   NA                  1      
#> 
#> JPN
#>   n                 0                   2                   0                   2      
#>   M                 -                2 (100%)               -                2 (100%)  
#> 
#> GBR
#>   n                 0                   0                   0                   0      
#> 
#> CAN
#>   n                 0                   1                   0                   1      
#>   M                 -                1 (100%)               -                1 (100%)  
#> 
#> CHE
#>   n                 0                   0                   0                   0      

ADSL <- ADSL %>% select(STUDYID, USUBJID, ARMCD)

ADQS <- radqs(ADSL, seed = 2)
ADQS_f <- ADQS %>% filter(PARAMCD=="BFIALL")

t_summary_by(
 x = ADQS_f$AVAL,
 by = ADQS_f$AVISIT,
 col_by = ADQS_f$ARMCD,
 col_N = table(ADSL$ARMCD),
 total = "All Patients"
)

#> ADQS_f$AVISIT               ARM A                ARM B                ARM C            All Patients 
#>   Analysis Value            (N=7)               (N=14)                (N=9)               (N=30)    
#> ----------------------------------------------------------------------------------------------------
#> SCREENING
#>   n                           7                   14                    9                   30      
#>   Mean (SD)             55.28 (6.89)         51.07 (9.18)         50.33 (11.54)        51.83 (9.39) 
#>   Median                    53.34                50.88                52.32                51.9     
#>   Min - Max             48.08 - 65.88        31.51 - 66.73        30.39 - 64.26        30.39 - 66.73
#> 
#> BASELINE
#>   n                           7                   14                    9                   30      
#>   Mean (SD)             53.72 (8.01)         50.46 (7.99)         45.43 (6.63)         49.71 (7.99) 
#>   Median                    53.71                50.62                44.78                51.3     
#>   Min - Max             39.26 - 66.62        33.21 - 60.22        37.01 - 56.72        33.21 - 66.62
#> 
#> WEEK 1 DAY 8
#>   n                           7                   14                    9                   30      
#>   Mean (SD)             55.59 (5.26)         50.53 (6.46)          51.08 (9.1)         51.88 (7.18) 
#>   Median                    57.07                51.25                50.71                52.64    
#>   Min - Max             44.55 - 59.93        34.92 - 62.02        34.21 - 65.27        34.21 - 65.27
#> 
#> WEEK 2 DAY 15
#>   n                           7                   14                    9                   30      
#>   Mean (SD)             53.24 (6.52)          47.74 (8.7)         51.91 (7.36)         50.27 (7.98) 
#>   Median                    51.63                47.47                50.93                49.57    
#>   Min - Max             47.85 - 66.62        30.58 - 62.65        43.61 - 61.59        30.58 - 66.62
#> 
#> WEEK 3 DAY 22
#>   n                           7                   14                    9                   30      
#>   Mean (SD)             52.45 (8.09)         49.91 (6.64)         47.19 (7.58)         49.69 (7.28) 
#>   Median                    56.61                49.91                50.95                50.79    
#>   Min - Max             42.76 - 62.69        39.83 - 60.45        35.34 - 56.09        35.34 - 62.69
#> 
#> WEEK 4 DAY 29
#>   n                           7                   14                    9                   30      
#>   Mean (SD)             47.94 (4.22)         47.11 (6.81)         50.88 (8.85)         48.43 (6.99) 
#>   Median                    50.02                46.77                49.87                48.73    
#>   Min - Max             40.44 - 51.64        34.84 - 59.08        37.92 - 70.83        34.84 - 70.83
#> 
#> WEEK 5 DAY 36
#>   n                           7                   14                    9                   30      
#>   Mean (SD)             50.85 (8.03)         54.02 (7.34)          53.23 (5.5)         53.04 (6.89) 
#>   Median                    50.02                52.38                54.62                51.97    
#>   Min - Max             42.44 - 67.4         44.26 - 74.07        40.93 - 58.53        40.93 - 74.07

ADQS_f$AVALCAT1 <- factor(ifelse(ADQS_f$AVAL >= 0, "Positive", "Negative"),
  levels = c("Positive", "Negative"))
ADQS_f <- var_relabel(ADQS_f, AVALCAT1 = "Result" )

t_summary_by(
 x = ADQS_f$AVALCAT1,
 by = ADQS_f$AVISIT,
 col_by = ADQS_f$ARMCD,
 col_N = table(ADSL$ARMCD),
 total = "All Patients"
)

#> ADQS_f$AVISIT           ARM A               ARM B               ARM C            All Patients
#>   Result                (N=7)               (N=14)              (N=9)               (N=30)   
#> ---------------------------------------------------------------------------------------------
#> SCREENING
#>   n                       7                   14                  9                   30     
#>   Positive             7 (100%)           14 (100%)            9 (100%)           30 (100%)  
#>   Negative                -                   -                   -                   -      
#> 
#> BASELINE
#>   n                       7                   14                  9                   30     
#>   Positive             7 (100%)           14 (100%)            9 (100%)           30 (100%)  
#>   Negative                -                   -                   -                   -      
#> 
#> WEEK 1 DAY 8
#>   n                       7                   14                  9                   30     
#>   Positive             7 (100%)           14 (100%)            9 (100%)           30 (100%)  
#>   Negative                -                   -                   -                   -      
#> 
#> WEEK 2 DAY 15
#>   n                       7                   14                  9                   30     
#>   Positive             7 (100%)           14 (100%)            9 (100%)           30 (100%)  
#>   Negative                -                   -                   -                   -      
#> 
#> WEEK 3 DAY 22
#>   n                       7                   14                  9                   30     
#>   Positive             7 (100%)           14 (100%)            9 (100%)           30 (100%)  
#>   Negative                -                   -                   -                   -      
#> 
#> WEEK 4 DAY 29
#>   n                       7                   14                  9                   30     
#>   Positive             7 (100%)           14 (100%)            9 (100%)           30 (100%)  
#>   Negative                -                   -                   -                   -      
#> 
#> WEEK 5 DAY 36
#>   n                       7                   14                  9                   30     
#>   Positive             7 (100%)           14 (100%)            9 (100%)           30 (100%)  
#>   Negative                -                   -                   -                   -