Summarize an Object for Different Groups with by Variable

Summarize an Object for Different Groups with by Variable

t_summary_by(x, by, col_by, col_N, total = NULL, ...)

Arguments

x
vector

by
a factor of length nrow(x) with no missing values. The levels of by define the summary sub-groups in the table.

col_by
a factor of length nrow(x) with no missing values. The levels of col_by define the columns in the table.

col_N
a table object with the reference population used for the header of the table. See examples below.

total
character string that will be used as a label for a column with pooled total population. If the levels of col_by are the only columns of interest then total should be NULL.

...
arguments passed on to methods

Examples

t_summary_by(
  x = sample(c(TRUE, FALSE), 100, TRUE),
  by = factor(sample(paste("day", 1:5), 100, TRUE)),
  col_by = factor(sample(paste("ARM", LETTERS[1:3]), 100, TRUE)),
  col_N = table(rep(paste("ARM", LETTERS[1:3]), 10)),
  total = "All Patients"
)

#>                   ARM A               ARM B               ARM C            All Patients
#>                   (N=10)              (N=10)              (N=10)              (N=30)   
#> ---------------------------------------------------------------------------------------
#> day 1
#>   n                 5                   7                   5                   17     
#>   TRUE           1 (20%)            2 (28.57%)           4 (80%)            7 (41.18%) 
#>   FALSE          4 (80%)            5 (71.43%)           1 (20%)           10 (58.82%) 
#> 
#> day 2
#>   n                 6                   11                  7                   24     
#>   TRUE          2 (33.33%)          7 (63.64%)          5 (71.43%)         14 (58.33%) 
#>   FALSE         4 (66.67%)          4 (36.36%)          2 (28.57%)         10 (41.67%) 
#> 
#> day 3
#>   n                 6                   5                   5                   16     
#>   TRUE           3 (50%)             2 (40%)             2 (40%)            7 (43.75%) 
#>   FALSE          3 (50%)             3 (60%)             3 (60%)            9 (56.25%) 
#> 
#> day 4
#>   n                 5                   12                  6                   23     
#>   TRUE           3 (60%)            7 (58.33%)           3 (50%)           13 (56.52%) 
#>   FALSE          2 (40%)            5 (41.67%)           3 (50%)           10 (43.48%) 
#> 
#> day 5
#>   n                 8                   6                   6                   20     
#>   TRUE           6 (75%)             3 (50%)             3 (50%)             12 (60%)  
#>   FALSE          2 (25%)             3 (50%)             3 (50%)             8 (40%)   

t_summary_by(
  x = sample(c(TRUE, FALSE), 100, TRUE),
  by = factor(sample(paste("day", 1:5), 100, TRUE)),
  col_by = factor(sample(paste("ARM", LETTERS[1:3]), 100, TRUE)),
  col_N = table(rep(paste("ARM", LETTERS[1:3]), 10)),
  total = "All Patients",
  denominator = "N"
)

#>                   ARM A               ARM B               ARM C            All Patients
#>                   (N=10)              (N=10)              (N=10)              (N=30)   
#> ---------------------------------------------------------------------------------------
#> day 1
#>   n                 2                   4                   2                   8      
#>   TRUE           1 (10%)             2 (20%)             1 (10%)            4 (13.33%) 
#>   FALSE          1 (10%)             2 (20%)             1 (10%)            4 (13.33%) 
#> 
#> day 2
#>   n                 3                   12                  11                  26     
#>   TRUE           1 (10%)             7 (70%)             7 (70%)             15 (50%)  
#>   FALSE          2 (20%)             5 (50%)             4 (40%)           11 (36.67%) 
#> 
#> day 3
#>   n                 10                  6                   9                   25     
#>   TRUE           7 (70%)             1 (10%)             5 (50%)           13 (43.33%) 
#>   FALSE          3 (30%)             5 (50%)             4 (40%)             12 (40%)  
#> 
#> day 4
#>   n                 7                   6                   9                   22     
#>   TRUE           4 (40%)                -                4 (40%)            8 (26.67%) 
#>   FALSE          3 (30%)             6 (60%)             5 (50%)           14 (46.67%) 
#> 
#> day 5
#>   n                 5                   8                   6                   19     
#>   TRUE           4 (40%)             3 (30%)             2 (20%)             9 (30%)   
#>   FALSE          1 (10%)             5 (50%)             4 (40%)           10 (33.33%) 

ANL <- data.frame(
 aval = sample(letters[1:5], 100, TRUE),
 avisit = factor(sample(c(
   "Screening", "Day 5", "Day 10"
 ), 100, TRUE), levels = c("Screening", "Day 5", "Day 10")),
 col_by = factor(sample(paste("ARM", LETTERS[1:3]), 100, TRUE))
)
ANL <- var_relabel(ANL, aval = "Analysis Value", avisit = "Analysis Visit" )
levels(ANL$aval) <- c(levels(ANL$aval), "z")

t_summary_by(
  x = ANL$aval,
  by = ANL$avisit,
  col_by = ANL$col_by,
  col_N = c(10, 15, 20)
)

#> Analysis Visit            ARM A             ARM B             ARM C   
#>   Analysis Value          (N=10)            (N=15)            (N=20)  
#> ----------------------------------------------------------------------
#> Screening
#>   n                         12                10                8     
#>   a                     4 (33.33%)         3 (30%)          1 (12.5%) 
#>   b                     1 (8.33%)          2 (20%)           2 (25%)  
#>   c                     4 (33.33%)            -              4 (50%)  
#>   d                     1 (8.33%)          2 (20%)              -     
#>   e                     2 (16.67%)         3 (30%)          1 (12.5%) 
#>   z                         -                 -                 -     
#> 
#> Day 5
#>   n                         12                19                8     
#>   a                     1 (8.33%)         4 (21.05%)        3 (37.5%) 
#>   b                      3 (25%)          3 (15.79%)            -     
#>   c                     4 (33.33%)        3 (15.79%)        1 (12.5%) 
#>   d                      3 (25%)          2 (10.53%)        1 (12.5%) 
#>   e                     1 (8.33%)         7 (36.84%)        3 (37.5%) 
#>   z                         -                 -                 -     
#> 
#> Day 10
#>   n                         12                10                9     
#>   a                      3 (25%)           2 (20%)              -     
#>   b                      3 (25%)              -             2 (22.22%)
#>   c                     1 (8.33%)          2 (20%)          1 (11.11%)
#>   d                     2 (16.67%)         2 (20%)          5 (55.56%)
#>   e                      3 (25%)           4 (40%)          1 (11.11%)
#>   z                         -                 -                 -     

t_summary_by(
  x = ANL$aval,
  by = ANL$avisit,
  col_by = ANL$col_by,
  col_N = c(10, 15, 20),
  total = "All",
  denominator = "N",
  drop_levels = TRUE
)

#> Analysis Visit             ARM A              ARM B              ARM C               All    
#>   Analysis Value          (N=10)             (N=15)             (N=20)             (N=45)   
#> --------------------------------------------------------------------------------------------
#> Screening
#>   n                         12                 10                  8                 30     
#>   a                       4 (40%)            3 (20%)            1 (5%)           8 (17.78%) 
#>   b                       1 (10%)          2 (13.33%)           2 (10%)          5 (11.11%) 
#>   c                       4 (40%)               -               4 (20%)          8 (17.78%) 
#>   d                       1 (10%)          2 (13.33%)              -              3 (6.67%) 
#>   e                       2 (20%)            3 (20%)            1 (5%)           6 (13.33%) 
#> 
#> Day 5
#>   n                         12                 19                  8                 39     
#>   a                       1 (10%)          4 (26.67%)           3 (15%)          8 (17.78%) 
#>   b                       3 (30%)            3 (20%)               -             6 (13.33%) 
#>   c                       4 (40%)            3 (20%)            1 (5%)           8 (17.78%) 
#>   d                       3 (30%)          2 (13.33%)           1 (5%)           6 (13.33%) 
#>   e                       1 (10%)          7 (46.67%)           3 (15%)          11 (24.44%)
#> 
#> Day 10
#>   n                         12                 10                  9                 31     
#>   a                       3 (30%)          2 (13.33%)              -             5 (11.11%) 
#>   b                       3 (30%)               -               2 (10%)          5 (11.11%) 
#>   c                       1 (10%)          2 (13.33%)           1 (5%)            4 (8.89%) 
#>   d                       2 (20%)          2 (13.33%)           5 (25%)            9 (20%)  
#>   e                       3 (30%)          4 (26.67%)           1 (5%)           8 (17.78%)