Time to Event Table

The time to event table summarizes time to event data with different models as described in the details section.

t_tte(formula, data, event_descr, time_points, time_unit = "month",
  ties = "exact")

Arguments

formula
a survival formula, the arm variable needs to be wrapped in arm(). The strata() special will only be used for the stratified analysis. If there is not strata specification then the stratified analysis is omitted.

data
a data.frame with all the variable that are used in formula

event_descr
a factor that partitions the the events into earliest contributing event. The variable name can be provided unquoted in which case it is looked up in data and then the calling environment.

time_points
numeric vector with times displayed in the time point analysis, if NULL this section of the table will be omitted

time_unit
a string with the unit of the tte argument

ties
passed forward to coxph

Value

an rtable object

Details

The time to event section is derived from a kaplan meier estimator for the formula argument with the strata special dropped.

The stratified and unstratified analysis is evaluated pair-wise (reference to comparison) and survdiff is used to get the p-value whereas coxph is used to calculate the hazard ratio and confidence interval.

The time point analysis is based on the kaplan meier estimator.

Examples

library(random.cdisc.data)

ASL <- radam("ASL", start_with = list(ARM = paste("ARM", LETTERS[1:3])))
ATE <- radam("ATE", ADSL = ASL)

ASL$ARM <- as.factor(ASL$ARM)
ATE_f <- subset(ATE, PARAMCD == "OS")

ANL <- merge(ASL, ATE_f, all.x =TRUE, all.y = FALSE, by = c("USUBJID", "STUDYID"))

tbl <- t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM) + strata(SEX),
  data = ANL,
  event_descr = factor(EVNTDESC),
  time_points = c(6, 2000),
  time_unit = "month"
)

tbl

#>                                               ARM A                   ARM B                   ARM C      
#>                                               (N=36)                  (N=25)                  (N=39)     
#> ---------------------------------------------------------------------------------------------------------
#> Patients without event (%)                 19 (52.78%)               12 (48%)              20 (51.28%)   
#>   Earliest Contributing Event
#>     Death                                       7                       10                      8        
#>     Disease Progression                         5                       2                       6        
#>     First Response                              7                       0                       6        
#> Patients without event (%)                 17 (47.22%)               13 (52%)              19 (48.72%)   
#> 
#> Time to Event (months)
#>   Median                                      176.16                  201.06                  106.33     
#>     95% CI                                (147.2, 218.6)          (77.9, 231.1)           (43.5, 149.5)  
#>   25% and 75%−ile                          108.4, 220.5            76.3, 231.1             35.7, 236.9   
#>   Range                                     0 to 480.6              0 to 241.2              0 to 308.8   
#> 
#> Unstratified Analysis
#>   p-value (log-rank)                                                  0.6022                  0.1406     
#> 
#>   Hazard Ratio                                                        1.221                   1.6172     
#>     95% CI                                                       (0.5755, 2.5905)        (0.8483, 3.0829)
#> 
#> Stratified Analysis
#>   p-value (log-rank)                                                  0.794                   0.1843     
#> 
#>   Hazard Ratio                                                        1.108                   1.558      
#>     95% CI                                                       (0.5129, 2.3938)        (0.8057, 3.0126)
#> 
#> Time Point Analysis
#>   6 months
#>     Patients remaining at risk                  32                      23                      36       
#>     Event Free Rate (%)                       97.14%                  95.83%                  97.3%      
#>       95% CI                               (91.62, 100)            (87.84, 100)            (92.07, 100)  
#>     Difference in Event Free Rate                                     -1.31                    0.15      
#>       95% CI                                                      (-11.02, 8.41)          (-7.45, 7.75)  
#>     p-value (Z-test)                                                  0.7916                  0.9682     
#> 
#>   2000 months
#>     -- no data