Time to Event Table

The time to event table summarizes time to event data with different models as described in the details section.

t_tte(
  formula,
  data,
  col_N,
  event_descr = NULL,
  time_points = NULL,
  time_unit = "month",
  conf_level = 0.95,
  conf_type_survfit = c("plain", "none", "log", "log-log"),
  probs_survfit = c(0.25, 0.75),
  pval_method_coxph = c("log-rank", "wald", "likelihood"),
  ties_coxph = c("efron", "breslow", "exact"),
  table_tree = FALSE
)

Arguments

formula
(formula)
Specifies survival model. The arm variable needs to be wrapped in arm. The strata special will only be used for the stratified analysis. If there is not strata specification then the stratified analysis is omitted.

data
(data.frame)
Contains all the variable that are used in formula

col_N
(integer vector)
Contains the reference population per level of col_by. This by default is often set to get_N(col_by). Names of the list are ignored, they are added in the order they appear. The default may not appropriate and should be modified by the user when col_by contains more than one record per subject.

event_descr
(factor vector) that partitions the the events into earliest contributing event. The variable name can be provided unquoted in which case it is looked up in data and then the calling environment.

time_points
(numeric vector) with times displayed in the time point analysis, if NULL this section of the table will be omitted

time_unit
(character value) with the unit of the tte argument

conf_level
(numeric value or named numeric vector) either a single number or a named vector of confidence levels where the names are survfit, coxph, and ztest

conf_type_survfit
(one of "plain" (default), "none", "log", "log-log"). Used in conf.type in survfit

probs_survfit
(numeric vector of length two) to specify the quantiles in quantile

pval_method_coxph
passed as argument pval_method to function s_coxph_pairwise

ties_coxph
passed as argument ties to coxph

table_tree
(logical value)
If FALSE then a single rtable gets returned, if TRUE a node object gets returned with the elementary rtable objects.

Value

an rtable object

Details

The time to event section is derived from a Kaplan-Meier estimator for the formula argument with the strata special dropped.

The stratified and unstratified analysis is evaluated pair-wise (reference to comparison) and coxph is used to get the p-value, calculate the hazard ratio and confidence interval.

The time point analysis is based on the Kaplan-Meier estimator.

Examples

library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)

ADTTE <- radtte(cached = TRUE)
ADTTE_f <- subset(ADTTE, PARAMCD == "EFS" & BMEASIFL == "Y")

t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM),
  data = ADTTE_f,
  col_N = table(ADSL$ARM)
)

#>                                      A: Drug X               B: Placebo            C: Combination 
#>                                       (N=134)                 (N=134)                 (N=132)     
#> --------------------------------------------------------------------------------------------------
#> Patients with event (%)             49 (72.06%)             67 (91.78%)             44 (70.97%)   
#> Patients without event (%)          19 (27.94%)              6 (8.22%)              18 (29.03%)   
#> 
#> Time to Event (months)
#>   Median                               70.18                   99.46                   44.58      
#>     95% CI                          (51, 127.5)            (76.3, 121.3)            (29.2, 66.5)  
#>   25% and 75%-ile                   25.3, 186.7             30.6, 147.1              17.1, 93.9   
#>   Range (censored)                 12.4 to 241.5            0.2 to 108.8              1 to 184    
#>   Range (event)                     1.5 to 540.2            1.2 to 287.8            2.1 to 190.4  
#> 
#> Unstratified Analysis
#>   p-value (Log-rank)                                           0.3102                  0.0038     
#> 
#>   Hazard Ratio                                                 1.2143                  1.8698     
#>     95% CI                                                (0.8341, 1.7678)        (1.2163, 2.8747)

t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM),
  data = ADTTE_f,
  col_N = table(ADSL$ARM),
  event_descr = factor(ADTTE_f$EVNTDESC)
)

#>                                           A: Drug X               B: Placebo            C: Combination 
#>                                            (N=134)                 (N=134)                 (N=132)     
#> -------------------------------------------------------------------------------------------------------
#> Patients with event (%)                  49 (72.06%)             67 (91.78%)             44 (70.97%)   
#>   Earliest Contributing Event
#>     Adverse Event                             6                       11                      11       
#>     Death                                     14                      17                      8        
#>     Disease Progression                       8                       14                      6        
#>     Last Date Known To Be Alive               12                      16                      7        
#>     Last Tumor Assessment                     9                       9                       12       
#> Patients without event (%)               19 (27.94%)              6 (8.22%)              18 (29.03%)   
#> 
#> Time to Event (months)
#>   Median                                    70.18                   99.46                   44.58      
#>     95% CI                               (51, 127.5)            (76.3, 121.3)            (29.2, 66.5)  
#>   25% and 75%-ile                        25.3, 186.7             30.6, 147.1              17.1, 93.9   
#>   Range (censored)                      12.4 to 241.5            0.2 to 108.8              1 to 184    
#>   Range (event)                          1.5 to 540.2            1.2 to 287.8            2.1 to 190.4  
#> 
#> Unstratified Analysis
#>   p-value (Log-rank)                                                0.3102                  0.0038     
#> 
#>   Hazard Ratio                                                      1.2143                  1.8698     
#>     95% CI                                                     (0.8341, 1.7678)        (1.2163, 2.8747)

t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM),
  data = ADTTE_f,
  col_N = table(ADSL$ARM),
  event_descr = factor(ADTTE_f$EVNTDESC),
  time_points = c(6, 12, 360),
  time_unit = "month",
)

#>                                                 A: Drug X               B: Placebo            C: Combination 
#>                                                  (N=134)                 (N=134)                 (N=132)     
#> -------------------------------------------------------------------------------------------------------------
#> Patients with event (%)                        49 (72.06%)             67 (91.78%)             44 (70.97%)   
#>   Earliest Contributing Event
#>     Adverse Event                                   6                       11                      11       
#>     Death                                           14                      17                      8        
#>     Disease Progression                             8                       14                      6        
#>     Last Date Known To Be Alive                     12                      16                      7        
#>     Last Tumor Assessment                           9                       9                       12       
#> Patients without event (%)                     19 (27.94%)              6 (8.22%)              18 (29.03%)   
#> 
#> Time to Event (months)
#>   Median                                          70.18                   99.46                   44.58      
#>     95% CI                                     (51, 127.5)            (76.3, 121.3)            (29.2, 66.5)  
#>   25% and 75%-ile                              25.3, 186.7             30.6, 147.1              17.1, 93.9   
#>   Range (censored)                            12.4 to 241.5            0.2 to 108.8              1 to 184    
#>   Range (event)                                1.5 to 540.2            1.2 to 287.8            2.1 to 190.4  
#> 
#> Unstratified Analysis
#>   p-value (Log-rank)                                                      0.3102                  0.0038     
#> 
#>   Hazard Ratio                                                            1.2143                  1.8698     
#>     95% CI                                                           (0.8341, 1.7678)        (1.2163, 2.8747)
#> 
#> Time Point Analysis
#>   6 months
#>         Patients remaining at risk                  61                      67                      51       
#>         Event Free Rate (%)                       89.71%                  93.06%                  89.75%     
#>           95% CI                              (82.48, 96.93)          (87.18, 98.93)          (81.98, 97.53) 
#>         Difference in Event Free Rate                                      3.35                    0.05      
#>           95% CI                                                      (-5.96, 12.66)         (-10.56, 10.66) 
#>         p-value (Z-test)                                                  0.4806                  0.993      
#> 
#>   12 months
#>         Patients remaining at risk                  60                      64                      44       
#>         Event Free Rate (%)                       88.24%                  90.26%                  80.8%      
#>           95% CI                              (80.58, 95.89)          (83.4, 97.11)           (70.58, 91.02) 
#>         Difference in Event Free Rate                                      2.02                   -7.43      
#>           95% CI                                                      (-8.26, 12.3)           (-20.2, 5.34)  
#>         p-value (Z-test)                                                   0.7                    0.2541     
#> 
#>   360 months
#>         -- not enough data

t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM),
  data = ADTTE_f,
  col_N = table(ADSL$ARM),
  event_descr = factor(ADTTE_f$EVNTDESC),
  time_points = c(6, 12, 360),
  conf_level = c(survfit = 0.91, coxph = 0.95, ztest = 0.99)
)

#>                                                 A: Drug X               B: Placebo            C: Combination 
#>                                                  (N=134)                 (N=134)                 (N=132)     
#> -------------------------------------------------------------------------------------------------------------
#> Patients with event (%)                        49 (72.06%)             67 (91.78%)             44 (70.97%)   
#>   Earliest Contributing Event
#>     Adverse Event                                   6                       11                      11       
#>     Death                                           14                      17                      8        
#>     Disease Progression                             8                       14                      6        
#>     Last Date Known To Be Alive                     12                      16                      7        
#>     Last Tumor Assessment                           9                       9                       12       
#> Patients without event (%)                     19 (27.94%)              6 (8.22%)              18 (29.03%)   
#> 
#> Time to Event (months)
#>   Median                                          70.18                   99.46                   44.58      
#>     91% CI                                    (51.6, 115.2)           (81.8, 118.6)            (31.6, 61.4)  
#>   25% and 75%-ile                              25.3, 186.7             30.6, 147.1              17.1, 93.9   
#>   Range (censored)                            12.4 to 241.5            0.2 to 108.8              1 to 184    
#>   Range (event)                                1.5 to 540.2            1.2 to 287.8            2.1 to 190.4  
#> 
#> Unstratified Analysis
#>   p-value (Log-rank)                                                      0.3102                  0.0038     
#> 
#>   Hazard Ratio                                                            1.2143                  1.8698     
#>     95% CI                                                           (0.8341, 1.7678)        (1.2163, 2.8747)
#> 
#> Time Point Analysis
#>   6 months
#>         Patients remaining at risk                  61                      67                      51       
#>         Event Free Rate (%)                       89.71%                  93.06%                  89.75%     
#>           91% CI                              (83.46, 95.95)          (87.98, 98.13)          (83.03, 96.48) 
#>         Difference in Event Free Rate                                      3.35                    0.05      
#>           99% CI                                                      (-8.88, 15.58)          (-13.9, 13.99) 
#>         p-value (Z-test)                                                  0.4806                  0.993      
#> 
#>   12 months
#>         Patients remaining at risk                  60                      64                      44       
#>         Event Free Rate (%)                       88.24%                  90.26%                  80.8%      
#>           91% CI                              (81.61, 94.86)          (84.32, 96.19)          (71.96, 89.65) 
#>         Difference in Event Free Rate                                      2.02                   -7.43      
#>           99% CI                                                     (-11.49, 15.53)          (-24.21, 9.35) 
#>         p-value (Z-test)                                                   0.7                    0.2541     
#> 
#>   360 months
#>         -- not enough data

t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM) + strata(SEX),
  data = ADTTE_f,
  col_N = table(ADSL$ARM),
  event_descr = factor(ADTTE_f$EVNTDESC),
  time_points = c(6, 12, 24),
  time_unit = "month",
  pval_method_coxph = "wald",
  conf_type_survfit = "log-log",
  ties_coxph = "exact"
)

#>                                                 A: Drug X               B: Placebo            C: Combination 
#>                                                  (N=134)                 (N=134)                 (N=132)     
#> -------------------------------------------------------------------------------------------------------------
#> Patients with event (%)                        49 (72.06%)             67 (91.78%)             44 (70.97%)   
#>   Earliest Contributing Event
#>     Adverse Event                                   6                       11                      11       
#>     Death                                           14                      17                      8        
#>     Disease Progression                             8                       14                      6        
#>     Last Date Known To Be Alive                     12                      16                      7        
#>     Last Tumor Assessment                           9                       9                       12       
#> Patients without event (%)                     19 (27.94%)              6 (8.22%)              18 (29.03%)   
#> 
#> Time to Event (months)
#>   Median                                          70.18                   99.46                   44.58      
#>     95% CI                                    (47.6, 127.5)           (75.7, 121.3)            (29.2, 61.4)  
#>   25% and 75%-ile                              25.3, 186.7             30.6, 147.1              17.1, 93.9   
#>   Range (censored)                            12.4 to 241.5            0.2 to 108.8              1 to 184    
#>   Range (event)                                1.5 to 540.2            1.2 to 287.8            2.1 to 190.4  
#> 
#> Unstratified Analysis
#>   p-value (Wald)                                                          0.3109                  0.0043     
#> 
#>   Hazard Ratio                                                            1.2143                  1.8698     
#>     95% CI                                                           (0.8341, 1.7678)        (1.2163, 2.8747)
#> 
#> Stratified Analysis
#>   p-value (Wald)                                                          0.2505                  0.0029     
#> 
#>   Hazard Ratio                                                            1.2499                  1.9432     
#>     95% CI                                                           (0.8543, 1.8287)        (1.2549, 3.0091)
#> 
#> Time Point Analysis
#>   6 months
#>         Patients remaining at risk                  61                      67                      51       
#>         Event Free Rate (%)                       89.71%                  93.06%                  89.75%     
#>           95% CI                              (79.62, 94.95)          (84.12, 97.05)          (78.59, 95.26) 
#>         Difference in Event Free Rate                                      3.35                    0.05      
#>           95% CI                                                      (-5.96, 12.66)         (-10.56, 10.66) 
#>         p-value (Z-test)                                                  0.4806                  0.993      
#> 
#>   12 months
#>         Patients remaining at risk                  60                      64                      44       
#>         Event Free Rate (%)                       88.24%                  90.26%                  80.8%      
#>           95% CI                              (77.85, 93.94)          (80.64, 95.23)          (67.99, 88.89) 
#>         Difference in Event Free Rate                                      2.02                   -7.43      
#>           95% CI                                                      (-8.26, 12.3)           (-20.2, 5.34)  
#>         p-value (Z-test)                                                   0.7                    0.2541     
#> 
#>   24 months
#>         Patients remaining at risk                  51                      57                      34       
#>         Event Free Rate (%)                       77.68%                  80.38%                  69.09%     
#>           95% CI                              (65.72, 85.9)           (69.13, 87.88)          (54.94, 79.59) 
#>         Difference in Event Free Rate                                      2.7                    -8.59      
#>           95% CI                                                     (-10.87, 16.28)          (-24.44, 7.25) 
#>         p-value (Z-test)                                                  0.6962                  0.2878     

# table_tree = TRUE
tbl3 <- t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM) + strata(SEX),
  data = ADTTE_f,
  col_N = table(ADSL$ARM),
  event_descr = factor(ADTTE_f$EVNTDESC),
  time_points = c(6, 2000),
  time_unit = "month",
  table_tree = TRUE
)
summary(tbl3)

#>                                            index            content             dim            hide label          format  
#> ---------------------------------------------------------------------------------------------------------------------------
#> root                                                          NULL                                TRUE                     
#>   Patient events                             1                NULL                                TRUE               0     
#>     Patients with event                     1, 1             rtable            1 x 3              TRUE              0, 1   
#>       Earliest Contributing Event         1, 1, 1            rtable            5 x 3             FALSE                     
#>     Patients without events                 1, 2             rtable            1 x 3              TRUE                     
#>   Time to Event (months)                     2               rtable            5 x 3             FALSE                     
#>   Unstratified Analysis                      3                NULL                               FALSE                     
#>     p-value                                 3, 1             rtable            1 x 3              TRUE                     
#>     Hazard Ratio                            3, 2             rtable            2 x 3              TRUE                     
#>   Stratified Analysis                        4                NULL                               FALSE                     
#>     p-value                                 4, 1             rtable            1 x 3              TRUE                     
#>     Hazard Ratio                            4, 2             rtable            2 x 3              TRUE                     
#>   Time Point Analysis                        5                NULL                               FALSE                     
#>     6 months                                5, 1             rtable            6 x 3             FALSE                     
#>     2000 months                             5, 2             rtable            1 x 3             FALSE                     

to_rtable(tbl3)

#>                                                 A: Drug X               B: Placebo            C: Combination 
#>                                                  (N=134)                 (N=134)                 (N=132)     
#> -------------------------------------------------------------------------------------------------------------
#> Patients with event (%)                        49 (72.06%)             67 (91.78%)             44 (70.97%)   
#>   Earliest Contributing Event
#>     Adverse Event                                   6                       11                      11       
#>     Death                                           14                      17                      8        
#>     Disease Progression                             8                       14                      6        
#>     Last Date Known To Be Alive                     12                      16                      7        
#>     Last Tumor Assessment                           9                       9                       12       
#> Patients without event (%)                     19 (27.94%)              6 (8.22%)              18 (29.03%)   
#> 
#> Time to Event (months)
#>   Median                                          70.18                   99.46                   44.58      
#>     95% CI                                     (51, 127.5)            (76.3, 121.3)            (29.2, 66.5)  
#>   25% and 75%-ile                              25.3, 186.7             30.6, 147.1              17.1, 93.9   
#>   Range (censored)                            12.4 to 241.5            0.2 to 108.8              1 to 184    
#>   Range (event)                                1.5 to 540.2            1.2 to 287.8            2.1 to 190.4  
#> 
#> Unstratified Analysis
#>   p-value (Log-rank)                                                      0.3102                  0.0038     
#> 
#>   Hazard Ratio                                                            1.2143                  1.8698     
#>     95% CI                                                           (0.8341, 1.7678)        (1.2163, 2.8747)
#> 
#> Stratified Analysis
#>   p-value (Log-rank)                                                      0.2496                  0.0025     
#> 
#>   Hazard Ratio                                                            1.2499                  1.9432     
#>     95% CI                                                           (0.8543, 1.8287)        (1.2549, 3.0091)
#> 
#> Time Point Analysis
#>   6 months
#>         Patients remaining at risk                  61                      67                      51       
#>         Event Free Rate (%)                       89.71%                  93.06%                  89.75%     
#>           95% CI                              (82.48, 96.93)          (87.18, 98.93)          (81.98, 97.53) 
#>         Difference in Event Free Rate                                      3.35                    0.05      
#>           95% CI                                                      (-5.96, 12.66)         (-10.56, 10.66) 
#>         p-value (Z-test)                                                  0.4806                  0.993      
#> 
#>   2000 months
#>         -- no data