Time to Event Table

The time to event table summarizes time to event data with different models as described in the details section.

t_tte(
  formula,
  data,
  event_descr,
  time_points,
  time_unit = "month",
  ties = "exact",
  table_tree = FALSE
)

Arguments

formula
a survival formula, the arm variable needs to be wrapped in arm. The strata special will only be used for the stratified analysis. If there is not strata specification then the stratified analysis is omitted.

data
a data.frame with all the variable that are used in formula

event_descr
a factor that partitions the the events into earliest contributing event. The variable name can be provided unquoted in which case it is looked up in data and then the calling environment.

time_points
numeric vector with times displayed in the time point analysis, if NULL this section of the table will be omitted

time_unit
a string with the unit of the tte argument

ties
passed forward to coxph

table_tree
logical, if FALSE then a single rtable gets returned, if TRUE a node object gets returned with the elementary rtable objects.

Value

an rtable object

Details

The time to event section is derived from a Kaplan-Meier estimator for the formula argument with the strata special dropped.

The stratified and unstratified analysis is evaluated pair-wise (reference to comparison) and survdiff is used to get the p-value whereas coxph is used to calculate the hazard ratio and confidence interval.

The time point analysis is based on the Kaplan-Meier estimator.

See also

Examples

# \donttest{
library(random.cdisc.data)
ADSL <- cadsl

ADTTE <- cadtte
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")

tbl <- t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM) + strata(SEX),
  data = ADTTE_f,
  event_descr = factor(EVNTDESC),
  time_points = c(6, 2000),
  time_unit = "month"
)

tbl

#>                                                 A: Drug X               B: Placebo            C: Combination 
#> -------------------------------------------------------------------------------------------------------------
#> Patients with event (%)                        82 (61.19%)             100 (74.63%)            89 (67.42%)   
#>   Earliest Contributing Event
#>     Adverse Event                                   16                      16                      13       
#>     Death                                           17                      26                      18       
#>     Disease Progression                             15                      20                      19       
#>     Last Date Known To Be Alive                     17                      18                      19       
#>     Last Tumor Assessment                           17                      20                      20       
#> Patients without event (%)                     52 (38.81%)             34 (25.37%)             43 (32.58%)   
#> 
#> Time to Event (months)
#>   Median                                          96.86                   92.31                   79.23      
#>     95% CI                                     (74, 125.3)            (71.2, 102.7)           (55.8, 107.4)  
#>   25% and 75%-ile                              37.5, 184.6             32.6, 174.7             26.3, 155.7   
#>   Range (censored)                             0.5 to 235.8             0.4 to 386             1.9 to 140.5  
#>   Range (event)                                1.7 to 479.2            1.8 to 395.9            2.7 to 326.1  
#> 
#> Unstratified Analysis
#>   p-value (log-rank)                                                      0.8512                  0.0918     
#> 
#>   Hazard Ratio                                                            1.0287                  1.2955     
#>     95% CI                                                           (0.7652, 1.3831)        (0.958, 1.7519) 
#> 
#> Stratified Analysis
#>   p-value (log-rank)                                                      0.8392                  0.0926     
#> 
#>   Hazard Ratio                                                            1.0315                  1.3096     
#>     95% CI                                                           (0.7648, 1.391)         (0.9555, 1.7948)
#> 
#> Time Point Analysis
#>   6 months
#>         Patients remaining at risk                 124                     126                     120       
#>         Event Free Rate (%)                       96.92%                  95.47%                  94.53%     
#>           95% CI                              (93.95, 99.89)          (91.93, 99.01)          (90.59, 98.47) 
#>         Difference in Event Free Rate                                     -1.45                   -2.39      
#>           95% CI                                                      (-6.07, 3.17)           (-7.33, 2.54)  
#>         p-value (Z-test)                                                  0.5386                  0.3422     
#> 
#>   2000 months
#>         -- no data

Viewer(tbl)

# table_tree = TRUE
tbl <- t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM) + strata(SEX),
  data = ADTTE_f,
  event_descr = factor(EVNTDESC),
  time_points = c(6, 2000),
  time_unit = "month",
  table_tree = TRUE
)
summary(tbl)

#>                                            index            content             dim            hide label          format  
#> ---------------------------------------------------------------------------------------------------------------------------
#> root                                                          NULL                                TRUE                     
#>   Patient events                             1                NULL                                TRUE               0     
#>     Patients with event                     1, 1             rtable            1 x 3              TRUE              0, 1   
#>       Earliest Contributing Event         1, 1, 1            rtable            5 x 3             FALSE                     
#>     Patients without events                 1, 2             rtable            1 x 3              TRUE                     
#>   Time to Event (months)                     2               rtable            5 x 3             FALSE                     
#>   Unstratified Analysis                      3                NULL                               FALSE                     
#>     p-value                                 3, 1             rtable            1 x 3              TRUE                     
#>     Hazard Ratio                            3, 2             rtable            2 x 3              TRUE                     
#>   Stratified Analysis                        4                NULL                               FALSE                     
#>     p-value                                 4, 1             rtable            1 x 3              TRUE                     
#>     Hazard Ratio                            4, 2             rtable            2 x 3              TRUE                     
#>   Time Point Analysis                        5                NULL                               FALSE                     
#>     6 months                                5, 1             rtable            6 x 3             FALSE                     
#>     2000 months                             5, 2             rtable            1 x 3             FALSE                     

# }