Time to Event Table

The time to event table summarizes time to event data with different models as described in the details section.

t_tte(formula, data, event_descr, time_points, time_unit = "month",
  ties = "exact")

Arguments

formula
a survival formula, the arm variable needs to be wrapped in arm. The strata special will only be used for the stratified analysis. If there is not strata specification then the stratified analysis is omitted.

data
a data.frame with all the variable that are used in formula

event_descr
a factor that partitions the the events into earliest contributing event. The variable name can be provided unquoted in which case it is looked up in data and then the calling environment.

time_points
numeric vector with times displayed in the time point analysis, if NULL this section of the table will be omitted

time_unit
a string with the unit of the tte argument

ties
passed forward to coxph

Value

an rtable object

Details

The time to event section is derived from a Kaplan-Meier estimator for the formula argument with the strata special dropped.

The stratified and unstratified analysis is evaluated pair-wise (reference to comparison) and survdiff is used to get the p-value whereas coxph is used to calculate the hazard ratio and confidence interval.

The time point analysis is based on the Kaplan-Meier estimator.

See also

Examples

library(random.cdisc.data)
ADSL <- cadsl

ADTTE <- cadtte
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")

tbl <- t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM) + strata(SEX),
  data = ADTTE_f,
  event_descr = factor(EVNTDESC),
  time_points = c(6, 2000),
  time_unit = "month"
)

tbl

#>                                             A: Drug X               B: Placebo            C: Combination 
#>                                              (N=142)                 (N=133)                 (N=125)     
#> ---------------------------------------------------------------------------------------------------------
#> Patients with event (%)                    88 (61.97%)             105 (78.95%)             69 (55.2%)   
#>   Earliest Contributing Event
#>     Adverse Event                               18                      15                      11       
#>     Death                                       17                      20                      15       
#>     Disease Progression                         15                      21                      8        
#>     Last Date Known To Be Alive                 13                      23                      14       
#>     Last Tumor Assessment                       25                      26                      21       
#> Patients without event (%)                 54 (38.03%)             28 (21.05%)              56 (44.8%)   
#> 
#> Time to Event (months)
#>   Median                                      88.67                    77.4                   72.41      
#>     95% CI                                (61.2, 124.8)           (56.7, 104.8)            (60.7, 90.1)  
#>   25% and 75%-ile                          30.5, 176.3             27.6, 152.7              40, 135.2    
#>   Range (censored)                          1 to 338.8             0.2 to 372.9            2.3 to 395.1  
#>   Range (event)                            0.2 to 521.8            0.4 to 535.1            0.1 to 244.2  
#> 
#> Unstratified Analysis
#>   p-value (log-rank)                                                  0.3759                  0.6758     
#> 
#>   Hazard Ratio                                                        1.1372                  1.0701     
#>     95% CI                                                       (0.8554, 1.5118)        (0.7789, 1.4703)
#> 
#> Stratified Analysis
#>   p-value (log-rank)                                                  0.5044                  0.8566     
#> 
#>   Hazard Ratio                                                        1.1034                  1.0303     
#>     95% CI                                                       (0.8265, 1.4731)        (0.7453, 1.4244)
#> 
#> Time Point Analysis
#>   6 months
#>     Patients remaining at risk                 127                     123                     110       
#>     Event Free Rate (%)                       92.86%                  93.89%                  93.47%     
#>       95% CI                              (88.59, 97.12)          (89.79, 97.99)          (89.09, 97.85) 
#>     Difference in Event Free Rate                                      1.04                    0.61      
#>       95% CI                                                      (-4.88, 6.95)           (-5.51, 6.72)  
#>     p-value (Z-test)                                                  0.7317                  0.8452     
#> 
#>   2000 months
#>     -- no data

Viewer(tbl)