Time to Event Table

The time to event table summarizes time to event data with different models as described in the details section.

t_tte(formula, data, event_descr, time_points, time_unit = "month",
  ties = "exact")

Arguments

formula
a survival formula, the arm variable needs to be wrapped in arm. The strata special will only be used for the stratified analysis. If there is not strata specification then the stratified analysis is omitted.

data
a data.frame with all the variable that are used in formula

event_descr
a factor that partitions the the events into earliest contributing event. The variable name can be provided unquoted in which case it is looked up in data and then the calling environment.

time_points
numeric vector with times displayed in the time point analysis, if NULL this section of the table will be omitted

time_unit
a string with the unit of the tte argument

ties
passed forward to coxph

Value

an rtable object

Details

The time to event section is derived from a kaplan meier estimator for the formula argument with the strata special dropped.

The stratified and unstratified analysis is evaluated pair-wise (reference to comparison) and survdiff is used to get the p-value whereas coxph is used to calculate the hazard ratio and confidence interval.

The time point analysis is based on the kaplan meier estimator.

See also

Examples

library(random.cdisc.data)
ADSL <- radsl(seed = 1)

ADTTE <- radtte(ADSL, seed = 2)
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")

tbl <- t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM) + strata(SEX),
  data = ADTTE_f,
  event_descr = factor(EVNTDESC),
  time_points = c(6, 2000),
  time_unit = "month"
)

tbl

#>                                             A: Drug X               B: Placebo            C: Combination 
#>                                              (N=142)                 (N=133)                 (N=125)     
#> ---------------------------------------------------------------------------------------------------------
#> Patients with event (%)                    97 (68.31%)             112 (84.21%)             76 (60.8%)   
#>   Earliest Contributing Event
#>     Adverse Event                               23                      20                      13       
#>     Death                                       17                      18                      16       
#>     Disease Progression                         15                      19                      13       
#>     Last Date Known To Be Alive                 23                      27                      20       
#>     Last Tumor Assessment                       19                      28                      14       
#> Patients without event (%)                 45 (31.69%)             21 (15.79%)              49 (39.2%)   
#> 
#> Time to Event (months)
#>   Median                                      79.52                   72.34                   64.69      
#>     95% CI                                (60.6, 119.2)            (55.7, 94.9)            (54.6, 87.7)  
#>   25% and 75%-ile                          36.1, 159.8             31.8, 144.7             35.2, 150.3   
#>   Range (censored)                         0.2 to 333.3            3.7 to 318.9             0.3 to 316   
#>   Range (event)                             0.8 to 389             0.2 to 491.5              1 to 248    
#> 
#> Unstratified Analysis
#>   p-value (log-rank)                                                  0.2414                  0.3386     
#> 
#>   Hazard Ratio                                                        1.1769                   1.16      
#>     95% CI                                                       (0.8959, 1.5461)        (0.8557, 1.5724)
#> 
#> Stratified Analysis
#>   p-value (log-rank)                                                  0.1348                  0.3446     
#> 
#>   Hazard Ratio                                                        1.2347                  1.1591     
#>     95% CI                                                       (0.9362, 1.6283)        (0.8532, 1.5748)
#> 
#> Time Point Analysis
#>   6 months
#>     Patients remaining at risk                 130                     127                     110       
#>     Event Free Rate (%)                       94.26%                  96.23%                  92.63%     
#>       95% CI                              (90.4, 98.12)           (92.99, 99.47)           (88, 97.27)   
#>     Difference in Event Free Rate                                      1.97                   -1.63      
#>       95% CI                                                      (-3.08, 7.01)            (-7.66, 4.4)  
#>     p-value (Z-test)                                                  0.4447                  0.5964     
#> 
#>   2000 months
#>     -- no data

# NOT RUN {
Viewer(tbl)
# }