Time to Event Table

The time to event table summarizes time to event data with different models as described in the details section.

t_tte(formula, data, event_descr, time_points, time_unit = "month",
  ties = "exact")

Arguments

formula
a survival formula, the arm variable needs to be wrapped in arm. The strata special will only be used for the stratified analysis. If there is not strata specification then the stratified analysis is omitted.

data
a data.frame with all the variable that are used in formula

event_descr
a factor that partitions the the events into earliest contributing event. The variable name can be provided unquoted in which case it is looked up in data and then the calling environment.

time_points
numeric vector with times displayed in the time point analysis, if NULL this section of the table will be omitted

time_unit
a string with the unit of the tte argument

ties
passed forward to coxph

Value

an rtable object

Details

The time to event section is derived from a kaplan meier estimator for the formula argument with the strata special dropped.

The stratified and unstratified analysis is evaluated pair-wise (reference to comparison) and survdiff is used to get the p-value whereas coxph is used to calculate the hazard ratio and confidence interval.

The time point analysis is based on the kaplan meier estimator.

See also

Examples

library(random.cdisc.data)
ADSL <- radsl()

ADTTE <- radtte(ADSL)
ADTTE_f <- subset(ADTTE, PARAMCD == "OS")

tbl <- t_tte(
  formula = Surv(AVAL, !CNSR) ~ arm(ARM) + strata(SEX),
  data = ADTTE_f,
  event_descr = factor(EVNTDESC),
  time_points = c(6, 2000),
  time_unit = "month"
)

tbl

#>                                             A: Drug X               B: Placebo            C: Combination 
#>                                              (N=136)                 (N=127)                 (N=137)     
#> ---------------------------------------------------------------------------------------------------------
#> Patients with event (%)                    93 (68.38%)             97 (76.38%)             72 (52.55%)   
#>   Earliest Contributing Event
#>     Adverse Event                               16                      18                      18       
#>     Death                                       22                      22                      8        
#>     Disease Progression                         14                      26                      17       
#>     Last Date Known To Be Alive                 21                      15                      17       
#>     Last Tumor Assessment                       20                      16                      12       
#> Patients without event (%)                 43 (31.62%)             30 (23.62%)             65 (47.45%)   
#> 
#> Time to Event (months)
#>   Median                                      72.87                   78.67                   70.63      
#>     95% CI                                 (57.2, 97.1)            (62.1, 98.4)           (53.3, 101.9)  
#>   25% and 75%-ile                          41.3, 128.1             33.4, 149.7             25.2, 152.6   
#>   Range (censored)                         0.4 to 417.7            1.7 to 370.5            0.4 to 165.9  
#>   Range (event)                            0.4 to 330.7            1.6 to 466.8             0 to 316.2   
#> 
#> Unstratified Analysis
#>   p-value (log-rank)                                                  0.4651                  0.8457     
#> 
#>   Hazard Ratio                                                        0.8988                  1.0315     
#>     95% CI                                                       (0.675, 1.1969)         (0.7548, 1.4095)
#> 
#> Stratified Analysis
#>   p-value (log-rank)                                                  0.538                   0.9427     
#> 
#>   Hazard Ratio                                                        0.9124                  1.0115     
#>     95% CI                                                       (0.6814, 1.2217)         (0.74, 1.3827) 
#> 
#> Time Point Analysis
#>   6 months
#>     Patients remaining at risk                 125                     120                     122       
#>     Event Free Rate (%)                       96.99%                  96.82%                  97.76%     
#>       95% CI                              (94.09, 99.9)           (93.75, 99.89)           (95.26, 100)  
#>     Difference in Event Free Rate                                     -0.17                    0.77      
#>       95% CI                                                       (-4.4, 4.05)           (-3.06, 4.61)  
#>     p-value (Z-test)                                                  0.9363                  0.6923     
#> 
#>   2000 months
#>     -- no data

# NOT RUN {
Viewer(tbl)
# }