Tabulation of MMRM Results

These functions can be used to produce tables from a fitted MMRM produced with fit_mmrm().

# S3 method for mmrm
as.rtable(x, type = c("fixed", "cov", "diagnostic"), ...)

h_mmrm_fixed(x, format = "xx.xxxx")

h_mmrm_cov(x, format = "xx.xxxx")

h_mmrm_diagnostic(x, format = "xx.xxxx")

# S3 method for mmrm
tidy(x)

s_mmrm_lsmeans(df, .in_ref_col, show_relative = c("reduction", "increase"))

a_mmrm_lsmeans(df, .in_ref_col, show_relative = c("reduction", "increase"))

s_mmrm_lsmeans_single(df)

a_mmrm_lsmeans_single(df)

summarize_lsmeans(
  lyt,
  arms = TRUE,
  ...,
  table_names = "lsmeans_summary",
  .stats = NULL,
  .formats = NULL,
  .indent_mods = NULL,
  .labels = NULL
)

Arguments

x
(mmrm)
the original MMRM fit object.

type
(string)
type of table which should be returned.

...
additional argument format for controlling the numeric format.

format
(string)
format for the numbers in the table.

df
(data frame)
data set containing all analysis variables.

.in_ref_col
(logical)
TRUE when working with the reference level, FALSE otherwise.

show_relative
should the "reduction" (control - treatment, default) or the "increase" (treatment - control) be shown for the relative change from baseline?

lyt
(layout)
input layout where analyses will be added to.

arms
(flag)
should treatment variable be considered when using summarize_lsmeans layout generating function.

table_names
(character)
this can be customized in case that the same vars are analyzed multiple times, to avoid warnings from rtables.

.stats
(character)
statistics to select for the table.

.formats
(named character or list)
formats for the statistics.

.indent_mods
(named integer)
indent modifiers for the labels.

.labels
(named character)
labels for the statistics (without indent).

Value

as.rtable.mmrm() returns the fixed effects, covariance estimate or diagnostic statistics tables.

Functions

as.rtable.mmrm: Produce simple MMRM tables via the generic as.rtable().

h_mmrm_fixed: Helper function to produce fixed effects table.

h_mmrm_cov: Helper function to produce covariance matrix table.

h_mmrm_diagnostic: Helper function to produce diagnostic statistics table.

tidy.mmrm: Helper method (for broom::tidy()) to prepare a data frame from an mmrm object containing the LS means and contrasts.

s_mmrm_lsmeans: Statistics function which is extracting estimates from a tidied LS means data frame.

a_mmrm_lsmeans: Formatted Analysis function which can be further customized by calling rtables::make_afun() on it. It is used as afun in rtables::analyze().

s_mmrm_lsmeans_single: Statistics function which is extracting estimates from a tidied LS means data frame when ARM is not considered in the model.

a_mmrm_lsmeans_single: Formatted Analysis function (when ARM is not considered in the model) which can be further customized by calling rtables::make_afun() on it. It is used as afun in rtables::analyze().

summarize_lsmeans: Analyze function for tabulating LS means estimates from tidied mmrm results.

Examples

library(dplyr)
library(lme4)

#> Loading required package: Matrix

dat <- sleepstudy %>%
  mutate(
    group = factor(rep(c("A", "B"), length = nrow(sleepstudy))),
    days_grouped = cut(
      Days,
      breaks = quantile(Days, probs = seq(0, 1, length = 5)),
      include.lowest = TRUE
    )
  )
result <- fit_mmrm(
  vars = list(
    response = "Reaction",
    covariates = c(),
    id = "Subject",
    arm = "group",
    visit = "days_grouped"
  ),
  data = dat,
  cor_struct = "compound-symmetry",
  parallel = TRUE
)
as.rtable(result, type = "cov", format = "xx.x")

#>              [0,2]    [0,2]    [0,2]    (2,4.5]   (2,4.5]   (4.5,7]   (4.5,7]   (4.5,7]   (7,9]    (7,9] 
#> ---------------------------------------------------------------------------------------------------------
#> X.0.2.        2361    1375.7   1375.7   1375.7    1375.7    1375.7    1375.7    1375.7    1375.7   1375.7
#> X.0.2..1     1375.7    2361    1375.7   1375.7    1375.7    1375.7    1375.7    1375.7    1375.7   1375.7
#> X.0.2..2     1375.7   1375.7    2361    1375.7    1375.7    1375.7    1375.7    1375.7    1375.7   1375.7
#> X.2.4.5.     1375.7   1375.7   1375.7    2361     1375.7    1375.7    1375.7    1375.7    1375.7   1375.7
#> X.2.4.5..1   1375.7   1375.7   1375.7   1375.7     2361     1375.7    1375.7    1375.7    1375.7   1375.7
#> X.4.5.7.     1375.7   1375.7   1375.7   1375.7    1375.7     2361     1375.7    1375.7    1375.7   1375.7
#> X.4.5.7..1   1375.7   1375.7   1375.7   1375.7    1375.7    1375.7     2361     1375.7    1375.7   1375.7
#> X.4.5.7..2   1375.7   1375.7   1375.7   1375.7    1375.7    1375.7    1375.7     2361     1375.7   1375.7
#> X.7.9.       1375.7   1375.7   1375.7   1375.7    1375.7    1375.7    1375.7    1375.7     2361    1375.7
#> X.7.9..1     1375.7   1375.7   1375.7   1375.7    1375.7    1375.7    1375.7    1375.7    1375.7    2361 

result_no_arm <- fit_mmrm(
  vars = list(
    response = "Reaction",
    covariates = c(),
    id = "Subject",
    visit = "days_grouped"
  ),
  data = dat,
  cor_struct = "compound-symmetry",
  parallel = TRUE
)
as.rtable(result_no_arm, type = "cov", format = "xx.x")

#>              [0,2]    [0,2]    [0,2]    (2,4.5]   (2,4.5]   (4.5,7]   (4.5,7]   (4.5,7]   (7,9]    (7,9] 
#> ---------------------------------------------------------------------------------------------------------
#> X.0.2.       2351.6   1376.7   1376.7   1376.7    1376.7    1376.7    1376.7    1376.7    1376.7   1376.7
#> X.0.2..1     1376.7   2351.6   1376.7   1376.7    1376.7    1376.7    1376.7    1376.7    1376.7   1376.7
#> X.0.2..2     1376.7   1376.7   2351.6   1376.7    1376.7    1376.7    1376.7    1376.7    1376.7   1376.7
#> X.2.4.5.     1376.7   1376.7   1376.7   2351.6    1376.7    1376.7    1376.7    1376.7    1376.7   1376.7
#> X.2.4.5..1   1376.7   1376.7   1376.7   1376.7    2351.6    1376.7    1376.7    1376.7    1376.7   1376.7
#> X.4.5.7.     1376.7   1376.7   1376.7   1376.7    1376.7    2351.6    1376.7    1376.7    1376.7   1376.7
#> X.4.5.7..1   1376.7   1376.7   1376.7   1376.7    1376.7    1376.7    2351.6    1376.7    1376.7   1376.7
#> X.4.5.7..2   1376.7   1376.7   1376.7   1376.7    1376.7    1376.7    1376.7    2351.6    1376.7   1376.7
#> X.7.9.       1376.7   1376.7   1376.7   1376.7    1376.7    1376.7    1376.7    1376.7    2351.6   1376.7
#> X.7.9..1     1376.7   1376.7   1376.7   1376.7    1376.7    1376.7    1376.7    1376.7    1376.7   2351.6

library(broom)
df <- tidy(result)
df_no_arm <- tidy(result_no_arm)

s_mmrm_lsmeans(df[8, ], .in_ref_col = FALSE)

#> $n
#> [1] 18
#> 
#> $adj_mean_se
#> [1] 350.85122  11.45285
#> 
#> $adj_mean_ci
#> [1] 327.7374 373.9650
#> attr(,"label")
#> [1] "95% CI"
#> 
#> $diff_mean_se
#> [1] 14.22172 10.46332
#> 
#> $diff_mean_ci
#> [1] -6.447392 34.890825
#> attr(,"label")
#> [1] "95% CI"
#> 
#> $change
#> [1] -0.04224739
#> attr(,"label")
#> [1] "Relative Reduction (%)"
#> 
#> $p_value
#> [1] 0.1760584
#> 

s_mmrm_lsmeans_single(df_no_arm[4, ])

#> $n
#> [1] 36
#> 
#> $adj_mean_se
#> [1] 343.74036  10.17671
#> 
#> $adj_mean_ci
#> [1] 322.8588 364.6219
#> attr(,"label")
#> [1] "95% CI"
#> 

dat_adsl <- dat %>%
  select(Subject, group) %>%
  unique
basic_table() %>%
  split_cols_by("group", ref_group = result$ref_level) %>%
  add_colcounts() %>%
  split_rows_by("days_grouped") %>%
  summarize_lsmeans(show_relative = "increase") %>%
  build_table(
    df = tidy(result),
    alt_counts_df = dat_adsl
  )

#>                                               A                    B         
#>                                             (N=18)               (N=18)      
#> -----------------------------------------------------------------------------
#> [0,2]                                                                        
#>   n                                           36                   18        
#>   Adjusted Mean (SE)                   261.007 (10.188)     264.496 (11.453) 
#>     95% CI                            (240.106, 281.908)   (241.382, 287.61) 
#>   Difference in Adjusted Means (SE)                          3.489 (9.062)   
#>     95% CI                                                 (-14.411, 21.389) 
#>     Relative Increase (%)                                         1.3%       
#>     p-value (MMRM)                                               0.7007      
#> (2,4.5]                                                                      
#>   n                                           18                   18        
#>   Adjusted Mean (SE)                   288.649 (11.453)     282.992 (11.453) 
#>     95% CI                            (265.536, 311.763)   (259.878, 306.106)
#>   Difference in Adjusted Means (SE)                         -5.657 (10.463)  
#>     95% CI                                                 (-26.327, 15.012) 
#>     Relative Increase (%)                                         -2%        
#>     p-value (MMRM)                                               0.5895      
#> (4.5,7]                                                                      
#>   n                                           18                   36        
#>   Adjusted Mean (SE)                   312.178 (11.453)     313.635 (10.188) 
#>     95% CI                            (289.064, 335.292)   (292.734, 334.535)
#>   Difference in Adjusted Means (SE)                          1.456 (9.062)   
#>     95% CI                                                 (-16.444, 19.356) 
#>     Relative Increase (%)                                         0.5%       
#>     p-value (MMRM)                                               0.8725      
#> (7,9]                                                                        
#>   n                                           18                   18        
#>   Adjusted Mean (SE)                   336.630 (11.453)     350.851 (11.453) 
#>     95% CI                            (313.516, 359.743)   (327.737, 373.965)
#>   Difference in Adjusted Means (SE)                         14.222 (10.463)  
#>     95% CI                                                  (-6.447, 34.891) 
#>     Relative Increase (%)                                         4.2%       
#>     p-value (MMRM)                                               0.1761      

basic_table() %>%
  split_rows_by("days_grouped") %>%
  summarize_lsmeans(arms = FALSE) %>%
  build_table(
    df = tidy(result_no_arm),
    alt_counts_df = dat_adsl
  )

#>                             all obs      
#> -----------------------------------------
#> [0,2]                                    
#>   n                            54        
#>   Adjusted Mean (SE)    262.170 (9.723)  
#>     95% CI             (242.036, 282.304)
#> (2,4.5]                                  
#>   n                            36        
#>   Adjusted Mean (SE)    285.821 (10.177) 
#>     95% CI             (264.939, 306.702)
#> (4.5,7]                                  
#>   n                            54        
#>   Adjusted Mean (SE)    313.149 (9.723)  
#>     95% CI             (293.015, 333.283)
#> (7,9]                                    
#>   n                            36        
#>   Adjusted Mean (SE)    343.740 (10.177) 
#>     95% CI             (322.859, 364.622)