Calculate statistics Pairwise with reference level

Calculate statistics Pairwise with reference level

tabulate_pairwise(x, col_by, FUN, row.name, format = NULL, indent = 0)

Arguments

x
a vector

col_by
a factor first level is taken as reference level

FUN
a function with two arguments, first argument is the subset of x and the second argument is a factor with two levels from col_by

row.name
string with row name

format
a valid format string or a format function for rcells. To get a list of all valid format strings use list_rcell_format_labels. If format is NULL then the elements of a cell get pasted separated by a comma.

indent
non-negative integer where 0 means that the row should not be indented

Value

an rtable

Examples

tabulate_pairwise(
  x = iris$Sepal.Length,
  col_by = iris$Species,
  FUN = function(xi, col_by_i) diff(tapply(xi, col_by_i, mean)),
  row.name = "diff mean"
)

#>                    setosa          versicolor        virginica 
#> ---------------------------------------------------------------
#> diff mean                             0.93             1.582   

tabulate_pairwise(
   x = data.frame(
        val = sample(c("X1", "X2"), 100, replace = TRUE),
        strata = sample(c("A", "B"), 100, replace = TRUE)
   ),
  col_by = factor(sample(c("BY1", "BY2", "BY3"), 100, replace = TRUE)),
  FUN = function(xi, col_by_i) {
            mhp <- mantelhaen.test(xi$val, col_by_i, xi$strata)
            mhp$p.value
            },
  row.name = "MH pvalue: stratified"
)

#>                                     BY1                      BY2                      BY3       
#> ------------------------------------------------------------------------------------------------
#> MH pvalue: stratified                                 0.619948820668433        0.86073671523636