Time to Event Table

This is the description of the function

time_to_event_table(time_to_event, event, arm, earliest_contributing_event,
  strata_data, time_points, time_unit = "month", ties = "exact")

Arguments

time_to_event
time to event data

is_event
boolean with TRUE if event and FALSE if censored

Value

a named list with one element per row where each element contains the data for the row

Examples

library(random.cdisc.data)
library(forcats)

ASL <- radam("ASL")
ATE <- merge(ASL, radam("ATE", ADSL = ASL))

ATE_f <- subset(ATE, PARAMCD == "OS")

tbl <- time_to_event_table(
  time_to_event = ATE_f$AVAL,
  event = ATE_f$CNSR == 0,
  earliest_contributing_event = ATE_f$EVNTDESC,
  arm = factor(ATE_f$ARM),
  strata_data = ATE_f %>% select(SEX, RACE),
  time_points = 6,
  time_unit = "month"
)

tbl

#>                                                                          ARM A                   ARM B      
#>                                                                          (N=54)                  (N=46)     
#> ------------------------------------------------------------------------------------------------------------
#> Patients with event (%)                                                 27 (50%)              21 (45.65%)   
#>   Earliest contributing event
#>     Alive                                                                  5                       6        
#>     Death                                                                  8                       9        
#>     Disease Progression                                                    9                       6        
#>     Disease Progression or Death Censored for Missing Visits               6                       8        
#>     First Response                                                         10                      6        
#>     Last tumor assessment                                                  6                       6        
#>     Randomization                                                          10                      5        
#> Patients without event (%)                                              27 (50%)              25 (54.35%)   
#> 
#> Time to Event (months)
#>   Median                                                                 99.224                 160.799     
#>     95% CI                                                           (63.8, 180.2)           (106.3, 225.7) 
#>   25% and 75%−ile                                                     32.1, 251.3             63.2, 231.1   
#>   Range                                                               1.2 to 501.2             4 to 384.5   
#> 
#> Unstratified Analysis
#>   p-value (log-rank)                                                                             0.4208     
#> 
#>   Hazard Ratio                                                                                   0.7902     
#>     95% CI                                                                                  (0.4448, 1.4037)
#> 
#> Stratified Analysis
#>   p-value (log-rank)                                                                             0.2512     
#> 
#>   Hazard Ratio                                                                                   0.6774     
#>     95% CI                                                                                  (0.3472, 1.3218)
#> 
#> Time Point Analysis
#>   6 months
#>     Patients remaining at risk                                             48                      44       
#>     Event Free Rate (%)                                                  96.04%                  97.78%     
#>       95% CI                                                          (90.66, 100)            (93.47, 100)  
#>     Difference in Event Free Rate                                                                 1.74      
#>       95% CI                                                                                 (-5.15, 8.63)  
#>     p-value (Z-test)                                                                             0.6209     

# NOT RUN {
library(atezo.data)
library(dplyr)
library(forcats)

ASL <- asl(com.roche.cdt30019.go29436.re)
ATE <- ate(com.roche.cdt30019.go29436.re)

tte_tbl_stream <- get_time_to_event_table(com.roche.cdt30019.go29436.re)
Viewer(tte_tbl_stream)

ATE_f <- ATE %>%
 filter(ITTFL == "Y", PARAMCD == "OS") %>%
 mutate(ARM_ANL = fct_relevel(ARM1, "DUMMY C", "DUMMY B", "DUMMY A"))

tbl <- time_to_event_table(
  time_to_event = ATE_f$AVAL,
  event = ATE_f$CNSR == 0,
  earliest_contributing_event = ATE_f$EVNTDESC,
  arm = ATE_f$ARM_ANL,
  strata_data = ATE_f %>% select(SEX, MLIVER, TCICLVL2),
  time_points = 6,
  time_unit = "month"
)

Viewer(tbl)

Viewer(tbl, tte_tbl_stream)

compare_rtables(tbl, tte_tbl_stream, comp.attr = FALSE)

tbl[13,1]
tte_tbl_stream[13,1]
# }